Copula transformation from an Exponential to a standard Normal distribution

Copula transformation from an Exponential to a standard Normal distribution

Usage

inv_exp_t(x, rate)

Arguments

x

values from an Exponential distribution, vector.

rate

rate of the Exponential distribution, scalar.

Value

values from a standard Normal distribution, vector same length as x

See also