Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample.GR.magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample.GR.magnitudes(n=100, beta.p=log(10), M0=2.5)
#>   [1] 2.951052 2.517751 3.262311 2.761735 2.572211 2.748229 3.082083 2.981837
#>   [9] 3.084726 3.017070 3.965078 2.677753 4.057012 3.767840 3.759855 3.229996
#>  [17] 2.887959 2.509857 2.555098 2.856710 2.718500 3.936816 2.763119 2.678506
#>  [25] 3.880528 2.613455 2.628142 3.467257 3.048572 3.052241 2.716938 3.142438
#>  [33] 2.590194 3.591001 2.613393 4.199242 2.713440 3.257295 3.130297 3.250555
#>  [41] 2.726538 3.124819 2.710872 2.554182 3.763783 2.946770 2.642187 3.195128
#>  [49] 2.834657 3.285251 2.503827 2.553280 2.585008 2.652971 2.643366 2.877713
#>  [57] 2.677703 2.697693 2.955373 2.911249 3.184472 3.230706 2.511016 3.104148
#>  [65] 3.397171 3.273349 2.507339 3.225107 2.688290 2.510148 2.811191 2.846604
#>  [73] 3.014292 2.893772 2.855041 2.894221 2.567907 3.013147 2.635766 2.601207
#>  [81] 2.916538 2.700897 2.715520 3.574893 3.000419 2.724885 2.779180 3.286212
#>  [89] 2.911153 2.668083 2.764333 2.645195 3.215647 3.222793 3.290826 2.632491
#>  [97] 2.816113 2.691619 2.669201 2.509982