Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   500.1445   2.595705
#> 2   503.0222   3.962625
#> 3   500.9908   2.650241
#> 4   684.7832   6.278675
#> 5   834.5312   3.534595
#> 6   500.0628   2.646332
#> 7   606.1127   2.512617
#> 8   523.8618   4.098005
#> 9   500.0283   2.659983
#> 10  513.5174   3.287048
#> 11  500.0514   2.682724
#> 12  503.5913   2.519925
#> 13  526.2965   3.265852
#> 14  516.2213   3.670657
#> 15  503.1130   2.728408
#> 16  500.0237   3.139859
#> 17  500.2841   3.161934
#> 18  515.0587   2.949678
#> 19  501.3967   4.113800
#> 20  543.0389   3.334431
#> 21  501.7146   2.541763
#> 22  518.5454   2.747014
#> 23  520.7581   2.547648
#> 24  500.1265   3.273549
#> 25  582.1418   2.808275
#> 26  741.7228   2.567974
#> 27  600.8604   2.821258
#> 28  549.9574   3.908040
#> 29  500.7734   2.767625
#> 30  503.3624   2.840795
#> 31  502.9346   2.886910
#> 32  506.9112   2.825016
#> 33  500.0322   2.848848
#> 34  521.9205   5.925717
#> 35  955.9170   3.218100
#> 36  506.8748   2.745055
#> 37  508.5620   3.308591
#> 38  590.2509   3.500565
#> 39  513.9351   2.512656
#> 40  517.4492   2.708942
#> 41  501.0839   3.841467
#> 42  547.0293   2.584805
#> 43  500.0554   2.552244
#> 44  500.4119   2.978447
#> 45  501.0117   2.807220
#> 46  500.1889   2.858498
#> 47  707.9037   3.228977
#> 48  501.4503   3.414654
#> 49  725.1352   2.715345
#> 50  629.0522   3.413356
#> 51  501.2250   4.328253
#> 52  500.9075   2.682980
#> 53  641.7304   3.047812
#> 54  500.1686   3.448500
#> 55  921.8307   2.605804
#> 56  500.9524   2.739857
#> 57  512.8728   2.626858
#> 58  500.3349   2.609280
#> 59  505.9636   2.639705
#> 60  500.0606   3.176869
#> 61  501.6038   2.557256
#> 62  500.3107   2.502575
#> 63  503.0122   2.637747
#> 64  515.0808   2.803536
#> 65  515.1819   2.508496
#> 66  719.4444   2.601842
#> 67  925.7278   2.775147
#> 68  500.0039   2.511064
#> 69  500.0109   2.938412
#> 70  500.7723   2.990892
#> 71  506.0450   2.956004
#> 72  515.0187   3.690086
#> 73  508.7494   2.746397
#> 74  502.3413   2.674990
#> 75  525.9357   3.047594
#> 76  504.8220   2.644188
#> 77  553.6033   3.235350
#> 78  500.2842   4.735906
#> 79  867.4921   2.501887
#> 80  500.2943   2.564496
#> 81  526.8301   3.029997
#> 82  613.4612   3.117563
#> 83  500.1767   2.594777
#> 84  501.5593   3.231615
#> 85  500.9216   3.367359
#> 86  518.0402   2.790461
#> 87  500.5089   3.586291
#> 88  502.9153   2.743666
#> 89  500.5677   2.530390
#> 90  500.8191   2.918502
#> 91  500.1837   3.381243
#> 92  500.0120   3.485153
#> 93  501.4772   2.617538
#> 94  500.5231   2.884974
#> 95  547.8542   2.572768
#> 96  501.2121   3.075135
#> 97  500.1097   4.184473
#> 98  500.6208   2.791875
#> 99  502.2276   3.202420
#> 100 500.0468   2.985594
#> 101 550.6672   3.209537
#> 102 501.0620   3.109446
#> 103 501.2330   3.004163
#> 104 545.1454   2.941859
#> 105 575.3785   2.628275
#> 106 500.1060   2.859936
#> 107 500.2963   2.615161
#> 108 504.5803   2.738120
#> 109 731.4749   4.114868
#> 110 509.1951   2.655721
#> 111 502.7297   2.747220
#> 112 500.3275   2.576155
#> 113 500.6539   3.902728
#> 114 500.2954   3.398695
#> 115 500.0043   2.748708
#> 116 501.0781   2.745140
#> 117 504.5267   2.777853
#> 118 532.3882   2.537685
#> 119 506.6964   4.524239
#> 120 500.0761   3.587271
#> 121 500.5206   3.705967
#> 122 774.4289   3.254642
#> 123 529.6088   3.109218
#> 124 500.0547   2.805212
#> 125 500.2322   2.572092
#> 126 500.0450   3.173127
#> 127 500.6315   2.969817
#> 128 500.0063   2.581969
#> 129 503.6928   2.530173
#> 130 502.6900   2.765539
#> 131 508.1218   2.523131
#> 132 500.2042   3.196645
#> 133 503.2469   2.645813
#> 134 625.4334   3.231364
#> 135 522.8832   2.586381
#> 136 500.4209   3.426975
#> 137 500.0135   2.796463
#> 138 554.6244   3.003719
#> 139 518.1109   2.730601
#> 140 504.1818   2.500245
#> 141 644.7904   3.637263
#> 142 507.2718   2.823167
#> 143 598.8764   2.659806
#> 144 500.5484   2.529877
#> 145 665.4192   2.602809
#> 146 503.5492   3.205627
#> 147 784.6145   2.540598
#> 148 594.3221   2.710376
#> 149 531.8496   2.516698
#> 150 547.6042   2.941895
#> 151 500.8479   2.920356
#> 152 697.9851   2.818272
#> 153 507.1210   2.572764
#> 154 500.4773   2.667682
#> 155 748.3703   2.839388
#> 156 650.1676   3.276466
#> 157 523.7309   2.628065
#> 158 996.9888   2.681343
#> 159 521.1723   2.965735
#> 160 505.8542   2.560872
#> 161 743.9679   3.615335
#> 162 500.3636   2.893726
#> 163 506.1738   2.747591
#> 164 580.8099   2.835417
#> 165 505.2268   3.263530
#> 166 591.4223   2.787687
#> 167 500.1912   2.508489
#> 168 500.1718   3.297916
#> 169 500.1572   3.404104
#> 170 500.4756   2.608254
#> 171 703.0345   3.577653
#> 172 501.5895   2.558854
#> 173 558.3118   2.795368
#> 174 506.5139   3.788691
#> 175 502.3534   3.381475
#> 176 500.6540   2.871208
#> 177 718.0800   2.537543
#> 178 501.0912   2.705559
#> 179 756.0110   3.081541
#> 180 725.9611   2.987528
#> 181 517.1323   2.543233
#> 182 503.5022   2.518861
#> 183 500.2116   3.565664
#> 184 500.1554   2.794451
#> 185 515.3421   2.711121
#> 186 500.2773   2.765058
#> 187 501.4442   2.852949
#> 188 924.6563   2.555420
#> 189 509.7962   2.817427
#> 190 515.9590   3.319962
#> 191 500.2409   2.647542
#> 192 938.7019   2.965851
#> 193 813.9643   3.221187
#> 194 500.2026   3.351878
#> 195 564.6286   2.616488
#> 196 500.8146   3.176364
#> 197 504.8922   2.663986
#> 198 500.0760   2.665894
#> 199 505.5388   2.807791
#> 200 500.6988   3.015467
#> 201 500.2776   2.654978
#> 202 501.7976   4.683881
#> 203 521.4058   3.800139
#> 204 969.2014   2.817744
#> 205 678.4644   2.515190
#> 206 521.9619   2.650946
#> 207 501.9439   3.484823
#> 208 502.5858   3.306545
#> 209 517.3460   3.282417
#> 210 644.8633   2.739717
#> 211 844.2740   2.540635
#> 212 500.0849   2.571521
#> 213 505.2427   3.794351
#> 214 541.4267   2.800599
#> 215 618.8709   3.241464
#> 216 500.1254   2.733237
#> 217 530.6290   2.607321
#> 218 500.1432   2.678420
#> 219 510.6269   2.638452
#> 220 500.3751   2.555558
#> 221 505.7939   2.546939
#> 222 501.2606   3.156045
#> 223 500.9841   2.554906
#> 224 508.5731   2.552491
#> 225 500.3940   3.186025
#> 226 507.9527   3.048254
#> 227 501.3121   3.381695
#> 228 500.9469   2.970810
#> 229 917.9590   2.599906
#> 230 504.1990   2.572102
#> 231 527.8752   2.508154
#> 232 525.2179   3.003888
#> 233 500.0686   2.846503
#> 234 523.9503   3.752082
#> 235 500.1300   2.952238
#> 236 561.0529   3.305488
#> 237 501.9295   3.192028
#> 238 505.3901   2.816556
#> 239 864.5098   3.285882
#> 240 563.4172   2.675126
#> 241 523.4207   3.081212
#> 242 559.6127   2.683751
#> 243 500.0618   2.500807
#> 244 500.0721   2.688390
#> 245 687.4353   3.183526
#> 246 502.5049   3.099789
#> 247 500.1579   2.530999
#> 248 507.5423   2.789089
#> 249 500.3494   2.660148
#> 250 500.3380   2.706885
#> 251 502.6328   3.022428
#> 252 502.1110   2.927644
#> 253 516.6416   2.708907
#> 254 500.2644   2.627297
#> 255 500.9055   2.508822
#> 256 539.9303   4.497929
#> 257 500.0261   2.906647
#> 258 609.6518   3.209685
#> 259 812.1282   3.287568
#> 260 520.6200   2.973947
#> 261 500.8214   2.890483
#> 262 500.1078   2.597663
#> 263 501.5605   2.618015
#> 264 511.3743   2.607096
#> 265 500.1943   2.689927
#> 266 636.8504   2.552029
#> 267 518.2169   3.124931
#> 268 500.2041   2.738304
#> 269 534.3893   2.556655
#> 270 507.1871   2.677506
#> 271 500.0980   2.630775
#> 272 500.1081   2.573945
#> 273 514.2243   2.569141
#> 274 501.0283   2.512007
#> 275 924.5091   3.358331
#> 276 614.7762   2.537032
#> 277 508.4274   2.614355
#> 278 661.3738   3.598587
#> 279 506.3931   4.495863
#> 280 839.8309   2.773329
#> 281 518.7075   2.538646
#> 282 532.7785   3.007672
#> 283 500.3868   2.513558
#> 284 532.2760   2.581168
#> 285 505.2590   2.692187
#> 286 501.0050   2.838054
#> 287 502.7435   2.782367
#> 288 509.8310   3.273508
#> 289 650.7578   2.688764
#> 290 500.1518   2.993816
#> 291 500.2542   2.762067
#> 292 504.0656   2.894717
#> 293 522.3142   3.978396
#> 294 500.1014   3.357332
#> 295 555.1256   3.768912
#> 296 504.0290   2.962710
#> 297 504.7356   2.585776
#> 298 507.4217   3.007687
#> 299 501.9410   2.555987
#> 300 500.3744   2.564532
#> 301 500.4581   2.967785
#> 302 504.3464   2.776667
#> 303 504.4910   2.642019
#> 304 505.7853   2.930285
#> 305 515.2905   3.023415
#> 306 500.0844   2.769412
#> 307 500.7277   2.567046
#> 308 564.9298   2.519327
#> 309 500.0551   2.746193
#> 310 507.8528   2.812407
#> 311 504.5031   3.198633
#> 312 517.2369   2.811591
#> 313 501.9804   4.650724
#> 314 504.5731   2.885476
#> 315 501.8393   2.823103
#> 316 504.3696   3.150553
#> 317 507.7455   2.503886
#> 318 500.3501   4.131781
#> 319 765.0890   3.162191
#> 320 528.5472   2.518813
#> 321 689.6735   2.963275
#> 322 923.3039   3.293124
#> 323 505.3783   2.821100
#> 324 505.6532   3.105463
#> 325 505.0992   3.006007
#> 326 503.2385   3.779892
#> 327 585.4050   2.697577
#> 328 542.2318   3.251986
#> 329 535.2513   3.095676
#> 330 513.7712   2.552795
#> 331 512.3694   2.607161
#> 332 503.0852   2.708040
#> 333 504.7830   3.324836
#> 334 526.7314   3.890561
#> 335 631.2467   2.529080
#> 336 501.1822   2.518243
#> 337 507.0753   2.581116
#> 338 502.1695   3.653767
#> 339 500.0662   2.620749
#> 340 500.1150   2.857291
#> 341 505.9046   3.186640
#> 342 731.3847   2.699396
#> 343 500.9108   3.471327
#> 344 500.0840   3.397828
#> 345 500.0165   2.867350
#> 346 653.2478   3.480261
#> 347 515.1946   3.560843
#> 348 610.7193   2.537937
#> 349 500.1748   3.870002
#> 350 500.1078   3.018354
#> 351 554.2926   2.512729
#> 352 502.0846   4.014414
#> 353 501.8289   2.893018
#> 354 704.1225   2.506635
#> 355 889.7952   2.795932
#> 356 506.2174   3.146190
#> 357 508.2776   3.412089
#> 358 517.8056   2.777192
#> 359 500.2953   2.538962
#> 360 500.5440   2.557428
#> 361 500.1392   2.772374
#> 362 845.9034   2.504631
#> 363 502.5144   2.546224
#> 364 508.5010   3.671666
#> 365 884.3817   2.556350
#> 366 556.9981   2.603995
#> 367 799.0432   2.877772
#> 368 500.6700   2.521773
#> 369 520.0188   2.850549
#> 370 500.5023   3.082195
#> 371 500.5514   2.723796
#> 372 693.3634   2.749111
#> 373 504.9293   2.533933
#> 374 501.1577   3.590953
#> 375 512.6457   3.405092
#> 376 500.5210   3.080482
#> 377 835.3236   3.335654
#> 378 500.2265   2.529106
#> 379 501.8961   3.435749
#> 380 514.3639   3.286604
#> 381 500.5993   2.555863
#> 382 546.8250   3.072020
#> 383 504.3923   2.800255
#> 384 500.3332   2.662916
#> 385 512.6329   2.777107
#> 386 500.6080   2.630987
#> 387 501.3353   2.627735
#> 388 504.8904   3.978281
#> 389 655.0771   2.691767
#> 390 503.8207   2.981112
#> 391 501.7019   2.791369
#> 392 500.1037   2.881398
#> 393 589.2967   2.791253
#> 394 643.6983   2.731029
#> 395 501.6799   3.129433
#> 396 516.1650   2.735041
#> 397 500.5640   2.760325
#> 398 584.9350   2.784731
#> 399 500.1827   2.718826
#> 400 500.8808   2.566982
#> 401 512.7382   4.399562
#> 402 501.3827   3.138188
#> 403 540.9216   2.508789
#> 404 501.8984   2.757411
#> 405 500.7382   3.040211
#> 406 500.0140   2.660772
#> 407 500.7016   2.500629
#> 408 530.1412   2.653068
#> 409 526.3745   2.945242
#> 410 564.3737   2.757071
#> 411 500.2781   2.924408
#> 412 500.0053   3.679290
#> 413 500.2261   2.586159
#> 414 500.2068   2.548102
#> 415 500.0068   2.890633
#> 416 500.4766   2.779187
#> 417 500.1206   2.535773
#> 418 570.9273   2.523574
#> 419 685.4188   2.609452
#> 420 502.5070   3.406523
#> 421 501.2974   2.888285
#> 422 501.2912   2.610189
#> 423 518.2579   3.134637
#> 424 500.4224   2.710212
#> 425 500.1469   2.627636
#> 426 504.6266   3.136865
#> 427 500.0240   2.773853
#> 428 500.4822   2.539858
#> 429 535.1694   4.066572
#> 430 500.9906   3.444244
#> 431 500.6900   3.158963
#> 432 507.8978   2.525312
#> 433 503.3733   2.650247
#> 434 500.0904   2.814928
#> 435 500.1259   3.295816
#> 436 501.0426   2.795517
#> 437 500.5850   3.783063
#> 438 503.9557   2.721528
#> 439 540.5176   2.582360
#> 440 645.5399   3.631512
#> 441 635.0739   2.789615
#> 442 500.3747   2.591869
#> 443 502.2151   2.605015
#> 444 500.9249   2.566002
#> 445 509.3895   2.843461
#> 446 516.5783   4.056932
#> 447 501.2621   2.519859
#> 448 501.1557   2.809506
#> 449 759.2999   2.888788
#> 450 500.6217   2.820911
#> 451 506.6963   2.880687
#> 452 500.0238   3.150657
#> 453 500.1728   2.843955
#> 454 540.2661   2.829409
#> 455 528.8237   3.230259
#> 456 653.5357   3.000747
#> 457 503.5872   3.866162
#> 458 500.2182   2.587548
#> 459 501.2582   3.489765
#> 460 500.4904   2.722705
#> 461 500.0164   3.277002
#> 462 500.1909   2.789439
#> 463 506.8555   3.337003
#> 464 508.3938   3.586261
#> 465 500.0071   3.790670
#> 466 851.1560   3.726249
#> 467 500.0148   5.031110
#> 468 603.0988   4.322832
#> 469 502.7686   3.194785
#> 470 505.1000   2.603482
#> 471 505.6719   2.689450
#> 472 501.9080   2.694823
#> 473 500.3077   2.750940
#> 474 500.1049   2.774732
#> 475 500.1626   3.297780
#> 476 500.1339   2.546057
#> 477 570.0206   3.444746
#> 478 500.2370   2.890198
#> 479 500.0058   2.592764
#> 480 897.4790   2.751389
#> 481 500.1059   2.806247
#> 482 903.7323   2.734303
#> 483 501.1389   2.582548
#> 484 888.4109   3.538379
#> 485 522.1228   2.792852
#> 486 503.5836   2.855262
#> 487 704.0705   2.561898
#> 488 500.1655   2.955263
#> 489 504.2595   2.801142
#> 490 518.7731   3.019551
#> 491 502.1034   2.509775
#> 492 706.4314   2.923660
#> 493 598.9357   3.556650
#> 494 672.2682   3.098103
#> 495 558.7162   2.605118
#> 496 531.6561   2.978036
#> 497 500.2691   2.600093
#> 498 505.0792   3.310041
#> 499 509.6520   3.887030
#> 500 528.4144   2.753782
#> 501 502.3738   3.063779
#> 502 509.5120   2.701744
#> 503 574.7948   3.112034
#> 504 500.3405   2.505761
#> 505 504.0172   3.188285
#> 506 514.7855   2.818504
#> 507 500.6190   3.533658
#> 508 774.4464   2.703328
#> 509 500.2061   2.870027
#> 510 500.1408   2.792999
#> 511 500.0168   2.519986
#> 512 710.8696   2.702510
#> 513 791.0115   2.699075
#> 514 517.6038   3.515400
#> 515 500.7557   3.030122
#> 516 685.1618   2.894453
#> 517 500.0398   2.851959
#> 518 502.8470   3.055296
#> 519 551.1212   3.156918
#> 520 634.4516   3.089351
#> 521 568.2645   3.155047
#> 522 500.3974   3.061375
#> 523 500.1297   2.946171
#> 524 503.0989   3.382467
#> 525 505.6552   2.613351
#> 526 503.5137   3.036766
#> 527 501.9921   4.093136
#> 528 500.9406   3.062844
#> 529 637.4783   4.120887
#> 530 510.8934   2.713475
#> 531 500.9150   2.563578
#> 532 584.1626   2.671258
#> 533 514.0673   2.802246
#> 534 500.7071   4.621692
#> 535 500.1109   2.997892
#> 536 645.5834   2.771133
#> 537 584.4625   3.533808
#> 538 776.5845   2.752834
#> 539 500.4207   2.542180
#> 540 506.3565   2.639516
#> 541 718.2736   2.735380
#> 542 500.0448   2.585095
#> 543 500.2997   3.231523
#> 544 781.9502   2.617985
#> 545 501.1687   3.041308
#> 546 502.5678   2.655451
#> 547 501.4245   3.083250
#> 548 500.2258   3.010694
#> 549 500.0662   2.662713
#> 550 502.7654   2.641082
#> 551 500.0957   2.684856
#> 552 500.3847   3.037369
#> 553 500.5872   3.056615
#> 554 500.0063   3.332910
#> 555 500.1199   2.558812
#> 556 500.0461   3.382568
#> 557 500.2284   2.604126
#> 558 863.1336   3.140912
#> 559 516.9402   2.617516
#> 560 682.0795   2.714107
#> 561 504.6052   2.913066
#> 562 501.5058   2.546529
#> 563 500.0863   2.701887
#> 564 500.0358   2.612725
#> 565 500.2831   2.949161
#> 566 513.4262   2.713488
#> 567 500.6789   2.736745
#> 568 501.5960   3.119163
#> 569 553.7046   3.558396
#> 570 500.2012   3.406560
#> 571 501.7273   2.642342
#> 572 500.6112   4.004553
#> 573 549.9928   2.993953
#> 574 536.6022   2.629589
#> 575 513.1049   3.148852
#> 576 501.7021   3.488535
#> 577 501.5090   2.726894
#> 578 502.4735   2.549821
#> 579 788.9485   2.836525
#> 580 529.0970   3.638196
#> 581 505.2213   2.722785
#> 582 501.1364   2.708570
#> 583 503.9357   2.705956
#> 584 500.2095   2.565915
#> 585 645.3803   2.827409
#> 586 503.5366   2.972593
#> 587 500.3615   2.619932
#> 588 501.8487   2.635190
#> 589 513.2653   3.225770
#> 590 541.5708   3.385924
#> 591 684.8274   2.854219
#> 592 500.0034   3.184554
#> 593 627.4891   2.943703
#> 594 507.7104   2.515706
#> 595 506.0303   2.828239
#> 596 500.4976   2.684461
#> 597 924.2047   2.589464
#> 598 500.4879   3.602717
#> 599 500.0872   3.015635
#> 600 500.4566   2.829448
#> 601 509.9738   2.662407
#> 602 501.4792   3.122899
#> 603 500.9868   2.757062
#> 604 500.0430   3.670199
#> 605 507.3882   3.070904
#> 606 500.1534   2.897948
#> 607 500.3845   2.500098
#> 608 503.6899   2.681888
#> 609 500.2183   2.821300
#> 610 577.6334   3.555773
#> 611 502.5608   2.775819
#> 612 636.5272   2.764236
#> 613 550.1502   3.672664
#> 614 954.1871   3.559443
#> 615 501.1573   3.809482
#> 616 986.3867   2.683279
#> 617 534.0354   2.507483
#> 618 523.4829   2.552904
#> 619 590.0064   3.197332
#> 620 501.2860   2.591879
#> 621 500.0632   2.527591
#> 622 500.8818   2.738021
#> 623 504.7692   3.019965
#> 624 568.6117   2.554618
#> 625 500.1997   2.628751
#> 626 533.8372   2.773736
#> 627 593.7129   2.585593
#> 628 501.4023   2.942975
#> 629 530.1652   3.089171
#> 630 522.3927   2.622757
#> 631 553.2716   2.614065
#> 632 525.0235   3.162644
#> 633 520.6498   2.540104
#> 634 522.9515   2.831242
#> 635 500.2536   3.928980
#> 636 519.2513   2.837503
#> 637 796.3506   3.215037
#> 638 500.0885   2.693720
#> 639 501.6654   4.129798
#> 640 504.7938   2.764893
#> 641 642.6121   2.522153
#> 642 501.0562   3.413900
#> 643 501.1600   2.765840
#> 644 500.0237   2.573687
#> 645 500.0223   2.841880
#> 646 680.9983   2.660471
#> 647 500.0073   2.760830
#> 648 503.0537   4.930447
#> 649 500.0336   3.886689
#> 650 503.7247   3.564956
#> 651 725.9411   3.258698
#> 652 501.1721   2.593571
#> 653 501.3759   2.771507
#> 654 502.3114   2.505485
#> 655 500.1331   3.951605
#> 656 755.5547   2.994158
#> 657 506.4726   4.825923
#> 658 667.3831   3.132777
#> 659 664.3685   2.760716
#> 660 501.0383   3.414331
#> 661 502.4975   2.530470
#> 662 552.4011   2.624185
#> 663 503.1717   2.768580
#> 664 516.7074   2.620989
#> 665 501.2530   2.535695
#> 666 500.5523   2.956301
#> 667 520.9606   2.650663
#> 668 500.6006   2.575366
#> 669 500.0116   2.656653
#> 670 508.7702   3.265251
#> 671 667.6781   3.260592
#> 672 500.0575   3.527649
#> 673 530.7839   3.278988
#> 674 501.8288   2.923270
#> 675 500.5016   3.160251
#> 676 505.6555   2.717064
#> 677 500.0436   2.588924
#> 678 519.8528   2.701851
#> 679 500.9480   3.192662
#> 680 511.7814   2.822992
#> 681 500.0787   2.983148
#> 682 500.1879   2.825023
#> 683 517.8458   2.805416
#> 684 537.1429   3.272207
#> 685 500.0035   3.233509
#> 686 502.8646   2.722065
#> 687 501.0085   2.855174
#> 688 529.4324   2.846729
#> 689 574.2406   3.028131
#> 690 502.1060   3.476865
#> 691 500.1946   3.044643
#> 692 500.1026   2.633016
#> 693 500.0340   2.558490
#> 694 644.7429   3.501483
#> 695 500.5988   3.896168
#> 696 839.5979   2.712358
#> 697 601.2795   2.706671
#> 698 500.9285   2.901470
#> 699 592.3069   2.901037
#> 700 519.7596   2.838274
#> 701 573.3415   3.111699
#> 702 500.7106   2.670036
#> 703 731.9169   2.541586
#> 704 500.0431   3.311267
#> 705 500.2931   2.528028
#> 706 500.0914   2.817363
#> 707 504.8301   3.090268
#> 708 503.6879   2.563228
#> 709 500.8858   2.562128
#> 710 689.6046   2.549002
#> 711 500.5520   3.051422
#> 712 500.0380   2.697132
#> 713 501.0827   2.524948
#> 714 501.7661   2.557791
#> 715 563.9836   2.826365
#> 716 501.8688   4.418085
#> 717 506.1712   3.186754
#> 718 500.1565   2.945995
#> 719 520.1685   2.696058
#> 720 533.6814   2.785309
#> 721 500.6893   2.525302
#> 722 647.4237   2.656621
#> 723 501.9579   3.101293
#> 724 514.1626   3.020837
#> 725 502.0136   2.555088
#> 726 500.0774   2.833317
#> 727 501.4604   2.749315
#> 728 533.5070   2.648219
#> 729 501.9611   3.643992
#> 730 500.0302   2.675051
#> 731 500.3889   3.198143
#> 732 526.8054   2.647663
#> 733 505.5940   3.173965
#> 734 505.0749   2.526761
#> 735 527.3764   3.082736
#> 736 503.9721   2.648771
#> 737 545.8191   2.812214
#> 738 700.0428   3.097541
#> 739 500.0987   2.940998
#> 740 500.3771   2.718689
#> 741 513.3814   2.819740
#> 742 503.1237   2.891572
#> 743 556.7733   3.715339
#> 744 956.1182   2.662038
#> 745 674.5788   2.714741
#> 746 651.2138   2.724072
#> 747 500.2636   3.029145
#> 748 508.1572   2.772271
#> 749 503.1264   2.655760
#> 750 500.6868   4.130425
#> 751 500.7653   2.565526
#> 752 500.9266   2.504306
#> 753 501.0377   2.758441
#> 754 768.1599   3.970003
#> 755 648.6069   2.584142
#> 756 501.0747   3.212444
#> 757 505.7792   2.529936
#> 758 535.1111   2.619414
#> 759 864.7149   3.176273
#> 760 500.0366   2.540744
#> 761 618.4485   2.724210
#> 762 503.4597   3.498497
#> 763 511.8121   2.796852
#> 764 500.3647   2.748265
#> 765 506.7936   2.722773
#> 766 502.8888   2.763684
#> 767 515.1851   2.738085
#> 768 500.1378   3.148780
#> 769 500.4189   3.346211
#> 770 502.3838   2.659754
#> 771 503.8549   2.524034
#> 772 546.7640   3.121122
#> 773 500.3653   2.711048
#> 774 500.0812   2.502423
#> 775 500.6397   2.583003
#> 776 567.4601   3.019117
#> 777 732.1522   2.719320
#> 778 500.3905   2.801079
#> 779 627.3440   2.637587
#> 780 509.9169   2.516678
#> 781 500.3203   2.566700
#> 782 521.1957   3.062970
#> 783 506.8844   2.856581
#> 784 501.1661   3.144583
#> 785 510.6161   2.591080
#> 786 892.0379   2.636178
#> 787 694.0463   2.796394
#> 788 507.6817   2.636982
#> 789 500.2509   2.553355
#> 790 519.8998   2.839770
#> 791 521.6527   2.891001
#> 792 501.0656   2.623627
#> 793 503.3407   2.591750
#> 794 501.1167   2.919425
#> 795 500.1121   2.570115
#> 796 500.1645   2.779103
#> 797 500.2376   2.580054
#> 798 509.0001   3.534412
#> 799 600.9407   3.735113
#> 800 505.9777   2.543405
#> 801 500.4717   2.885641
#> 802 805.5049   2.642057
#> 803 662.7184   2.823749
#> 804 798.6965   2.787486
#> 805 500.1301   3.275909
#> 806 521.0647   2.998048
#> 807 500.1244   2.501575
#> 808 650.2082   2.782451
#> 809 501.5745   2.909522
#> 810 544.7135   4.254605
#> 811 523.9448   2.997780
#> 812 660.9385   2.611659
#> 813 501.1187   2.512040
#> 814 501.7089   4.239594
#> 815 909.6275   2.605364
#> 816 500.4318   2.578764
#> 817 500.2361   3.017769
#> 818 503.0997   2.759010
#> 819 616.9588   2.646442
#> 820 597.9909   2.624811
#> 821 514.7256   3.104756
#> 822 509.6627   2.632185
#> 823 500.3679   3.443349
#> 824 758.6937   2.501705
#> 825 500.0812   3.527684
#> 826 502.1300   3.396584
#> 827 977.8177   2.616832
#> 828 502.2495   2.587175
#> 829 520.3300   2.603539
#> 830 552.8090   2.616245
#> 831 501.1909   2.500490
#> 832 501.4912   2.632875
#> 833 500.0681   2.738362
#> 834 500.2358   2.683320
#> 835 500.0265   2.637498
#> 836 512.5978   2.518131
#> 837 529.4823   2.763832
#> 838 774.9861   3.213532
#> 839 501.9929   2.815276
#> 840 536.8549   2.745868
#> 841 500.4378   3.045365
#> 842 698.5188   3.679910
#> 843 508.4474   2.922922
#> 844 500.3261   3.209645
#> 845 504.4408   2.733059
#> 846 655.7651   2.691385
#> 847 527.2017   3.382798
#> 848 509.8153   2.668053
#> 849 529.4328   4.078035
#> 850 500.0266   3.038701
#> 851 532.9201   2.773234
#> 852 500.4456   2.561894
#> 853 636.0373   2.884072
#> 854 513.1427   2.780029
#> 855 607.6659   2.838620
#> 856 567.9169   2.639571
#> 857 508.0120   2.854304
#> 858 612.4562   2.517884
#> 859 500.4220   3.321424