Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample_GR_magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample_GR_magnitudes(n = 100, beta.p = log(10), M0 = 2.5)
#>   [1] 2.746146 2.562176 2.561313 3.984621 2.722299 2.594637 2.626181 3.057158
#>   [9] 2.752198 3.232922 2.858353 3.161408 2.660147 3.410272 2.547676 2.859701
#>  [17] 2.503060 3.191125 2.761455 3.676954 2.662438 3.484869 2.769538 2.875426
#>  [25] 2.713642 3.547245 2.847216 3.104771 2.587755 3.159540 2.841103 3.043215
#>  [33] 3.038719 2.506939 2.859663 2.620257 2.574499 3.236216 3.076653 2.508638
#>  [41] 2.695986 2.648841 2.559916 2.618111 2.857076 2.562750 2.506449 2.732399
#>  [49] 4.732754 2.733752 2.615634 2.713362 3.237883 2.969714 2.684804 3.924010
#>  [57] 2.639796 3.171622 3.110303 4.390560 2.588366 2.598878 2.888738 3.143304
#>  [65] 4.061472 3.164223 2.748119 2.981427 2.879014 2.543463 2.663403 3.089421
#>  [73] 3.198233 2.585408 3.109466 2.850106 2.508669 3.287061 2.965376 2.553902
#>  [81] 3.158213 2.697405 3.139631 2.619705 2.656498 2.522264 2.776095 2.618224
#>  [89] 3.010431 2.529279 2.530985 2.564467 2.646009 2.829461 2.574533 2.726321
#>  [97] 2.874516 2.511848 2.578883 2.541028