Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   511.5104   3.216161
#> 2   500.2119   2.577017
#> 3   500.2631   3.022529
#> 4   501.0166   2.540982
#> 5   793.4279   3.152873
#> 6   501.6994   2.994217
#> 7   500.0599   3.266710
#> 8   503.9822   2.564542
#> 9   517.5664   3.067743
#> 10  531.3203   2.911913
#> 11  502.7274   2.980515
#> 12  522.6113   3.390487
#> 13  502.5953   2.702909
#> 14  500.0062   2.751957
#> 15  506.4837   4.315351
#> 16  500.5863   2.562791
#> 17  535.1667   2.834229
#> 18  503.1536   2.700580
#> 19  503.4230   2.817330
#> 20  584.5985   2.982110
#> 21  570.4157   3.431468
#> 22  500.6393   2.517449
#> 23  502.7219   2.709232
#> 24  502.3093   2.900245
#> 25  793.3879   2.803483
#> 26  501.7084   2.614111
#> 27  500.1103   2.940362
#> 28  500.0105   2.791807
#> 29  584.7333   2.745584
#> 30  500.4649   2.775748
#> 31  500.2310   3.718891
#> 32  594.0769   3.085882
#> 33  778.7055   2.649213
#> 34  501.5423   2.567450
#> 35  500.0558   3.969120
#> 36  503.7817   2.836115
#> 37  500.1370   2.643095
#> 38  594.6082   3.010727
#> 39  502.0827   3.385293
#> 40  500.5550   4.633612
#> 41  502.0288   3.091795
#> 42  958.2125   3.134989
#> 43  500.0159   2.682975
#> 44  500.6614   3.190006
#> 45  500.4122   3.013843
#> 46  500.3123   3.111057
#> 47  765.6281   2.951632
#> 48  501.5783   3.025316
#> 49  500.1423   3.255327
#> 50  533.7273   2.611314
#> 51  548.5103   2.524530
#> 52  509.8525   2.688922
#> 53  505.6902   3.408392
#> 54  501.0595   3.241252
#> 55  502.7388   2.847734
#> 56  500.2254   2.590591
#> 57  531.3845   2.827626
#> 58  500.0180   3.413206
#> 59  594.1078   2.773103
#> 60  813.2661   2.573985
#> 61  759.6646   3.047879
#> 62  500.1394   3.008278
#> 63  524.0802   2.996702
#> 64  506.5707   3.068506
#> 65  500.0858   2.887708
#> 66  500.0799   3.184661
#> 67  656.7335   2.515270
#> 68  500.3193   3.309524
#> 69  562.2398   4.689544
#> 70  502.1679   3.103039
#> 71  518.4465   2.528322
#> 72  508.8701   3.417780
#> 73  529.9271   2.556189
#> 74  811.5626   2.814693
#> 75  501.7374   2.731150
#> 76  500.9562   2.528113
#> 77  514.7414   2.646292
#> 78  500.2057   3.153417
#> 79  500.7545   2.607288
#> 80  503.9051   2.895905
#> 81  500.4814   2.781493
#> 82  625.2084   2.511138
#> 83  623.1691   2.759510
#> 84  500.8296   2.583512
#> 85  560.9345   4.061661
#> 86  500.8226   2.523334
#> 87  517.0326   2.940151
#> 88  500.2257   2.560951
#> 89  509.3135   2.921668
#> 90  544.0172   2.655795
#> 91  500.9022   2.768523
#> 92  511.1522   2.769244
#> 93  505.6472   2.636748
#> 94  500.0458   2.609965
#> 95  503.1138   3.206608
#> 96  502.7639   2.716873
#> 97  500.6023   2.626089
#> 98  662.6035   2.705365
#> 99  500.2199   2.640873
#> 100 564.5665   2.884123
#> 101 500.4226   2.537523
#> 102 500.0465   2.591192
#> 103 501.0253   2.576793
#> 104 671.4808   2.534024
#> 105 500.0384   3.100313
#> 106 500.1720   2.826050
#> 107 500.0309   2.535450
#> 108 980.9079   2.672086
#> 109 502.1665   2.583070
#> 110 517.0649   2.804942
#> 111 626.9252   3.103584
#> 112 527.6743   2.591619
#> 113 502.2212   2.879076
#> 114 535.8690   3.042198
#> 115 730.1615   2.559771
#> 116 500.1093   3.268053
#> 117 501.9401   4.074575
#> 118 500.4998   3.687221
#> 119 506.0732   2.530701
#> 120 500.0734   3.393489
#> 121 755.3559   3.508916
#> 122 552.4004   3.108562
#> 123 500.8567   2.561628
#> 124 503.4295   3.666922
#> 125 503.1608   6.648042
#> 126 500.0099   2.791841
#> 127 500.1945   2.593896
#> 128 508.0617   2.595982
#> 129 500.0001   2.620966
#> 130 782.2195   2.672548
#> 131 500.0360   2.846353
#> 132 501.1853   2.912856
#> 133 501.4160   3.079759
#> 134 511.0176   2.553691
#> 135 500.2268   4.348430
#> 136 734.5950   3.526905
#> 137 649.8215   2.904409
#> 138 500.5394   2.877411
#> 139 500.7836   2.674856
#> 140 531.8149   3.085803
#> 141 508.5288   3.173065
#> 142 549.9360   2.688034
#> 143 616.4628   2.730718
#> 144 545.5275   2.931302
#> 145 508.5623   3.807732
#> 146 500.1169   2.536227
#> 147 500.3930   3.178631
#> 148 500.0392   3.420284
#> 149 500.2266   2.633629
#> 150 502.8925   2.763390
#> 151 500.0168   2.613732
#> 152 506.6578   2.543740
#> 153 500.4123   3.040966
#> 154 500.2851   2.686035
#> 155 663.4040   2.643265
#> 156 555.4141   2.567252
#> 157 504.7608   3.319503
#> 158 682.3131   3.145208
#> 159 515.4939   2.595347
#> 160 598.0124   2.529986
#> 161 581.7018   3.098903
#> 162 710.6423   3.135897
#> 163 500.0555   3.160539
#> 164 500.2914   2.558172
#> 165 512.4951   3.634159
#> 166 847.7388   2.606043
#> 167 818.4063   2.509009
#> 168 500.2592   3.942870
#> 169 536.8360   3.369169
#> 170 548.8185   2.835620
#> 171 518.2714   2.938121
#> 172 504.1649   3.635914
#> 173 502.0500   2.583144
#> 174 514.2455   3.839350
#> 175 502.3922   2.693060
#> 176 572.1858   2.973899
#> 177 550.3904   2.819308
#> 178 847.8181   2.670526
#> 179 501.1379   3.627638
#> 180 538.3350   2.662009
#> 181 500.6572   2.955154
#> 182 502.5689   3.069827
#> 183 500.6509   3.509225
#> 184 564.6384   2.745342
#> 185 500.1031   3.349521
#> 186 503.7664   2.641993
#> 187 505.5664   2.770698
#> 188 500.2053   3.514719
#> 189 539.4019   2.648069
#> 190 689.7936   2.586500
#> 191 668.1849   3.441879
#> 192 759.4215   3.121966
#> 193 928.8793   2.509402
#> 194 500.5458   2.931673
#> 195 537.0561   3.185773
#> 196 500.7527   2.725797
#> 197 500.0047   2.670937
#> 198 507.4158   2.615381
#> 199 500.3425   2.864116
#> 200 779.1151   3.063487
#> 201 501.9759   2.599326
#> 202 536.5466   2.549255
#> 203 823.6121   2.950110
#> 204 501.1944   2.564882
#> 205 501.5281   2.932287
#> 206 503.6123   2.988504
#> 207 566.0458   2.929201
#> 208 500.4268   3.327919
#> 209 503.6735   2.627539
#> 210 500.0665   3.170930
#> 211 665.2159   2.795392
#> 212 540.2870   2.535862
#> 213 500.6434   2.600328
#> 214 500.6148   3.578669
#> 215 500.1366   2.670321
#> 216 547.9781   2.662258
#> 217 505.4931   2.576451
#> 218 501.7770   3.819759
#> 219 502.9158   2.503156
#> 220 500.5593   2.806969
#> 221 530.1772   2.640875
#> 222 969.0048   2.680890
#> 223 501.0399   2.602237
#> 224 503.0176   3.446158
#> 225 814.0108   2.817929
#> 226 500.2895   5.932841
#> 227 501.4923   3.385368
#> 228 516.0071   3.600709
#> 229 531.4598   2.559803
#> 230 500.0403   3.518334
#> 231 965.0523   2.635256
#> 232 766.9113   2.705394
#> 233 542.1637   3.143639
#> 234 502.5600   3.115340
#> 235 500.0208   3.513995
#> 236 502.0002   3.180506
#> 237 501.0476   3.774927
#> 238 527.5633   2.842798
#> 239 503.8212   3.082657
#> 240 502.6217   3.686858
#> 241 695.8796   2.582501
#> 242 501.7885   2.598638
#> 243 964.9930   3.197343
#> 244 507.0023   2.515380
#> 245 500.2682   4.135805
#> 246 811.7989   3.041226
#> 247 583.6765   2.679559
#> 248 516.7726   2.563940
#> 249 908.3557   2.910802
#> 250 502.0910   2.848199
#> 251 500.0450   2.618934
#> 252 501.3800   2.506877
#> 253 500.6085   3.051646
#> 254 572.4347   2.698053
#> 255 503.0391   2.934212
#> 256 501.7716   2.866486
#> 257 834.2749   2.815328
#> 258 500.0030   2.642781
#> 259 564.4532   3.045235
#> 260 509.7537   2.515170
#> 261 500.7870   2.599890
#> 262 500.1547   3.980093
#> 263 510.9030   2.710871
#> 264 523.2566   2.956725
#> 265 500.1094   4.655841
#> 266 513.4791   2.669195
#> 267 573.1435   2.852630
#> 268 500.1023   2.735979
#> 269 715.0486   4.000236
#> 270 690.2529   2.728234
#> 271 511.7204   2.516885
#> 272 500.8612   2.594014
#> 273 679.1196   2.838153
#> 274 527.2326   2.691379
#> 275 500.2468   2.596131
#> 276 512.0524   3.208039
#> 277 513.8091   2.661701
#> 278 500.4386   2.818445
#> 279 639.6829   2.613431
#> 280 501.1034   2.927115
#> 281 502.1588   3.140958
#> 282 500.0276   2.554110
#> 283 500.0154   2.696456
#> 284 500.4012   4.764574
#> 285 845.9337   2.830933
#> 286 500.0101   2.910739
#> 287 503.9497   4.133605
#> 288 622.1848   2.683174
#> 289 505.0668   2.930864
#> 290 507.6078   2.682470
#> 291 503.3308   2.684295
#> 292 986.9197   2.815851
#> 293 521.8524   3.028304
#> 294 828.0488   3.508822
#> 295 504.6946   2.732477
#> 296 500.1768   4.024932
#> 297 500.2292   3.082538
#> 298 513.6165   3.190979
#> 299 500.8066   2.504701
#> 300 500.0089   2.763578
#> 301 510.6305   3.256834
#> 302 527.0912   3.354884
#> 303 549.5356   2.522444
#> 304 504.5664   2.640512
#> 305 524.6935   2.776477
#> 306 500.0615   2.727874
#> 307 550.3662   2.811511
#> 308 820.7529   2.525654
#> 309 574.3886   3.957050
#> 310 500.5798   2.564635
#> 311 502.2818   2.519396
#> 312 530.3826   2.578690
#> 313 500.3152   2.554255
#> 314 500.1775   2.558363
#> 315 500.6446   2.566827
#> 316 511.9161   2.695907
#> 317 500.0785   2.975775
#> 318 528.2908   2.912309
#> 319 500.4155   2.586819
#> 320 564.5464   2.531428
#> 321 501.2505   2.590498
#> 322 500.4221   2.858048
#> 323 518.3673   3.556737
#> 324 506.9352   2.513451
#> 325 651.3294   2.695437
#> 326 815.1773   2.637216
#> 327 509.8095   2.694214
#> 328 501.4871   2.613842
#> 329 501.6090   2.523372
#> 330 500.9323   2.540115
#> 331 500.0911   3.086457
#> 332 501.8549   3.057209
#> 333 500.0573   2.686101
#> 334 729.4481   3.217138
#> 335 500.5339   2.926407
#> 336 506.1326   3.138461
#> 337 500.2908   2.501378
#> 338 506.8103   2.806101
#> 339 572.0182   3.141013
#> 340 500.3216   2.594367
#> 341 597.6224   2.529350
#> 342 502.9388   3.590617
#> 343 728.9134   3.751040
#> 344 500.1708   5.224709
#> 345 613.9733   3.464065
#> 346 501.6871   2.850943
#> 347 500.0023   2.852076
#> 348 500.3665   2.877489
#> 349 502.3499   2.901879
#> 350 506.4915   2.898307
#> 351 505.3092   2.866735
#> 352 624.8190   3.561134
#> 353 500.0638   2.711585
#> 354 500.3825   2.540147
#> 355 501.5804   3.697070
#> 356 500.4625   2.675730
#> 357 633.2239   3.293606
#> 358 500.6940   3.139219
#> 359 500.7838   4.197526
#> 360 666.2653   3.516201
#> 361 583.4044   4.596785
#> 362 500.1658   2.544690
#> 363 876.1512   3.055516
#> 364 500.8983   4.356598
#> 365 500.6492   2.577883
#> 366 651.6932   2.893614
#> 367 507.6264   2.617503
#> 368 500.4712   3.435290
#> 369 501.3893   2.714628
#> 370 507.1614   2.516639
#> 371 530.8705   2.509707
#> 372 500.8424   2.592376
#> 373 507.4032   2.959918
#> 374 500.2714   3.184842
#> 375 502.4289   2.533969
#> 376 506.1478   2.601545
#> 377 500.0085   2.698632
#> 378 500.0173   2.683699
#> 379 532.7365   2.686505
#> 380 500.2526   2.808092
#> 381 510.4193   2.634328
#> 382 609.7925   2.639351
#> 383 500.8332   2.682482
#> 384 584.2978   2.558066
#> 385 502.5805   2.708771
#> 386 500.9434   2.752990
#> 387 998.5845   2.505183
#> 388 500.6589   3.264446
#> 389 583.8995   2.619737
#> 390 501.2040   3.407662
#> 391 500.0980   2.775703
#> 392 502.9725   2.730506
#> 393 502.3410   3.077085
#> 394 500.4773   2.611824
#> 395 500.2679   2.761940
#> 396 500.0040   3.208463
#> 397 852.9596   3.210312
#> 398 508.7518   2.999189
#> 399 520.7244   2.773947
#> 400 500.6525   2.706088
#> 401 500.9128   2.743186
#> 402 500.5738   2.545476
#> 403 501.1213   2.509752
#> 404 728.3770   2.588067
#> 405 512.4975   2.509079
#> 406 500.5410   3.661294
#> 407 593.6279   2.712552
#> 408 520.4057   3.089694
#> 409 501.1982   2.651832
#> 410 511.8898   2.545210
#> 411 504.1072   2.620563
#> 412 534.2131   2.547463
#> 413 507.1192   3.009759
#> 414 607.8742   2.615662
#> 415 501.3560   4.238893
#> 416 505.1034   4.115193
#> 417 575.0968   4.315695
#> 418 500.0795   2.510344
#> 419 500.0123   2.888382
#> 420 511.7721   3.064922
#> 421 581.5224   2.986045
#> 422 572.2866   3.568941
#> 423 502.1591   2.696120
#> 424 500.2150   2.566535
#> 425 500.2429   2.638492
#> 426 531.4972   2.932251
#> 427 504.8429   2.559972
#> 428 522.9144   3.307244
#> 429 501.3000   2.584994
#> 430 501.2161   2.538769
#> 431 501.8396   2.811911
#> 432 500.0022   2.606378
#> 433 516.5521   4.177215
#> 434 503.3865   2.741632
#> 435 500.0362   2.850691
#> 436 586.8631   2.518288
#> 437 520.3351   3.244464
#> 438 547.8837   2.510957
#> 439 500.6698   2.801086
#> 440 552.8519   2.605132
#> 441 500.6519   2.640561
#> 442 502.9039   3.423046
#> 443 502.2295   3.350262
#> 444 505.2456   3.092652
#> 445 501.8015   3.221801
#> 446 504.4942   2.664273
#> 447 501.8160   2.675877
#> 448 500.1046   2.651561
#> 449 501.3656   2.613528
#> 450 500.1371   2.820953
#> 451 500.6667   3.164382
#> 452 500.3776   2.718887
#> 453 512.2230   2.615546
#> 454 500.8861   2.713709
#> 455 821.0990   2.795728
#> 456 979.0764   2.695006
#> 457 501.4603   2.669929
#> 458 709.9679   3.107049
#> 459 596.6453   4.003731
#> 460 652.4337   3.425841
#> 461 504.4303   2.553294
#> 462 500.4337   2.546526
#> 463 501.2144   2.599781
#> 464 500.2081   2.678210
#> 465 729.1423   2.719747
#> 466 516.4291   2.875424
#> 467 516.7728   3.115639
#> 468 501.8842   3.285022
#> 469 522.9384   3.548483
#> 470 509.1259   2.607681
#> 471 501.4066   2.761160
#> 472 500.0540   2.731186
#> 473 500.1524   2.559546
#> 474 504.1125   2.586171
#> 475 502.8247   2.708039
#> 476 500.1008   2.707393
#> 477 503.4045   3.100780
#> 478 501.3237   2.647055
#> 479 500.8968   2.748472
#> 480 618.1459   3.069139
#> 481 508.5551   3.373183
#> 482 624.4541   2.581561
#> 483 500.0480   3.092579
#> 484 529.8418   3.354652
#> 485 500.0066   2.582277
#> 486 797.0302   2.942209
#> 487 500.2807   2.876505
#> 488 508.9209   3.058851
#> 489 504.7563   2.881952
#> 490 653.3802   3.457551
#> 491 583.3180   3.267110
#> 492 707.6690   2.619292
#> 493 500.1496   2.547679
#> 494 502.4892   3.303758
#> 495 534.6270   3.724352
#> 496 501.3877   2.772342
#> 497 500.3805   2.590519
#> 498 500.1042   2.630246
#> 499 500.0537   3.211843
#> 500 502.6087   2.551679
#> 501 776.0038   2.873313
#> 502 526.0540   3.364643
#> 503 505.3775   3.367867
#> 504 500.7355   3.337684
#> 505 502.1615   3.587870
#> 506 503.3799   2.576121
#> 507 535.5242   4.037205
#> 508 500.8133   2.819496
#> 509 505.3898   2.854586
#> 510 528.6848   2.526635
#> 511 500.6728   2.718676
#> 512 501.0099   2.775887
#> 513 517.6545   3.928047
#> 514 501.0204   2.886794
#> 515 507.0226   2.715321
#> 516 500.1218   2.638906
#> 517 500.5093   2.525571
#> 518 513.0025   2.603261
#> 519 500.0356   2.819571
#> 520 506.2413   2.736823
#> 521 500.4630   2.528356
#> 522 500.5015   2.821590
#> 523 502.8538   2.526110
#> 524 549.3314   3.746427
#> 525 512.8534   3.121514
#> 526 539.4745   4.413380
#> 527 500.1139   3.251308
#> 528 502.8917   3.031986
#> 529 500.0498   3.484627
#> 530 503.0991   3.327706
#> 531 501.9684   2.619970
#> 532 500.1245   4.042627
#> 533 537.9879   2.894365
#> 534 512.2221   2.835798
#> 535 500.5869   2.705286
#> 536 519.3428   2.682903
#> 537 617.5956   3.308815
#> 538 546.9289   3.074614
#> 539 503.9994   2.784810
#> 540 520.6556   2.811307
#> 541 564.1830   2.758841
#> 542 635.8136   2.689337
#> 543 552.8372   3.147128
#> 544 508.8449   2.680267
#> 545 501.4789   3.859988
#> 546 500.0152   3.347346
#> 547 500.9230   2.590936
#> 548 500.5373   3.436899
#> 549 500.2113   2.626858
#> 550 500.1013   2.678928
#> 551 503.6536   2.504907
#> 552 507.2159   3.655967
#> 553 513.9797   2.530702
#> 554 510.0277   2.685465
#> 555 500.9417   2.665883
#> 556 532.7413   2.996309
#> 557 681.2667   2.557453
#> 558 500.4161   2.634594
#> 559 501.1653   3.066137
#> 560 505.9706   3.088631
#> 561 500.0182   2.882179
#> 562 516.6212   3.170233
#> 563 519.3047   2.881450
#> 564 512.8110   2.563193
#> 565 500.7499   3.496301
#> 566 502.7817   2.689311
#> 567 510.7753   2.605442
#> 568 515.0941   2.857430
#> 569 500.5491   3.590777
#> 570 787.5416   2.806822
#> 571 500.5932   3.713590
#> 572 513.9085   2.640235
#> 573 551.6208   2.765123
#> 574 504.4061   2.517654
#> 575 513.8403   2.513987
#> 576 500.5258   2.836948
#> 577 551.4629   3.210496
#> 578 500.6576   2.539529
#> 579 571.1851   2.557781
#> 580 571.0049   3.301174
#> 581 504.6978   3.298390
#> 582 500.0183   2.615151
#> 583 500.1832   2.750986
#> 584 813.7300   2.756789
#> 585 519.8361   2.846517
#> 586 540.4086   3.992822
#> 587 515.6798   3.142089
#> 588 602.5729   2.523684
#> 589 523.4332   2.805971
#> 590 509.0446   2.850158
#> 591 501.1533   3.038811
#> 592 502.0222   3.220949
#> 593 501.6712   2.563623
#> 594 519.8071   3.520528
#> 595 500.1782   2.767672
#> 596 500.0033   2.521691
#> 597 500.3251   2.534014
#> 598 504.0864   3.421916
#> 599 511.0493   2.632066
#> 600 504.8722   3.193828
#> 601 505.9411   3.620304
#> 602 500.4883   2.839501
#> 603 514.9391   3.426735
#> 604 500.0745   3.011084
#> 605 509.3029   2.681259
#> 606 500.1232   3.166545
#> 607 877.5475   2.574372
#> 608 518.6620   2.858938
#> 609 524.4813   3.553802
#> 610 519.3751   3.612848
#> 611 955.1702   3.108578
#> 612 500.2402   2.557393
#> 613 555.5535   2.864766
#> 614 557.9902   2.859037
#> 615 500.8823   3.293131
#> 616 559.7770   3.243002
#> 617 500.1459   2.547337
#> 618 502.6338   2.702517
#> 619 606.3248   3.189620
#> 620 500.4231   3.041930
#> 621 509.4276   3.001983
#> 622 504.8139   2.835182
#> 623 562.9246   3.503306
#> 624 500.1491   2.590059
#> 625 500.0043   3.024481
#> 626 500.8027   3.728639
#> 627 572.1663   2.774854
#> 628 500.4813   2.675637
#> 629 500.5097   2.570777
#> 630 500.0352   2.726691
#> 631 855.1348   2.607833
#> 632 502.6367   2.576488
#> 633 500.3053   3.478616
#> 634 502.2251   2.612193
#> 635 500.2918   2.770118
#> 636 503.3099   4.099240
#> 637 597.6808   2.670807
#> 638 501.4262   2.551436
#> 639 500.1815   2.746972
#> 640 516.1417   3.115087
#> 641 598.2149   4.422863
#> 642 502.9470   2.586535
#> 643 633.2954   2.598787
#> 644 500.1373   4.104191
#> 645 500.7737   2.637198
#> 646 775.5148   4.824742
#> 647 551.0415   2.892311
#> 648 500.8663   2.534408
#> 649 500.5275   2.620543
#> 650 745.6666   2.910577
#> 651 500.8448   2.567525
#> 652 500.0941   2.608767
#> 653 505.0166   2.670257
#> 654 500.0918   3.226578
#> 655 504.3403   2.624930
#> 656 500.1881   2.638942
#> 657 500.1198   2.950781
#> 658 506.8166   2.611053
#> 659 511.4088   2.834604
#> 660 502.2484   6.613596
#> 661 512.0738   2.776936
#> 662 500.1865   3.070504
#> 663 507.7255   2.911470
#> 664 501.9513   3.107747
#> 665 646.4549   4.229542
#> 666 500.1292   2.559308
#> 667 500.0313   2.918216
#> 668 501.8902   3.763885
#> 669 508.1670   2.632329
#> 670 567.3351   2.584661
#> 671 502.8540   2.881685
#> 672 501.5506   3.077795
#> 673 501.7202   2.724143
#> 674 500.0792   2.524223
#> 675 781.9924   2.530726
#> 676 509.3973   2.958851
#> 677 602.2485   2.802201
#> 678 640.6687   2.817829
#> 679 514.4825   3.294306
#> 680 512.7210   2.545415
#> 681 538.2259   3.337448
#> 682 501.5315   2.569899
#> 683 622.0934   2.560790
#> 684 503.1589   2.576633
#> 685 500.2295   2.826097
#> 686 500.5333   2.640529
#> 687 503.6177   2.783896
#> 688 685.7665   2.693241
#> 689 501.4992   3.044876
#> 690 508.2206   2.568751
#> 691 500.0882   2.819524
#> 692 981.1917   2.869620
#> 693 506.6462   3.441550
#> 694 532.1611   3.960389
#> 695 505.4754   2.540042
#> 696 501.9967   2.694443
#> 697 517.6814   3.339636
#> 698 500.0228   2.746003
#> 699 500.0154   2.569264
#> 700 500.6996   3.055225
#> 701 514.1682   2.922705
#> 702 500.0515   2.865449
#> 703 509.1944   2.642516
#> 704 855.8446   3.103746
#> 705 503.8067   2.711891
#> 706 500.0262   2.880091
#> 707 500.2152   3.779106
#> 708 789.7936   2.728343
#> 709 511.2355   2.680554
#> 710 509.2348   2.804681
#> 711 527.2810   2.806932
#> 712 500.4642   2.531719
#> 713 626.8218   2.682748
#> 714 529.6630   2.700326
#> 715 501.9729   2.590607
#> 716 505.0189   2.565559
#> 717 531.3880   2.618740
#> 718 500.1060   2.752812
#> 719 500.1500   3.112541
#> 720 502.3806   3.204667
#> 721 553.5260   2.661492
#> 722 502.5499   3.493389
#> 723 500.9710   2.855268
#> 724 500.5107   3.376506
#> 725 500.2357   2.794792
#> 726 500.2848   3.196858
#> 727 521.3601   2.676900
#> 728 520.2574   2.594461
#> 729 537.5693   2.517738
#> 730 501.8036   3.024865
#> 731 500.0785   3.529764
#> 732 500.0435   2.869262
#> 733 500.1160   3.079279
#> 734 511.0418   3.311022
#> 735 578.4368   2.527860
#> 736 504.0980   3.783818
#> 737 825.2891   2.582863
#> 738 500.0013   2.774966
#> 739 500.0023   2.740902
#> 740 738.1569   3.357560
#> 741 768.1765   2.523363
#> 742 967.0611   2.816706
#> 743 504.1802   2.584723
#> 744 855.2015   2.667911
#> 745 543.7216   2.528533
#> 746 501.9649   3.295944
#> 747 501.0781   2.536510
#> 748 577.5913   2.534791
#> 749 684.0712   2.723693
#> 750 512.6676   2.796730
#> 751 505.8134   3.269807
#> 752 500.2264   3.286645
#> 753 506.4096   2.565410
#> 754 500.0983   2.578664
#> 755 573.7877   2.506222
#> 756 500.0133   2.609200
#> 757 500.5699   2.666364
#> 758 690.9775   4.196101
#> 759 504.1636   3.206655
#> 760 523.6136   2.837629
#> 761 500.8952   2.583660
#> 762 578.4894   4.022933
#> 763 508.2258   2.780845
#> 764 502.2208   2.646245
#> 765 503.1885   2.942174
#> 766 600.6257   2.822863
#> 767 501.9915   2.854484
#> 768 688.5084   2.710134
#> 769 501.0040   2.688474
#> 770 519.5510   2.719088
#> 771 500.8198   2.625317
#> 772 577.4915   2.849546
#> 773 500.2730   3.249756
#> 774 500.0579   2.587771
#> 775 501.5013   2.993414
#> 776 518.0105   2.657702
#> 777 500.0790   3.731816
#> 778 501.1681   3.192684
#> 779 500.0567   2.504107
#> 780 500.2634   2.980315
#> 781 503.9002   3.655635
#> 782 500.2884   2.636098
#> 783 527.1355   3.118423
#> 784 500.0371   2.802091
#> 785 544.7835   2.838681
#> 786 500.1402   3.061282
#> 787 501.8089   2.683965
#> 788 507.6435   2.637030
#> 789 503.6274   2.502448
#> 790 501.0213   2.928847
#> 791 501.2645   3.542008
#> 792 545.0739   2.724162
#> 793 500.2905   3.141131
#> 794 525.6797   3.980590
#> 795 500.3796   3.785167
#> 796 564.6710   2.507612
#> 797 501.4881   2.702932
#> 798 500.0343   2.823017
#> 799 554.3089   2.769758
#> 800 500.7248   3.116549
#> 801 500.2239   3.052880
#> 802 500.3988   2.675249
#> 803 518.5509   2.603691
#> 804 500.1630   2.786240
#> 805 500.2969   2.514825
#> 806 500.1219   2.630744
#> 807 535.9117   2.658024
#> 808 595.3708   3.775064
#> 809 559.0869   2.756799
#> 810 500.1045   2.863837
#> 811 512.8838   3.657095
#> 812 744.3751   3.341498
#> 813 500.9024   3.192056
#> 814 500.0886   2.949773
#> 815 500.1770   2.528638
#> 816 627.5897   2.621356
#> 817 503.3943   2.903972
#> 818 502.6968   2.577398
#> 819 525.8965   2.534752
#> 820 500.8489   2.981054
#> 821 500.1912   2.524346
#> 822 504.4221   2.657163
#> 823 503.2053   3.225497
#> 824 500.8805   2.535749
#> 825 502.3058   2.717983
#> 826 523.9183   2.738441
#> 827 508.9490   2.597832
#> 828 509.7762   2.592174
#> 829 500.2220   3.022020
#> 830 500.0156   2.567602
#> 831 500.1538   3.853850
#> 832 500.9615   3.363092
#> 833 501.1334   2.638313
#> 834 701.3338   2.646367
#> 835 586.3477   3.123776
#> 836 507.6459   2.804888
#> 837 500.0770   2.795851
#> 838 504.5755   2.837914
#> 839 509.5407   3.081143
#> 840 531.3151   2.573705
#> 841 501.8354   2.573555
#> 842 504.2537   3.155677
#> 843 502.1801   2.810942
#> 844 500.4295   2.704461
#> 845 601.8467   2.509760
#> 846 907.6421   3.152596
#> 847 502.0528   2.814181
#> 848 526.8533   2.764076
#> 849 501.3914   2.969454
#> 850 500.2536   2.693262
#> 851 630.9453   3.242568
#> 852 500.0101   3.591708
#> 853 580.0037   3.144657
#> 854 500.1565   2.878147
#> 855 501.2534   2.516064
#> 856 506.9478   2.592922
#> 857 565.2137   2.522288
#> 858 501.5795   2.963539
#> 859 695.9417   2.548836
#> 860 500.1263   3.068475
#> 861 648.7639   4.535817
#> 862 500.4322   2.599053
#> 863 500.0275   3.728043
#> 864 561.5305   2.975146
#> 865 676.4664   3.247448
#> 866 514.2803   2.941194
#> 867 518.4877   2.636677
#> 868 502.5920   2.690943
#> 869 511.1947   2.880448
#> 870 502.3754   2.663082
#> 871 500.7974   2.919648
#> 872 574.4087   2.640070
#> 873 501.6239   2.958747
#> 874 500.3538   3.587772
#> 875 502.2506   2.554261