Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample.GR.magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample.GR.magnitudes(n=100, beta.p=log(10), M0=2.5)
#>   [1] 2.625821 2.704320 2.708853 3.938084 2.590899 2.983556 2.507833 2.685962
#>   [9] 2.897804 3.587020 2.764607 2.653352 2.563043 2.552746 2.548978 2.734631
#>  [17] 4.084863 3.030668 2.787362 3.616258 3.083258 3.737160 2.685440 2.517355
#>  [25] 2.995004 2.889504 2.547649 2.572937 2.542308 3.137860 2.946325 3.409274
#>  [33] 2.529947 3.513635 2.674804 2.909130 2.597242 2.775857 2.695347 3.278008
#>  [41] 2.737670 2.626954 2.762171 2.994040 2.624154 2.702055 2.973641 2.808548
#>  [49] 2.921655 2.855653 3.105940 2.613839 2.905353 3.302029 2.841419 2.513704
#>  [57] 2.955832 2.684601 3.099950 2.689200 2.780473 2.522815 2.911181 3.447662
#>  [65] 2.513745 2.896375 2.504765 2.931729 2.703170 2.572613 2.606587 3.221394
#>  [73] 2.658793 5.304698 3.187805 2.881374 2.597898 3.488908 2.853696 2.847437
#>  [81] 2.756490 2.649395 3.948544 3.010460 2.547103 2.589134 2.540458 3.940298
#>  [89] 3.094565 2.816646 2.723966 2.828596 2.558099 4.301393 2.510203 3.025937
#>  [97] 2.795758 3.419612 3.432195 3.241129