Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample.GR.magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample.GR.magnitudes(n=100, beta.p=log(10), M0=2.5)
#>   [1] 3.804468 2.772865 2.539528 2.747005 2.561155 2.575896 3.571892 2.655136
#>   [9] 3.750650 2.566016 2.502325 2.545616 3.304445 2.817967 2.872623 2.824522
#>  [17] 3.682095 2.516008 3.149382 2.537316 2.565724 2.595560 2.840604 2.768582
#>  [25] 2.552313 3.517061 2.974919 2.603314 3.403507 2.559879 2.693523 2.501184
#>  [33] 2.739113 2.806565 2.970249 3.210149 3.093397 2.819538 2.761085 2.761193
#>  [41] 2.560894 2.736833 2.865075 2.669356 3.195525 2.525958 3.610762 2.614398
#>  [49] 3.018200 3.735262 2.683840 3.018139 2.822647 2.771518 2.589079 2.756012
#>  [57] 2.563601 2.567163 2.709967 2.698455 2.666415 2.710179 2.600497 3.286986
#>  [65] 2.634985 3.461382 3.128044 2.907212 2.569365 2.588889 2.623758 2.786431
#>  [73] 3.065191 2.905985 3.293738 2.887709 2.542378 3.157412 3.212740 2.755912
#>  [81] 3.452721 2.631028 2.910951 2.626846 2.615919 2.880636 2.656838 3.421244
#>  [89] 4.910065 3.329992 2.590063 3.106817 2.955035 3.067758 3.428315 2.991033
#>  [97] 2.522240 3.959695 2.845557 2.501132