Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample.GR.magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample.GR.magnitudes(n=100, beta.p=log(10), M0=2.5)
#>   [1] 2.641933 3.129423 2.603200 2.822155 3.104633 2.590570 3.087346 3.065790
#>   [9] 2.697443 3.154398 3.523040 2.808663 2.738108 3.134920 2.679909 3.211363
#>  [17] 2.503164 3.088190 2.788848 2.594367 2.561584 4.860109 2.536216 2.930382
#>  [25] 3.981899 2.527621 3.080592 2.590994 3.181434 4.248675 3.534735 2.512854
#>  [33] 2.889331 3.671727 2.753200 3.612292 2.752012 2.765502 2.512567 2.750001
#>  [41] 3.071141 3.045273 2.596226 3.255760 2.804602 2.790563 3.369945 2.687815
#>  [49] 2.560033 2.621135 3.149738 3.030031 2.563609 2.529354 3.064888 2.813328
#>  [57] 3.049537 2.608344 2.543545 3.347524 2.816901 2.721514 2.820777 2.519242
#>  [65] 2.788716 3.891934 3.344792 3.030377 2.675773 2.588388 2.664468 2.554327
#>  [73] 3.034192 3.399116 2.955684 3.288447 2.763819 2.941204 3.470106 2.753878
#>  [81] 2.515954 2.985931 2.899087 3.291367 3.215908 2.659627 3.365296 2.748815
#>  [89] 2.594763 2.851491 2.591799 2.521553 3.088666 2.987512 2.639378 2.672261
#>  [97] 2.779778 3.935615 2.507209 3.389363