Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample.GR.magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample.GR.magnitudes(n=100, beta.p=log(10), M0=2.5)
#>   [1] 2.722270 2.811399 3.112606 2.907237 3.003420 3.241585 2.624158 2.514424
#>   [9] 3.440371 2.792377 2.697279 3.397633 2.756946 2.669865 3.178565 2.752937
#>  [17] 2.716542 3.229406 2.822981 2.640293 2.542094 2.873914 2.910538 2.671658
#>  [25] 2.500780 2.527655 2.530694 2.671805 2.585761 2.924360 3.160607 3.115632
#>  [33] 2.599147 3.041687 2.529623 2.519223 3.615039 3.245961 2.629507 4.398769
#>  [41] 2.875168 3.092817 2.744037 4.925330 2.646771 3.268544 2.791737 3.080759
#>  [49] 3.136716 2.825063 4.223541 2.739388 2.565096 3.180023 2.885289 3.134267
#>  [57] 2.948688 3.448626 3.183174 2.672957 3.091346 2.691273 2.691958 3.403208
#>  [65] 2.643230 3.004027 3.103590 2.734765 2.625371 2.914707 2.842808 2.527542
#>  [73] 2.522452 3.256528 2.861350 2.738529 2.812750 2.991394 2.827083 2.765862
#>  [81] 2.711710 2.795817 2.857062 3.693442 2.523821 2.992262 2.599010 3.559388
#>  [89] 2.549589 2.732997 2.981093 2.505927 2.815736 3.073771 4.755590 2.555566
#>  [97] 2.741735 2.516690 2.958570 2.584690