Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample.GR.magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample.GR.magnitudes(n=100, beta.p=log(10), M0=2.5)
#>   [1] 2.504221 3.215335 3.498684 2.662734 3.830758 2.782812 2.576073 2.786066
#>   [9] 2.821545 2.798651 2.872886 2.577325 3.096484 2.626068 4.189134 2.607921
#>  [17] 3.251300 3.951588 2.533779 3.401893 3.071859 2.914953 3.152533 2.527979
#>  [25] 2.694716 2.546001 2.943172 3.549874 2.640376 2.573585 2.799706 2.816290
#>  [33] 3.140626 3.345591 2.816208 2.534738 2.924766 2.734472 3.202612 2.509332
#>  [41] 3.462869 2.565869 2.876925 3.666788 2.621451 2.838755 2.648517 3.310032
#>  [49] 2.885058 2.880203 2.606110 3.519191 3.021882 2.754348 2.829723 2.638029
#>  [57] 3.129539 3.449598 3.413089 3.061994 3.544151 2.823115 2.678619 2.840585
#>  [65] 3.254094 2.575382 3.088894 2.710512 3.024288 2.729793 2.640687 2.790156
#>  [73] 2.672338 2.700188 3.078284 3.533543 2.510224 2.532865 3.182135 3.021373
#>  [81] 2.864574 2.670788 2.684464 2.982497 3.140527 2.658296 2.531398 2.915516
#>  [89] 2.538768 2.715821 2.896481 2.897643 2.999098 2.622826 3.103048 2.662051
#>  [97] 2.810888 2.550062 2.571664 2.690831