Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample.GR.magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample.GR.magnitudes(n=100, beta.p=log(10), M0=2.5)
#>   [1] 3.104617 2.615985 3.555158 3.056580 2.544345 3.403611 2.860058 2.505429
#>   [9] 3.310656 2.693577 2.604681 2.558904 2.509721 2.757715 2.577848 3.100577
#>  [17] 2.877556 2.661824 2.516907 2.613581 2.538398 2.911824 2.736476 3.769126
#>  [25] 2.687616 2.871114 3.210685 3.481290 2.798975 2.561986 2.768773 3.033386
#>  [33] 2.865438 2.928737 2.967915 3.021199 2.572427 2.597177 2.614528 2.505996
#>  [41] 2.596735 2.951962 2.753198 2.580859 3.603527 2.881803 2.801642 2.553726
#>  [49] 2.778738 2.653432 2.502857 3.030181 2.792077 2.582973 2.881507 2.981916
#>  [57] 3.961750 2.555641 2.596809 2.626569 2.552383 2.515867 2.832378 4.518288
#>  [65] 2.904996 2.979037 2.816225 3.189173 2.622115 2.512142 3.467054 2.721075
#>  [73] 3.032648 2.570222 3.009049 2.532518 2.694465 3.544810 3.817668 4.551874
#>  [81] 3.111290 3.218980 2.689653 4.142692 3.041180 2.770528 2.836850 3.355917
#>  [89] 2.524099 2.553150 3.142243 2.679644 2.558853 2.858503 2.582395 3.254803
#>  [97] 2.645223 3.617317 3.263867 2.614143