Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample.GR.magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample.GR.magnitudes(n=100, beta.p=log(10), M0=2.5)
#>   [1] 2.986648 2.507549 2.707647 2.632715 2.660338 2.734978 2.892317 2.775035
#>   [9] 3.039808 2.620818 2.958050 2.928378 2.705734 2.653577 3.035954 2.548132
#>  [17] 2.615714 3.087897 2.720818 2.650190 2.524098 3.110814 2.881590 2.807347
#>  [25] 2.571364 3.864141 2.513055 2.684588 2.738656 2.544974 2.651987 2.601976
#>  [33] 3.080689 2.760007 2.603361 2.902366 2.674844 2.766649 2.762485 2.653589
#>  [41] 2.908786 3.607065 3.000579 3.560960 2.809835 3.126010 2.947250 3.104886
#>  [49] 2.511963 2.545452 3.428720 3.491074 2.762027 5.500447 3.113366 2.622011
#>  [57] 2.556641 2.714417 3.158317 3.112835 2.670267 2.643520 3.303508 4.384122
#>  [65] 2.629005 2.604587 2.508208 2.733269 3.828914 3.801590 2.720255 2.683804
#>  [73] 3.192543 2.781884 3.035059 3.025682 2.919062 3.320019 2.654699 2.802128
#>  [81] 2.708304 2.666890 3.142423 2.703104 2.568321 3.223348 3.069880 2.606515
#>  [89] 2.845702 2.514042 3.510212 2.602760 2.658186 2.895951 2.671422 3.374693
#>  [97] 3.489008 3.417384 2.924998 2.860075