Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample.GR.magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample.GR.magnitudes(n=100, beta.p=log(10), M0=2.5)
#>   [1] 2.738956 3.496005 2.897551 2.549515 3.413896 2.619381 3.311707 2.530441
#>   [9] 2.917650 3.335500 2.532280 3.084162 2.887626 3.000099 2.699376 3.046723
#>  [17] 2.882465 3.043022 2.904025 2.654271 2.674906 2.602232 3.166487 2.566053
#>  [25] 3.940057 3.029932 2.564086 3.125482 3.599049 2.805254 4.370911 4.121608
#>  [33] 3.033746 2.670631 2.696712 2.851387 2.821576 2.718268 2.887329 2.757361
#>  [41] 2.921780 2.718727 2.633083 2.745473 3.323175 2.835460 3.494314 2.753394
#>  [49] 2.809045 3.236426 2.751994 3.714741 2.739482 2.828881 2.818022 2.796508
#>  [57] 2.637914 2.701755 2.849583 2.584326 3.083807 2.889771 4.001880 3.031926
#>  [65] 2.520085 4.561250 3.000268 3.593214 2.955786 2.536793 2.561121 2.574096
#>  [73] 3.200423 2.510505 2.629272 2.599650 3.200217 2.852945 3.897371 3.108564
#>  [81] 2.635659 2.547647 3.821204 3.872536 2.691694 2.589711 2.534007 2.942948
#>  [89] 2.738939 2.849028 3.148703 2.614107 2.534811 3.079371 2.871601 3.060573
#>  [97] 3.049881 2.669105 2.580350 3.127450