Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample.GR.magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample.GR.magnitudes(n=100, beta.p=log(10), M0=2.5)
#>   [1] 3.139132 2.671270 3.333273 2.777410 3.181213 3.199895 2.751083 3.032898
#>   [9] 2.689418 2.678191 3.645903 2.733156 2.543067 2.691124 2.505913 2.612252
#>  [17] 2.840534 2.636416 2.627737 3.071147 3.154230 3.391868 2.504021 2.769365
#>  [25] 3.688227 3.427384 2.814434 2.660315 2.678407 2.927171 2.905411 2.665005
#>  [33] 2.627608 2.742457 2.689395 2.655666 2.880207 2.644979 3.211683 2.945097
#>  [41] 3.009646 2.922205 2.666276 3.144313 2.553398 3.589146 3.165256 2.745339
#>  [49] 2.684746 2.726500 5.254403 3.614027 2.539844 2.782432 3.072493 2.512166
#>  [57] 2.592147 2.528909 2.839211 3.134971 2.571199 2.721018 3.163206 3.069819
#>  [65] 3.179256 2.945555 2.558312 2.855424 3.121183 2.730115 3.148589 3.113592
#>  [73] 3.210059 2.643757 2.881594 3.006852 2.739245 2.653943 2.732402 2.562279
#>  [81] 2.886316 2.671038 2.503786 3.533309 3.062450 3.232747 2.561529 2.612327
#>  [89] 2.631763 3.115633 2.716752 3.150979 4.752654 2.549360 2.609100 3.750689
#>  [97] 2.521426 2.664725 2.734917 2.621464