Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample.GR.magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample.GR.magnitudes(n=100, beta.p=log(10), M0=2.5)
#>   [1] 3.116167 2.651309 2.572510 4.349061 2.501816 3.259691 2.815905 3.339228
#>   [9] 2.804889 2.777180 2.525243 4.121814 2.897900 2.661060 2.811804 2.685721
#>  [17] 2.813788 3.013615 3.880385 2.559628 3.772163 3.989779 2.740104 3.297152
#>  [25] 3.102574 2.523255 3.840687 2.602773 2.794405 3.270553 2.661416 2.880601
#>  [33] 2.795870 2.517200 2.582705 2.965909 2.976986 2.769597 3.002287 3.072382
#>  [41] 2.879984 2.884961 3.057581 2.516455 3.102390 3.653282 2.566877 2.890182
#>  [49] 3.515829 3.170546 2.509185 2.643612 2.619722 2.933255 2.653391 2.900887
#>  [57] 3.049925 4.317750 4.485920 3.503028 3.005916 2.535577 2.591199 2.865821
#>  [65] 2.872411 2.905679 3.262262 2.618965 2.671186 2.904446 2.669291 2.659310
#>  [73] 3.471162 2.799704 2.553919 2.717861 2.580606 4.320058 2.523802 3.050556
#>  [81] 2.858635 3.105097 3.100213 2.722098 2.689094 2.849473 2.847501 2.737235
#>  [89] 2.686150 2.555939 2.924905 2.582632 2.750614 3.092189 2.740198 2.582977
#>  [97] 3.305783 2.574484 2.983758 2.746166