Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample.GR.magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample.GR.magnitudes(n=100, beta.p=log(10), M0=2.5)
#>   [1] 3.571820 2.935396 2.743097 2.571339 2.598547 2.949744 4.763442 3.157411
#>   [9] 2.561705 2.874231 2.584831 3.967150 2.762877 3.164911 3.314899 2.567076
#>  [17] 2.879965 3.055174 2.728155 2.727057 2.838445 2.730079 3.402296 2.983613
#>  [25] 2.588297 2.851474 2.670257 2.616299 2.674499 2.557116 2.643375 2.550376
#>  [33] 2.738083 3.370499 2.530162 2.845641 2.760028 2.598826 3.101651 2.782533
#>  [41] 2.525730 2.898976 2.610171 2.762103 4.360188 2.926610 2.593621 2.516868
#>  [49] 3.514829 2.981055 2.806361 2.774680 3.194065 2.629254 3.715630 2.667977
#>  [57] 2.540442 2.905799 3.213017 2.574717 3.725924 2.780106 3.737800 2.651173
#>  [65] 2.692309 3.163801 2.981469 2.579608 2.566907 3.185208 2.505765 2.587311
#>  [73] 2.530758 2.529253 2.587034 3.027547 2.854068 3.316678 2.575271 2.581127
#>  [81] 2.925582 2.668956 2.998703 2.641236 2.580165 4.250870 2.926341 2.824753
#>  [89] 3.064862 2.578934 2.694354 3.270057 2.751736 2.600147 2.795395 2.859480
#>  [97] 3.342436 2.570461 2.732300 3.057887