Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample.GR.magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample.GR.magnitudes(n=100, beta.p=log(10), M0=2.5)
#>   [1] 2.701922 2.600809 3.250465 2.847835 2.510330 3.006694 2.788459 2.583130
#>   [9] 2.771211 3.617630 3.062035 4.021136 2.665172 2.624588 3.208512 2.650408
#>  [17] 2.510439 2.654583 2.653394 2.642407 3.933202 4.061134 2.508642 3.186060
#>  [25] 2.524593 2.766806 2.905972 3.603305 2.935860 2.893676 2.547818 2.983435
#>  [33] 2.816696 2.689690 2.614234 2.644607 3.422608 3.214321 2.816206 2.595849
#>  [41] 2.967712 2.654589 3.115165 2.910889 3.333009 3.062772 2.725829 2.585953
#>  [49] 2.905441 2.956238 2.731907 4.090162 2.726357 2.955421 5.153429 3.245895
#>  [57] 2.627869 2.742824 2.605739 2.661493 2.528069 3.005319 2.792170 2.511067
#>  [65] 2.625084 3.750913 3.747512 2.700710 2.586817 2.594656 2.525588 2.937642
#>  [73] 2.614057 2.515252 2.639846 4.566991 2.770978 2.539684 2.514551 3.547376
#>  [81] 2.517597 2.792528 2.526583 2.724844 3.017155 2.530251 2.878939 3.588476
#>  [89] 2.664087 2.656270 2.746589 3.312808 2.731262 2.857949 2.727322 2.732816
#>  [97] 3.980873 2.763366 2.643353 4.171585