Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample.GR.magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample.GR.magnitudes(n=100, beta.p=log(10), M0=2.5)
#>   [1] 2.532657 3.137604 3.458405 2.593052 2.677763 2.667010 3.026865 2.547445
#>   [9] 2.921626 3.096645 3.520929 3.076732 2.835385 2.623896 2.916768 2.531311
#>  [17] 2.607561 3.231742 2.689543 2.619162 2.712099 2.566453 3.177633 2.509608
#>  [25] 3.142731 2.520530 2.761804 2.936529 2.947024 2.844503 3.083568 4.616704
#>  [33] 2.791429 2.794218 2.619484 2.927169 3.026839 3.380834 2.782483 2.668974
#>  [41] 3.184134 2.502948 2.870465 2.904849 3.031261 2.523764 2.837044 2.517793
#>  [49] 4.183848 2.631635 2.619292 2.682635 2.579109 3.019795 2.581168 4.009626
#>  [57] 2.672908 2.741583 3.487133 2.852743 2.738816 2.563498 3.634966 3.521560
#>  [65] 2.808824 2.739920 2.547796 3.684711 3.057313 2.688271 2.918077 2.556372
#>  [73] 2.998605 2.821072 2.572285 3.427292 2.601505 3.556461 2.549411 2.614516
#>  [81] 3.527814 2.976220 2.609159 2.622938 3.188275 3.156254 2.844808 4.150868
#>  [89] 2.997358 2.516481 2.839963 2.526273 2.991236 3.257566 3.117180 2.616361
#>  [97] 2.915532 2.567500 3.026105 2.852908