Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample.GR.magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample.GR.magnitudes(n=100, beta.p=log(10), M0=2.5)
#>   [1] 2.649871 2.726535 2.814775 2.657365 2.801376 2.772375 3.002251 3.049210
#>   [9] 2.547331 3.060097 2.995285 3.350704 3.322403 3.008919 2.602129 2.936369
#>  [17] 3.408675 2.981154 3.147116 2.606232 2.519350 2.935860 2.983897 2.995153
#>  [25] 2.512064 2.893688 3.259871 2.596651 3.034455 2.532740 2.825723 3.267887
#>  [33] 2.504347 3.579178 3.915072 2.635716 2.522382 2.765220 2.615030 2.662592
#>  [41] 3.812887 2.767281 2.910474 3.959442 2.639326 2.783723 3.027146 2.809854
#>  [49] 3.554904 2.595904 3.246595 2.717088 2.660325 2.761260 2.778123 2.687257
#>  [57] 2.779635 2.572263 3.375357 2.564506 3.157808 3.032825 2.714499 2.945786
#>  [65] 2.513018 3.817412 3.202078 2.509670 3.159532 3.164303 3.226536 2.609248
#>  [73] 2.525225 2.676882 2.884937 2.870375 2.649976 2.549870 2.545659 2.606970
#>  [81] 3.698739 2.783163 2.758282 2.605799 2.508039 2.811716 3.360051 2.769109
#>  [89] 2.982594 3.345944 3.145329 2.896660 4.172763 3.177542 2.836538 2.895953
#>  [97] 2.707469 5.121307 2.577326 2.547027