Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample.GR.magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample.GR.magnitudes(n=100, beta.p=log(10), M0=2.5)
#>   [1] 2.628424 2.603155 2.658392 3.197318 3.266479 2.819594 2.609339 3.166154
#>   [9] 2.738211 2.581725 2.602850 2.781169 2.922208 2.733333 2.549241 2.623098
#>  [17] 2.911192 3.840351 2.581854 3.370045 2.974825 2.787459 2.757894 3.457450
#>  [25] 2.904131 2.674880 2.743515 3.491671 2.528550 2.784183 2.783154 3.106137
#>  [33] 2.670396 3.769808 2.818129 2.674760 2.562631 3.031442 2.570240 3.109593
#>  [41] 2.859589 3.087629 2.536944 2.805289 2.570549 2.646264 2.732522 2.932195
#>  [49] 2.591647 2.766897 2.684901 2.565135 4.309415 3.440930 2.595653 2.500972
#>  [57] 3.255625 2.538130 2.973504 3.250170 2.728691 2.867497 2.737155 2.906091
#>  [65] 2.713054 2.589946 3.199352 2.608779 3.639775 4.270911 2.954448 3.928663
#>  [73] 4.015621 2.648991 2.867644 3.329276 2.530419 2.745838 2.646105 4.092225
#>  [81] 2.524025 3.579738 2.666115 2.523642 2.641405 2.631839 3.519134 2.535573
#>  [89] 2.876850 2.601944 2.892313 2.607488 2.651285 2.667419 2.741875 2.619812
#>  [97] 2.985389 2.593431 4.619659 2.713448