Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample.GR.magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample.GR.magnitudes(n=100, beta.p=log(10), M0=2.5)
#>   [1] 2.586895 2.756041 2.609535 2.705473 2.550962 2.757021 3.363081 3.157977
#>   [9] 2.670478 2.594263 2.672836 2.552466 2.687712 2.851205 2.686568 3.554178
#>  [17] 2.708816 2.779994 2.557998 3.642494 2.807824 2.753509 4.501633 2.562751
#>  [25] 2.777693 2.836889 2.521489 2.502923 2.938838 2.523370 2.523909 2.532601
#>  [33] 3.085058 2.565719 2.573414 2.956698 2.518916 2.666477 3.890548 3.266856
#>  [41] 2.887183 2.920907 3.833776 3.978541 2.691393 2.613259 2.711971 2.507522
#>  [49] 3.143771 2.632595 3.139803 2.707419 2.836379 2.732766 2.515060 2.639220
#>  [57] 2.575365 2.551277 3.656329 2.548577 2.793004 3.548576 2.540549 2.694647
#>  [65] 2.701708 2.775668 2.595356 2.871938 2.587793 2.891752 2.747554 2.855988
#>  [73] 2.521393 3.092556 3.008357 2.515908 3.161103 2.694146 2.563460 2.992627
#>  [81] 4.087905 2.961271 3.373445 2.878870 3.294644 2.870001 2.642915 3.273241
#>  [89] 2.911734 3.002077 3.297436 3.076920 2.944610 2.685285 3.470909 2.552779
#>  [97] 2.762958 4.298066 2.779919 2.517449