Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample.GR.magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample.GR.magnitudes(n=100, beta.p=log(10), M0=2.5)
#>   [1] 2.750919 2.564597 3.497209 3.323585 3.073029 2.718400 2.914888 3.634408
#>   [9] 2.508962 2.732354 3.764584 2.601709 3.251434 3.341822 2.501381 3.666550
#>  [17] 3.556261 2.825769 2.832301 2.519709 2.753578 3.201026 4.119520 2.603120
#>  [25] 2.985476 2.599043 3.239654 4.561642 3.069043 2.615173 2.894275 3.254457
#>  [33] 2.652517 2.959715 2.953597 2.729350 2.713198 2.957895 2.837870 2.523978
#>  [41] 3.013657 3.033500 2.817061 2.782608 2.930182 3.074191 3.355891 2.874134
#>  [49] 2.801820 2.933373 4.508992 3.459534 2.558619 2.537877 3.203837 2.749624
#>  [57] 3.117454 2.593413 2.914609 3.386157 2.604173 2.507104 2.980104 2.594063
#>  [65] 2.506381 2.517086 3.038366 2.585707 2.685965 2.856774 2.958114 2.918611
#>  [73] 2.506883 2.503242 3.611606 2.728177 2.569516 3.227183 2.975752 2.692171
#>  [81] 3.430593 2.660138 2.898613 2.657324 2.559279 2.600198 2.769967 2.722342
#>  [89] 3.233070 2.717408 2.838678 4.638995 2.564435 3.033314 2.844012 2.591887
#>  [97] 3.090787 3.060711 2.861669 3.010364