Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   506.2506   2.858338
#> 2   500.4800   2.529172
#> 3   846.9243   2.537127
#> 4   531.1376   2.676338
#> 5   540.5341   3.394109
#> 6   503.1689   2.602960
#> 7   514.9434   2.557719
#> 8   533.1002   3.197033
#> 9   500.0773   2.570053
#> 10  500.0125   3.665042
#> 11  579.5744   2.855781
#> 12  500.1115   3.023366
#> 13  500.7841   3.187489
#> 14  502.0393   2.694030
#> 15  503.7770   2.510817
#> 16  503.7547   2.825064
#> 17  500.4305   2.517940
#> 18  530.6164   2.705511
#> 19  507.8578   2.588102
#> 20  500.0466   2.531411
#> 21  506.4375   2.594300
#> 22  947.8373   2.745557
#> 23  529.8781   2.638319
#> 24  779.2459   2.934068
#> 25  500.3968   2.952326
#> 26  500.1437   2.564974
#> 27  567.3968   3.009420
#> 28  500.7902   3.024501
#> 29  500.1047   2.592361
#> 30  501.4890   2.752394
#> 31  500.0792   3.657874
#> 32  500.3762   2.642545
#> 33  868.1383   2.563467
#> 34  500.7903   2.709369
#> 35  988.6839   2.645662
#> 36  540.4362   2.776099
#> 37  500.0276   2.705471
#> 38  664.8202   2.887857
#> 39  656.2959   2.643373
#> 40  573.2853   3.528863
#> 41  500.0962   3.021025
#> 42  503.1642   2.712154
#> 43  577.3765   3.104599
#> 44  502.8240   2.578552
#> 45  520.1529   2.779153
#> 46  500.0664   3.529908
#> 47  500.0515   3.500459
#> 48  502.9824   3.275814
#> 49  500.1316   2.502130
#> 50  766.5947   2.706922
#> 51  517.4401   2.898283
#> 52  500.6818   2.624516
#> 53  565.9777   2.963733
#> 54  500.2388   3.088745
#> 55  523.5409   2.514345
#> 56  502.4829   3.243977
#> 57  706.5810   2.763944
#> 58  500.5268   2.933230
#> 59  500.4078   2.963034
#> 60  502.0502   3.626844
#> 61  500.3915   2.505085
#> 62  500.0739   2.723151
#> 63  505.1112   2.861567
#> 64  606.0631   2.513359
#> 65  505.7371   3.805348
#> 66  620.0391   2.646568
#> 67  500.4975   3.379668
#> 68  501.6317   2.611469
#> 69  528.8107   2.621511
#> 70  504.3879   2.650151
#> 71  500.3857   2.706672
#> 72  500.0423   2.737814
#> 73  501.6277   2.947174
#> 74  500.2818   3.173710
#> 75  500.5566   2.618483
#> 76  500.0793   3.140476
#> 77  548.6578   2.571547
#> 78  503.4806   2.784998
#> 79  500.0067   2.629106
#> 80  500.8150   2.866556
#> 81  500.3154   3.823016
#> 82  500.2939   2.672736
#> 83  500.4070   2.543608
#> 84  576.3657   2.871635
#> 85  644.3407   3.400296
#> 86  500.0900   2.504248
#> 87  500.4426   2.744238
#> 88  582.3342   2.760206
#> 89  541.9174   2.952867
#> 90  500.5159   2.629554
#> 91  520.2598   2.672598
#> 92  515.9771   2.703000
#> 93  500.2102   3.245944
#> 94  975.6362   3.666084
#> 95  510.3399   2.674704
#> 96  613.4754   2.579489
#> 97  991.9650   2.590698
#> 98  504.3038   3.338183
#> 99  500.2357   2.771608
#> 100 502.6732   2.631577
#> 101 523.9535   2.637833
#> 102 501.7604   2.829312
#> 103 510.2942   2.652171
#> 104 538.2909   2.673951
#> 105 501.1371   3.569693
#> 106 668.6175   2.507691
#> 107 500.0153   4.008561
#> 108 501.9435   2.871454
#> 109 801.8506   3.323982
#> 110 506.6593   3.667901
#> 111 505.1367   2.543428
#> 112 518.2654   2.703542
#> 113 510.5584   3.435135
#> 114 500.0443   3.051640
#> 115 500.3763   3.807983
#> 116 500.1227   2.868556
#> 117 501.0403   2.729631
#> 118 508.9306   2.852210
#> 119 622.0910   2.811923
#> 120 501.0862   2.913273
#> 121 500.2224   2.705384
#> 122 503.0059   2.558899
#> 123 500.6892   2.629644
#> 124 510.6980   3.451173
#> 125 503.8286   2.607932
#> 126 501.2916   2.863994
#> 127 500.2583   2.685735
#> 128 555.1826   2.960338
#> 129 500.6584   2.853537
#> 130 554.2470   2.588135
#> 131 517.4991   2.550223
#> 132 501.5636   3.422012
#> 133 501.5239   3.103004
#> 134 503.5992   3.028164
#> 135 709.1313   2.583381
#> 136 502.4225   2.518025
#> 137 505.5062   2.507322
#> 138 500.1657   3.430298
#> 139 501.0064   3.027312
#> 140 878.5611   2.590659
#> 141 500.4567   2.589784
#> 142 500.8093   2.965997
#> 143 754.7213   2.788433
#> 144 500.1467   2.917186
#> 145 522.9310   2.623958
#> 146 500.1766   2.961535
#> 147 502.5403   3.118602
#> 148 500.6965   3.676349
#> 149 500.4000   2.896153
#> 150 733.7289   2.595721
#> 151 507.9665   2.518870
#> 152 506.7741   3.905198
#> 153 501.8937   3.249537
#> 154 500.7029   2.909143
#> 155 767.4710   2.784823
#> 156 548.2646   2.920461
#> 157 500.0796   3.396206
#> 158 500.2582   2.649389
#> 159 557.6735   2.672389
#> 160 500.2671   2.876717
#> 161 501.6078   3.275071
#> 162 556.4961   3.128949
#> 163 856.2286   2.789089
#> 164 500.1966   3.080877
#> 165 527.2984   3.396095
#> 166 503.3510   2.679360
#> 167 500.2195   3.119834
#> 168 511.4276   2.554099
#> 169 730.5424   2.675285
#> 170 505.3212   2.520058
#> 171 870.1111   2.512638
#> 172 552.8895   2.986351
#> 173 500.6233   2.964052
#> 174 500.4648   3.648975
#> 175 519.8692   2.741407
#> 176 500.0597   3.095605
#> 177 500.5724   2.803877
#> 178 774.8195   2.892848
#> 179 503.7196   2.577818
#> 180 647.9523   2.913822
#> 181 500.0870   3.432055
#> 182 503.9946   2.845472
#> 183 500.3819   2.963270
#> 184 506.0273   2.525883
#> 185 627.0684   2.579179
#> 186 597.5456   2.631734
#> 187 500.3149   2.800278
#> 188 501.1071   2.719906
#> 189 615.5500   2.697558
#> 190 500.4602   3.899052
#> 191 512.3658   3.243641
#> 192 500.3661   2.820963
#> 193 678.2256   2.911235
#> 194 860.2871   3.216326
#> 195 502.6461   2.793833
#> 196 500.5791   3.270792
#> 197 500.0859   2.548950
#> 198 500.5205   2.967112
#> 199 500.2573   3.057301
#> 200 500.4065   2.526953
#> 201 532.1155   3.394920
#> 202 500.1234   3.275768
#> 203 509.5757   2.630256
#> 204 501.3689   3.312515
#> 205 502.6526   2.602474
#> 206 592.7349   2.599775
#> 207 760.6670   2.847096
#> 208 502.9676   2.547447
#> 209 690.1029   3.044853
#> 210 500.2148   2.682146
#> 211 513.3608   2.961277
#> 212 501.0182   2.741854
#> 213 937.4246   2.691180
#> 214 504.4160   2.562177
#> 215 500.3559   2.573867
#> 216 501.3929   2.643594
#> 217 520.3801   2.536445
#> 218 500.1058   2.803787
#> 219 500.0062   2.625581
#> 220 562.6063   2.571730
#> 221 501.3847   2.715160
#> 222 501.2307   2.699003
#> 223 702.7427   2.785099
#> 224 500.2616   4.477164
#> 225 500.3935   3.082981
#> 226 500.6725   2.638887
#> 227 500.8171   2.645604
#> 228 500.5462   2.544674
#> 229 513.3483   2.912323
#> 230 544.2314   2.838285
#> 231 935.9354   2.699010
#> 232 500.3110   2.719750
#> 233 500.2410   2.525226
#> 234 757.0744   2.573654
#> 235 612.2691   2.852019
#> 236 700.4760   2.564203
#> 237 504.1605   2.657490
#> 238 506.7641   4.059130
#> 239 505.5198   2.956955
#> 240 718.9479   2.781372
#> 241 500.0494   2.785915
#> 242 506.8617   2.757029
#> 243 523.8411   2.885513
#> 244 527.6662   3.276875
#> 245 554.1086   3.260529
#> 246 676.1427   2.662700
#> 247 500.5238   3.012015
#> 248 500.1965   3.175779
#> 249 648.7108   2.520097
#> 250 500.3334   2.693460
#> 251 505.2091   3.710722
#> 252 500.1972   2.538123
#> 253 502.6610   2.956921
#> 254 652.2472   2.848805
#> 255 500.2628   2.718572
#> 256 501.6359   2.951337
#> 257 501.0521   3.054640
#> 258 572.1193   2.527845
#> 259 509.1530   3.007881
#> 260 518.9811   3.291163
#> 261 546.7484   2.553960
#> 262 500.1051   2.781614
#> 263 500.0310   2.934642
#> 264 513.0290   2.581150
#> 265 520.1054   3.814233
#> 266 529.0612   2.661163
#> 267 502.0990   2.629013
#> 268 505.8210   2.987517
#> 269 576.6543   2.668743
#> 270 500.3912   2.528499
#> 271 503.6988   2.570833
#> 272 502.3765   3.378223
#> 273 533.6725   2.646258
#> 274 500.3157   3.386879
#> 275 500.2990   3.751233
#> 276 500.8516   3.346623
#> 277 671.2811   2.553088
#> 278 509.3647   4.092801
#> 279 501.2337   2.642447
#> 280 500.8364   2.751705
#> 281 570.2675   2.797305
#> 282 502.4563   2.638580
#> 283 514.9408   4.371569
#> 284 505.0430   3.427158
#> 285 507.7284   3.901245
#> 286 500.5869   2.721467
#> 287 523.9324   2.759407
#> 288 501.8136   3.319463
#> 289 500.7962   2.890698
#> 290 500.0296   2.901674
#> 291 500.0646   2.727476
#> 292 500.2536   2.836119
#> 293 604.3086   3.566603
#> 294 521.4873   2.525182
#> 295 500.3251   2.503011
#> 296 500.2260   2.819309
#> 297 501.2568   2.526531
#> 298 520.4869   3.654357
#> 299 612.4601   2.663133
#> 300 518.6309   2.672984
#> 301 564.4144   3.098958
#> 302 501.2147   2.632237
#> 303 501.2109   2.534725
#> 304 514.2688   2.773069
#> 305 523.5171   4.183782
#> 306 500.1046   3.616272
#> 307 502.6010   2.729911
#> 308 779.0600   3.271369
#> 309 532.5462   2.690452
#> 310 500.0059   3.056848
#> 311 500.1989   2.951192
#> 312 500.8651   2.849886
#> 313 500.7185   3.496130
#> 314 504.1975   2.720467
#> 315 500.1123   2.794389
#> 316 500.3194   2.723520
#> 317 525.8132   2.841908
#> 318 745.3062   3.259826
#> 319 500.0332   2.661801
#> 320 552.1344   3.238128
#> 321 555.6932   4.860447
#> 322 500.2097   3.324992
#> 323 520.9886   2.565568
#> 324 500.0404   2.771713
#> 325 517.5542   3.403920
#> 326 527.5672   3.204019
#> 327 500.2964   2.759354
#> 328 500.4719   2.605827
#> 329 515.3133   2.883038
#> 330 500.2252   2.650026
#> 331 501.4551   2.959263
#> 332 606.3145   3.380452
#> 333 515.3867   2.784832
#> 334 500.0829   3.408638
#> 335 586.7637   2.894116
#> 336 512.7602   2.951512
#> 337 588.2022   2.858127
#> 338 500.3139   2.540774
#> 339 539.8917   2.712088
#> 340 500.2866   2.748360
#> 341 507.2095   3.735399
#> 342 509.8805   4.236744
#> 343 659.4966   2.722132
#> 344 501.0009   2.982357
#> 345 500.0424   3.540532
#> 346 526.6119   2.898130
#> 347 732.2927   2.891649
#> 348 502.8469   2.657279
#> 349 575.1598   3.844470
#> 350 500.6513   2.917131
#> 351 500.1943   2.678940
#> 352 502.3301   3.204102
#> 353 785.6664   2.905517
#> 354 500.2567   2.699173
#> 355 639.6547   2.971291
#> 356 733.2136   2.797967
#> 357 502.5655   2.875545
#> 358 500.0598   2.540746
#> 359 686.8295   2.612105
#> 360 571.9767   3.201154
#> 361 510.0649   2.501329
#> 362 766.6189   2.760715
#> 363 500.5042   2.558704
#> 364 503.1715   2.698752
#> 365 500.0455   3.537648
#> 366 934.0833   3.234611
#> 367 500.0003   2.673364
#> 368 507.1175   2.747455
#> 369 525.0229   2.560307
#> 370 500.1844   3.013319
#> 371 500.2316   3.258699
#> 372 506.3150   2.549810
#> 373 501.3986   3.408295
#> 374 502.0963   3.015533
#> 375 502.7659   2.589057
#> 376 775.8211   4.067364
#> 377 505.9341   2.536917
#> 378 501.4462   3.422305
#> 379 504.9061   3.429464
#> 380 516.7817   3.348808
#> 381 500.1779   2.663103
#> 382 729.3753   2.534754
#> 383 532.6566   3.760116
#> 384 500.2107   2.723093
#> 385 524.5287   2.758769
#> 386 500.0017   2.888494
#> 387 500.1649   2.508207
#> 388 507.7556   3.393677
#> 389 500.1707   2.568286
#> 390 687.8636   2.592288
#> 391 503.2121   3.498202
#> 392 500.0686   2.672492
#> 393 583.7956   2.773153
#> 394 540.2200   2.549342
#> 395 522.7443   2.620632
#> 396 500.1084   3.032268
#> 397 890.3959   2.525417
#> 398 506.2866   2.517357
#> 399 726.8106   2.714916
#> 400 509.5015   3.161290
#> 401 500.3213   2.717321
#> 402 822.5787   2.714936
#> 403 501.2454   2.530689
#> 404 557.4620   4.591458
#> 405 500.0432   2.513654
#> 406 501.1335   3.075343
#> 407 501.1670   2.561871
#> 408 607.1376   2.554593
#> 409 536.2645   2.962570
#> 410 500.3524   2.829283
#> 411 502.0739   2.607984
#> 412 508.0433   3.165784
#> 413 500.0056   3.366104
#> 414 500.7619   2.928557
#> 415 532.8044   2.528516
#> 416 557.3345   2.991757
#> 417 527.8187   3.272570
#> 418 504.0058   2.994638
#> 419 500.1315   2.594878
#> 420 500.4861   2.935656
#> 421 576.3063   2.806355
#> 422 500.0867   2.579339
#> 423 500.3675   2.888516
#> 424 500.4651   3.267869
#> 425 617.4655   3.551875
#> 426 500.1784   2.595726
#> 427 514.2711   3.455624
#> 428 500.2997   3.043452
#> 429 578.0471   2.592531
#> 430 511.8827   2.503725
#> 431 503.2330   3.941723
#> 432 564.8075   2.645906
#> 433 502.7576   2.898197
#> 434 528.1458   2.839734
#> 435 500.4654   3.002580
#> 436 587.4475   3.343087
#> 437 500.1449   2.623662
#> 438 501.3182   2.813709
#> 439 500.0041   3.135119
#> 440 687.6430   3.570226
#> 441 500.1510   2.608045
#> 442 500.4683   2.883678
#> 443 510.6831   2.594668
#> 444 510.3427   2.926497
#> 445 503.9321   3.257546
#> 446 555.3954   2.621087
#> 447 516.4849   2.604492
#> 448 535.4356   2.899570
#> 449 500.0267   2.863248
#> 450 500.5233   2.620919
#> 451 500.0394   3.056602
#> 452 500.1534   2.566710
#> 453 606.5518   2.931079
#> 454 637.8435   2.599341
#> 455 508.1848   3.152296
#> 456 523.1308   2.690204
#> 457 502.1495   2.507317
#> 458 785.2307   2.739956
#> 459 502.5918   2.974956
#> 460 500.6804   3.226251
#> 461 509.3756   3.146206
#> 462 500.8265   3.085674
#> 463 500.3898   2.784292
#> 464 547.5849   2.802410
#> 465 522.0109   2.915647
#> 466 510.3846   3.100838
#> 467 508.5005   3.040351
#> 468 526.8564   3.322307
#> 469 503.4410   2.596570
#> 470 502.0250   2.875697
#> 471 502.2825   3.346050
#> 472 509.2122   2.704285
#> 473 500.0801   2.728719
#> 474 500.4858   2.516168
#> 475 535.1666   3.203880
#> 476 552.0778   2.752134
#> 477 513.6643   3.432680
#> 478 879.4973   3.306429
#> 479 871.2471   2.882689
#> 480 500.8172   2.906504
#> 481 501.7333   2.512670
#> 482 636.4636   3.322419
#> 483 500.4231   3.303477
#> 484 502.8863   2.947002
#> 485 500.3268   2.709624
#> 486 505.2537   6.057711
#> 487 502.6317   3.910788
#> 488 511.5239   2.591047
#> 489 500.6655   2.659737
#> 490 510.5058   3.025005
#> 491 500.7597   2.533524
#> 492 667.9662   2.918371
#> 493 515.8760   3.808678
#> 494 504.3983   2.511065
#> 495 511.4156   2.707551
#> 496 559.5636   3.706644
#> 497 516.6940   2.767499
#> 498 500.3504   3.448808
#> 499 500.3608   3.404834
#> 500 518.2579   2.940672
#> 501 500.1357   2.824092
#> 502 515.5712   2.659426
#> 503 538.8773   2.600610
#> 504 703.8086   2.673382
#> 505 503.4410   2.761246
#> 506 509.0626   3.144855
#> 507 818.8232   3.267913
#> 508 619.2064   3.548817
#> 509 535.9463   3.105742
#> 510 500.0439   2.682606
#> 511 571.3875   3.589533
#> 512 517.8749   2.872688
#> 513 500.0254   2.609692
#> 514 521.2756   3.272506
#> 515 654.1226   2.511441
#> 516 513.5554   2.814320
#> 517 500.6259   2.903739
#> 518 529.2568   2.631711
#> 519 500.7593   3.458049
#> 520 502.1399   2.835125
#> 521 500.6227   2.822638
#> 522 683.0121   4.102198
#> 523 508.8898   3.301043
#> 524 500.0390   3.230658
#> 525 500.0758   2.503700
#> 526 500.3235   2.689737
#> 527 566.9311   2.684986
#> 528 505.1508   3.064188
#> 529 606.3105   2.606439
#> 530 502.3574   2.820029
#> 531 522.8854   2.505356
#> 532 500.0012   2.774752
#> 533 568.1739   3.182297
#> 534 500.0011   2.658866
#> 535 517.2576   2.870938
#> 536 502.6793   3.062298
#> 537 509.3746   2.958577
#> 538 501.3558   2.684852
#> 539 502.4803   2.531226
#> 540 500.3118   3.521466
#> 541 500.0957   2.860010
#> 542 508.3609   2.509814
#> 543 811.6988   2.855841
#> 544 500.7275   2.536612
#> 545 501.9406   3.746826
#> 546 501.6137   3.105374
#> 547 625.1042   3.702928
#> 548 607.3009   2.722173
#> 549 506.9187   2.886407
#> 550 512.9958   2.645032
#> 551 500.6040   2.635505
#> 552 500.2689   2.638638
#> 553 500.3775   2.763892
#> 554 500.2649   2.724767
#> 555 502.7783   2.585131
#> 556 500.1001   2.668806
#> 557 513.9757   2.807754
#> 558 500.4674   2.747626
#> 559 504.4822   2.955800
#> 560 502.2594   2.773140
#> 561 500.0944   2.695753
#> 562 677.5242   2.599921
#> 563 986.7342   2.682192
#> 564 500.0908   2.539333
#> 565 507.0972   4.372040
#> 566 607.8064   2.830337
#> 567 510.9360   3.100416
#> 568 500.3868   2.932579
#> 569 500.3742   2.741657
#> 570 500.3192   3.770634
#> 571 514.3207   2.656346
#> 572 500.0428   2.525609
#> 573 503.9024   2.921220
#> 574 501.1924   3.367022
#> 575 500.1389   3.001239
#> 576 500.8141   2.585824
#> 577 500.1578   2.573685
#> 578 892.6978   2.877742
#> 579 507.1257   4.624129
#> 580 540.2920   3.726984
#> 581 509.0503   3.324444
#> 582 713.2454   2.547173
#> 583 500.7523   2.720025
#> 584 508.2531   3.325979
#> 585 500.0197   3.653588
#> 586 527.7082   2.586018
#> 587 500.5228   2.936455
#> 588 514.1031   2.656240
#> 589 515.8511   2.534041
#> 590 500.0004   2.676372
#> 591 587.5762   3.887816
#> 592 516.4694   3.028193
#> 593 500.3745   4.584050
#> 594 500.1403   2.984514
#> 595 500.2016   3.417480
#> 596 879.6228   2.656553
#> 597 509.4052   2.712133
#> 598 500.0303   2.927768
#> 599 511.2197   3.119275
#> 600 500.3800   3.306476
#> 601 500.9804   2.791384
#> 602 752.1458   2.645725
#> 603 502.6133   2.731175
#> 604 505.1998   2.867698
#> 605 501.5332   3.257506
#> 606 500.0277   3.140012
#> 607 500.1159   2.638687
#> 608 500.0089   2.725798
#> 609 500.0493   2.986186
#> 610 500.3569   3.152806
#> 611 966.8479   2.559676
#> 612 580.3207   2.588528
#> 613 692.9372   3.162870
#> 614 643.7400   3.773253
#> 615 527.4813   3.199346
#> 616 500.1850   4.147863
#> 617 500.0010   2.516917
#> 618 741.7148   2.689014
#> 619 500.9181   2.725266
#> 620 500.0743   3.454757
#> 621 500.1534   2.768593
#> 622 500.0064   2.898675
#> 623 501.4411   3.224920
#> 624 500.1505   2.836280
#> 625 500.8776   2.674413
#> 626 524.7526   2.915277
#> 627 656.9534   3.133026
#> 628 500.8343   2.929230
#> 629 579.9104   2.897119
#> 630 500.5531   2.502806
#> 631 500.0101   2.743769
#> 632 500.2607   3.774525
#> 633 500.4672   3.532405
#> 634 518.6853   2.744954
#> 635 500.1918   2.840000
#> 636 589.0493   3.197349
#> 637 500.3277   3.096829
#> 638 518.8061   2.698796
#> 639 501.0164   2.510908
#> 640 508.0174   2.515638
#> 641 503.2806   2.620519
#> 642 500.3695   3.131042
#> 643 500.4929   2.764619
#> 644 500.1694   2.698976
#> 645 507.8289   2.585912
#> 646 500.0554   2.817003
#> 647 500.1004   2.929894
#> 648 500.7422   2.636435
#> 649 500.5128   2.567678
#> 650 582.3531   2.623436
#> 651 574.8352   2.939864
#> 652 823.7787   3.343836
#> 653 503.1220   2.869535
#> 654 693.5528   3.472850
#> 655 539.7949   2.735846
#> 656 523.0704   2.716234
#> 657 500.0153   2.983420
#> 658 500.0288   2.519403
#> 659 828.1682   2.748313
#> 660 500.1407   2.560595
#> 661 500.1405   3.068721
#> 662 500.2882   2.930134
#> 663 500.1018   2.633511
#> 664 500.2593   2.718625
#> 665 501.5953   3.115973
#> 666 518.1457   2.594038
#> 667 519.5798   3.340872
#> 668 533.7352   3.626364
#> 669 504.1236   2.755936
#> 670 501.5746   2.510784
#> 671 556.2626   2.666550
#> 672 500.0806   3.318746
#> 673 500.3812   2.703555
#> 674 500.0665   2.684401
#> 675 500.0173   2.648805
#> 676 501.5482   3.150739
#> 677 643.8117   2.582477
#> 678 501.0807   2.690465
#> 679 500.6052   4.304290
#> 680 512.8943   3.320836
#> 681 504.6511   2.565702
#> 682 614.1273   2.983075
#> 683 516.6834   2.660058
#> 684 501.1781   3.035195
#> 685 501.2615   3.007366
#> 686 584.0937   2.829718
#> 687 500.0167   2.741480
#> 688 500.3600   2.788959
#> 689 545.3109   2.692452
#> 690 509.6513   2.984035
#> 691 501.4091   2.532496
#> 692 509.1636   3.648381
#> 693 502.1875   3.780382
#> 694 504.7502   3.427004
#> 695 505.9124   2.522759
#> 696 501.1954   2.547780
#> 697 765.7955   2.679921
#> 698 546.5644   2.833959
#> 699 500.1555   2.675320
#> 700 501.4951   2.558636
#> 701 500.0965   2.721843
#> 702 507.4091   2.867699
#> 703 500.1955   3.023393
#> 704 623.9313   2.817122
#> 705 500.0306   2.509919
#> 706 560.6546   4.117229
#> 707 500.1332   2.678733
#> 708 596.0558   2.542467
#> 709 512.2022   3.685846
#> 710 500.0107   3.152428
#> 711 752.3014   3.025107
#> 712 500.0328   2.614909
#> 713 522.4672   3.582790
#> 714 508.7127   2.665260
#> 715 687.2505   2.699140
#> 716 511.4282   2.605811
#> 717 500.6938   3.029115
#> 718 509.0782   2.813886
#> 719 537.7350   2.793432
#> 720 590.8102   2.625829
#> 721 560.1496   3.800084
#> 722 526.8520   2.633998
#> 723 502.4884   2.536955
#> 724 549.2559   3.324684
#> 725 560.2730   3.568144
#> 726 691.7003   2.597451
#> 727 500.0067   2.815757
#> 728 500.3878   3.204307
#> 729 537.2324   2.737285
#> 730 500.0031   3.619649
#> 731 649.0011   2.678522
#> 732 503.4847   2.896119
#> 733 541.4615   2.587746
#> 734 528.5530   2.866253
#> 735 504.9102   3.243389
#> 736 500.0602   3.040751
#> 737 508.9706   2.777121
#> 738 500.1645   3.100892
#> 739 501.5796   3.216338
#> 740 515.6644   3.418208
#> 741 500.8871   2.895528
#> 742 500.1882   3.190660
#> 743 774.4251   2.846003
#> 744 560.0887   2.588752
#> 745 741.7930   2.522133
#> 746 501.4086   2.822527
#> 747 501.1585   2.955408
#> 748 505.4410   2.664830
#> 749 521.7042   2.630187
#> 750 501.1590   3.313277
#> 751 500.7306   3.327465
#> 752 501.2593   2.667086
#> 753 593.9295   2.604455
#> 754 500.2825   3.551387
#> 755 500.1552   2.586670
#> 756 503.1505   4.880047
#> 757 500.2263   3.318365
#> 758 514.1917   3.627781
#> 759 581.0818   3.396764
#> 760 738.2086   3.521618
#> 761 581.4327   3.126403
#> 762 500.2298   3.190774
#> 763 509.3317   2.727440
#> 764 500.1674   2.784295
#> 765 500.3100   3.044930
#> 766 511.4868   2.815538
#> 767 503.7854   2.949142
#> 768 500.0055   2.610497
#> 769 506.5704   2.703736
#> 770 575.6887   2.629377
#> 771 524.5857   3.066876
#> 772 513.3259   2.760468
#> 773 713.3994   2.624583
#> 774 501.2491   3.327550
#> 775 512.3616   2.695270
#> 776 500.5856   2.882335
#> 777 531.2040   2.719105
#> 778 695.6898   3.202435
#> 779 500.3776   3.112215
#> 780 511.1254   2.950419
#> 781 500.0737   2.895272
#> 782 505.6784   2.917265
#> 783 645.1471   2.747559
#> 784 593.7349   2.934650
#> 785 542.8190   2.500951
#> 786 500.0113   2.810928
#> 787 500.0290   2.611285
#> 788 502.9112   2.514447
#> 789 520.9231   2.910430
#> 790 500.1278   2.628509
#> 791 516.5121   2.946219
#> 792 557.2475   2.577833
#> 793 500.7829   2.610033
#> 794 500.2886   2.714062
#> 795 553.2674   2.820992
#> 796 500.3890   2.929911
#> 797 502.6209   2.929806
#> 798 504.4998   2.694629
#> 799 502.2775   2.816812
#> 800 505.7777   3.017860
#> 801 500.1435   2.955511
#> 802 663.3108   2.728119
#> 803 507.3858   4.424119
#> 804 842.0979   2.650794
#> 805 623.7869   2.673377
#> 806 500.1287   3.086719
#> 807 640.0407   3.023210
#> 808 633.6072   2.570194
#> 809 515.2348   2.644386
#> 810 500.1880   2.571352
#> 811 500.3435   2.895028
#> 812 500.2228   2.563165
#> 813 537.5764   2.831134
#> 814 512.3853   2.856817
#> 815 867.8547   2.660135
#> 816 500.1527   3.335587
#> 817 501.9941   2.540951
#> 818 598.2243   3.299689
#> 819 500.0323   2.897908
#> 820 501.7819   2.595732
#> 821 500.5400   3.753827
#> 822 501.4519   2.925063
#> 823 500.1770   2.751757
#> 824 500.7541   2.992194
#> 825 519.0642   3.013797
#> 826 500.2927   2.531670
#> 827 500.4202   2.633509
#> 828 563.4502   3.290460
#> 829 503.1647   3.045037
#> 830 557.8778   3.120501
#> 831 666.9453   2.651374
#> 832 507.4563   2.732469
#> 833 500.2971   2.560338
#> 834 500.9447   2.871462
#> 835 520.4859   2.602248
#> 836 500.0646   2.703475
#> 837 500.1273   2.974587
#> 838 507.5925   3.054988
#> 839 503.3438   3.157970
#> 840 500.8712   2.857354
#> 841 666.9471   3.137563
#> 842 502.7744   2.587561
#> 843 511.7033   2.969747
#> 844 501.9353   2.775263
#> 845 500.0989   3.619746
#> 846 530.8761   2.579398
#> 847 506.3800   3.045201
#> 848 500.5727   2.809204
#> 849 513.3905   2.706668
#> 850 502.4124   3.010446
#> 851 500.0385   3.270896
#> 852 504.6084   2.925349
#> 853 512.9225   2.607522
#> 854 525.7647   3.165176
#> 855 517.5453   3.114777
#> 856 504.7681   2.764767
#> 857 500.6945   2.644268
#> 858 501.8526   3.654941
#> 859 545.7623   2.814261
#> 860 518.4895   3.285723
#> 861 634.5951   3.938109
#> 862 508.1947   3.542167
#> 863 601.6544   2.577362
#> 864 500.1053   3.204463
#> 865 523.6123   2.761289
#> 866 501.2173   2.675244
#> 867 984.1908   2.545431
#> 868 500.6139   2.736600
#> 869 505.1551   3.455479
#> 870 528.4832   2.521613
#> 871 503.6857   2.906381
#> 872 537.7996   3.974323
#> 873 501.5968   2.566326
#> 874 500.0482   3.213483
#> 875 503.2442   2.649155
#> 876 548.2699   2.758144
#> 877 502.7199   2.654135
#> 878 500.0127   3.699054
#> 879 548.3765   4.010756
#> 880 500.2656   2.535818
#> 881 500.0634   2.680059
#> 882 742.7763   3.461393
#> 883 501.2397   2.779663
#> 884 623.6156   3.130158
#> 885 505.9165   3.389741
#> 886 501.8762   3.059335
#> 887 500.7237   2.895858
#> 888 573.2243   3.023919