Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   502.7483   3.311227
#> 2   500.4916   3.222430
#> 3   983.0114   2.704471
#> 4   500.5592   2.864086
#> 5   500.0047   2.584595
#> 6   549.5256   2.858121
#> 7   783.0687   2.573435
#> 8   500.1107   3.071178
#> 9   508.6635   2.508219
#> 10  500.3328   2.767538
#> 11  535.0533   2.572803
#> 12  500.0759   2.915345
#> 13  505.3859   3.094060
#> 14  500.0126   3.584723
#> 15  796.8481   2.584175
#> 16  526.6339   2.878389
#> 17  501.0742   2.589342
#> 18  500.0936   2.617918
#> 19  526.1046   2.963353
#> 20  503.1663   2.730176
#> 21  518.2395   2.737245
#> 22  500.2271   2.567950
#> 23  502.7666   3.944387
#> 24  501.2271   2.600483
#> 25  506.2856   2.560117
#> 26  529.2492   2.616033
#> 27  583.5346   2.736124
#> 28  500.7485   3.321142
#> 29  500.5957   2.891552
#> 30  528.0660   3.193368
#> 31  500.1608   2.680376
#> 32  504.0397   2.935467
#> 33  500.6583   3.348128
#> 34  510.6596   2.708611
#> 35  505.8009   2.693940
#> 36  500.0547   3.003808
#> 37  502.2842   2.566886
#> 38  503.3811   2.631706
#> 39  502.7200   2.626514
#> 40  500.5342   3.270516
#> 41  500.1558   2.828697
#> 42  500.4966   3.358763
#> 43  724.0596   3.245473
#> 44  508.6072   2.519224
#> 45  500.1117   3.281554
#> 46  501.5879   2.603468
#> 47  500.1690   2.772064
#> 48  510.8406   4.573832
#> 49  500.9538   3.780578
#> 50  585.5147   2.809633
#> 51  501.7402   2.868295
#> 52  531.1458   2.828812
#> 53  501.3594   2.558777
#> 54  570.3275   2.892221
#> 55  507.2646   2.514476
#> 56  501.4930   2.820169
#> 57  530.5371   2.945320
#> 58  500.3218   3.116784
#> 59  532.8767   2.549206
#> 60  511.5648   2.775722
#> 61  503.1790   3.589587
#> 62  500.7736   2.566737
#> 63  573.6697   2.653041
#> 64  598.1587   4.121899
#> 65  500.2590   2.569280
#> 66  505.5193   2.838531
#> 67  621.5416   2.979199
#> 68  512.5812   3.619888
#> 69  508.5569   3.015163
#> 70  510.0442   2.726671
#> 71  808.8971   2.623672
#> 72  566.7409   3.568100
#> 73  605.8989   2.640141
#> 74  500.0540   2.527228
#> 75  599.2556   3.959172
#> 76  553.4084   3.161117
#> 77  500.0274   2.815007
#> 78  520.4800   3.359138
#> 79  760.5425   2.812016
#> 80  501.0404   2.532657
#> 81  706.7452   2.545120
#> 82  510.7525   2.666383
#> 83  934.1649   4.278852
#> 84  532.4389   2.624137
#> 85  754.6080   3.223389
#> 86  504.3520   2.531022
#> 87  618.6013   2.958193
#> 88  565.0456   2.545719
#> 89  500.1100   2.631834
#> 90  500.1751   2.530257
#> 91  500.2586   2.913836
#> 92  500.0109   2.746663
#> 93  504.2737   2.622814
#> 94  516.0573   4.349108
#> 95  501.2661   2.552374
#> 96  693.7957   3.209064
#> 97  517.0474   2.741204
#> 98  500.3573   3.102933
#> 99  518.4710   2.861166
#> 100 502.7006   2.684259
#> 101 622.5241   2.559402
#> 102 504.0032   3.007940
#> 103 500.4246   3.491866
#> 104 503.8503   2.761207
#> 105 686.3908   2.892505
#> 106 989.4267   3.393030
#> 107 550.8956   3.080303
#> 108 502.9869   2.808853
#> 109 511.7054   2.693366
#> 110 548.5731   2.784652
#> 111 530.8073   2.799110
#> 112 501.1121   2.578141
#> 113 500.0015   2.746054
#> 114 500.3434   2.517030
#> 115 605.3828   2.740547
#> 116 507.4973   3.317880
#> 117 504.0876   3.405782
#> 118 500.6439   4.168501
#> 119 500.0065   4.518767
#> 120 500.1352   2.965414
#> 121 522.4952   2.734155
#> 122 503.2781   2.769527
#> 123 502.2897   2.619707
#> 124 533.5467   2.580970
#> 125 538.1949   2.717885
#> 126 500.2317   3.382869
#> 127 749.7518   2.736904
#> 128 737.3941   2.750600
#> 129 507.7162   2.570756
#> 130 500.2089   3.021316
#> 131 524.3780   2.977945
#> 132 719.8271   2.573347
#> 133 500.3311   3.586700
#> 134 620.7983   3.455166
#> 135 557.1722   2.535811
#> 136 507.6534   2.538407
#> 137 558.1696   2.597837
#> 138 500.6125   2.837366
#> 139 500.2150   2.870598
#> 140 558.5101   2.569128
#> 141 535.7275   2.557754
#> 142 507.7912   2.633835
#> 143 505.1144   2.985358
#> 144 503.7650   2.620639
#> 145 513.7825   2.888181
#> 146 511.4746   2.869266
#> 147 510.2328   2.599178
#> 148 623.5750   3.166862
#> 149 500.2801   3.395912
#> 150 707.5436   2.708103
#> 151 500.2499   2.680799
#> 152 505.6518   2.607978
#> 153 500.1780   2.615665
#> 154 638.6676   2.585377
#> 155 500.5846   2.616191
#> 156 500.9373   2.543324
#> 157 500.1141   2.610787
#> 158 603.6853   2.755528
#> 159 543.4232   3.882325
#> 160 500.8185   2.804817
#> 161 544.7558   2.844445
#> 162 504.3844   2.652375
#> 163 501.3772   2.756174
#> 164 989.8255   2.732114
#> 165 500.5129   2.638759
#> 166 500.3069   3.053197
#> 167 509.4786   2.526994
#> 168 516.6709   3.712433
#> 169 566.7037   3.250919
#> 170 833.5700   3.813251
#> 171 500.0065   3.367203
#> 172 511.9821   2.659300
#> 173 508.9829   2.836182
#> 174 500.2341   2.926980
#> 175 500.2062   2.547353
#> 176 500.0687   3.066546
#> 177 521.5042   3.131309
#> 178 500.2242   3.000520
#> 179 500.1463   3.183522
#> 180 516.4656   2.600691
#> 181 812.2899   2.994533
#> 182 500.1550   3.012782
#> 183 500.0143   2.620475
#> 184 755.6758   2.679050
#> 185 500.5187   2.530243
#> 186 521.6973   3.091944
#> 187 505.1082   2.618967
#> 188 508.3000   3.569580
#> 189 509.0866   2.582200
#> 190 500.2329   4.090714
#> 191 503.0976   3.233222
#> 192 501.0444   2.525318
#> 193 661.4180   2.953195
#> 194 500.5086   2.654536
#> 195 500.5189   2.581441
#> 196 522.3919   2.726295
#> 197 508.6368   2.656834
#> 198 502.5037   2.582470
#> 199 517.8827   2.556320
#> 200 502.0156   3.292108
#> 201 500.0136   2.526710
#> 202 522.6597   3.059028
#> 203 510.1899   2.809777
#> 204 503.8411   2.890690
#> 205 823.8641   2.655695
#> 206 500.1350   2.807169
#> 207 653.5010   2.761073
#> 208 500.1690   3.019452
#> 209 500.1440   2.678989
#> 210 841.1985   3.668707
#> 211 761.6022   3.054581
#> 212 500.2805   2.696619
#> 213 501.4397   3.130924
#> 214 540.5302   2.563992
#> 215 508.1430   3.062446
#> 216 500.3378   2.754045
#> 217 500.5754   2.562395
#> 218 500.0588   3.343087
#> 219 505.8684   3.782621
#> 220 500.4199   3.330597
#> 221 500.1195   2.774889
#> 222 506.8943   2.978737
#> 223 500.4747   2.549010
#> 224 500.8658   2.591088
#> 225 522.6702   2.785724
#> 226 501.4303   2.652711
#> 227 500.2834   2.604921
#> 228 702.5919   3.234422
#> 229 632.2757   3.805749
#> 230 504.6489   3.616905
#> 231 879.1647   2.672602
#> 232 500.6391   2.744297
#> 233 501.2091   2.706279
#> 234 508.1932   2.615872
#> 235 501.1455   2.582621
#> 236 500.8238   2.514089
#> 237 501.1720   3.427939
#> 238 500.2543   2.687755
#> 239 520.0313   3.181224
#> 240 667.2470   2.516498
#> 241 640.8114   2.533124
#> 242 503.4257   2.811349
#> 243 650.4870   2.871631
#> 244 502.7089   2.640194
#> 245 500.4571   3.186568
#> 246 500.6153   2.935473
#> 247 501.4840   2.552276
#> 248 500.7575   3.666023
#> 249 500.2765   3.116499
#> 250 507.8452   2.907807
#> 251 542.9463   3.138498
#> 252 500.0475   2.874024
#> 253 830.7451   2.560930
#> 254 500.0542   2.635543
#> 255 505.3714   2.637408
#> 256 526.8370   3.101860
#> 257 881.6852   3.290345
#> 258 502.5477   3.714619
#> 259 500.0713   2.546850
#> 260 547.3204   2.562002
#> 261 500.9993   2.853014
#> 262 500.6232   3.206153
#> 263 500.0047   2.756716
#> 264 840.9690   3.399514
#> 265 501.8467   3.818429
#> 266 500.2355   2.565585
#> 267 500.3286   2.750653
#> 268 707.3151   2.605674
#> 269 500.8566   2.652098
#> 270 500.5614   2.879699
#> 271 503.6929   2.501779
#> 272 500.0400   2.774938
#> 273 501.0369   2.542915
#> 274 500.1103   3.229727
#> 275 500.2179   3.225560
#> 276 500.0258   2.656729
#> 277 500.3196   2.737116
#> 278 508.6371   2.638077
#> 279 668.0502   3.393017
#> 280 500.5166   2.507981
#> 281 500.7738   2.677410
#> 282 501.9321   2.743603
#> 283 500.5258   2.576059
#> 284 500.4321   3.077657
#> 285 502.7811   2.906911
#> 286 536.7862   2.865075
#> 287 512.5296   2.612738
#> 288 723.6991   3.521624
#> 289 702.9024   2.696542
#> 290 512.3867   2.889094
#> 291 551.9395   4.093041
#> 292 504.3185   3.012111
#> 293 711.1374   2.610945
#> 294 731.8650   3.049490
#> 295 500.3047   3.785602
#> 296 506.0465   3.284930
#> 297 515.4929   2.777897
#> 298 500.7938   2.568054
#> 299 962.4724   2.712767
#> 300 500.2340   2.636447
#> 301 539.5819   2.855807
#> 302 500.0979   3.081770
#> 303 501.3191   2.965589
#> 304 500.0233   2.766600
#> 305 503.4412   3.517381
#> 306 501.9203   3.925858
#> 307 501.5411   2.581134
#> 308 857.5978   2.826161
#> 309 500.1028   2.722401
#> 310 609.7729   2.641764
#> 311 540.8039   2.970314
#> 312 500.3862   2.751286
#> 313 619.7480   3.026452
#> 314 556.7158   2.776777
#> 315 500.0403   3.277686
#> 316 524.9032   2.761357
#> 317 500.4193   3.035164
#> 318 722.1913   2.795418
#> 319 501.0000   2.607874
#> 320 500.8639   2.547457
#> 321 501.7541   3.776029
#> 322 506.6094   2.974404
#> 323 510.6805   2.517480
#> 324 680.8868   3.104629
#> 325 500.2458   3.126601
#> 326 500.0499   3.356127
#> 327 500.0996   3.611668
#> 328 505.3083   2.580145
#> 329 502.9271   2.790072
#> 330 579.3033   3.356430
#> 331 514.7081   3.289151
#> 332 506.4005   2.980467
#> 333 500.3431   2.520572
#> 334 502.2832   2.682358
#> 335 867.6832   2.611289
#> 336 503.4226   3.174792
#> 337 537.5635   2.789987
#> 338 572.8779   2.583627
#> 339 500.6634   2.670656
#> 340 500.5693   3.203791
#> 341 510.0557   2.932734
#> 342 500.6647   4.104570
#> 343 501.2396   2.634149
#> 344 507.6149   2.613455
#> 345 500.4286   4.207104
#> 346 500.1969   2.750985
#> 347 501.3942   3.210160
#> 348 515.0924   2.885070
#> 349 501.6499   3.862736
#> 350 505.9049   2.805480
#> 351 500.8493   3.223879
#> 352 507.8381   2.820802
#> 353 502.9935   2.979980
#> 354 515.5084   3.070176
#> 355 500.0200   2.745803
#> 356 507.1893   2.759462
#> 357 500.0045   2.582191
#> 358 500.2021   4.153367
#> 359 519.9970   3.352457
#> 360 500.1567   2.611109
#> 361 515.2498   3.223130
#> 362 508.5421   2.537620
#> 363 625.4518   3.373326
#> 364 502.7267   2.567572
#> 365 526.2237   2.582261
#> 366 503.2638   4.781309
#> 367 513.0514   3.121556
#> 368 532.0827   2.620591
#> 369 538.0179   2.594475
#> 370 527.7125   2.672274
#> 371 500.3173   2.506176
#> 372 500.5763   4.030804
#> 373 501.4661   2.546108
#> 374 500.0372   3.132594
#> 375 505.4925   2.676875
#> 376 500.4155   3.440435
#> 377 502.7790   2.807729
#> 378 500.1117   3.442725
#> 379 500.0109   2.506926
#> 380 500.1900   2.608190
#> 381 506.9117   4.016671
#> 382 519.8682   3.105996
#> 383 523.2922   3.643217
#> 384 500.1485   2.558382
#> 385 502.0677   3.178565
#> 386 534.9830   3.503966
#> 387 500.0045   2.558622
#> 388 597.2748   2.885318
#> 389 648.2554   2.508490
#> 390 681.8867   2.810931
#> 391 509.5657   3.184126
#> 392 511.7250   2.516053
#> 393 520.5351   3.368461
#> 394 502.6955   3.350031
#> 395 877.4721   2.711276
#> 396 500.9326   2.850589
#> 397 500.0814   3.059567
#> 398 507.5417   2.816954
#> 399 504.5459   2.657605
#> 400 505.6367   2.562700
#> 401 500.1636   2.981691
#> 402 500.2845   2.729565
#> 403 507.8646   2.736419
#> 404 500.2605   2.910226
#> 405 510.4733   3.851658
#> 406 561.6647   2.610157
#> 407 500.0109   2.665462
#> 408 503.3137   2.791858
#> 409 503.4302   2.565880
#> 410 503.2569   2.728751
#> 411 692.5115   2.756346
#> 412 503.9618   2.620501
#> 413 506.2968   2.716141
#> 414 503.9859   2.783800
#> 415 549.9319   3.050002
#> 416 934.3300   4.084820
#> 417 528.2905   3.767471
#> 418 500.5104   2.644067
#> 419 500.6354   2.644241
#> 420 564.8837   2.848288
#> 421 502.3817   3.371927
#> 422 514.7704   2.904688
#> 423 500.9871   2.848438
#> 424 501.1793   2.691350
#> 425 500.0264   2.864869
#> 426 500.6514   2.884868
#> 427 503.8954   2.526655
#> 428 509.7340   3.289709
#> 429 512.3333   3.503668
#> 430 511.6589   2.582844
#> 431 500.6236   3.167476
#> 432 500.0735   3.117215
#> 433 519.5615   3.093597
#> 434 500.1927   2.505752
#> 435 500.0069   2.512088
#> 436 641.1013   2.811926
#> 437 500.3981   3.153520
#> 438 500.5820   3.284884
#> 439 500.2834   2.902422
#> 440 758.0884   2.528046
#> 441 830.8857   2.728553
#> 442 501.9096   3.148719
#> 443 500.3532   2.770868
#> 444 520.1629   3.375576
#> 445 591.8101   2.542478
#> 446 500.2483   2.929659
#> 447 517.9670   2.513722
#> 448 514.1564   3.107003
#> 449 843.5084   2.513278
#> 450 528.7309   2.639386
#> 451 706.7837   2.885165
#> 452 773.8393   2.752013
#> 453 500.0188   2.828075
#> 454 546.1054   4.243093
#> 455 500.1523   4.487130
#> 456 719.1447   2.599041
#> 457 518.3462   2.942811
#> 458 500.2614   3.423212
#> 459 502.9258   2.755147
#> 460 531.5570   2.574498
#> 461 529.3158   2.995649
#> 462 501.3366   3.121941
#> 463 531.2423   3.231214
#> 464 501.5684   2.961913
#> 465 526.0282   3.012446
#> 466 500.5880   2.511621
#> 467 500.7035   2.659268
#> 468 679.6778   2.599667
#> 469 599.0810   2.562222
#> 470 560.4991   3.318351
#> 471 500.2317   2.559481
#> 472 500.2347   2.598034
#> 473 500.0867   2.596882
#> 474 500.2275   2.639592
#> 475 568.1678   2.515731
#> 476 500.0341   2.829896
#> 477 505.9466   2.512813
#> 478 500.1543   2.550748
#> 479 532.8648   2.709454
#> 480 500.7409   2.740152
#> 481 504.5002   2.562558
#> 482 501.2304   2.866996
#> 483 535.3151   2.667076
#> 484 500.8291   3.276150
#> 485 506.4481   2.666984
#> 486 520.8271   3.071382
#> 487 507.6048   3.473247
#> 488 509.9214   2.546568
#> 489 938.7324   3.501689
#> 490 503.1149   2.596573
#> 491 500.0900   3.510655
#> 492 557.4368   2.572466
#> 493 718.9545   2.524461
#> 494 734.4412   2.526208
#> 495 500.0237   2.762597
#> 496 503.0105   3.671144
#> 497 510.4865   2.558134
#> 498 503.7277   2.573528
#> 499 500.0641   2.624863
#> 500 500.9865   2.554950
#> 501 504.4037   2.984757
#> 502 501.6476   2.622845
#> 503 500.3630   4.476140
#> 504 538.8667   2.758538
#> 505 500.4451   2.518068
#> 506 559.2540   2.628243
#> 507 544.3947   3.140561
#> 508 500.4736   2.523517
#> 509 509.9962   2.630860
#> 510 503.6043   2.798773
#> 511 522.5861   2.613267
#> 512 503.0809   2.991143
#> 513 505.7032   3.167762
#> 514 500.0671   3.272785
#> 515 507.7449   3.158814
#> 516 563.9759   2.564001
#> 517 770.8127   2.960443
#> 518 500.0001   3.178510
#> 519 568.3579   2.506697
#> 520 501.2396   2.905145
#> 521 501.2248   3.343556
#> 522 500.6592   3.821060
#> 523 500.8414   2.995755
#> 524 500.5785   2.655760
#> 525 500.6883   3.261693
#> 526 524.6280   2.507430
#> 527 500.4779   2.716101
#> 528 505.5236   2.604053
#> 529 500.0436   2.729813
#> 530 583.3712   2.559757
#> 531 511.2176   2.752390
#> 532 500.9187   2.977460
#> 533 537.2154   2.580071
#> 534 534.1701   2.688541
#> 535 506.6911   2.811400
#> 536 500.0774   3.198443
#> 537 641.0568   2.547988
#> 538 500.2532   2.862876
#> 539 505.6216   2.893473
#> 540 502.0154   3.332837
#> 541 500.4794   2.695558
#> 542 500.0624   2.750025
#> 543 501.3939   3.830622
#> 544 500.0149   2.740758
#> 545 503.3657   2.715697
#> 546 500.7663   3.177737
#> 547 503.3812   2.785168
#> 548 500.0542   2.905711
#> 549 500.2960   2.750589
#> 550 500.0504   2.884739
#> 551 500.2429   2.625301
#> 552 604.2714   2.748133
#> 553 500.1764   4.206282
#> 554 501.0406   2.553286
#> 555 504.1438   2.750990
#> 556 500.0921   2.987055
#> 557 500.4661   2.538101
#> 558 500.1394   2.916834
#> 559 800.3977   2.520703
#> 560 500.4067   2.694607
#> 561 506.7790   3.372774
#> 562 884.9906   3.073261
#> 563 588.8322   2.699177
#> 564 501.1807   3.066234
#> 565 720.2573   3.680480
#> 566 500.0841   3.009204
#> 567 503.1639   2.833769
#> 568 507.7603   2.509783
#> 569 501.0699   2.614749
#> 570 660.4900   2.513785
#> 571 500.1521   3.709825
#> 572 500.1787   2.521239
#> 573 500.8894   2.623912
#> 574 500.6297   2.711776
#> 575 500.4614   2.680991
#> 576 501.0414   2.613903
#> 577 529.8365   2.865225
#> 578 518.2160   2.607656
#> 579 501.1371   2.552598
#> 580 658.6104   2.825720
#> 581 543.5876   2.668356
#> 582 660.9839   2.618746
#> 583 765.8755   2.713863
#> 584 542.9457   2.854670
#> 585 500.7753   2.584288
#> 586 857.3380   2.529985
#> 587 513.2899   3.084992
#> 588 500.0015   2.865997
#> 589 508.1245   3.173525
#> 590 500.2586   2.708914
#> 591 532.5227   2.886688
#> 592 515.2560   2.773911
#> 593 665.9450   2.751977
#> 594 528.0997   2.604177
#> 595 502.8672   2.599017
#> 596 529.4594   3.068633
#> 597 506.9956   2.748208
#> 598 503.8257   2.645798
#> 599 501.4771   2.776279
#> 600 500.7279   3.033228
#> 601 528.9131   2.538487
#> 602 514.4773   2.881527
#> 603 500.3714   2.675128
#> 604 607.3883   2.973552
#> 605 808.3655   2.713797
#> 606 604.0258   2.886333
#> 607 515.1463   2.811120
#> 608 681.1648   2.611798
#> 609 500.0719   4.044074
#> 610 500.6269   3.081017
#> 611 569.7355   2.752001
#> 612 906.8052   2.555257
#> 613 500.6855   2.760792
#> 614 560.8389   2.802297
#> 615 587.8638   2.680763
#> 616 772.9376   2.979320
#> 617 501.9350   2.553140
#> 618 619.2110   3.310407
#> 619 909.7572   3.065648
#> 620 501.6758   2.917654
#> 621 500.0357   3.173587
#> 622 576.4896   2.624021
#> 623 505.0634   2.964981
#> 624 502.3377   2.722900
#> 625 554.9547   4.189624
#> 626 500.3333   3.295934
#> 627 501.6164   2.520677
#> 628 500.0974   3.186577
#> 629 578.6794   3.088640
#> 630 533.1244   2.895795
#> 631 500.1797   2.836535
#> 632 500.0540   2.628277
#> 633 704.1102   2.631495
#> 634 500.1563   2.761051
#> 635 501.0356   2.533903
#> 636 519.7147   2.840823
#> 637 501.6272   3.085186
#> 638 555.6791   2.801856
#> 639 510.0853   3.986383
#> 640 991.6982   2.652317
#> 641 501.1171   2.676994
#> 642 603.8988   2.531711
#> 643 500.0883   2.560014
#> 644 639.9131   2.619812
#> 645 523.6793   2.563209
#> 646 500.0641   2.504940
#> 647 516.5253   2.641642
#> 648 502.6450   2.991156
#> 649 514.8927   2.529994
#> 650 501.2090   2.874853
#> 651 500.7000   2.554911
#> 652 692.5202   5.275256
#> 653 535.2018   2.833737
#> 654 500.3876   3.125468
#> 655 570.8128   2.624499
#> 656 511.5588   3.478233
#> 657 503.0103   2.629033
#> 658 503.4946   2.529189
#> 659 512.4185   2.625386
#> 660 500.1357   2.725284
#> 661 502.7967   2.546771
#> 662 500.6498   2.623323
#> 663 557.6824   2.671921
#> 664 500.1538   2.511816
#> 665 698.0985   2.632624
#> 666 500.4589   3.461002
#> 667 503.0541   2.749968
#> 668 671.8144   3.434208
#> 669 500.0941   2.616886
#> 670 506.3616   2.684771
#> 671 501.2993   2.995236
#> 672 637.2605   2.767930
#> 673 502.4694   4.059671
#> 674 506.7075   2.526195
#> 675 525.7996   2.608161
#> 676 506.2052   2.525917
#> 677 500.0968   3.830553
#> 678 508.9935   3.527950
#> 679 812.9395   2.563898
#> 680 832.3391   3.308996
#> 681 958.4393   2.720085
#> 682 500.1363   3.153909
#> 683 592.2832   2.526494
#> 684 791.5589   2.681905
#> 685 501.8360   3.114742
#> 686 526.5194   2.543103
#> 687 504.4479   2.519867
#> 688 500.1392   2.504768
#> 689 500.2138   2.623895
#> 690 504.0055   2.809845
#> 691 583.9964   2.909282
#> 692 580.3261   2.569449
#> 693 662.8243   2.786310
#> 694 503.5217   2.671053
#> 695 501.5153   3.022844
#> 696 502.7935   2.767948
#> 697 502.1428   2.639376
#> 698 500.0365   3.655708
#> 699 500.4260   2.705632
#> 700 501.4389   2.524987
#> 701 505.3226   3.728306
#> 702 500.3808   2.549710
#> 703 500.4504   2.876140
#> 704 500.0168   2.756153
#> 705 500.0557   3.138351
#> 706 584.5558   3.286981
#> 707 500.5888   3.103288
#> 708 534.2411   3.073027
#> 709 534.9453   2.850324
#> 710 587.0771   2.533337
#> 711 500.1446   2.613040
#> 712 500.7330   3.049012
#> 713 500.0637   2.672741
#> 714 500.5472   2.608513
#> 715 500.1963   3.316804
#> 716 500.6160   2.563178
#> 717 501.2342   3.425370
#> 718 502.1350   2.623187
#> 719 501.9764   3.233785
#> 720 507.3838   3.524289
#> 721 761.6785   2.839575
#> 722 581.8493   2.923374
#> 723 501.7072   5.681429
#> 724 501.3449   2.525688
#> 725 508.8042   2.503560
#> 726 604.7145   3.285418
#> 727 702.6810   3.260623
#> 728 868.9641   2.874739
#> 729 735.5074   4.218679
#> 730 505.3631   3.368229
#> 731 500.0631   2.768637
#> 732 500.7800   2.791053
#> 733 503.0108   2.670629
#> 734 506.2623   3.171385
#> 735 501.9127   2.553981
#> 736 794.1226   2.658126
#> 737 501.9573   3.134832
#> 738 594.7845   2.616715
#> 739 562.4380   2.727224
#> 740 657.8829   2.601988
#> 741 508.9964   3.063182
#> 742 955.4609   3.022805
#> 743 537.6198   3.029561
#> 744 500.0329   3.029965
#> 745 500.2526   2.717685
#> 746 500.2495   2.697113
#> 747 797.6969   2.620901
#> 748 500.2130   3.897745
#> 749 646.4041   2.520593
#> 750 513.8046   2.922954
#> 751 514.5860   2.512132
#> 752 500.1362   4.335154
#> 753 874.6862   2.741446
#> 754 500.9861   2.629494
#> 755 500.0206   3.021875
#> 756 531.2350   2.833617
#> 757 500.7220   2.726884
#> 758 520.5503   2.597981
#> 759 500.1504   2.991157
#> 760 509.5137   2.798358
#> 761 501.9206   2.556422
#> 762 597.7114   3.192941
#> 763 500.1228   2.598427
#> 764 500.1078   2.729221
#> 765 546.7819   3.771577
#> 766 500.6463   3.193465
#> 767 500.1651   2.646342
#> 768 500.0067   4.908649
#> 769 599.6905   2.691880
#> 770 502.0137   3.500581
#> 771 505.5484   2.524802
#> 772 553.1437   3.277542
#> 773 545.0935   3.130544
#> 774 630.6902   2.738970
#> 775 510.2559   3.386034
#> 776 500.2489   2.738748
#> 777 535.3619   2.617980
#> 778 504.6203   2.755000
#> 779 615.5367   4.088661
#> 780 505.3670   2.754544
#> 781 503.7824   2.983815
#> 782 506.6641   3.052397
#> 783 501.5595   2.610247
#> 784 501.3515   2.692891
#> 785 500.4517   3.432422
#> 786 503.0170   2.734524
#> 787 500.8648   2.918837
#> 788 512.5064   2.783707
#> 789 605.5151   2.696373
#> 790 500.3492   2.793605
#> 791 608.2922   2.558718
#> 792 501.0702   2.892599
#> 793 508.9016   3.009640
#> 794 741.5196   4.112304
#> 795 501.6699   2.717193
#> 796 500.8556   4.851823
#> 797 514.5392   2.913421
#> 798 550.3907   2.533269
#> 799 526.3427   3.554888
#> 800 531.7370   3.193458
#> 801 507.4680   3.854253
#> 802 536.4600   2.716648
#> 803 525.3639   3.057532
#> 804 507.6756   2.902542
#> 805 502.2817   3.349189
#> 806 559.4346   2.780588
#> 807 578.0880   2.991988
#> 808 781.1317   3.011280
#> 809 502.7304   2.763533
#> 810 500.1089   2.782848
#> 811 500.1556   2.548279
#> 812 502.1577   2.706097
#> 813 624.8980   2.843581
#> 814 539.9467   2.844752
#> 815 869.5665   3.245301
#> 816 502.0638   3.841331
#> 817 506.8494   2.767276
#> 818 500.0991   2.549900
#> 819 506.7821   3.094264
#> 820 557.8802   2.988838
#> 821 500.0017   2.868224
#> 822 825.9353   3.542638
#> 823 500.0438   3.173110
#> 824 502.5262   2.813608
#> 825 511.6409   2.760370
#> 826 504.2679   2.872267
#> 827 539.1442   2.828548
#> 828 531.4260   2.551719
#> 829 565.1874   2.905166
#> 830 506.1731   2.742121
#> 831 500.4287   2.797890
#> 832 502.2003   3.015909
#> 833 500.3496   2.513395
#> 834 500.0172   2.579948
#> 835 501.3830   2.628340
#> 836 500.0500   3.165760
#> 837 565.1067   2.983328
#> 838 724.8212   3.209683
#> 839 500.3468   2.681710
#> 840 755.5627   3.440876
#> 841 548.9217   2.743914
#> 842 502.0879   2.797809
#> 843 512.6273   2.905380
#> 844 500.1407   2.861215
#> 845 681.2199   2.547015
#> 846 500.0406   3.000116
#> 847 657.7456   2.697410
#> 848 501.5530   2.711009
#> 849 555.9048   2.668194
#> 850 513.3276   2.981765
#> 851 510.3008   2.816749
#> 852 500.7352   3.261868
#> 853 512.9953   5.833627
#> 854 500.0683   2.942087
#> 855 500.4251   3.331880
#> 856 580.3654   2.540702
#> 857 500.3342   3.152127
#> 858 500.4699   3.248239
#> 859 501.1890   2.828527
#> 860 500.7980   3.818787
#> 861 833.5380   3.230268
#> 862 503.6814   2.942974
#> 863 500.0483   2.863603
#> 864 501.2707   2.575266
#> 865 520.9770   3.158244
#> 866 513.2065   2.574877
#> 867 500.2283   2.825735
#> 868 500.5267   3.672800
#> 869 577.5217   2.585654
#> 870 537.3094   2.648337
#> 871 525.3259   3.827160
#> 872 947.3981   2.924965
#> 873 500.0278   2.788236
#> 874 524.3356   2.641712
#> 875 503.8643   2.750000
#> 876 503.4751   2.582433
#> 877 507.5321   3.030775
#> 878 507.5786   2.566431
#> 879 500.4232   2.649834
#> 880 598.5573   3.274716
#> 881 501.0010   3.506540
#> 882 596.1129   4.256009
#> 883 577.3631   2.612356
#> 884 535.1183   2.743252
#> 885 626.9556   2.714453
#> 886 906.2099   2.500040
#> 887 622.0671   4.113296
#> 888 500.1042   2.930052
#> 889 610.3481   2.997880
#> 890 500.8667   4.138198
#> 891 505.0125   3.018867
#> 892 501.1311   3.768261
#> 893 501.8443   2.846314
#> 894 500.0068   2.658031
#> 895 500.1161   2.764556
#> 896 524.8326   3.148014
#> 897 664.7185   2.689097
#> 898 557.3118   2.741180
#> 899 530.1063   2.889554
#> 900 522.4590   3.377961
#> 901 834.2311   2.715366
#> 902 603.8458   3.073105
#> 903 511.9996   2.608361
#> 904 503.4640   2.883821
#> 905 500.2766   3.375271
#> 906 604.5666   2.889575
#> 907 502.4649   2.826398