Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   510.1308   2.843732
#> 2   602.7416   2.932635
#> 3   500.0025   2.879973
#> 4   501.6953   2.564078
#> 5   500.1764   3.230178
#> 6   934.2789   2.608410
#> 7   529.6142   2.537313
#> 8   501.7300   2.586686
#> 9   859.2826   3.713809
#> 10  500.2463   3.270067
#> 11  614.4298   4.321688
#> 12  686.1960   3.178277
#> 13  504.6300   2.763524
#> 14  632.5772   2.610330
#> 15  500.0483   2.961181
#> 16  500.0371   2.623677
#> 17  501.4898   2.687993
#> 18  500.2764   2.560236
#> 19  500.3436   2.536310
#> 20  509.0683   2.698055
#> 21  500.1897   3.886506
#> 22  500.6769   2.813046
#> 23  955.0327   2.625712
#> 24  501.6183   3.541502
#> 25  501.9450   2.709770
#> 26  666.8218   3.149374
#> 27  502.1148   2.580968
#> 28  546.8518   2.904517
#> 29  501.1302   2.549421
#> 30  838.8566   3.997767
#> 31  503.4041   2.715547
#> 32  665.9791   3.572160
#> 33  530.0054   2.656781
#> 34  524.3298   2.889020
#> 35  517.7843   2.524858
#> 36  500.2907   4.129182
#> 37  500.1911   3.378853
#> 38  500.1359   3.824920
#> 39  530.5037   3.433604
#> 40  519.9807   3.029178
#> 41  500.5610   3.212754
#> 42  505.7685   2.517798
#> 43  515.7939   2.764305
#> 44  505.0250   2.825853
#> 45  500.0269   3.166115
#> 46  500.4217   4.813314
#> 47  500.0214   3.171434
#> 48  500.0581   2.628385
#> 49  503.5849   4.177283
#> 50  500.2788   3.132618
#> 51  508.1840   2.937114
#> 52  570.4895   2.647221
#> 53  523.5084   3.549614
#> 54  501.8681   2.570612
#> 55  978.8658   3.226045
#> 56  501.1370   2.938948
#> 57  625.0732   3.044028
#> 58  718.2780   2.550630
#> 59  500.7412   2.780992
#> 60  501.7270   2.518858
#> 61  502.3690   2.632687
#> 62  503.0276   3.168272
#> 63  500.0261   2.888555
#> 64  552.5805   3.282100
#> 65  500.1958   3.243265
#> 66  525.6567   2.941945
#> 67  530.0864   3.583501
#> 68  501.0823   2.701741
#> 69  546.3808   2.563018
#> 70  500.2197   3.552353
#> 71  523.4163   3.229170
#> 72  525.9599   2.679136
#> 73  500.4753   3.022206
#> 74  872.5863   2.837238
#> 75  500.1590   2.531622
#> 76  500.4230   2.693049
#> 77  535.2704   2.819536
#> 78  500.0142   3.051631
#> 79  504.4877   3.007114
#> 80  511.8214   2.781581
#> 81  512.4501   2.800336
#> 82  500.3070   3.416008
#> 83  503.3245   3.841422
#> 84  940.5872   2.981201
#> 85  500.2877   3.010982
#> 86  501.5551   2.560340
#> 87  505.2276   2.925298
#> 88  506.9419   3.194514
#> 89  500.4561   3.766719
#> 90  500.0305   3.327997
#> 91  902.7005   4.248081
#> 92  978.1651   2.519319
#> 93  597.7299   2.814098
#> 94  500.0247   2.755121
#> 95  801.3791   3.269910
#> 96  503.4820   2.591364
#> 97  500.8286   2.711179
#> 98  500.0346   3.269984
#> 99  502.4588   2.576411
#> 100 549.1492   3.043474
#> 101 510.3950   2.603797
#> 102 500.0457   2.694491
#> 103 967.3709   2.689852
#> 104 515.5525   2.574966
#> 105 887.2334   2.523676
#> 106 500.0584   2.580230
#> 107 628.8781   2.723167
#> 108 534.3321   3.338562
#> 109 843.1472   2.851843
#> 110 500.3502   2.513853
#> 111 500.0602   2.672766
#> 112 500.3900   2.615519
#> 113 505.3187   2.657364
#> 114 505.5784   3.241914
#> 115 501.6157   2.603224
#> 116 794.6861   2.626757
#> 117 500.0055   3.013540
#> 118 500.6129   3.182340
#> 119 500.0269   3.063945
#> 120 503.5192   4.040300
#> 121 500.0629   2.992469
#> 122 500.0605   2.737498
#> 123 500.0142   3.180354
#> 124 502.8874   2.639679
#> 125 501.4239   2.868700
#> 126 500.0035   2.543987
#> 127 502.0630   2.662122
#> 128 559.7047   2.772860
#> 129 601.8194   2.858495
#> 130 974.5153   3.221206
#> 131 872.5471   2.677235
#> 132 500.4641   2.888858
#> 133 504.5493   2.749199
#> 134 500.2838   3.241130
#> 135 500.1640   2.712927
#> 136 500.2777   2.870860
#> 137 501.3463   3.624028
#> 138 657.9032   2.660039
#> 139 845.1000   2.591833
#> 140 500.1271   3.072358
#> 141 514.3348   3.070547
#> 142 500.1738   3.026217
#> 143 509.4962   2.603607
#> 144 510.4914   2.978182
#> 145 500.3896   2.698021
#> 146 500.9590   2.572620
#> 147 502.5063   2.672988
#> 148 502.1425   2.619577
#> 149 528.4307   2.935854
#> 150 502.3942   2.955409
#> 151 500.0007   2.577141
#> 152 501.2166   2.559206
#> 153 659.3601   3.036741
#> 154 563.4419   2.981219
#> 155 500.0299   3.177001
#> 156 502.1847   2.810828
#> 157 528.2913   2.870967
#> 158 500.0579   2.868464
#> 159 514.7090   3.076094
#> 160 556.5414   2.785492
#> 161 500.4437   3.371589
#> 162 548.6896   2.767655
#> 163 500.6237   2.600123
#> 164 500.0860   3.030113
#> 165 534.8958   3.259900
#> 166 500.4673   2.717580
#> 167 505.9546   2.652049
#> 168 500.5930   3.671683
#> 169 778.9260   2.688302
#> 170 500.2476   2.564843
#> 171 537.0045   4.244134
#> 172 501.0201   3.196650
#> 173 549.5264   3.461792
#> 174 511.3862   3.863550
#> 175 502.5835   2.656188
#> 176 522.7262   2.679829
#> 177 682.5461   3.633751
#> 178 592.3685   2.592574
#> 179 500.5860   3.052769
#> 180 519.6573   4.232021
#> 181 500.8578   2.699940
#> 182 514.7064   2.909125
#> 183 500.0245   3.045523
#> 184 504.6612   2.530326
#> 185 500.1626   2.685138
#> 186 501.3683   2.682281
#> 187 507.7984   2.671881
#> 188 517.1016   2.713125
#> 189 500.0638   2.529366
#> 190 502.1362   2.704028
#> 191 500.6569   2.595502
#> 192 503.1339   3.857298
#> 193 544.9958   2.813516
#> 194 504.3773   3.100234
#> 195 510.4589   3.092131
#> 196 500.1052   3.175099
#> 197 501.8277   2.655790
#> 198 500.1578   2.686924
#> 199 503.2059   2.527890
#> 200 500.2342   2.566590
#> 201 505.1381   2.542741
#> 202 654.3047   2.918124
#> 203 504.8247   3.523292
#> 204 500.0109   4.815757
#> 205 584.1157   3.086171
#> 206 539.8183   2.938693
#> 207 501.0259   2.614882
#> 208 559.8662   2.558626
#> 209 518.5279   2.678371
#> 210 621.5318   2.982542
#> 211 653.5325   3.389034
#> 212 530.6685   2.829690
#> 213 509.4311   2.598129
#> 214 682.5405   3.243493
#> 215 500.0497   2.576698
#> 216 500.0770   3.247245
#> 217 500.4126   2.572805
#> 218 501.8163   4.363424
#> 219 500.1314   2.633750
#> 220 520.9077   2.530962
#> 221 500.0113   2.500439
#> 222 500.0366   2.968220
#> 223 500.3390   2.690192
#> 224 501.7223   2.555564
#> 225 533.5385   3.027135
#> 226 730.4977   2.760089
#> 227 573.7064   3.355229
#> 228 500.0419   2.631283
#> 229 500.9748   3.016678
#> 230 780.4924   3.246798
#> 231 501.5001   2.556540
#> 232 501.6884   2.617430
#> 233 592.1899   2.526380
#> 234 500.1016   2.515309
#> 235 502.3602   2.618203
#> 236 501.4324   2.832848
#> 237 500.0309   2.810710
#> 238 517.1125   4.720042
#> 239 500.0245   2.770691
#> 240 500.0048   2.976786
#> 241 500.2230   2.878951
#> 242 500.3220   2.732210
#> 243 501.1331   2.588127
#> 244 500.5831   2.619140
#> 245 992.5764   2.714249
#> 246 500.0048   2.600289
#> 247 533.1534   2.539886
#> 248 526.4912   2.663365
#> 249 504.5200   2.518769
#> 250 504.0904   3.713965
#> 251 512.0322   4.260347
#> 252 526.8771   3.051690
#> 253 500.1291   2.715327
#> 254 597.7814   3.442916
#> 255 799.7191   3.063357
#> 256 503.6425   2.685565
#> 257 500.0673   2.747877
#> 258 500.9070   2.623509
#> 259 661.7208   2.504047
#> 260 503.3992   3.511006
#> 261 501.3322   3.210318
#> 262 500.7175   3.333273
#> 263 704.9964   3.258059
#> 264 506.7477   2.574473
#> 265 541.7090   3.756374
#> 266 583.2267   3.297144
#> 267 500.0889   2.896064
#> 268 506.3971   2.756915
#> 269 963.4991   3.544303
#> 270 517.8026   3.647069
#> 271 529.7166   2.885688
#> 272 512.3021   2.969324
#> 273 500.6936   2.637508
#> 274 567.5154   3.079890
#> 275 500.3632   2.875412
#> 276 500.1967   2.906603
#> 277 500.0014   2.633053
#> 278 776.0459   3.436767
#> 279 500.2258   2.686430
#> 280 728.8576   2.816772
#> 281 669.9114   3.052865
#> 282 500.3711   2.745749
#> 283 500.2863   2.650207
#> 284 501.3534   2.916641
#> 285 575.4279   2.575620
#> 286 500.4325   2.732031
#> 287 506.5788   2.725046
#> 288 500.2074   2.549363
#> 289 625.1328   2.554585
#> 290 501.8583   2.679796
#> 291 900.3203   4.304962
#> 292 500.1090   2.642551
#> 293 510.4946   2.509036
#> 294 666.1294   2.897813
#> 295 500.8674   3.248189
#> 296 786.7523   4.638551
#> 297 501.0232   3.720391
#> 298 510.7625   4.181392
#> 299 902.2376   3.146848
#> 300 541.7095   2.748599
#> 301 514.1788   2.789329
#> 302 500.1550   2.884507
#> 303 501.1726   3.078412
#> 304 500.3722   2.509921
#> 305 579.8401   3.313676
#> 306 502.0815   2.675868
#> 307 504.9525   2.892821
#> 308 500.3246   2.686797
#> 309 503.4202   2.559467
#> 310 501.3822   3.780783
#> 311 501.6813   2.519691
#> 312 684.5970   4.474349
#> 313 627.1712   3.144251
#> 314 500.2313   3.127094
#> 315 500.5435   2.551161
#> 316 502.6563   2.822457
#> 317 500.1147   2.519243
#> 318 500.0140   3.412111
#> 319 500.5260   3.148565
#> 320 500.2609   3.002291
#> 321 742.1933   4.047293
#> 322 773.4582   2.576981
#> 323 518.9115   2.682583
#> 324 504.7081   3.221260
#> 325 519.0270   3.234266
#> 326 500.4507   2.509886
#> 327 543.2315   2.895424
#> 328 531.5049   2.562964
#> 329 615.1418   3.166931
#> 330 500.1247   3.086815
#> 331 510.9765   3.455362
#> 332 505.9694   3.185490
#> 333 502.2374   3.100041
#> 334 502.9550   2.578099
#> 335 504.3740   2.783182
#> 336 572.9685   2.617809
#> 337 500.5054   2.742425
#> 338 505.2498   2.810423
#> 339 501.1553   2.769532
#> 340 500.6674   3.025471
#> 341 501.6625   2.616169
#> 342 560.7801   3.308877
#> 343 564.7644   3.192148
#> 344 512.1281   2.994886
#> 345 501.0618   2.920291
#> 346 616.0290   3.000666
#> 347 501.1478   2.770364
#> 348 502.9351   2.528768
#> 349 500.1583   2.713884
#> 350 502.4583   2.649647
#> 351 500.0094   5.164752
#> 352 500.4049   2.847215
#> 353 500.0753   2.792459
#> 354 503.6698   2.793780
#> 355 500.4543   2.737658
#> 356 501.3327   2.698746
#> 357 500.0324   3.577245
#> 358 508.0138   2.675612
#> 359 511.0180   2.710909
#> 360 529.8165   2.699650
#> 361 937.7355   2.611530
#> 362 587.6876   2.887561
#> 363 500.2565   2.932163
#> 364 500.1899   2.936847
#> 365 500.1410   2.501072
#> 366 500.7896   2.856490
#> 367 501.4672   2.680576
#> 368 500.4445   2.752853
#> 369 726.6122   2.618551
#> 370 500.2023   2.562663
#> 371 500.0246   3.077399
#> 372 502.0970   2.592762
#> 373 510.8799   2.683571
#> 374 500.1639   2.519325
#> 375 545.4811   3.182218
#> 376 506.4145   2.543602
#> 377 500.4295   2.520067
#> 378 504.1827   4.036522
#> 379 515.2630   2.734145
#> 380 500.0241   2.555735
#> 381 500.0799   3.293107
#> 382 503.0478   2.725405
#> 383 614.5560   2.613959
#> 384 645.1742   2.541244
#> 385 500.6228   2.584872
#> 386 500.4800   2.932521
#> 387 502.2708   2.555376
#> 388 502.9473   2.875082
#> 389 539.2376   3.076202
#> 390 500.6459   3.828020
#> 391 500.5959   2.919069
#> 392 536.0277   4.659289
#> 393 517.2144   3.668986
#> 394 501.4191   3.097253
#> 395 500.0272   2.665941
#> 396 540.9611   2.889614
#> 397 501.8709   2.660801
#> 398 503.4960   2.929351
#> 399 761.6451   2.984501
#> 400 500.0665   2.649978
#> 401 502.2492   3.474771
#> 402 542.0971   3.265182
#> 403 500.3324   2.984972
#> 404 501.4988   3.368117
#> 405 500.1225   3.563447
#> 406 501.2913   2.805046
#> 407 500.0087   3.099173
#> 408 500.0350   2.569736
#> 409 500.5573   2.734030
#> 410 502.3445   2.896859
#> 411 709.5583   2.958387
#> 412 566.5442   3.522704
#> 413 500.0365   2.799310
#> 414 627.1182   3.081352
#> 415 500.1063   2.591791
#> 416 536.1483   2.735140
#> 417 500.0341   2.594908
#> 418 500.0169   2.813953
#> 419 506.3890   2.599130
#> 420 507.5848   2.679235
#> 421 502.4133   3.207358
#> 422 500.0409   2.578803
#> 423 514.6585   2.824159
#> 424 515.3204   3.605470
#> 425 510.4944   2.535303
#> 426 508.4983   2.732123
#> 427 502.0156   2.676714
#> 428 545.3200   2.735922
#> 429 516.3541   3.002728
#> 430 741.0510   4.177202
#> 431 501.0593   2.538995
#> 432 501.8276   4.006957
#> 433 500.0186   2.652003
#> 434 530.0585   2.981044
#> 435 500.4556   2.690560
#> 436 500.5062   2.847517
#> 437 544.4701   2.716932
#> 438 507.3326   2.673389
#> 439 615.7134   2.691948
#> 440 501.2619   3.766109
#> 441 500.0753   2.521514
#> 442 645.0565   2.796913
#> 443 500.0425   2.542899
#> 444 500.0298   2.655765
#> 445 545.9657   3.238803
#> 446 521.6118   2.534031
#> 447 528.9947   2.772691
#> 448 501.5507   2.639423
#> 449 703.4957   2.504659
#> 450 611.7123   2.678694
#> 451 507.9287   2.643672
#> 452 500.0003   2.578482
#> 453 500.0373   3.080340
#> 454 505.6523   2.862347
#> 455 509.8640   3.179897
#> 456 645.4304   2.700502
#> 457 500.0441   3.241829
#> 458 508.0398   3.027450
#> 459 506.6898   2.975491
#> 460 771.0280   2.553057
#> 461 747.2714   2.822668
#> 462 515.9806   3.487889
#> 463 500.8618   2.518740
#> 464 500.0922   3.127185
#> 465 570.1165   2.747126
#> 466 500.3319   2.665733
#> 467 500.0311   2.605957
#> 468 548.0360   3.285710
#> 469 501.6862   2.659082
#> 470 596.1789   2.699822
#> 471 501.6291   2.525669
#> 472 500.3176   2.843687
#> 473 500.1442   3.111392
#> 474 545.6983   2.714671
#> 475 500.5682   3.215120
#> 476 501.1935   2.672597
#> 477 537.8298   2.624263
#> 478 500.2596   2.769168
#> 479 500.0363   2.584487
#> 480 500.1564   2.903763
#> 481 502.1140   2.521022
#> 482 510.9499   2.689546
#> 483 500.9068   2.513200
#> 484 501.0671   3.104166
#> 485 505.8362   2.832007
#> 486 756.1322   2.532897
#> 487 522.4803   2.969774
#> 488 701.5055   2.709509
#> 489 514.6896   3.099370
#> 490 500.2168   2.537184
#> 491 500.2422   2.603457
#> 492 500.2948   3.196477
#> 493 501.3543   2.628606
#> 494 540.5520   3.726757
#> 495 500.0309   2.740079
#> 496 572.8210   2.858282
#> 497 502.0540   2.808541
#> 498 507.8372   2.547561
#> 499 520.8354   2.706888
#> 500 903.1111   2.877853
#> 501 500.5331   3.042495
#> 502 501.5104   2.618282
#> 503 516.6597   2.584803
#> 504 500.5359   2.834715
#> 505 500.3745   2.756879
#> 506 533.7274   2.617559
#> 507 502.4377   3.542249
#> 508 552.7504   2.648258
#> 509 500.0320   3.174332
#> 510 502.2235   2.857737
#> 511 582.2062   2.870436
#> 512 851.3566   2.617181
#> 513 500.4940   2.731528
#> 514 500.5660   4.027310
#> 515 500.0213   4.287169
#> 516 500.2559   3.089179
#> 517 572.2863   2.727097
#> 518 501.1309   3.388663
#> 519 500.6292   3.256787
#> 520 570.8487   2.586343
#> 521 533.8996   3.402283
#> 522 501.2435   2.729276
#> 523 500.0532   2.811336
#> 524 500.0206   2.535345
#> 525 500.2543   2.954695
#> 526 526.8510   2.772412
#> 527 574.9920   2.707184
#> 528 500.2369   3.431606
#> 529 500.5453   2.937535
#> 530 501.1261   2.897791
#> 531 500.5944   3.055380
#> 532 523.0296   2.846412
#> 533 554.2064   3.195784
#> 534 501.5365   2.881880
#> 535 502.4756   2.520642
#> 536 500.0015   3.941110
#> 537 502.3732   3.723644
#> 538 504.6212   3.001512
#> 539 500.1964   2.869999
#> 540 541.5256   2.603918
#> 541 501.9586   2.840781
#> 542 535.6581   2.616238
#> 543 558.7091   3.035379
#> 544 501.3313   3.315165
#> 545 537.2706   3.070859
#> 546 500.5304   2.960003
#> 547 500.2002   2.560794
#> 548 542.6851   4.133416
#> 549 502.2730   2.604375
#> 550 593.4111   2.896954
#> 551 780.7743   2.625636
#> 552 500.0246   2.868068
#> 553 501.0596   2.564200
#> 554 501.1787   3.685020
#> 555 500.0093   2.972998
#> 556 506.0409   2.735664
#> 557 500.1045   2.569912
#> 558 635.1293   3.038821
#> 559 513.8169   2.700902
#> 560 502.5498   3.150009
#> 561 597.6471   3.143416
#> 562 507.0633   2.651268
#> 563 503.0126   4.441244
#> 564 513.0478   2.575759
#> 565 501.9814   2.540948
#> 566 501.7325   2.561836
#> 567 500.5846   2.580300
#> 568 696.3560   3.015269
#> 569 549.4038   2.839325
#> 570 516.8428   2.586644
#> 571 501.4561   2.869730
#> 572 500.0285   2.629836
#> 573 574.5336   2.843144
#> 574 500.0103   2.972829
#> 575 506.0361   2.661143
#> 576 500.0299   2.939730
#> 577 531.9452   2.516456
#> 578 500.1675   2.591339
#> 579 547.5888   2.574144
#> 580 500.6770   2.591427
#> 581 503.5319   2.778592
#> 582 500.0760   3.524678
#> 583 505.9963   2.538721
#> 584 500.4997   3.191570
#> 585 559.7101   3.117146
#> 586 500.0681   2.944607
#> 587 517.4068   2.643084
#> 588 511.1606   3.347359
#> 589 500.1711   3.555337
#> 590 502.5638   2.903563
#> 591 501.8190   3.692311
#> 592 500.0223   2.567685
#> 593 500.2394   2.719276
#> 594 500.5064   2.681335
#> 595 500.7115   2.741733
#> 596 575.3527   2.820236
#> 597 537.0731   2.959526
#> 598 503.6330   2.894643
#> 599 510.4952   2.972548
#> 600 577.1158   2.645021
#> 601 505.5120   2.578683
#> 602 505.5642   3.040238
#> 603 504.1109   2.791420
#> 604 500.7939   2.853910
#> 605 506.3572   2.619041
#> 606 836.6481   2.839379
#> 607 500.0103   3.126857
#> 608 500.2575   2.551500
#> 609 501.1591   2.734099
#> 610 584.4265   3.937213
#> 611 507.5055   2.847949
#> 612 621.7509   2.729334
#> 613 500.1339   2.533092
#> 614 501.7999   2.728598
#> 615 500.3562   2.875765
#> 616 567.4540   3.854869
#> 617 514.7754   2.638146
#> 618 575.6553   2.724766
#> 619 506.4005   3.060026
#> 620 500.4911   3.977646
#> 621 541.5770   2.533775
#> 622 543.2014   2.785831
#> 623 500.0834   2.931259
#> 624 591.1781   3.470698
#> 625 521.5646   3.366173
#> 626 501.1635   4.564327
#> 627 501.9151   3.695217
#> 628 502.9206   2.577212
#> 629 501.2683   3.851397
#> 630 518.4139   2.961802
#> 631 505.5621   2.701230
#> 632 500.2967   3.213722
#> 633 500.3456   2.724876
#> 634 786.3093   2.608006
#> 635 560.3617   3.126395
#> 636 500.5882   2.512768
#> 637 765.6895   2.543511
#> 638 502.3626   3.560708
#> 639 500.0599   2.772147
#> 640 510.8934   2.512513
#> 641 598.4933   2.535195
#> 642 501.7704   2.942070
#> 643 500.0359   2.831770
#> 644 576.4046   2.660812
#> 645 869.2515   3.017647
#> 646 500.1110   2.802375
#> 647 500.4417   2.939786
#> 648 511.8917   2.873515
#> 649 505.2518   2.708365
#> 650 555.1475   2.750124
#> 651 646.7939   3.037074
#> 652 500.7271   3.614341
#> 653 774.3886   2.609872
#> 654 508.1802   3.150137
#> 655 500.0546   3.766550
#> 656 500.4249   2.857494
#> 657 604.1783   2.702877
#> 658 506.4644   2.842498
#> 659 500.6825   2.624348
#> 660 549.3461   2.546999
#> 661 746.6608   2.832700
#> 662 503.2868   2.836165
#> 663 634.0131   4.331041
#> 664 508.0022   2.527933
#> 665 500.0528   4.278492
#> 666 500.1909   2.652989
#> 667 501.2966   3.659623
#> 668 528.6344   4.073891
#> 669 500.4895   3.334083
#> 670 501.3605   2.585858
#> 671 513.7579   2.825847
#> 672 501.0163   2.647577
#> 673 500.7524   2.855146
#> 674 518.9835   2.621849
#> 675 517.6644   2.546801
#> 676 500.4757   2.580475
#> 677 506.1845   2.999047
#> 678 569.4897   2.639932
#> 679 500.8347   3.100311
#> 680 711.9119   2.600925
#> 681 500.4836   3.228592
#> 682 583.4508   2.907312
#> 683 501.9053   2.546716
#> 684 503.9886   2.719429
#> 685 502.8387   2.654443
#> 686 507.3307   2.567214
#> 687 774.7792   2.791889
#> 688 528.9680   4.416732
#> 689 582.9590   2.836280
#> 690 506.7064   2.717716
#> 691 500.0549   2.692303
#> 692 500.9320   2.736786
#> 693 500.0162   3.485039
#> 694 500.1387   3.063889
#> 695 500.1729   2.560071
#> 696 661.6382   3.079844
#> 697 500.3332   5.103122
#> 698 551.3537   2.517772
#> 699 500.0985   3.223136
#> 700 500.3648   2.665975
#> 701 500.2256   2.506920
#> 702 501.6107   2.887378
#> 703 500.1436   2.559310
#> 704 500.7708   3.138101
#> 705 656.7589   3.888904
#> 706 500.4514   2.544336
#> 707 564.7098   2.906737
#> 708 526.7118   3.101611
#> 709 500.0341   2.547466
#> 710 510.6250   2.851815
#> 711 508.0476   2.923125
#> 712 504.3064   3.061299
#> 713 734.7251   2.711975
#> 714 500.3726   2.773351
#> 715 532.2074   2.580460
#> 716 761.3344   2.774138
#> 717 500.4857   2.745501
#> 718 501.1705   2.523744
#> 719 513.4194   2.731923
#> 720 503.7453   2.682468
#> 721 513.2934   2.678264
#> 722 707.7650   3.184641
#> 723 563.1206   2.569978
#> 724 504.5367   2.538841
#> 725 500.0078   2.511133
#> 726 500.9109   3.325298
#> 727 500.3305   3.727600
#> 728 501.3168   2.619200
#> 729 500.8195   2.550082
#> 730 554.5728   3.351321
#> 731 500.0490   2.664997
#> 732 579.5847   2.515938
#> 733 500.5930   2.558173
#> 734 509.9840   3.193900
#> 735 500.0003   3.147308
#> 736 500.8648   2.544711
#> 737 976.4255   2.625961
#> 738 500.6974   2.738234
#> 739 505.9635   2.646017
#> 740 950.8137   2.884146
#> 741 514.7862   2.527361
#> 742 507.5968   2.652454
#> 743 508.9206   3.772110
#> 744 501.3492   2.736430
#> 745 522.4463   2.565004
#> 746 501.0142   2.990613
#> 747 500.1029   2.570419
#> 748 574.2899   2.608012
#> 749 557.1737   4.019144
#> 750 509.8974   2.759682
#> 751 501.8647   3.127627
#> 752 504.6043   3.001643
#> 753 595.3744   2.691514
#> 754 502.2663   3.537724
#> 755 749.7712   2.705106
#> 756 762.7027   2.616389
#> 757 558.7258   2.548859
#> 758 546.1931   4.876216
#> 759 605.1964   2.945797
#> 760 527.5190   2.754183
#> 761 741.3122   3.857693
#> 762 717.1882   2.976528
#> 763 629.9091   2.805861
#> 764 595.8046   2.585527
#> 765 500.3052   2.706805
#> 766 500.0063   2.722956
#> 767 503.6012   2.589271
#> 768 500.1766   2.722206
#> 769 500.7697   2.503701
#> 770 503.3464   2.883476
#> 771 500.1754   3.857136
#> 772 644.7058   2.515536
#> 773 500.0037   2.581154
#> 774 500.7016   2.584516
#> 775 500.7703   2.708139
#> 776 503.3332   3.436936
#> 777 574.0592   2.584454
#> 778 500.7575   2.886339
#> 779 500.0363   2.741074
#> 780 500.0227   2.831783
#> 781 500.2771   3.063214
#> 782 502.1128   2.892441
#> 783 500.3641   2.624589
#> 784 511.7345   3.320480
#> 785 546.6960   2.872096
#> 786 500.6304   2.531650
#> 787 518.6248   2.732545
#> 788 503.5186   2.573966
#> 789 501.4219   3.193597
#> 790 500.5257   3.670618
#> 791 500.1143   2.974357
#> 792 500.1086   3.540896
#> 793 500.7543   2.694343
#> 794 500.4792   3.073238
#> 795 501.1999   2.594381
#> 796 500.0534   2.630922
#> 797 502.5072   2.719789
#> 798 686.3595   2.546237
#> 799 503.5651   3.565040
#> 800 501.4276   2.665828
#> 801 501.0356   2.613262
#> 802 577.0212   2.612884
#> 803 516.1091   3.035823
#> 804 506.2787   3.645023
#> 805 572.1721   3.690148
#> 806 500.0007   2.634749
#> 807 502.4789   2.683227
#> 808 504.7459   2.688115
#> 809 500.9616   2.552452
#> 810 508.5353   3.352755
#> 811 532.6548   2.919225
#> 812 752.2409   2.545970
#> 813 500.1712   2.537679
#> 814 524.8569   3.711742
#> 815 501.6538   2.995753
#> 816 515.8639   3.157809
#> 817 926.1310   3.571850
#> 818 509.6243   2.822453
#> 819 501.2859   3.519595
#> 820 500.0292   2.902842
#> 821 500.0300   2.638331
#> 822 509.6593   3.294056
#> 823 535.0071   3.406181
#> 824 545.2392   3.010907
#> 825 503.3325   2.870831
#> 826 544.9367   3.158280
#> 827 511.6168   2.561739
#> 828 500.2049   2.523979
#> 829 500.0916   2.740882
#> 830 528.8518   3.304547
#> 831 500.3077   2.939029
#> 832 596.9178   2.722316
#> 833 513.5161   3.037711
#> 834 513.5991   2.564287
#> 835 500.9335   2.657481
#> 836 500.2623   2.510081
#> 837 500.3357   2.697816
#> 838 505.4415   3.924459
#> 839 501.0217   3.599470
#> 840 500.0057   2.678811
#> 841 500.0987   2.734850
#> 842 500.2159   2.679270
#> 843 662.4322   2.905783
#> 844 500.2505   3.399672
#> 845 987.2406   2.666462
#> 846 573.2321   3.035238
#> 847 501.7202   3.028965
#> 848 512.4735   2.816396
#> 849 501.0237   2.769264
#> 850 505.8777   2.611352
#> 851 512.6702   2.909819
#> 852 503.9886   3.003719
#> 853 502.7736   3.270178
#> 854 503.6902   2.663057
#> 855 500.2875   2.944508
#> 856 501.9910   3.074166
#> 857 690.9226   2.542477
#> 858 515.5781   3.960045
#> 859 603.1641   4.675154
#> 860 504.7345   3.773598
#> 861 651.7090   3.094492
#> 862 518.5475   2.654823
#> 863 503.4086   2.854445
#> 864 583.6092   2.751118
#> 865 500.1181   2.662005
#> 866 501.3364   2.803424
#> 867 505.0773   3.320999
#> 868 545.6209   2.620126
#> 869 500.1405   2.971300
#> 870 735.8308   2.781270
#> 871 500.8733   2.711047
#> 872 914.7023   2.506944
#> 873 559.7949   2.578441
#> 874 520.6679   3.194524
#> 875 523.6601   2.506786
#> 876 577.3601   3.142684
#> 877 500.1698   2.844865
#> 878 501.4870   3.751512
#> 879 500.0577   2.728920
#> 880 500.0160   2.812629
#> 881 501.6431   2.946834
#> 882 551.9258   3.825484
#> 883 500.1160   3.256402
#> 884 500.7182   2.633942
#> 885 875.2825   2.630643
#> 886 501.1369   2.939328
#> 887 501.6548   2.856028
#> 888 500.1519   2.889030
#> 889 510.8936   3.552495
#> 890 528.3677   2.590968
#> 891 500.5060   2.678220
#> 892 520.1528   3.077157
#> 893 511.7021   3.458622
#> 894 501.6589   3.106940
#> 895 500.4399   2.774577
#> 896 505.6662   2.829639
#> 897 512.0260   2.672067
#> 898 500.7917   2.512378
#> 899 500.4467   2.637697
#> 900 686.0078   2.682048
#> 901 500.0334   3.078832
#> 902 727.8433   2.610580
#> 903 500.0659   2.573531
#> 904 536.4560   3.069848
#> 905 500.3172   2.742166
#> 906 519.0733   3.106030
#> 907 500.0712   3.618081
#> 908 513.3209   3.284031
#> 909 501.6542   3.546985
#> 910 661.6870   2.515090
#> 911 632.8370   2.602195
#> 912 529.3359   3.148679
#> 913 500.1315   2.703956
#> 914 898.3862   2.630816
#> 915 529.9802   3.061367
#> 916 531.8733   2.509485
#> 917 503.4048   2.586083
#> 918 500.3036   2.561499
#> 919 509.9457   2.786651
#> 920 501.4164   3.928199
#> 921 500.1142   2.955139
#> 922 504.5239   2.992129
#> 923 981.9918   2.565651
#> 924 626.1714   4.635319
#> 925 500.0157   2.563380
#> 926 656.8729   3.101241
#> 927 500.1709   3.428308
#> 928 914.5038   2.533657
#> 929 502.9220   2.655259
#> 930 524.8069   2.905132
#> 931 502.2142   2.625469
#> 932 501.0485   4.401849
#> 933 893.1828   3.100768
#> 934 500.0390   2.540230
#> 935 518.0477   2.593165
#> 936 512.0512   2.554971
#> 937 547.1614   2.541540
#> 938 548.3175   2.571506
#> 939 500.0263   3.281536
#> 940 878.2081   2.553749
#> 941 501.1854   3.030046
#> 942 500.9311   3.197979
#> 943 510.6918   2.761530
#> 944 599.9895   2.582214
#> 945 634.5292   2.766095
#> 946 500.2511   3.262066
#> 947 995.4824   3.711386
#> 948 500.4369   2.716710
#> 949 623.7226   2.605875
#> 950 501.3865   3.029313
#> 951 751.8221   3.049326
#> 952 500.1162   2.980742
#> 953 527.0473   3.168016
#> 954 724.3832   2.704490
#> 955 858.3199   2.542590
#> 956 500.2879   2.872783
#> 957 668.0310   2.677076
#> 958 611.7815   3.148129
#> 959 559.4957   2.637029
#> 960 500.2457   2.978105
#> 961 962.0430   3.320410
#> 962 500.3113   2.562139
#> 963 501.1104   3.643523
#> 964 605.2216   2.873303
#> 965 500.5442   3.276302