Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   509.8825   2.911575
#> 2   524.9614   2.537007
#> 3   876.5580   2.590428
#> 4   500.0405   2.808169
#> 5   521.0750   3.505783
#> 6   500.2789   2.528168
#> 7   548.3566   2.522803
#> 8   967.6978   2.587419
#> 9   952.8057   4.085631
#> 10  500.1064   2.836703
#> 11  500.7514   2.977892
#> 12  501.4652   2.576035
#> 13  502.3674   2.834369
#> 14  788.6454   2.533851
#> 15  501.8174   2.562047
#> 16  630.8006   3.128006
#> 17  588.3968   2.762754
#> 18  601.1203   2.972128
#> 19  573.5997   2.722826
#> 20  500.2702   2.618540
#> 21  500.0524   2.661973
#> 22  500.0134   3.583332
#> 23  500.2361   2.714007
#> 24  501.8648   2.587794
#> 25  500.2458   2.698945
#> 26  502.6025   2.719371
#> 27  501.7840   3.135610
#> 28  500.4675   2.927317
#> 29  500.0144   2.521697
#> 30  501.5168   2.658880
#> 31  500.2380   2.617843
#> 32  500.1159   3.971735
#> 33  500.0034   3.219477
#> 34  556.0398   2.723383
#> 35  500.1623   3.122929
#> 36  512.1671   2.866135
#> 37  503.6233   2.601168
#> 38  591.7540   4.540991
#> 39  510.4883   3.999886
#> 40  500.0246   2.865550
#> 41  500.2089   3.146282
#> 42  502.9707   3.812165
#> 43  587.0777   2.981643
#> 44  513.6023   2.605339
#> 45  504.5332   2.591999
#> 46  524.9451   2.934802
#> 47  626.4438   3.023356
#> 48  501.8910   2.668250
#> 49  507.6332   3.363237
#> 50  500.0933   2.917413
#> 51  504.9964   2.694688
#> 52  501.0263   2.934693
#> 53  501.6185   3.162691
#> 54  518.0447   2.594651
#> 55  579.2627   3.083163
#> 56  502.8968   2.507897
#> 57  500.5137   2.610846
#> 58  757.6012   5.721503
#> 59  500.2916   2.582397
#> 60  936.3780   3.023932
#> 61  500.6983   2.545288
#> 62  500.3049   2.902349
#> 63  501.4552   2.581310
#> 64  509.1209   2.536603
#> 65  638.5464   2.521908
#> 66  506.8167   2.551041
#> 67  619.3421   2.608284
#> 68  500.5964   2.838071
#> 69  502.2417   2.950313
#> 70  569.3172   2.575812
#> 71  530.4573   2.727012
#> 72  633.5969   3.231260
#> 73  588.9651   2.972221
#> 74  518.1722   2.640606
#> 75  502.4711   3.080018
#> 76  500.3446   2.763993
#> 77  689.4588   2.558744
#> 78  523.9255   4.235498
#> 79  500.4262   3.372200
#> 80  500.2451   2.693458
#> 81  725.7049   2.626657
#> 82  506.7867   2.538518
#> 83  503.6723   2.785754
#> 84  500.2485   3.176926
#> 85  501.4407   2.634155
#> 86  501.7312   2.691137
#> 87  500.6833   3.114378
#> 88  500.5294   3.018040
#> 89  850.8263   3.091171
#> 90  503.4077   2.573535
#> 91  501.3159   2.843544
#> 92  500.6169   2.685429
#> 93  560.7977   3.728944
#> 94  557.6698   3.137485
#> 95  501.6210   2.527379
#> 96  507.7987   3.865528
#> 97  500.0483   2.618635
#> 98  502.2780   2.723512
#> 99  500.5763   3.218691
#> 100 500.0004   3.782526
#> 101 500.2839   4.279315
#> 102 519.7445   3.429316
#> 103 502.2355   3.222793
#> 104 500.0319   2.599102
#> 105 501.0753   2.798709
#> 106 500.0711   2.765521
#> 107 505.9226   4.553467
#> 108 502.5190   2.712382
#> 109 564.3798   2.502090
#> 110 514.4452   2.873257
#> 111 500.0686   2.521596
#> 112 500.0077   3.576341
#> 113 656.9293   2.715686
#> 114 500.0429   2.625259
#> 115 500.6291   2.513991
#> 116 503.5790   2.505081
#> 117 508.6323   2.595292
#> 118 500.4184   2.593685
#> 119 500.2127   2.567900
#> 120 501.4794   2.502594
#> 121 949.1399   2.805998
#> 122 685.8791   3.026034
#> 123 732.7596   2.829141
#> 124 503.1788   2.584516
#> 125 500.2001   2.962313
#> 126 500.7099   3.022217
#> 127 999.2395   3.714060
#> 128 866.5328   2.579123
#> 129 592.3427   2.765707
#> 130 522.7108   2.916373
#> 131 500.0131   2.979106
#> 132 646.5144   3.132042
#> 133 549.2846   3.199447
#> 134 529.6958   2.518119
#> 135 592.5896   2.500525
#> 136 500.3680   2.947817
#> 137 500.2620   3.640587
#> 138 555.6161   2.552776
#> 139 962.8929   3.871940
#> 140 881.2854   3.852562
#> 141 504.5500   2.876374
#> 142 500.1736   3.096536
#> 143 532.1667   2.780032
#> 144 547.2481   2.880141
#> 145 566.7249   3.126692
#> 146 502.5585   3.116732
#> 147 515.2060   2.810240
#> 148 676.7366   2.580164
#> 149 500.5871   2.544656
#> 150 648.2701   2.648990
#> 151 500.1957   3.450638
#> 152 894.7913   2.936295
#> 153 504.3705   2.886436
#> 154 500.0551   3.044326
#> 155 625.6039   3.206529
#> 156 564.9069   2.839494
#> 157 503.4850   4.113795
#> 158 507.5588   2.856194
#> 159 503.1307   2.556346
#> 160 528.2750   3.037281
#> 161 557.5802   2.526661
#> 162 500.3118   3.919442
#> 163 519.7640   2.518469
#> 164 500.1926   2.949629
#> 165 506.2593   2.518768
#> 166 695.5322   2.952133
#> 167 578.5427   3.057262
#> 168 500.0509   2.584265
#> 169 500.8845   2.782797
#> 170 561.6984   3.039050
#> 171 518.2424   2.782947
#> 172 500.1959   3.057816
#> 173 500.2665   2.800096
#> 174 501.6492   3.119888
#> 175 857.5672   3.062907
#> 176 502.2885   2.760779
#> 177 601.9274   2.921315
#> 178 682.9660   3.048581
#> 179 673.7672   2.631048
#> 180 883.8707   2.906626
#> 181 500.0770   2.940856
#> 182 511.8343   4.137781
#> 183 500.0059   3.167240
#> 184 502.4667   2.842817
#> 185 502.8943   2.509908
#> 186 510.8810   2.650154
#> 187 776.5249   2.713412
#> 188 502.3421   2.988341
#> 189 895.8321   2.773302
#> 190 532.3128   2.904028
#> 191 500.3360   2.715179
#> 192 502.6679   2.843867
#> 193 500.8891   3.715334
#> 194 500.5947   2.800872
#> 195 501.0297   2.610117
#> 196 523.8524   3.596962
#> 197 501.5122   3.081058
#> 198 514.8059   2.618130
#> 199 777.4883   3.673821
#> 200 505.5390   3.100560
#> 201 503.5084   2.600254
#> 202 618.8672   2.701883
#> 203 506.3955   2.528773
#> 204 503.5381   3.461336
#> 205 594.4333   2.660543
#> 206 502.6774   2.638992
#> 207 729.1656   2.732658
#> 208 500.0815   2.621169
#> 209 586.9833   2.655923
#> 210 500.2679   2.556213
#> 211 500.8543   3.016657
#> 212 500.4813   2.725068
#> 213 575.7875   2.791323
#> 214 679.3538   3.856987
#> 215 500.6905   2.709407
#> 216 508.0600   2.542798
#> 217 635.4399   3.665250
#> 218 660.4193   2.958553
#> 219 500.0729   2.649061
#> 220 515.0394   2.502606
#> 221 500.2897   2.676139
#> 222 502.5379   2.529917
#> 223 501.1432   2.868554
#> 224 500.3800   2.721751
#> 225 501.1700   2.630881
#> 226 500.4005   2.764022
#> 227 976.4057   2.740005
#> 228 559.2253   3.113768
#> 229 522.0909   6.718677
#> 230 516.1454   3.406765
#> 231 500.9643   2.566769
#> 232 500.2232   3.023756
#> 233 502.1503   2.790300
#> 234 516.7112   2.686785
#> 235 500.0787   2.683276
#> 236 500.5098   2.613815
#> 237 500.7665   2.824359
#> 238 553.0768   2.613750
#> 239 500.7447   2.947518
#> 240 509.9187   2.865062
#> 241 500.0171   2.636229
#> 242 500.1295   2.554973
#> 243 508.4389   2.878899
#> 244 518.0489   3.289311
#> 245 502.4114   2.553744
#> 246 636.0436   2.884081
#> 247 525.1207   2.869034
#> 248 848.6083   2.587938
#> 249 505.9231   2.916114
#> 250 500.5893   3.099179
#> 251 501.8276   3.085116
#> 252 500.9035   2.648451
#> 253 502.7636   2.688664
#> 254 503.4072   2.730347
#> 255 835.5007   2.690186
#> 256 647.1471   2.501373
#> 257 500.4962   3.260223
#> 258 500.8477   2.701401
#> 259 502.4876   2.578583
#> 260 714.5754   2.595972
#> 261 650.5192   2.749315
#> 262 500.0762   3.005177
#> 263 657.4624   2.862920
#> 264 500.9300   2.523161
#> 265 509.1532   3.352820
#> 266 501.5941   2.741670
#> 267 510.5654   2.646469
#> 268 512.6112   2.582192
#> 269 688.0015   3.077472
#> 270 500.0233   2.635032
#> 271 509.9162   2.798247
#> 272 501.6574   2.564698
#> 273 507.9813   2.644622
#> 274 501.6606   2.812564
#> 275 500.5917   3.213891
#> 276 500.3491   2.739304
#> 277 610.0686   2.692218
#> 278 515.6850   2.536050
#> 279 500.9366   3.130085
#> 280 501.6703   2.783883
#> 281 543.4276   3.332259
#> 282 686.1674   3.059442
#> 283 505.8139   2.843745
#> 284 777.9029   2.589435
#> 285 500.3963   2.758208
#> 286 500.6604   2.675134
#> 287 571.0922   4.687748
#> 288 500.7256   2.715916
#> 289 501.0210   2.621914
#> 290 500.8026   2.698732
#> 291 513.1874   3.493286
#> 292 501.4419   2.963566
#> 293 505.1781   2.827429
#> 294 500.2853   2.634259
#> 295 500.0731   2.525233
#> 296 584.5048   2.515878
#> 297 513.6126   3.303804
#> 298 972.9714   3.113930
#> 299 504.9111   2.562400
#> 300 512.5815   2.665316
#> 301 500.2359   2.646366
#> 302 517.2820   2.721786
#> 303 593.4667   2.797376
#> 304 633.5807   2.522082
#> 305 500.6330   2.507086
#> 306 509.8996   2.901816
#> 307 500.2006   2.693870
#> 308 501.3147   3.121773
#> 309 516.6242   3.430259
#> 310 500.6819   2.703445
#> 311 512.5684   2.599519
#> 312 501.0070   2.745712
#> 313 500.0739   3.241135
#> 314 500.6518   2.539517
#> 315 501.4844   3.875726
#> 316 500.4889   3.086095
#> 317 504.1738   2.978624
#> 318 500.7982   2.546813
#> 319 500.0733   3.301457
#> 320 500.0494   4.581403
#> 321 564.9209   2.904219
#> 322 827.3750   3.283050
#> 323 509.1949   2.936765
#> 324 501.7033   2.503394
#> 325 500.0891   3.034077
#> 326 501.8420   2.929476
#> 327 500.5194   2.745786
#> 328 500.1763   2.736240
#> 329 502.3017   2.656989
#> 330 512.4512   2.967643
#> 331 500.1195   3.049013
#> 332 524.9083   2.695419
#> 333 500.1028   3.666050
#> 334 501.8884   2.554700
#> 335 500.9997   2.726737
#> 336 502.7062   3.180539
#> 337 501.4238   2.519175
#> 338 500.3626   2.941058
#> 339 510.9051   2.720563
#> 340 506.5860   2.693314
#> 341 501.7444   3.569501
#> 342 530.0460   2.790227
#> 343 819.5988   2.793112
#> 344 502.2750   2.559518
#> 345 504.1197   2.563593
#> 346 591.4693   3.736032
#> 347 503.0606   2.648445
#> 348 518.4081   2.886785
#> 349 503.2709   3.398185
#> 350 502.8124   2.843537
#> 351 543.3548   3.693118
#> 352 501.0735   2.605802
#> 353 872.8792   3.032337
#> 354 502.6163   2.781807
#> 355 505.4816   2.938772
#> 356 574.7077   2.637320
#> 357 507.6630   3.162720
#> 358 949.5520   4.007111
#> 359 888.6444   3.402492
#> 360 500.4865   2.684874
#> 361 500.1009   3.250356
#> 362 500.3032   2.627231
#> 363 500.5507   2.813211
#> 364 500.3010   4.526793
#> 365 505.7731   2.958285
#> 366 674.8086   2.692641
#> 367 541.8604   2.635020
#> 368 500.0001   2.864608
#> 369 548.7239   2.756126
#> 370 503.2860   2.670301
#> 371 503.6618   2.565811
#> 372 896.1176   2.631811
#> 373 542.4513   2.558110
#> 374 500.3280   3.052365
#> 375 504.1547   2.556322
#> 376 500.3715   2.526449
#> 377 511.2375   2.516530
#> 378 501.2162   2.750256
#> 379 513.7259   2.629529
#> 380 504.7484   3.479794
#> 381 547.6747   3.033255
#> 382 747.7788   2.611869
#> 383 507.6109   2.810712
#> 384 507.6090   2.551757
#> 385 504.6887   2.696775
#> 386 608.5685   3.335758
#> 387 511.5368   2.586690
#> 388 543.8115   3.419754
#> 389 508.7186   4.125911
#> 390 595.6589   2.507011
#> 391 520.8266   2.520210
#> 392 503.3062   2.549951
#> 393 500.7221   3.305117
#> 394 500.2199   2.929863
#> 395 505.3301   2.670448
#> 396 513.8144   2.790315
#> 397 501.9723   2.659948
#> 398 596.9689   3.256897
#> 399 500.1698   2.611411
#> 400 813.9379   2.641415
#> 401 507.8303   3.311103
#> 402 513.0349   3.354914
#> 403 500.6729   2.818153
#> 404 503.4157   2.505479
#> 405 503.5632   2.589787
#> 406 515.7789   2.558819
#> 407 500.6232   3.116436
#> 408 650.8999   2.669367
#> 409 500.7412   2.508192
#> 410 536.9280   2.753518
#> 411 500.0156   2.585947
#> 412 670.9449   2.799156
#> 413 501.2598   3.144725
#> 414 660.7997   2.557172
#> 415 500.6254   2.655837
#> 416 687.5806   2.665712
#> 417 500.1895   2.817732
#> 418 500.1326   2.554389
#> 419 500.2843   2.605328
#> 420 765.7225   4.459905
#> 421 502.6128   3.170080
#> 422 500.1147   2.851115
#> 423 575.8265   3.276256
#> 424 547.7890   2.723770
#> 425 500.8965   2.637875
#> 426 500.1150   2.727774
#> 427 500.7948   2.824107
#> 428 515.4027   2.763018
#> 429 798.1892   4.258841
#> 430 511.1000   2.584343
#> 431 500.0037   2.819756
#> 432 515.2471   2.666833
#> 433 596.4358   2.604712
#> 434 501.1999   3.046171
#> 435 501.4715   2.531119
#> 436 500.7019   3.606086
#> 437 504.5162   2.664535
#> 438 503.5861   2.527116
#> 439 617.5045   2.713717
#> 440 504.4578   2.592959
#> 441 509.9372   2.713850
#> 442 605.7939   2.662932
#> 443 502.1158   3.094000
#> 444 737.5585   2.633542
#> 445 609.0916   2.670220
#> 446 500.0224   2.806367
#> 447 500.0743   3.144890
#> 448 500.0866   2.650663
#> 449 500.1204   2.587545
#> 450 800.3715   2.595197
#> 451 505.9138   2.896006
#> 452 514.7303   2.681024
#> 453 685.7128   2.553823
#> 454 722.4779   2.662175
#> 455 571.4222   2.853455
#> 456 500.0228   3.002202
#> 457 501.4216   2.507768
#> 458 500.3576   2.544319
#> 459 500.4474   3.637400
#> 460 925.9801   2.553601
#> 461 522.0830   2.578033
#> 462 973.5624   2.869307
#> 463 503.9580   2.600331
#> 464 504.0340   3.368849
#> 465 572.8360   2.734578
#> 466 507.5532   2.748962
#> 467 634.7776   3.115575
#> 468 502.1497   2.595937
#> 469 535.9540   2.856922
#> 470 500.0458   4.027270
#> 471 511.8257   2.755477
#> 472 500.0916   2.851531
#> 473 515.1852   2.693126
#> 474 500.3701   3.167840
#> 475 502.9710   2.847950
#> 476 501.7851   2.879512
#> 477 512.1930   2.670026
#> 478 503.7366   2.650384
#> 479 522.9256   2.590900
#> 480 500.0354   2.654701
#> 481 534.2410   3.165769
#> 482 630.2928   3.624521
#> 483 690.9557   2.582854
#> 484 500.7006   2.711712
#> 485 511.8864   2.876100
#> 486 530.6719   2.633045
#> 487 514.5296   2.872556
#> 488 501.4142   3.032985
#> 489 502.4966   3.894351
#> 490 503.9514   3.712135
#> 491 500.2081   2.580956
#> 492 886.1420   2.541509
#> 493 502.6524   3.095679
#> 494 500.8949   2.674464
#> 495 681.9438   2.961247
#> 496 501.8975   2.799583
#> 497 500.2906   3.022585
#> 498 500.8082   3.161787
#> 499 500.0200   3.498248
#> 500 545.4392   2.629688
#> 501 951.6382   3.291458
#> 502 560.9584   2.819220
#> 503 641.4574   2.728101
#> 504 500.1361   2.544872
#> 505 508.5734   2.530041
#> 506 500.3671   3.042571
#> 507 500.0061   2.827768
#> 508 543.2669   3.134267
#> 509 607.0718   2.897286
#> 510 501.1655   2.964525
#> 511 501.0794   2.611930
#> 512 748.8360   2.525007
#> 513 500.0810   3.954769
#> 514 502.7916   2.745040
#> 515 502.5122   2.634034
#> 516 500.8314   2.763815
#> 517 501.5197   2.751265
#> 518 500.1143   2.882348
#> 519 502.0473   2.962770
#> 520 529.9288   3.015639
#> 521 500.7531   2.561109
#> 522 500.7432   2.673054
#> 523 504.4530   2.526543
#> 524 514.3983   2.565666
#> 525 599.2251   2.862586
#> 526 505.7962   3.858414
#> 527 616.7251   2.636249
#> 528 515.1379   2.623877
#> 529 500.3581   2.512052
#> 530 510.6500   3.225935
#> 531 500.2465   2.739577
#> 532 501.5652   2.920305
#> 533 501.0069   2.567501
#> 534 519.2828   2.697379
#> 535 505.0307   3.153048
#> 536 505.8024   2.639690
#> 537 537.4819   2.747657
#> 538 504.1489   3.011952
#> 539 585.7927   3.391836
#> 540 507.0517   3.032644
#> 541 512.5328   2.519167
#> 542 503.7718   2.794145
#> 543 503.0768   3.104574
#> 544 506.1416   2.598902
#> 545 500.3610   2.584015
#> 546 523.9187   2.839581
#> 547 502.7850   2.775544
#> 548 601.1123   2.853494
#> 549 501.0302   2.526043
#> 550 500.1413   2.719092
#> 551 531.3751   2.664749
#> 552 500.0750   3.173044
#> 553 502.8706   4.436046
#> 554 500.0754   2.998595
#> 555 500.6534   2.673174
#> 556 500.7485   3.375157
#> 557 504.4766   2.590945
#> 558 515.9051   2.599486
#> 559 514.8254   2.558642
#> 560 500.1367   2.664465
#> 561 500.0766   2.528720
#> 562 519.0550   2.648173
#> 563 806.9582   2.668234
#> 564 531.1753   3.043504
#> 565 536.0539   2.530494
#> 566 851.8548   4.037109
#> 567 500.0856   2.804783
#> 568 540.8532   2.729627
#> 569 555.1716   2.595921
#> 570 509.0375   3.772690
#> 571 500.0932   2.744396
#> 572 728.9714   2.564513
#> 573 714.4211   2.558640
#> 574 500.0624   3.954729
#> 575 539.3716   3.271822
#> 576 508.7292   2.575836
#> 577 500.7018   4.775124
#> 578 500.3641   2.600380
#> 579 511.1785   3.293524
#> 580 500.4307   2.533680
#> 581 575.7601   2.586785
#> 582 500.6789   2.636590
#> 583 702.1070   2.667479
#> 584 500.1442   2.980597
#> 585 583.4834   3.264392
#> 586 508.0947   2.510698
#> 587 500.9603   2.691227
#> 588 509.9543   2.877425
#> 589 503.4505   2.920350
#> 590 500.3140   3.032087
#> 591 500.2297   3.063747
#> 592 774.9331   3.342652
#> 593 500.0182   2.537073
#> 594 500.2441   2.709841
#> 595 500.0772   2.684230
#> 596 681.4037   2.686101
#> 597 502.6569   3.193888
#> 598 500.9010   2.724858
#> 599 820.7498   2.671086
#> 600 500.8988   3.145684
#> 601 501.7411   2.958012
#> 602 542.8520   2.846124
#> 603 500.1189   3.264265
#> 604 557.0612   2.608168
#> 605 500.3619   2.697540
#> 606 624.5335   2.523355
#> 607 500.2475   2.655133
#> 608 587.7504   2.741735
#> 609 500.7963   2.519514
#> 610 500.1689   3.254570
#> 611 504.3830   2.914934
#> 612 877.2703   2.510239
#> 613 500.2072   2.848524
#> 614 759.8577   2.635957
#> 615 503.8574   3.308407
#> 616 508.5480   2.506675
#> 617 503.9617   2.599384
#> 618 544.6573   2.692088
#> 619 500.7099   2.816673
#> 620 500.0410   2.917258
#> 621 502.5344   3.585222
#> 622 502.4291   2.581079
#> 623 777.9439   2.651233
#> 624 500.6279   2.835036
#> 625 500.9869   3.004799
#> 626 512.8115   2.551335
#> 627 576.2101   3.798969
#> 628 641.8298   2.519280
#> 629 500.4688   3.039898
#> 630 504.9291   3.344329
#> 631 516.6893   3.497006
#> 632 514.2589   2.656431
#> 633 501.0131   2.568951
#> 634 512.9940   3.570284
#> 635 819.4619   3.059489
#> 636 500.4883   3.537227
#> 637 500.0705   2.796259
#> 638 511.6500   3.326463
#> 639 653.8599   2.583864
#> 640 530.6980   2.501039
#> 641 500.0623   2.858050
#> 642 502.7173   3.304997
#> 643 500.5226   2.554824
#> 644 500.4224   2.599306
#> 645 500.2659   2.584873
#> 646 500.0395   2.610483
#> 647 653.0693   2.734434
#> 648 500.0121   3.496654
#> 649 500.1635   2.806191
#> 650 504.7956   2.515975
#> 651 501.5748   2.544413
#> 652 513.7165   2.517954
#> 653 500.8932   2.601021
#> 654 500.3880   2.701809
#> 655 510.9030   2.624734
#> 656 526.8143   2.829297
#> 657 500.0375   3.224079
#> 658 520.2556   3.798624
#> 659 500.0640   2.692485
#> 660 500.0118   2.638232
#> 661 500.5888   2.569159
#> 662 581.3983   2.989829
#> 663 510.5298   2.508841
#> 664 502.1778   2.881610
#> 665 579.6151   3.478005
#> 666 517.1223   2.834736
#> 667 500.3900   2.744437
#> 668 502.0725   2.545612
#> 669 500.1512   2.569163
#> 670 532.6236   2.692935
#> 671 502.6279   2.661644
#> 672 500.2910   2.770005
#> 673 566.1035   3.005325
#> 674 500.9475   2.588441
#> 675 501.3132   2.541948
#> 676 544.4973   3.834595
#> 677 500.0311   2.658518
#> 678 500.0036   3.102940
#> 679 633.8952   2.799002
#> 680 501.9686   2.725153
#> 681 505.9960   2.954384
#> 682 583.2019   2.914920
#> 683 806.5732   3.187443
#> 684 525.0672   2.651048
#> 685 500.7329   2.669996
#> 686 500.0728   3.063247
#> 687 504.7735   2.821323
#> 688 501.0295   3.995058
#> 689 589.7770   3.541562
#> 690 501.5794   2.531764
#> 691 506.9034   2.761483
#> 692 971.3604   2.564253
#> 693 501.7264   2.774359
#> 694 500.2389   2.614024
#> 695 500.0637   2.602758
#> 696 503.8646   2.660120
#> 697 500.1742   3.053388
#> 698 500.5018   2.605065
#> 699 504.1722   2.904489
#> 700 500.7272   3.704126
#> 701 500.4260   3.354388
#> 702 501.4130   2.545607
#> 703 524.8998   2.706930
#> 704 502.6125   2.617341
#> 705 507.8711   2.601082
#> 706 556.7232   3.053718
#> 707 500.7068   3.078764
#> 708 500.8277   3.143656
#> 709 501.4633   3.108706
#> 710 959.9013   2.839726
#> 711 500.0352   2.772241
#> 712 500.0045   2.707522
#> 713 500.5604   2.772907
#> 714 500.1846   2.627867
#> 715 503.1533   3.852442
#> 716 623.2365   2.785431
#> 717 807.8000   2.555537
#> 718 503.0866   2.516796
#> 719 513.3439   2.728540
#> 720 532.1550   2.905153
#> 721 507.7025   2.549208
#> 722 500.0028   2.660694
#> 723 531.2905   2.921295
#> 724 500.9611   3.058908
#> 725 535.1843   3.138048
#> 726 568.9438   3.022210
#> 727 500.1819   3.121842
#> 728 507.3085   2.587776
#> 729 500.1039   3.918544
#> 730 503.8040   2.832585
#> 731 508.7191   2.851776
#> 732 501.4341   2.553613
#> 733 500.2526   2.515451
#> 734 500.5605   2.694267
#> 735 501.7700   2.561175
#> 736 500.5309   2.558133
#> 737 503.4800   3.180981
#> 738 502.3381   2.583907
#> 739 904.8011   2.526141
#> 740 500.1413   3.421068
#> 741 500.2674   3.403756
#> 742 500.1475   2.538552
#> 743 500.1984   2.628289
#> 744 501.9082   2.835495
#> 745 969.3534   2.539737
#> 746 506.1343   2.736410
#> 747 963.3814   3.188268
#> 748 507.0703   3.153798
#> 749 500.1746   2.502672
#> 750 500.8019   3.065145
#> 751 506.6043   2.770881
#> 752 500.2551   2.825129
#> 753 500.0538   2.809658
#> 754 512.0426   2.850996
#> 755 500.6079   2.881787
#> 756 500.2531   3.003344
#> 757 512.7592   2.734502
#> 758 500.0370   2.678084
#> 759 500.0814   2.547192
#> 760 500.1731   2.673697
#> 761 528.4251   2.537660
#> 762 503.2815   2.640860
#> 763 501.8064   2.940643
#> 764 502.8070   3.606486
#> 765 504.1938   2.684901
#> 766 500.0393   2.662854
#> 767 640.8294   2.871955
#> 768 669.0130   2.611137
#> 769 501.4522   2.760082
#> 770 785.5219   2.610722
#> 771 500.1021   2.519037
#> 772 500.0663   2.603206
#> 773 524.1342   2.649237
#> 774 502.8228   2.707459
#> 775 539.8816   2.714256
#> 776 500.2158   2.515781
#> 777 500.6524   3.796506
#> 778 625.0024   3.988788
#> 779 500.2125   2.678603
#> 780 866.8703   2.604340
#> 781 500.1375   2.686730
#> 782 783.9653   3.219253
#> 783 502.1540   2.824108
#> 784 509.8472   2.514949
#> 785 542.2842   2.761605
#> 786 507.8999   3.071793
#> 787 500.3692   3.026829
#> 788 543.1920   2.553997
#> 789 502.9158   2.555428
#> 790 501.0013   2.885655
#> 791 502.9951   2.715324
#> 792 522.5473   2.944422
#> 793 501.2884   2.760226
#> 794 779.2558   2.769017
#> 795 745.7107   3.310824
#> 796 500.0039   2.514823
#> 797 502.0228   2.581352
#> 798 500.3669   2.519177
#> 799 506.5394   4.220843
#> 800 619.0701   3.221797
#> 801 858.3552   2.940572
#> 802 804.0053   3.076291
#> 803 501.1311   2.500445
#> 804 516.6935   2.814032
#> 805 500.0174   2.593983
#> 806 527.0601   3.132849
#> 807 500.0910   2.574530
#> 808 500.8039   2.864986
#> 809 535.8036   3.123274
#> 810 511.0184   2.852077
#> 811 520.1656   2.580322
#> 812 500.4410   2.512588
#> 813 507.4694   3.069995
#> 814 553.6956   2.907298
#> 815 509.1729   2.733670
#> 816 500.8423   2.610486
#> 817 500.4334   4.473671
#> 818 577.1937   2.820606
#> 819 507.1694   3.072634
#> 820 501.8475   2.790955
#> 821 550.1435   3.392400
#> 822 562.5409   3.145829
#> 823 564.4007   3.263952
#> 824 501.2773   2.509880
#> 825 501.2077   3.170169
#> 826 520.6432   2.629081
#> 827 525.0087   4.175485
#> 828 534.8473   2.567649
#> 829 516.6056   3.076639
#> 830 522.9161   3.064892
#> 831 613.6291   2.517980
#> 832 502.3887   3.180348
#> 833 526.6214   2.626285
#> 834 504.2300   2.928880
#> 835 501.2566   2.512047
#> 836 500.2408   2.624616
#> 837 548.7487   2.578061
#> 838 500.0337   4.037770
#> 839 526.8231   3.011150
#> 840 505.4315   2.615841
#> 841 528.7246   3.411909
#> 842 508.7475   2.686854
#> 843 500.1527   3.243769
#> 844 500.9322   2.879128
#> 845 503.6475   2.521081
#> 846 500.0502   2.756660
#> 847 500.1048   2.640532
#> 848 541.4631   2.882039
#> 849 501.2597   2.685075
#> 850 670.8614   3.385218
#> 851 500.2888   2.588731
#> 852 500.4043   2.678178
#> 853 705.4403   2.736031
#> 854 500.8501   2.989718
#> 855 513.8183   2.957557
#> 856 520.1796   2.543413
#> 857 673.0293   2.536784
#> 858 500.2105   2.723340
#> 859 503.4874   2.927125
#> 860 528.6721   2.620728
#> 861 605.1313   3.154144
#> 862 501.1411   2.534793
#> 863 505.0348   3.069127
#> 864 501.5452   3.363536
#> 865 507.2309   2.604425
#> 866 517.4913   2.629133
#> 867 525.0711   2.834733
#> 868 542.3126   2.623534
#> 869 500.7869   2.507371
#> 870 533.6654   2.922291
#> 871 504.2520   2.544435
#> 872 500.0851   2.512181
#> 873 682.7079   2.652949
#> 874 503.0911   2.837658
#> 875 500.0603   2.785051
#> 876 500.0176   2.744351
#> 877 500.5992   2.865601
#> 878 502.0465   2.792393
#> 879 500.2715   2.695790
#> 880 508.4025   2.609406
#> 881 500.3087   3.797506
#> 882 500.8838   2.690734
#> 883 500.2058   2.509665
#> 884 842.7710   3.673296
#> 885 509.3461   5.164861
#> 886 551.5114   2.565523
#> 887 515.4238   2.700826
#> 888 500.1612   2.545448
#> 889 500.2037   2.585997
#> 890 513.2489   2.723405
#> 891 500.1507   3.134137
#> 892 549.0856   2.899582
#> 893 501.1266   2.507450
#> 894 500.1667   2.610305
#> 895 500.9292   2.665127
#> 896 500.0363   3.564629
#> 897 501.6979   2.538782
#> 898 500.2280   2.827221
#> 899 500.7399   4.584739
#> 900 500.7935   2.708218
#> 901 501.8798   3.110935
#> 902 512.1373   2.652689
#> 903 500.7796   2.787998
#> 904 691.7809   2.534930
#> 905 526.2261   3.588137
#> 906 500.0480   2.643277
#> 907 517.8567   2.511191
#> 908 532.3843   2.720731
#> 909 618.6437   3.125586
#> 910 509.5984   2.759520
#> 911 502.8964   2.569354
#> 912 662.6134   2.576413
#> 913 646.4929   4.453147
#> 914 862.6884   2.585397
#> 915 500.6324   2.633943
#> 916 501.8935   2.733857
#> 917 512.8938   3.433154
#> 918 501.3147   2.639090
#> 919 501.7324   2.797348
#> 920 686.6266   2.611572
#> 921 832.8488   3.090142