Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   508.8629   2.613067
#> 2   500.4151   3.370275
#> 3   500.9204   2.993812
#> 4   500.7842   2.728455
#> 5   504.4852   4.249926
#> 6   510.0361   2.662017
#> 7   502.9883   2.706600
#> 8   546.5740   2.596780
#> 9   500.2038   2.818021
#> 10  502.7749   2.543301
#> 11  500.0971   2.613443
#> 12  630.9887   2.854819
#> 13  501.3271   2.628370
#> 14  500.1833   3.717105
#> 15  501.4450   2.518459
#> 16  501.6942   3.006857
#> 17  500.4953   2.989674
#> 18  506.2737   2.595192
#> 19  515.1403   2.575367
#> 20  500.5075   2.833331
#> 21  501.9540   3.119444
#> 22  868.8978   2.615981
#> 23  500.7195   2.731397
#> 24  505.9040   2.877471
#> 25  500.0989   2.526075
#> 26  500.0804   2.635050
#> 27  500.6075   3.354095
#> 28  513.6977   3.736241
#> 29  500.8709   2.545100
#> 30  503.2345   2.673891
#> 31  501.1158   3.303477
#> 32  500.3578   2.638206
#> 33  512.7735   2.542470
#> 34  519.1628   2.891267
#> 35  500.5985   2.971835
#> 36  502.3365   2.512429
#> 37  500.1335   2.949137
#> 38  527.0345   3.446214
#> 39  502.3692   2.650416
#> 40  501.3432   2.762658
#> 41  781.7354   2.819948
#> 42  500.4004   2.654603
#> 43  500.8002   2.624636
#> 44  502.6124   2.593687
#> 45  836.9082   3.099457
#> 46  567.7522   2.920675
#> 47  503.0396   3.032622
#> 48  502.5058   2.988942
#> 49  551.2225   2.682887
#> 50  502.3764   3.741298
#> 51  538.0780   3.249558
#> 52  531.0655   2.769608
#> 53  501.7914   2.637486
#> 54  506.0894   2.844097
#> 55  500.1814   2.718443
#> 56  501.9396   3.001827
#> 57  571.7299   2.751945
#> 58  501.4768   2.617931
#> 59  627.7842   3.004260
#> 60  535.0099   3.270226
#> 61  501.1746   3.052523
#> 62  503.6845   2.581044
#> 63  880.6604   2.801392
#> 64  500.0866   3.080459
#> 65  504.3556   2.531579
#> 66  501.0814   2.748690
#> 67  502.4440   2.846142
#> 68  501.1697   3.126881
#> 69  500.0303   2.569875
#> 70  500.0197   2.560667
#> 71  548.6336   4.094843
#> 72  500.0265   4.050202
#> 73  500.7102   2.671354
#> 74  501.4161   2.586554
#> 75  505.1173   2.521673
#> 76  524.8186   2.569349
#> 77  825.3249   2.554313
#> 78  542.7275   2.648353
#> 79  500.1022   2.557190
#> 80  501.2920   2.744131
#> 81  503.9092   3.353428
#> 82  500.1778   2.738981
#> 83  529.2298   2.921469
#> 84  554.3301   2.933745
#> 85  897.4744   6.336187
#> 86  504.8110   2.502090
#> 87  500.1152   2.966682
#> 88  500.1760   3.047527
#> 89  500.0163   3.179724
#> 90  500.6478   2.546793
#> 91  500.0214   2.771147
#> 92  604.9569   3.515318
#> 93  501.3415   2.766883
#> 94  501.0241   3.189385
#> 95  500.0493   2.585405
#> 96  500.5816   2.502528
#> 97  500.2938   3.189432
#> 98  500.3068   2.689593
#> 99  605.7446   2.683952
#> 100 505.9753   4.006019
#> 101 501.4773   2.918570
#> 102 500.6434   2.928975
#> 103 500.2765   2.846372
#> 104 500.1327   2.783748
#> 105 510.8688   3.212882
#> 106 500.8177   2.621592
#> 107 560.4132   3.681156
#> 108 500.0028   2.731482
#> 109 586.8304   2.623744
#> 110 500.0189   4.372268
#> 111 500.1022   2.622186
#> 112 500.0548   4.333583
#> 113 500.8667   3.201525
#> 114 500.8986   2.977905
#> 115 525.9449   2.578644
#> 116 501.6733   2.994320
#> 117 500.4441   2.526729
#> 118 544.4388   3.867801
#> 119 500.4980   2.788371
#> 120 500.1244   2.731932
#> 121 500.0398   4.276355
#> 122 503.5925   2.552064
#> 123 507.5830   2.695094
#> 124 501.4247   3.257380
#> 125 507.2610   2.819651
#> 126 500.1580   2.635350
#> 127 500.0473   2.608908
#> 128 501.4597   2.580249
#> 129 539.1065   2.815858
#> 130 500.4984   3.241489
#> 131 500.1002   2.883505
#> 132 504.4432   3.155928
#> 133 500.1463   3.341860
#> 134 633.9930   2.783423
#> 135 500.0579   3.130412
#> 136 502.8277   2.660522
#> 137 500.6123   3.237116
#> 138 500.0261   3.487928
#> 139 775.8610   3.355378
#> 140 500.0150   2.776821
#> 141 500.0914   2.692580
#> 142 777.9831   2.854818
#> 143 500.8607   2.877303
#> 144 500.3834   2.944347
#> 145 562.8650   2.929045
#> 146 500.4133   2.559882
#> 147 501.0841   3.113672
#> 148 500.1081   2.613961
#> 149 526.4264   3.781222
#> 150 500.3442   4.010832
#> 151 507.7813   3.932210
#> 152 500.0541   3.023124
#> 153 540.4329   3.109695
#> 154 501.7326   3.374131
#> 155 501.9331   3.229783
#> 156 746.0221   2.908735
#> 157 500.5559   3.249396
#> 158 502.8504   3.018769
#> 159 502.0357   2.916286
#> 160 500.3149   2.677417
#> 161 500.0434   2.875218
#> 162 500.5052   2.986915
#> 163 539.9659   2.504740
#> 164 500.6186   3.476246
#> 165 631.4130   2.528480
#> 166 506.9769   3.279835
#> 167 500.1535   2.652434
#> 168 644.0575   4.249701
#> 169 616.7926   3.149232
#> 170 517.6848   2.743317
#> 171 500.1410   2.624620
#> 172 500.0006   2.613255
#> 173 500.2127   2.604349
#> 174 582.6291   2.648191
#> 175 504.3712   2.861264
#> 176 665.9205   3.479513
#> 177 500.3300   2.804971
#> 178 603.7936   2.785367
#> 179 509.8666   3.158597
#> 180 500.1060   2.544570
#> 181 504.6525   2.532209
#> 182 512.5938   2.621759
#> 183 502.1554   3.421871
#> 184 542.1827   2.537919
#> 185 614.0791   2.688166
#> 186 501.2447   2.588008
#> 187 500.4940   2.921381
#> 188 558.2299   2.670716
#> 189 500.0665   2.845611
#> 190 508.6299   2.506526
#> 191 500.6030   2.550312
#> 192 508.6550   3.607531
#> 193 507.1628   3.330182
#> 194 586.0605   2.581990
#> 195 503.4557   2.755403
#> 196 500.1746   2.597600
#> 197 853.5046   3.361191
#> 198 671.7955   3.306882
#> 199 792.4758   2.773212
#> 200 500.0854   2.688764
#> 201 545.2322   3.140626
#> 202 524.4500   2.632309
#> 203 674.3580   2.551313
#> 204 504.6765   2.647173
#> 205 500.2920   2.997335
#> 206 500.2993   2.575611
#> 207 809.6423   2.582961
#> 208 500.1016   2.535269
#> 209 505.8251   2.637034
#> 210 511.7765   3.231597
#> 211 500.7674   3.407313
#> 212 545.0331   2.666555
#> 213 500.0852   3.120136
#> 214 906.5462   2.832591
#> 215 541.6511   3.156621
#> 216 828.5275   2.834739
#> 217 500.8763   3.010871
#> 218 500.0980   3.240027
#> 219 535.5807   2.815355
#> 220 568.4194   2.541836
#> 221 500.0097   2.713857
#> 222 502.4185   3.272974
#> 223 500.0995   2.569687
#> 224 500.0204   2.603403
#> 225 537.9280   3.120271
#> 226 500.1651   2.659455
#> 227 511.8364   2.905874
#> 228 503.9182   2.562140
#> 229 500.0936   3.114814
#> 230 500.2050   2.655937
#> 231 500.0139   2.985825
#> 232 659.1105   2.599915
#> 233 515.9641   2.834734
#> 234 500.0480   2.627578
#> 235 504.7026   2.691167
#> 236 501.0337   2.831623
#> 237 713.1939   2.640314
#> 238 503.1025   3.340394
#> 239 500.0404   3.144919
#> 240 500.4489   2.852073
#> 241 500.4914   2.689534
#> 242 500.5276   3.642068
#> 243 513.3485   3.095030
#> 244 500.0843   3.925092
#> 245 500.6593   2.966490
#> 246 512.9232   3.779840
#> 247 500.1975   2.567268
#> 248 574.2329   2.509050
#> 249 563.3959   3.011875
#> 250 500.2320   2.541713
#> 251 514.9413   2.795272
#> 252 500.0041   2.654689
#> 253 576.6976   2.871769
#> 254 500.1276   2.588238
#> 255 505.6699   3.271952
#> 256 600.7766   3.107046
#> 257 773.7486   2.625473
#> 258 500.3906   3.031134
#> 259 522.2545   2.884212
#> 260 501.3251   2.829425
#> 261 500.0108   3.673554
#> 262 502.0700   2.619347
#> 263 500.1321   2.630659
#> 264 500.0700   4.679622
#> 265 503.3571   2.780428
#> 266 500.2101   2.706693
#> 267 500.0448   3.609899
#> 268 624.1488   2.502644
#> 269 500.0397   2.714956
#> 270 505.9467   3.469295
#> 271 510.4908   3.757120
#> 272 500.0081   2.551622
#> 273 535.7453   2.898031
#> 274 747.0203   3.053628
#> 275 537.8646   2.742398
#> 276 535.1632   2.526657
#> 277 556.0373   3.567189
#> 278 502.3813   2.793258
#> 279 505.7180   3.330369
#> 280 509.2461   2.562392
#> 281 796.6501   2.904456
#> 282 530.7902   2.718242
#> 283 500.0299   2.530113
#> 284 509.8429   2.763392
#> 285 500.2306   2.808154
#> 286 500.1292   3.745485
#> 287 504.8069   3.082166
#> 288 500.0093   2.660850
#> 289 500.3785   3.243968
#> 290 506.4375   3.620205
#> 291 504.8062   3.076089
#> 292 502.2012   2.951651
#> 293 503.9029   2.772963
#> 294 500.0716   2.984832
#> 295 511.5176   3.490924
#> 296 500.5187   2.780875
#> 297 533.9994   3.740282
#> 298 502.4660   2.753479
#> 299 555.4381   2.768204
#> 300 511.0058   2.558016
#> 301 517.6692   2.501329
#> 302 533.4650   2.531120
#> 303 518.5456   2.726352
#> 304 511.5890   3.343720
#> 305 500.5696   2.594848
#> 306 656.0038   2.690125
#> 307 500.2312   2.507128
#> 308 501.5986   2.676562
#> 309 554.1734   2.500517
#> 310 714.4172   2.759969
#> 311 573.4264   2.542071
#> 312 618.5026   2.592178
#> 313 500.0133   2.530538
#> 314 651.9622   3.921963
#> 315 560.9661   2.538470
#> 316 602.3775   2.534141
#> 317 500.3491   3.615102
#> 318 504.5329   2.792303
#> 319 553.1758   2.922754
#> 320 500.1169   2.925327
#> 321 500.1146   2.884524
#> 322 500.2548   2.850439
#> 323 750.6907   3.204502
#> 324 522.7510   2.536762
#> 325 502.0934   2.564966
#> 326 500.0392   3.135130
#> 327 500.0527   3.616106
#> 328 504.3258   3.185346
#> 329 551.6503   2.795781
#> 330 530.6920   2.797190
#> 331 500.8529   2.577583
#> 332 500.4083   2.749011
#> 333 500.3455   3.207734
#> 334 501.6781   3.580226
#> 335 500.0738   2.571693
#> 336 505.2900   2.505818
#> 337 501.2020   2.565943
#> 338 504.1577   2.632976
#> 339 500.2813   2.803215
#> 340 502.4715   2.719927
#> 341 500.5174   2.509165
#> 342 500.9927   2.715253
#> 343 734.3589   2.886804
#> 344 505.4781   3.215075
#> 345 506.1380   3.239240
#> 346 546.7637   2.545885
#> 347 520.7107   3.011314
#> 348 544.5940   2.524861
#> 349 503.1571   2.624041
#> 350 506.1185   2.883942
#> 351 503.4004   3.240532
#> 352 501.8534   2.568525
#> 353 766.2930   2.696651
#> 354 500.6890   2.735243
#> 355 505.9477   2.818861
#> 356 519.0817   3.178590
#> 357 506.0935   2.648352
#> 358 545.5096   3.099028
#> 359 501.5254   2.650101
#> 360 526.7305   2.505855
#> 361 500.1445   3.119337
#> 362 507.8064   2.993514
#> 363 500.3358   3.119652
#> 364 546.4687   2.525146
#> 365 500.1511   2.534567
#> 366 500.5233   2.631152
#> 367 500.1465   2.958293
#> 368 511.0153   2.771919
#> 369 500.1515   2.546792
#> 370 504.1234   3.173800
#> 371 504.2219   4.363851
#> 372 501.5753   2.742498
#> 373 500.1893   2.633828
#> 374 500.8382   3.221530
#> 375 549.8636   2.654344
#> 376 500.3866   3.171860
#> 377 704.2345   2.718185
#> 378 521.7021   3.321984
#> 379 500.1343   2.850746
#> 380 500.0176   2.781885
#> 381 500.8243   3.493136
#> 382 500.0431   2.631098
#> 383 590.0978   2.609755
#> 384 500.1078   2.576004
#> 385 500.5455   2.512577
#> 386 534.0111   3.411227
#> 387 501.7164   2.598946
#> 388 506.1429   2.510281
#> 389 500.1054   2.666052
#> 390 794.3703   2.995190
#> 391 501.4564   4.908539
#> 392 504.0866   3.357914
#> 393 616.4072   4.501702
#> 394 503.1464   3.916839
#> 395 500.8910   2.704119
#> 396 500.7806   2.817043
#> 397 500.2437   2.602585
#> 398 516.9375   2.655679
#> 399 501.9964   3.352162
#> 400 500.2453   2.879599
#> 401 504.5208   3.327444
#> 402 501.2500   2.991641
#> 403 501.4846   2.710789
#> 404 500.2294   4.088463
#> 405 501.7006   2.901869
#> 406 753.1053   3.404148
#> 407 501.9823   5.435391
#> 408 525.2132   3.322299
#> 409 838.4871   2.655355
#> 410 613.1461   3.700228
#> 411 863.8270   3.033212
#> 412 500.6258   2.935399
#> 413 500.5585   2.649342
#> 414 500.3869   2.642587
#> 415 511.1041   2.883386
#> 416 524.6420   2.772868
#> 417 514.0575   3.395153
#> 418 509.8062   2.658973
#> 419 575.1422   3.256719
#> 420 500.5514   3.082501
#> 421 504.7047   2.730004
#> 422 515.3387   2.920576
#> 423 500.2635   2.579329
#> 424 500.9168   2.628282
#> 425 501.7035   3.041581
#> 426 512.3370   3.020897
#> 427 500.1547   2.658721
#> 428 677.4510   2.973013
#> 429 500.0490   2.962529
#> 430 506.8189   2.950868
#> 431 754.9929   2.533261
#> 432 500.1147   2.868211
#> 433 500.3804   2.601437
#> 434 502.4205   2.585951
#> 435 547.7379   2.503597
#> 436 500.1220   2.694892
#> 437 500.4302   3.366305
#> 438 501.4587   3.160896
#> 439 500.0326   3.040911
#> 440 500.2059   2.989652
#> 441 500.1262   2.645799
#> 442 500.0126   2.926540
#> 443 500.0023   2.718817
#> 444 500.3045   3.115468
#> 445 507.0385   3.186789
#> 446 503.0185   2.926571
#> 447 505.4763   3.393429
#> 448 502.7353   2.853843
#> 449 502.2014   2.633060
#> 450 500.1985   3.481068
#> 451 870.7485   3.419259
#> 452 681.7305   3.337137
#> 453 522.7596   2.505421
#> 454 501.6904   2.598658
#> 455 500.3975   2.898914
#> 456 500.9674   2.996298
#> 457 500.0446   2.561218
#> 458 551.1467   2.956596
#> 459 501.7383   2.791494
#> 460 529.5771   2.559771
#> 461 531.8774   2.880243
#> 462 501.9848   3.602197
#> 463 570.3202   2.684936
#> 464 500.7685   2.701131
#> 465 509.0766   2.722488
#> 466 508.1089   2.595530
#> 467 535.7120   2.518571
#> 468 517.4717   2.658889
#> 469 503.1993   2.595098
#> 470 500.6176   2.525065
#> 471 500.5410   2.556299
#> 472 539.3761   2.610800
#> 473 653.3068   4.269492
#> 474 618.6900   3.583827
#> 475 669.7763   2.738043
#> 476 501.0040   2.768986
#> 477 502.2124   2.626058
#> 478 507.7229   2.963919
#> 479 501.3987   3.553016
#> 480 500.1294   3.605365
#> 481 500.1899   2.665432
#> 482 500.6408   3.205589
#> 483 524.7176   2.840210
#> 484 500.2229   3.126139
#> 485 500.0062   3.111857
#> 486 502.8502   3.323241
#> 487 500.0852   2.725777
#> 488 709.6484   2.562941
#> 489 512.8570   3.495969
#> 490 500.2082   4.280204
#> 491 774.2940   3.542887
#> 492 831.1882   2.624953
#> 493 505.2082   3.214173
#> 494 500.9033   2.683291
#> 495 500.1288   2.705728
#> 496 808.2320   3.155799
#> 497 500.2859   2.509645
#> 498 522.2588   2.679120
#> 499 500.0937   2.813704
#> 500 500.4809   3.022985
#> 501 500.2997   2.633564
#> 502 516.6057   2.667051
#> 503 564.0261   2.592577
#> 504 500.3302   3.593538
#> 505 689.7652   2.525431
#> 506 500.2218   2.667608
#> 507 500.7680   2.700534
#> 508 500.3519   2.537406
#> 509 506.2806   3.244635
#> 510 531.4364   2.606297
#> 511 500.1717   3.105365
#> 512 535.4390   2.736141
#> 513 510.6025   2.562545
#> 514 500.9287   2.542826
#> 515 500.0321   2.554429
#> 516 575.3948   2.922467
#> 517 529.3939   5.337965
#> 518 528.8998   3.528402
#> 519 500.1602   2.513650
#> 520 502.1111   3.461545
#> 521 500.0242   2.515730
#> 522 735.4850   2.679767
#> 523 501.3896   3.365137
#> 524 502.4350   2.970624
#> 525 504.6648   2.563068
#> 526 501.2715   5.065599
#> 527 504.7455   2.627036
#> 528 508.8460   3.220330
#> 529 500.4282   2.561660
#> 530 563.8683   2.899957
#> 531 569.3734   2.566105
#> 532 500.0352   2.686333
#> 533 556.0026   2.837990
#> 534 504.6272   3.189054
#> 535 507.9333   2.861729
#> 536 595.7655   2.785512
#> 537 571.2516   2.959472
#> 538 500.4902   2.572677
#> 539 510.1936   3.127299
#> 540 507.6335   2.909990
#> 541 538.0804   3.149361
#> 542 500.4866   3.258711
#> 543 506.2566   2.753892
#> 544 508.1574   2.572986
#> 545 585.2624   2.597614
#> 546 535.4526   2.911675
#> 547 571.0310   3.167707
#> 548 826.5556   3.050151
#> 549 500.2616   2.789649
#> 550 500.0444   2.565844
#> 551 501.1253   2.904953
#> 552 509.2379   2.571412
#> 553 502.5448   2.926504
#> 554 500.7172   2.911274
#> 555 500.0243   2.634458
#> 556 500.5729   2.834917
#> 557 982.0503   3.119116
#> 558 509.7887   2.647953
#> 559 546.0672   4.053283
#> 560 500.8077   2.929338
#> 561 502.1906   3.128537
#> 562 500.0814   2.549464
#> 563 500.1853   3.062522
#> 564 531.2842   3.306219
#> 565 511.5088   2.610618
#> 566 500.2480   2.778269
#> 567 502.9858   2.531577
#> 568 501.8905   2.501683
#> 569 500.8059   3.383061
#> 570 531.5617   2.890324
#> 571 505.4416   3.179663
#> 572 500.7487   2.554700
#> 573 500.0721   2.542773
#> 574 755.2528   3.139972
#> 575 500.5565   3.295567
#> 576 506.4566   2.959975
#> 577 501.9051   2.624520
#> 578 500.0214   3.319614
#> 579 542.3989   2.511606
#> 580 528.6935   3.162055
#> 581 500.1626   2.653547
#> 582 514.0854   2.810986
#> 583 500.0188   2.542153
#> 584 516.2613   2.500920
#> 585 501.3101   2.742812
#> 586 502.8927   2.522202
#> 587 500.0480   2.592779
#> 588 500.0040   2.526819
#> 589 500.1740   2.668406
#> 590 511.3039   3.155031
#> 591 501.9907   2.765282
#> 592 500.4933   2.557422
#> 593 503.6616   2.964237
#> 594 501.3897   2.927581
#> 595 501.1059   2.664357
#> 596 502.1217   2.557195
#> 597 521.2979   2.636822
#> 598 607.6209   2.747066
#> 599 501.2965   2.629557
#> 600 790.6827   2.590660
#> 601 500.2064   2.673762
#> 602 513.6900   2.778036
#> 603 642.0948   2.950736
#> 604 500.2226   2.543900
#> 605 511.6498   2.818256
#> 606 731.9465   3.178506
#> 607 505.0961   2.563475
#> 608 504.4534   2.730635
#> 609 500.0025   3.915492
#> 610 500.1242   2.750962
#> 611 500.2497   3.283858
#> 612 500.0138   2.646227
#> 613 504.5799   2.746565
#> 614 506.7282   2.578995
#> 615 500.8399   3.435308
#> 616 509.5907   3.241903
#> 617 919.4270   3.186322
#> 618 504.4125   3.831428
#> 619 544.5769   3.172664
#> 620 518.9110   2.513982
#> 621 500.1816   2.782740
#> 622 500.8186   2.682501
#> 623 581.1112   2.932449
#> 624 508.0613   2.932305
#> 625 509.2271   3.934743
#> 626 500.0288   2.660273
#> 627 500.0526   2.633864
#> 628 596.4951   2.630753
#> 629 510.3790   2.640350
#> 630 501.8555   2.568277
#> 631 500.2808   2.919444
#> 632 503.2260   2.769979
#> 633 500.0195   3.460470
#> 634 500.0200   3.331529
#> 635 556.3170   2.819809
#> 636 529.5507   2.872016
#> 637 872.9734   2.873542
#> 638 501.5230   3.068211
#> 639 500.7131   2.668032
#> 640 539.8525   3.816181
#> 641 549.1906   2.563835
#> 642 504.9591   2.824866
#> 643 501.0516   3.065419
#> 644 500.6471   2.925546
#> 645 500.1099   2.887873
#> 646 500.7248   2.646842
#> 647 500.0588   2.856143
#> 648 536.7571   2.699908
#> 649 500.0395   2.631373
#> 650 664.5764   3.242541
#> 651 590.2566   2.702539
#> 652 502.0757   3.188716
#> 653 500.1987   2.788267
#> 654 519.7212   4.024450
#> 655 506.7879   2.508237
#> 656 502.2389   3.080282
#> 657 517.3626   2.993065
#> 658 586.0277   2.642857
#> 659 500.0989   2.596313
#> 660 500.0621   3.212725
#> 661 508.0508   2.610814
#> 662 500.5524   2.777870
#> 663 501.2468   2.599397
#> 664 509.6035   3.004881
#> 665 525.6320   2.648686
#> 666 535.6249   2.751848
#> 667 516.8063   2.758204
#> 668 500.9781   2.544567
#> 669 506.1047   3.105736
#> 670 754.7283   2.764003
#> 671 503.7064   2.517402
#> 672 500.2373   2.534368
#> 673 528.2478   2.658438
#> 674 500.1405   3.148993
#> 675 518.0367   2.515271
#> 676 501.7635   2.633192
#> 677 545.8599   2.523906
#> 678 504.3299   3.393652
#> 679 500.7511   2.622978
#> 680 509.1611   2.699558
#> 681 534.0206   2.679254
#> 682 500.1312   3.170862
#> 683 853.4432   2.774307
#> 684 500.1489   2.578079
#> 685 523.6768   2.876329
#> 686 502.3241   2.602415
#> 687 577.4472   3.136136
#> 688 567.0872   2.546729
#> 689 744.9754   2.668827
#> 690 507.9320   3.472533
#> 691 678.6108   3.238362
#> 692 501.5491   3.466610
#> 693 508.9931   3.076194
#> 694 538.9740   2.880199
#> 695 501.5459   3.083079
#> 696 500.0352   2.821976
#> 697 500.0515   3.062867
#> 698 508.5947   2.811946
#> 699 500.0188   3.483926
#> 700 504.1106   2.981421
#> 701 500.0416   3.091378
#> 702 500.9593   3.375127
#> 703 500.9798   2.770185
#> 704 501.1378   2.537449
#> 705 924.5563   3.425389
#> 706 500.0710   2.627833
#> 707 546.2939   3.447812
#> 708 639.4806   2.532433
#> 709 506.6186   2.655283
#> 710 503.5397   3.785793
#> 711 504.9874   2.967203
#> 712 560.7945   2.545940
#> 713 520.4698   2.727849
#> 714 622.8540   2.826901
#> 715 504.5787   2.504482
#> 716 500.8212   3.101936
#> 717 522.4047   3.670818
#> 718 500.1855   2.527796
#> 719 509.9642   3.302372
#> 720 501.4399   3.164850
#> 721 500.7304   2.578168
#> 722 556.5171   2.625550
#> 723 651.7353   3.110609
#> 724 505.6407   2.892225
#> 725 517.2681   3.895330
#> 726 631.4975   2.980130
#> 727 501.0960   3.082606
#> 728 603.9464   3.216186
#> 729 500.0441   3.015376
#> 730 501.7511   3.265475
#> 731 504.0674   3.264487
#> 732 505.2488   3.140414
#> 733 500.9611   2.533548
#> 734 588.1050   3.914974
#> 735 503.7796   2.657335
#> 736 534.7556   2.550406
#> 737 502.5064   2.838038
#> 738 515.5299   3.315790
#> 739 502.2034   2.557009
#> 740 534.3214   2.649470
#> 741 501.6104   2.658201
#> 742 503.4384   2.790490
#> 743 543.4635   3.220254
#> 744 524.3612   2.922428
#> 745 500.3157   2.764311
#> 746 524.3444   2.540152
#> 747 514.6042   2.739356
#> 748 500.0322   3.005987
#> 749 502.5981   2.916202
#> 750 918.2913   3.107891
#> 751 501.8160   2.504355
#> 752 501.9907   3.705141
#> 753 503.2500   2.607655
#> 754 719.1785   2.715893
#> 755 500.0296   2.758771
#> 756 500.2782   2.562344
#> 757 500.0597   3.498676
#> 758 664.3229   2.775222
#> 759 533.2935   3.000501
#> 760 504.7466   3.015557
#> 761 534.1213   2.537053
#> 762 517.7610   2.548211
#> 763 503.8258   2.743320
#> 764 502.5561   2.913540
#> 765 504.9149   2.794055
#> 766 731.2398   2.526851
#> 767 500.0476   2.841903
#> 768 504.4101   2.577343
#> 769 500.0285   2.682245
#> 770 502.8746   3.327123
#> 771 509.8717   3.075664
#> 772 504.0716   2.986041
#> 773 523.4902   2.583553
#> 774 500.4123   3.367139
#> 775 502.3127   2.721212
#> 776 554.7074   2.730376
#> 777 557.4275   2.538644
#> 778 502.1922   2.845405
#> 779 500.2297   2.636789
#> 780 631.1832   2.648494
#> 781 504.3285   3.253310
#> 782 517.5254   2.767018
#> 783 818.9704   2.616627
#> 784 500.0262   2.895129
#> 785 500.0448   2.872696
#> 786 500.2377   2.507153
#> 787 531.8091   3.079527
#> 788 500.2466   3.353018
#> 789 501.2845   2.972080
#> 790 507.2330   3.539796
#> 791 512.3444   2.755763
#> 792 500.2711   2.565637
#> 793 501.3071   2.728718
#> 794 500.9253   2.991407
#> 795 501.3086   2.516817
#> 796 510.5263   3.011640
#> 797 500.3670   2.593151
#> 798 907.6037   2.541951
#> 799 512.7024   2.521985
#> 800 500.1078   3.091420
#> 801 500.1796   2.926833
#> 802 808.3307   3.408469
#> 803 655.7466   3.039052
#> 804 847.2532   2.910034
#> 805 872.1772   2.824254
#> 806 500.0129   2.745570
#> 807 505.5691   2.526651
#> 808 500.4321   2.514403
#> 809 500.6239   2.866423
#> 810 501.1099   2.797615
#> 811 501.5871   2.631321
#> 812 592.5498   3.151158
#> 813 518.9996   2.641186
#> 814 504.3859   2.553023
#> 815 894.0404   2.893012
#> 816 504.5855   3.118263
#> 817 519.9761   2.914273
#> 818 500.9184   3.838013
#> 819 500.6637   4.483022
#> 820 580.0860   2.913065
#> 821 500.0057   4.652204
#> 822 502.6493   2.673492
#> 823 619.9212   3.467568
#> 824 500.2372   3.290311
#> 825 500.0562   2.563676
#> 826 588.4509   2.597224
#> 827 500.8141   2.975700
#> 828 585.4034   2.653859
#> 829 633.7081   2.532938
#> 830 500.2853   3.437805
#> 831 902.3555   2.621662
#> 832 500.0578   4.433389
#> 833 501.5679   2.683211
#> 834 518.0087   3.086895
#> 835 525.8290   2.963071
#> 836 500.5191   2.504054
#> 837 502.1371   2.530771
#> 838 529.2221   3.828299
#> 839 500.0051   2.726192
#> 840 963.6056   2.607911
#> 841 500.0525   2.812881
#> 842 501.1239   3.271893
#> 843 526.8052   3.500696
#> 844 500.9613   2.534271
#> 845 500.0787   3.131208
#> 846 905.8187   3.320283
#> 847 500.2192   3.518547
#> 848 500.0563   2.672025
#> 849 520.4493   3.272579
#> 850 500.3989   2.597396
#> 851 500.0850   2.877466
#> 852 627.4105   3.698919
#> 853 508.2148   2.664270
#> 854 528.9106   3.169582
#> 855 500.2905   2.730075
#> 856 508.5677   3.039557
#> 857 500.0901   3.352188
#> 858 501.7296   2.524880
#> 859 502.4754   2.799562
#> 860 639.4147   2.881875
#> 861 501.3751   2.981078
#> 862 500.7985   4.869129
#> 863 881.7650   3.268078
#> 864 515.0824   3.048304
#> 865 500.1796   2.816293
#> 866 500.5373   2.646665
#> 867 500.2927   2.730916
#> 868 514.3316   2.645240
#> 869 506.0565   2.568867
#> 870 521.5085   3.798636
#> 871 509.3366   3.522519
#> 872 610.3500   2.920214
#> 873 506.1732   2.691778
#> 874 502.6916   3.617106
#> 875 500.4379   3.180091
#> 876 520.1328   2.575223
#> 877 500.7251   2.526568
#> 878 500.5390   2.896989
#> 879 503.7380   2.743765
#> 880 693.1424   3.849732
#> 881 507.0487   2.798492
#> 882 502.1570   3.251996
#> 883 501.8999   2.555594
#> 884 500.1387   2.864377
#> 885 558.0026   3.299096
#> 886 500.0533   2.539150
#> 887 500.5632   4.355307
#> 888 571.0110   2.598046
#> 889 555.3722   2.817628
#> 890 543.1773   2.542575
#> 891 854.8849   3.623033
#> 892 500.0337   2.661552
#> 893 733.6545   2.944078
#> 894 500.9966   2.537909
#> 895 652.5617   2.756606
#> 896 500.0232   3.141472
#> 897 603.9062   2.822482
#> 898 506.8127   2.657838
#> 899 617.6342   2.782936
#> 900 500.0102   2.723543
#> 901 500.5680   3.372455
#> 902 500.0360   3.435615
#> 903 607.4882   2.688913
#> 904 503.3642   2.548435
#> 905 784.4870   2.590532
#> 906 500.8375   3.340925
#> 907 500.0984   4.793979
#> 908 674.2099   2.678865
#> 909 501.7457   3.399954
#> 910 501.5168   2.703883
#> 911 500.1537   2.907914
#> 912 533.7871   2.736784
#> 913 526.0860   2.753470
#> 914 586.1974   3.207571
#> 915 503.9464   3.091888
#> 916 567.3594   2.766214
#> 917 530.0429   3.374684
#> 918 501.5006   2.935492
#> 919 506.7145   2.835900
#> 920 500.6311   2.545134
#> 921 500.8415   2.562510
#> 922 511.3159   2.664475
#> 923 500.2122   2.552744
#> 924 564.8421   2.720718
#> 925 500.5351   3.810642
#> 926 500.1098   2.681299
#> 927 500.4900   3.520237
#> 928 604.9284   2.863331
#> 929 500.1396   2.663925
#> 930 516.9393   2.537643
#> 931 658.4163   2.639163
#> 932 500.0037   3.944410
#> 933 523.7780   2.973697
#> 934 501.9131   3.164460
#> 935 503.2370   2.938161
#> 936 876.0437   2.831762
#> 937 500.1600   2.711797
#> 938 555.0054   3.385159
#> 939 763.5198   3.546814
#> 940 503.6986   2.853874
#> 941 500.0261   2.873934
#> 942 540.3695   3.351381
#> 943 613.7344   2.780542
#> 944 796.0563   3.098987
#> 945 961.6801   2.706142
#> 946 580.1485   2.698773
#> 947 502.4131   2.669366
#> 948 508.3678   2.750793