Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   500.5125   2.983035
#> 2   501.4853   2.500712
#> 3   502.7169   2.695222
#> 4   500.7982   2.748364
#> 5   548.9912   2.529714
#> 6   550.2886   3.009036
#> 7   799.6660   3.379864
#> 8   500.0622   2.813263
#> 9   500.5873   2.815440
#> 10  504.5627   2.653741
#> 11  502.8356   2.906894
#> 12  500.5294   2.588111
#> 13  503.8683   2.673044
#> 14  500.1178   3.233787
#> 15  546.8915   2.537346
#> 16  500.3346   2.677496
#> 17  500.3742   2.853744
#> 18  500.4089   2.636862
#> 19  501.2447   2.558370
#> 20  626.3292   2.903425
#> 21  503.6448   2.817533
#> 22  541.5256   3.070301
#> 23  503.1787   2.899953
#> 24  632.3192   2.998008
#> 25  517.6771   2.633742
#> 26  554.0038   2.623505
#> 27  501.5022   2.747047
#> 28  515.0616   2.707031
#> 29  500.0127   3.131769
#> 30  915.6705   2.601029
#> 31  577.9458   3.280537
#> 32  500.0211   2.977170
#> 33  509.9427   3.000371
#> 34  558.1831   2.980505
#> 35  539.3781   3.071029
#> 36  646.9274   2.805914
#> 37  501.1383   2.680063
#> 38  511.5079   2.657608
#> 39  500.2048   2.704821
#> 40  558.6459   3.082892
#> 41  871.3511   2.581780
#> 42  500.6890   2.880932
#> 43  727.3551   2.541140
#> 44  502.7392   2.772357
#> 45  500.0517   2.589088
#> 46  505.4543   3.269473
#> 47  500.1724   3.053140
#> 48  971.2314   2.779433
#> 49  607.0462   2.865268
#> 50  500.2813   2.927477
#> 51  578.1485   2.556697
#> 52  867.2984   2.680610
#> 53  500.1070   3.322726
#> 54  500.0250   3.020691
#> 55  530.4745   4.067198
#> 56  501.0284   3.508529
#> 57  500.2266   2.908121
#> 58  500.8218   2.560721
#> 59  511.4370   3.497408
#> 60  500.0068   2.676529
#> 61  541.0305   2.767125
#> 62  500.0555   2.585606
#> 63  501.9118   2.713311
#> 64  673.4488   3.252347
#> 65  500.0612   2.995552
#> 66  630.4690   4.377531
#> 67  500.6808   3.072757
#> 68  522.1958   2.716922
#> 69  533.7246   4.039090
#> 70  502.0306   2.690102
#> 71  500.2595   2.598049
#> 72  500.0141   2.690653
#> 73  500.2454   3.159170
#> 74  505.8063   2.884670
#> 75  516.2688   2.560398
#> 76  500.0575   2.677669
#> 77  525.3397   3.887930
#> 78  500.4579   3.125809
#> 79  502.4493   2.995165
#> 80  557.0730   2.655050
#> 81  500.6615   2.551596
#> 82  500.1929   2.797351
#> 83  535.7722   3.179843
#> 84  505.8398   2.868180
#> 85  508.2618   3.089268
#> 86  501.6246   2.515476
#> 87  500.1123   3.543887
#> 88  501.1127   2.751550
#> 89  500.3201   2.553335
#> 90  501.0988   2.762983
#> 91  933.4328   2.748211
#> 92  500.7023   3.550542
#> 93  500.0598   2.810477
#> 94  506.5470   2.527899
#> 95  503.9173   2.589434
#> 96  500.0148   3.284076
#> 97  737.2730   3.221854
#> 98  500.2978   2.518327
#> 99  512.3349   4.753291
#> 100 500.4192   3.225676
#> 101 502.7168   4.493233
#> 102 503.2269   2.647913
#> 103 611.0807   2.562353
#> 104 501.2672   2.631731
#> 105 500.7409   4.793283
#> 106 506.6856   2.786794
#> 107 501.3080   2.533376
#> 108 507.9903   2.642690
#> 109 512.2346   2.687008
#> 110 500.0054   2.531205
#> 111 500.0571   2.859156
#> 112 858.7432   2.975540
#> 113 503.7172   2.555693
#> 114 529.1293   2.733291
#> 115 895.4384   2.915311
#> 116 502.3808   3.098947
#> 117 627.2531   3.529400
#> 118 502.7246   3.020476
#> 119 500.0780   3.581368
#> 120 502.2737   2.801313
#> 121 501.1870   2.623250
#> 122 550.9072   2.802869
#> 123 505.9341   2.684791
#> 124 500.6373   3.453485
#> 125 500.0602   3.155892
#> 126 500.0335   2.579295
#> 127 973.4228   4.136083
#> 128 516.4402   2.656081
#> 129 502.5098   3.211724
#> 130 517.5822   3.755978
#> 131 876.8586   2.509475
#> 132 501.5934   2.572652
#> 133 500.1417   3.498850
#> 134 500.1204   2.641783
#> 135 632.4951   2.609246
#> 136 500.6736   2.648042
#> 137 500.1716   2.805923
#> 138 503.3373   2.988583
#> 139 574.8024   3.253486
#> 140 668.3422   4.311072
#> 141 513.5769   2.766597
#> 142 597.2184   2.925932
#> 143 500.4762   3.203652
#> 144 548.6917   2.586197
#> 145 500.7622   2.671950
#> 146 506.3725   2.779053
#> 147 500.0964   3.035920
#> 148 507.7840   2.735892
#> 149 613.6340   2.723765
#> 150 503.5591   3.090693
#> 151 732.6157   2.986500
#> 152 500.1039   3.108380
#> 153 503.3988   2.859685
#> 154 500.4229   2.671201
#> 155 500.2032   2.693058
#> 156 976.1168   2.519803
#> 157 501.6260   2.667462
#> 158 500.0496   3.351679
#> 159 501.6767   3.430193
#> 160 500.1866   2.575762
#> 161 500.1904   2.633427
#> 162 520.8391   2.715753
#> 163 566.8032   2.814625
#> 164 500.9310   2.853758
#> 165 502.7406   2.771901
#> 166 500.0571   2.720995
#> 167 522.3139   2.979313
#> 168 501.7036   3.265787
#> 169 868.5000   2.883278
#> 170 500.8922   2.891489
#> 171 547.4512   2.586434
#> 172 500.8860   2.574847
#> 173 500.3284   2.781227
#> 174 502.8220   3.323568
#> 175 508.0046   2.592977
#> 176 551.0151   2.896796
#> 177 780.0601   2.917065
#> 178 502.1392   3.143302
#> 179 500.2672   3.465770
#> 180 519.9177   2.653771
#> 181 707.1013   2.752269
#> 182 501.2679   3.270025
#> 183 505.1445   2.636563
#> 184 608.0245   3.057968
#> 185 500.0248   2.662689
#> 186 612.2438   2.716744
#> 187 504.7457   2.939597
#> 188 503.8157   3.045630
#> 189 500.0783   2.813235
#> 190 517.9124   2.743915
#> 191 923.4611   3.389158
#> 192 501.1791   2.604030
#> 193 500.2134   2.668179
#> 194 505.5414   2.839404
#> 195 567.1872   3.166091
#> 196 890.3908   2.714142
#> 197 888.2715   2.502157
#> 198 510.7018   2.979174
#> 199 591.5598   2.804599
#> 200 505.1068   2.773689
#> 201 506.9265   3.023561
#> 202 501.0342   3.021376
#> 203 500.2391   2.973074
#> 204 501.0323   3.303928
#> 205 533.0918   2.633192
#> 206 534.5426   2.644720
#> 207 784.5069   2.564446
#> 208 501.2711   2.701608
#> 209 674.9214   2.947946
#> 210 500.5412   3.965441
#> 211 592.1025   2.593686
#> 212 500.4545   2.590868
#> 213 500.2701   3.063995
#> 214 500.2570   2.533036
#> 215 540.5790   2.503112
#> 216 501.6979   2.511692
#> 217 500.2570   2.620529
#> 218 501.1641   2.825234
#> 219 500.1135   3.062151
#> 220 500.2214   2.537342
#> 221 500.0549   3.415601
#> 222 609.4923   2.563743
#> 223 546.5309   2.816646
#> 224 502.0918   2.513789
#> 225 500.9167   2.520935
#> 226 500.1216   2.712684
#> 227 502.8112   3.429069
#> 228 504.1183   2.836371
#> 229 501.3991   3.316069
#> 230 559.6050   2.563299
#> 231 859.4553   2.754438
#> 232 503.3971   3.141403
#> 233 500.7809   3.036310
#> 234 500.1848   2.973531
#> 235 500.0742   2.608234
#> 236 507.8249   3.007712
#> 237 500.6175   3.174409
#> 238 500.0903   2.796707
#> 239 570.6243   2.708009
#> 240 500.0306   2.860510
#> 241 501.6301   3.173912
#> 242 500.9603   2.825512
#> 243 500.2254   2.988488
#> 244 501.0410   3.480708
#> 245 502.1403   2.700051
#> 246 500.4819   2.649844
#> 247 508.3093   2.952144
#> 248 504.4065   2.538535
#> 249 503.3404   4.180306
#> 250 514.2847   3.020535
#> 251 500.4526   3.727653
#> 252 666.5770   2.990255
#> 253 500.2582   2.725507
#> 254 501.0563   2.779719
#> 255 551.3397   3.707371
#> 256 505.9198   2.508470
#> 257 513.9812   2.502359
#> 258 683.3111   2.545808
#> 259 576.7765   2.853455
#> 260 500.0402   2.509584
#> 261 558.0429   2.665121
#> 262 532.3538   2.833872
#> 263 502.0129   2.698788
#> 264 500.1595   2.918725
#> 265 505.6398   2.969611
#> 266 552.1889   3.157179
#> 267 545.7123   3.806073
#> 268 500.9061   2.902957
#> 269 500.0425   3.445517
#> 270 500.1249   2.996171
#> 271 525.0408   3.492181
#> 272 501.1788   2.726074
#> 273 502.8262   2.542303
#> 274 504.9069   2.753336
#> 275 502.6298   3.387763
#> 276 720.1703   2.821027
#> 277 502.2915   3.261138
#> 278 760.1321   2.684050
#> 279 501.0758   2.811762
#> 280 503.9406   2.844761
#> 281 502.4629   3.113185
#> 282 501.0683   2.599805
#> 283 500.0275   2.507448
#> 284 628.5905   2.692292
#> 285 624.1553   2.615087
#> 286 623.7164   2.516300
#> 287 500.6324   2.585615
#> 288 501.3981   2.711281
#> 289 515.7540   4.297169
#> 290 500.0957   2.529016
#> 291 520.4864   2.650671
#> 292 501.3102   3.508617
#> 293 718.6519   2.849753
#> 294 500.0961   3.569672
#> 295 500.4688   3.787201
#> 296 504.5147   2.838760
#> 297 507.4933   3.045033
#> 298 554.2485   2.815432
#> 299 530.2197   2.584886
#> 300 517.6162   2.752622
#> 301 518.5257   2.608862
#> 302 500.1651   2.760184
#> 303 502.1732   2.510979
#> 304 520.9002   2.744652
#> 305 501.4290   2.564838
#> 306 521.6193   3.530087
#> 307 501.6422   2.601875
#> 308 500.0877   2.553632
#> 309 531.4035   3.460237
#> 310 604.1897   2.644500
#> 311 500.0790   2.841257
#> 312 500.4714   2.913509
#> 313 500.1420   2.722385
#> 314 501.1307   2.693901
#> 315 501.0127   3.058845
#> 316 501.8894   3.482987
#> 317 500.2517   2.602414
#> 318 501.5655   2.612072
#> 319 501.0973   2.667608
#> 320 500.0928   3.066005
#> 321 501.2627   2.866450
#> 322 504.3286   2.767708
#> 323 569.3075   3.039274
#> 324 502.9553   2.636988
#> 325 500.1218   2.655091
#> 326 514.3316   3.485126
#> 327 545.4319   2.746217
#> 328 509.1265   3.051021
#> 329 568.1068   3.330550
#> 330 500.1386   3.026731
#> 331 502.6478   2.660161
#> 332 500.8995   2.541107
#> 333 506.0151   3.162606
#> 334 509.7195   3.040901
#> 335 501.9573   3.658622
#> 336 682.4919   3.691594
#> 337 500.4409   2.519327
#> 338 501.5328   3.150671
#> 339 595.1607   3.065656
#> 340 509.3435   3.503776
#> 341 502.7755   2.702723
#> 342 563.2913   2.590582
#> 343 677.4197   3.196469
#> 344 502.7982   3.491928
#> 345 500.0092   2.538240
#> 346 567.6060   2.916507
#> 347 501.1513   2.572677
#> 348 500.0122   3.023794
#> 349 659.1096   2.616516
#> 350 501.0038   2.989052
#> 351 500.8519   3.231025
#> 352 781.0079   3.278324
#> 353 570.6705   2.901113
#> 354 500.2740   3.280638
#> 355 502.7496   2.767877
#> 356 500.0925   2.647772
#> 357 658.9670   2.748758
#> 358 500.2199   2.599895
#> 359 501.2529   2.817131
#> 360 507.3768   2.545975
#> 361 500.4914   2.579900
#> 362 532.5600   2.538479
#> 363 500.2651   3.133937
#> 364 502.9844   2.723627
#> 365 501.2707   4.285273
#> 366 535.9332   3.330405
#> 367 500.4403   2.656719
#> 368 504.2474   2.641804
#> 369 500.0888   2.633291
#> 370 532.8877   2.710885
#> 371 504.4447   3.444071
#> 372 500.1168   2.532709
#> 373 500.2702   2.554794
#> 374 505.8338   2.630437
#> 375 500.0789   2.521185
#> 376 621.2842   2.746345
#> 377 500.2961   2.779111
#> 378 588.3158   2.892219
#> 379 543.8976   2.503941
#> 380 500.2706   2.931340
#> 381 500.3907   2.700041
#> 382 500.2847   2.628397
#> 383 604.1644   2.528312
#> 384 500.4017   2.897426
#> 385 530.8640   4.197905
#> 386 500.0350   2.836201
#> 387 504.8801   2.721385
#> 388 500.2703   2.885381
#> 389 508.1888   2.598458
#> 390 501.3964   3.334347
#> 391 540.2406   2.830884
#> 392 619.9088   2.989928
#> 393 513.3367   3.153466
#> 394 580.9105   3.339668
#> 395 526.7962   2.602193
#> 396 588.8629   2.566229
#> 397 501.2128   2.758138
#> 398 513.9446   2.739674
#> 399 505.8729   2.608972
#> 400 514.3591   3.001366
#> 401 500.2389   3.328905
#> 402 515.5012   3.014518
#> 403 968.2117   3.686866
#> 404 501.7952   3.095157
#> 405 707.5547   2.686232
#> 406 500.0378   2.975222
#> 407 500.3592   3.171486
#> 408 500.6516   2.538609
#> 409 507.4845   3.826468
#> 410 588.3412   2.639406
#> 411 500.5290   2.893517
#> 412 709.2224   3.206887
#> 413 529.2756   2.514356
#> 414 503.0571   2.796998
#> 415 501.1313   2.612999
#> 416 500.7049   4.654299
#> 417 636.5280   2.742719
#> 418 984.1959   3.183183
#> 419 510.6033   2.718937
#> 420 508.8120   2.990521
#> 421 500.2015   2.517480
#> 422 500.3866   4.115333
#> 423 504.4098   2.748348
#> 424 507.6562   3.244641
#> 425 952.5260   3.805949
#> 426 508.6032   2.861333
#> 427 500.0202   2.784371
#> 428 504.1592   2.524609
#> 429 590.1198   3.295994
#> 430 541.5263   2.574575
#> 431 502.2534   3.426728
#> 432 503.2341   3.197394
#> 433 501.8555   2.521305
#> 434 521.9681   2.528591
#> 435 517.5809   2.717060
#> 436 508.1069   2.903928
#> 437 500.3059   3.103333
#> 438 501.0443   3.442304
#> 439 501.8956   2.764547
#> 440 501.0875   3.209824
#> 441 505.6213   3.286321
#> 442 500.5841   2.526658
#> 443 500.5045   3.509479
#> 444 514.6050   2.585449
#> 445 508.7954   2.634568
#> 446 509.0819   3.190261
#> 447 500.7210   5.481539
#> 448 503.0481   2.604810
#> 449 522.3253   2.517197
#> 450 608.1579   2.572838
#> 451 500.0760   3.222734
#> 452 862.0207   2.538087
#> 453 500.5548   2.796715
#> 454 966.6085   3.524582
#> 455 521.9117   2.635926
#> 456 516.1699   3.399067
#> 457 500.0095   2.921751
#> 458 500.0264   2.991390
#> 459 500.5284   3.200373
#> 460 614.3877   2.661088
#> 461 505.0830   2.987424
#> 462 690.4305   2.845673
#> 463 514.5393   2.648297
#> 464 694.9623   3.157457
#> 465 504.0750   2.642844
#> 466 515.8881   3.166288
#> 467 716.1489   2.905892
#> 468 971.1090   2.524033
#> 469 500.0512   3.200995
#> 470 500.0823   3.250205
#> 471 500.0631   3.051374
#> 472 502.0044   2.572086
#> 473 506.3509   3.577059
#> 474 503.7558   2.546685
#> 475 524.1051   2.824565
#> 476 840.3011   2.964182
#> 477 524.4421   2.541774
#> 478 511.2512   2.835575
#> 479 500.4121   2.824949
#> 480 500.2232   3.413730
#> 481 648.3957   3.145947
#> 482 501.3331   3.053385
#> 483 500.5077   2.711560
#> 484 500.7305   5.565617
#> 485 503.8757   3.073043
#> 486 511.1155   3.342836
#> 487 500.4358   2.532653
#> 488 500.1002   2.933448
#> 489 573.1984   2.573387
#> 490 501.8396   2.576448
#> 491 504.6937   2.514131
#> 492 699.6812   2.549895
#> 493 500.4045   2.614381
#> 494 500.1405   2.799260
#> 495 596.8847   4.143322
#> 496 544.1597   2.571267
#> 497 505.8829   2.646287
#> 498 507.6372   2.662342
#> 499 516.8886   2.658929
#> 500 512.7028   3.581678
#> 501 502.1782   2.974232
#> 502 966.9054   2.931962
#> 503 500.7048   3.386402
#> 504 500.4374   2.740040
#> 505 501.3255   2.559301
#> 506 702.6608   2.525936
#> 507 500.7047   4.132988
#> 508 514.6885   2.662591
#> 509 537.5921   2.971113
#> 510 502.7057   3.709412
#> 511 501.2898   2.861133
#> 512 514.2069   2.573283
#> 513 500.2222   3.092056
#> 514 515.2017   2.523029
#> 515 531.1815   2.553438
#> 516 500.1870   2.700717
#> 517 500.2319   2.933612
#> 518 500.1029   2.792688
#> 519 500.9155   2.737155
#> 520 517.3914   3.241562
#> 521 506.3168   2.681289
#> 522 515.4182   2.611919
#> 523 500.9732   2.717212
#> 524 501.4401   2.661162
#> 525 508.1761   2.623884
#> 526 500.2394   3.054625
#> 527 501.7580   3.045823
#> 528 500.1502   2.841891
#> 529 501.3427   2.633498
#> 530 526.1297   3.196268
#> 531 500.2054   2.616693
#> 532 500.0006   2.825616
#> 533 500.0041   3.223493
#> 534 588.1099   2.992998
#> 535 501.8974   2.590847
#> 536 500.1533   2.857885
#> 537 514.4000   2.978096
#> 538 516.2472   2.718589
#> 539 503.5740   4.003261
#> 540 507.4230   2.724801
#> 541 508.5355   2.637317
#> 542 591.1050   3.021583
#> 543 500.8078   2.898033
#> 544 503.0440   2.897964
#> 545 502.4260   2.844420
#> 546 502.2595   3.566268
#> 547 583.3051   2.649798
#> 548 949.5680   3.065174
#> 549 542.6639   2.582204
#> 550 674.5702   2.595194
#> 551 501.1553   2.751697
#> 552 525.5726   2.532241
#> 553 500.1381   2.629204
#> 554 560.7507   2.601941
#> 555 502.2942   2.712948
#> 556 501.1520   3.219877
#> 557 500.3422   3.177213
#> 558 500.3343   2.549881
#> 559 506.5315   2.556743
#> 560 589.4957   2.983456
#> 561 502.7269   3.545745
#> 562 543.4971   3.093612
#> 563 500.2850   3.436167
#> 564 670.5113   2.608895
#> 565 500.4320   3.326603
#> 566 500.0130   2.844037
#> 567 967.8810   3.045240
#> 568 500.7313   2.505913
#> 569 500.3710   2.553861
#> 570 501.6436   2.525308
#> 571 860.5101   3.752707
#> 572 510.2074   2.879066
#> 573 501.1252   2.568183
#> 574 775.9249   4.384156
#> 575 716.5862   2.870799
#> 576 533.6842   3.744730
#> 577 500.6519   3.268339
#> 578 500.0006   2.753820
#> 579 808.0098   2.594334
#> 580 501.2737   2.652394
#> 581 500.1491   2.602256
#> 582 753.7766   3.025322
#> 583 505.2249   2.676269
#> 584 500.3524   3.238584
#> 585 506.3571   3.502606
#> 586 501.0448   3.184506
#> 587 957.3190   2.892705
#> 588 505.8992   3.038414
#> 589 500.0137   2.868920
#> 590 500.0913   2.581247
#> 591 539.3634   2.670716
#> 592 500.8171   3.613688
#> 593 503.3453   2.789212
#> 594 500.0717   2.753257
#> 595 502.6036   3.549032
#> 596 503.1358   2.920225
#> 597 501.0966   2.841884
#> 598 500.0175   2.588315
#> 599 594.1929   2.676530
#> 600 512.4768   2.704948
#> 601 500.0142   2.847673
#> 602 500.1234   3.163954
#> 603 501.5617   3.123337
#> 604 500.0102   3.075924
#> 605 502.0483   3.140342
#> 606 643.2921   2.562814
#> 607 505.3393   3.385994
#> 608 939.4704   2.634736
#> 609 510.6247   2.596982
#> 610 500.0310   2.794467
#> 611 501.6592   2.856822
#> 612 580.5308   2.614137
#> 613 503.7549   3.415172
#> 614 500.1511   2.783269
#> 615 501.5280   2.713144
#> 616 531.1490   3.002978
#> 617 603.3996   2.819804
#> 618 501.0356   3.030825
#> 619 504.8781   2.533174
#> 620 500.0272   3.389404
#> 621 508.0499   2.790668
#> 622 578.2022   2.556351
#> 623 681.1637   3.403093
#> 624 502.6064   3.702995
#> 625 605.3122   2.505179
#> 626 500.1067   3.228933
#> 627 500.0081   2.899266
#> 628 533.1555   2.904328
#> 629 503.7954   2.879735
#> 630 508.8576   2.523294
#> 631 625.3261   4.929247
#> 632 978.2660   2.973000
#> 633 690.0219   2.589338
#> 634 610.9145   3.115555
#> 635 561.1171   2.670877
#> 636 501.7525   2.516552
#> 637 504.3125   3.376268
#> 638 514.3610   2.627005
#> 639 500.2343   3.363937
#> 640 501.8846   2.968418
#> 641 500.4085   2.538665
#> 642 504.3360   2.524539
#> 643 688.8599   2.724395
#> 644 681.4743   2.872799
#> 645 500.8532   3.571229
#> 646 504.6417   2.632515
#> 647 568.3027   3.884004
#> 648 501.1242   2.846944
#> 649 578.9807   2.578885
#> 650 649.1276   2.683458
#> 651 504.2564   2.694889
#> 652 501.1410   2.565344
#> 653 503.8322   2.752635
#> 654 500.7210   2.572399
#> 655 508.9053   3.342127
#> 656 504.9217   2.679570
#> 657 529.2388   2.632440
#> 658 501.5983   2.688785
#> 659 500.1840   3.139073
#> 660 501.7440   2.679419
#> 661 502.0770   2.839477
#> 662 519.6268   3.077168
#> 663 500.5784   3.304429
#> 664 500.8101   2.816399
#> 665 504.3425   2.546702
#> 666 500.2595   2.751692
#> 667 674.5009   2.963638
#> 668 506.8844   2.668487
#> 669 514.3450   3.302261
#> 670 616.2245   2.560848
#> 671 591.3139   2.606910
#> 672 939.7696   4.454509
#> 673 500.6346   3.276571
#> 674 500.5284   2.570827
#> 675 500.1097   3.164894
#> 676 624.2805   2.792542
#> 677 570.9284   2.917034
#> 678 504.2237   4.051431
#> 679 532.8274   2.566525
#> 680 508.2267   3.598780
#> 681 501.2001   2.844643
#> 682 501.3900   2.605276
#> 683 504.3549   2.892957
#> 684 500.3360   2.690274
#> 685 506.9796   2.956510
#> 686 504.7152   2.957332
#> 687 685.3055   4.943013
#> 688 500.6455   2.613626
#> 689 599.5031   2.936479
#> 690 500.6471   3.100624
#> 691 761.2002   3.422942
#> 692 500.4792   3.381646
#> 693 501.1035   3.057455
#> 694 977.7339   2.597829
#> 695 511.2749   2.894928
#> 696 767.5132   3.044895
#> 697 500.9371   3.387702
#> 698 500.1780   2.760975
#> 699 500.0423   2.754259
#> 700 503.2334   2.658766
#> 701 722.5085   2.673002
#> 702 500.3142   2.936924
#> 703 533.0689   2.538186
#> 704 503.1196   2.790980
#> 705 501.8737   2.802908
#> 706 500.1208   2.746134
#> 707 500.1472   3.160365
#> 708 581.1363   2.504592
#> 709 565.2097   3.316319
#> 710 501.8916   2.896153
#> 711 611.6136   2.639892
#> 712 866.2399   2.593157
#> 713 500.7501   2.929227
#> 714 503.9775   3.319302
#> 715 996.7076   3.103863
#> 716 693.0761   3.284034
#> 717 965.2037   2.585830
#> 718 500.9123   2.504107
#> 719 500.4532   3.252283
#> 720 500.6170   2.917841
#> 721 514.9006   3.885144
#> 722 502.9467   2.724205
#> 723 683.2944   2.824470
#> 724 510.1808   2.879732
#> 725 736.6314   2.540849
#> 726 500.4573   2.942298
#> 727 500.0557   2.655805
#> 728 500.1635   2.546002
#> 729 816.7920   2.642375
#> 730 833.8261   2.508690
#> 731 503.9490   3.307650
#> 732 596.2853   3.290286
#> 733 500.8552   3.347017
#> 734 917.9144   2.513277
#> 735 531.2215   2.911877
#> 736 504.5017   3.998498
#> 737 983.3883   2.807578
#> 738 500.1447   2.836284
#> 739 520.0421   2.514158
#> 740 500.1337   2.771117
#> 741 500.7442   2.527526
#> 742 534.7460   2.710782
#> 743 519.6147   2.998565
#> 744 533.4744   2.692981
#> 745 500.9791   3.328020
#> 746 500.1977   2.931895
#> 747 501.3819   3.785449
#> 748 540.8999   3.342239
#> 749 500.2567   2.558541
#> 750 504.7884   3.265436
#> 751 501.2220   2.854743
#> 752 608.9113   2.619764
#> 753 752.8747   2.814629
#> 754 503.6969   3.159448
#> 755 510.6367   2.595959
#> 756 500.3900   2.646259
#> 757 502.4162   2.668851
#> 758 614.8562   2.747545
#> 759 500.1655   2.697496
#> 760 512.8803   2.530442
#> 761 500.1239   2.619435
#> 762 548.4232   2.604498
#> 763 500.9060   3.157643
#> 764 505.9069   3.078652
#> 765 564.8471   2.912722
#> 766 506.2908   3.544844
#> 767 833.7420   2.894678
#> 768 737.2277   2.776007
#> 769 539.0638   2.600719
#> 770 502.3915   2.517064
#> 771 581.6031   2.949520
#> 772 500.3612   2.588469
#> 773 532.1989   2.700807
#> 774 522.0530   3.531325
#> 775 531.8174   2.575122
#> 776 503.0431   2.745387
#> 777 500.9028   2.602288
#> 778 553.0505   2.507315
#> 779 550.3474   2.557959
#> 780 938.0162   2.623061
#> 781 500.2164   2.654229
#> 782 500.0327   5.556853
#> 783 500.5905   2.887127
#> 784 539.2318   3.195113
#> 785 500.0022   3.255260
#> 786 500.0097   2.902448
#> 787 512.0591   2.817144
#> 788 529.9720   2.775808
#> 789 500.1597   3.307069
#> 790 517.6727   3.855995
#> 791 517.7945   6.397422
#> 792 501.5097   2.823333
#> 793 508.6645   2.745514
#> 794 508.5626   2.509216
#> 795 500.9133   2.630001
#> 796 500.0358   2.901499
#> 797 501.7900   2.965468
#> 798 500.9594   2.501714
#> 799 894.8328   3.028040
#> 800 505.3270   2.817675
#> 801 500.0815   2.972173
#> 802 650.2617   3.343763
#> 803 502.5688   2.892676
#> 804 720.1668   2.607689
#> 805 563.5079   3.793118
#> 806 516.6113   3.467362
#> 807 501.3855   3.154793
#> 808 500.2915   2.663457
#> 809 631.0581   2.892915
#> 810 516.7166   3.177592
#> 811 504.6414   2.884541
#> 812 500.6578   2.604767
#> 813 500.0246   2.556406
#> 814 806.3307   2.718241
#> 815 500.1386   2.567970
#> 816 504.1508   2.722398
#> 817 500.7456   3.010380
#> 818 735.5055   2.938435
#> 819 518.9282   2.614241
#> 820 707.2263   2.662859
#> 821 514.2754   3.224940
#> 822 500.0873   2.832653
#> 823 500.6588   2.705871
#> 824 500.6402   3.207551
#> 825 500.0613   2.521510
#> 826 504.1006   2.717278
#> 827 502.2516   2.836631
#> 828 500.3529   2.692425
#> 829 500.0208   3.294622
#> 830 500.3415   3.078058
#> 831 517.6377   2.590407
#> 832 531.0889   2.503022
#> 833 500.0267   3.854136
#> 834 509.7581   2.902201
#> 835 533.9978   2.572361
#> 836 513.4057   2.949261
#> 837 500.6221   2.645340
#> 838 503.8568   2.687819
#> 839 500.1078   2.864896
#> 840 503.6907   2.520045
#> 841 549.7868   3.297312
#> 842 501.7997   2.792055
#> 843 500.0779   3.012656
#> 844 501.3190   2.861086
#> 845 500.0039   3.064218
#> 846 510.1549   2.562403
#> 847 504.2267   2.523521
#> 848 688.6992   2.619568
#> 849 503.0335   2.849710
#> 850 508.4555   2.684954
#> 851 500.0713   3.553388
#> 852 503.3063   2.933045
#> 853 501.4357   2.885117
#> 854 526.5778   2.768379
#> 855 509.1838   2.772993
#> 856 507.1213   2.747748
#> 857 509.5149   2.535504
#> 858 500.3160   2.621366
#> 859 500.0333   4.144640
#> 860 559.5123   2.870805
#> 861 914.2365   2.564572
#> 862 514.2604   2.644929
#> 863 502.1620   2.910329
#> 864 520.5091   2.925887
#> 865 502.4272   2.711175
#> 866 504.0228   2.759550
#> 867 500.0222   2.923174
#> 868 510.8050   2.769636
#> 869 506.3131   2.677149
#> 870 520.3001   3.002619
#> 871 501.7160   3.990695
#> 872 515.8748   2.606145
#> 873 501.2844   2.500062
#> 874 534.7545   2.683669
#> 875 503.2009   2.787643
#> 876 523.5394   2.951246
#> 877 500.0196   2.903731
#> 878 978.8850   2.543771
#> 879 500.4425   2.527260
#> 880 502.2334   2.754000
#> 881 510.0310   2.969139
#> 882 600.2873   2.964972
#> 883 981.5968   3.503967
#> 884 500.1958   2.580196
#> 885 590.9354   3.351733
#> 886 503.7673   3.176505
#> 887 500.8903   3.038975
#> 888 691.9431   3.078293
#> 889 500.1079   2.540647
#> 890 506.0885   2.611692
#> 891 508.5840   2.621436
#> 892 500.0991   4.407322
#> 893 620.7619   2.670815
#> 894 505.6286   2.720153
#> 895 502.9867   2.768105
#> 896 786.3926   2.615524
#> 897 506.6017   3.657049
#> 898 522.2408   2.663746
#> 899 500.4639   2.712590
#> 900 694.0169   4.709080
#> 901 509.3404   3.032827
#> 902 501.0339   2.599007
#> 903 502.8319   2.609298
#> 904 668.1838   2.779605
#> 905 502.5515   3.220992
#> 906 500.0169   2.858815
#> 907 516.6260   3.844570
#> 908 500.2162   2.911247
#> 909 500.0183   2.787826
#> 910 502.5647   2.672592
#> 911 500.1162   2.715948
#> 912 684.8416   3.231857
#> 913 520.0285   2.613544
#> 914 500.1162   3.528191
#> 915 557.1609   2.971758
#> 916 501.6467   2.939428
#> 917 501.2382   4.338195
#> 918 500.5681   3.119566
#> 919 525.5246   3.009810
#> 920 590.0326   2.702471
#> 921 500.3549   2.861727
#> 922 505.8388   2.727789
#> 923 676.9616   2.500737
#> 924 500.3061   2.552670
#> 925 546.0615   2.807237
#> 926 505.8149   2.523395
#> 927 501.5072   2.879807