Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   505.1982   2.680356
#> 2   542.2974   3.996927
#> 3   534.0852   2.828456
#> 4   510.8334   2.561829
#> 5   500.0572   2.684984
#> 6   504.5649   3.179061
#> 7   500.0988   2.893970
#> 8   737.5871   3.025216
#> 9   500.2036   2.577405
#> 10  506.7509   3.013859
#> 11  500.1012   3.063674
#> 12  500.7325   2.796748
#> 13  520.7911   2.565698
#> 14  506.7735   2.976560
#> 15  523.0100   2.866393
#> 16  628.2496   2.612134
#> 17  523.9099   2.535061
#> 18  500.4762   2.649734
#> 19  509.7343   2.534492
#> 20  515.6732   2.765852
#> 21  500.1248   2.660579
#> 22  500.4804   4.318610
#> 23  500.2948   2.620744
#> 24  500.2248   2.558336
#> 25  500.3027   2.861923
#> 26  500.6785   2.855739
#> 27  514.7129   2.630079
#> 28  500.2424   2.867423
#> 29  501.3239   2.824464
#> 30  500.2336   2.518529
#> 31  500.0211   2.563888
#> 32  535.9818   2.573194
#> 33  520.6246   2.660459
#> 34  502.6493   2.657242
#> 35  504.8999   2.929953
#> 36  900.3106   2.731808
#> 37  509.8071   3.154331
#> 38  502.2250   4.285756
#> 39  500.1037   2.532495
#> 40  508.7392   2.601937
#> 41  500.3751   3.065814
#> 42  708.1996   2.517386
#> 43  526.4370   2.948507
#> 44  500.0739   2.970320
#> 45  520.7766   3.039197
#> 46  500.6157   2.514507
#> 47  656.8266   2.572473
#> 48  525.0063   2.892695
#> 49  862.2636   2.682288
#> 50  500.1231   3.004780
#> 51  510.2133   4.414585
#> 52  511.7687   2.914501
#> 53  507.7466   2.842260
#> 54  532.7852   2.515329
#> 55  950.1788   2.550227
#> 56  500.1630   2.556033
#> 57  503.7050   2.524293
#> 58  612.7940   2.504824
#> 59  503.5190   2.731596
#> 60  500.0451   3.083486
#> 61  659.6580   2.787325
#> 62  500.2737   2.719716
#> 63  500.0258   3.743588
#> 64  500.0654   2.510421
#> 65  500.6615   2.548238
#> 66  784.2591   4.339113
#> 67  716.2828   2.530293
#> 68  500.7497   3.607745
#> 69  613.0134   2.676396
#> 70  513.2543   2.609083
#> 71  500.0040   5.129058
#> 72  517.3983   3.121939
#> 73  520.1479   2.618554
#> 74  501.1743   2.731538
#> 75  511.5701   2.782118
#> 76  501.5530   2.828389
#> 77  518.7042   2.580591
#> 78  502.9897   2.859054
#> 79  507.4267   2.841375
#> 80  500.9667   2.529147
#> 81  502.4311   3.012343
#> 82  500.0908   2.635086
#> 83  500.0321   2.572263
#> 84  503.0180   2.554624
#> 85  500.1028   3.394754
#> 86  503.9943   2.940543
#> 87  500.2632   2.551907
#> 88  507.7927   2.626271
#> 89  517.1877   2.661104
#> 90  505.5156   2.559839
#> 91  500.0127   2.889208
#> 92  523.6728   2.508367
#> 93  500.5147   2.696627
#> 94  818.4735   2.592796
#> 95  500.0257   3.104553
#> 96  508.9203   2.503931
#> 97  501.4585   2.906827
#> 98  500.8286   2.772245
#> 99  880.3414   2.503068
#> 100 506.3534   2.856551
#> 101 500.0555   2.898925
#> 102 576.7361   3.220518
#> 103 507.9120   2.639382
#> 104 500.3314   5.534646
#> 105 570.8934   2.573372
#> 106 736.1915   5.240890
#> 107 501.2450   3.084447
#> 108 505.6698   2.953249
#> 109 500.2348   2.657101
#> 110 896.4033   2.520926
#> 111 500.4256   2.723646
#> 112 529.4302   2.900908
#> 113 504.0882   3.322141
#> 114 511.4957   2.580203
#> 115 501.3365   2.696642
#> 116 526.0285   2.548982
#> 117 506.2137   3.062686
#> 118 507.9045   3.012994
#> 119 500.5719   3.186983
#> 120 503.0510   2.518726
#> 121 500.4870   2.539078
#> 122 502.3099   2.609855
#> 123 531.2529   2.684765
#> 124 517.1336   3.671150
#> 125 501.3730   2.840055
#> 126 699.4616   2.568095
#> 127 500.5674   3.025889
#> 128 500.0075   2.896690
#> 129 603.5190   3.485453
#> 130 500.4443   2.941493
#> 131 900.4043   2.719551
#> 132 523.2558   2.952121
#> 133 500.0316   2.972318
#> 134 531.1596   3.603935
#> 135 967.6366   3.258839
#> 136 504.2773   2.796322
#> 137 503.4767   2.680730
#> 138 642.4724   2.550523
#> 139 589.8032   2.687628
#> 140 502.0438   2.741497
#> 141 718.3530   2.745745
#> 142 500.1890   4.437273
#> 143 501.2586   3.190657
#> 144 503.9160   3.089591
#> 145 562.4939   3.074329
#> 146 502.1845   2.696364
#> 147 500.0768   2.695441
#> 148 500.0539   2.547409
#> 149 656.1647   2.656424
#> 150 521.4890   2.616387
#> 151 506.0109   2.898402
#> 152 508.2799   2.639834
#> 153 500.2502   3.191278
#> 154 511.9620   2.952881
#> 155 560.6337   2.903716
#> 156 502.5060   2.774478
#> 157 545.1495   2.603348
#> 158 566.8351   2.769767
#> 159 500.0173   3.173201
#> 160 569.2819   2.552064
#> 161 500.1553   2.628412
#> 162 500.2059   3.178765
#> 163 503.7275   3.510194
#> 164 666.5756   2.860119
#> 165 502.6785   3.039291
#> 166 712.8899   2.614692
#> 167 504.5025   2.808638
#> 168 500.3799   3.038598
#> 169 501.5692   2.820968
#> 170 581.2475   2.615801
#> 171 524.1782   2.759976
#> 172 544.0228   2.780036
#> 173 502.3489   2.687948
#> 174 506.7064   2.504373
#> 175 500.6561   3.155847
#> 176 554.7501   2.621509
#> 177 524.0140   2.786599
#> 178 500.5837   2.639470
#> 179 551.7636   3.125202
#> 180 754.7324   2.624251
#> 181 505.5242   3.425893
#> 182 501.3238   2.716702
#> 183 500.0961   3.308457
#> 184 508.7480   3.113474
#> 185 534.8725   3.499786
#> 186 517.8746   2.538063
#> 187 590.9906   2.670028
#> 188 501.0310   2.914594
#> 189 638.0001   2.934929
#> 190 500.0888   2.851532
#> 191 500.5932   3.666377
#> 192 504.9742   3.484749
#> 193 581.8649   2.541017
#> 194 500.9170   2.804932
#> 195 502.8024   2.957736
#> 196 647.0160   2.576301
#> 197 500.4473   2.835700
#> 198 508.8714   3.398558
#> 199 532.7745   2.506832
#> 200 504.4066   2.509984
#> 201 500.6410   2.888198
#> 202 566.4770   2.546294
#> 203 504.7450   3.195571
#> 204 500.0903   2.845765
#> 205 794.2764   2.670927
#> 206 674.7563   2.539742
#> 207 510.9768   3.391883
#> 208 500.3867   3.043401
#> 209 507.9156   3.190308
#> 210 599.8375   2.636404
#> 211 501.3591   2.816920
#> 212 506.6930   2.757760
#> 213 501.2974   2.515350
#> 214 504.0380   3.351697
#> 215 502.3339   3.371341
#> 216 500.2550   3.122488
#> 217 502.3523   2.533056
#> 218 766.2596   2.874092
#> 219 524.4447   3.167589
#> 220 503.0852   2.738484
#> 221 500.0467   2.549824
#> 222 617.4624   2.658100
#> 223 500.7965   2.690428
#> 224 500.0085   2.644114
#> 225 500.5482   2.527209
#> 226 500.0424   2.937174
#> 227 536.0594   2.582199
#> 228 500.3656   4.234302
#> 229 500.4551   2.903105
#> 230 504.4016   2.936274
#> 231 500.0207   2.535741
#> 232 500.0805   2.998506
#> 233 970.9229   2.621301
#> 234 530.7803   2.813365
#> 235 500.9856   4.021607
#> 236 505.5024   3.259226
#> 237 634.2160   3.272783
#> 238 505.1091   2.892908
#> 239 500.2251   2.660340
#> 240 690.8866   2.943845
#> 241 616.0715   4.191519
#> 242 520.1663   2.860404
#> 243 506.0920   2.559175
#> 244 504.9380   3.135190
#> 245 546.4750   2.979677
#> 246 506.0548   2.997407
#> 247 500.0168   2.518640
#> 248 503.6262   3.113018
#> 249 509.8285   2.507850
#> 250 654.6796   3.670440
#> 251 500.0621   2.639273
#> 252 671.2169   3.366981
#> 253 500.0717   2.699967
#> 254 758.1382   2.966704
#> 255 502.0195   2.643966
#> 256 503.5875   2.941445
#> 257 502.3580   2.536316
#> 258 548.4344   3.334085
#> 259 527.3746   2.576707
#> 260 500.0008   2.678864
#> 261 513.4723   2.804851
#> 262 517.0255   2.596257
#> 263 530.3493   2.715464
#> 264 677.6616   2.815301
#> 265 500.2354   2.687330
#> 266 985.2082   3.552054
#> 267 506.8025   3.346031
#> 268 586.9583   2.615196
#> 269 530.7016   2.581663
#> 270 514.5040   2.541442
#> 271 517.9492   4.079905
#> 272 500.3980   2.654928
#> 273 500.5045   2.509620
#> 274 978.3046   3.108632
#> 275 500.1238   2.825176
#> 276 500.1040   2.770609
#> 277 500.0267   2.518399
#> 278 583.7222   2.525075
#> 279 601.4637   2.802600
#> 280 501.9994   3.917258
#> 281 515.3403   2.640554
#> 282 500.0876   2.776061
#> 283 500.0180   2.955226
#> 284 501.3585   3.159510
#> 285 500.7317   2.612794
#> 286 502.3614   3.610004
#> 287 547.3224   2.844972
#> 288 500.4675   3.211323
#> 289 500.1167   2.609567
#> 290 500.0005   2.564237
#> 291 500.5312   2.559605
#> 292 939.6363   3.347405
#> 293 500.0353   3.284235
#> 294 504.6764   2.507638
#> 295 739.7934   2.533419
#> 296 613.7916   2.726599
#> 297 504.7487   2.713326
#> 298 848.4005   2.789544
#> 299 500.5967   2.595265
#> 300 514.6082   2.575534
#> 301 501.9848   2.945643
#> 302 509.8895   3.406725
#> 303 557.4972   2.564272
#> 304 500.1215   2.506360
#> 305 505.8772   2.678093
#> 306 500.0025   2.511711
#> 307 500.3030   2.508928
#> 308 500.0676   3.065464
#> 309 504.8822   2.550871
#> 310 538.8455   3.251490
#> 311 662.9058   2.631154
#> 312 500.1589   2.731663
#> 313 672.0180   2.814078
#> 314 748.9027   2.797779
#> 315 503.9419   2.576573
#> 316 500.1810   2.735992
#> 317 501.5867   2.810109
#> 318 501.5592   2.791985
#> 319 509.5506   2.611550
#> 320 503.9685   2.762557
#> 321 753.7409   3.372628
#> 322 522.5790   2.873506
#> 323 553.6131   2.639049
#> 324 530.9650   2.632774
#> 325 500.0454   2.910369
#> 326 790.9365   3.031278
#> 327 500.1977   2.623971
#> 328 500.0365   2.527267
#> 329 579.4140   3.409166
#> 330 706.3216   2.724683
#> 331 519.5400   4.923360
#> 332 500.6406   2.511321
#> 333 921.7564   2.943612
#> 334 851.3906   2.626328
#> 335 500.1906   2.873784
#> 336 500.5381   2.880230
#> 337 502.6934   2.854657
#> 338 530.9474   2.596647
#> 339 500.6167   2.949488
#> 340 500.2134   3.235738
#> 341 500.0926   2.700534
#> 342 502.6690   2.550199
#> 343 567.6531   2.573056
#> 344 500.0688   2.589657
#> 345 500.1379   2.774950
#> 346 500.2040   2.529745
#> 347 501.6774   3.671696
#> 348 500.4380   2.926910
#> 349 669.9432   3.003778
#> 350 500.9193   2.552848
#> 351 501.3850   3.734625
#> 352 500.1258   2.588488
#> 353 824.3454   2.827147
#> 354 500.0843   4.961662
#> 355 528.7582   2.745893
#> 356 516.0606   2.787700
#> 357 612.0706   2.880068
#> 358 508.9121   2.582504
#> 359 501.3040   2.673939
#> 360 500.0163   2.520313
#> 361 784.3216   2.599688
#> 362 500.2511   2.529576
#> 363 501.0870   2.842143
#> 364 501.2348   2.656527
#> 365 500.1245   2.528117
#> 366 519.6098   2.679818
#> 367 500.4164   3.342883
#> 368 514.0432   2.514397
#> 369 546.1263   2.728279
#> 370 500.0134   2.623256
#> 371 574.3356   2.734243
#> 372 501.0281   3.227874
#> 373 500.7620   2.712953
#> 374 618.1984   2.562501
#> 375 500.4606   2.517495
#> 376 500.9290   3.499800
#> 377 500.0560   3.287571
#> 378 509.8226   2.680295
#> 379 571.1609   2.509260
#> 380 533.5612   2.714379
#> 381 502.9210   3.577345
#> 382 597.4634   2.773189
#> 383 507.2172   2.562649
#> 384 500.3310   2.684528
#> 385 834.6225   2.885671
#> 386 506.6929   3.155671
#> 387 502.9220   2.995397
#> 388 510.7360   2.614042
#> 389 929.5379   2.539835
#> 390 500.0882   2.602822
#> 391 509.0225   2.561634
#> 392 500.5093   3.300204
#> 393 504.9327   3.865677
#> 394 500.0426   2.774907
#> 395 766.1890   3.033732
#> 396 505.0802   4.375072
#> 397 507.4788   3.174221
#> 398 500.0472   2.573052
#> 399 542.4595   2.788584
#> 400 521.3693   3.120966
#> 401 545.9769   2.836002
#> 402 519.2247   2.988852
#> 403 953.8406   3.213166
#> 404 507.6426   3.736654
#> 405 500.2331   4.368626
#> 406 646.7498   3.606547
#> 407 882.5571   2.672032
#> 408 500.7946   2.597287
#> 409 556.8075   2.518935
#> 410 500.2742   3.069833
#> 411 571.7442   2.834078
#> 412 841.8767   2.652412
#> 413 507.7911   2.842436
#> 414 537.3149   3.080878
#> 415 539.1879   3.332698
#> 416 501.8766   2.627172
#> 417 605.5075   2.627091
#> 418 502.0000   2.797457
#> 419 532.3986   2.500434
#> 420 503.7092   4.128556
#> 421 504.6417   3.480641
#> 422 501.1460   2.934827
#> 423 500.4710   2.532501
#> 424 500.6960   2.948237
#> 425 500.3773   3.387255
#> 426 503.5483   2.516336
#> 427 524.4267   2.520886
#> 428 595.3528   2.675834
#> 429 580.5930   2.822412
#> 430 516.9584   2.718875
#> 431 687.3284   2.682179
#> 432 946.8998   4.124535
#> 433 501.9228   2.647309
#> 434 503.6139   3.298282
#> 435 500.0937   2.539466
#> 436 500.3732   2.769716
#> 437 538.5405   3.093215
#> 438 500.2984   3.649789
#> 439 608.5285   2.518917
#> 440 501.6478   2.692397
#> 441 504.0722   2.634534
#> 442 513.1023   3.017647
#> 443 803.6005   2.605040
#> 444 500.3285   2.621908
#> 445 791.2481   2.590414
#> 446 500.3959   2.752138
#> 447 501.0231   3.126682
#> 448 503.8677   2.976641
#> 449 501.8079   2.916428
#> 450 500.3384   2.802624
#> 451 500.4891   2.618574
#> 452 500.3466   2.580608
#> 453 501.0873   3.910988
#> 454 500.0746   3.353860
#> 455 507.9465   2.962436
#> 456 872.4813   2.841001
#> 457 500.4841   3.542973
#> 458 500.6329   2.764842
#> 459 500.2590   3.055827
#> 460 500.1764   2.812655
#> 461 507.0049   2.666694
#> 462 505.4926   2.866632
#> 463 500.0826   2.887669
#> 464 501.8437   2.599053
#> 465 504.3465   2.567333
#> 466 505.5221   2.831132
#> 467 506.6464   2.805806
#> 468 519.7664   2.965881
#> 469 500.4763   2.629557
#> 470 502.4922   2.630421
#> 471 608.8858   2.907037
#> 472 907.0661   2.568125
#> 473 531.1271   2.633274
#> 474 501.7751   3.837507
#> 475 553.9833   3.443292
#> 476 500.1224   2.696737
#> 477 514.8278   2.685651
#> 478 505.5659   3.684711
#> 479 500.0556   3.688802
#> 480 500.1037   3.088456
#> 481 501.6108   3.402630
#> 482 574.7504   2.889087
#> 483 500.0868   3.109026
#> 484 744.0685   2.813505
#> 485 503.5715   3.556911
#> 486 500.0036   2.535825
#> 487 500.0792   2.657214
#> 488 899.2031   2.840267
#> 489 523.9075   2.545590
#> 490 502.7373   2.651814
#> 491 776.0916   3.189646
#> 492 500.1999   3.433406
#> 493 675.1314   2.699604
#> 494 504.2969   2.805369
#> 495 504.6101   2.979061
#> 496 505.5334   4.449750
#> 497 500.4845   2.670203
#> 498 504.9331   3.120428
#> 499 513.8275   2.845871
#> 500 500.0347   2.537483
#> 501 500.5631   3.943964
#> 502 910.1859   2.664428
#> 503 503.8974   2.813869
#> 504 505.0346   3.331140
#> 505 500.0457   3.139662
#> 506 502.2339   3.091828
#> 507 525.6761   2.617252
#> 508 505.2784   2.519989
#> 509 746.4206   2.679821
#> 510 500.5481   2.985970
#> 511 526.1523   3.638331
#> 512 500.3015   2.727917
#> 513 503.1185   4.732020
#> 514 502.6147   2.600678
#> 515 544.9507   2.902021
#> 516 504.7896   2.583778
#> 517 500.0024   5.153761
#> 518 500.0309   2.850082
#> 519 501.7486   2.607193
#> 520 546.7962   2.975316
#> 521 501.2750   3.830292
#> 522 730.2172   2.781052
#> 523 507.0761   2.888037
#> 524 500.3547   2.502076
#> 525 683.0467   2.752234
#> 526 500.4798   2.613994
#> 527 501.1189   3.211021
#> 528 502.5152   2.834684
#> 529 527.6184   2.995023
#> 530 505.9545   2.626761
#> 531 500.9859   2.741048
#> 532 500.9479   3.091670
#> 533 508.2339   3.250758
#> 534 503.0704   3.476914
#> 535 500.6661   3.411405
#> 536 601.0715   2.962577
#> 537 704.2938   3.877389
#> 538 500.1132   2.727258
#> 539 500.7101   2.804966
#> 540 502.3012   3.181189
#> 541 714.2650   2.688133
#> 542 500.6588   2.693054
#> 543 525.5185   2.936177
#> 544 507.6289   2.811812
#> 545 500.0712   2.539720
#> 546 501.2995   2.989579
#> 547 570.9190   2.716579
#> 548 675.2236   3.555195
#> 549 654.5765   3.146006
#> 550 500.8373   2.526331
#> 551 500.0409   4.035466
#> 552 500.0862   3.137496
#> 553 875.7355   2.604987
#> 554 501.2929   2.545911
#> 555 500.0704   3.051526
#> 556 502.3908   2.717446
#> 557 500.0471   2.593536
#> 558 500.0794   2.550433
#> 559 500.9936   2.621016
#> 560 500.1898   2.699944
#> 561 500.4276   3.005741
#> 562 501.5146   3.066702
#> 563 501.1630   2.556111
#> 564 500.7898   3.640754
#> 565 501.0779   3.032459
#> 566 500.1448   2.824398
#> 567 503.8798   2.714881
#> 568 512.2400   3.068545
#> 569 507.6859   2.672214
#> 570 555.5679   2.767992
#> 571 558.4449   2.573242
#> 572 501.7862   3.190192
#> 573 500.3395   2.517609
#> 574 518.4808   3.028611
#> 575 500.6873   3.174442
#> 576 501.1910   2.818289
#> 577 579.6430   2.601017
#> 578 500.0098   3.426131
#> 579 500.0242   2.745041
#> 580 500.0824   2.917303
#> 581 512.9833   2.795181
#> 582 518.3844   2.909104
#> 583 507.3654   2.915785
#> 584 500.6709   2.829078
#> 585 500.4477   2.615890
#> 586 515.6876   2.626205
#> 587 510.9103   2.832794
#> 588 937.1023   3.659154
#> 589 500.9746   3.101195
#> 590 500.1513   2.563620
#> 591 506.2502   2.640780
#> 592 500.3526   2.776300
#> 593 501.4616   2.743853
#> 594 500.3712   2.754239
#> 595 957.5946   2.936429
#> 596 500.3039   2.557898
#> 597 508.6988   2.553292
#> 598 559.3072   2.794395
#> 599 500.0549   2.682648
#> 600 511.8940   2.516191
#> 601 526.7725   3.095298
#> 602 500.2479   3.113199
#> 603 500.3877   2.525102
#> 604 500.2278   2.675625
#> 605 513.8282   3.013893
#> 606 502.5450   3.291694
#> 607 500.7201   2.612842
#> 608 517.7046   2.647834
#> 609 500.1873   2.919709
#> 610 500.0024   2.984620
#> 611 500.6292   3.086466
#> 612 645.6292   3.616087
#> 613 500.5758   2.702547
#> 614 502.2910   3.138790
#> 615 500.0467   3.029986
#> 616 500.1009   2.610282
#> 617 500.4100   2.994976
#> 618 586.6403   3.078217
#> 619 501.6494   2.737019
#> 620 516.8324   2.761719
#> 621 507.2878   3.373749
#> 622 519.4153   3.369319
#> 623 513.7560   2.538825
#> 624 500.6714   2.847054
#> 625 500.8539   3.335280
#> 626 500.2546   2.884222
#> 627 823.3817   2.553163
#> 628 500.8091   2.720993
#> 629 500.0676   2.905917
#> 630 610.6447   2.553559
#> 631 502.3937   3.395599
#> 632 545.4727   2.616525
#> 633 506.5216   3.681803
#> 634 529.4674   3.023979
#> 635 520.6151   2.931714
#> 636 530.9128   2.973966
#> 637 588.2788   2.661571
#> 638 500.9843   2.749524
#> 639 530.8194   3.302121
#> 640 519.8668   2.820608
#> 641 500.2473   2.530615
#> 642 500.0530   2.977671
#> 643 501.7338   2.593841
#> 644 510.4582   2.754771
#> 645 500.0616   2.815946
#> 646 508.2554   2.715680
#> 647 520.7535   4.109385
#> 648 500.2221   4.651821
#> 649 645.8628   2.553023
#> 650 907.2799   2.884804
#> 651 662.5990   2.701015
#> 652 506.2520   2.609972
#> 653 732.5116   2.862257
#> 654 504.5304   2.723290
#> 655 501.4528   3.040381
#> 656 942.5618   2.841635
#> 657 500.5186   2.592879
#> 658 510.5140   3.652628
#> 659 558.3123   2.855294
#> 660 500.4265   3.099909
#> 661 506.6234   2.604535
#> 662 500.4239   2.554233
#> 663 500.1818   2.723398
#> 664 500.0715   2.515457
#> 665 500.1917   2.578110
#> 666 500.0571   2.581058
#> 667 505.0349   2.755803
#> 668 500.0224   2.548414
#> 669 633.7693   2.608240
#> 670 527.3400   3.280146
#> 671 500.4726   2.839589
#> 672 500.2161   2.512901
#> 673 512.6843   2.953670
#> 674 500.9560   2.920213
#> 675 500.0271   2.985577
#> 676 508.3898   2.961507
#> 677 503.0505   3.038639
#> 678 500.0198   3.322034
#> 679 500.1473   2.827856
#> 680 510.8479   2.528288
#> 681 692.2495   2.635020
#> 682 500.6056   3.289603
#> 683 502.9374   3.383193
#> 684 893.1668   2.802761
#> 685 500.0294   5.008387
#> 686 500.0816   2.577985
#> 687 560.6410   2.904880
#> 688 501.5845   3.035740
#> 689 500.2717   2.508025
#> 690 502.5903   2.606512
#> 691 563.8255   2.925492
#> 692 507.4849   2.590346
#> 693 505.6827   2.974425
#> 694 501.1841   2.549573
#> 695 500.0744   2.631337
#> 696 702.5909   2.981481
#> 697 500.5552   3.132101
#> 698 514.8118   5.433466
#> 699 508.9703   3.103453
#> 700 500.7126   2.637195
#> 701 730.7118   2.603206
#> 702 504.0715   2.855460
#> 703 503.7912   3.143694
#> 704 584.6996   2.769370
#> 705 500.0605   2.738236
#> 706 508.6068   3.893668
#> 707 501.5839   3.146917
#> 708 516.8324   3.177636
#> 709 501.5930   3.217008
#> 710 502.0935   3.017433
#> 711 502.6480   2.512381
#> 712 508.1117   3.104803
#> 713 501.9200   3.208364
#> 714 500.7379   2.627613
#> 715 519.3694   2.993348
#> 716 501.0686   2.937076
#> 717 525.2585   2.893418
#> 718 500.5570   2.574969
#> 719 500.4086   3.136571
#> 720 501.7483   2.582069
#> 721 502.6023   3.602700
#> 722 500.1083   2.611218
#> 723 500.0064   3.262740
#> 724 505.9797   2.822707
#> 725 500.8371   3.021611
#> 726 599.3194   3.773076
#> 727 500.9366   3.347553
#> 728 501.1960   2.631575
#> 729 502.8351   2.615535
#> 730 503.4101   2.551816
#> 731 537.6700   3.321940
#> 732 757.5358   2.962641
#> 733 500.2906   3.334634
#> 734 518.0481   2.524140
#> 735 501.4046   2.540834
#> 736 500.0034   3.007259
#> 737 559.1664   2.669301
#> 738 500.2681   2.907264
#> 739 500.0429   3.334862
#> 740 512.0391   2.990475
#> 741 504.0632   2.520919
#> 742 500.2640   2.560045
#> 743 500.0795   2.655815
#> 744 752.3980   2.767912
#> 745 504.9441   2.594353
#> 746 500.1202   2.717178
#> 747 513.4302   2.909346
#> 748 846.5442   2.697268
#> 749 503.3874   3.461442
#> 750 500.0854   2.694197
#> 751 503.4797   3.983817
#> 752 568.4713   2.763779
#> 753 534.1121   3.288694
#> 754 502.8898   2.739439
#> 755 500.0632   2.823439
#> 756 503.4441   2.961945
#> 757 517.1165   2.926931
#> 758 504.5100   3.392189
#> 759 500.3808   2.698208
#> 760 592.1230   2.539325
#> 761 500.0537   3.237641
#> 762 502.1589   2.533627
#> 763 510.0300   2.687994
#> 764 500.4609   2.859371
#> 765 500.0591   2.921305
#> 766 564.0262   2.658415
#> 767 542.6703   2.996797
#> 768 552.8321   2.933146
#> 769 502.7346   2.569020
#> 770 517.5979   2.637549
#> 771 500.4379   2.544037
#> 772 500.1447   2.842951
#> 773 511.2090   2.947277
#> 774 502.3113   2.953190
#> 775 500.7149   3.548058
#> 776 516.8203   3.147700
#> 777 511.6236   2.511668
#> 778 610.4417   2.542380
#> 779 500.3408   2.522781
#> 780 509.2601   2.676928
#> 781 511.9720   3.403344
#> 782 506.3223   2.540645
#> 783 675.0997   2.876215
#> 784 513.7522   2.598915
#> 785 503.3333   2.552153
#> 786 501.5075   4.506246
#> 787 520.1698   3.699631
#> 788 500.1015   2.765705
#> 789 505.6853   2.792891
#> 790 509.6625   2.508053
#> 791 502.7281   2.633459
#> 792 500.5940   2.672824
#> 793 501.4571   2.915902
#> 794 542.9963   2.719940
#> 795 523.9590   2.514545
#> 796 500.0629   2.746982
#> 797 504.4664   3.103363
#> 798 500.7855   3.326133
#> 799 500.4749   3.007256
#> 800 500.0898   3.302867
#> 801 500.4699   2.540601
#> 802 501.2670   2.667889
#> 803 581.1009   2.652112
#> 804 501.2809   2.823601
#> 805 500.1454   2.547560
#> 806 500.4880   2.742388
#> 807 509.9357   2.599831
#> 808 500.0606   2.507588
#> 809 501.2867   3.117556
#> 810 500.0718   2.519203
#> 811 500.8390   2.998032
#> 812 500.7365   2.503538
#> 813 509.2469   2.524416
#> 814 501.2229   3.147394
#> 815 500.4151   3.101241
#> 816 504.1635   2.601516
#> 817 505.5528   2.548886
#> 818 974.3202   2.965100
#> 819 500.5134   2.551220
#> 820 505.7777   2.729730
#> 821 545.4181   2.831551
#> 822 517.7411   3.750608
#> 823 500.4009   3.123754
#> 824 523.2659   3.929800
#> 825 508.6928   3.047887
#> 826 502.8046   2.517990
#> 827 505.6446   3.829228
#> 828 501.3815   2.611566
#> 829 500.1454   2.827234
#> 830 518.1142   2.708463
#> 831 500.0404   2.582103
#> 832 565.1904   2.727181
#> 833 503.5025   3.363179
#> 834 754.9334   2.703057
#> 835 524.2842   2.871537
#> 836 501.2335   2.556232
#> 837 500.0631   2.585027
#> 838 500.9474   3.662109
#> 839 501.6351   2.658245
#> 840 500.1857   2.851055
#> 841 500.0530   2.657852
#> 842 608.7305   3.713428
#> 843 500.2962   2.987619
#> 844 502.3753   2.726099
#> 845 501.1305   2.530396
#> 846 502.5176   2.677240
#> 847 504.7721   3.319054
#> 848 547.5602   2.755768
#> 849 500.8080   2.796693
#> 850 761.6121   3.097222
#> 851 742.4771   3.678314
#> 852 574.1564   2.638574
#> 853 612.1677   2.574614
#> 854 505.3807   2.548643
#> 855 513.4368   2.946479
#> 856 571.4898   2.612959
#> 857 511.9218   2.568621