Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   501.8739   3.019016
#> 2   500.1874   2.807476
#> 3   501.1321   2.574050
#> 4   999.9662   2.683818
#> 5   500.0051   2.521771
#> 6   500.0072   2.555124
#> 7   501.6012   3.125708
#> 8   500.1488   3.561006
#> 9   500.0018   3.802680
#> 10  500.0971   3.434184
#> 11  500.5089   2.534278
#> 12  500.9956   2.581759
#> 13  500.0825   2.769294
#> 14  502.6099   2.648093
#> 15  500.4678   2.582005
#> 16  657.4705   2.836876
#> 17  500.5574   3.765684
#> 18  500.0828   4.404316
#> 19  500.1203   3.072371
#> 20  500.2867   2.989116
#> 21  500.8699   2.999944
#> 22  512.3624   3.131579
#> 23  500.0103   2.820732
#> 24  532.1056   2.639445
#> 25  502.5559   2.575494
#> 26  543.4584   3.520016
#> 27  500.0412   3.830681
#> 28  500.1774   2.518926
#> 29  502.8690   2.546212
#> 30  500.1448   2.792081
#> 31  500.1425   2.990463
#> 32  510.4979   2.581126
#> 33  500.0274   4.173176
#> 34  500.0116   2.873652
#> 35  505.9910   3.323144
#> 36  540.0071   3.469159
#> 37  511.7544   2.832008
#> 38  500.2865   3.000234
#> 39  571.1979   3.231915
#> 40  521.5601   2.944153
#> 41  500.2347   4.073278
#> 42  500.8383   3.361809
#> 43  500.2284   2.666853
#> 44  516.8578   2.677661
#> 45  501.5910   2.742617
#> 46  502.5267   2.779576
#> 47  500.3314   5.406642
#> 48  500.4656   2.863828
#> 49  502.6970   2.544686
#> 50  522.0070   2.721903
#> 51  500.4135   3.001449
#> 52  500.9383   3.116525
#> 53  501.7184   2.760474
#> 54  500.0547   2.910222
#> 55  512.2528   3.411063
#> 56  763.3540   2.869459
#> 57  544.9921   2.845458
#> 58  563.2723   3.292056
#> 59  500.0626   2.954551
#> 60  623.7322   2.609845
#> 61  500.1172   2.525130
#> 62  595.6972   2.752011
#> 63  520.0721   2.839777
#> 64  500.2331   2.810693
#> 65  500.1131   2.824048
#> 66  501.0003   2.508711
#> 67  507.8140   2.513911
#> 68  500.1730   3.019398
#> 69  790.6383   3.886196
#> 70  722.8561   3.136234
#> 71  730.3478   2.613660
#> 72  506.7778   4.046862
#> 73  500.6627   4.641825
#> 74  501.0647   3.107212
#> 75  500.0483   2.782986
#> 76  500.0519   2.574513
#> 77  500.1797   2.958076
#> 78  500.0767   2.526125
#> 79  884.9177   2.704531
#> 80  502.0620   2.716278
#> 81  603.0818   2.586397
#> 82  504.6219   2.966391
#> 83  500.4579   2.603055
#> 84  641.2354   2.676765
#> 85  510.8391   2.593592
#> 86  500.0261   2.800921
#> 87  512.2286   2.904770
#> 88  500.6579   2.732935
#> 89  623.2607   3.028068
#> 90  709.9774   2.913621
#> 91  997.7580   2.632652
#> 92  502.0401   2.505566
#> 93  500.1986   3.055442
#> 94  501.3228   2.591467
#> 95  510.0687   2.875665
#> 96  507.7413   3.245648
#> 97  500.0259   3.029634
#> 98  654.6172   2.577842
#> 99  549.8048   3.013669
#> 100 501.2622   2.840218
#> 101 639.2012   3.052565
#> 102 531.8303   2.550576
#> 103 535.0338   2.554273
#> 104 857.2541   3.086461
#> 105 501.0923   2.632896
#> 106 500.0377   3.204794
#> 107 504.5107   3.247487
#> 108 500.1022   2.557288
#> 109 504.2148   2.864843
#> 110 500.2914   2.645358
#> 111 696.4290   2.938046
#> 112 815.5086   2.543972
#> 113 574.0926   2.619702
#> 114 500.4659   3.592165
#> 115 503.0579   2.657991
#> 116 951.8923   2.502606
#> 117 512.0924   2.786831
#> 118 510.3764   3.672487
#> 119 500.4409   2.700798
#> 120 500.4593   3.098248
#> 121 516.8477   3.494899
#> 122 631.6068   2.716271
#> 123 500.3082   2.605028
#> 124 501.3745   2.678686
#> 125 500.0853   2.509288
#> 126 500.9694   2.602921
#> 127 501.2305   2.726001
#> 128 501.0016   2.879079
#> 129 500.2818   2.865475
#> 130 551.9946   3.117847
#> 131 532.4732   2.889624
#> 132 500.9913   3.226533
#> 133 743.1725   2.878123
#> 134 667.8734   2.566059
#> 135 554.0789   2.998596
#> 136 510.4428   3.530419
#> 137 549.7999   2.725582
#> 138 500.0154   2.710652
#> 139 501.0979   2.511453
#> 140 561.5141   2.607965
#> 141 516.4181   2.574285
#> 142 617.5145   2.612715
#> 143 500.1700   2.534227
#> 144 508.4536   2.608281
#> 145 682.1534   2.516584
#> 146 956.7810   2.613528
#> 147 577.3919   2.603279
#> 148 511.0268   2.697304
#> 149 678.2729   2.641945
#> 150 526.8252   3.391557
#> 151 501.2625   2.693060
#> 152 501.3754   2.518945
#> 153 515.4996   2.782024
#> 154 500.3825   3.001085
#> 155 594.7677   2.601265
#> 156 541.3823   2.841210
#> 157 503.6134   3.620741
#> 158 501.0193   2.910974
#> 159 521.6219   2.798343
#> 160 516.7550   3.680829
#> 161 507.1090   2.599386
#> 162 506.0869   2.620029
#> 163 501.1886   2.560873
#> 164 544.1810   2.569706
#> 165 500.1980   3.112117
#> 166 545.2216   2.922723
#> 167 622.2914   3.059976
#> 168 767.7043   2.511050
#> 169 508.5130   2.709106
#> 170 502.8518   3.707857
#> 171 500.1840   2.659846
#> 172 512.4871   2.609500
#> 173 501.1237   4.383024
#> 174 500.0113   2.607999
#> 175 500.7806   2.801779
#> 176 533.6348   4.005857
#> 177 505.2412   2.902331
#> 178 507.3816   3.835324
#> 179 860.3280   2.579907
#> 180 500.3595   2.576176
#> 181 509.2513   3.083944
#> 182 501.0935   3.290022
#> 183 501.6364   2.635213
#> 184 566.9185   4.850010
#> 185 900.8183   2.984673
#> 186 539.3677   2.942449
#> 187 502.5477   2.864687
#> 188 500.0908   3.028263
#> 189 521.1640   3.316029
#> 190 500.0025   3.635897
#> 191 502.5726   3.718206
#> 192 500.5706   3.067363
#> 193 504.3739   2.869186
#> 194 500.8117   3.114873
#> 195 500.9586   2.710888
#> 196 500.0241   2.874340
#> 197 511.6440   3.437235
#> 198 500.2362   2.996147
#> 199 500.2001   2.536216
#> 200 508.1357   4.220898
#> 201 501.2104   2.543557
#> 202 500.3743   2.684597
#> 203 503.1789   2.765421
#> 204 510.8942   2.802877
#> 205 505.0064   3.965094
#> 206 559.4698   2.995601
#> 207 500.0564   2.543614
#> 208 569.4225   2.607065
#> 209 500.9079   2.562694
#> 210 502.5943   3.619596
#> 211 501.1079   3.221108
#> 212 500.8964   2.688340
#> 213 507.1640   2.686802
#> 214 500.5172   2.999426
#> 215 504.7280   3.090933
#> 216 500.3507   2.681074
#> 217 500.4567   2.834999
#> 218 500.5649   3.806001
#> 219 597.6862   2.644643
#> 220 521.9281   2.774063
#> 221 516.5177   5.046363
#> 222 500.3711   3.150303
#> 223 507.9122   2.764028
#> 224 516.0169   3.236398
#> 225 500.3118   3.255155
#> 226 502.2768   2.699105
#> 227 549.1640   2.620881
#> 228 707.0577   2.820319
#> 229 969.3095   2.840398
#> 230 580.4162   3.315133
#> 231 502.5315   3.245189
#> 232 500.1709   2.550260
#> 233 500.3059   2.642057
#> 234 515.8277   2.954583
#> 235 500.4986   3.021866
#> 236 500.2589   2.666192
#> 237 500.0730   2.873658
#> 238 500.2717   2.982213
#> 239 500.0783   2.982921
#> 240 504.4403   2.775956
#> 241 500.5791   3.073139
#> 242 503.9921   2.545633
#> 243 696.7621   2.823506
#> 244 500.0102   2.610149
#> 245 508.0989   3.301544
#> 246 556.8134   3.575538
#> 247 500.0425   2.634945
#> 248 909.1179   2.571810
#> 249 529.6157   2.881046
#> 250 549.1223   3.360321
#> 251 510.0172   3.055747
#> 252 500.6860   3.160753
#> 253 504.7680   2.670977
#> 254 681.9534   3.419490
#> 255 500.3711   2.676113
#> 256 501.5014   2.527975
#> 257 501.3289   2.659639
#> 258 501.1241   3.728831
#> 259 500.6446   3.811975
#> 260 500.0092   3.305117
#> 261 500.7124   2.942389
#> 262 527.5654   3.041652
#> 263 514.9696   2.685614
#> 264 501.1534   2.537293
#> 265 500.0012   2.860265
#> 266 507.7738   2.994702
#> 267 501.9038   2.542630
#> 268 500.4807   2.854674
#> 269 500.8532   2.504835
#> 270 504.0102   2.552376
#> 271 501.1421   2.516310
#> 272 504.0117   3.192655
#> 273 500.3649   2.542807
#> 274 593.1990   2.596165
#> 275 633.1008   3.307323
#> 276 501.8147   2.733320
#> 277 500.0209   3.168642
#> 278 504.2433   2.648935
#> 279 543.0049   2.900199
#> 280 500.0739   2.589249
#> 281 508.8997   2.772132
#> 282 521.6257   2.509149
#> 283 500.0042   2.930175
#> 284 500.7389   2.670360
#> 285 526.4911   2.768259
#> 286 500.2942   2.798545
#> 287 502.9423   2.947009
#> 288 502.5125   2.720817
#> 289 501.4217   2.680579
#> 290 509.2340   2.611053
#> 291 503.5329   2.777783
#> 292 500.0846   2.501089
#> 293 500.6385   2.883302
#> 294 504.0272   3.038815
#> 295 505.0080   2.534293
#> 296 500.9997   2.729830
#> 297 546.3126   2.947309
#> 298 656.9835   2.502150
#> 299 587.2736   3.095855
#> 300 555.7898   3.545380
#> 301 501.2976   2.574387
#> 302 575.9768   2.529845
#> 303 500.8688   2.578050
#> 304 501.0515   2.556724
#> 305 500.1029   3.003252
#> 306 742.6695   2.661454
#> 307 500.0513   2.830172
#> 308 595.5198   2.532695
#> 309 500.1480   2.666834
#> 310 517.7149   2.663532
#> 311 500.0405   3.449708
#> 312 501.4680   3.760422
#> 313 500.4646   2.504620
#> 314 507.7670   2.642360
#> 315 523.8466   2.902216
#> 316 501.1874   2.621880
#> 317 500.9035   2.932149
#> 318 500.6518   2.910921
#> 319 500.4875   2.547628
#> 320 505.3340   2.543928
#> 321 503.1092   2.960981
#> 322 512.7750   2.589839
#> 323 500.8886   2.763456
#> 324 941.7330   2.998819
#> 325 500.8201   2.615337
#> 326 500.3331   3.050835
#> 327 611.9046   2.738285
#> 328 505.6973   2.696093
#> 329 525.4445   2.692583
#> 330 500.0307   2.579221
#> 331 502.5213   2.669503
#> 332 500.4186   2.686875
#> 333 500.0383   2.963271
#> 334 502.2908   2.724484
#> 335 501.1579   2.721710
#> 336 500.6858   3.426973
#> 337 584.8711   2.614516
#> 338 531.8172   2.548890
#> 339 500.6108   2.737830
#> 340 977.7525   2.725753
#> 341 500.4826   2.509489
#> 342 752.8726   3.575237
#> 343 500.5884   2.548564
#> 344 504.5977   3.032538
#> 345 501.4113   3.238383
#> 346 500.0624   3.260661
#> 347 501.2563   2.559160
#> 348 500.1161   2.858530
#> 349 500.8025   2.524200
#> 350 500.5334   2.524737
#> 351 500.2569   3.240540
#> 352 500.0830   2.943454
#> 353 535.6838   3.107515
#> 354 572.9298   2.843017
#> 355 608.5097   2.606018
#> 356 650.2411   2.504048
#> 357 520.7398   2.675834
#> 358 500.5084   3.164143
#> 359 609.6662   4.084919
#> 360 500.1393   3.107595
#> 361 503.0952   2.656630
#> 362 539.7049   3.095321
#> 363 500.0493   3.212388
#> 364 669.4740   2.596663
#> 365 501.0778   2.956878
#> 366 634.7695   2.514456
#> 367 503.4918   2.860786
#> 368 521.5005   2.772889
#> 369 552.5509   3.046236
#> 370 500.0844   2.894127
#> 371 502.4816   2.738198
#> 372 664.3895   4.626017
#> 373 500.8816   2.600718
#> 374 500.6402   2.517027
#> 375 522.7046   2.858483
#> 376 501.1235   3.136010
#> 377 528.2427   2.733103
#> 378 502.2011   3.130539
#> 379 889.7822   2.705010
#> 380 500.5307   3.734974
#> 381 501.2224   3.329423
#> 382 502.5933   2.702733
#> 383 512.4090   2.664454
#> 384 500.2704   2.509671
#> 385 510.0382   2.784771
#> 386 538.0727   3.044334
#> 387 500.9543   2.957462
#> 388 502.8620   2.613833
#> 389 504.1193   3.166587
#> 390 500.8036   2.541273
#> 391 500.7110   2.706778
#> 392 500.4112   2.504649
#> 393 509.8082   3.665948
#> 394 500.2718   3.309735
#> 395 728.3363   3.200571
#> 396 500.0611   2.646135
#> 397 524.5779   3.195834
#> 398 545.3996   3.423276
#> 399 500.1560   2.597550
#> 400 500.2232   2.711164
#> 401 500.4803   3.482682
#> 402 532.3584   2.618432
#> 403 500.7195   2.820981
#> 404 523.8036   2.582611
#> 405 614.4012   2.879424
#> 406 500.4202   2.664908
#> 407 500.5889   2.508876
#> 408 507.6538   4.131458
#> 409 504.6809   3.229791
#> 410 500.0528   3.343474
#> 411 509.5553   2.562051
#> 412 500.3141   3.166560
#> 413 500.2647   2.595638
#> 414 503.5945   2.877179
#> 415 500.3087   4.286918
#> 416 500.0829   2.566181
#> 417 502.8674   3.293385
#> 418 502.4651   2.786469
#> 419 500.0700   2.577149
#> 420 592.6993   2.562839
#> 421 648.0546   3.009394
#> 422 500.4361   3.022917
#> 423 500.5532   4.352975
#> 424 501.8317   2.730383
#> 425 501.5269   2.917269
#> 426 603.0126   3.766758
#> 427 508.2251   2.652090
#> 428 510.4750   2.782862
#> 429 507.5771   2.792060
#> 430 510.4822   2.752147
#> 431 972.6485   4.221811
#> 432 500.0054   2.571985
#> 433 500.2854   3.533736
#> 434 500.8728   3.380665
#> 435 500.2504   2.674459
#> 436 866.2563   3.371300
#> 437 588.5243   2.878982
#> 438 500.0657   2.521209
#> 439 502.6371   5.221312
#> 440 667.4577   2.509825
#> 441 710.2188   2.924041
#> 442 500.0962   3.040640
#> 443 517.2462   2.640923
#> 444 500.7472   2.650103
#> 445 500.0823   2.806235
#> 446 502.0058   2.673055
#> 447 571.2977   3.082639
#> 448 502.0026   2.648755
#> 449 500.5318   2.566611
#> 450 500.4305   3.082665
#> 451 500.7152   5.604222
#> 452 500.0348   2.533841
#> 453 638.9204   2.841043
#> 454 500.7818   2.565898
#> 455 548.6488   3.543244
#> 456 723.8589   2.737657
#> 457 509.7115   3.151583
#> 458 500.8166   2.693957
#> 459 524.7147   3.540646
#> 460 500.4666   2.576974
#> 461 500.3576   2.768381
#> 462 517.6559   2.519794
#> 463 500.2957   3.211555
#> 464 505.4094   2.751895
#> 465 501.5166   3.796245
#> 466 762.4324   2.624918
#> 467 500.4007   2.921779
#> 468 500.2948   2.853148
#> 469 502.1667   3.463631
#> 470 517.9185   3.476025
#> 471 503.2523   2.929664
#> 472 508.7091   2.934244
#> 473 516.8872   3.585471
#> 474 517.8199   2.558196
#> 475 500.2502   2.723136
#> 476 500.2820   3.290500
#> 477 874.1253   2.700250
#> 478 710.9244   2.659801
#> 479 500.0873   2.542604
#> 480 500.5582   2.649274
#> 481 500.6475   2.709224
#> 482 500.9378   2.974757
#> 483 500.4845   2.503393
#> 484 870.9236   2.890757
#> 485 516.1750   2.583641
#> 486 510.7328   2.574544
#> 487 537.7145   2.966074
#> 488 500.2758   2.839460
#> 489 500.0350   2.865464
#> 490 877.2857   2.573972
#> 491 500.0555   2.837586
#> 492 500.2198   2.667723
#> 493 500.5947   2.886740
#> 494 500.2402   2.573777
#> 495 500.0112   2.706525
#> 496 500.0733   2.508609
#> 497 502.1683   2.511451
#> 498 504.0270   3.048499
#> 499 523.9426   3.113269
#> 500 500.0529   2.690610
#> 501 514.8145   3.113278
#> 502 505.9405   3.105520
#> 503 853.8069   2.762502
#> 504 500.3806   3.963256
#> 505 501.3461   2.511210
#> 506 507.3007   2.832531
#> 507 507.6441   2.616269
#> 508 500.2472   3.081193
#> 509 647.8493   3.389750
#> 510 501.1863   2.728404
#> 511 679.7815   2.555761
#> 512 989.4656   2.961300
#> 513 500.0531   2.997228
#> 514 501.8295   3.027275
#> 515 500.2113   2.562754
#> 516 500.2147   2.881318
#> 517 530.4291   2.521648
#> 518 500.2937   2.666480
#> 519 500.7255   2.963472
#> 520 599.0770   2.569741
#> 521 548.4970   2.528048
#> 522 536.4888   3.505735
#> 523 669.7760   2.586813
#> 524 500.3495   3.025251
#> 525 707.9390   2.843130
#> 526 500.6820   2.585052
#> 527 500.1135   2.568164
#> 528 748.3605   3.131095
#> 529 516.8923   2.661195
#> 530 876.2211   2.865710
#> 531 522.3933   2.551354
#> 532 622.7442   3.036748
#> 533 501.0808   2.694509
#> 534 500.9689   2.787491
#> 535 500.3775   2.553593
#> 536 507.2280   2.589190
#> 537 504.9941   2.591248
#> 538 512.9369   2.571860
#> 539 503.3387   3.550494
#> 540 988.4587   2.840336
#> 541 511.5910   4.332695
#> 542 513.5999   2.678112
#> 543 500.0965   2.851765
#> 544 500.0784   2.571542
#> 545 500.0219   2.897231
#> 546 512.7002   2.581527
#> 547 617.9577   3.930565
#> 548 502.3616   2.728208
#> 549 523.1187   2.641111
#> 550 500.5433   3.136735
#> 551 502.4452   2.893596
#> 552 575.4303   2.697000
#> 553 591.7608   2.840279
#> 554 607.4651   3.178429
#> 555 526.5066   2.652877
#> 556 560.3559   3.442396
#> 557 882.7210   3.120658
#> 558 515.5999   2.630995
#> 559 710.1939   2.986874
#> 560 504.3492   3.042567
#> 561 549.5675   2.922197
#> 562 502.2113   2.982756
#> 563 500.5054   2.545997
#> 564 500.0824   2.985312
#> 565 581.7597   3.035604
#> 566 511.4835   2.508872
#> 567 559.6869   2.927182
#> 568 500.5463   4.807674
#> 569 517.3316   3.194583
#> 570 500.1071   2.748565
#> 571 502.1381   2.639703
#> 572 511.8171   3.208030
#> 573 500.0096   3.004812
#> 574 500.3060   2.740664
#> 575 502.8401   2.705235
#> 576 502.1156   3.290935
#> 577 566.6977   2.834195
#> 578 511.5786   3.351863
#> 579 500.3013   2.838283
#> 580 574.3746   2.835060
#> 581 627.4228   4.526658
#> 582 500.5281   3.679685
#> 583 558.8768   3.080196
#> 584 500.1356   2.779499
#> 585 641.1161   2.914348
#> 586 968.3932   3.702292
#> 587 507.7032   2.953238
#> 588 684.3879   2.706958
#> 589 507.6270   3.094799
#> 590 500.5152   3.775079
#> 591 581.5748   3.002962
#> 592 624.7685   2.662382
#> 593 888.4960   3.638240
#> 594 500.4051   2.922492
#> 595 506.3166   3.801753
#> 596 500.7018   2.639039
#> 597 500.6454   2.973816
#> 598 548.2757   3.012267
#> 599 500.3429   4.148510
#> 600 862.3563   2.747917
#> 601 541.7003   2.607046
#> 602 502.7165   3.038302
#> 603 503.5870   2.572260
#> 604 679.1935   2.605044
#> 605 508.8949   4.033274
#> 606 761.9425   2.551084
#> 607 501.7358   3.135334
#> 608 504.1079   3.153733
#> 609 528.8674   4.133153
#> 610 502.2365   2.960147
#> 611 500.3312   2.776858
#> 612 583.7913   2.678314
#> 613 504.4609   2.540751
#> 614 501.0486   2.674728
#> 615 500.6841   2.819534
#> 616 500.5001   2.522350
#> 617 709.7701   3.041960
#> 618 500.0132   2.641229
#> 619 500.9395   2.593534
#> 620 505.1057   3.219627
#> 621 637.7431   2.606672
#> 622 500.1493   2.844555
#> 623 562.2381   2.593099
#> 624 501.1539   2.548301
#> 625 520.7688   3.229318
#> 626 513.4237   3.395776
#> 627 500.0375   2.607489
#> 628 633.0611   3.194585
#> 629 559.1059   2.522211
#> 630 516.8971   2.873230
#> 631 624.6532   2.850312
#> 632 500.1785   3.340909
#> 633 542.1197   2.921357
#> 634 853.8119   3.002868
#> 635 501.5102   2.680335
#> 636 505.1054   3.309213
#> 637 500.0114   3.269672
#> 638 506.5069   3.063963
#> 639 653.4299   2.586099
#> 640 500.1473   2.936620
#> 641 510.3782   3.643735
#> 642 583.8360   2.640281
#> 643 500.0675   2.752992
#> 644 501.9044   2.603880
#> 645 593.0741   2.509110
#> 646 508.3218   3.204698
#> 647 572.9922   3.160667
#> 648 500.6118   2.952240
#> 649 501.3203   3.512597
#> 650 500.0048   3.713559
#> 651 505.6727   3.189830
#> 652 500.9601   2.630621
#> 653 749.1999   3.136970
#> 654 500.7744   2.748050
#> 655 500.2095   4.133876
#> 656 500.0353   3.975300
#> 657 528.4462   2.909439
#> 658 682.8280   3.267103
#> 659 504.7638   3.067856
#> 660 504.2794   2.588241
#> 661 553.1921   2.574766
#> 662 500.2640   3.024038
#> 663 847.8842   3.532489
#> 664 500.9605   2.779316
#> 665 505.7733   4.944437
#> 666 802.8097   3.824335
#> 667 514.3078   2.977208
#> 668 501.0753   3.531013
#> 669 500.2454   2.986666
#> 670 503.4260   3.061654
#> 671 553.0522   3.495556
#> 672 504.0929   2.809161
#> 673 501.3203   2.927798
#> 674 942.6655   2.556937
#> 675 814.1090   2.580461
#> 676 537.3946   2.569683
#> 677 599.5603   2.795420
#> 678 500.2622   3.670002
#> 679 500.1472   3.244468
#> 680 500.0462   3.106288
#> 681 504.5293   2.856207
#> 682 505.8953   2.707683
#> 683 500.2301   2.910398
#> 684 504.7454   2.828094
#> 685 756.6070   3.901249
#> 686 502.0004   2.704222
#> 687 996.2219   2.596375
#> 688 500.0419   3.063832
#> 689 535.3985   2.515686
#> 690 509.0774   3.489607
#> 691 502.3441   2.655648
#> 692 667.3174   2.505718
#> 693 500.0420   2.588214
#> 694 501.6352   3.482807
#> 695 809.4499   2.797861
#> 696 500.6098   2.608514
#> 697 560.1308   2.587161
#> 698 543.3371   3.274634
#> 699 660.3714   2.531693
#> 700 508.2827   2.776321
#> 701 500.5935   2.668614
#> 702 500.0747   2.830530
#> 703 500.5495   2.605539
#> 704 520.1079   2.521804
#> 705 516.7413   2.931123
#> 706 500.1564   2.984422
#> 707 508.8904   3.999950
#> 708 505.4671   2.969549
#> 709 500.3807   2.526956
#> 710 516.1183   2.589034
#> 711 500.0295   2.806082
#> 712 500.1255   3.695964
#> 713 507.8731   2.520242
#> 714 500.7473   2.766731
#> 715 515.9556   3.877220
#> 716 943.1475   2.780786
#> 717 581.7229   2.539650
#> 718 965.5356   2.755210
#> 719 500.8326   2.549045
#> 720 500.1159   4.278266
#> 721 850.0554   2.571998
#> 722 517.1221   2.991377
#> 723 503.2320   2.974481
#> 724 500.3820   3.391495
#> 725 500.6665   2.555459
#> 726 500.2264   3.338654
#> 727 500.0805   3.195022
#> 728 597.6765   3.384956
#> 729 510.0068   2.821045
#> 730 500.2483   2.524941
#> 731 548.5956   2.570811
#> 732 500.2872   2.789840
#> 733 500.3619   2.581841
#> 734 597.8858   3.817552
#> 735 505.8470   2.828748
#> 736 500.2279   2.630447
#> 737 500.6094   3.217303
#> 738 501.0065   3.642088
#> 739 507.6239   4.657202
#> 740 500.1248   3.170208
#> 741 500.9242   3.251926
#> 742 526.5903   3.131304
#> 743 521.0556   3.484016
#> 744 501.1466   3.035272
#> 745 521.4607   2.884647
#> 746 531.4900   2.562344
#> 747 867.5526   2.557448
#> 748 500.4962   3.436371
#> 749 762.2203   2.720440
#> 750 503.8661   2.636041
#> 751 500.7407   2.915377
#> 752 500.0679   2.737553
#> 753 535.8835   2.892823
#> 754 500.0874   2.663518
#> 755 529.0487   2.540376
#> 756 641.9036   2.520200
#> 757 969.8789   2.842410
#> 758 519.2832   2.636490
#> 759 541.2999   2.509730
#> 760 501.3500   2.642812
#> 761 500.2301   2.842025
#> 762 500.5932   2.964056
#> 763 500.4391   2.609617
#> 764 500.5518   2.731008
#> 765 521.2044   2.733250
#> 766 500.2131   2.933871
#> 767 500.1552   3.207005
#> 768 516.8164   3.555449
#> 769 500.0294   2.718311
#> 770 500.4713   2.813356
#> 771 515.4319   2.818772
#> 772 522.1055   2.642332
#> 773 500.0084   3.971192
#> 774 500.2641   3.138455
#> 775 676.7812   2.880803
#> 776 630.0890   2.967622
#> 777 500.0609   2.506290
#> 778 500.1284   2.710352
#> 779 500.1380   2.768942
#> 780 540.1759   2.582113
#> 781 525.3980   2.972289
#> 782 500.1301   3.323375
#> 783 531.6496   2.751271
#> 784 506.6190   2.826838
#> 785 511.7505   4.048510
#> 786 504.3123   3.335443
#> 787 637.9754   3.677727
#> 788 500.0609   2.546124
#> 789 827.7184   2.625189
#> 790 501.1008   2.644928
#> 791 500.1722   2.880257
#> 792 500.0413   2.949531
#> 793 500.1185   2.674144
#> 794 537.8764   2.816689
#> 795 614.8569   2.523546
#> 796 579.0128   2.563032
#> 797 501.2541   2.633444
#> 798 507.4179   3.716425
#> 799 507.9581   2.880073
#> 800 511.2333   2.604892
#> 801 588.1437   2.661595
#> 802 501.4563   3.175500
#> 803 569.4731   3.583019
#> 804 513.0315   3.013329
#> 805 520.1915   2.947681
#> 806 520.0555   2.612243
#> 807 501.1584   2.703160
#> 808 500.4884   3.556575
#> 809 507.6239   2.817461
#> 810 500.6386   2.652878
#> 811 517.9196   3.181217
#> 812 536.1151   3.971005
#> 813 692.7264   4.166309
#> 814 500.2478   2.658598
#> 815 549.1126   2.610149
#> 816 512.3451   2.634972
#> 817 500.4275   2.596859
#> 818 639.3596   3.097933
#> 819 517.8475   3.200819
#> 820 505.8206   3.278254
#> 821 502.9761   3.148651
#> 822 502.8474   2.685457
#> 823 500.1511   2.852653
#> 824 501.2913   3.386621
#> 825 567.6083   2.557414
#> 826 500.1658   2.875435
#> 827 592.8305   3.225041
#> 828 855.3117   2.836933
#> 829 883.9293   3.096151
#> 830 500.1665   2.609527
#> 831 500.9265   3.496600
#> 832 503.0395   2.594537
#> 833 503.8083   2.843457
#> 834 500.0251   3.128209
#> 835 501.6746   2.508176
#> 836 500.1186   2.538142
#> 837 500.0661   2.741344
#> 838 500.1490   2.566320
#> 839 504.9051   3.483546
#> 840 500.1614   2.791884
#> 841 500.9351   2.691563
#> 842 507.1896   2.500620
#> 843 500.5781   2.508176
#> 844 706.9817   3.227426
#> 845 503.8583   2.508516
#> 846 500.8118   2.625577
#> 847 533.0314   2.866720
#> 848 500.9280   2.606240
#> 849 500.6397   3.930162
#> 850 501.6293   3.204982
#> 851 504.9042   2.794719
#> 852 500.0203   2.910939
#> 853 509.3112   2.834793
#> 854 502.7028   2.911097
#> 855 729.7494   3.134517
#> 856 522.4928   3.011864
#> 857 500.0116   2.576071
#> 858 504.4724   2.622157
#> 859 500.8683   2.650755
#> 860 500.0232   2.701276
#> 861 500.4582   2.608500
#> 862 505.4704   3.125260
#> 863 500.1484   3.662371
#> 864 501.2646   3.348512
#> 865 500.0578   2.699833
#> 866 526.6186   3.179976
#> 867 500.0283   2.998327
#> 868 505.5305   2.550955
#> 869 523.7700   3.231307
#> 870 508.9451   2.611700
#> 871 503.6170   2.982971
#> 872 500.9710   2.744785
#> 873 701.1121   2.720935
#> 874 500.3308   2.508242
#> 875 544.9505   3.168673
#> 876 501.6967   2.584234
#> 877 648.5130   3.117998
#> 878 500.5188   4.621931
#> 879 514.4288   3.561837
#> 880 500.2543   2.518430
#> 881 505.1534   2.824141
#> 882 729.8013   2.531060
#> 883 500.0444   4.299797
#> 884 524.0525   3.303187
#> 885 633.8488   2.575838
#> 886 556.9370   2.670572
#> 887 719.4416   2.516064
#> 888 501.4979   3.248031
#> 889 500.5719   2.590398
#> 890 500.6488   2.826818
#> 891 515.6885   2.513401
#> 892 525.5133   2.509395
#> 893 504.2299   2.818634
#> 894 509.2209   2.743893
#> 895 960.7014   3.901638
#> 896 500.0982   3.509827
#> 897 549.5920   2.699082
#> 898 595.1543   2.649054
#> 899 504.7267   2.562964
#> 900 501.0516   2.690967
#> 901 515.9149   3.119738
#> 902 724.9167   2.853963
#> 903 526.0737   2.906257
#> 904 500.6070   2.625726
#> 905 500.7888   2.563303
#> 906 503.2443   2.647114
#> 907 507.3745   2.562582
#> 908 500.3082   2.518899
#> 909 623.5272   2.508300
#> 910 556.3135   2.659548
#> 911 504.0551   2.501637
#> 912 503.1453   3.638450
#> 913 702.8870   2.512546
#> 914 500.0299   2.753365
#> 915 531.0884   4.360794
#> 916 501.9271   3.416107
#> 917 500.4818   2.523089
#> 918 500.0478   2.746233
#> 919 501.4659   2.542369