Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   516.8127   4.404978
#> 2   503.7640   3.257399
#> 3   503.0805   2.730029
#> 4   500.2011   2.530536
#> 5   500.2154   2.779226
#> 6   500.0612   2.593967
#> 7   534.4506   3.433146
#> 8   502.7587   2.502514
#> 9   501.3633   2.631252
#> 10  502.3383   2.572304
#> 11  610.0412   4.117650
#> 12  500.7947   2.553937
#> 13  503.4381   2.591205
#> 14  518.5937   2.626237
#> 15  778.3543   2.932288
#> 16  653.4505   2.651316
#> 17  502.2957   3.128254
#> 18  531.7094   2.869628
#> 19  500.6088   3.249734
#> 20  502.7752   4.293711
#> 21  598.4875   3.856586
#> 22  507.5108   2.887224
#> 23  675.6808   2.649926
#> 24  500.2327   2.513605
#> 25  506.1689   2.681294
#> 26  502.8777   3.376269
#> 27  501.3870   3.027355
#> 28  503.5013   2.688065
#> 29  500.4398   2.753994
#> 30  500.1282   2.666769
#> 31  522.8205   2.544360
#> 32  508.6679   2.597445
#> 33  512.6935   2.889084
#> 34  503.2746   3.656244
#> 35  502.7731   3.922432
#> 36  503.8334   2.896392
#> 37  781.3821   2.671227
#> 38  503.7638   2.552335
#> 39  500.2246   3.242929
#> 40  502.1725   2.809254
#> 41  500.0903   3.632785
#> 42  520.6760   3.519601
#> 43  502.6100   3.741277
#> 44  500.6622   2.822371
#> 45  500.0472   2.678565
#> 46  501.0714   3.722561
#> 47  602.4322   3.220247
#> 48  590.1069   2.706009
#> 49  510.0241   2.663435
#> 50  500.2847   3.133715
#> 51  501.3771   3.088935
#> 52  533.9745   2.587032
#> 53  501.2507   2.944042
#> 54  591.3309   2.546319
#> 55  566.2502   2.644685
#> 56  705.4759   2.963465
#> 57  516.1035   2.675863
#> 58  551.0817   2.593045
#> 59  500.1231   2.596789
#> 60  636.9143   2.918801
#> 61  501.2790   2.693242
#> 62  500.0577   2.914524
#> 63  500.2546   3.812593
#> 64  519.2699   2.628393
#> 65  500.9113   3.299547
#> 66  851.9883   3.077321
#> 67  501.2579   2.590755
#> 68  500.8507   2.741580
#> 69  504.4416   2.551082
#> 70  619.3222   2.838104
#> 71  500.8076   2.586056
#> 72  816.6281   2.627006
#> 73  562.4354   4.206417
#> 74  509.5882   3.252123
#> 75  504.2610   3.158078
#> 76  500.0821   4.326985
#> 77  509.9721   3.394716
#> 78  565.1832   2.500947
#> 79  500.2437   2.883237
#> 80  501.7576   3.000588
#> 81  509.8076   2.769495
#> 82  514.0113   2.588051
#> 83  503.6729   2.541192
#> 84  506.6571   3.034734
#> 85  501.6344   2.811197
#> 86  505.8863   2.773618
#> 87  830.8816   4.021630
#> 88  502.7446   2.541056
#> 89  500.0571   2.995893
#> 90  500.5044   3.849598
#> 91  502.1886   2.525620
#> 92  634.8754   2.527197
#> 93  500.8264   2.655202
#> 94  500.2763   2.983316
#> 95  942.5848   2.770475
#> 96  505.4042   2.859263
#> 97  500.5452   2.627117
#> 98  624.0698   2.621106
#> 99  500.0729   2.529686
#> 100 609.8235   2.726116
#> 101 500.6375   2.909351
#> 102 594.6337   2.508598
#> 103 500.0692   3.058401
#> 104 502.9540   2.733647
#> 105 509.1052   2.793511
#> 106 500.3764   3.362382
#> 107 504.8028   2.833192
#> 108 500.6875   2.568472
#> 109 500.1035   2.621463
#> 110 967.2789   2.724965
#> 111 500.4339   2.956051
#> 112 505.5673   3.448699
#> 113 501.5997   2.581141
#> 114 513.0010   3.049450
#> 115 500.2231   2.529203
#> 116 504.2682   3.004759
#> 117 500.1747   2.899506
#> 118 500.2280   2.968833
#> 119 502.3891   2.661883
#> 120 895.8330   2.640849
#> 121 503.4456   2.521821
#> 122 504.9396   2.714867
#> 123 504.6917   3.675853
#> 124 513.2172   2.633115
#> 125 518.6625   2.554255
#> 126 505.2660   2.588694
#> 127 550.3631   2.996422
#> 128 501.3162   2.954746
#> 129 508.4591   2.573419
#> 130 500.0491   2.996775
#> 131 501.5663   2.525854
#> 132 519.2173   3.063076
#> 133 500.0054   2.811842
#> 134 784.1417   2.865150
#> 135 505.8813   2.596373
#> 136 500.3233   2.941175
#> 137 501.0681   2.639344
#> 138 502.1519   2.788910
#> 139 500.7871   2.539589
#> 140 818.8420   3.409674
#> 141 501.0329   2.997786
#> 142 882.6152   2.560810
#> 143 500.4447   2.524677
#> 144 504.6924   3.201666
#> 145 566.7571   2.998360
#> 146 818.9993   2.711470
#> 147 563.0160   2.594318
#> 148 536.2900   2.947732
#> 149 603.0334   2.737391
#> 150 500.0071   3.827984
#> 151 521.2119   2.668628
#> 152 500.0188   2.535352
#> 153 500.0926   2.583732
#> 154 502.2090   2.619325
#> 155 500.1383   2.550677
#> 156 501.3186   4.060860
#> 157 504.6399   3.469206
#> 158 500.6689   2.525247
#> 159 500.3189   3.124942
#> 160 513.5477   2.599593
#> 161 514.2761   2.537182
#> 162 507.2542   2.665247
#> 163 518.7736   2.800288
#> 164 500.8477   2.808205
#> 165 512.0306   2.518306
#> 166 500.0330   3.113960
#> 167 602.3950   3.214213
#> 168 500.0378   2.785430
#> 169 500.3868   3.079638
#> 170 503.3011   2.998481
#> 171 971.5620   2.864646
#> 172 517.8975   2.922149
#> 173 500.2065   2.778583
#> 174 502.3615   3.028328
#> 175 501.0772   2.788426
#> 176 510.2595   3.081778
#> 177 500.9778   2.613119
#> 178 502.7769   2.876451
#> 179 507.7923   2.743559
#> 180 534.1790   3.419019
#> 181 582.0623   2.919681
#> 182 500.0573   2.503567
#> 183 527.5181   2.613479
#> 184 500.0321   4.141405
#> 185 546.0632   2.550159
#> 186 515.3594   2.503514
#> 187 507.8656   3.638397
#> 188 501.5087   3.329957
#> 189 500.1674   2.789178
#> 190 504.5648   2.534536
#> 191 573.2370   2.513811
#> 192 500.4295   2.565934
#> 193 525.5959   2.773161
#> 194 507.9685   2.956719
#> 195 546.5439   2.792627
#> 196 501.4484   3.075379
#> 197 777.2062   2.926360
#> 198 500.8308   3.669251
#> 199 560.8554   2.626021
#> 200 500.4288   4.139154
#> 201 500.2432   3.322952
#> 202 501.9045   2.954531
#> 203 603.4334   2.976213
#> 204 518.5004   2.631252
#> 205 542.5143   2.556557
#> 206 500.0175   3.877362
#> 207 770.6772   2.611997
#> 208 742.2295   2.656851
#> 209 516.5121   3.032899
#> 210 529.1524   2.533115
#> 211 787.5458   2.685914
#> 212 500.1396   3.000365
#> 213 556.5310   4.465354
#> 214 500.4108   2.505250
#> 215 822.2611   3.495327
#> 216 529.4587   2.945239
#> 217 509.6591   2.759181
#> 218 500.9109   2.506752
#> 219 572.0018   3.090158
#> 220 505.3381   2.861501
#> 221 804.4609   2.547074
#> 222 506.3163   2.899244
#> 223 506.4635   2.810058
#> 224 537.6997   2.587027
#> 225 508.3303   2.991797
#> 226 539.7949   2.993426
#> 227 515.0269   2.503958
#> 228 517.4454   2.596376
#> 229 518.1419   2.550279
#> 230 642.3748   2.545142
#> 231 517.5717   2.741153
#> 232 501.6285   3.089920
#> 233 575.8554   2.708616
#> 234 509.1861   2.607874
#> 235 503.5088   3.809174
#> 236 501.0886   2.872047
#> 237 502.3495   4.311918
#> 238 500.0395   3.081013
#> 239 500.6072   3.142144
#> 240 500.3573   3.904915
#> 241 610.7473   2.917636
#> 242 501.6062   2.607253
#> 243 501.9650   2.973553
#> 244 500.0123   3.169479
#> 245 536.3300   2.675070
#> 246 544.0687   3.544835
#> 247 500.2099   3.271848
#> 248 500.3707   3.112100
#> 249 501.3046   2.743317
#> 250 500.0213   2.726309
#> 251 575.1502   2.671521
#> 252 501.7892   2.749953
#> 253 536.9543   2.640985
#> 254 700.4288   3.974666
#> 255 500.0452   2.911090
#> 256 500.4722   2.576861
#> 257 782.9471   3.271085
#> 258 502.2352   6.565488
#> 259 534.8066   2.858679
#> 260 600.9585   3.442051
#> 261 622.5911   3.428760
#> 262 576.3228   3.443326
#> 263 696.5208   4.023791
#> 264 512.4759   2.577302
#> 265 501.6832   3.012706
#> 266 500.0570   5.099362
#> 267 505.0078   2.643757
#> 268 503.9873   2.696857
#> 269 502.6222   2.852000
#> 270 500.0760   2.881990
#> 271 500.0198   2.590309
#> 272 513.5751   2.716946
#> 273 501.0164   2.551045
#> 274 532.9148   2.964501
#> 275 506.3254   2.684390
#> 276 500.1840   2.704046
#> 277 500.0789   3.147925
#> 278 500.0031   2.677314
#> 279 500.1195   2.759762
#> 280 504.2920   3.336846
#> 281 782.8664   3.340681
#> 282 635.9061   3.590713
#> 283 776.2280   2.736434
#> 284 552.0054   2.652102
#> 285 606.5908   2.659296
#> 286 502.0416   2.872366
#> 287 623.0852   2.778537
#> 288 501.5924   3.362898
#> 289 538.7707   3.172622
#> 290 526.0699   2.531445
#> 291 500.2576   3.014666
#> 292 500.5835   3.078618
#> 293 500.9479   3.371418
#> 294 500.0514   2.635279
#> 295 548.5235   3.004906
#> 296 826.2112   2.736114
#> 297 591.7813   2.733488
#> 298 500.2410   2.500547
#> 299 619.0517   2.606334
#> 300 500.1319   3.040255
#> 301 500.8502   2.932048
#> 302 500.5356   2.587000
#> 303 500.0134   2.699076
#> 304 797.1285   3.256102
#> 305 564.7914   2.586326
#> 306 503.5821   4.272775
#> 307 529.1589   2.944670
#> 308 500.0542   4.471995
#> 309 815.4057   2.669753
#> 310 767.4660   2.912442
#> 311 517.3923   2.555017
#> 312 500.0067   3.258776
#> 313 505.8343   3.148086
#> 314 505.3683   3.313988
#> 315 590.6508   2.869161
#> 316 511.1965   2.609373
#> 317 501.4426   2.920002
#> 318 710.4798   2.625074
#> 319 977.1647   2.913226
#> 320 540.7125   2.889035
#> 321 547.3383   2.874098
#> 322 556.5640   2.702956
#> 323 503.0553   2.543449
#> 324 500.0996   3.297701
#> 325 501.3835   3.212264
#> 326 504.3333   2.583444
#> 327 500.1903   3.456331
#> 328 505.9469   2.796445
#> 329 500.1197   2.594191
#> 330 503.8732   2.579417
#> 331 769.4077   2.615334
#> 332 510.4064   2.685280
#> 333 501.0484   3.534905
#> 334 618.9477   2.753254
#> 335 514.0492   2.790971
#> 336 545.2109   2.702407
#> 337 885.3834   3.075128
#> 338 500.4680   2.545233
#> 339 579.0388   2.607301
#> 340 500.1269   2.623408
#> 341 500.0076   2.565286
#> 342 500.2664   3.292277
#> 343 500.1525   2.707497
#> 344 500.5478   2.874002
#> 345 500.5514   3.320716
#> 346 504.4750   2.914910
#> 347 500.1240   2.965758
#> 348 500.4811   3.049668
#> 349 500.1565   2.505346
#> 350 505.7580   2.502202
#> 351 500.2395   3.005470
#> 352 500.1989   3.342037
#> 353 525.0100   2.620850
#> 354 500.9172   2.734983
#> 355 737.1310   2.718951
#> 356 519.8146   3.363173
#> 357 647.8420   2.524426
#> 358 500.3105   3.370598
#> 359 506.6321   2.724234
#> 360 500.7152   2.854078
#> 361 838.5456   3.440908
#> 362 500.0694   2.578072
#> 363 501.1656   2.649668
#> 364 500.0810   3.193528
#> 365 502.4187   3.137606
#> 366 519.4612   3.173055
#> 367 500.3442   2.642493
#> 368 902.8686   2.545269
#> 369 546.2052   2.771469
#> 370 500.6229   2.713620
#> 371 599.4387   2.782806
#> 372 502.7710   2.926747
#> 373 502.6241   2.998881
#> 374 515.3678   3.067564
#> 375 500.6902   2.722892
#> 376 626.5374   3.374183
#> 377 500.3227   2.582679
#> 378 502.6995   2.999606
#> 379 823.1517   2.502251
#> 380 502.0576   2.690944
#> 381 515.6279   2.536072
#> 382 500.0752   3.035576
#> 383 500.0285   2.667695
#> 384 528.4845   3.304043
#> 385 616.2280   2.671174
#> 386 533.3797   3.254660
#> 387 525.2588   3.451287
#> 388 512.6418   3.534326
#> 389 866.3081   2.502219
#> 390 502.7823   2.917737
#> 391 539.3252   3.270494
#> 392 864.9045   2.753042
#> 393 502.1702   2.698931
#> 394 501.1890   2.583814
#> 395 527.1308   2.745395
#> 396 506.0204   2.638060
#> 397 500.8237   3.276579
#> 398 513.2839   3.376526
#> 399 521.8849   3.528356
#> 400 506.6799   3.941313
#> 401 859.0114   2.651876
#> 402 825.0169   3.699731
#> 403 514.7071   3.047811
#> 404 500.0159   3.033335
#> 405 682.2128   2.528478
#> 406 590.9703   2.991318
#> 407 500.3965   2.621056
#> 408 595.6122   3.008490
#> 409 671.2820   2.687460
#> 410 542.9287   2.667600
#> 411 630.8403   2.521392
#> 412 510.2465   3.312912
#> 413 503.4565   2.725719
#> 414 958.7282   3.249329
#> 415 501.6021   3.582076
#> 416 518.2235   2.500338
#> 417 523.4177   2.651005
#> 418 504.4543   2.608591
#> 419 529.6898   3.071966
#> 420 513.9284   2.581202
#> 421 500.7562   2.676387
#> 422 908.0633   3.057965
#> 423 500.4159   3.327387
#> 424 500.1848   2.834816
#> 425 500.5933   2.906395
#> 426 556.6900   2.580748
#> 427 708.5071   2.561389
#> 428 509.5893   3.033533
#> 429 500.0223   2.534301
#> 430 556.6892   2.630295
#> 431 500.2348   3.062091
#> 432 672.7572   2.566107
#> 433 511.8229   2.683750
#> 434 514.6759   2.528831
#> 435 502.8986   2.642769
#> 436 500.9710   3.185860
#> 437 500.0179   2.816546
#> 438 601.6070   2.516184
#> 439 500.0217   2.502973
#> 440 500.1599   3.020931
#> 441 817.1048   3.011500
#> 442 591.6678   3.304491
#> 443 556.7551   3.112114
#> 444 500.0246   2.570561
#> 445 500.8716   2.593960
#> 446 532.1803   3.181014
#> 447 501.1488   2.697016
#> 448 500.1431   2.903608
#> 449 500.3460   2.566800
#> 450 612.6959   3.113422
#> 451 500.1308   4.045765
#> 452 515.6677   2.610252
#> 453 503.0581   2.959149
#> 454 505.9782   2.708251
#> 455 511.0181   2.927006
#> 456 749.8226   2.678397
#> 457 503.3362   3.290456
#> 458 587.8726   2.572954
#> 459 834.6409   2.675053
#> 460 500.1023   2.913150
#> 461 626.6574   3.181602
#> 462 698.6101   3.058737
#> 463 500.6431   2.800331
#> 464 577.8852   2.608829
#> 465 500.0251   3.534912
#> 466 502.4090   3.503685
#> 467 551.0609   3.498955
#> 468 880.5635   3.954790
#> 469 547.6551   2.845872
#> 470 500.1442   2.828335
#> 471 500.0170   2.544150
#> 472 501.0233   3.809868
#> 473 537.3718   2.507466
#> 474 500.7712   2.848191
#> 475 702.8488   2.648144
#> 476 500.3432   2.628043
#> 477 500.0600   2.711353
#> 478 506.4800   3.456743
#> 479 501.1643   4.224205
#> 480 503.4835   2.628057
#> 481 500.5105   2.517149
#> 482 500.8175   3.396428
#> 483 503.6783   2.615927
#> 484 500.2603   2.558590
#> 485 500.5894   2.764356
#> 486 500.1004   2.735245
#> 487 500.3476   3.338670
#> 488 501.3120   2.653219
#> 489 500.1018   2.566442
#> 490 500.3609   3.602609
#> 491 500.7084   3.287931
#> 492 515.4372   2.577270
#> 493 511.0671   2.699813
#> 494 500.8151   2.554660
#> 495 589.0777   2.689280
#> 496 514.3762   2.714516
#> 497 500.5389   2.753711
#> 498 500.0574   2.639614
#> 499 506.4858   2.805263
#> 500 629.6314   2.651803
#> 501 920.0002   2.505457
#> 502 506.5858   2.879372
#> 503 518.8503   2.528907
#> 504 512.4074   3.623512
#> 505 667.2299   2.904953
#> 506 595.9613   2.684774
#> 507 505.2693   3.549310
#> 508 953.0433   2.797273
#> 509 738.9893   2.877005
#> 510 500.0894   2.649463
#> 511 525.2355   3.017563
#> 512 500.2961   2.908107
#> 513 502.0391   2.678340
#> 514 502.5299   3.112075
#> 515 500.7233   2.604859
#> 516 521.5120   2.527664
#> 517 500.0414   3.454312
#> 518 563.4836   2.541576
#> 519 502.7707   4.146163
#> 520 507.9307   2.633137
#> 521 500.3779   2.833527
#> 522 526.2194   3.199532
#> 523 534.0759   4.358088
#> 524 502.7953   3.125582
#> 525 500.5954   2.869336
#> 526 500.5387   3.058404
#> 527 512.5339   3.510874
#> 528 503.1997   3.901224
#> 529 531.2203   2.999855
#> 530 699.4539   2.735902
#> 531 500.9500   2.878828
#> 532 644.3673   3.517611
#> 533 697.5489   2.813864
#> 534 643.2152   2.549364
#> 535 502.6400   2.796504
#> 536 728.0634   4.066576
#> 537 500.1353   2.783682
#> 538 503.1721   3.443273
#> 539 814.7238   2.522689
#> 540 504.7190   2.951358
#> 541 500.0094   2.571254
#> 542 500.4990   4.640998
#> 543 501.3191   3.007693
#> 544 501.4382   4.226781
#> 545 501.1432   2.793620
#> 546 517.1667   2.743132
#> 547 501.3598   2.851410
#> 548 529.1227   3.382620
#> 549 500.5432   2.894802
#> 550 500.0360   3.438387
#> 551 500.0478   2.813093
#> 552 500.3006   2.556968
#> 553 500.7531   2.579049
#> 554 500.0069   4.000675
#> 555 500.0313   3.432317
#> 556 513.2475   3.008641
#> 557 521.9144   2.651009
#> 558 500.5453   4.348931
#> 559 500.1752   2.641085
#> 560 520.7023   3.669088
#> 561 500.1367   2.519089
#> 562 500.5942   2.669117
#> 563 506.1491   2.515381
#> 564 502.3960   3.306130
#> 565 500.0506   2.985909
#> 566 500.6433   2.637423
#> 567 503.9507   2.733022
#> 568 500.0819   2.793928
#> 569 500.9964   3.230234
#> 570 500.0469   2.587096
#> 571 501.0520   2.514919
#> 572 524.0030   2.539593
#> 573 710.7302   3.173725
#> 574 745.9601   2.609320
#> 575 500.4320   2.944534
#> 576 687.9773   3.147746
#> 577 502.5754   2.782967
#> 578 509.8538   2.545989
#> 579 547.1378   2.661599
#> 580 549.4186   2.757589
#> 581 500.0140   3.004360
#> 582 500.7822   4.366450
#> 583 526.1576   2.923737
#> 584 502.5175   3.051851
#> 585 529.5838   2.644615
#> 586 573.3980   2.601023
#> 587 546.1838   2.808213
#> 588 501.4553   2.733268
#> 589 501.9960   2.541394
#> 590 556.2280   2.940139
#> 591 530.4431   3.183125
#> 592 500.3787   2.856013
#> 593 500.0370   2.882700
#> 594 509.4703   3.467696
#> 595 502.7581   2.570290
#> 596 887.5420   2.533565
#> 597 500.0191   2.720445
#> 598 704.9651   4.312093
#> 599 500.0943   2.740078
#> 600 500.1175   3.027231
#> 601 501.8786   2.716758
#> 602 500.3354   2.780149
#> 603 509.8641   3.163694
#> 604 501.3395   2.647224
#> 605 505.4417   3.281297
#> 606 506.0987   2.777949
#> 607 507.8719   2.802116
#> 608 541.8146   2.874941
#> 609 503.9444   2.528479
#> 610 667.0154   2.504330
#> 611 500.0296   2.562793
#> 612 500.1897   3.786562
#> 613 513.6174   3.272557
#> 614 504.8641   2.753316
#> 615 513.6631   3.110050
#> 616 742.3816   3.008248
#> 617 500.0802   2.545516
#> 618 524.0078   2.819261
#> 619 502.6017   3.803646
#> 620 501.8384   3.275169
#> 621 500.0450   2.902378
#> 622 501.8571   3.016336
#> 623 973.7757   3.444393
#> 624 534.3174   2.982945
#> 625 608.8849   3.418131
#> 626 500.1830   2.999653
#> 627 503.0327   2.906292
#> 628 500.4421   2.604331
#> 629 500.0084   2.667416
#> 630 500.0422   2.545541
#> 631 666.6159   2.620073
#> 632 503.3018   3.080996
#> 633 531.4513   3.158294
#> 634 505.1215   2.861633
#> 635 672.7412   2.673237
#> 636 500.6240   3.013574
#> 637 501.5916   2.712014
#> 638 618.2194   3.345241
#> 639 500.4611   3.369016
#> 640 503.0262   3.212071
#> 641 500.0566   3.066425
#> 642 501.8405   3.068563
#> 643 699.0653   2.625402
#> 644 500.0882   2.636524
#> 645 518.2166   2.825200
#> 646 521.7636   3.180442
#> 647 512.7553   2.723712
#> 648 500.0014   2.531956
#> 649 533.9310   2.621478
#> 650 554.5170   2.809426
#> 651 519.5165   2.635141
#> 652 619.8849   2.574170
#> 653 500.3247   2.865618
#> 654 559.7959   2.525804
#> 655 500.0870   2.666386
#> 656 503.3624   2.576061
#> 657 555.3159   3.693821
#> 658 504.6192   3.459004
#> 659 500.0857   2.543865
#> 660 501.2431   3.277326
#> 661 504.6727   2.742318
#> 662 504.4847   3.248160
#> 663 809.0237   3.185796
#> 664 500.3172   2.850440
#> 665 505.1682   2.614085
#> 666 664.1533   2.599892
#> 667 503.9901   2.579056
#> 668 500.0506   2.516391
#> 669 500.8766   2.878564
#> 670 524.4448   2.979741
#> 671 500.4558   2.693657
#> 672 500.5088   3.379567
#> 673 514.0949   2.632906
#> 674 500.5599   2.605817
#> 675 537.1606   2.532047
#> 676 500.2123   3.535745
#> 677 501.4928   2.737709
#> 678 500.1096   2.957375
#> 679 784.8739   2.520835
#> 680 515.8893   4.054504
#> 681 511.8822   3.699789
#> 682 614.4530   2.794564
#> 683 500.0791   2.592888
#> 684 507.0113   2.583168
#> 685 500.1204   4.371867
#> 686 500.0585   2.658893
#> 687 916.2054   2.596312
#> 688 500.3253   2.913369
#> 689 500.5136   3.689645
#> 690 501.7926   2.623105
#> 691 554.7882   2.512524
#> 692 504.2131   3.108942
#> 693 558.5822   3.013887
#> 694 500.4592   2.608656
#> 695 509.6722   2.669166
#> 696 500.1247   3.622915
#> 697 500.0870   2.694868
#> 698 500.1488   2.801661
#> 699 511.8695   2.688679
#> 700 500.7184   4.317846
#> 701 594.8233   2.828270
#> 702 503.8710   3.734790
#> 703 510.0460   3.379327
#> 704 502.0840   2.537011
#> 705 502.4971   3.251219
#> 706 507.4176   3.133068
#> 707 500.0166   3.332975
#> 708 500.1327   3.100145
#> 709 514.7314   3.421961
#> 710 560.0527   3.855283
#> 711 535.0142   2.545876
#> 712 500.1162   2.502826
#> 713 506.4724   2.809484
#> 714 500.0356   2.729790
#> 715 501.5448   2.983394
#> 716 511.4687   2.565813
#> 717 500.2264   3.542259
#> 718 511.2775   2.819194
#> 719 500.1200   3.185110
#> 720 504.9392   3.141852
#> 721 500.0136   2.550096
#> 722 738.0203   2.596787
#> 723 500.0006   3.028564
#> 724 594.7575   2.829346
#> 725 652.5599   2.730492
#> 726 500.4122   2.789080
#> 727 500.9505   2.947055
#> 728 501.5424   2.811540
#> 729 504.7741   2.602494
#> 730 755.1827   3.064304
#> 731 739.0258   2.580998
#> 732 500.9409   3.559304
#> 733 506.6392   2.682974
#> 734 500.1179   2.979964
#> 735 512.0389   3.414379
#> 736 736.3357   2.807221
#> 737 500.0590   3.241635
#> 738 655.5558   2.500306
#> 739 501.3385   2.633738
#> 740 501.3138   4.421272
#> 741 500.1350   3.133635
#> 742 508.5369   3.030881
#> 743 535.4461   4.704492
#> 744 500.6339   2.567188
#> 745 602.5302   3.576907
#> 746 506.0091   2.786474
#> 747 500.0077   2.525035
#> 748 500.4268   2.818877
#> 749 500.1580   2.555584
#> 750 501.4171   2.715164
#> 751 502.5627   2.792229
#> 752 500.1404   2.658819
#> 753 500.1629   2.577503
#> 754 500.7533   3.122801
#> 755 503.1439   2.813137
#> 756 506.0090   2.645876
#> 757 500.5606   3.032639
#> 758 500.7936   3.134293
#> 759 504.5585   2.533037
#> 760 505.9740   2.859667
#> 761 524.3543   3.234773
#> 762 500.0466   3.169469
#> 763 570.1348   3.718048
#> 764 515.4444   2.506448
#> 765 507.4374   3.984807
#> 766 899.5614   2.650586
#> 767 508.3179   3.255302
#> 768 513.8701   2.987713
#> 769 540.9042   2.788297
#> 770 501.9138   2.992267
#> 771 520.3253   2.724414
#> 772 503.4111   2.527060
#> 773 515.1584   2.532457
#> 774 505.6077   3.238973
#> 775 887.4097   2.614652
#> 776 546.8603   2.929017
#> 777 500.9477   3.299464
#> 778 721.2157   2.922609
#> 779 502.4570   3.113413
#> 780 500.5629   2.963495
#> 781 500.9420   2.946116
#> 782 756.7673   3.148563
#> 783 519.4007   2.587636
#> 784 502.0495   3.091621
#> 785 504.6981   2.777719
#> 786 500.0397   2.589688
#> 787 502.3991   2.526200
#> 788 504.0075   3.112591
#> 789 500.5880   2.571817
#> 790 622.6454   3.147587
#> 791 503.2282   3.294823
#> 792 500.0366   2.942434
#> 793 500.2296   3.529822
#> 794 813.7390   2.767264
#> 795 503.6005   2.717944
#> 796 502.2192   2.998328
#> 797 506.8228   2.525095
#> 798 575.6820   3.291282
#> 799 502.8560   2.849088
#> 800 515.6675   2.845788
#> 801 504.1673   2.624926
#> 802 501.5748   2.714077
#> 803 500.3624   4.495026
#> 804 502.2349   2.626924
#> 805 960.9409   3.345411
#> 806 500.1687   3.154167
#> 807 500.0792   2.526061
#> 808 752.4110   3.175266
#> 809 500.0006   2.580336
#> 810 512.6079   3.199305
#> 811 533.5303   2.723239
#> 812 500.5178   2.535381
#> 813 745.1118   2.949566
#> 814 572.5677   2.683665
#> 815 718.1787   2.828011
#> 816 500.8307   3.394748
#> 817 500.3285   2.671529
#> 818 629.4562   2.872837
#> 819 518.5984   2.734331
#> 820 507.9943   2.508428
#> 821 500.1389   2.857442
#> 822 500.0044   3.428586
#> 823 534.1911   2.582742
#> 824 604.7818   2.990945
#> 825 500.0494   3.306444
#> 826 500.2425   3.051369
#> 827 500.3169   3.034839
#> 828 502.0667   2.595026
#> 829 500.0648   2.891291
#> 830 508.5167   2.662130
#> 831 507.5302   3.449588
#> 832 567.5329   2.564389
#> 833 509.1223   2.783824
#> 834 500.0398   2.836707
#> 835 502.0623   2.540404
#> 836 500.0399   2.962583
#> 837 504.0390   2.723310
#> 838 637.8905   3.281795
#> 839 500.4473   2.874010
#> 840 501.9492   2.531177
#> 841 500.0866   2.955414
#> 842 500.4779   2.745140
#> 843 520.6797   2.514975
#> 844 501.4079   2.923378
#> 845 702.2622   2.962836
#> 846 500.0282   2.892263
#> 847 501.2830   3.029204
#> 848 891.5701   3.943892
#> 849 818.2310   2.710681
#> 850 940.9769   2.832894
#> 851 500.3365   3.213019
#> 852 500.4039   2.681341
#> 853 503.9609   2.658927
#> 854 501.1977   2.600310
#> 855 555.7444   2.619880
#> 856 500.9013   2.592517
#> 857 500.4284   2.879719
#> 858 501.2571   2.928983
#> 859 500.0470   2.771831
#> 860 532.0575   2.626692
#> 861 500.1010   2.748274
#> 862 552.4157   4.425939
#> 863 504.2617   4.205346
#> 864 500.8623   2.528791
#> 865 500.0730   2.525938
#> 866 786.4259   2.627137
#> 867 502.8291   2.529033
#> 868 515.2879   3.073682
#> 869 500.0442   3.243652
#> 870 518.0146   2.598863
#> 871 500.8813   2.577977
#> 872 546.1018   2.675691
#> 873 520.1694   2.896957
#> 874 501.0066   2.715113
#> 875 561.1852   3.010073
#> 876 516.1782   3.291207
#> 877 518.9802   3.183803
#> 878 645.4219   2.832034
#> 879 511.3062   2.722222
#> 880 500.1921   2.825627
#> 881 558.8986   3.490123
#> 882 501.3855   3.226871
#> 883 500.2063   3.516703
#> 884 505.7516   2.574376
#> 885 558.7166   4.649071
#> 886 502.0688   3.197467
#> 887 500.0190   3.231708
#> 888 500.6433   2.900935
#> 889 505.5487   3.127428
#> 890 727.1521   2.962217
#> 891 583.0393   3.056631
#> 892 500.0107   2.657305
#> 893 502.3317   2.825712
#> 894 514.3783   2.956270
#> 895 501.4239   2.935467
#> 896 500.7000   2.827901
#> 897 512.7969   3.022938
#> 898 500.0525   2.793824
#> 899 774.8616   2.932009
#> 900 501.4047   2.901644
#> 901 667.1321   3.134572
#> 902 508.5664   2.637973
#> 903 500.1129   2.629916
#> 904 799.6708   3.264953
#> 905 568.3893   2.561218
#> 906 500.0234   2.653213
#> 907 569.6091   2.663019
#> 908 503.0308   2.877583
#> 909 500.5315   2.785788
#> 910 502.1154   3.826547
#> 911 500.1033   3.111010
#> 912 500.0652   2.550863
#> 913 500.0776   2.807206
#> 914 518.3456   3.369521
#> 915 502.0034   3.071455
#> 916 500.0994   3.322483
#> 917 506.4036   2.543305
#> 918 500.9650   3.599766
#> 919 561.3502   2.590974
#> 920 500.0715   2.679387
#> 921 887.1680   2.606581
#> 922 502.8083   2.592372