Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   500.4115   2.675800
#> 2   501.7278   2.735421
#> 3   500.1565   3.007948
#> 4   828.4857   2.720913
#> 5   504.4757   3.269020
#> 6   584.0728   2.694462
#> 7   560.2062   3.158375
#> 8   500.7977   3.356217
#> 9   885.1256   3.200761
#> 10  502.5360   3.052376
#> 11  512.6010   3.137443
#> 12  500.1032   2.958667
#> 13  501.6572   3.686672
#> 14  500.4377   2.535179
#> 15  589.0466   4.535544
#> 16  511.0327   2.772302
#> 17  540.5590   2.629203
#> 18  500.0297   3.272606
#> 19  505.2330   3.036383
#> 20  507.0058   2.977167
#> 21  875.7524   3.156177
#> 22  769.8608   2.769196
#> 23  500.8111   2.519284
#> 24  502.3551   2.592466
#> 25  500.4557   2.771901
#> 26  693.8037   2.759506
#> 27  501.6588   3.698309
#> 28  502.8192   2.799100
#> 29  500.4488   2.732002
#> 30  504.2236   2.555304
#> 31  500.6041   3.483802
#> 32  608.8651   3.077043
#> 33  500.3718   2.576404
#> 34  500.2431   3.792296
#> 35  500.5299   4.048087
#> 36  500.0823   3.057230
#> 37  519.3888   2.616355
#> 38  527.2288   3.746373
#> 39  862.2777   3.301119
#> 40  513.2504   3.037758
#> 41  639.1991   2.711886
#> 42  500.3211   2.594856
#> 43  501.2875   3.056610
#> 44  501.2991   2.548128
#> 45  500.4179   2.826880
#> 46  519.2552   2.637575
#> 47  500.6598   3.244934
#> 48  501.7222   2.888848
#> 49  555.2076   3.026346
#> 50  502.8774   2.661845
#> 51  639.5869   3.673031
#> 52  500.1497   2.908984
#> 53  500.0926   2.907582
#> 54  500.2434   2.811565
#> 55  524.3026   2.528454
#> 56  500.5203   2.512126
#> 57  600.7389   2.718000
#> 58  527.1483   2.596356
#> 59  522.2819   3.409576
#> 60  620.9206   2.873873
#> 61  500.9855   2.725165
#> 62  503.7980   2.860150
#> 63  502.0626   3.387587
#> 64  916.9483   2.873618
#> 65  504.1818   2.544307
#> 66  500.0415   2.859441
#> 67  567.3876   3.525217
#> 68  541.1744   2.687066
#> 69  504.6512   2.786444
#> 70  613.2602   2.687012
#> 71  538.4411   2.563885
#> 72  747.2411   2.516350
#> 73  500.0317   2.960652
#> 74  622.0527   3.689371
#> 75  502.0146   2.672993
#> 76  540.9059   2.829352
#> 77  685.9297   2.958478
#> 78  511.2983   3.510537
#> 79  511.7433   3.377979
#> 80  500.2499   3.534063
#> 81  565.1585   2.643565
#> 82  964.2774   4.196790
#> 83  540.3660   2.950037
#> 84  501.0646   2.500095
#> 85  963.3856   2.693104
#> 86  503.1093   2.582407
#> 87  500.2347   2.611006
#> 88  503.2022   2.566536
#> 89  501.7829   2.588610
#> 90  688.7009   2.953881
#> 91  500.4651   2.726287
#> 92  500.0779   3.079299
#> 93  512.6363   2.851263
#> 94  657.8746   2.595031
#> 95  501.1274   2.622047
#> 96  543.3673   3.372556
#> 97  501.0272   3.255655
#> 98  538.5331   2.794436
#> 99  500.6467   2.565485
#> 100 500.3358   3.024852
#> 101 500.3167   2.947738
#> 102 604.9063   2.916687
#> 103 531.6192   2.760210
#> 104 612.0310   3.140536
#> 105 500.4233   2.510406
#> 106 500.9221   3.020441
#> 107 504.1752   3.244189
#> 108 565.6967   2.605548
#> 109 500.6044   3.366677
#> 110 501.0160   2.581820
#> 111 500.3384   3.115728
#> 112 523.0088   2.777876
#> 113 500.2709   3.596843
#> 114 500.7442   2.881457
#> 115 500.2993   2.605790
#> 116 500.0254   2.703123
#> 117 503.6837   3.258372
#> 118 500.2814   2.802091
#> 119 502.1149   2.666360
#> 120 536.4513   3.160188
#> 121 505.1939   2.527381
#> 122 500.0427   2.550085
#> 123 500.0308   2.570666
#> 124 551.4018   2.605805
#> 125 503.3737   2.808758
#> 126 500.7039   2.810969
#> 127 500.1516   2.503467
#> 128 500.6414   2.992009
#> 129 500.6014   2.928890
#> 130 897.9064   2.597459
#> 131 505.9083   2.951571
#> 132 500.0094   2.592988
#> 133 500.0075   2.836894
#> 134 689.4708   2.657158
#> 135 504.3750   2.899772
#> 136 501.3773   3.623683
#> 137 743.1446   3.351987
#> 138 516.6321   2.522341
#> 139 500.3398   2.587354
#> 140 502.9627   2.788415
#> 141 506.8996   3.436149
#> 142 500.1226   2.617902
#> 143 500.0587   2.754691
#> 144 503.7742   2.669095
#> 145 539.8794   4.084750
#> 146 500.0147   2.675906
#> 147 517.7462   2.667699
#> 148 500.2006   2.823042
#> 149 503.1902   3.070132
#> 150 547.8944   2.575100
#> 151 502.6847   3.404577
#> 152 502.5133   3.207459
#> 153 525.7288   2.880035
#> 154 572.9845   2.786915
#> 155 566.1602   2.862001
#> 156 500.3906   2.563002
#> 157 500.0631   4.989202
#> 158 511.4272   2.979173
#> 159 530.5972   2.547686
#> 160 500.1733   3.125728
#> 161 584.9975   2.590176
#> 162 551.3575   3.042973
#> 163 514.7340   2.515424
#> 164 501.5016   2.733688
#> 165 773.3294   2.710594
#> 166 500.0201   2.532355
#> 167 584.3781   3.030789
#> 168 649.7379   2.893969
#> 169 528.7351   3.179303
#> 170 504.9706   2.615203
#> 171 501.4609   3.108349
#> 172 559.7776   2.689065
#> 173 500.1710   2.585033
#> 174 500.0650   2.780429
#> 175 500.7941   3.913776
#> 176 500.0845   2.936342
#> 177 585.5146   3.689419
#> 178 503.2242   3.108287
#> 179 507.8065   3.276437
#> 180 501.0806   2.571097
#> 181 500.0271   2.628006
#> 182 502.6410   2.742822
#> 183 501.0626   2.966001
#> 184 510.8906   5.063634
#> 185 613.5169   2.899272
#> 186 525.8502   2.949893
#> 187 500.3306   2.804807
#> 188 500.0872   2.716361
#> 189 701.8488   2.559734
#> 190 561.9227   2.685268
#> 191 505.5759   2.590665
#> 192 522.8952   2.998865
#> 193 713.3533   2.555359
#> 194 531.2512   2.683340
#> 195 501.1105   2.572214
#> 196 507.4877   2.993986
#> 197 501.6532   3.046503
#> 198 544.8712   2.838664
#> 199 613.2138   2.606567
#> 200 500.6230   2.757901
#> 201 500.0481   3.164049
#> 202 500.0948   2.873767
#> 203 522.5665   2.835334
#> 204 504.0870   2.755992
#> 205 502.4621   2.535322
#> 206 529.2693   2.650726
#> 207 506.6172   4.374153
#> 208 501.1366   3.874912
#> 209 500.3778   2.816096
#> 210 500.3339   3.065446
#> 211 582.7649   3.903565
#> 212 500.4640   2.538534
#> 213 506.4233   2.679158
#> 214 501.7931   3.142552
#> 215 500.4416   2.568130
#> 216 584.1282   2.501219
#> 217 500.1381   2.691254
#> 218 500.2752   2.800200
#> 219 500.0398   3.122140
#> 220 500.0398   2.515168
#> 221 627.5554   3.022155
#> 222 500.8497   2.726757
#> 223 502.3041   3.073026
#> 224 627.8313   2.523131
#> 225 500.1394   2.769138
#> 226 711.0903   3.049148
#> 227 539.0036   2.563995
#> 228 560.9659   2.534867
#> 229 829.2429   2.998695
#> 230 500.1821   3.135507
#> 231 522.2220   2.750311
#> 232 729.1458   2.704441
#> 233 796.8924   2.938148
#> 234 500.6439   3.019775
#> 235 515.0191   2.894404
#> 236 786.7314   2.588919
#> 237 501.9464   2.771844
#> 238 500.2616   2.949062
#> 239 878.1948   2.672697
#> 240 503.4690   3.224327
#> 241 512.8138   2.593821
#> 242 548.9580   3.691850
#> 243 516.4232   2.606383
#> 244 502.6869   2.527708
#> 245 503.4509   2.665694
#> 246 504.8242   2.656445
#> 247 503.8914   2.577494
#> 248 500.8886   2.846253
#> 249 579.8139   3.007354
#> 250 510.7243   2.574950
#> 251 500.8628   2.668319
#> 252 500.0493   3.276654
#> 253 500.3280   2.845645
#> 254 501.9202   3.044616
#> 255 500.0392   3.113765
#> 256 509.0216   2.745123
#> 257 526.9831   2.961447
#> 258 501.2724   2.806209
#> 259 507.0074   2.830009
#> 260 501.4057   2.783671
#> 261 500.7672   2.952420
#> 262 734.4194   2.870341
#> 263 525.1568   3.575927
#> 264 500.3427   2.851849
#> 265 500.0041   4.048183
#> 266 501.9524   3.034939
#> 267 500.0292   2.870436
#> 268 502.8281   4.782495
#> 269 500.1527   3.135196
#> 270 501.1045   2.853647
#> 271 538.1339   3.247657
#> 272 538.9037   3.112044
#> 273 515.1794   2.521870
#> 274 517.4954   3.238190
#> 275 500.1014   2.893902
#> 276 500.3038   4.700184
#> 277 500.0041   2.514467
#> 278 500.4027   2.673713
#> 279 500.1902   3.077723
#> 280 635.6877   2.586091
#> 281 564.8130   2.634672
#> 282 948.0140   3.321507
#> 283 600.7418   2.770609
#> 284 626.0891   2.709476
#> 285 524.9781   3.182520
#> 286 504.2730   3.265859
#> 287 501.6177   2.648058
#> 288 571.3225   2.638858
#> 289 506.0688   3.007200
#> 290 603.6522   2.523101
#> 291 500.4120   2.821839
#> 292 501.7608   2.768450
#> 293 500.0877   2.568672
#> 294 500.0176   2.581129
#> 295 500.3017   2.511475
#> 296 539.6511   2.648710
#> 297 500.2701   3.060510
#> 298 953.7999   3.194788
#> 299 644.5207   2.819102
#> 300 500.3246   3.094875
#> 301 500.1168   3.356259
#> 302 500.0654   2.887098
#> 303 501.9188   2.537946
#> 304 500.0032   2.684748
#> 305 504.3890   2.721514
#> 306 719.9524   3.199256
#> 307 568.5273   2.853321
#> 308 501.8484   2.596944
#> 309 500.0708   3.010593
#> 310 518.6392   2.724406
#> 311 501.0037   2.785828
#> 312 515.6106   2.766628
#> 313 565.0049   2.973940
#> 314 600.2879   3.608625
#> 315 961.3134   2.510578
#> 316 551.7629   2.778325
#> 317 759.3152   2.946009
#> 318 534.7237   2.789639
#> 319 500.6660   2.905140
#> 320 502.5203   2.584070
#> 321 500.0826   3.692599
#> 322 500.0817   2.562065
#> 323 502.4760   3.427998
#> 324 515.4894   2.963905
#> 325 500.4266   2.723205
#> 326 503.3824   3.136629
#> 327 500.0116   2.595042
#> 328 519.6452   2.653839
#> 329 500.2669   2.838380
#> 330 551.4883   2.954766
#> 331 503.0091   2.714122
#> 332 500.5035   2.581197
#> 333 505.3160   3.100336
#> 334 518.2877   3.697653
#> 335 502.7081   2.578207
#> 336 500.3709   2.682353
#> 337 504.4427   2.843378
#> 338 509.7577   2.988759
#> 339 604.3838   2.622778
#> 340 501.6530   2.837350
#> 341 520.8546   2.791904
#> 342 500.0006   2.601290
#> 343 505.3490   2.522212
#> 344 500.4303   2.537164
#> 345 500.1416   2.529575
#> 346 508.9776   3.590604
#> 347 928.9156   2.634661
#> 348 501.4980   2.646237
#> 349 535.5345   2.607390
#> 350 500.1013   3.725516
#> 351 500.0078   3.255564
#> 352 958.3429   2.796285
#> 353 500.5539   3.269587
#> 354 735.6211   2.743769
#> 355 500.6248   3.039971
#> 356 536.8920   3.899643
#> 357 500.1302   3.756570
#> 358 500.2002   2.875685
#> 359 533.4088   2.682340
#> 360 561.7406   2.711468
#> 361 507.2848   2.512666
#> 362 928.6239   2.610838
#> 363 790.0849   2.967143
#> 364 851.9435   2.583630
#> 365 500.5799   2.554688
#> 366 500.6463   2.702576
#> 367 517.9816   4.015040
#> 368 500.3332   2.774284
#> 369 502.5306   3.520621
#> 370 500.0133   3.012029
#> 371 511.4093   2.565943
#> 372 500.0681   2.751224
#> 373 500.7473   2.546544
#> 374 501.1139   3.085616
#> 375 929.6832   2.894251
#> 376 500.3079   3.398821
#> 377 505.6383   2.827505
#> 378 500.7677   2.787750
#> 379 500.0232   2.647506
#> 380 503.6450   2.674070
#> 381 757.3974   2.841201
#> 382 983.4714   2.708477
#> 383 500.0678   3.806193
#> 384 500.7950   2.710208
#> 385 500.5138   2.858598
#> 386 563.9384   2.563334
#> 387 501.5749   3.340236
#> 388 513.0810   3.187978
#> 389 531.0089   2.571283
#> 390 502.8494   2.638818
#> 391 502.7411   2.570997
#> 392 500.3730   2.844660
#> 393 998.8255   2.798641
#> 394 500.1357   3.369532
#> 395 512.1640   3.088007
#> 396 500.3076   2.816477
#> 397 574.9412   2.526669
#> 398 505.0004   2.583056
#> 399 555.3555   2.569792
#> 400 500.1356   3.033957
#> 401 500.3885   3.010650
#> 402 529.3891   3.890800
#> 403 503.2030   2.542284
#> 404 500.7793   2.613591
#> 405 500.7969   3.862790
#> 406 530.2224   2.510314
#> 407 501.5934   3.342887
#> 408 501.1230   2.636874
#> 409 500.0409   3.440776
#> 410 500.4492   2.540498
#> 411 500.0063   3.211821
#> 412 516.9495   2.594551
#> 413 500.0842   2.706371
#> 414 502.9035   2.647856
#> 415 658.9821   2.654442
#> 416 507.4870   2.500555
#> 417 500.9385   2.709712
#> 418 526.7721   2.975663
#> 419 558.7049   3.098906
#> 420 536.2048   2.530455
#> 421 855.4169   3.778282
#> 422 548.9383   3.190608
#> 423 518.1459   2.642318
#> 424 597.6053   2.535854
#> 425 500.7588   3.539907
#> 426 500.0527   2.845104
#> 427 500.2708   2.620188
#> 428 517.4040   3.040552
#> 429 503.3558   2.567766
#> 430 509.8335   2.501804
#> 431 500.1290   2.569074
#> 432 500.0078   2.810393
#> 433 500.9519   2.672880
#> 434 500.0059   3.336948
#> 435 500.1721   2.927505
#> 436 508.0137   3.686151
#> 437 500.3075   2.628214
#> 438 521.6056   2.802945
#> 439 500.0219   2.831120
#> 440 501.8173   2.784299
#> 441 514.9370   2.654740
#> 442 503.0441   2.577355
#> 443 507.1848   2.512551
#> 444 500.0892   2.508665
#> 445 500.2912   3.606082
#> 446 500.4934   2.726503
#> 447 752.3506   2.792995
#> 448 586.4527   2.625120
#> 449 584.9172   3.030170
#> 450 543.1161   2.725885
#> 451 500.1822   2.914979
#> 452 511.5726   2.646806
#> 453 502.5770   2.552229
#> 454 500.0038   3.514007
#> 455 522.1060   2.650407
#> 456 507.1164   2.540865
#> 457 500.4203   2.729389
#> 458 956.9040   2.766859
#> 459 502.5044   2.917724
#> 460 538.5727   4.127896
#> 461 503.8489   3.289678
#> 462 500.1644   3.076530
#> 463 501.2145   2.552377
#> 464 518.9433   2.579427
#> 465 500.2277   2.730062
#> 466 505.0632   2.567107
#> 467 505.6430   2.610073
#> 468 818.6252   2.772770
#> 469 504.8877   2.555648
#> 470 500.1483   2.787203
#> 471 502.9650   3.469269
#> 472 500.7410   2.693961
#> 473 502.1727   3.287622
#> 474 501.1876   2.803032
#> 475 500.8550   2.511690
#> 476 648.5937   2.909097
#> 477 536.7880   2.541271
#> 478 543.2578   4.783708
#> 479 500.7575   2.573598
#> 480 501.3594   2.911831
#> 481 503.6296   2.640234
#> 482 503.5585   3.106887
#> 483 695.0321   3.100242
#> 484 506.1125   2.835285
#> 485 833.9440   2.771059
#> 486 576.7856   2.940701
#> 487 504.5177   2.504658
#> 488 500.2723   2.536608
#> 489 500.7344   2.843525
#> 490 502.3981   2.533754
#> 491 501.3896   2.765365
#> 492 501.0851   2.764104
#> 493 770.2055   2.794722
#> 494 501.9539   2.673556
#> 495 502.5428   2.757415
#> 496 529.0375   2.615523
#> 497 513.4926   3.690126
#> 498 500.2772   3.105190
#> 499 500.0707   2.608794
#> 500 557.4739   3.192478
#> 501 500.1723   2.688705
#> 502 500.0271   2.870525
#> 503 501.7843   3.047493
#> 504 500.6569   2.814425
#> 505 500.2224   2.842720
#> 506 547.6340   2.597390
#> 507 500.9592   2.854344
#> 508 525.5192   2.854551
#> 509 871.2766   3.373769
#> 510 504.2709   3.330242
#> 511 500.7559   2.575743
#> 512 500.1075   2.565359
#> 513 500.0462   2.519570
#> 514 500.3690   2.515102
#> 515 821.8877   3.008881
#> 516 500.5186   3.365029
#> 517 503.3045   2.658670
#> 518 514.2489   3.258769
#> 519 873.5672   2.570770
#> 520 524.6676   3.382553
#> 521 504.0363   4.078146
#> 522 500.0157   2.873368
#> 523 501.0031   3.406656
#> 524 507.0906   2.910057
#> 525 502.5410   2.715246
#> 526 561.9314   2.845890
#> 527 710.6997   3.099288
#> 528 513.6544   2.877395
#> 529 534.8354   2.744840
#> 530 500.3151   3.193131
#> 531 521.1494   3.196115
#> 532 512.4769   2.583853
#> 533 500.0451   2.597545
#> 534 643.8197   2.667300
#> 535 500.5572   3.662211
#> 536 500.2389   4.071261
#> 537 781.6124   2.712608
#> 538 660.8908   2.890428
#> 539 510.4434   2.789711
#> 540 500.0204   2.576934
#> 541 506.7302   2.620858
#> 542 500.0214   2.686616
#> 543 824.5862   3.277875
#> 544 500.0155   3.001533
#> 545 618.3102   3.476038
#> 546 500.1186   3.108287
#> 547 500.1797   2.674980
#> 548 502.2489   4.791502
#> 549 500.4842   2.568251
#> 550 502.6961   3.025497
#> 551 539.6553   2.739920
#> 552 558.1673   3.192248
#> 553 518.4675   2.949317
#> 554 500.1754   2.829446
#> 555 501.7502   3.055530
#> 556 503.5665   3.473054
#> 557 500.3795   2.919931
#> 558 502.8177   2.729041
#> 559 502.3476   2.501466
#> 560 500.2874   2.614370
#> 561 500.1185   2.667942
#> 562 500.2057   3.425313
#> 563 500.0290   2.500862
#> 564 508.2240   2.914833
#> 565 500.1229   2.900380
#> 566 500.8121   2.577687
#> 567 525.0204   3.116452
#> 568 501.6103   2.942292
#> 569 670.9916   2.623568
#> 570 531.9978   3.329588
#> 571 502.0361   2.875235
#> 572 501.0267   2.546356
#> 573 601.0330   2.509677
#> 574 507.8756   2.510097
#> 575 500.0750   2.677886
#> 576 500.8902   3.248901
#> 577 586.2789   2.568371
#> 578 528.8091   2.549535
#> 579 500.0003   2.794043
#> 580 500.0843   2.557950
#> 581 500.0017   2.779075
#> 582 503.0333   2.675663
#> 583 500.1256   2.583299
#> 584 501.2108   2.608019
#> 585 554.1544   2.590539
#> 586 500.5109   2.999617
#> 587 500.8054   2.609084
#> 588 500.0487   2.549782
#> 589 811.4013   2.810362
#> 590 500.0821   2.643076
#> 591 505.8616   3.661457
#> 592 764.6838   3.705092
#> 593 516.9157   2.564117
#> 594 612.8453   2.668391
#> 595 517.7706   2.862749
#> 596 505.8328   2.817715
#> 597 500.0895   2.684283
#> 598 508.0526   3.308050
#> 599 500.1793   2.904123
#> 600 544.1833   2.555460
#> 601 749.1281   2.770214
#> 602 693.1479   2.962807
#> 603 606.9539   3.412027
#> 604 500.0563   2.789456
#> 605 526.7843   2.735813
#> 606 500.0394   3.539253
#> 607 503.6191   3.522412
#> 608 500.8472   2.924351
#> 609 502.6025   2.913043
#> 610 504.6835   3.752058
#> 611 507.5162   3.161417
#> 612 950.1043   3.347822
#> 613 502.3809   2.543706
#> 614 500.0665   2.955401
#> 615 504.7309   3.133416
#> 616 500.1229   2.635982
#> 617 500.4549   3.127028
#> 618 503.7816   2.574996
#> 619 501.4291   2.947259
#> 620 676.4446   2.773836
#> 621 500.0606   2.832691
#> 622 512.5080   2.944278
#> 623 510.4372   2.633288
#> 624 510.2841   4.202440
#> 625 516.5858   2.943970
#> 626 500.6234   2.724173
#> 627 504.9579   3.262287
#> 628 518.8778   2.742651
#> 629 505.6483   2.552457
#> 630 500.3486   3.755510
#> 631 500.0670   2.886693
#> 632 521.6278   2.706837
#> 633 502.6559   3.193575
#> 634 500.1309   2.765020
#> 635 503.7479   2.579247
#> 636 643.2968   2.576460
#> 637 503.3781   2.609092
#> 638 588.4280   2.625425
#> 639 506.8997   2.956244
#> 640 500.0560   2.866297
#> 641 523.6293   3.055746
#> 642 502.7032   2.596412
#> 643 808.5267   2.502103
#> 644 500.0045   2.865651
#> 645 500.1223   4.511530
#> 646 668.5588   2.699690
#> 647 948.4992   3.104267
#> 648 510.2024   2.947951
#> 649 500.1502   2.534200
#> 650 506.8566   2.537956
#> 651 500.4037   2.560943
#> 652 529.7182   2.676969
#> 653 502.1746   2.667764
#> 654 531.3759   2.902154
#> 655 617.0992   3.465868
#> 656 506.1908   2.666046
#> 657 500.2094   2.521216
#> 658 709.1614   2.831658
#> 659 584.5268   2.604467
#> 660 525.9408   2.517289
#> 661 501.5567   3.040694
#> 662 500.0172   3.393541
#> 663 703.0707   3.446486
#> 664 500.9448   3.447598
#> 665 502.8813   2.617463
#> 666 510.5245   2.607480
#> 667 500.0952   3.001497
#> 668 611.1584   2.717232
#> 669 513.5492   3.472828
#> 670 503.6716   2.712420
#> 671 536.2598   2.510020
#> 672 513.4680   3.285477
#> 673 500.6943   2.709056
#> 674 500.7504   2.927052
#> 675 500.1268   2.571160
#> 676 511.2754   3.266056
#> 677 501.8250   2.550806
#> 678 502.4420   3.590065
#> 679 500.0053   2.815290
#> 680 512.4364   2.802723
#> 681 501.5737   2.702333
#> 682 500.2366   3.148614
#> 683 505.9513   2.701077
#> 684 502.6188   3.107723
#> 685 785.1802   2.657007
#> 686 500.1845   2.884473
#> 687 500.9780   2.732835
#> 688 502.9706   3.021026
#> 689 592.4608   3.280970
#> 690 507.8013   2.648249
#> 691 536.4288   2.951606
#> 692 500.1747   3.457160
#> 693 501.6873   2.664169
#> 694 506.3185   3.058236
#> 695 663.9923   2.953908
#> 696 531.4412   2.838065
#> 697 534.2859   4.450651
#> 698 640.2608   2.670970
#> 699 500.1237   2.601397
#> 700 500.1433   3.482617
#> 701 598.8562   2.538949
#> 702 749.3530   3.370833
#> 703 500.5067   3.183832
#> 704 504.4903   3.016671
#> 705 500.3642   2.551489
#> 706 535.5805   2.505375
#> 707 637.6515   2.611314
#> 708 501.3748   2.931479
#> 709 557.3819   3.197732
#> 710 506.9761   4.218003
#> 711 511.2688   3.075726
#> 712 500.0089   2.717340
#> 713 500.7346   3.354327
#> 714 579.0208   2.523576
#> 715 500.9198   2.810899
#> 716 744.6705   3.030830
#> 717 500.7067   3.107639
#> 718 500.4673   3.096878
#> 719 500.4409   2.588659
#> 720 504.2760   3.964454
#> 721 500.7372   2.529885
#> 722 545.2747   4.085350
#> 723 517.3903   3.210893
#> 724 500.0780   2.604361
#> 725 502.8028   2.840160
#> 726 524.5213   3.441980
#> 727 501.4883   2.858648
#> 728 501.3011   2.749455
#> 729 500.5038   2.516900
#> 730 500.3099   3.848364
#> 731 856.2740   4.150764
#> 732 550.6330   2.592212
#> 733 504.3223   2.604596
#> 734 500.0017   3.152420
#> 735 506.5429   2.663603
#> 736 500.0561   4.900237
#> 737 501.1822   2.962339
#> 738 516.1769   2.656450
#> 739 501.6582   2.655605
#> 740 501.0132   2.620255
#> 741 528.9660   2.740820
#> 742 500.2665   4.388275
#> 743 575.8998   2.771323
#> 744 543.6970   2.834938
#> 745 506.2265   3.083767
#> 746 510.0336   3.308279
#> 747 590.4766   2.778272
#> 748 500.2320   2.989184
#> 749 528.5487   2.874087
#> 750 500.6094   2.531500
#> 751 500.2068   2.582461
#> 752 578.8038   2.534154
#> 753 512.4281   2.539756
#> 754 502.8908   2.975473
#> 755 506.2236   3.040854
#> 756 500.6708   3.016803
#> 757 868.7337   3.453063
#> 758 650.5360   2.950858
#> 759 500.9261   2.585863
#> 760 500.0069   2.856523
#> 761 955.8855   2.520732
#> 762 500.3820   3.913144
#> 763 500.1328   5.300353
#> 764 500.8278   4.302822
#> 765 783.7107   5.541524
#> 766 546.2007   3.019461
#> 767 501.5607   2.616079
#> 768 500.0146   2.577052
#> 769 546.8099   3.286308
#> 770 645.0797   2.593248
#> 771 501.5098   2.542471
#> 772 521.3319   2.879871
#> 773 960.0032   2.619999
#> 774 518.3101   2.610412
#> 775 550.4262   2.515480
#> 776 758.3540   2.722676
#> 777 506.7086   3.601989
#> 778 500.2069   2.909800
#> 779 500.2451   2.598349
#> 780 507.8207   2.697325
#> 781 500.3705   2.530432
#> 782 505.8546   2.634311
#> 783 509.7048   3.522547
#> 784 593.0509   3.347223
#> 785 500.2529   2.680093
#> 786 502.4119   2.641877
#> 787 500.0711   2.700406
#> 788 500.1389   3.116507
#> 789 500.0719   3.829807
#> 790 500.1780   2.972388
#> 791 500.0716   2.520476
#> 792 692.4816   2.706579
#> 793 520.9273   3.628936
#> 794 501.1979   2.953150
#> 795 500.0957   3.065568
#> 796 673.7821   2.743107
#> 797 500.0469   3.096818
#> 798 500.7073   2.689305
#> 799 500.4046   3.012348
#> 800 581.1075   2.589581
#> 801 717.8754   2.746364
#> 802 506.7849   2.832028
#> 803 660.5825   2.714468
#> 804 506.6302   2.852207
#> 805 500.0944   3.588240
#> 806 500.0980   2.767660
#> 807 503.8452   2.529760
#> 808 500.0359   3.422018
#> 809 500.9223   2.851965
#> 810 521.5426   2.623827
#> 811 653.4763   2.890568
#> 812 513.6592   2.651607
#> 813 508.0827   2.659736
#> 814 501.0132   3.601622
#> 815 537.6142   2.560262
#> 816 512.5927   2.852893
#> 817 506.7019   3.676856
#> 818 502.4086   2.858189
#> 819 636.1742   2.838209
#> 820 523.8786   2.911473
#> 821 546.9973   2.704256
#> 822 811.4592   4.186930
#> 823 502.7421   2.605631
#> 824 500.4111   2.924852
#> 825 506.0562   2.777324
#> 826 500.5339   2.963160
#> 827 500.4990   3.292374
#> 828 501.1567   2.914097
#> 829 500.3597   2.556612
#> 830 557.7314   2.973383
#> 831 523.3542   2.896475
#> 832 521.6537   3.711232
#> 833 500.3730   2.603001
#> 834 500.2114   2.739709
#> 835 500.3843   3.866630
#> 836 556.6455   2.782169
#> 837 502.7345   2.829279
#> 838 500.2333   2.794970
#> 839 617.8303   3.063996
#> 840 500.1669   2.547272
#> 841 510.7259   2.505417
#> 842 500.2390   2.845965
#> 843 505.7899   2.925648
#> 844 513.7519   2.637657
#> 845 554.7760   3.435727
#> 846 500.2124   2.578798
#> 847 502.5119   2.596787
#> 848 501.0724   2.662625
#> 849 501.0351   3.239996
#> 850 504.1596   3.414102
#> 851 518.3435   3.538604
#> 852 500.0002   2.812127
#> 853 502.7798   3.071287
#> 854 500.7633   2.746827
#> 855 529.2983   2.946250
#> 856 500.6676   2.796432
#> 857 500.0770   2.627593
#> 858 746.7911   2.766361
#> 859 500.0807   3.457031
#> 860 501.1456   2.890034
#> 861 521.4420   4.887519
#> 862 514.4815   3.399969
#> 863 501.5987   2.569969
#> 864 507.5185   2.938975
#> 865 500.0621   3.026262
#> 866 500.0473   3.704274
#> 867 500.4134   2.637551
#> 868 551.9876   2.716852
#> 869 500.7888   2.576101
#> 870 500.0339   3.068431
#> 871 500.2668   3.748123
#> 872 500.7200   3.022649
#> 873 500.6208   2.700780
#> 874 501.9676   2.932343
#> 875 546.6328   2.672404
#> 876 502.2216   2.617265
#> 877 507.2224   3.012202
#> 878 907.0814   2.771984
#> 879 500.8947   2.555150
#> 880 791.7874   2.700009
#> 881 500.0175   2.901865
#> 882 687.0980   3.134001
#> 883 614.5119   2.641821
#> 884 503.5646   3.187194
#> 885 500.7940   2.712539
#> 886 500.9779   2.647755
#> 887 501.8642   2.540842
#> 888 562.2829   2.560353
#> 889 501.1257   3.073723
#> 890 634.6350   3.123336
#> 891 675.0688   2.571248
#> 892 508.3422   2.863126
#> 893 500.0845   2.678934
#> 894 575.9243   2.615059
#> 895 500.0252   3.190855
#> 896 500.0115   3.084163
#> 897 652.9091   3.769669
#> 898 500.2108   2.520869
#> 899 501.1187   2.674049
#> 900 546.6107   2.758925
#> 901 500.0581   3.137532
#> 902 501.0608   3.106399
#> 903 864.1151   2.906805
#> 904 588.2530   2.579004
#> 905 502.4405   2.582140
#> 906 949.0444   2.866622
#> 907 553.3287   2.516170
#> 908 690.4320   2.896189
#> 909 500.0014   2.968505
#> 910 503.5574   2.595440
#> 911 500.7101   2.548239
#> 912 500.0279   2.932466
#> 913 646.1144   3.085262
#> 914 500.6762   2.705513
#> 915 991.9560   2.807699
#> 916 696.6936   2.692080