Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   930.0734   2.935538
#> 2   500.2427   2.597609
#> 3   501.4332   2.930135
#> 4   500.3133   2.723306
#> 5   516.9320   4.058366
#> 6   500.1605   2.617624
#> 7   518.4555   3.464361
#> 8   529.1734   3.219082
#> 9   545.7116   2.882327
#> 10  500.6366   2.582693
#> 11  500.1245   3.296128
#> 12  500.0749   2.753372
#> 13  676.4414   3.208905
#> 14  500.0468   3.761977
#> 15  693.4006   2.724698
#> 16  500.9655   3.329999
#> 17  500.5748   2.787701
#> 18  508.7905   2.654646
#> 19  500.1468   2.718469
#> 20  500.6441   2.540546
#> 21  500.2275   2.617836
#> 22  500.0176   2.674375
#> 23  504.9501   2.541918
#> 24  500.0139   2.754810
#> 25  500.0654   3.407886
#> 26  500.1898   2.704265
#> 27  500.1814   2.560756
#> 28  503.0913   2.535394
#> 29  518.1552   3.127165
#> 30  674.8632   3.296021
#> 31  555.9885   2.661782
#> 32  509.5056   3.058564
#> 33  603.9160   2.526193
#> 34  500.1798   2.513927
#> 35  515.6211   2.573566
#> 36  500.0579   2.619098
#> 37  500.0518   2.519871
#> 38  511.5662   3.195734
#> 39  500.1744   2.683572
#> 40  552.4826   2.631810
#> 41  550.8773   2.509225
#> 42  574.2765   2.546606
#> 43  533.7448   3.689907
#> 44  509.3951   2.751333
#> 45  500.6128   3.037898
#> 46  532.8589   3.268824
#> 47  781.8020   4.020187
#> 48  524.9309   2.509096
#> 49  502.1901   3.058717
#> 50  500.0251   3.773437
#> 51  500.3802   2.689033
#> 52  501.4802   4.153938
#> 53  501.1934   2.578379
#> 54  500.0649   3.357161
#> 55  501.8607   2.966701
#> 56  518.9758   3.061261
#> 57  500.0127   2.559636
#> 58  501.5971   2.853455
#> 59  500.7025   3.087030
#> 60  503.6323   3.836111
#> 61  503.2956   2.597527
#> 62  504.9250   3.461577
#> 63  522.3458   2.921624
#> 64  509.2292   2.834049
#> 65  514.7872   2.670667
#> 66  745.4038   3.467338
#> 67  964.5902   2.521012
#> 68  500.1123   2.824562
#> 69  500.2667   3.222564
#> 70  503.8820   4.141772
#> 71  504.0909   3.591804
#> 72  528.8813   2.600639
#> 73  514.1740   3.579824
#> 74  501.6977   3.615025
#> 75  990.2467   2.607731
#> 76  500.0779   3.024223
#> 77  500.0638   2.655362
#> 78  500.2961   2.591056
#> 79  503.2872   2.812805
#> 80  829.1690   3.131877
#> 81  507.6125   2.706937
#> 82  526.3944   2.744656
#> 83  500.8361   2.778638
#> 84  534.9648   3.544868
#> 85  694.3762   2.520061
#> 86  500.1792   3.322458
#> 87  503.1927   3.047074
#> 88  500.3305   2.671148
#> 89  510.2520   2.884219
#> 90  516.6080   2.926490
#> 91  515.8990   2.525689
#> 92  518.3978   3.872098
#> 93  500.3206   2.832880
#> 94  500.9076   2.763558
#> 95  580.1573   2.588307
#> 96  603.3588   3.043037
#> 97  500.0175   2.643793
#> 98  612.7055   3.165589
#> 99  501.4736   2.574232
#> 100 668.8568   2.951990
#> 101 500.0600   2.766272
#> 102 501.9511   2.859002
#> 103 500.3340   2.621716
#> 104 503.1063   3.419296
#> 105 503.9838   2.576265
#> 106 558.5270   3.103581
#> 107 501.4671   2.615823
#> 108 515.5554   2.858428
#> 109 500.2175   3.854436
#> 110 500.0568   2.678757
#> 111 500.1827   2.725214
#> 112 552.9327   3.570643
#> 113 503.1714   3.202221
#> 114 508.6961   3.076210
#> 115 565.9846   2.766701
#> 116 501.0331   2.923564
#> 117 737.7922   2.596087
#> 118 511.2511   2.859958
#> 119 500.0584   2.710440
#> 120 500.3112   2.983092
#> 121 500.1096   3.492388
#> 122 502.2923   2.602391
#> 123 536.4597   2.769074
#> 124 500.1770   2.759242
#> 125 502.0010   2.676241
#> 126 501.0416   2.551138
#> 127 501.5783   3.045233
#> 128 500.0227   2.704340
#> 129 731.6733   2.723805
#> 130 500.0414   2.669730
#> 131 500.0052   2.931731
#> 132 685.7713   2.632651
#> 133 527.8891   3.114333
#> 134 500.7872   2.952047
#> 135 709.9286   3.719226
#> 136 500.1224   2.504482
#> 137 566.5929   2.707011
#> 138 512.1201   2.565602
#> 139 501.5025   4.606873
#> 140 500.3566   2.934139
#> 141 500.5997   2.779555
#> 142 645.7402   3.843740
#> 143 501.3680   2.531941
#> 144 528.5021   3.729722
#> 145 500.7611   2.765236
#> 146 697.6485   3.270910
#> 147 501.1606   2.581483
#> 148 607.0775   2.544302
#> 149 551.8670   2.569143
#> 150 500.0141   2.887911
#> 151 500.1376   2.732792
#> 152 636.3094   2.877246
#> 153 513.1264   2.834987
#> 154 501.2530   4.822587
#> 155 760.1534   2.640340
#> 156 502.3187   3.112679
#> 157 848.5001   3.078917
#> 158 500.1356   2.971242
#> 159 540.4075   2.552104
#> 160 689.8970   2.701324
#> 161 528.3644   2.641649
#> 162 505.6626   3.360469
#> 163 503.9716   2.788745
#> 164 544.2772   2.611482
#> 165 501.2572   3.618877
#> 166 500.6724   2.842628
#> 167 526.6232   3.061270
#> 168 684.4296   3.186351
#> 169 502.8865   2.676020
#> 170 540.5493   3.938761
#> 171 504.1177   3.153854
#> 172 500.0243   3.395198
#> 173 672.7927   2.644896
#> 174 500.9081   4.398084
#> 175 503.4632   2.708462
#> 176 514.8412   3.282930
#> 177 680.4162   2.837222
#> 178 522.6952   3.573348
#> 179 500.0946   2.552918
#> 180 500.0107   2.954511
#> 181 924.5777   2.784626
#> 182 557.3477   2.512719
#> 183 502.5038   2.724139
#> 184 500.0803   2.870363
#> 185 500.0305   2.621761
#> 186 501.9587   2.641980
#> 187 515.9682   3.002198
#> 188 504.6695   2.616263
#> 189 604.2203   2.752847
#> 190 500.2571   2.830102
#> 191 500.2568   3.389302
#> 192 500.6630   3.062489
#> 193 500.1126   2.503979
#> 194 500.1781   2.649337
#> 195 505.5795   2.560199
#> 196 629.5287   2.796844
#> 197 557.5603   2.691157
#> 198 504.6136   2.867809
#> 199 500.4765   3.430815
#> 200 532.9497   2.523508
#> 201 511.9043   2.661946
#> 202 569.4022   3.396579
#> 203 500.3801   3.156891
#> 204 546.4820   2.660339
#> 205 501.5625   2.513410
#> 206 506.0442   5.398781
#> 207 501.1954   2.530736
#> 208 772.2597   2.512419
#> 209 501.6370   2.589966
#> 210 527.0034   2.535311
#> 211 660.7105   2.599185
#> 212 508.4338   2.597626
#> 213 553.1466   3.595074
#> 214 500.0726   2.610349
#> 215 504.6207   2.748982
#> 216 501.5636   4.597608
#> 217 631.8396   3.206544
#> 218 501.6757   2.563887
#> 219 697.5568   3.250618
#> 220 583.5965   2.960885
#> 221 546.5367   2.907483
#> 222 502.1435   3.086523
#> 223 672.6485   2.563065
#> 224 500.4593   3.711941
#> 225 561.8812   2.595596
#> 226 500.0060   2.748724
#> 227 500.2513   2.539889
#> 228 500.2074   3.099950
#> 229 512.6404   2.633350
#> 230 502.4506   3.185781
#> 231 788.8860   2.524650
#> 232 721.7596   2.539361
#> 233 500.0042   3.094230
#> 234 571.9944   2.717395
#> 235 500.9600   2.565672
#> 236 519.5542   4.091529
#> 237 506.8369   3.375523
#> 238 502.1270   3.025342
#> 239 500.3160   2.683299
#> 240 500.0031   3.112896
#> 241 573.1871   2.579053
#> 242 504.1073   2.929031
#> 243 500.3545   2.535842
#> 244 500.0935   2.886748
#> 245 907.4743   3.130542
#> 246 501.1902   2.757643
#> 247 525.2310   2.704093
#> 248 500.1957   2.717250
#> 249 500.0373   2.993191
#> 250 513.6024   2.992958
#> 251 500.1098   3.004034
#> 252 500.0110   3.704485
#> 253 503.2731   2.778365
#> 254 913.2722   2.580872
#> 255 505.5185   2.848198
#> 256 501.0396   2.616364
#> 257 503.1556   2.977184
#> 258 501.5702   2.715454
#> 259 500.2629   3.285104
#> 260 500.7377   3.700090
#> 261 607.9870   3.153105
#> 262 500.2246   3.319687
#> 263 577.3032   2.892353
#> 264 501.4709   2.950730
#> 265 580.0433   2.683329
#> 266 521.0155   2.809246
#> 267 500.2262   2.608689
#> 268 802.3605   3.115681
#> 269 507.0925   2.681368
#> 270 505.0079   4.681328
#> 271 502.9625   2.958658
#> 272 513.0651   2.646460
#> 273 502.7089   2.586424
#> 274 546.5809   4.127889
#> 275 500.0289   2.593214
#> 276 575.7622   2.934665
#> 277 506.5164   3.120349
#> 278 540.6398   2.643029
#> 279 633.7986   3.049341
#> 280 525.1770   2.744342
#> 281 500.0998   2.706604
#> 282 627.6404   2.778899
#> 283 502.1595   2.596969
#> 284 503.7696   2.552035
#> 285 535.9715   2.863873
#> 286 535.8848   3.240798
#> 287 515.4259   3.296593
#> 288 500.2225   2.646437
#> 289 505.4881   2.916733
#> 290 504.0942   2.568920
#> 291 502.7429   3.210312
#> 292 823.2809   2.511300
#> 293 501.4871   3.025482
#> 294 502.1307   2.601850
#> 295 500.7615   2.504171
#> 296 693.0881   3.084908
#> 297 785.1007   3.135371
#> 298 500.1015   2.710836
#> 299 505.6525   2.675158
#> 300 500.0139   2.668113
#> 301 547.1863   2.993614
#> 302 500.1523   3.278638
#> 303 500.3296   2.791101
#> 304 687.2874   2.615886
#> 305 527.3076   2.856613
#> 306 500.3516   2.758169
#> 307 500.2254   2.833420
#> 308 521.6709   2.669021
#> 309 525.1671   3.090279
#> 310 630.2092   2.737496
#> 311 909.9463   2.686611
#> 312 500.0675   2.951612
#> 313 500.6331   2.554305
#> 314 509.6597   2.988415
#> 315 532.2397   2.519599
#> 316 500.0071   2.558420
#> 317 500.0259   2.740172
#> 318 652.0202   3.853970
#> 319 505.5455   2.521007
#> 320 502.7755   2.832757
#> 321 500.4592   2.614035
#> 322 536.0816   2.921935
#> 323 547.3074   2.569665
#> 324 510.0068   2.574292
#> 325 589.8364   2.556685
#> 326 500.1175   2.631550
#> 327 500.8555   3.485779
#> 328 579.2770   2.684221
#> 329 503.6282   2.653786
#> 330 584.6928   3.133509
#> 331 596.2255   3.005303
#> 332 500.5160   2.899803
#> 333 535.8998   2.719442
#> 334 502.0077   3.279316
#> 335 500.3528   2.602089
#> 336 599.2665   2.570401
#> 337 501.2583   2.988942
#> 338 521.1396   2.530052
#> 339 500.1692   3.140249
#> 340 500.3818   2.708688
#> 341 992.5339   2.742354
#> 342 506.1786   3.620481
#> 343 501.5541   2.924681
#> 344 528.4207   3.301145
#> 345 500.0002   2.689818
#> 346 591.3798   2.549533
#> 347 500.1103   3.068094
#> 348 513.7397   2.510118
#> 349 503.6584   3.038819
#> 350 675.4027   2.826741
#> 351 501.4590   2.637199
#> 352 500.1004   3.283206
#> 353 526.5518   2.591249
#> 354 541.2205   3.597649
#> 355 500.1042   2.934570
#> 356 522.4453   2.904796
#> 357 500.1278   3.194945
#> 358 653.6156   2.518152
#> 359 511.2146   3.299579
#> 360 613.4387   3.001052
#> 361 500.0963   2.683734
#> 362 500.2240   2.816255
#> 363 500.5181   3.230499
#> 364 500.3234   2.529971
#> 365 516.2259   3.496760
#> 366 908.1128   2.782949
#> 367 500.4523   4.459445
#> 368 594.0273   3.306686
#> 369 578.8088   2.719256
#> 370 500.0023   2.779125
#> 371 500.6315   2.572730
#> 372 502.2330   3.257786
#> 373 500.2006   3.282161
#> 374 503.9157   3.293239
#> 375 517.0566   2.530438
#> 376 549.5456   2.823388
#> 377 500.7470   2.700199
#> 378 501.1046   2.567344
#> 379 523.0456   2.747349
#> 380 501.0465   2.706999
#> 381 521.3173   2.710878
#> 382 525.6260   3.208067
#> 383 500.8227   2.669145
#> 384 631.8282   2.702875
#> 385 702.2950   3.187054
#> 386 538.9970   2.914828
#> 387 501.6592   2.941423
#> 388 534.0120   2.985694
#> 389 500.3686   2.651035
#> 390 544.4685   3.509085
#> 391 631.6319   3.225933
#> 392 500.4291   2.633699
#> 393 525.9930   2.762528
#> 394 503.1718   2.744984
#> 395 509.5797   2.650066
#> 396 501.6968   2.551769
#> 397 500.5842   2.642008
#> 398 500.0084   2.747584
#> 399 761.0935   4.007441
#> 400 549.9570   3.220281
#> 401 505.6920   3.311200
#> 402 580.3726   3.168058
#> 403 561.6441   2.608748
#> 404 500.0532   2.917046
#> 405 650.3129   2.863160
#> 406 500.6468   3.470478
#> 407 501.4527   2.674659
#> 408 523.6996   2.926341
#> 409 505.7076   2.852002
#> 410 500.1498   2.580974
#> 411 895.3851   2.679285
#> 412 500.0429   2.886389
#> 413 500.4682   2.814147
#> 414 507.9233   3.954127
#> 415 501.6951   2.861686
#> 416 546.3060   2.587253
#> 417 500.3965   4.085402
#> 418 500.0716   2.890982
#> 419 667.0775   3.708727
#> 420 500.5267   2.792810
#> 421 501.8562   2.559838
#> 422 518.9189   2.713308
#> 423 506.5986   2.551919
#> 424 623.7020   2.569605
#> 425 500.0617   2.573765
#> 426 662.8782   2.766093
#> 427 500.1758   2.620357
#> 428 545.1437   2.613158
#> 429 501.1203   2.622475
#> 430 504.1485   2.807494
#> 431 845.8397   3.661338
#> 432 500.2127   2.783694
#> 433 504.9828   3.327849
#> 434 504.7273   2.565141
#> 435 500.0121   3.845943
#> 436 639.0289   2.552071
#> 437 501.2775   3.933985
#> 438 504.6172   2.693941
#> 439 549.3754   3.248815
#> 440 500.0459   2.528480
#> 441 503.8033   2.594090
#> 442 514.7651   3.410370
#> 443 594.4065   3.385691
#> 444 500.0632   2.646801
#> 445 535.3999   2.946822
#> 446 717.2839   2.614857
#> 447 503.9904   2.836639
#> 448 500.1228   2.597271
#> 449 501.7448   3.460998
#> 450 518.8468   2.941435
#> 451 500.0156   2.555890
#> 452 500.0620   4.384584
#> 453 500.3076   2.720830
#> 454 500.0474   3.583491
#> 455 605.1019   2.587962
#> 456 513.6478   2.900945
#> 457 506.8042   3.700119
#> 458 500.2030   2.920390
#> 459 706.7170   2.689128
#> 460 500.1235   2.943913
#> 461 580.2743   3.573680
#> 462 518.4897   2.709606
#> 463 529.0820   3.513749
#> 464 501.3384   2.650086
#> 465 504.2283   2.796256
#> 466 592.6678   3.097967
#> 467 500.0638   2.710895
#> 468 544.5460   2.793836
#> 469 507.1861   2.641249
#> 470 508.0475   2.919891
#> 471 500.2449   2.777902
#> 472 504.3307   2.959665
#> 473 539.4181   3.039326
#> 474 632.1680   2.626005
#> 475 500.3274   2.935596
#> 476 500.4643   2.578435
#> 477 500.8864   2.606799
#> 478 500.0411   2.827127
#> 479 502.7118   2.975143
#> 480 500.0140   2.616003
#> 481 690.8174   2.722743
#> 482 500.7249   2.548576
#> 483 507.7999   3.005446
#> 484 941.8353   2.724114
#> 485 500.1653   2.511262
#> 486 505.7705   3.017197
#> 487 507.5618   3.770693
#> 488 501.9479   2.697187
#> 489 500.0239   3.034830
#> 490 583.8364   2.742551
#> 491 591.5866   2.780506
#> 492 500.1229   2.742779
#> 493 652.7405   2.566117
#> 494 500.1288   2.527943
#> 495 521.0010   2.577191
#> 496 500.1767   2.543712
#> 497 529.6839   2.593668
#> 498 500.1100   2.729590
#> 499 500.8262   2.617455
#> 500 501.5012   3.346771
#> 501 507.8906   2.637210
#> 502 500.5844   2.918462
#> 503 523.7582   2.830856
#> 504 502.3076   4.604829
#> 505 517.4629   3.047826
#> 506 571.1533   2.666882
#> 507 579.6862   2.505253
#> 508 557.5653   2.630922
#> 509 501.0245   2.727952
#> 510 500.0427   3.008355
#> 511 546.7742   3.585679
#> 512 500.3176   2.611466
#> 513 515.8340   2.632099
#> 514 500.1513   2.758513
#> 515 500.7002   2.897957
#> 516 511.7640   2.579078
#> 517 500.7795   2.916665
#> 518 500.2300   3.498487
#> 519 500.2301   2.571758
#> 520 501.0758   2.808840
#> 521 500.2298   2.799677
#> 522 501.0668   2.796433
#> 523 511.4299   2.883420
#> 524 623.2343   2.582778
#> 525 500.0731   2.943475
#> 526 504.3141   3.531394
#> 527 537.9783   3.152020
#> 528 500.0146   3.821769
#> 529 500.5395   2.537624
#> 530 805.3330   2.682085
#> 531 509.8506   3.786216
#> 532 500.1337   3.846453
#> 533 585.2054   2.831423
#> 534 561.9736   2.619434
#> 535 500.0116   2.692205
#> 536 500.0576   2.631420
#> 537 500.0679   3.385556
#> 538 891.9147   2.610929
#> 539 500.2881   3.561861
#> 540 500.4588   2.938967
#> 541 511.7746   3.645744
#> 542 516.4733   2.725821
#> 543 547.4604   2.548839
#> 544 502.8022   3.089854
#> 545 500.2815   2.987728
#> 546 500.0712   2.539715
#> 547 505.5639   4.092286
#> 548 505.8107   2.674272
#> 549 501.3841   3.164362
#> 550 720.8113   3.087868
#> 551 501.0435   3.141734
#> 552 500.1461   3.013627
#> 553 500.4953   2.796583
#> 554 500.7640   3.060307
#> 555 501.8559   2.942844
#> 556 517.1882   3.368736
#> 557 852.3053   3.238984
#> 558 505.4031   2.709501
#> 559 995.2032   2.725866
#> 560 500.3438   2.644444
#> 561 637.7595   3.743323
#> 562 959.6527   2.665953
#> 563 709.6193   3.067447
#> 564 506.8681   3.408978
#> 565 500.1426   2.779912
#> 566 594.0559   2.749106
#> 567 516.3636   2.669175
#> 568 502.2536   2.627246
#> 569 507.0895   3.299271
#> 570 966.2920   3.091461
#> 571 500.3051   3.096696
#> 572 500.0995   3.280690
#> 573 501.1457   3.065348
#> 574 578.4747   2.738430
#> 575 500.0026   2.523924
#> 576 500.0593   2.555682
#> 577 502.0443   2.930231
#> 578 500.2027   2.534633
#> 579 500.0563   3.926807
#> 580 602.7374   2.731752
#> 581 500.4803   2.843543
#> 582 547.3097   2.747558
#> 583 501.2397   2.505657
#> 584 501.5129   3.233877
#> 585 504.3047   2.918882
#> 586 500.0595   2.529328
#> 587 500.1948   2.679609
#> 588 500.2295   2.555520
#> 589 518.8843   3.202581
#> 590 515.5330   2.543205
#> 591 520.1924   2.517669
#> 592 502.0160   2.798044
#> 593 501.8355   2.937352
#> 594 500.0555   2.814186
#> 595 507.6742   3.082691
#> 596 503.2627   2.695125
#> 597 522.4805   2.602990
#> 598 501.8141   3.210462
#> 599 711.7080   2.683932
#> 600 516.1200   2.577339
#> 601 540.3115   2.607229
#> 602 508.1702   2.753735
#> 603 504.5658   2.836265
#> 604 512.0760   2.973251
#> 605 506.1311   2.675784
#> 606 697.2477   2.894401
#> 607 509.2929   2.561967
#> 608 505.7783   3.143738
#> 609 500.4062   2.722476
#> 610 511.8286   3.511339
#> 611 500.9343   3.446196
#> 612 518.6832   3.093050
#> 613 500.0256   2.822948
#> 614 505.7255   2.532002
#> 615 502.0299   2.615225
#> 616 500.0061   3.040044
#> 617 500.1531   2.657609
#> 618 586.7077   2.550090
#> 619 610.0272   2.784401
#> 620 556.9351   3.229674
#> 621 500.1055   2.793058
#> 622 530.8814   4.073049
#> 623 566.1910   2.729042
#> 624 546.5304   2.629085
#> 625 669.4595   2.794279
#> 626 501.5003   2.561008
#> 627 518.4450   2.651850
#> 628 500.3397   3.098031
#> 629 862.7765   2.740570
#> 630 516.9648   2.500632
#> 631 501.6778   2.661910
#> 632 500.0575   2.772762
#> 633 691.8946   2.853653
#> 634 500.0086   3.204455
#> 635 505.1549   2.836843
#> 636 546.3442   2.564889
#> 637 500.7588   2.679004
#> 638 500.6493   3.095781
#> 639 500.5084   3.449338
#> 640 500.8231   2.629143
#> 641 500.2114   3.102765
#> 642 817.9268   2.673337
#> 643 516.1448   2.710754
#> 644 573.3646   2.527677
#> 645 723.3542   3.682003
#> 646 822.7142   2.925834
#> 647 500.0298   2.685609
#> 648 503.5633   4.171272
#> 649 501.1518   3.570262
#> 650 505.2252   2.648722
#> 651 500.8316   3.487640
#> 652 933.5196   2.573759
#> 653 527.1538   2.811246
#> 654 613.4003   2.659947
#> 655 682.6467   2.645983
#> 656 507.5310   3.142471
#> 657 500.3229   2.660703
#> 658 500.0371   2.503665
#> 659 501.0173   3.884831
#> 660 503.0351   2.587946
#> 661 573.5412   3.123306
#> 662 555.5142   2.613594
#> 663 501.9484   2.595085
#> 664 541.3187   2.770402
#> 665 500.0716   2.505968
#> 666 568.4222   2.534600
#> 667 546.4011   2.730883
#> 668 500.0241   2.666469
#> 669 505.6061   3.252310
#> 670 751.5433   2.859765
#> 671 500.0792   3.552298
#> 672 500.6135   2.747821
#> 673 500.3632   3.240342
#> 674 500.8759   2.824258
#> 675 503.9714   2.622701
#> 676 521.4870   2.955987
#> 677 500.1461   2.895782
#> 678 500.8147   2.510477
#> 679 536.7300   3.987434
#> 680 501.1993   2.504786
#> 681 511.4498   2.567883
#> 682 846.8008   3.335665
#> 683 743.0909   2.684106
#> 684 509.0941   2.909795
#> 685 500.0364   3.015705
#> 686 509.4157   2.566280
#> 687 503.1889   3.013570
#> 688 557.5383   2.689743
#> 689 500.8552   2.521490
#> 690 500.3771   2.514464
#> 691 664.3214   2.598097
#> 692 500.3502   2.549517
#> 693 500.8371   3.305845
#> 694 512.5088   3.389275
#> 695 889.0302   3.110077
#> 696 574.1594   2.922418
#> 697 511.7597   2.545979
#> 698 502.8630   2.690784
#> 699 500.9709   2.513212
#> 700 503.5749   3.212059
#> 701 556.5881   4.362353
#> 702 503.2841   3.441047
#> 703 502.3384   2.793113
#> 704 621.2062   2.549735
#> 705 545.7012   2.674322
#> 706 500.0757   3.380747
#> 707 628.8781   2.691114
#> 708 548.5711   2.862250
#> 709 500.0458   2.517310
#> 710 500.1615   2.923385
#> 711 502.6699   4.018917
#> 712 501.2793   2.906577
#> 713 500.0141   2.835501
#> 714 503.1374   2.740166
#> 715 500.0805   2.566899
#> 716 500.0480   3.421867
#> 717 500.2036   2.734016
#> 718 500.1128   3.017169
#> 719 528.1167   4.261396
#> 720 736.5069   2.560718
#> 721 500.3026   3.257912
#> 722 502.4253   2.523145
#> 723 500.6368   2.869363
#> 724 525.3890   2.576331
#> 725 504.1694   2.778170
#> 726 500.5535   2.880459
#> 727 673.1914   3.031192
#> 728 500.0100   3.221293
#> 729 501.3429   2.656254
#> 730 937.2557   3.079579
#> 731 521.7105   2.595673
#> 732 524.4621   2.755962
#> 733 501.2530   2.667527
#> 734 583.8413   4.066459
#> 735 500.0182   2.594344
#> 736 500.2547   2.722400
#> 737 504.9499   3.134120
#> 738 502.0997   4.600359
#> 739 524.3566   2.614805
#> 740 883.2714   2.638592
#> 741 503.2315   2.846659
#> 742 734.3339   2.947359
#> 743 502.4520   2.540878
#> 744 600.0893   4.837302
#> 745 601.9490   3.461797
#> 746 500.4788   2.819574
#> 747 506.7651   3.119003
#> 748 500.4499   2.656478
#> 749 566.7551   2.554204
#> 750 515.5638   4.733083
#> 751 500.5737   2.950215
#> 752 500.1004   2.715491
#> 753 505.8335   2.567412
#> 754 519.9745   2.644496
#> 755 511.6415   2.651108
#> 756 500.9999   2.656868
#> 757 500.0663   3.222663
#> 758 520.8398   2.521887
#> 759 512.6158   2.903629
#> 760 505.9062   3.650381
#> 761 500.2890   2.859852
#> 762 502.2581   2.591392
#> 763 610.4518   3.441501
#> 764 500.9122   2.717847
#> 765 789.2991   2.564649
#> 766 500.0008   3.044484
#> 767 582.2783   3.512979
#> 768 500.0307   3.692443
#> 769 500.2515   3.197240
#> 770 500.9143   3.355437
#> 771 504.9730   3.845061
#> 772 501.0387   2.576628
#> 773 500.5348   2.849514
#> 774 500.1161   3.287031
#> 775 642.3059   2.663969
#> 776 502.7759   2.748615
#> 777 500.0183   3.594022
#> 778 503.4665   3.022295
#> 779 544.0364   2.690448
#> 780 501.2048   2.534020
#> 781 500.9622   3.042955
#> 782 681.1519   3.054871
#> 783 500.7981   2.750583
#> 784 508.0765   2.623034
#> 785 500.0558   2.755056
#> 786 500.6031   3.542946
#> 787 543.8706   3.272458
#> 788 510.5412   2.838577
#> 789 500.0185   2.693645
#> 790 502.4574   2.537957
#> 791 500.0059   2.720090
#> 792 521.5623   2.866300
#> 793 500.5606   2.595647
#> 794 505.3817   2.543880
#> 795 529.5528   3.550927
#> 796 628.3691   2.643161
#> 797 500.3469   2.710833
#> 798 501.3234   3.376016
#> 799 500.0210   2.605582
#> 800 502.1436   2.617490
#> 801 697.3896   2.608938
#> 802 626.9739   3.302330
#> 803 501.2574   3.195659
#> 804 500.3599   2.733819
#> 805 501.6952   2.542535
#> 806 500.2514   2.950680
#> 807 500.0602   3.178937
#> 808 501.2708   2.680637
#> 809 500.6526   3.403342
#> 810 574.6143   3.017861
#> 811 500.7052   2.931182
#> 812 568.0980   2.537754
#> 813 558.7267   2.674520
#> 814 502.7656   2.906857
#> 815 509.0867   2.756451
#> 816 500.6589   2.752900
#> 817 502.6090   2.823142
#> 818 519.7099   3.167266
#> 819 520.3616   2.538404
#> 820 500.8430   2.500405
#> 821 516.4014   2.603320
#> 822 500.5551   3.578910
#> 823 569.5447   2.803652
#> 824 509.4388   2.706039
#> 825 500.9468   3.004732
#> 826 503.3682   3.013293
#> 827 501.4414   3.210733
#> 828 500.6347   2.790324
#> 829 500.2767   2.759815
#> 830 506.3921   3.115890
#> 831 502.8875   2.678435
#> 832 502.5634   2.929655
#> 833 501.0138   2.635206
#> 834 549.7951   2.818559
#> 835 500.7123   2.873865
#> 836 644.1043   2.556357
#> 837 504.1903   3.554867
#> 838 502.3116   2.624882
#> 839 500.0172   3.049461
#> 840 501.3843   2.589992
#> 841 519.3348   2.634338
#> 842 536.4943   2.528477
#> 843 514.2735   2.576248
#> 844 504.4964   2.566286
#> 845 500.0749   2.796151
#> 846 505.2528   2.696281
#> 847 961.7088   2.899430
#> 848 525.6523   2.540389
#> 849 500.0983   2.968865
#> 850 507.9587   2.777157
#> 851 522.7163   4.390834
#> 852 500.0429   2.622898
#> 853 500.8370   2.679260
#> 854 570.1032   3.176690
#> 855 505.8100   2.664555
#> 856 788.4287   2.656218
#> 857 500.2878   2.516367
#> 858 526.9656   2.775164
#> 859 501.2065   2.591187
#> 860 605.9125   2.900702
#> 861 500.1476   2.838957
#> 862 567.8286   2.996067
#> 863 508.0054   2.622224
#> 864 500.3095   2.565905
#> 865 902.8908   2.624962
#> 866 574.2525   4.026932
#> 867 505.3864   2.818291
#> 868 503.3163   2.577082
#> 869 863.7247   2.668593
#> 870 500.1746   3.174164
#> 871 502.5631   2.544345
#> 872 500.0918   2.700787
#> 873 502.5883   2.955494
#> 874 541.7618   2.516407
#> 875 500.0352   3.279687
#> 876 500.4764   2.785951
#> 877 959.9745   3.172387
#> 878 500.0807   3.032721
#> 879 501.3720   2.667803
#> 880 501.0208   2.941329
#> 881 538.1633   2.963972
#> 882 617.9001   3.973082
#> 883 500.0714   3.017363
#> 884 500.0531   3.659458
#> 885 500.8001   3.074192
#> 886 505.0659   3.303532
#> 887 504.4566   2.663040
#> 888 604.8667   2.682239
#> 889 505.1684   2.651243
#> 890 504.7585   2.801252