Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   979.1842   3.370910
#> 2   510.4544   2.684358
#> 3   500.0517   2.655394
#> 4   527.2554   2.503430
#> 5   500.1537   2.778110
#> 6   500.4556   3.151196
#> 7   503.6451   3.210943
#> 8   500.3383   2.941251
#> 9   512.6128   2.537724
#> 10  984.3894   2.698260
#> 11  500.0579   2.654810
#> 12  505.3471   2.777665
#> 13  502.0142   2.685597
#> 14  507.2329   2.916243
#> 15  757.7778   4.500103
#> 16  904.0723   2.718782
#> 17  500.3001   2.992810
#> 18  758.2217   2.656910
#> 19  500.0288   3.374334
#> 20  500.0294   2.699478
#> 21  500.7989   2.736477
#> 22  505.1020   3.988998
#> 23  500.0253   3.228663
#> 24  501.3198   3.279906
#> 25  506.4343   2.804521
#> 26  533.4308   3.732121
#> 27  500.1668   3.337831
#> 28  501.5358   2.810592
#> 29  500.0609   2.546234
#> 30  971.2548   2.587584
#> 31  502.5195   3.043048
#> 32  500.2077   3.020832
#> 33  500.6825   2.901059
#> 34  500.4825   2.787973
#> 35  838.0591   3.211443
#> 36  501.3820   3.160134
#> 37  501.1932   2.539597
#> 38  500.1495   2.693776
#> 39  571.2437   3.023342
#> 40  519.1039   2.769564
#> 41  506.3958   3.298357
#> 42  507.1092   3.521330
#> 43  508.6387   3.527447
#> 44  507.4190   2.923330
#> 45  747.1630   2.595716
#> 46  613.6086   2.621529
#> 47  502.1117   3.308580
#> 48  500.3237   2.584286
#> 49  503.6807   2.683124
#> 50  586.1199   2.669872
#> 51  507.0241   2.638889
#> 52  502.3265   2.745960
#> 53  500.0836   2.570283
#> 54  617.2199   3.589238
#> 55  502.2215   2.651804
#> 56  500.1720   3.080424
#> 57  500.8040   5.163953
#> 58  500.0506   2.585540
#> 59  521.9976   2.793206
#> 60  500.5083   3.212755
#> 61  626.8597   2.986158
#> 62  501.0991   2.596422
#> 63  536.1623   2.642952
#> 64  502.3830   2.664211
#> 65  501.0733   2.626013
#> 66  500.7615   3.038236
#> 67  514.3828   2.677968
#> 68  609.7040   2.824212
#> 69  883.4389   2.715589
#> 70  512.5589   2.578779
#> 71  507.3231   3.163014
#> 72  689.8541   2.833180
#> 73  504.7071   2.786164
#> 74  613.3103   3.499868
#> 75  500.0530   2.617234
#> 76  701.3911   2.905059
#> 77  579.5897   2.589384
#> 78  500.9180   2.807277
#> 79  500.1709   2.680434
#> 80  723.6173   2.591747
#> 81  500.8351   2.669179
#> 82  500.0597   3.425428
#> 83  510.5269   2.787727
#> 84  504.2706   2.651792
#> 85  500.1957   3.205715
#> 86  500.1135   2.812968
#> 87  503.7099   2.784836
#> 88  500.1372   2.608970
#> 89  507.9068   3.248510
#> 90  919.7665   3.502496
#> 91  500.0067   2.866289
#> 92  522.2716   3.009600
#> 93  671.8424   2.742292
#> 94  504.2798   2.939320
#> 95  500.8255   2.937489
#> 96  500.0245   2.989554
#> 97  510.5653   2.615060
#> 98  515.8566   2.574600
#> 99  500.2313   2.778763
#> 100 500.0171   2.556709
#> 101 502.4160   3.584730
#> 102 502.1815   2.730147
#> 103 504.0168   3.469936
#> 104 500.0619   2.667339
#> 105 518.8651   2.825658
#> 106 660.0446   2.579533
#> 107 501.5291   2.730145
#> 108 501.0831   2.798688
#> 109 500.0376   3.015693
#> 110 500.1328   2.665494
#> 111 500.0088   2.680500
#> 112 525.1498   2.526681
#> 113 508.7145   3.134480
#> 114 547.3500   2.627915
#> 115 502.6489   2.582155
#> 116 501.8514   2.677487
#> 117 500.0301   3.057920
#> 118 509.7521   3.056389
#> 119 500.8635   2.906505
#> 120 574.2228   3.339514
#> 121 500.1158   2.978583
#> 122 550.7554   2.965168
#> 123 735.1280   2.511238
#> 124 500.3263   2.608127
#> 125 530.4369   2.524479
#> 126 574.1570   2.526135
#> 127 501.6016   3.917734
#> 128 500.3714   2.558612
#> 129 573.3444   3.228804
#> 130 503.1106   2.615118
#> 131 501.7702   3.394900
#> 132 514.9634   3.768926
#> 133 501.9322   2.625203
#> 134 555.3250   2.559042
#> 135 500.5372   2.908924
#> 136 500.3290   3.589811
#> 137 500.1197   2.714877
#> 138 567.8888   2.504110
#> 139 506.0279   2.651197
#> 140 540.0715   2.820942
#> 141 619.0280   2.767629
#> 142 503.8411   2.505657
#> 143 601.0082   3.003465
#> 144 503.4934   2.855547
#> 145 902.1619   3.178813
#> 146 500.2308   3.679623
#> 147 502.0068   2.634651
#> 148 505.2928   3.446411
#> 149 505.4998   3.251487
#> 150 500.1328   2.700712
#> 151 500.6676   2.649046
#> 152 544.8820   2.520257
#> 153 553.7296   2.945037
#> 154 509.2204   3.037936
#> 155 519.0808   4.318383
#> 156 501.6058   2.758185
#> 157 517.4289   2.853119
#> 158 590.2305   2.940456
#> 159 503.3798   3.118403
#> 160 500.6471   2.646102
#> 161 527.5206   2.716077
#> 162 534.2063   2.854347
#> 163 530.1028   2.595755
#> 164 500.3951   2.917558
#> 165 500.7982   2.751955
#> 166 519.3654   2.592406
#> 167 500.4005   3.178724
#> 168 645.4722   2.529241
#> 169 501.8302   2.945603
#> 170 500.3943   3.563738
#> 171 502.5396   2.849440
#> 172 521.4990   3.302553
#> 173 500.4148   3.395863
#> 174 500.5141   2.785161
#> 175 526.9527   2.936014
#> 176 513.8505   2.741299
#> 177 706.5095   3.067433
#> 178 500.0245   3.258683
#> 179 500.0191   2.808665
#> 180 504.1608   2.954216
#> 181 500.0424   2.657376
#> 182 502.5331   2.703980
#> 183 501.5004   2.689889
#> 184 504.8073   2.648171
#> 185 502.2407   2.556733
#> 186 510.5957   2.749635
#> 187 565.0887   4.311125
#> 188 505.0722   2.959369
#> 189 501.9813   3.063239
#> 190 659.7629   3.025281
#> 191 500.8442   2.662406
#> 192 534.6958   3.385271
#> 193 769.2880   2.565546
#> 194 501.9790   3.402363
#> 195 545.6390   2.540160
#> 196 509.2435   2.699657
#> 197 519.5363   2.582528
#> 198 840.2617   4.671050
#> 199 500.1038   2.512329
#> 200 505.8653   2.557482
#> 201 500.8652   2.972141
#> 202 500.1714   2.812051
#> 203 500.1987   3.022335
#> 204 501.6350   3.110265
#> 205 500.6897   2.525544
#> 206 503.8034   2.716592
#> 207 666.7590   3.060659
#> 208 602.0992   2.906113
#> 209 506.4427   3.116694
#> 210 500.2301   3.557408
#> 211 533.1931   2.877574
#> 212 500.0106   3.067595
#> 213 504.9097   2.512165
#> 214 500.6049   3.105834
#> 215 500.1019   3.391764
#> 216 567.1993   2.548652
#> 217 500.0047   2.578595
#> 218 562.1146   4.725487
#> 219 500.3585   2.692807
#> 220 504.7951   3.458045
#> 221 783.3339   3.005870
#> 222 500.1375   3.166365
#> 223 521.7657   2.712569
#> 224 520.8342   2.592248
#> 225 500.8919   2.621402
#> 226 502.4526   2.686237
#> 227 511.3814   2.829483
#> 228 510.3089   3.009830
#> 229 588.6774   2.553364
#> 230 599.5595   2.560505
#> 231 502.7412   2.813664
#> 232 548.5116   3.231857
#> 233 500.1639   2.887894
#> 234 511.1415   2.628664
#> 235 503.4625   3.737081
#> 236 500.4406   2.546539
#> 237 500.0399   2.605083
#> 238 500.2390   3.201522
#> 239 840.6836   2.761870
#> 240 507.4167   2.576316
#> 241 500.1979   4.548342
#> 242 501.7382   2.701947
#> 243 501.7378   3.203371
#> 244 500.8846   2.680308
#> 245 500.8665   2.551174
#> 246 500.5651   2.593686
#> 247 710.5508   2.576382
#> 248 906.9729   3.251441
#> 249 500.9504   2.641211
#> 250 500.0868   2.983026
#> 251 508.8539   2.750656
#> 252 548.0214   2.626316
#> 253 512.3072   3.083194
#> 254 500.0754   2.709554
#> 255 500.7239   2.552316
#> 256 533.8704   2.528972
#> 257 795.0002   2.570857
#> 258 510.4289   3.011255
#> 259 507.0066   3.077439
#> 260 513.3181   3.548895
#> 261 503.7006   2.769748
#> 262 505.1144   2.654606
#> 263 500.1023   2.564687
#> 264 504.7714   2.780047
#> 265 501.7612   2.715192
#> 266 501.5540   3.380245
#> 267 698.9783   2.854508
#> 268 500.0175   2.610997
#> 269 500.2598   2.563268
#> 270 504.6479   3.268335
#> 271 521.9812   2.921459
#> 272 504.1621   2.630310
#> 273 535.9465   2.576636
#> 274 542.9028   2.721021
#> 275 726.7090   3.688894
#> 276 501.6925   2.510659
#> 277 502.4344   3.896787
#> 278 506.0216   2.567637
#> 279 502.7697   2.686741
#> 280 500.1901   2.752044
#> 281 555.0672   2.710396
#> 282 623.0242   3.544855
#> 283 573.4948   3.170834
#> 284 500.0107   2.887731
#> 285 523.3849   2.601873
#> 286 500.1883   3.155977
#> 287 698.3056   2.886309
#> 288 505.6693   2.668297
#> 289 500.0533   2.910527
#> 290 502.4475   2.635624
#> 291 500.2160   3.156071
#> 292 887.2578   3.238672
#> 293 523.4458   3.219913
#> 294 757.3188   3.255498
#> 295 505.6185   2.820905
#> 296 500.4963   2.744374
#> 297 545.3152   3.497741
#> 298 500.4256   2.767301
#> 299 502.4562   2.823124
#> 300 539.3363   3.537445
#> 301 502.3035   2.792691
#> 302 698.6216   2.630735
#> 303 500.1824   3.026681
#> 304 500.6825   4.150430
#> 305 694.6220   2.813954
#> 306 500.2716   2.603159
#> 307 511.7995   2.873936
#> 308 516.2294   2.655439
#> 309 551.4054   4.432273
#> 310 502.4808   3.109663
#> 311 500.4394   3.208171
#> 312 546.2914   3.349918
#> 313 500.1678   2.588370
#> 314 500.5917   2.650808
#> 315 500.0364   2.917277
#> 316 526.4191   2.707867
#> 317 729.4887   3.578954
#> 318 501.3292   2.852321
#> 319 503.6131   2.681583
#> 320 505.6418   3.681780
#> 321 757.6987   2.618128
#> 322 500.6274   3.077197
#> 323 544.4421   3.471772
#> 324 500.1990   2.863164
#> 325 500.1527   2.694446
#> 326 500.2042   2.611448
#> 327 629.4996   2.617652
#> 328 500.1166   2.532272
#> 329 500.9793   2.778456
#> 330 548.1865   2.509586
#> 331 652.7067   2.649253
#> 332 503.7728   2.524505
#> 333 557.5270   3.625619
#> 334 502.4418   2.808556
#> 335 500.1168   2.506213
#> 336 500.0869   3.046448
#> 337 504.0119   3.007726
#> 338 508.8705   2.586011
#> 339 500.1851   4.367265
#> 340 507.2268   3.273983
#> 341 501.3849   3.175542
#> 342 759.1787   3.120609
#> 343 513.4142   2.998835
#> 344 500.0447   3.033104
#> 345 514.0390   2.815663
#> 346 501.4704   2.694958
#> 347 501.0479   2.529010
#> 348 912.2763   2.574871
#> 349 866.5519   2.700462
#> 350 505.1383   3.141087
#> 351 577.4448   2.885388
#> 352 511.6468   2.667402
#> 353 505.8929   2.736206
#> 354 500.2738   2.778529
#> 355 565.7462   2.705153
#> 356 509.5322   3.014321
#> 357 500.5937   3.569704
#> 358 511.3108   2.872529
#> 359 514.2864   2.917660
#> 360 675.9726   3.326322
#> 361 512.2026   2.988478
#> 362 500.7438   2.718485
#> 363 541.6341   3.473907
#> 364 500.0125   2.564333
#> 365 704.4024   2.714387
#> 366 586.3651   2.553982
#> 367 500.0960   2.735953
#> 368 527.5641   2.854397
#> 369 503.0122   3.340191
#> 370 500.6594   2.890957
#> 371 510.7187   3.460091
#> 372 707.0701   2.785528
#> 373 500.3109   2.600806
#> 374 500.0962   2.662468
#> 375 500.3995   3.659115
#> 376 500.0026   2.663537
#> 377 500.1613   2.716190
#> 378 501.0742   2.567987
#> 379 505.3722   3.025044
#> 380 695.4563   2.666045
#> 381 924.5694   2.581853
#> 382 500.1486   3.556402
#> 383 614.9239   2.769395
#> 384 500.9664   2.936100
#> 385 500.0080   3.484851
#> 386 512.0099   3.642468
#> 387 504.0317   3.627565
#> 388 502.9400   2.558520
#> 389 500.3556   2.824341
#> 390 502.6419   3.185291
#> 391 619.4741   2.508626
#> 392 500.0968   3.442721
#> 393 502.9472   3.171076
#> 394 500.1191   2.613032
#> 395 500.7001   2.762353
#> 396 503.3317   3.569779
#> 397 500.1279   2.750216
#> 398 537.8555   2.696453
#> 399 500.1564   2.926074
#> 400 789.9767   2.677680
#> 401 500.4959   2.780476
#> 402 500.0783   2.782822
#> 403 516.5387   2.832590
#> 404 502.1072   3.043077
#> 405 500.1445   2.556633
#> 406 500.2700   3.147581
#> 407 503.8544   3.059224
#> 408 734.1466   3.129255
#> 409 504.1373   3.211370
#> 410 500.0668   2.822460
#> 411 880.4557   2.846989
#> 412 500.8857   2.551635
#> 413 500.0290   2.544739
#> 414 502.1965   2.671109
#> 415 500.6384   3.426084
#> 416 504.2707   3.090490
#> 417 716.0901   2.896089
#> 418 506.1545   2.746786
#> 419 784.9541   2.595167
#> 420 514.4724   2.515399
#> 421 501.2424   2.520497
#> 422 500.2662   2.666556
#> 423 500.0210   3.035561
#> 424 516.9446   2.639623
#> 425 500.0139   2.698470
#> 426 500.0536   2.733453
#> 427 637.2185   2.683338
#> 428 503.7346   2.650306
#> 429 500.1700   3.385235
#> 430 530.0106   2.537621
#> 431 828.6931   3.203922
#> 432 503.9006   2.915528
#> 433 500.0676   2.720499
#> 434 500.4631   2.898015
#> 435 515.3528   4.736013
#> 436 501.2935   3.235519
#> 437 501.2831   3.346198
#> 438 500.2400   2.593390
#> 439 585.4286   3.546881
#> 440 501.6065   2.683361
#> 441 500.0946   2.699461
#> 442 568.1656   2.815281
#> 443 500.7225   3.409275
#> 444 658.6655   2.961909
#> 445 652.3656   3.747804
#> 446 500.5575   3.629736
#> 447 500.0932   2.651245
#> 448 500.0844   2.816895
#> 449 500.0075   2.685076
#> 450 500.6505   2.953801
#> 451 502.8992   2.658306
#> 452 500.2369   2.554958
#> 453 507.1516   2.574769
#> 454 500.9022   3.126266
#> 455 508.2177   2.876972
#> 456 500.3410   3.005479
#> 457 588.0040   2.671211
#> 458 501.3838   2.551554
#> 459 581.2253   2.579532
#> 460 501.4254   3.923201
#> 461 501.3959   3.417079
#> 462 501.1040   3.987187
#> 463 614.5624   2.526611
#> 464 546.5800   3.096853
#> 465 500.3574   2.743247
#> 466 527.4907   2.518634
#> 467 510.0778   2.535482
#> 468 500.6870   2.607747
#> 469 724.9346   2.919494
#> 470 530.6153   2.579308
#> 471 517.6860   3.026503
#> 472 500.1980   2.975557
#> 473 502.1985   2.539606
#> 474 523.8918   2.543346
#> 475 548.3210   3.076896
#> 476 727.3783   3.256530
#> 477 520.7154   3.322701
#> 478 543.6720   2.771361
#> 479 500.4659   2.564447
#> 480 500.0846   2.501465
#> 481 500.0224   3.946607
#> 482 542.5480   2.661892
#> 483 507.4525   2.584697
#> 484 500.1373   2.739758
#> 485 501.8754   3.024980
#> 486 500.4887   2.733931
#> 487 513.0176   2.607000
#> 488 792.5288   2.988862
#> 489 500.2433   3.175417
#> 490 689.0020   2.852388
#> 491 514.2857   3.229133
#> 492 542.8210   2.688059
#> 493 500.0250   2.862809
#> 494 612.4832   3.174033
#> 495 505.1137   2.509044
#> 496 868.8301   2.906656
#> 497 502.4478   3.234456
#> 498 599.4928   3.404556
#> 499 501.6364   2.591314
#> 500 535.7592   2.767673
#> 501 559.6777   2.599806
#> 502 593.9205   2.934212
#> 503 770.7121   2.830476
#> 504 501.9867   3.989585
#> 505 500.2002   2.546041
#> 506 506.0015   3.029068
#> 507 502.7534   3.249614
#> 508 513.3412   2.523747
#> 509 500.1060   2.617688
#> 510 550.1902   2.620446
#> 511 500.0542   2.962975
#> 512 505.0850   3.142718
#> 513 611.5583   2.741982
#> 514 501.7250   4.883543
#> 515 501.3113   2.632066
#> 516 500.3719   3.353656
#> 517 502.8891   3.898326
#> 518 501.3663   2.692035
#> 519 690.7676   2.744796
#> 520 703.0739   2.962863
#> 521 505.9761   2.584657
#> 522 877.2194   3.234206
#> 523 500.9558   2.865451
#> 524 500.0808   2.872654
#> 525 500.1279   3.091574
#> 526 502.1386   3.688032
#> 527 500.0052   2.649202
#> 528 513.3894   2.687778
#> 529 503.4875   3.203380
#> 530 501.0083   2.766131
#> 531 647.6775   3.080622
#> 532 586.4474   2.614747
#> 533 500.2883   2.500433
#> 534 521.6757   2.927292
#> 535 709.3158   2.670706
#> 536 504.6342   3.273032
#> 537 501.8665   2.685845
#> 538 509.3757   3.231951
#> 539 500.3017   3.143064
#> 540 507.0220   3.855021
#> 541 501.2669   2.655005
#> 542 581.6589   2.824087
#> 543 530.2412   2.607435
#> 544 500.0811   3.518233
#> 545 502.0269   2.728823
#> 546 510.7311   2.819031
#> 547 502.3449   3.108682
#> 548 500.5536   3.187938
#> 549 503.1480   2.614281
#> 550 500.0211   3.421263
#> 551 546.9774   3.177140
#> 552 500.1895   3.612877
#> 553 609.8994   2.782660
#> 554 775.1816   2.888993
#> 555 500.1237   2.593153
#> 556 500.2739   2.535042
#> 557 501.1332   2.631267
#> 558 514.8069   3.043175
#> 559 501.4404   2.667143
#> 560 500.3634   2.795378
#> 561 511.3555   3.412081
#> 562 500.0877   2.611626
#> 563 500.2051   3.238262
#> 564 509.3648   2.584474
#> 565 500.3929   2.612724
#> 566 501.0066   3.712235
#> 567 507.2266   2.746578
#> 568 509.4700   2.623875
#> 569 515.1209   2.874216
#> 570 554.8563   2.543066
#> 571 558.5261   3.231005
#> 572 505.7473   3.875049
#> 573 500.1218   2.787287
#> 574 507.3140   3.104554
#> 575 500.0570   2.636800
#> 576 500.3548   2.512086
#> 577 500.1825   3.136173
#> 578 502.5096   2.527957
#> 579 500.1280   3.299498
#> 580 580.6155   2.547312
#> 581 502.3294   2.676343
#> 582 500.0634   2.527091
#> 583 500.0008   2.808864
#> 584 500.6506   2.521292
#> 585 541.0821   2.905536
#> 586 549.3926   2.541228
#> 587 500.6406   2.543823
#> 588 502.0157   2.930166
#> 589 503.6222   2.875505
#> 590 521.3811   2.778595
#> 591 522.1306   2.594186
#> 592 847.4334   2.722791
#> 593 502.1318   2.893348
#> 594 500.0085   3.139893
#> 595 686.4928   2.748755
#> 596 533.0213   3.535147
#> 597 500.5286   3.097745
#> 598 664.1967   3.946853
#> 599 500.3002   2.522585
#> 600 500.2362   3.988798
#> 601 500.7711   3.051823
#> 602 500.0629   3.164599
#> 603 501.1102   2.802875
#> 604 505.4183   3.282811
#> 605 505.3748   2.804054
#> 606 500.3132   2.884112
#> 607 544.8132   2.624997
#> 608 534.7207   3.184398
#> 609 504.3237   3.655504
#> 610 847.9420   2.744048
#> 611 522.4756   2.737546
#> 612 500.4000   2.540316
#> 613 500.5000   2.519528
#> 614 527.4661   3.136603
#> 615 504.3152   3.614298
#> 616 508.5954   2.559653
#> 617 500.9719   4.276402
#> 618 500.7440   3.028857
#> 619 653.7366   2.564936
#> 620 500.1449   2.978915
#> 621 519.8471   3.143137
#> 622 537.2766   2.844741
#> 623 500.2054   2.576817
#> 624 500.0526   3.252172
#> 625 500.6969   3.041526
#> 626 501.0287   2.890913
#> 627 501.4717   2.563248
#> 628 503.1846   3.026208
#> 629 500.1027   2.888716
#> 630 501.3196   2.527580
#> 631 508.5406   3.691155
#> 632 500.0346   2.927744
#> 633 519.9085   2.638372
#> 634 540.8908   3.018455
#> 635 500.4701   2.773400
#> 636 500.9936   2.597023
#> 637 500.7050   2.779798
#> 638 520.6902   3.153498
#> 639 500.8897   3.278699
#> 640 500.5705   2.592791
#> 641 571.0556   2.710940
#> 642 540.5326   2.923490
#> 643 504.8492   2.776693
#> 644 807.2205   2.717240
#> 645 502.7739   2.650043
#> 646 500.0365   3.313438
#> 647 501.9579   2.634559
#> 648 519.2660   2.826016
#> 649 501.9132   2.826247
#> 650 500.8674   2.800533
#> 651 500.0632   2.535066
#> 652 500.1294   2.544866
#> 653 513.1301   3.032208
#> 654 684.9115   2.746973
#> 655 561.6004   3.763719
#> 656 500.1689   2.687734
#> 657 501.2420   3.044545
#> 658 502.2606   2.981860
#> 659 500.5809   2.504233
#> 660 514.8597   2.601066
#> 661 506.7733   2.863540
#> 662 505.9374   2.563822
#> 663 501.6729   2.675703
#> 664 501.6984   3.317409
#> 665 506.6091   3.690683
#> 666 574.6636   2.962805
#> 667 673.4518   3.108331
#> 668 501.8094   3.438564
#> 669 500.3619   2.616636
#> 670 512.8125   3.379312
#> 671 534.5365   2.612001
#> 672 567.0090   2.526450
#> 673 503.3234   3.660963
#> 674 500.0788   2.555770
#> 675 500.9820   2.626942
#> 676 500.1760   2.914861
#> 677 518.4004   2.801546
#> 678 501.3265   2.564838
#> 679 500.5891   2.697058
#> 680 511.1043   2.784219
#> 681 509.4530   2.604653
#> 682 600.3613   3.087678
#> 683 510.6066   2.861768
#> 684 721.7192   2.615747
#> 685 583.3768   3.806606
#> 686 778.1167   2.678955
#> 687 511.1546   2.527646
#> 688 565.5002   2.640080
#> 689 501.9072   2.564208
#> 690 500.0013   2.559167
#> 691 527.6915   2.949927
#> 692 502.4838   2.512931
#> 693 534.8028   2.818522
#> 694 500.7643   2.954193
#> 695 523.7940   3.034203
#> 696 500.3739   2.883473
#> 697 500.3811   3.071787
#> 698 663.7795   2.697817
#> 699 646.8001   2.662160
#> 700 545.5182   2.581170
#> 701 500.5705   3.097964
#> 702 500.5338   2.753341
#> 703 517.9791   2.711664
#> 704 884.5602   2.656728
#> 705 666.8171   4.463188
#> 706 641.0273   2.595455
#> 707 501.2876   2.566210
#> 708 500.6807   2.829436
#> 709 500.5732   2.695486
#> 710 510.4894   2.770804
#> 711 520.7630   4.073803
#> 712 501.7818   3.143742
#> 713 500.7767   3.111402
#> 714 501.4266   4.670364
#> 715 503.5524   2.647878
#> 716 581.8416   3.039990
#> 717 501.4514   2.767255
#> 718 927.8177   2.623679
#> 719 530.4258   2.909832
#> 720 501.2809   2.855809
#> 721 506.4658   2.581294
#> 722 525.2669   3.212424
#> 723 524.1441   2.599873
#> 724 500.2466   2.679717
#> 725 584.9693   3.341062
#> 726 500.9214   2.559986
#> 727 500.1452   2.539430
#> 728 500.1345   2.850873
#> 729 500.0291   2.539875
#> 730 508.8557   2.588858
#> 731 506.6725   2.901927
#> 732 585.9455   2.546026
#> 733 500.6176   3.075579
#> 734 504.3269   3.447097
#> 735 500.4693   3.910371
#> 736 707.1854   2.953318
#> 737 572.2617   2.501561
#> 738 501.4444   3.751595
#> 739 500.3640   2.603599
#> 740 500.6052   2.705759
#> 741 564.3361   2.511630
#> 742 500.4833   3.294597
#> 743 506.2800   2.972232
#> 744 500.1416   2.581150
#> 745 500.4483   2.553083
#> 746 713.6730   3.074278
#> 747 523.5653   2.549478
#> 748 589.4461   2.596908
#> 749 501.5654   2.545152
#> 750 688.6990   4.417171
#> 751 500.2166   2.552355
#> 752 500.7440   2.672464
#> 753 501.1224   2.527820
#> 754 503.8103   3.281438
#> 755 500.7999   3.413402
#> 756 506.2782   2.695671
#> 757 529.4213   2.778560
#> 758 500.0995   2.929261
#> 759 500.0002   3.718645
#> 760 500.0591   2.976607
#> 761 511.1592   2.681333
#> 762 500.5563   3.809336
#> 763 500.0387   3.128301
#> 764 500.0003   3.052843
#> 765 500.1642   2.540130
#> 766 508.0452   2.700682
#> 767 758.1305   2.657299
#> 768 501.0124   2.847119
#> 769 500.0080   2.558669
#> 770 501.9305   2.531726
#> 771 500.1012   3.753525
#> 772 500.4538   3.370279
#> 773 588.6670   2.723593
#> 774 880.3725   2.633279
#> 775 518.2009   2.800168
#> 776 507.1071   2.655651
#> 777 502.5650   3.380078
#> 778 516.8521   2.757233
#> 779 521.3774   3.173315
#> 780 500.2156   2.667849
#> 781 515.1041   4.708115
#> 782 667.3332   2.826429
#> 783 907.8326   2.881905
#> 784 606.6672   2.953809
#> 785 529.9128   2.633801
#> 786 500.0046   2.964889
#> 787 501.1401   3.304449
#> 788 500.1247   2.713749
#> 789 500.2287   2.569160
#> 790 506.0628   3.053466
#> 791 500.5629   2.706245
#> 792 562.0047   2.590829
#> 793 527.8402   3.265837
#> 794 510.8174   2.543084
#> 795 501.5775   2.573678
#> 796 587.1913   2.995012
#> 797 504.5126   2.762332
#> 798 500.4293   2.773574
#> 799 504.6094   4.480473
#> 800 500.4298   2.597339
#> 801 500.9211   2.537072
#> 802 500.0273   2.952144
#> 803 585.2035   2.664740
#> 804 500.9244   2.686732
#> 805 500.0152   2.672475
#> 806 506.3670   3.516976
#> 807 500.1291   2.583520
#> 808 500.0205   2.570181
#> 809 512.7548   2.715593
#> 810 501.0198   3.205600
#> 811 520.8446   2.887876
#> 812 519.3949   2.710329
#> 813 500.9450   2.824205
#> 814 507.9081   2.585825
#> 815 501.1158   3.112166
#> 816 500.0108   2.984891
#> 817 503.5768   3.685789
#> 818 500.2703   2.573542
#> 819 507.2102   2.766803
#> 820 500.5954   2.691429
#> 821 511.7644   2.609984
#> 822 500.0619   2.575382
#> 823 500.0315   2.712071
#> 824 500.2315   3.454543
#> 825 504.4553   3.041158
#> 826 501.6536   2.631382
#> 827 546.8296   2.585632
#> 828 507.6078   2.761685
#> 829 500.3122   3.524688
#> 830 500.0112   2.793548
#> 831 503.7087   2.519628
#> 832 500.2208   2.889873
#> 833 505.2563   3.957477
#> 834 500.1480   2.582624
#> 835 500.0806   2.779910
#> 836 554.8753   2.604535
#> 837 572.1098   2.569296
#> 838 500.5032   2.545551
#> 839 510.3460   2.512060
#> 840 500.2516   2.517740
#> 841 685.2333   2.575359
#> 842 500.8617   2.732951
#> 843 500.1647   2.810752
#> 844 500.8275   2.620561
#> 845 501.2527   2.660381
#> 846 533.5311   2.763106
#> 847 663.5449   3.022927
#> 848 500.6577   3.072743
#> 849 509.8689   2.621821
#> 850 502.9527   3.234026
#> 851 501.2958   2.978016
#> 852 501.2895   2.586148
#> 853 501.4421   2.549090
#> 854 500.6329   2.524656
#> 855 500.1114   2.504622
#> 856 500.4060   2.555914
#> 857 500.0265   2.568735
#> 858 560.5907   3.000868
#> 859 508.7141   2.679652
#> 860 520.5836   2.650146
#> 861 619.4058   2.963228
#> 862 500.1427   3.115562
#> 863 558.9680   3.257267
#> 864 564.6218   2.944030
#> 865 500.0263   2.946035
#> 866 502.9373   2.606988
#> 867 500.1756   2.566542
#> 868 501.1001   3.343581
#> 869 500.9191   3.047467
#> 870 710.2021   4.507613
#> 871 567.1063   3.171316
#> 872 501.2809   2.623467
#> 873 500.1636   3.394253
#> 874 500.4602   2.615209
#> 875 567.2978   2.563354
#> 876 544.3933   2.538216
#> 877 505.3608   3.223260
#> 878 784.5133   2.513925
#> 879 522.7584   2.609735
#> 880 501.2119   3.028523
#> 881 500.7393   2.716690
#> 882 500.0453   2.593702
#> 883 514.6716   2.579170
#> 884 500.1708   4.096664
#> 885 500.0009   2.506152
#> 886 557.9836   2.541245
#> 887 851.8294   2.781051
#> 888 742.5112   2.513471
#> 889 500.3475   2.508480
#> 890 500.0117   2.576699
#> 891 504.6707   3.056257
#> 892 538.3456   2.900023
#> 893 501.0851   2.985442
#> 894 500.6791   4.155221
#> 895 560.8625   2.653605
#> 896 512.2231   2.887132
#> 897 537.5525   2.843947
#> 898 548.2044   3.840889
#> 899 501.0933   3.561428
#> 900 503.9978   2.506707
#> 901 531.8863   2.775646
#> 902 560.3010   3.492969
#> 903 500.2654   2.904437
#> 904 502.0762   3.314256
#> 905 534.9903   2.571313
#> 906 500.0201   2.727084