Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   500.1579   2.586273
#> 2   500.0596   3.151212
#> 3   507.0035   3.144787
#> 4   500.4836   2.734993
#> 5   584.8129   3.981304
#> 6   535.5281   2.929539
#> 7   500.1406   3.790651
#> 8   500.2805   2.972028
#> 9   549.5307   3.649012
#> 10  501.4204   2.533452
#> 11  500.0594   2.641807
#> 12  505.3476   2.909948
#> 13  509.8662   2.543779
#> 14  524.7139   2.670487
#> 15  502.2238   3.255343
#> 16  820.3583   2.709673
#> 17  513.4207   2.797023
#> 18  645.0611   2.925255
#> 19  638.2757   3.499499
#> 20  906.2715   2.762384
#> 21  510.0915   2.985506
#> 22  500.1208   3.220342
#> 23  548.4965   2.713879
#> 24  500.1317   3.042659
#> 25  729.0645   3.423004
#> 26  500.3253   3.337780
#> 27  500.0693   2.621256
#> 28  502.6408   2.506630
#> 29  518.7012   2.602772
#> 30  501.3994   2.858007
#> 31  639.9010   2.537033
#> 32  500.0687   2.518679
#> 33  536.5771   2.562899
#> 34  500.3280   2.956464
#> 35  500.1801   2.582095
#> 36  572.3305   2.754481
#> 37  505.1597   2.946640
#> 38  537.7853   2.626256
#> 39  566.5676   2.744355
#> 40  517.6259   2.518698
#> 41  500.4290   2.787037
#> 42  505.1641   2.780692
#> 43  509.7786   2.794865
#> 44  501.6406   2.590783
#> 45  504.2487   2.650820
#> 46  513.5700   3.479669
#> 47  513.9002   2.789166
#> 48  525.1110   3.451814
#> 49  501.7181   2.559620
#> 50  779.7830   2.646091
#> 51  500.0616   4.115692
#> 52  500.2336   2.871079
#> 53  508.2060   2.853731
#> 54  500.2709   2.539742
#> 55  503.5252   3.088707
#> 56  546.9256   2.509351
#> 57  500.9892   2.809125
#> 58  508.2300   2.598760
#> 59  503.2076   2.646398
#> 60  544.8539   2.925545
#> 61  503.6440   2.752743
#> 62  500.0786   2.566161
#> 63  515.7627   3.344713
#> 64  500.0458   2.675589
#> 65  500.7782   2.625347
#> 66  501.6477   3.538355
#> 67  631.2959   2.843967
#> 68  643.3306   2.985958
#> 69  595.2395   2.684250
#> 70  957.6285   2.545594
#> 71  508.3758   2.699310
#> 72  523.7471   2.802046
#> 73  521.0483   2.656751
#> 74  500.0552   2.714579
#> 75  542.6057   2.864505
#> 76  512.8747   2.610471
#> 77  511.3767   3.763107
#> 78  500.1147   2.501082
#> 79  534.0468   3.843542
#> 80  500.3737   2.832658
#> 81  500.0786   3.106594
#> 82  593.0003   2.699164
#> 83  598.5936   3.995994
#> 84  755.2704   2.581768
#> 85  500.4745   4.671943
#> 86  503.0408   2.720627
#> 87  545.8842   2.842114
#> 88  505.4433   2.637512
#> 89  501.4953   2.558258
#> 90  786.5851   4.267201
#> 91  604.3837   3.345079
#> 92  501.7506   2.637598
#> 93  500.2344   2.532670
#> 94  574.9109   3.279572
#> 95  509.2598   2.638375
#> 96  614.3652   3.080496
#> 97  501.5218   3.207955
#> 98  500.2470   2.509394
#> 99  511.3688   2.959775
#> 100 500.6548   3.277221
#> 101 500.1427   2.652827
#> 102 505.8114   2.575367
#> 103 501.0019   2.787305
#> 104 648.7633   3.442569
#> 105 500.2146   2.655774
#> 106 511.6001   3.317347
#> 107 529.4877   2.912144
#> 108 501.3656   2.591773
#> 109 523.3024   2.584788
#> 110 500.6833   3.697183
#> 111 951.4624   2.596434
#> 112 506.8382   3.315528
#> 113 514.9243   3.314101
#> 114 500.3854   2.534799
#> 115 508.0529   2.522536
#> 116 501.3770   2.889054
#> 117 500.3252   2.722141
#> 118 589.3414   2.807576
#> 119 758.5961   2.643080
#> 120 507.3513   2.638440
#> 121 500.0475   3.171709
#> 122 501.0604   2.626579
#> 123 500.4486   3.855516
#> 124 556.9782   2.915180
#> 125 503.7061   2.788711
#> 126 504.9424   2.959177
#> 127 500.0614   2.769897
#> 128 567.9332   3.350983
#> 129 673.5947   3.961195
#> 130 641.0923   2.714452
#> 131 519.0723   3.177846
#> 132 670.1718   3.187236
#> 133 606.7272   3.442867
#> 134 837.1955   3.225666
#> 135 524.4772   2.520172
#> 136 502.0659   2.524901
#> 137 526.7083   2.513362
#> 138 500.3950   2.907699
#> 139 500.4011   2.639328
#> 140 821.2774   2.675107
#> 141 500.2006   2.500909
#> 142 749.0093   2.885888
#> 143 502.7664   2.539714
#> 144 907.0141   2.755182
#> 145 500.0460   2.552711
#> 146 500.8688   3.059685
#> 147 502.6017   2.578567
#> 148 500.0439   3.027914
#> 149 501.1497   2.796061
#> 150 500.2937   3.067803
#> 151 541.1919   2.751656
#> 152 507.1702   3.050654
#> 153 500.2708   2.834145
#> 154 501.4577   2.860495
#> 155 500.2144   2.860697
#> 156 501.3234   3.509888
#> 157 501.1442   2.922684
#> 158 505.5216   2.539225
#> 159 537.9520   2.796707
#> 160 503.7108   3.146424
#> 161 501.5156   2.849297
#> 162 543.5813   2.814774
#> 163 500.4805   2.526693
#> 164 502.7577   2.672221
#> 165 553.4765   2.543364
#> 166 500.3563   2.871561
#> 167 500.6447   3.234774
#> 168 597.7573   2.883755
#> 169 500.1781   3.189075
#> 170 612.6978   2.609059
#> 171 501.1276   2.940290
#> 172 687.5593   4.056901
#> 173 500.2855   3.247298
#> 174 500.0958   2.584765
#> 175 500.0072   3.299104
#> 176 501.3287   2.520405
#> 177 502.7689   2.670114
#> 178 504.1791   2.854641
#> 179 528.6181   2.888118
#> 180 645.7973   2.659519
#> 181 500.2699   2.925624
#> 182 500.2446   3.387611
#> 183 500.0597   2.551865
#> 184 502.2008   2.871756
#> 185 514.0001   2.834605
#> 186 517.9468   2.511921
#> 187 564.3668   2.609668
#> 188 509.4168   3.138490
#> 189 575.3034   2.877125
#> 190 501.9626   2.701294
#> 191 500.7552   2.683967
#> 192 709.7068   2.547273
#> 193 595.9107   2.700679
#> 194 500.9169   2.819431
#> 195 579.8367   2.842897
#> 196 581.7026   2.576799
#> 197 580.6140   2.734843
#> 198 500.0435   2.519383
#> 199 502.9537   2.548363
#> 200 508.3080   3.584451
#> 201 500.0845   2.593320
#> 202 520.8775   2.723069
#> 203 513.7793   2.787295
#> 204 510.3167   3.280875
#> 205 695.0025   2.545597
#> 206 548.7467   2.599191
#> 207 557.8270   2.773967
#> 208 503.3948   2.515801
#> 209 500.1921   2.639125
#> 210 500.8028   2.691929
#> 211 545.0189   2.590359
#> 212 520.6145   2.612931
#> 213 500.1024   3.792639
#> 214 500.4082   3.003996
#> 215 501.1996   2.756740
#> 216 564.8317   2.790349
#> 217 500.3697   2.994228
#> 218 520.6305   3.009421
#> 219 500.0004   3.332523
#> 220 500.5589   2.600425
#> 221 500.7705   3.257957
#> 222 525.3800   2.514716
#> 223 502.0737   2.548460
#> 224 500.6218   2.528685
#> 225 510.9120   2.814422
#> 226 502.2238   2.674486
#> 227 539.4931   2.527815
#> 228 500.5493   2.818659
#> 229 500.0188   2.613905
#> 230 518.6521   3.803893
#> 231 516.4349   2.964856
#> 232 502.4844   2.528211
#> 233 536.7141   2.574774
#> 234 500.3355   2.737712
#> 235 511.9583   2.782401
#> 236 512.9244   3.214348
#> 237 528.0589   4.091187
#> 238 514.7447   4.420574
#> 239 500.7701   2.725410
#> 240 501.4780   3.137540
#> 241 500.4711   2.697049
#> 242 504.1457   3.018798
#> 243 732.1777   2.523462
#> 244 643.7904   3.850060
#> 245 501.0647   2.524561
#> 246 512.3234   2.618673
#> 247 511.5563   2.700906
#> 248 589.2539   2.618794
#> 249 504.4728   2.690493
#> 250 500.8652   2.886716
#> 251 580.5381   2.880225
#> 252 583.3174   3.765498
#> 253 794.8483   2.665839
#> 254 550.8316   2.584025
#> 255 500.0022   2.744447
#> 256 522.8072   3.906884
#> 257 504.3650   2.511652
#> 258 570.7684   2.503668
#> 259 500.1591   2.876471
#> 260 515.6118   2.575133
#> 261 500.0237   2.961104
#> 262 564.2527   3.007292
#> 263 503.6959   2.519888
#> 264 501.9604   2.539196
#> 265 500.5274   2.532723
#> 266 500.1908   2.760184
#> 267 558.4455   2.664851
#> 268 628.9108   3.789656
#> 269 577.5903   2.722498
#> 270 500.3703   2.762747
#> 271 502.2678   2.683500
#> 272 500.0422   2.639686
#> 273 502.7480   2.505073
#> 274 695.3857   2.559604
#> 275 557.4591   2.941820
#> 276 521.0111   2.603154
#> 277 648.8248   3.746586
#> 278 521.5002   3.155612
#> 279 500.1005   2.504770
#> 280 502.1625   2.752909
#> 281 506.6664   3.482864
#> 282 510.7835   2.691889
#> 283 500.6192   2.869099
#> 284 504.8414   3.245736
#> 285 500.1963   2.803778
#> 286 506.9132   3.407563
#> 287 642.3715   3.135610
#> 288 578.3286   2.546407
#> 289 734.2476   2.521474
#> 290 613.6308   3.398708
#> 291 500.5748   3.286229
#> 292 504.3448   3.300613
#> 293 528.6509   2.542122
#> 294 501.2834   3.205167
#> 295 512.9770   2.581994
#> 296 500.1335   2.729864
#> 297 506.2840   2.766644
#> 298 500.2796   2.900371
#> 299 500.2653   3.487870
#> 300 500.8313   2.640230
#> 301 508.9058   2.816712
#> 302 506.8248   2.514769
#> 303 512.8395   2.662165
#> 304 517.1408   2.739855
#> 305 909.4557   2.675634
#> 306 555.9972   2.603624
#> 307 503.8764   3.552102
#> 308 504.9388   3.195595
#> 309 500.0080   2.512380
#> 310 502.4801   2.713163
#> 311 511.5681   2.965069
#> 312 879.6222   3.534915
#> 313 502.2547   2.722599
#> 314 514.2322   2.522635
#> 315 500.6926   2.572082
#> 316 698.7853   2.964362
#> 317 509.4906   2.623331
#> 318 501.6083   2.616268
#> 319 500.3251   2.936320
#> 320 500.9281   2.536355
#> 321 500.9608   4.109881
#> 322 508.5686   2.778764
#> 323 502.4216   2.542339
#> 324 500.0118   4.057149
#> 325 651.5970   4.212903
#> 326 530.0161   2.803013
#> 327 500.8990   2.533293
#> 328 511.7182   2.782062
#> 329 554.1847   3.519139
#> 330 766.3804   2.885744
#> 331 503.0833   3.423180
#> 332 500.3101   3.941590
#> 333 606.4388   3.019162
#> 334 547.9905   3.595571
#> 335 500.6893   2.950440
#> 336 586.8887   3.744682
#> 337 723.4859   3.218051
#> 338 517.6775   2.672097
#> 339 517.4741   2.850532
#> 340 589.4133   2.765812
#> 341 500.2376   2.591320
#> 342 554.9800   3.222414
#> 343 502.6237   2.823297
#> 344 500.7116   3.265686
#> 345 537.6681   3.145444
#> 346 574.5238   2.671509
#> 347 523.0129   2.765365
#> 348 500.0172   2.585508
#> 349 510.5114   2.847195
#> 350 500.0698   2.692055
#> 351 500.7277   3.994354
#> 352 521.3096   2.579060
#> 353 626.0603   2.865698
#> 354 534.1325   2.818346
#> 355 558.7198   2.685824
#> 356 503.1932   4.922818
#> 357 509.4269   2.641877
#> 358 500.1172   2.878185
#> 359 501.6803   2.759625
#> 360 518.9452   3.134536
#> 361 516.6681   4.071205
#> 362 530.1184   2.530463
#> 363 613.0880   2.917570
#> 364 652.7433   3.844277
#> 365 500.6449   2.652936
#> 366 600.0041   2.823914
#> 367 560.3212   3.334634
#> 368 500.6160   2.638282
#> 369 500.8733   2.542565
#> 370 500.5667   4.168755
#> 371 500.3329   3.288058
#> 372 986.3557   2.557237
#> 373 539.7545   2.540701
#> 374 506.5840   3.596037
#> 375 508.8311   2.595952
#> 376 502.1289   4.040382
#> 377 500.0814   2.867764
#> 378 808.8464   3.033554
#> 379 550.8951   2.527490
#> 380 500.5434   2.695186
#> 381 504.7047   2.868630
#> 382 500.3432   3.206563
#> 383 500.8892   2.672301
#> 384 500.7037   2.850795
#> 385 553.5270   2.865739
#> 386 509.6307   3.772550
#> 387 505.6777   3.812549
#> 388 504.9985   2.638014
#> 389 536.2181   2.695494
#> 390 500.1097   2.581603
#> 391 531.1355   2.645562
#> 392 516.6848   2.772607
#> 393 511.4217   3.188598
#> 394 500.2965   2.736800
#> 395 516.3789   2.706517
#> 396 504.4392   2.557702
#> 397 500.6015   2.547004
#> 398 504.7026   2.729563
#> 399 544.5216   3.212471
#> 400 500.0441   2.703925
#> 401 514.4388   3.549110
#> 402 504.0031   3.414359
#> 403 503.3058   2.561711
#> 404 500.1135   2.875166
#> 405 755.0286   2.633819
#> 406 503.3724   2.973379
#> 407 500.1025   2.932422
#> 408 500.5140   2.744492
#> 409 548.7105   2.515774
#> 410 505.1383   2.753938
#> 411 500.1476   2.550701
#> 412 501.6739   2.741021
#> 413 555.2129   3.199751
#> 414 864.1555   3.594834
#> 415 500.6887   2.749043
#> 416 720.3822   2.860947
#> 417 500.0012   3.357545
#> 418 501.9611   3.223948
#> 419 502.6093   2.579013
#> 420 512.5102   3.301651
#> 421 578.4764   2.712533
#> 422 566.2235   3.032869
#> 423 500.3129   2.860301
#> 424 502.1700   2.818771
#> 425 592.7666   3.114021
#> 426 506.1281   2.621565
#> 427 577.4171   2.687889
#> 428 502.4930   3.043886
#> 429 500.1630   2.532253
#> 430 500.1379   3.097638
#> 431 534.2019   3.022296
#> 432 605.8041   2.917784
#> 433 566.6908   2.870786
#> 434 603.4933   3.040188
#> 435 500.0596   2.548876
#> 436 508.7113   3.306855
#> 437 503.0042   2.617049
#> 438 509.8291   2.802502
#> 439 500.1859   3.021231
#> 440 505.0738   2.551178
#> 441 500.3694   3.038565
#> 442 500.4177   2.587314
#> 443 669.4128   2.947846
#> 444 500.5362   2.895486
#> 445 933.6785   2.540379
#> 446 585.3974   2.764440
#> 447 500.8493   2.530503
#> 448 805.4196   4.038231
#> 449 500.6560   2.995543
#> 450 500.5034   2.683804
#> 451 511.5996   2.994030
#> 452 517.4422   2.693288
#> 453 500.0775   2.807968
#> 454 500.2447   2.591274
#> 455 500.0439   2.575335
#> 456 500.0838   2.547355
#> 457 561.8496   2.546761
#> 458 550.9074   2.611402
#> 459 502.4933   2.939456
#> 460 690.8655   2.640939
#> 461 500.6989   3.662481
#> 462 530.3905   2.799487
#> 463 500.2786   2.591295
#> 464 517.1715   2.866493
#> 465 515.9174   2.578650
#> 466 962.8863   2.922535
#> 467 515.9334   2.887259
#> 468 555.0651   4.300231
#> 469 500.9224   3.600249
#> 470 996.9733   2.615714
#> 471 502.2943   2.729073
#> 472 540.1352   2.504465
#> 473 507.1850   3.468265
#> 474 624.0225   3.374617
#> 475 500.0634   2.949946
#> 476 500.1819   3.311794
#> 477 508.5289   3.034655
#> 478 500.2080   2.822434
#> 479 500.5822   2.728052
#> 480 508.3462   2.827815
#> 481 500.0607   2.517767
#> 482 508.6219   2.557944
#> 483 526.6304   3.854871
#> 484 500.9565   2.608404
#> 485 500.1077   4.303726
#> 486 500.0260   3.134629
#> 487 500.1474   3.512159
#> 488 599.4931   3.753418
#> 489 726.8059   2.648618
#> 490 500.0335   3.131994
#> 491 694.1606   2.586669
#> 492 500.9481   2.537713
#> 493 518.3108   2.550075
#> 494 500.3074   3.426936
#> 495 503.1169   2.558354
#> 496 515.5427   2.733508
#> 497 501.0304   3.626760
#> 498 500.4327   2.919595
#> 499 500.3896   2.566712
#> 500 508.7885   3.788185
#> 501 500.0094   2.619232
#> 502 539.0343   3.743597
#> 503 500.1580   2.637107
#> 504 501.3919   2.737373
#> 505 549.7361   2.634238
#> 506 632.7357   3.675413
#> 507 501.0946   2.531408
#> 508 631.6766   2.843905
#> 509 500.3740   2.532631
#> 510 501.1632   3.069404
#> 511 504.5446   3.045344
#> 512 511.5342   4.031800
#> 513 500.9255   2.675607
#> 514 500.1141   3.494400
#> 515 507.4492   2.804750
#> 516 500.5241   2.593554
#> 517 503.4781   2.599415
#> 518 500.4526   2.869915
#> 519 500.3689   2.655036
#> 520 632.5650   2.927633
#> 521 569.6960   2.606985
#> 522 589.5921   2.823620
#> 523 506.5296   2.897019
#> 524 500.0258   2.725258
#> 525 667.5391   2.528320
#> 526 517.5607   2.914134
#> 527 501.2518   2.685731
#> 528 996.1198   2.614191
#> 529 695.5979   2.598357
#> 530 535.8018   2.888552
#> 531 500.4094   2.657050
#> 532 526.5444   3.049876
#> 533 792.3181   2.994292
#> 534 505.5471   2.958702
#> 535 609.3644   2.748311
#> 536 503.3910   2.847702
#> 537 500.0012   2.636806
#> 538 500.0461   3.189568
#> 539 500.1279   3.547542
#> 540 500.1421   2.833461
#> 541 513.4338   2.558202
#> 542 929.9970   2.612602
#> 543 500.2465   2.504569
#> 544 500.1725   2.824778
#> 545 954.6261   2.747555
#> 546 500.0348   2.626644
#> 547 527.5710   3.512667
#> 548 501.7802   2.638947
#> 549 500.7050   2.823473
#> 550 500.2653   2.739385
#> 551 502.1033   3.015220
#> 552 509.2566   4.272627
#> 553 505.4799   2.850249
#> 554 530.0152   3.057026
#> 555 512.6308   3.031575
#> 556 502.1501   3.343777
#> 557 500.1632   2.735484
#> 558 505.5041   2.920300
#> 559 500.2422   2.812614
#> 560 500.0798   2.641911
#> 561 531.4035   2.880127
#> 562 516.0547   2.967870
#> 563 501.3974   2.546217
#> 564 500.0558   2.724339
#> 565 555.1345   2.599298
#> 566 503.3805   2.996929
#> 567 511.7304   2.652830
#> 568 500.1197   2.565001
#> 569 500.0427   2.585044
#> 570 500.1855   3.202039
#> 571 500.3956   2.633827
#> 572 581.5643   2.787614
#> 573 500.1001   2.753008
#> 574 508.8748   2.794631
#> 575 502.7003   2.981320
#> 576 500.1048   3.596673
#> 577 501.6816   4.215987
#> 578 501.5978   2.518155
#> 579 676.9943   2.729262
#> 580 501.1223   3.006772
#> 581 656.2443   2.666841
#> 582 500.0767   4.480061
#> 583 973.2032   2.704652
#> 584 500.1624   2.909783
#> 585 508.6575   2.719502
#> 586 500.1963   2.778635
#> 587 500.0659   2.527453
#> 588 504.0921   3.564817
#> 589 614.4732   2.973916
#> 590 500.4228   2.570378
#> 591 523.9057   3.255190
#> 592 500.9730   2.659962
#> 593 513.1630   3.246191
#> 594 503.0289   2.516113
#> 595 500.0126   3.334004
#> 596 519.1234   2.545385
#> 597 560.3216   2.659394
#> 598 500.7839   3.400108
#> 599 500.5582   2.532703
#> 600 505.1024   4.236873
#> 601 920.9832   4.255388
#> 602 500.3090   2.769450
#> 603 626.9575   2.793652
#> 604 500.2845   2.798240
#> 605 500.6865   2.630525
#> 606 713.2973   3.834460
#> 607 500.4313   2.932227
#> 608 500.2741   2.704363
#> 609 857.5247   2.997197
#> 610 500.6583   2.668227
#> 611 503.3958   3.691535
#> 612 500.2581   3.152854
#> 613 500.0207   2.594446
#> 614 585.5217   2.568907
#> 615 522.5428   3.366290
#> 616 500.3827   2.886151
#> 617 500.8057   2.669918
#> 618 502.0144   2.975548
#> 619 503.5476   2.964491
#> 620 565.9109   3.050089
#> 621 550.3448   3.126622
#> 622 579.5138   2.572136
#> 623 500.4384   3.056232
#> 624 523.4854   2.729972
#> 625 580.7754   2.796316
#> 626 509.5429   2.576727
#> 627 546.6642   2.514548
#> 628 780.6650   2.800870
#> 629 677.4934   2.901510
#> 630 500.3042   3.590816
#> 631 576.3980   2.874683
#> 632 507.3416   2.919321
#> 633 501.1301   2.778180
#> 634 508.0873   2.949869
#> 635 562.8905   2.654926
#> 636 500.3171   3.087561
#> 637 500.2923   2.697796
#> 638 513.9690   3.443873
#> 639 504.8849   3.132811
#> 640 501.8858   2.504801
#> 641 502.8110   2.596967
#> 642 500.6890   2.942408
#> 643 500.8565   2.517475
#> 644 500.5545   2.669551
#> 645 501.5051   2.948988
#> 646 500.2409   2.524178
#> 647 500.6475   3.551563
#> 648 518.8616   3.039723
#> 649 500.5265   3.746594
#> 650 501.2488   2.752053
#> 651 500.0779   2.787924
#> 652 500.3295   2.910966
#> 653 512.4537   2.912630
#> 654 923.5327   2.777027
#> 655 549.1322   3.033427
#> 656 502.0870   2.597769
#> 657 763.5401   2.609082
#> 658 764.8366   2.659536
#> 659 500.0374   3.712700
#> 660 514.5672   2.934669
#> 661 577.2543   2.581723
#> 662 500.3591   2.541080
#> 663 500.1988   2.519403
#> 664 602.5533   2.602645
#> 665 619.5664   2.585685
#> 666 547.0319   2.941311
#> 667 501.2984   2.716858
#> 668 500.5547   2.692749
#> 669 500.0406   3.053153
#> 670 514.1731   2.665993
#> 671 500.0716   2.647230
#> 672 540.1335   2.731665
#> 673 501.7608   2.593112
#> 674 501.0045   2.710676
#> 675 500.7099   2.924132
#> 676 500.0350   3.133112
#> 677 506.0117   3.746024
#> 678 958.0573   2.575540
#> 679 503.9076   3.191586
#> 680 508.2086   2.595611
#> 681 575.4260   4.211570
#> 682 514.6234   2.776619
#> 683 500.9757   2.842518
#> 684 500.2440   2.703939
#> 685 661.4199   2.532662
#> 686 518.0966   2.635246
#> 687 500.0801   2.798326
#> 688 501.8312   2.803411
#> 689 500.0009   2.679003
#> 690 500.0976   2.552754
#> 691 500.7974   2.759728
#> 692 503.7101   2.733704
#> 693 500.1609   3.468065
#> 694 500.0296   2.544609
#> 695 500.2844   2.596668
#> 696 504.5325   3.321839
#> 697 501.2153   2.656452
#> 698 500.2866   3.382090
#> 699 504.2435   2.854342
#> 700 500.6807   2.511132
#> 701 500.0841   2.508029
#> 702 501.1373   2.638464
#> 703 502.2100   3.514421
#> 704 504.7442   3.686203
#> 705 500.3502   3.595645
#> 706 500.0299   3.774945
#> 707 500.0484   2.805880
#> 708 508.3925   3.057190
#> 709 750.3779   2.564891
#> 710 501.5870   2.565552
#> 711 530.9544   2.578368
#> 712 500.0313   2.665177
#> 713 544.7883   3.208442
#> 714 881.1904   3.065991
#> 715 508.4346   2.743835
#> 716 500.3916   2.645578
#> 717 507.4445   3.580842
#> 718 501.3481   2.747818
#> 719 526.3255   3.130888
#> 720 500.5414   2.569968
#> 721 500.0874   2.673713
#> 722 503.9722   3.024368
#> 723 506.0057   2.699628
#> 724 506.0205   2.965338
#> 725 505.0501   2.790446
#> 726 520.0314   3.187454
#> 727 501.4296   3.035587
#> 728 515.0734   2.941864
#> 729 504.4653   2.535418
#> 730 503.8847   3.015618
#> 731 501.0275   2.666217
#> 732 544.1032   3.808007
#> 733 502.2268   2.618601
#> 734 500.0331   2.600642
#> 735 503.3878   3.532942
#> 736 796.1384   3.521112
#> 737 853.9552   2.940936
#> 738 500.1890   3.685965
#> 739 500.1365   3.298748
#> 740 500.2306   2.700472
#> 741 502.4093   3.132630
#> 742 504.8993   2.805060
#> 743 500.1953   4.113112
#> 744 502.2905   2.930011
#> 745 500.8735   2.524417
#> 746 500.5232   2.784875
#> 747 501.8204   2.630539
#> 748 500.2003   2.690943
#> 749 507.0228   2.519398
#> 750 714.5492   2.688293
#> 751 511.9372   2.673571
#> 752 500.1548   2.736661
#> 753 500.1771   2.627935
#> 754 572.5396   2.501895
#> 755 500.0333   4.299310
#> 756 654.5773   2.670658
#> 757 500.6289   2.578577
#> 758 500.0605   2.694626
#> 759 500.1994   2.868793
#> 760 597.5726   2.902843
#> 761 512.3118   2.914950
#> 762 521.9658   3.337509
#> 763 523.8428   2.543763
#> 764 500.0702   2.945864
#> 765 783.6696   3.197490
#> 766 522.1486   3.986417
#> 767 503.2188   3.101911
#> 768 500.0843   2.638357
#> 769 500.2463   2.539819
#> 770 502.8936   4.299904
#> 771 500.2645   2.693783
#> 772 516.3756   2.995529
#> 773 548.2245   2.750768
#> 774 745.2429   2.719104
#> 775 504.6913   2.665226
#> 776 500.9986   2.635729
#> 777 628.8836   2.596645
#> 778 500.4825   2.663597
#> 779 505.1306   2.516153
#> 780 500.1171   2.509878
#> 781 506.9480   2.640160
#> 782 509.7573   3.417907
#> 783 500.1693   2.816171
#> 784 500.4471   2.986095
#> 785 500.1233   3.069723
#> 786 501.4476   2.545155
#> 787 694.1134   2.823507
#> 788 507.5054   2.504855
#> 789 500.4033   3.034079
#> 790 515.9348   2.500932
#> 791 549.1887   2.605709
#> 792 697.1928   3.421192
#> 793 507.6980   3.120047
#> 794 500.0175   2.518980
#> 795 502.1330   2.867102
#> 796 643.4618   2.550959
#> 797 511.1309   2.688469
#> 798 505.5471   2.654501
#> 799 500.0057   3.135235
#> 800 968.3373   2.520079
#> 801 500.7818   2.546332
#> 802 500.3126   2.992257
#> 803 520.3410   2.737496
#> 804 521.6636   2.561245
#> 805 521.5311   2.765271
#> 806 500.8204   3.206782
#> 807 502.5392   2.567606
#> 808 500.7976   3.059764
#> 809 542.6291   3.178188
#> 810 505.8494   2.814438
#> 811 545.3866   2.982341
#> 812 899.9106   3.200013
#> 813 504.0200   2.777748
#> 814 500.2999   2.659866
#> 815 564.9360   3.149034
#> 816 500.2143   2.897049
#> 817 500.2436   2.765721
#> 818 512.4394   3.100194
#> 819 527.1105   2.527215
#> 820 500.0029   2.645487
#> 821 504.3219   3.532510
#> 822 502.5546   2.589462
#> 823 500.4641   2.758321
#> 824 503.8873   4.381104
#> 825 676.8746   2.825586
#> 826 500.9128   2.598005
#> 827 502.9321   3.137368
#> 828 500.2547   3.060262
#> 829 508.0037   3.384334
#> 830 500.0143   2.546243
#> 831 501.2178   2.881887
#> 832 613.5937   2.798934
#> 833 500.2889   2.928475
#> 834 500.4417   2.859488
#> 835 539.8929   2.591917
#> 836 500.5556   2.681099
#> 837 500.3416   2.822801
#> 838 500.3012   3.995372
#> 839 500.3214   2.684018
#> 840 500.0782   2.724594
#> 841 501.3836   2.629932
#> 842 501.3088   3.019233
#> 843 500.6926   2.712577
#> 844 522.3965   3.341205
#> 845 550.4194   2.562490
#> 846 500.1160   2.610812
#> 847 525.4181   4.306183
#> 848 502.7756   2.768975
#> 849 500.4452   3.381036
#> 850 501.5153   3.056986
#> 851 621.0076   2.778307
#> 852 500.0261   2.507296
#> 853 521.3780   2.515675
#> 854 518.4301   3.955538
#> 855 500.5574   2.568171
#> 856 504.2657   2.981748
#> 857 507.5062   2.656281
#> 858 511.7261   2.923798
#> 859 532.5851   2.649195
#> 860 500.4732   2.717712
#> 861 500.0666   2.728507
#> 862 513.6615   2.792679
#> 863 973.0774   3.961635
#> 864 574.4892   2.814491
#> 865 508.2849   2.675543
#> 866 501.1680   2.577231
#> 867 500.1550   2.600837
#> 868 540.9308   3.031227
#> 869 501.7787   2.673112
#> 870 511.5346   2.530882
#> 871 505.1040   2.540338
#> 872 544.3364   2.829439
#> 873 501.2103   3.503785
#> 874 698.6344   2.810734
#> 875 500.4145   2.798103
#> 876 500.5041   3.385801
#> 877 501.5051   2.633135
#> 878 500.5702   2.866253
#> 879 501.7609   2.760152
#> 880 502.3967   2.870567
#> 881 500.0617   2.719623
#> 882 500.5330   2.570657
#> 883 501.0950   2.786442
#> 884 512.0808   2.596320
#> 885 649.4374   3.587252
#> 886 500.0182   3.234138
#> 887 522.1925   3.251939
#> 888 560.8837   2.991976
#> 889 848.5686   3.269716
#> 890 500.0526   2.625076
#> 891 500.8743   3.067103
#> 892 500.0764   2.551812
#> 893 500.1207   2.749730
#> 894 504.1095   2.791634
#> 895 718.6598   2.539505
#> 896 501.4940   2.790936
#> 897 624.2160   2.722704