Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   509.4243   2.570890
#> 2   501.1417   3.228248
#> 3   512.6946   2.753688
#> 4   500.0691   2.734028
#> 5   588.2623   2.993001
#> 6   507.2696   4.921149
#> 7   500.8384   2.746833
#> 8   744.1429   2.702392
#> 9   500.3121   2.578817
#> 10  500.3163   2.565991
#> 11  547.5436   2.822272
#> 12  500.6647   2.784068
#> 13  522.6709   3.080450
#> 14  735.0726   2.674899
#> 15  501.9893   3.215245
#> 16  500.1477   2.570002
#> 17  631.6967   2.777013
#> 18  621.7773   2.601334
#> 19  526.7877   2.891990
#> 20  510.5186   2.508956
#> 21  585.4782   2.643013
#> 22  648.4193   2.684451
#> 23  519.1579   3.005377
#> 24  507.4803   2.574477
#> 25  501.5947   3.162469
#> 26  500.1108   2.522751
#> 27  500.6249   2.724361
#> 28  501.2864   3.119320
#> 29  621.8946   3.032182
#> 30  504.9186   3.023353
#> 31  500.0786   2.937013
#> 32  916.2235   3.190595
#> 33  500.1607   3.153467
#> 34  621.3455   2.573151
#> 35  637.7661   3.058483
#> 36  502.1523   3.130130
#> 37  502.1209   3.633698
#> 38  502.4298   2.854528
#> 39  529.4618   2.583702
#> 40  501.6232   2.507495
#> 41  500.3163   2.902319
#> 42  503.1442   2.689154
#> 43  657.1230   2.554866
#> 44  504.4878   2.590003
#> 45  500.6377   3.242282
#> 46  590.5986   3.195796
#> 47  525.3203   2.898272
#> 48  504.5534   2.566532
#> 49  507.3589   2.836782
#> 50  745.3735   2.632596
#> 51  501.1273   2.529354
#> 52  500.3062   2.633287
#> 53  649.0959   2.673200
#> 54  500.0591   2.959070
#> 55  500.1407   2.571737
#> 56  500.2004   2.740956
#> 57  630.6934   2.603981
#> 58  501.7472   2.561815
#> 59  500.0511   2.708181
#> 60  500.0572   2.623166
#> 61  525.1038   3.445037
#> 62  735.3274   3.663117
#> 63  500.7677   3.183936
#> 64  507.1604   3.049585
#> 65  500.6258   3.487862
#> 66  542.2107   4.110676
#> 67  500.3450   3.003350
#> 68  503.7470   2.713925
#> 69  512.1999   2.637731
#> 70  547.7236   2.673835
#> 71  534.8228   3.236982
#> 72  500.2100   2.539385
#> 73  502.9033   3.103089
#> 74  500.0098   2.529330
#> 75  500.1247   3.282067
#> 76  500.1771   2.827411
#> 77  500.2691   2.805021
#> 78  500.9155   2.783253
#> 79  500.4348   2.570968
#> 80  536.0070   2.997292
#> 81  504.0802   2.721264
#> 82  694.7187   3.599083
#> 83  605.0870   5.685256
#> 84  552.7032   3.498400
#> 85  535.6566   2.825018
#> 86  506.2893   3.158784
#> 87  512.1799   2.583218
#> 88  978.9056   2.636035
#> 89  512.8867   2.953019
#> 90  540.4301   3.168305
#> 91  582.0450   2.691841
#> 92  508.4994   2.822289
#> 93  500.4884   4.715834
#> 94  507.0114   2.937173
#> 95  500.0103   2.827615
#> 96  500.3328   2.861479
#> 97  983.6930   3.048593
#> 98  500.2666   2.703973
#> 99  500.0362   2.527947
#> 100 508.4608   3.609889
#> 101 795.8752   2.647980
#> 102 503.6610   2.501160
#> 103 762.6378   2.918126
#> 104 500.9365   3.580444
#> 105 516.8822   2.628573
#> 106 500.6657   2.639102
#> 107 540.3488   2.693617
#> 108 500.0301   2.944240
#> 109 500.0429   2.933620
#> 110 609.9978   2.594630
#> 111 526.3445   2.657268
#> 112 931.0687   2.528275
#> 113 500.0969   2.512892
#> 114 500.1037   3.287465
#> 115 556.7444   2.556125
#> 116 768.5907   2.842613
#> 117 503.0356   2.536348
#> 118 507.2633   4.891293
#> 119 554.6108   2.787243
#> 120 500.0023   2.595224
#> 121 500.6429   3.215685
#> 122 501.1791   2.514069
#> 123 615.5634   2.920622
#> 124 500.1420   2.740944
#> 125 717.5845   2.549633
#> 126 607.0410   4.000674
#> 127 502.0907   2.767644
#> 128 537.6149   2.653550
#> 129 500.0319   2.533646
#> 130 610.9648   2.651758
#> 131 570.2083   2.610887
#> 132 545.1023   2.809194
#> 133 500.0666   2.587343
#> 134 509.3909   3.443162
#> 135 503.8929   2.919702
#> 136 500.1238   2.510284
#> 137 517.0043   2.655937
#> 138 501.4511   2.767023
#> 139 553.9294   3.596755
#> 140 500.0287   2.637039
#> 141 504.6733   3.123145
#> 142 521.7130   2.902290
#> 143 503.3198   2.633060
#> 144 518.4301   2.603078
#> 145 572.8069   2.599365
#> 146 500.0025   3.083656
#> 147 646.9127   2.938245
#> 148 500.4876   2.674563
#> 149 537.2812   3.396668
#> 150 511.6197   2.918666
#> 151 502.8002   3.886803
#> 152 515.6350   2.700204
#> 153 528.8681   2.648805
#> 154 502.5031   2.592504
#> 155 503.5761   2.693470
#> 156 502.1409   3.386927
#> 157 514.1647   2.732133
#> 158 522.4884   3.899023
#> 159 512.3196   2.526170
#> 160 991.8917   2.569099
#> 161 519.2719   3.881548
#> 162 719.6254   2.969638
#> 163 502.8346   2.706855
#> 164 500.9976   2.834944
#> 165 501.8203   3.371291
#> 166 568.7925   3.163988
#> 167 500.1654   3.453313
#> 168 500.4532   2.511484
#> 169 500.0276   2.924739
#> 170 510.8539   3.218671
#> 171 521.1937   2.921654
#> 172 500.0303   2.797247
#> 173 669.1341   2.933015
#> 174 564.6509   2.950338
#> 175 730.0722   3.224893
#> 176 500.7889   3.004051
#> 177 523.5977   2.545743
#> 178 900.7442   2.898187
#> 179 506.2690   2.864568
#> 180 504.5843   2.901687
#> 181 501.3741   3.467786
#> 182 500.3876   2.614183
#> 183 700.2857   3.521128
#> 184 506.7552   3.989577
#> 185 500.0442   2.805442
#> 186 500.9825   3.182157
#> 187 500.0658   2.864397
#> 188 595.4025   2.610335
#> 189 500.1139   2.582476
#> 190 610.3798   2.668391
#> 191 511.3741   2.783600
#> 192 500.1411   2.625014
#> 193 507.7083   2.527351
#> 194 699.7458   2.712530
#> 195 500.5463   2.533259
#> 196 904.0640   3.198820
#> 197 765.0600   3.774200
#> 198 501.0852   2.549924
#> 199 500.5424   2.573958
#> 200 501.1943   2.818415
#> 201 500.2051   3.329762
#> 202 500.4908   5.020605
#> 203 533.0701   2.743359
#> 204 500.2637   2.830889
#> 205 500.2167   2.711442
#> 206 500.3038   3.166578
#> 207 500.0233   2.732782
#> 208 500.6302   3.266561
#> 209 500.0886   3.939218
#> 210 857.3247   4.310705
#> 211 500.0082   2.667581
#> 212 995.3371   2.710892
#> 213 544.6143   3.319124
#> 214 500.0028   3.354062
#> 215 619.4932   3.076337
#> 216 509.7861   3.048122
#> 217 500.2355   3.364516
#> 218 500.2620   2.979609
#> 219 948.1492   3.296134
#> 220 663.6789   3.314219
#> 221 519.1100   2.682878
#> 222 501.2526   3.563158
#> 223 702.3221   4.473811
#> 224 514.2439   2.635906
#> 225 501.1053   3.395992
#> 226 977.1814   2.655037
#> 227 500.0907   2.716502
#> 228 545.8744   2.660477
#> 229 508.1601   2.562097
#> 230 500.2335   2.567243
#> 231 500.1104   2.605922
#> 232 500.7245   3.083678
#> 233 502.0970   2.518006
#> 234 501.9468   3.412834
#> 235 500.1792   4.081885
#> 236 504.0480   2.584210
#> 237 503.7243   2.773328
#> 238 501.5937   2.792018
#> 239 509.8560   2.955811
#> 240 564.6857   2.567588
#> 241 502.7028   2.855796
#> 242 638.5755   2.710831
#> 243 500.3545   2.691513
#> 244 500.1216   3.187625
#> 245 521.3485   4.349500
#> 246 502.3040   2.512212
#> 247 502.6061   2.801381
#> 248 631.2687   2.919964
#> 249 500.2515   2.753554
#> 250 608.5770   2.702356
#> 251 500.5261   2.604495
#> 252 500.4867   2.917435
#> 253 691.7032   2.576179
#> 254 513.3916   3.019630
#> 255 687.5396   2.935986
#> 256 500.0714   2.664159
#> 257 508.4534   2.692109
#> 258 503.3004   2.766596
#> 259 500.0974   3.502221
#> 260 580.5779   2.646039
#> 261 509.1754   4.502516
#> 262 500.1825   2.680386
#> 263 500.4087   2.533407
#> 264 560.9860   3.926285
#> 265 500.6994   3.262502
#> 266 501.8601   3.235361
#> 267 669.0606   2.770212
#> 268 523.3207   4.054273
#> 269 685.3189   2.637393
#> 270 502.7283   2.855402
#> 271 500.2730   3.739505
#> 272 507.8756   2.553145
#> 273 569.6952   3.315062
#> 274 500.2042   2.503705
#> 275 645.5825   3.129120
#> 276 500.9043   2.578025
#> 277 555.4963   3.641431
#> 278 503.3144   2.608550
#> 279 500.0207   3.370050
#> 280 584.2628   2.619728
#> 281 501.7186   2.874866
#> 282 500.8058   3.537093
#> 283 500.2660   3.091514
#> 284 971.8806   2.527244
#> 285 506.8984   2.746215
#> 286 871.6327   4.072793
#> 287 935.3876   4.101772
#> 288 500.1694   3.257022
#> 289 505.1539   3.381117
#> 290 500.1863   3.796403
#> 291 500.3227   2.666590
#> 292 501.2540   2.801528
#> 293 508.4719   3.457990
#> 294 501.7146   3.578571
#> 295 626.6395   2.587094
#> 296 500.3881   4.046164
#> 297 500.0007   3.099854
#> 298 515.7982   2.679171
#> 299 642.3328   2.656428
#> 300 503.0034   2.615186
#> 301 512.2651   2.544823
#> 302 514.2552   2.539313
#> 303 500.1731   2.601308
#> 304 509.8676   2.652247
#> 305 502.5869   2.783191
#> 306 502.1457   2.576826
#> 307 535.6482   3.109960
#> 308 502.1478   3.299470
#> 309 501.3844   2.923322
#> 310 503.5668   3.586766
#> 311 509.1064   3.299095
#> 312 597.5961   2.602844
#> 313 500.0103   3.360580
#> 314 500.2410   2.810973
#> 315 500.9393   3.106976
#> 316 526.7510   3.061194
#> 317 501.8601   2.742241
#> 318 500.1243   2.698242
#> 319 518.3159   3.249916
#> 320 500.8450   2.557111
#> 321 503.3998   2.547959
#> 322 533.5518   2.623664
#> 323 500.1564   3.403375
#> 324 509.3869   2.776408
#> 325 500.0426   2.981267
#> 326 517.3627   2.903056
#> 327 520.5550   2.668160
#> 328 756.2243   3.606499
#> 329 500.0321   2.640229
#> 330 501.0807   2.534589
#> 331 501.7855   2.842235
#> 332 500.0262   2.635450
#> 333 500.7608   3.155148
#> 334 775.3187   2.504220
#> 335 500.3004   3.683169
#> 336 544.1900   2.634157
#> 337 500.2742   2.899251
#> 338 501.0264   4.025426
#> 339 586.5694   2.892842
#> 340 503.0719   2.901588
#> 341 501.1094   2.766178
#> 342 513.9550   3.037841
#> 343 500.0085   3.309773
#> 344 513.8454   2.659818
#> 345 501.1007   2.964261
#> 346 500.0442   3.646406
#> 347 549.7825   3.101556
#> 348 502.3174   2.620693
#> 349 500.9812   2.708553
#> 350 519.2547   2.533077
#> 351 500.8682   2.869265
#> 352 500.1020   3.056762
#> 353 609.2720   2.581260
#> 354 500.1295   2.607177
#> 355 684.5970   2.714292
#> 356 501.0262   2.584143
#> 357 500.2539   2.606112
#> 358 500.0002   2.673552
#> 359 510.1529   2.513119
#> 360 500.0677   2.914645
#> 361 501.6122   2.756466
#> 362 500.1822   3.487676
#> 363 501.6554   3.972536
#> 364 647.9499   4.648085
#> 365 502.8117   2.711066
#> 366 762.3141   2.920761
#> 367 500.0055   2.929953
#> 368 617.1718   2.776673
#> 369 501.9727   2.671073
#> 370 807.4235   3.509680
#> 371 500.3124   2.670154
#> 372 564.0879   2.840438
#> 373 501.1018   3.611413
#> 374 550.7151   2.971885
#> 375 504.4089   2.505495
#> 376 500.1272   3.417654
#> 377 501.2608   2.834199
#> 378 554.3371   2.915045
#> 379 509.7404   2.615792
#> 380 603.5184   2.854664
#> 381 510.5682   2.683330
#> 382 767.9173   2.668184
#> 383 500.8889   3.417096
#> 384 500.1117   2.619275
#> 385 500.0410   2.520707
#> 386 606.0819   2.714276
#> 387 503.0152   2.510004
#> 388 500.1872   2.998233
#> 389 500.0121   2.738392
#> 390 501.1894   3.203128
#> 391 504.3347   3.040031
#> 392 500.0167   2.980934
#> 393 658.8680   2.666183
#> 394 824.7324   2.924176
#> 395 501.2938   2.506119
#> 396 500.2226   2.647236
#> 397 512.5457   4.690126
#> 398 504.0525   2.791602
#> 399 760.4194   2.877611
#> 400 500.3262   3.951761
#> 401 500.8215   2.597191
#> 402 535.6459   2.717953
#> 403 504.4577   2.746989
#> 404 500.6106   2.617974
#> 405 512.1978   2.847019
#> 406 501.4425   3.053893
#> 407 931.1603   3.492561
#> 408 958.4257   3.093099
#> 409 506.4488   3.687310
#> 410 500.0429   3.290185
#> 411 500.0443   2.629208
#> 412 501.6389   2.608969
#> 413 500.7392   2.507207
#> 414 500.0117   2.756185
#> 415 501.5759   2.603589
#> 416 500.3980   3.191789
#> 417 500.3170   2.670720
#> 418 500.0905   2.693778
#> 419 501.7317   2.632103
#> 420 515.4958   2.987696
#> 421 500.0664   2.839803
#> 422 619.5750   3.593510
#> 423 500.5308   2.763167
#> 424 506.3994   2.779474
#> 425 501.1604   2.707888
#> 426 500.0080   2.733588
#> 427 500.0232   2.696854
#> 428 500.1761   2.867703
#> 429 500.7946   3.843827
#> 430 903.0962   2.706934
#> 431 504.2255   3.207776
#> 432 918.0541   2.626824
#> 433 510.2367   3.767094
#> 434 502.6981   2.720047
#> 435 510.7265   2.924351
#> 436 500.5192   3.306575
#> 437 658.8759   2.563014
#> 438 500.5078   3.118996
#> 439 514.7650   2.535459
#> 440 500.4817   3.072062
#> 441 500.6829   2.614787
#> 442 517.5946   2.590527
#> 443 940.7029   3.032573
#> 444 503.5098   2.666050
#> 445 500.0831   2.579946
#> 446 500.5501   2.716064
#> 447 644.8851   2.836020
#> 448 500.4239   3.725211
#> 449 569.3583   3.118728
#> 450 500.0343   2.556260
#> 451 512.2476   3.999259
#> 452 500.9800   2.541175
#> 453 500.8944   3.294603
#> 454 508.8929   2.504760
#> 455 503.1161   2.930131
#> 456 500.0743   2.644354
#> 457 549.7881   2.536830
#> 458 661.8665   2.527387
#> 459 547.7444   2.750479
#> 460 510.9957   2.863050
#> 461 953.5881   2.647987
#> 462 500.6524   2.777884
#> 463 527.6760   2.729040
#> 464 500.4809   2.583803
#> 465 514.8654   2.848139
#> 466 706.5622   3.326166
#> 467 502.6370   2.689587
#> 468 639.0989   2.898385
#> 469 522.8614   2.611516
#> 470 502.2372   3.464870
#> 471 500.0415   3.183932
#> 472 500.0492   2.674641
#> 473 500.9896   3.084784
#> 474 569.1246   3.559498
#> 475 502.5788   2.823692
#> 476 500.4059   2.573570
#> 477 501.8106   2.636619
#> 478 500.1653   2.599968
#> 479 504.1326   3.651153
#> 480 500.0680   2.624534
#> 481 539.8788   2.550445
#> 482 500.2298   2.671456
#> 483 508.1430   2.719016
#> 484 518.9790   2.559114
#> 485 507.5087   5.235633
#> 486 502.2855   2.565983
#> 487 500.3141   2.832379
#> 488 501.0375   2.609906
#> 489 526.9243   2.815890
#> 490 500.9403   2.996507
#> 491 856.9773   3.226795
#> 492 510.7212   2.574874
#> 493 513.2652   2.987957
#> 494 529.3505   2.682940
#> 495 500.4987   3.011760
#> 496 500.1292   3.422182
#> 497 501.3603   2.855495
#> 498 596.6739   2.752778
#> 499 676.6306   3.052926
#> 500 580.8281   2.610936
#> 501 517.7167   2.945107
#> 502 500.9155   2.896670
#> 503 500.5819   3.665663
#> 504 501.7966   3.560312
#> 505 500.4542   3.346600
#> 506 881.9766   2.656497
#> 507 500.0969   2.606759
#> 508 793.8475   2.638104
#> 509 517.1655   2.809307
#> 510 507.0963   3.045581
#> 511 908.8360   3.424419
#> 512 580.4171   3.150689
#> 513 548.6220   2.537200
#> 514 500.3357   3.629512
#> 515 503.6724   2.516637
#> 516 500.2822   2.771789
#> 517 500.0836   2.599466
#> 518 837.6726   3.126916
#> 519 500.9235   2.791776
#> 520 500.0428   2.534845
#> 521 500.6080   2.787053
#> 522 500.1731   2.515488
#> 523 500.2338   2.538828
#> 524 500.1673   2.554945
#> 525 501.8811   3.444759
#> 526 502.4467   2.824715
#> 527 501.0153   2.785706
#> 528 501.8677   3.184354
#> 529 502.7645   2.854878
#> 530 512.4521   2.798258
#> 531 760.4245   2.618053
#> 532 500.0522   2.575075
#> 533 512.8211   2.634495
#> 534 506.7407   3.439252
#> 535 512.1233   2.593877
#> 536 504.1276   2.628432
#> 537 507.7788   5.579883
#> 538 589.5110   2.819092
#> 539 543.4326   2.736973
#> 540 634.7147   2.568998
#> 541 530.6701   2.729550
#> 542 500.4062   2.563040
#> 543 563.6544   2.922660
#> 544 538.5348   2.536764
#> 545 506.2648   3.328674
#> 546 769.8407   3.227440
#> 547 506.6448   3.061861
#> 548 500.2753   3.047364
#> 549 577.2172   2.660172
#> 550 641.4700   2.502801
#> 551 934.7705   2.807282
#> 552 534.0333   2.574535
#> 553 522.2379   2.888031
#> 554 508.2015   2.666539
#> 555 500.0207   2.961802
#> 556 521.5858   2.566833
#> 557 501.4452   2.613379
#> 558 500.0693   3.646880
#> 559 669.2098   2.518179
#> 560 505.8767   3.101546
#> 561 518.9271   3.043492
#> 562 843.3530   2.540281
#> 563 504.4093   2.687333
#> 564 506.7826   2.994151
#> 565 504.7646   2.797853
#> 566 979.4441   2.715206
#> 567 528.7510   3.011205
#> 568 501.8414   2.585248
#> 569 503.8989   3.593595
#> 570 658.4761   2.576476
#> 571 501.6638   3.048430
#> 572 678.2688   2.801340
#> 573 869.8282   3.937618
#> 574 500.0038   2.910953
#> 575 503.7561   3.911105
#> 576 553.9433   2.976470
#> 577 500.0092   2.532848
#> 578 504.6906   3.109418
#> 579 548.6309   2.947706
#> 580 504.0066   3.775432
#> 581 504.5889   3.254244
#> 582 518.6751   3.371012
#> 583 512.7919   2.622884
#> 584 500.9214   2.952014
#> 585 500.7008   2.795921
#> 586 791.9867   3.529554
#> 587 500.6389   2.689845
#> 588 519.1421   2.997280
#> 589 712.7333   2.862410
#> 590 501.2203   3.135208
#> 591 526.7613   2.914855
#> 592 509.9965   2.748939
#> 593 511.1734   2.557357
#> 594 500.0109   2.609938
#> 595 580.7616   2.853633
#> 596 516.3346   4.451169
#> 597 506.4423   2.670125
#> 598 655.9762   3.796129
#> 599 500.0510   2.896959
#> 600 537.6568   2.578928
#> 601 500.0527   2.566783
#> 602 505.2612   2.924117
#> 603 500.4812   2.512652
#> 604 512.6174   3.029737
#> 605 500.2935   2.954153
#> 606 519.2009   3.356535
#> 607 755.6604   2.607572
#> 608 500.2081   3.573718
#> 609 500.2039   3.020908
#> 610 508.6232   2.798155
#> 611 500.3342   2.632576
#> 612 501.3593   3.229557
#> 613 500.8877   2.811588
#> 614 503.2145   2.666836
#> 615 513.1684   2.531629
#> 616 502.2996   2.880582
#> 617 500.2071   2.795770
#> 618 500.2320   3.478567
#> 619 549.2975   2.586431
#> 620 529.8329   2.777365
#> 621 515.4393   2.810698
#> 622 500.5445   2.815135
#> 623 500.0707   2.645181
#> 624 500.5240   2.927411
#> 625 500.3580   2.977154
#> 626 501.3878   3.061005
#> 627 510.6775   2.629624
#> 628 502.6555   3.753056
#> 629 529.1209   2.672406
#> 630 500.5066   3.316588
#> 631 582.3924   2.646111
#> 632 507.3479   2.620810
#> 633 701.5866   2.899824
#> 634 500.0652   2.566076
#> 635 508.1782   2.694589
#> 636 500.2454   2.620338
#> 637 518.4024   2.733039
#> 638 520.9594   2.568562
#> 639 510.6704   2.690818
#> 640 542.4785   2.553556
#> 641 500.2317   2.790622
#> 642 512.8386   2.538352
#> 643 501.5054   3.379896
#> 644 730.4303   3.051385
#> 645 500.0101   2.909592
#> 646 500.1331   2.876770
#> 647 573.3707   4.106691
#> 648 501.3785   2.656775
#> 649 562.4764   2.836979
#> 650 500.4547   2.879073
#> 651 500.3900   2.986948
#> 652 501.0802   2.599183
#> 653 500.4916   2.965470
#> 654 500.9258   2.523079
#> 655 500.2387   2.612966
#> 656 512.4142   2.527401
#> 657 950.5935   2.843746
#> 658 525.4928   3.175744
#> 659 500.8291   2.831659
#> 660 500.4364   3.513559
#> 661 501.4161   3.868002
#> 662 500.0487   3.178812
#> 663 509.0657   2.576349
#> 664 500.2473   2.562633
#> 665 507.1067   4.327743
#> 666 506.6747   2.508276
#> 667 605.7369   2.832439
#> 668 501.7800   2.802543
#> 669 505.3820   2.686103
#> 670 500.1316   2.882729
#> 671 500.1196   2.620545
#> 672 500.5314   2.803010
#> 673 501.0708   3.491199
#> 674 936.0776   2.610198
#> 675 513.7697   2.505908
#> 676 502.2730   2.681559
#> 677 500.0138   2.952368
#> 678 502.1930   2.620477
#> 679 615.4693   3.544016
#> 680 965.7482   3.842947
#> 681 500.1364   3.920002
#> 682 552.9174   2.585891
#> 683 500.6702   3.234510
#> 684 500.3706   3.011863
#> 685 500.1273   2.949744
#> 686 516.2425   3.266622
#> 687 508.5606   2.929107
#> 688 500.2820   3.458689
#> 689 649.8434   2.653441
#> 690 501.6572   2.579668
#> 691 500.5425   2.860183
#> 692 628.4412   2.627300
#> 693 500.5512   2.700083
#> 694 507.0729   2.552085
#> 695 502.1058   2.783015
#> 696 500.0767   2.535026
#> 697 645.9746   2.838751
#> 698 501.6568   2.834077
#> 699 835.8521   3.201550
#> 700 512.6458   4.327783
#> 701 577.7316   3.568148
#> 702 565.6658   2.981316
#> 703 509.9037   3.579057
#> 704 522.7247   2.766780
#> 705 501.9967   2.575695
#> 706 500.3849   2.913009
#> 707 611.6767   2.784124
#> 708 500.3488   2.917915
#> 709 706.7674   2.529018
#> 710 502.7757   2.641329
#> 711 978.8822   2.658164
#> 712 508.2561   2.561442
#> 713 507.4423   3.304278
#> 714 555.0199   2.703479
#> 715 500.0699   2.762847
#> 716 500.1023   2.658843
#> 717 500.0394   2.698158
#> 718 502.4189   3.451511
#> 719 500.0687   2.872131
#> 720 500.1454   2.558124
#> 721 504.0801   3.426294
#> 722 500.6098   2.908986
#> 723 571.5927   2.627946
#> 724 500.5540   2.891141
#> 725 501.7842   2.732431
#> 726 500.1127   3.523705
#> 727 502.6952   3.368192
#> 728 688.3188   2.961994
#> 729 503.1505   2.636830
#> 730 938.9439   2.941465
#> 731 506.5811   2.877122
#> 732 831.2336   2.632291
#> 733 501.2224   2.628560
#> 734 506.5202   2.955872
#> 735 500.1529   3.749464
#> 736 543.8763   3.376471
#> 737 667.8855   2.538957
#> 738 945.1974   2.997454
#> 739 516.1156   3.706028
#> 740 553.9779   2.769097
#> 741 515.6343   2.776965
#> 742 500.1128   2.924793
#> 743 500.8587   2.791369
#> 744 554.0785   2.506705
#> 745 534.8791   2.971392
#> 746 650.6846   2.898876
#> 747 526.4930   3.716037
#> 748 568.5071   2.563894
#> 749 500.0066   2.913289
#> 750 500.6466   2.931258
#> 751 502.3263   3.775120
#> 752 500.0281   2.584352
#> 753 522.9133   2.957724
#> 754 500.1952   2.566494
#> 755 506.7996   2.516449
#> 756 582.9009   2.561758
#> 757 523.0511   3.000421
#> 758 506.9964   2.835994
#> 759 502.2808   2.731377
#> 760 521.8564   2.664595
#> 761 514.2219   2.676275
#> 762 501.4658   2.571490
#> 763 500.2495   2.546153
#> 764 500.7706   2.615038
#> 765 500.0933   2.574062
#> 766 504.0710   3.399208
#> 767 500.4043   3.776411
#> 768 500.0671   3.377419
#> 769 500.0918   2.681782
#> 770 580.6860   2.935375
#> 771 506.6465   2.878988
#> 772 501.9243   3.638319
#> 773 500.1795   2.888039
#> 774 539.8018   3.301304
#> 775 531.7516   2.557270
#> 776 502.1367   2.650603
#> 777 500.0902   3.331862
#> 778 505.7640   3.882587
#> 779 510.1709   3.355497
#> 780 576.6238   2.886490
#> 781 629.0534   3.063479
#> 782 500.0739   2.684213
#> 783 518.1824   3.191225
#> 784 856.6078   2.524034
#> 785 615.0361   3.081996
#> 786 505.2452   2.536986
#> 787 546.1129   2.625774
#> 788 503.6255   2.638388
#> 789 512.7911   4.064197
#> 790 501.5973   2.505375
#> 791 514.7335   2.567831
#> 792 500.0246   2.505776
#> 793 500.0565   3.268564
#> 794 501.4017   2.519534
#> 795 504.2520   2.550841
#> 796 500.4237   2.748924
#> 797 500.1993   2.783344
#> 798 581.8937   2.601150
#> 799 500.3697   4.187815
#> 800 501.0323   2.511206
#> 801 538.7456   2.890904
#> 802 500.1680   3.955263
#> 803 500.0413   3.020490
#> 804 501.0289   2.924402
#> 805 513.2326   2.845040
#> 806 642.7773   3.185774
#> 807 502.2786   4.664000
#> 808 507.7581   3.582487
#> 809 917.1192   3.437629
#> 810 501.5773   2.500356
#> 811 500.0692   3.089788
#> 812 506.9626   2.535060
#> 813 529.2354   3.353523
#> 814 559.5994   3.297985
#> 815 527.9142   2.819773
#> 816 500.9495   2.986641
#> 817 504.3979   3.699505
#> 818 543.4940   2.826053
#> 819 549.0654   2.992077
#> 820 561.5780   2.756876
#> 821 521.0118   3.650027
#> 822 566.6448   3.652349
#> 823 500.0016   3.249046
#> 824 544.3214   3.249853
#> 825 559.0186   2.705220
#> 826 678.3478   3.787120
#> 827 502.4366   2.802877
#> 828 501.5186   2.991969
#> 829 572.6462   3.758163
#> 830 500.3868   3.110687
#> 831 572.6142   2.608922
#> 832 529.0304   2.604446
#> 833 500.2065   2.726113
#> 834 524.9840   2.946084
#> 835 512.3158   3.000634
#> 836 500.1202   2.516223
#> 837 797.5585   2.888643
#> 838 502.7499   3.324184
#> 839 506.5765   2.544991
#> 840 500.0696   2.588858
#> 841 502.7417   2.548530
#> 842 500.1439   2.585539
#> 843 501.2955   2.661066
#> 844 512.2851   3.436025
#> 845 679.2363   2.569421
#> 846 559.2035   2.813872
#> 847 500.1184   2.848462
#> 848 500.0932   3.055889
#> 849 502.2531   2.606098
#> 850 943.7614   2.572035
#> 851 505.2664   2.841417
#> 852 503.4368   2.802655
#> 853 512.7683   2.886409
#> 854 500.1508   4.039030
#> 855 505.9800   2.960378
#> 856 616.8600   3.161356
#> 857 501.0320   2.572408
#> 858 635.3630   2.612371
#> 859 500.2589   3.284970
#> 860 502.6903   3.225314
#> 861 503.2694   2.600810
#> 862 861.1049   2.887325
#> 863 501.6337   3.311883
#> 864 503.3179   3.893572
#> 865 501.5518   3.233238
#> 866 998.5124   2.585264
#> 867 510.6698   2.527559
#> 868 500.3874   2.628275
#> 869 552.2831   2.671910
#> 870 573.7868   2.796593
#> 871 500.0647   2.973329
#> 872 500.4435   2.997247
#> 873 553.0514   3.031662
#> 874 501.6793   2.766083
#> 875 500.0611   2.738973
#> 876 519.0686   3.220182
#> 877 500.2349   2.762949
#> 878 513.8304   2.563115
#> 879 500.0905   2.642342
#> 880 500.4175   2.546405
#> 881 512.4036   2.810309
#> 882 500.5983   2.688526
#> 883 500.4125   2.990572
#> 884 885.8887   3.433203
#> 885 504.3359   3.384176
#> 886 504.3745   2.762243
#> 887 508.5678   3.095422
#> 888 502.0122   3.388771
#> 889 500.0913   3.152245
#> 890 500.4614   3.416966
#> 891 570.0535   2.764198
#> 892 500.1501   2.548343
#> 893 500.0603   2.534589
#> 894 544.8727   2.724674
#> 895 573.1929   2.560394
#> 896 785.0969   2.549062
#> 897 500.0466   2.514976
#> 898 503.5084   2.755691
#> 899 502.9549   2.759189
#> 900 500.1492   2.801330
#> 901 562.3975   2.872258
#> 902 531.1336   2.696372
#> 903 500.3381   3.047744
#> 904 500.0091   2.517501
#> 905 500.0546   2.636527
#> 906 528.4430   2.546310
#> 907 503.7900   2.717019
#> 908 500.1268   3.083804
#> 909 508.7134   2.654051
#> 910 500.6779   2.614463
#> 911 500.9453   2.967705
#> 912 500.2821   2.554694
#> 913 500.1383   2.745727
#> 914 505.7006   2.638857
#> 915 500.0200   2.786162
#> 916 501.8587   2.586025