Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   506.3089   2.847428
#> 2   502.4048   2.515663
#> 3   504.0360   4.143893
#> 4   958.7583   2.642234
#> 5   501.6032   2.676955
#> 6   551.8091   2.508773
#> 7   531.0938   2.924221
#> 8   500.1681   2.808409
#> 9   518.7723   3.098167
#> 10  501.0815   2.517468
#> 11  502.1969   3.735191
#> 12  501.7122   2.505828
#> 13  511.5852   4.817668
#> 14  500.0309   2.830370
#> 15  640.3092   2.707311
#> 16  770.9952   2.721937
#> 17  508.1149   3.046749
#> 18  500.3030   2.692531
#> 19  500.0973   2.551088
#> 20  511.9255   2.738229
#> 21  500.0326   2.804130
#> 22  637.5162   3.374915
#> 23  500.7281   2.686736
#> 24  531.0709   2.633068
#> 25  585.5868   4.026620
#> 26  543.4399   3.130985
#> 27  500.5958   2.624127
#> 28  502.0608   3.081975
#> 29  500.2705   3.363248
#> 30  508.5019   2.696047
#> 31  501.9080   2.579489
#> 32  504.2727   3.124310
#> 33  502.5779   2.958376
#> 34  525.8013   2.706427
#> 35  510.5313   2.846011
#> 36  500.2677   2.816047
#> 37  501.9849   3.155989
#> 38  500.5100   3.030881
#> 39  500.2715   2.687420
#> 40  501.0919   3.279748
#> 41  501.7941   3.524541
#> 42  500.7641   4.000057
#> 43  534.9430   3.084164
#> 44  501.5362   3.170257
#> 45  500.0600   3.773073
#> 46  507.4623   2.565210
#> 47  931.9042   2.607073
#> 48  562.5536   2.691881
#> 49  826.7769   2.728731
#> 50  505.7843   2.999493
#> 51  502.9721   2.579307
#> 52  523.8204   2.643125
#> 53  756.4235   2.607371
#> 54  501.8853   2.711765
#> 55  500.5412   2.775695
#> 56  525.2295   2.859836
#> 57  500.5880   3.253699
#> 58  503.5982   3.107420
#> 59  500.4474   3.472428
#> 60  756.9799   2.619815
#> 61  500.0822   3.479422
#> 62  605.9630   3.873343
#> 63  502.4958   2.899204
#> 64  557.7446   2.604923
#> 65  500.0136   2.725191
#> 66  755.5755   3.137354
#> 67  527.1577   2.844300
#> 68  500.0214   3.882429
#> 69  500.0195   2.944273
#> 70  500.8681   3.217554
#> 71  578.4292   2.987446
#> 72  506.7349   2.899778
#> 73  659.0385   2.619482
#> 74  525.9449   2.560958
#> 75  500.0999   2.739872
#> 76  501.0658   2.773533
#> 77  500.8247   3.205313
#> 78  512.3396   2.879257
#> 79  517.9076   2.823210
#> 80  500.0247   2.963750
#> 81  508.5242   2.918844
#> 82  510.3840   2.706509
#> 83  574.9930   3.079453
#> 84  500.7231   2.672767
#> 85  500.0330   2.672527
#> 86  559.2701   3.243328
#> 87  500.0345   3.022882
#> 88  500.1127   2.663939
#> 89  514.4955   3.083424
#> 90  500.4884   2.788399
#> 91  500.3726   2.602498
#> 92  500.4728   2.940045
#> 93  501.4895   2.641308
#> 94  501.1680   2.549827
#> 95  505.8119   3.714915
#> 96  553.1376   2.942349
#> 97  529.5368   2.984095
#> 98  500.5216   3.254513
#> 99  500.0253   2.675832
#> 100 500.1853   5.349105
#> 101 529.6727   3.513942
#> 102 637.1832   2.979566
#> 103 504.2888   2.542685
#> 104 507.5699   2.544900
#> 105 500.1963   2.894298
#> 106 557.0708   2.737583
#> 107 514.7942   3.339838
#> 108 500.6337   2.738156
#> 109 510.8851   2.747365
#> 110 506.1236   3.506700
#> 111 502.2032   2.890876
#> 112 500.0144   2.707921
#> 113 684.0297   3.255691
#> 114 504.8632   3.636277
#> 115 502.7171   3.118073
#> 116 648.2544   3.094637
#> 117 500.1759   3.257312
#> 118 509.3991   2.645978
#> 119 500.0161   3.561457
#> 120 520.5052   2.936469
#> 121 580.1892   2.564885
#> 122 500.1502   3.152176
#> 123 558.0924   2.559865
#> 124 553.6339   2.661825
#> 125 500.3440   3.137900
#> 126 500.0280   2.642420
#> 127 506.1713   2.710178
#> 128 500.3079   2.611635
#> 129 500.6569   2.807405
#> 130 500.2574   2.949626
#> 131 520.4632   2.578232
#> 132 500.0059   2.752371
#> 133 500.6797   2.865063
#> 134 500.6997   3.341301
#> 135 530.2246   2.611717
#> 136 502.9313   2.632803
#> 137 501.3623   2.515600
#> 138 501.2677   3.007570
#> 139 500.6777   3.518799
#> 140 564.4730   2.522349
#> 141 671.3126   4.623771
#> 142 500.4075   2.708811
#> 143 500.7949   2.797029
#> 144 500.3646   2.921227
#> 145 850.0249   3.321026
#> 146 500.8079   2.662155
#> 147 532.8028   3.270807
#> 148 500.0096   3.329007
#> 149 500.2363   3.736363
#> 150 500.5468   3.208173
#> 151 581.4762   2.533034
#> 152 501.4914   2.801185
#> 153 500.5065   2.513661
#> 154 674.6965   2.622650
#> 155 913.5667   2.628990
#> 156 502.9431   2.927835
#> 157 530.9384   2.906812
#> 158 500.1842   2.854043
#> 159 500.0502   2.511644
#> 160 503.1848   3.115514
#> 161 501.3741   3.280668
#> 162 533.3771   2.913774
#> 163 738.2489   2.822135
#> 164 509.5225   2.560777
#> 165 500.3412   2.951236
#> 166 501.7075   3.307302
#> 167 699.5295   2.860088
#> 168 500.1510   3.092799
#> 169 500.3195   3.105655
#> 170 514.3512   2.745383
#> 171 523.6862   2.556808
#> 172 573.8115   2.791089
#> 173 855.5510   2.539714
#> 174 508.4473   2.748555
#> 175 543.5605   2.799090
#> 176 632.0594   2.951489
#> 177 501.7704   2.678951
#> 178 500.1202   2.670317
#> 179 610.3612   3.440824
#> 180 504.6851   3.271750
#> 181 501.1649   2.734621
#> 182 503.9356   2.541110
#> 183 502.4148   2.649320
#> 184 501.5237   2.688811
#> 185 506.4144   2.812474
#> 186 506.6759   2.749038
#> 187 505.2261   2.955404
#> 188 503.7512   3.425327
#> 189 500.0231   4.059260
#> 190 500.1720   2.652913
#> 191 500.4842   2.699602
#> 192 500.7597   2.687766
#> 193 500.4878   2.862470
#> 194 500.9590   4.499601
#> 195 511.8869   2.578246
#> 196 579.2644   3.231639
#> 197 500.7628   2.565376
#> 198 513.8841   2.551640
#> 199 500.4982   2.622511
#> 200 500.4038   3.465716
#> 201 794.8650   3.405070
#> 202 500.1704   2.941226
#> 203 500.0530   2.508733
#> 204 573.8788   3.277950
#> 205 500.4232   2.805638
#> 206 505.5516   3.119477
#> 207 913.4101   2.601900
#> 208 581.2638   2.761584
#> 209 500.0798   2.598472
#> 210 529.6143   2.756632
#> 211 782.7572   3.582488
#> 212 500.0794   3.069849
#> 213 518.4483   3.251959
#> 214 501.5233   3.088757
#> 215 500.3764   3.124461
#> 216 501.7022   4.029192
#> 217 756.0644   2.518496
#> 218 607.9008   3.163614
#> 219 971.4295   2.904551
#> 220 501.0468   2.641857
#> 221 500.9013   3.796099
#> 222 500.0285   2.733064
#> 223 501.0771   2.750495
#> 224 535.3087   2.718990
#> 225 512.9904   3.018163
#> 226 605.9800   2.700479
#> 227 508.3030   2.541000
#> 228 504.4382   2.636956
#> 229 537.8555   3.068537
#> 230 503.6038   2.565863
#> 231 500.5812   2.529446
#> 232 752.7462   3.905389
#> 233 508.7821   3.068612
#> 234 501.3443   3.959388
#> 235 500.6926   3.110866
#> 236 684.3539   2.660058
#> 237 500.3845   3.997252
#> 238 500.0159   4.250992
#> 239 534.7017   2.817342
#> 240 628.8893   3.611574
#> 241 500.1826   2.611369
#> 242 500.7352   2.946167
#> 243 500.0069   2.628813
#> 244 506.7004   2.634720
#> 245 516.4084   3.090404
#> 246 500.9969   2.614138
#> 247 500.4890   3.010720
#> 248 502.2784   2.507321
#> 249 537.7317   2.570600
#> 250 628.4940   3.361015
#> 251 500.7855   2.820141
#> 252 500.5965   3.117021
#> 253 500.0510   3.381635
#> 254 513.2531   3.102021
#> 255 500.8168   3.223985
#> 256 506.4262   2.891964
#> 257 500.2547   3.302553
#> 258 500.0165   3.420849
#> 259 500.2090   3.013658
#> 260 588.2785   2.569098
#> 261 743.0645   3.447374
#> 262 501.1247   2.624013
#> 263 500.0744   2.829902
#> 264 500.3839   2.818058
#> 265 501.2461   2.907176
#> 266 502.7826   3.321716
#> 267 598.4290   3.554113
#> 268 696.9037   2.550707
#> 269 504.6463   2.529365
#> 270 516.9949   2.819914
#> 271 500.5327   2.553356
#> 272 500.0889   2.568863
#> 273 501.4696   2.527098
#> 274 500.1466   3.239384
#> 275 500.7042   2.632575
#> 276 502.3441   3.451713
#> 277 508.3613   2.995792
#> 278 500.0893   2.768386
#> 279 515.3274   2.618782
#> 280 515.7432   2.546948
#> 281 504.0629   2.995830
#> 282 814.0254   3.224403
#> 283 515.8718   3.409832
#> 284 500.2239   2.530548
#> 285 503.6514   2.594401
#> 286 501.2282   2.640749
#> 287 500.0876   3.108882
#> 288 500.7955   2.607122
#> 289 516.1257   2.648582
#> 290 500.7980   2.832845
#> 291 518.6874   2.722912
#> 292 773.1541   2.588340
#> 293 501.4343   3.492710
#> 294 500.6074   3.466451
#> 295 500.7919   2.527062
#> 296 502.8499   2.836037
#> 297 502.0439   2.826771
#> 298 500.3263   2.978798
#> 299 509.8395   2.567659
#> 300 500.6107   2.580317
#> 301 501.9331   2.547006
#> 302 508.1094   2.532287
#> 303 597.9920   2.822797
#> 304 511.1019   2.580108
#> 305 513.1403   3.122809
#> 306 501.0743   4.023852
#> 307 527.4694   2.590671
#> 308 500.1527   4.766081
#> 309 505.0890   2.990751
#> 310 500.7999   2.544897
#> 311 503.1061   2.850939
#> 312 502.4623   2.697539
#> 313 500.0226   2.527774
#> 314 501.1079   3.009509
#> 315 500.3474   2.686191
#> 316 516.7005   2.937241
#> 317 500.0827   2.935386
#> 318 514.7577   2.695640
#> 319 540.9774   2.561132
#> 320 506.4174   2.779355
#> 321 500.0281   2.814725
#> 322 500.1672   2.813659
#> 323 501.0532   2.802727
#> 324 500.8244   2.838483
#> 325 779.2998   2.603490
#> 326 919.1611   2.834719
#> 327 500.4485   2.581358
#> 328 500.1147   2.549036
#> 329 500.3256   2.849615
#> 330 502.4013   3.207598
#> 331 777.1293   2.604011
#> 332 789.7436   2.702733
#> 333 501.1486   3.442686
#> 334 500.0189   3.333998
#> 335 889.4000   2.568088
#> 336 502.7954   2.709886
#> 337 671.8616   2.718318
#> 338 540.7818   2.521990
#> 339 500.0978   2.540027
#> 340 612.3529   2.620671
#> 341 506.4539   3.030236
#> 342 500.4302   2.942556
#> 343 507.4954   2.847846
#> 344 505.4322   2.739532
#> 345 501.1607   2.755977
#> 346 501.0544   2.655930
#> 347 507.8144   3.249740
#> 348 958.3535   2.768057
#> 349 505.5652   2.514264
#> 350 500.1244   2.648993
#> 351 567.2565   3.263720
#> 352 504.4438   2.992024
#> 353 509.4174   4.163106
#> 354 501.1450   2.658377
#> 355 500.0897   2.702577
#> 356 520.2827   2.917931
#> 357 507.7531   2.565361
#> 358 511.3624   3.430725
#> 359 504.0441   3.109022
#> 360 500.1495   2.551669
#> 361 500.0025   3.927146
#> 362 538.6670   3.535199
#> 363 520.6504   3.399488
#> 364 500.4950   3.126177
#> 365 501.8465   2.834763
#> 366 500.0081   4.725431
#> 367 831.7728   2.506805
#> 368 500.8138   2.931922
#> 369 500.0841   3.011089
#> 370 500.0333   2.923758
#> 371 506.1767   2.990983
#> 372 500.3735   3.233287
#> 373 500.0261   2.979625
#> 374 500.0013   2.921970
#> 375 572.6025   2.644664
#> 376 502.8188   2.662666
#> 377 500.3892   2.832356
#> 378 508.8375   2.827553
#> 379 501.4163   2.770268
#> 380 500.1227   2.874125
#> 381 556.1856   2.789831
#> 382 511.5318   2.739863
#> 383 556.7471   2.530029
#> 384 508.4127   2.531532
#> 385 541.1714   2.564008
#> 386 518.7442   2.894597
#> 387 500.0610   3.079958
#> 388 506.7279   2.726108
#> 389 539.3658   2.822564
#> 390 500.2409   3.223322
#> 391 578.4286   2.718062
#> 392 723.5136   2.984336
#> 393 655.4488   2.722941
#> 394 542.3706   2.885507
#> 395 516.6645   2.816569
#> 396 505.4575   3.398211
#> 397 509.1958   2.916141
#> 398 501.8595   3.042642
#> 399 507.4152   2.563072
#> 400 608.2938   2.544445
#> 401 532.4853   2.871950
#> 402 502.3940   3.255584
#> 403 503.7374   3.097443
#> 404 500.0984   3.547115
#> 405 500.4590   4.151932
#> 406 500.0086   2.583873
#> 407 500.1335   4.605899
#> 408 500.0206   2.949323
#> 409 501.8762   2.592003
#> 410 533.4961   3.602001
#> 411 500.4638   2.567679
#> 412 555.8039   2.663503
#> 413 500.0515   2.712177
#> 414 518.5581   4.322918
#> 415 691.3665   2.783569
#> 416 500.1916   4.003079
#> 417 505.1070   2.982730
#> 418 704.3518   2.512986
#> 419 503.1958   2.707597
#> 420 500.3258   3.781163
#> 421 530.3852   2.584605
#> 422 501.7858   2.736779
#> 423 504.5315   2.850077
#> 424 504.6982   4.007342
#> 425 505.5786   2.562148
#> 426 539.0689   3.032634
#> 427 506.6062   5.159776
#> 428 879.1208   2.635031
#> 429 500.0000   3.325916
#> 430 500.4063   2.815702
#> 431 552.1480   3.340051
#> 432 505.7880   2.941434
#> 433 500.0995   3.511776
#> 434 946.7859   3.161683
#> 435 500.7444   2.604523
#> 436 501.8324   3.957308
#> 437 500.0657   2.662918
#> 438 536.2720   2.860478
#> 439 579.7449   3.752098
#> 440 877.3135   2.681283
#> 441 709.3109   2.787276
#> 442 504.0902   2.584650
#> 443 854.6120   2.891355
#> 444 500.0292   4.239619
#> 445 580.9548   2.857619
#> 446 974.4748   2.874255
#> 447 500.1551   2.549248
#> 448 771.4155   2.856468
#> 449 500.0324   2.614346
#> 450 500.9191   2.629626
#> 451 550.0955   2.634504
#> 452 513.5132   3.130966
#> 453 530.3863   2.623251
#> 454 504.3040   3.204951
#> 455 501.1721   2.530244
#> 456 531.2271   2.639958
#> 457 500.0235   2.605792
#> 458 502.7449   2.612615
#> 459 500.4753   3.105446
#> 460 500.6582   2.700762
#> 461 536.2295   2.659149
#> 462 501.7310   3.012556
#> 463 516.6133   3.068670
#> 464 536.8370   2.744494
#> 465 500.8971   2.502641
#> 466 502.6408   2.500621
#> 467 501.9468   3.940133
#> 468 513.1371   3.960694
#> 469 505.1234   2.853665
#> 470 502.4673   2.563541
#> 471 844.7632   3.254760
#> 472 509.3039   2.676489
#> 473 993.6668   3.143602
#> 474 510.2710   3.087959
#> 475 530.0856   2.984071
#> 476 653.5544   2.615398
#> 477 508.9343   3.003197
#> 478 953.0037   2.936645
#> 479 612.1695   3.516020
#> 480 680.9776   3.060011
#> 481 582.7744   2.951042
#> 482 986.2721   2.912261
#> 483 501.3815   2.783770
#> 484 510.2498   2.736261
#> 485 537.6977   3.372413
#> 486 503.4614   2.671110
#> 487 501.2313   3.165466
#> 488 690.2003   2.838466
#> 489 555.1995   3.186736
#> 490 500.8982   2.565382
#> 491 551.5550   3.295241
#> 492 589.0322   2.800045
#> 493 500.2567   2.631556
#> 494 822.2813   2.650487
#> 495 500.1662   2.519037
#> 496 501.1414   3.081864
#> 497 500.5874   2.540117
#> 498 732.9938   3.120457
#> 499 513.9129   2.602005
#> 500 580.3012   3.195417
#> 501 500.3160   3.044145
#> 502 502.0315   2.622349
#> 503 500.3836   2.793472
#> 504 958.5575   2.961176
#> 505 500.2792   3.010176
#> 506 509.5592   2.839278
#> 507 500.2102   3.086365
#> 508 506.8589   2.534817
#> 509 500.9153   2.534489
#> 510 500.6781   2.546801
#> 511 502.2538   2.520926
#> 512 528.3670   2.603513
#> 513 500.4020   2.617612
#> 514 501.0176   2.531564
#> 515 572.5409   3.355653
#> 516 501.8362   2.530229
#> 517 562.2570   3.184048
#> 518 505.0361   2.608894
#> 519 500.3030   2.509914
#> 520 537.7907   2.564813
#> 521 503.1016   3.672011
#> 522 500.0614   3.442210
#> 523 858.1001   3.207663
#> 524 519.7916   3.107549
#> 525 502.1614   3.314945
#> 526 500.0156   2.970414
#> 527 500.4403   2.982844
#> 528 775.4007   2.590472
#> 529 657.2009   2.642268
#> 530 856.5439   2.532827
#> 531 500.8059   2.711705
#> 532 500.7514   3.072790
#> 533 558.6542   2.517224
#> 534 502.5362   4.162226
#> 535 539.1893   2.671023
#> 536 509.0735   3.561391
#> 537 516.7949   2.977043
#> 538 502.5326   2.718779
#> 539 500.1731   2.638420
#> 540 507.8075   3.351863
#> 541 501.0415   3.627555
#> 542 501.5759   3.013353
#> 543 500.6606   3.254515
#> 544 515.8314   3.377772
#> 545 500.0724   2.544097
#> 546 526.8477   2.825020
#> 547 891.0723   3.075072
#> 548 729.1903   3.204198
#> 549 501.1248   2.596083
#> 550 612.9479   2.640213
#> 551 589.2370   2.550940
#> 552 500.5518   2.714178
#> 553 636.8468   2.717177
#> 554 534.6046   2.540167
#> 555 502.9705   2.565391
#> 556 500.2278   2.597856
#> 557 620.4161   2.657289
#> 558 666.7709   2.634059
#> 559 500.2504   3.416692
#> 560 512.5744   3.642236
#> 561 500.0806   3.150721
#> 562 501.3390   3.370395
#> 563 508.4663   6.756116
#> 564 507.7643   2.559561
#> 565 500.5906   2.591428
#> 566 551.8333   4.244871
#> 567 500.5542   2.606341
#> 568 576.1828   3.491149
#> 569 789.2776   3.802425
#> 570 500.1576   2.613663
#> 571 508.4320   2.502442
#> 572 500.0655   2.820941
#> 573 521.1596   2.806829
#> 574 501.3201   3.924018
#> 575 502.8852   2.791769
#> 576 500.2146   2.843201
#> 577 500.0735   2.844431
#> 578 521.7533   2.993121
#> 579 535.1743   2.512207
#> 580 500.6746   2.656001
#> 581 694.8602   2.876186
#> 582 500.1464   3.722701
#> 583 852.3287   2.840378
#> 584 500.0701   2.681511
#> 585 525.2285   2.803797
#> 586 500.9457   2.628097
#> 587 500.0952   2.978650
#> 588 529.6939   2.703681
#> 589 653.7271   2.542904
#> 590 503.5272   2.840213
#> 591 505.7428   2.518123
#> 592 564.1989   3.356508
#> 593 502.0478   2.552628
#> 594 500.5057   3.358059
#> 595 786.4437   2.622866
#> 596 507.6976   2.599023
#> 597 500.0631   3.331590
#> 598 500.1579   2.659740
#> 599 567.3637   2.805795
#> 600 506.1860   3.546421
#> 601 591.7518   2.559761
#> 602 501.2559   2.509248
#> 603 503.9086   2.716710
#> 604 500.0858   2.956811
#> 605 502.1977   2.662132
#> 606 518.3082   2.612693
#> 607 502.8266   2.882760
#> 608 956.7328   2.680899
#> 609 500.2987   2.515224
#> 610 500.6861   3.169631
#> 611 501.3268   2.699391
#> 612 561.1094   2.527232
#> 613 519.3063   3.413931
#> 614 503.0737   4.382956
#> 615 530.1171   2.543331
#> 616 500.1014   3.269668
#> 617 501.1875   3.010627
#> 618 508.5965   2.616596
#> 619 501.1382   2.855430
#> 620 500.1085   2.710133
#> 621 502.8235   2.522178
#> 622 500.0602   3.784814
#> 623 500.2791   2.699416
#> 624 501.2208   2.616105
#> 625 519.4694   2.848747
#> 626 504.8830   2.867001
#> 627 563.3677   2.728699
#> 628 593.4723   3.053487
#> 629 500.8240   3.066476
#> 630 817.6088   2.561853
#> 631 519.1754   2.624569
#> 632 500.2110   2.936896
#> 633 500.4237   2.681106
#> 634 860.1371   2.515492
#> 635 738.0327   2.508102
#> 636 515.1397   3.010631
#> 637 500.4065   3.781688
#> 638 500.1043   2.758427
#> 639 554.8672   2.566672
#> 640 500.0270   2.636942
#> 641 500.0585   2.779837
#> 642 500.4980   2.634604
#> 643 500.1763   2.541687
#> 644 507.3816   2.607850
#> 645 500.3386   2.640753
#> 646 508.2651   2.928498
#> 647 500.0330   2.689676
#> 648 500.5299   2.692860
#> 649 500.0112   3.491351
#> 650 501.3063   3.158465
#> 651 500.0065   3.725067
#> 652 526.5277   2.715662
#> 653 500.3190   2.522520
#> 654 502.7816   3.142471
#> 655 576.8562   2.744665
#> 656 626.1289   2.878692
#> 657 502.4999   3.242233
#> 658 500.0692   2.934256
#> 659 804.4190   2.781325
#> 660 500.8530   2.684020
#> 661 500.4102   2.682159
#> 662 755.1642   2.512131
#> 663 500.0204   2.761924
#> 664 501.6795   2.732631
#> 665 500.0496   2.651055
#> 666 699.4039   4.076281
#> 667 518.3138   2.924861
#> 668 832.8516   2.713285
#> 669 527.2412   3.264225
#> 670 537.3421   2.773446
#> 671 501.9997   2.789621
#> 672 529.8486   2.906188
#> 673 501.3884   2.946015
#> 674 500.3146   3.777993
#> 675 504.4220   3.116312
#> 676 513.9083   2.828920
#> 677 502.0353   2.793740
#> 678 507.2647   2.678722
#> 679 724.6812   2.919525
#> 680 500.3261   3.092730
#> 681 650.9900   2.531855
#> 682 575.5126   2.504673
#> 683 601.2337   3.437698
#> 684 503.2038   2.601886
#> 685 507.3388   2.866232
#> 686 500.2291   2.505623
#> 687 783.7801   3.132279
#> 688 500.5164   4.075865
#> 689 650.7401   2.730630
#> 690 500.1814   3.216718
#> 691 999.1116   4.346085
#> 692 500.1467   3.137092
#> 693 903.5265   3.045458
#> 694 500.2432   5.744709
#> 695 603.0014   2.505420
#> 696 504.8806   2.997923
#> 697 514.7892   2.525396
#> 698 503.2550   2.730359
#> 699 560.2234   3.200392
#> 700 500.5895   3.414543
#> 701 627.5254   2.914096
#> 702 502.9877   3.484023
#> 703 878.4035   2.991205
#> 704 500.7436   2.924016
#> 705 779.6681   2.808842
#> 706 500.6283   2.753440
#> 707 500.0410   2.828617
#> 708 502.9102   2.736311
#> 709 540.4519   2.730393
#> 710 500.3916   2.531108
#> 711 537.2105   2.508279
#> 712 504.4570   2.510224
#> 713 500.1831   2.929363
#> 714 500.6238   2.958761
#> 715 500.1776   2.636824
#> 716 500.0279   2.744362
#> 717 571.1730   2.980865
#> 718 519.4450   3.167557
#> 719 567.4356   2.559070
#> 720 588.1927   2.845334
#> 721 622.9238   2.544652
#> 722 592.0263   2.967974
#> 723 500.0083   3.162935
#> 724 548.0653   2.659195
#> 725 500.1314   3.034435
#> 726 776.6467   2.665560
#> 727 503.3057   2.512493
#> 728 512.1585   2.523258
#> 729 549.4964   2.644330
#> 730 536.2983   2.630717
#> 731 500.9678   2.696202
#> 732 515.9047   2.645390
#> 733 500.0218   2.879855
#> 734 500.0566   2.642195
#> 735 508.9620   3.370151
#> 736 500.8682   2.638837
#> 737 500.6301   2.844785
#> 738 903.7587   3.245187
#> 739 500.1435   2.684733
#> 740 500.0119   2.728095
#> 741 500.1694   3.117932
#> 742 503.0493   3.238053
#> 743 500.0236   2.867033
#> 744 521.2914   2.599961
#> 745 503.6935   2.650278
#> 746 925.6435   2.877452
#> 747 540.5706   3.271760
#> 748 657.6485   2.605012
#> 749 500.6252   4.495434
#> 750 501.8026   2.739022
#> 751 523.6674   3.108861
#> 752 500.0145   3.064263
#> 753 517.0738   2.574282
#> 754 502.3885   2.876396
#> 755 509.4480   2.788091
#> 756 528.2780   2.856393
#> 757 538.6017   2.554740
#> 758 500.0205   2.500383
#> 759 504.7747   2.705478
#> 760 500.0261   2.664665
#> 761 501.8740   2.851755
#> 762 511.3596   3.086702
#> 763 500.1222   2.623838
#> 764 563.9313   2.931825
#> 765 500.1335   2.511158
#> 766 625.3564   2.939422
#> 767 500.0781   2.831137
#> 768 505.9501   2.570853
#> 769 515.8269   3.378449
#> 770 500.1963   3.749299
#> 771 501.5785   3.584822
#> 772 500.0642   2.639411
#> 773 500.5863   2.605126
#> 774 500.1099   2.833774
#> 775 504.1356   2.726494
#> 776 500.1423   3.237700
#> 777 541.9432   2.520684
#> 778 500.0080   2.767180
#> 779 500.1698   2.728695
#> 780 537.2027   3.178757
#> 781 510.5373   2.615895
#> 782 900.6656   2.794738
#> 783 522.2698   2.578317
#> 784 505.4701   2.501471
#> 785 541.1923   3.019102
#> 786 675.6673   2.773494
#> 787 500.1579   3.009403
#> 788 500.3705   3.094270
#> 789 510.5938   2.888269
#> 790 563.8823   2.868412
#> 791 500.0234   2.594179
#> 792 768.2027   4.275377
#> 793 500.1980   2.963533
#> 794 500.0467   3.688591
#> 795 523.8982   3.324172
#> 796 500.3708   3.009342
#> 797 501.1803   2.898128
#> 798 575.2224   2.527036
#> 799 501.0207   3.186292
#> 800 501.5033   3.396278
#> 801 500.0010   2.606217
#> 802 720.8174   2.693298
#> 803 500.7624   2.776920
#> 804 501.6041   2.555639
#> 805 582.2471   2.653804
#> 806 515.5620   2.927293
#> 807 845.5289   3.538210
#> 808 500.3270   3.867577
#> 809 500.6665   2.765486
#> 810 500.0012   3.022688
#> 811 502.4559   3.027124
#> 812 524.2044   3.606884
#> 813 513.2184   3.328013
#> 814 503.4305   2.811748
#> 815 513.9922   3.262013
#> 816 500.1126   3.496151
#> 817 575.1343   2.576379
#> 818 500.8307   2.531026
#> 819 501.7678   3.305076
#> 820 500.0326   2.853136
#> 821 508.3697   2.587965
#> 822 500.6109   3.161329
#> 823 602.0359   3.757838
#> 824 502.0281   2.517181
#> 825 566.6679   2.997765
#> 826 504.8896   2.606413
#> 827 607.0237   2.770957
#> 828 605.3716   2.575613
#> 829 568.3948   2.546493
#> 830 500.6796   3.589659
#> 831 574.4963   2.905534
#> 832 526.4884   2.853416
#> 833 509.6579   3.232290
#> 834 537.5111   2.859311
#> 835 502.9165   3.519196
#> 836 671.9049   2.506947
#> 837 753.0295   2.654682
#> 838 501.3289   2.650967
#> 839 516.7487   2.659467
#> 840 500.8877   3.556060
#> 841 500.7588   2.505167
#> 842 513.4634   2.968712
#> 843 548.5487   3.003669
#> 844 522.4389   2.521108
#> 845 519.3355   3.964693
#> 846 513.8058   2.718459
#> 847 501.3988   3.000297
#> 848 500.4442   3.052059
#> 849 611.7460   2.667302
#> 850 543.3394   2.595826
#> 851 500.2351   3.203667
#> 852 582.3245   2.563234
#> 853 571.5443   3.033391
#> 854 500.1858   2.641825
#> 855 601.3406   2.848383
#> 856 661.7966   3.363677
#> 857 785.1594   3.180701
#> 858 500.8189   2.812383
#> 859 500.0753   2.582149
#> 860 529.7896   2.572665
#> 861 606.1199   3.735304
#> 862 543.9042   2.507528
#> 863 918.6089   2.823711
#> 864 532.8833   2.741068
#> 865 500.7043   2.519013
#> 866 857.9549   2.836339
#> 867 500.6573   2.827366
#> 868 500.3011   3.321462
#> 869 500.1068   2.527450
#> 870 500.0510   2.536541
#> 871 500.3568   2.635997
#> 872 504.7856   2.729565
#> 873 508.7803   3.068952
#> 874 501.1036   3.127207
#> 875 500.7451   2.812643
#> 876 508.6828   4.235839
#> 877 503.2829   2.613587
#> 878 500.2575   2.919086
#> 879 929.9429   3.408866
#> 880 530.3304   2.713184
#> 881 500.1560   2.976241
#> 882 500.1553   2.749431
#> 883 501.3878   2.594300
#> 884 500.1096   2.705707
#> 885 540.9243   2.667870
#> 886 509.3592   2.532538
#> 887 500.4382   2.921213
#> 888 500.5565   2.647120
#> 889 500.2615   3.401399
#> 890 672.8739   2.554015
#> 891 972.6118   2.687583
#> 892 660.3734   3.390278
#> 893 506.5794   3.415285
#> 894 500.1266   3.532093
#> 895 500.2020   2.509794
#> 896 500.1528   2.991411
#> 897 560.4958   3.578279
#> 898 500.0759   2.515171
#> 899 500.5705   2.696887
#> 900 570.0650   2.670925
#> 901 520.6547   2.505946
#> 902 555.5394   2.580424
#> 903 500.5439   2.892472
#> 904 500.4518   3.285186
#> 905 500.4300   2.516999
#> 906 501.5098   3.230961
#> 907 500.5815   2.756029
#> 908 523.5140   2.992230
#> 909 500.0571   2.754879
#> 910 500.5472   3.059006
#> 911 592.5075   2.924088
#> 912 564.6098   3.263391
#> 913 500.2968   3.342347
#> 914 501.0869   2.658916