Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   500.9278   2.683948
#> 2   533.0772   2.592836
#> 3   550.0020   2.985128
#> 4   501.0480   2.917477
#> 5   508.8769   2.571587
#> 6   500.4581   2.945386
#> 7   527.6197   2.500793
#> 8   500.1013   2.514333
#> 9   549.7626   3.225350
#> 10  616.4736   2.590306
#> 11  501.6251   2.927888
#> 12  500.1162   2.530754
#> 13  535.0330   2.538086
#> 14  500.3997   2.593082
#> 15  500.0341   3.117281
#> 16  500.1204   3.161425
#> 17  500.6965   2.693995
#> 18  500.8398   2.844232
#> 19  507.3373   2.570096
#> 20  538.0033   3.472642
#> 21  508.4160   2.899214
#> 22  500.6038   3.021400
#> 23  506.8328   2.548260
#> 24  500.0761   2.606856
#> 25  508.4240   3.641548
#> 26  619.0158   2.710567
#> 27  515.2237   2.516483
#> 28  840.3670   2.519803
#> 29  500.5904   3.876210
#> 30  506.1918   2.654767
#> 31  566.4110   2.809295
#> 32  514.8809   2.966545
#> 33  558.5522   2.723986
#> 34  546.2037   2.831627
#> 35  516.5366   2.522016
#> 36  500.5101   2.752922
#> 37  503.0570   2.713703
#> 38  504.6806   2.561897
#> 39  500.4305   3.959726
#> 40  500.0446   2.796183
#> 41  506.2479   2.744436
#> 42  501.7505   3.654391
#> 43  501.1336   2.968424
#> 44  723.0518   3.456358
#> 45  500.2000   2.728132
#> 46  500.4960   2.677760
#> 47  500.5094   2.957222
#> 48  500.4229   3.695004
#> 49  628.7246   3.202533
#> 50  503.6948   2.767909
#> 51  532.1138   3.795185
#> 52  530.1511   2.881187
#> 53  500.0189   2.757291
#> 54  500.4181   2.814609
#> 55  500.1393   3.215290
#> 56  500.4188   3.666153
#> 57  735.8730   2.623522
#> 58  584.1505   2.743407
#> 59  502.6741   2.695818
#> 60  512.5448   3.031559
#> 61  513.3930   2.546127
#> 62  500.0969   2.917897
#> 63  500.9636   2.670566
#> 64  544.9545   2.561693
#> 65  503.2632   2.633692
#> 66  500.1447   4.283767
#> 67  512.2868   2.526524
#> 68  525.4905   2.853937
#> 69  502.2814   2.687252
#> 70  552.4612   2.879125
#> 71  518.4564   2.650157
#> 72  503.6189   3.782454
#> 73  737.3427   2.695345
#> 74  761.6273   2.510388
#> 75  751.2679   2.589895
#> 76  501.9045   2.558583
#> 77  721.9497   2.702591
#> 78  842.2351   2.768345
#> 79  537.9376   3.198348
#> 80  503.0090   2.627048
#> 81  887.5084   2.518188
#> 82  501.0347   2.641997
#> 83  504.2584   2.894652
#> 84  508.6149   2.867197
#> 85  501.8644   2.989865
#> 86  500.0647   2.687631
#> 87  500.3597   2.959745
#> 88  500.3838   2.665953
#> 89  705.5614   2.551969
#> 90  500.4191   2.577190
#> 91  500.1436   2.538768
#> 92  501.8445   2.981340
#> 93  641.8428   3.086291
#> 94  510.4937   2.983176
#> 95  540.8161   2.643397
#> 96  994.7269   2.777813
#> 97  502.4714   2.551223
#> 98  500.3604   2.533147
#> 99  500.2732   2.928904
#> 100 506.8724   3.916147
#> 101 633.7397   2.848033
#> 102 518.5910   2.565688
#> 103 502.9206   2.822300
#> 104 947.3857   3.654596
#> 105 500.6809   2.765182
#> 106 500.0663   3.849271
#> 107 501.3106   3.023054
#> 108 764.1933   2.568230
#> 109 530.9348   2.500361
#> 110 500.3473   2.635719
#> 111 507.7723   3.168206
#> 112 501.4411   2.851677
#> 113 641.0275   3.177078
#> 114 501.2412   4.238402
#> 115 500.6297   2.900432
#> 116 514.2153   2.752618
#> 117 500.3249   3.344531
#> 118 500.5574   2.615540
#> 119 500.2387   2.513463
#> 120 659.3747   3.112592
#> 121 500.7367   3.155981
#> 122 583.4007   2.879957
#> 123 506.5399   2.661630
#> 124 500.0614   3.112673
#> 125 500.1849   2.727790
#> 126 500.3987   2.669603
#> 127 506.6035   5.310733
#> 128 503.7919   2.534499
#> 129 503.8002   2.744259
#> 130 520.1767   2.704482
#> 131 500.6938   2.538281
#> 132 500.1122   2.802782
#> 133 630.2246   3.419679
#> 134 500.8061   3.007985
#> 135 500.1514   2.772417
#> 136 500.0696   2.718197
#> 137 500.0348   2.877381
#> 138 611.9489   2.886550
#> 139 560.8715   4.192572
#> 140 500.0031   3.154827
#> 141 501.2134   2.987366
#> 142 504.8227   2.587502
#> 143 500.7228   2.570696
#> 144 529.4137   3.962368
#> 145 500.4834   2.853474
#> 146 501.3286   2.546432
#> 147 532.6295   2.561988
#> 148 501.6858   2.656706
#> 149 502.3276   3.081124
#> 150 500.5440   3.328551
#> 151 503.6917   2.890197
#> 152 500.0070   3.003378
#> 153 500.0366   2.591081
#> 154 501.1128   2.741571
#> 155 500.9700   2.553622
#> 156 500.4398   2.704361
#> 157 500.5354   2.609453
#> 158 502.0000   3.239325
#> 159 599.3498   3.780673
#> 160 500.0633   2.872833
#> 161 501.5618   2.829713
#> 162 576.1436   2.750361
#> 163 500.0257   2.783783
#> 164 502.1582   3.130598
#> 165 500.5622   2.501297
#> 166 504.3729   2.834389
#> 167 929.1413   2.681232
#> 168 535.8057   2.806147
#> 169 683.7738   2.648884
#> 170 500.0573   2.636157
#> 171 515.0939   2.500115
#> 172 683.1283   2.541956
#> 173 500.0004   2.979766
#> 174 500.7114   2.536649
#> 175 500.4341   2.505155
#> 176 522.9313   2.821758
#> 177 505.0855   3.417552
#> 178 508.9379   2.818181
#> 179 506.1225   2.509372
#> 180 500.1779   2.705058
#> 181 500.0140   3.017568
#> 182 500.0362   4.005645
#> 183 535.1836   2.880573
#> 184 500.0974   2.513951
#> 185 637.1809   2.839422
#> 186 500.0531   2.798293
#> 187 606.5396   2.542121
#> 188 503.4056   3.238138
#> 189 500.0082   3.614848
#> 190 511.3465   2.694719
#> 191 500.1604   2.614368
#> 192 728.2695   2.586728
#> 193 504.0091   3.470846
#> 194 500.6054   2.678788
#> 195 521.0875   3.249768
#> 196 500.6848   2.533078
#> 197 534.0881   2.543155
#> 198 760.6100   2.702767
#> 199 500.2675   3.007154
#> 200 500.2995   2.610686
#> 201 500.3119   2.674176
#> 202 500.2476   4.877900
#> 203 501.5172   3.462480
#> 204 502.5095   2.830292
#> 205 500.3696   2.599860
#> 206 524.2344   2.995031
#> 207 628.7008   2.615900
#> 208 500.8540   2.672472
#> 209 500.0502   3.023209
#> 210 859.1321   2.576363
#> 211 503.4673   3.350228
#> 212 567.4796   2.859012
#> 213 551.0026   2.966446
#> 214 500.0541   2.626340
#> 215 537.4066   3.409433
#> 216 503.8579   2.631714
#> 217 500.5601   2.607276
#> 218 510.9839   2.556907
#> 219 520.8457   2.813692
#> 220 500.2276   2.988277
#> 221 508.2095   3.188027
#> 222 501.8491   2.722197
#> 223 500.0179   3.638727
#> 224 502.7992   2.805187
#> 225 500.6322   2.632905
#> 226 500.3523   2.897241
#> 227 500.5401   3.042039
#> 228 522.8062   2.652380
#> 229 504.3663   3.005811
#> 230 500.3170   2.663626
#> 231 507.2202   2.811335
#> 232 510.5892   3.203266
#> 233 508.2052   2.572272
#> 234 502.7420   2.765152
#> 235 500.6316   2.572873
#> 236 503.7914   3.139225
#> 237 500.2031   3.006675
#> 238 501.2723   3.011173
#> 239 500.6740   2.949809
#> 240 500.0042   3.106981
#> 241 500.6994   4.009536
#> 242 501.5148   2.553405
#> 243 546.6306   2.870317
#> 244 565.5850   3.085518
#> 245 500.3728   3.319800
#> 246 508.8181   2.653114
#> 247 552.9255   2.787977
#> 248 883.9792   2.564094
#> 249 500.3184   3.532684
#> 250 659.6166   2.509059
#> 251 510.1565   2.541902
#> 252 501.8656   2.751669
#> 253 535.4917   3.057435
#> 254 501.0872   2.868741
#> 255 503.0944   3.271503
#> 256 504.2131   2.881852
#> 257 516.4241   2.738627
#> 258 593.4448   2.848580
#> 259 500.1124   2.864834
#> 260 500.0124   3.821290
#> 261 561.3037   2.583923
#> 262 501.6310   2.537437
#> 263 507.6330   3.178737
#> 264 662.1568   2.677608
#> 265 500.6149   2.830032
#> 266 506.1020   2.506110
#> 267 649.4773   2.888427
#> 268 514.6103   2.687061
#> 269 511.5339   3.886377
#> 270 522.3908   2.937975
#> 271 501.3291   3.231059
#> 272 517.5347   2.505389
#> 273 508.6805   2.669510
#> 274 519.7577   3.385680
#> 275 500.8690   4.762729
#> 276 500.4917   3.140552
#> 277 520.3865   2.550570
#> 278 504.3546   2.883617
#> 279 809.4412   2.717606
#> 280 539.0671   2.538728
#> 281 505.8346   3.677235
#> 282 508.7431   3.145824
#> 283 500.0335   2.546722
#> 284 501.7706   2.535369
#> 285 581.6247   3.924067
#> 286 542.7514   3.602692
#> 287 500.0307   2.503350
#> 288 508.8601   3.165782
#> 289 501.1332   3.026149
#> 290 504.0833   2.682059
#> 291 500.0568   2.688975
#> 292 507.1884   4.387311
#> 293 514.0554   4.681187
#> 294 500.6770   2.710708
#> 295 502.8155   2.639475
#> 296 500.0111   3.028694
#> 297 502.1990   2.659404
#> 298 500.3599   2.654660
#> 299 500.1586   2.710665
#> 300 500.5954   3.016225
#> 301 500.0137   3.533726
#> 302 567.5884   3.525587
#> 303 501.1410   4.402730
#> 304 792.3457   2.911625
#> 305 732.8547   3.358561
#> 306 500.5025   2.548024
#> 307 512.7108   2.807875
#> 308 527.7517   3.198249
#> 309 650.7443   2.580237
#> 310 553.4499   2.551959
#> 311 569.3389   2.797425
#> 312 500.0996   2.919245
#> 313 504.1563   2.737357
#> 314 509.9774   2.964673
#> 315 501.7794   2.527365
#> 316 512.7926   2.679963
#> 317 500.1898   3.168092
#> 318 512.4383   2.833067
#> 319 537.8784   2.712220
#> 320 508.7070   5.830517
#> 321 505.6207   2.539017
#> 322 500.0920   2.985577
#> 323 527.6559   4.837483
#> 324 500.1855   2.719049
#> 325 500.7938   3.512047
#> 326 500.1732   2.541282
#> 327 507.5603   2.788110
#> 328 506.0257   2.615837
#> 329 500.4954   2.652614
#> 330 504.7275   2.674394
#> 331 500.0371   2.888699
#> 332 504.1508   2.611031
#> 333 500.3738   3.623742
#> 334 500.1041   2.855464
#> 335 500.0300   2.874031
#> 336 501.0930   2.542496
#> 337 500.5060   3.846567
#> 338 551.4222   2.952850
#> 339 500.3516   5.922959
#> 340 532.9818   2.512758
#> 341 500.1392   2.637518
#> 342 899.8025   2.826757
#> 343 509.5062   2.515249
#> 344 500.3081   2.560167
#> 345 500.1654   2.934012
#> 346 509.6734   4.395921
#> 347 500.3086   3.016173
#> 348 500.4965   2.506522
#> 349 500.8030   2.787367
#> 350 504.0455   3.197462
#> 351 606.2817   2.524064
#> 352 500.0113   2.820380
#> 353 500.0073   2.516008
#> 354 503.6141   2.760307
#> 355 502.8825   2.660383
#> 356 961.5931   2.916611
#> 357 789.6799   3.741298
#> 358 500.4049   2.942882
#> 359 738.4301   2.584279
#> 360 503.1309   2.512883
#> 361 500.4573   2.754888
#> 362 911.6110   3.346610
#> 363 558.4109   2.914538
#> 364 500.9332   3.018431
#> 365 643.4900   2.533964
#> 366 500.0157   3.243307
#> 367 501.3895   3.840714
#> 368 502.3306   2.716511
#> 369 505.0313   3.059635
#> 370 585.5840   2.726179
#> 371 500.3866   2.913202
#> 372 500.6818   2.593094
#> 373 500.0671   3.183717
#> 374 860.1455   2.929666
#> 375 502.3912   2.991974
#> 376 500.7187   2.758302
#> 377 502.7854   2.708332
#> 378 708.4659   2.941250
#> 379 639.2904   3.058592
#> 380 512.7772   2.568695
#> 381 500.0299   2.500687
#> 382 526.0853   3.942024
#> 383 517.9959   2.689280
#> 384 502.8153   2.733638
#> 385 654.0207   2.855325
#> 386 607.7807   2.574914
#> 387 516.2556   3.284178
#> 388 602.9148   3.386925
#> 389 509.8975   3.203514
#> 390 500.8095   3.196901
#> 391 633.5791   4.389359
#> 392 527.7515   2.532546
#> 393 500.0774   2.668089
#> 394 506.1268   2.529682
#> 395 500.0440   3.794836
#> 396 500.0210   2.833494
#> 397 500.0182   2.557173
#> 398 673.5297   2.916919
#> 399 500.1463   2.656859
#> 400 515.2171   3.315275
#> 401 500.0923   2.970949
#> 402 500.2408   2.576479
#> 403 503.9189   2.553942
#> 404 529.2892   4.297139
#> 405 500.0675   4.343283
#> 406 500.2511   3.091448
#> 407 784.4208   2.509078
#> 408 500.3080   3.110693
#> 409 771.3214   2.795084
#> 410 500.0019   2.815335
#> 411 531.8889   3.118666
#> 412 501.8668   2.810325
#> 413 575.6517   2.741435
#> 414 500.2056   3.188481
#> 415 501.7318   5.776258
#> 416 500.1010   2.538792
#> 417 500.8841   2.729570
#> 418 500.7379   2.558667
#> 419 500.2033   2.663990
#> 420 734.9626   2.677683
#> 421 626.0118   3.118419
#> 422 500.0522   2.551726
#> 423 517.2485   2.638151
#> 424 500.1102   2.782849
#> 425 522.2788   2.641984
#> 426 502.4235   2.772003
#> 427 562.7076   3.060893
#> 428 605.2626   3.133235
#> 429 531.3782   2.742509
#> 430 500.0915   2.598055
#> 431 515.7056   2.518868
#> 432 500.2305   2.616840
#> 433 500.9238   3.231372
#> 434 502.8078   2.628124
#> 435 500.6099   2.562862
#> 436 508.2552   3.203338
#> 437 578.0419   2.815590
#> 438 500.1732   2.700849
#> 439 501.0039   2.931642
#> 440 500.1037   3.849790
#> 441 501.3188   3.198373
#> 442 502.8434   3.611972
#> 443 518.4101   2.500010
#> 444 500.0059   2.548480
#> 445 502.0914   3.110716
#> 446 516.4260   3.120762
#> 447 503.0315   2.616300
#> 448 500.0052   2.682342
#> 449 524.6292   2.537052
#> 450 501.0961   3.696110
#> 451 500.3774   3.134856
#> 452 644.3458   2.611465
#> 453 500.0845   2.509388
#> 454 500.2635   2.646300
#> 455 512.6105   2.860104
#> 456 932.9100   3.204277
#> 457 500.0536   2.794928
#> 458 502.6610   2.505496
#> 459 519.9589   2.606467
#> 460 500.0103   3.431496
#> 461 587.1300   2.629639
#> 462 511.6851   3.105561
#> 463 500.0591   2.768192
#> 464 500.2569   3.325170
#> 465 500.4673   2.754426
#> 466 760.9527   2.628595
#> 467 500.2322   2.530467
#> 468 508.0240   3.578232
#> 469 500.4960   2.768659
#> 470 507.9848   2.507634
#> 471 579.3672   2.532926
#> 472 501.7318   3.522816
#> 473 500.3596   3.819355
#> 474 500.2145   2.833078
#> 475 500.0573   5.047700
#> 476 500.0228   3.539782
#> 477 558.5637   2.553391
#> 478 503.7576   2.631697
#> 479 881.1213   3.051909
#> 480 500.2418   2.548743
#> 481 535.9217   2.846111
#> 482 602.9585   3.527091
#> 483 504.7251   4.436306
#> 484 500.0516   2.534387
#> 485 501.0508   2.605889
#> 486 530.8171   2.886471
#> 487 500.1680   2.872570
#> 488 500.4294   2.602304
#> 489 500.1189   2.897941
#> 490 502.7155   2.786690
#> 491 501.5831   3.270856
#> 492 828.4759   3.338717
#> 493 578.7737   2.709680
#> 494 505.3344   3.041407
#> 495 500.0059   2.514244
#> 496 500.6671   3.195132
#> 497 555.8646   2.579374
#> 498 500.4799   2.723047
#> 499 500.2019   3.136584
#> 500 513.2912   2.661112
#> 501 520.8825   2.515643
#> 502 500.0016   3.637093
#> 503 525.4818   2.657686
#> 504 500.0591   2.526271
#> 505 571.5674   2.744245
#> 506 510.9024   3.001539
#> 507 500.0808   2.654907
#> 508 506.1287   3.820895
#> 509 501.0249   3.406560
#> 510 500.1286   3.289045
#> 511 513.0950   2.540394
#> 512 501.4298   2.586536
#> 513 506.2631   3.059737
#> 514 507.2650   2.571750
#> 515 509.1198   2.993106
#> 516 504.1687   3.326285
#> 517 770.4724   3.516674
#> 518 501.1768   2.829476
#> 519 838.8241   2.635478
#> 520 510.0432   3.021418
#> 521 500.0453   2.776523
#> 522 529.1873   2.536633
#> 523 667.2048   2.739643
#> 524 511.6156   3.079253
#> 525 501.5521   2.859634
#> 526 517.9070   2.726469
#> 527 715.7449   2.601613
#> 528 615.4636   3.890943
#> 529 502.3541   3.274105
#> 530 532.5903   2.510419
#> 531 500.5725   2.681325
#> 532 500.1245   2.798030
#> 533 694.7832   4.817312
#> 534 500.3517   3.310028
#> 535 500.8765   2.528057
#> 536 527.2673   2.698773
#> 537 976.8338   2.718777
#> 538 570.2520   2.852184
#> 539 500.9939   2.853491
#> 540 530.7230   3.740838
#> 541 500.0243   2.971174
#> 542 974.5708   2.735180
#> 543 514.9727   2.501067
#> 544 500.7306   2.938706
#> 545 500.1293   3.232796
#> 546 502.5305   3.111296
#> 547 533.6238   2.653016
#> 548 500.3378   4.767001
#> 549 515.9992   3.707616
#> 550 500.7761   2.928075
#> 551 670.4553   4.010465
#> 552 502.6458   2.644433
#> 553 506.1245   2.791465
#> 554 503.1276   2.575964
#> 555 504.1369   2.544555
#> 556 507.6344   4.293476
#> 557 500.6258   2.502719
#> 558 500.0684   2.932384
#> 559 500.5338   2.617241
#> 560 521.4641   3.182215
#> 561 500.0102   2.659358
#> 562 514.1095   2.928055
#> 563 500.1198   3.641871
#> 564 500.4189   3.190612
#> 565 500.1138   3.257097
#> 566 500.5400   4.244967
#> 567 502.4553   2.562561
#> 568 500.3311   2.710271
#> 569 500.7832   2.695527
#> 570 631.9195   2.769654
#> 571 500.6421   3.321978
#> 572 630.2285   3.318603
#> 573 551.0489   2.770317
#> 574 500.0269   3.063435
#> 575 552.5371   3.320521
#> 576 500.2233   2.797502
#> 577 830.9193   2.967765
#> 578 664.0956   2.535991
#> 579 502.2497   3.190168
#> 580 500.7262   2.917735
#> 581 502.6775   2.638262
#> 582 636.6413   2.863458
#> 583 501.4158   2.611861
#> 584 501.6369   2.555917
#> 585 617.5605   2.511149
#> 586 504.2532   3.125495
#> 587 613.9681   2.697195
#> 588 500.3217   2.502425
#> 589 510.5215   3.043619
#> 590 509.4795   3.198091
#> 591 509.7821   2.935365
#> 592 500.0633   3.217259
#> 593 501.0544   2.861106
#> 594 500.8337   2.625435
#> 595 500.2985   3.467155
#> 596 502.3307   2.776084
#> 597 502.5417   2.921910
#> 598 500.9942   2.844430
#> 599 534.3663   2.753811
#> 600 501.0040   2.957435
#> 601 507.9828   2.584782
#> 602 565.8022   2.916181
#> 603 500.4325   2.960907
#> 604 557.7970   3.263979
#> 605 500.0190   3.383293
#> 606 533.6655   2.820579
#> 607 522.9721   3.311581
#> 608 500.2504   3.084033
#> 609 515.7815   2.514176
#> 610 501.9914   3.429161
#> 611 500.4900   3.973570
#> 612 510.4830   3.353126
#> 613 509.0743   2.813526
#> 614 500.8143   2.942492
#> 615 615.4357   2.871420
#> 616 504.8343   2.540083
#> 617 544.2885   3.090780
#> 618 501.1877   2.540201
#> 619 502.2018   2.847449
#> 620 500.2284   3.041088
#> 621 500.0841   3.967963
#> 622 649.0694   3.030212
#> 623 614.8370   2.539452
#> 624 513.4357   2.628697
#> 625 617.3708   2.560646
#> 626 520.4768   2.567935
#> 627 509.0102   2.582222
#> 628 501.5192   2.751258
#> 629 502.7521   2.887106
#> 630 510.8465   2.582685
#> 631 500.0776   2.669850
#> 632 501.8395   2.885716
#> 633 502.4286   3.722068
#> 634 501.6413   3.465547
#> 635 500.4330   3.378347
#> 636 500.1172   2.543740
#> 637 630.4929   2.754545
#> 638 500.1535   2.594612
#> 639 791.0533   2.533803
#> 640 500.6461   2.918403
#> 641 501.1248   2.652497
#> 642 506.3067   2.515583
#> 643 500.3768   3.098579
#> 644 500.1916   2.577069
#> 645 683.8597   2.523856
#> 646 731.2894   2.771402
#> 647 500.5375   3.596720
#> 648 500.1832   3.053022
#> 649 500.9927   3.543836
#> 650 502.3104   3.077144
#> 651 500.1129   3.088158
#> 652 501.9823   2.889166
#> 653 644.1677   2.520573
#> 654 500.4388   2.503744
#> 655 501.1412   2.761624
#> 656 504.7320   3.502533
#> 657 503.1560   2.532823
#> 658 550.3806   3.714321
#> 659 619.6290   3.294011
#> 660 512.2410   2.787029
#> 661 978.4876   3.338285
#> 662 501.8708   2.645235
#> 663 500.1335   3.279722
#> 664 898.5574   2.680558
#> 665 535.0063   3.287500
#> 666 512.4161   2.679167
#> 667 500.0757   2.761604
#> 668 500.0877   2.805224
#> 669 500.4998   3.429199
#> 670 500.4904   2.708757
#> 671 502.1404   2.646360
#> 672 555.2137   2.806859
#> 673 599.3759   3.280390
#> 674 501.8258   2.691390
#> 675 564.6812   4.526323
#> 676 500.1390   2.792028
#> 677 501.9005   3.024365
#> 678 500.2497   2.799027
#> 679 505.9186   2.557851
#> 680 586.8056   3.627058
#> 681 512.9532   2.505467
#> 682 501.5091   2.520204
#> 683 545.5307   2.523820
#> 684 744.0951   3.220738
#> 685 500.0306   2.578023
#> 686 570.9854   2.729515
#> 687 500.0035   2.522780
#> 688 639.0899   2.702654
#> 689 500.1277   3.007357
#> 690 982.8887   2.632317
#> 691 548.3919   2.977981
#> 692 503.4332   2.668915
#> 693 500.0969   2.656745
#> 694 502.7470   2.514451
#> 695 501.2240   3.021017
#> 696 511.8226   3.592003
#> 697 500.6152   3.322880
#> 698 512.1199   3.188535
#> 699 500.1322   2.810215
#> 700 500.1690   2.702890
#> 701 500.0774   2.691416
#> 702 500.1616   3.426928
#> 703 503.4128   2.987831
#> 704 502.1243   2.670067
#> 705 500.6522   3.428431
#> 706 500.4986   3.149107
#> 707 500.9003   2.519103
#> 708 605.0331   3.262497
#> 709 554.4486   2.751862
#> 710 500.0809   2.948650
#> 711 500.3807   3.037382
#> 712 514.4021   2.678602
#> 713 783.1371   3.317485
#> 714 509.7592   2.969268
#> 715 513.4967   2.691907
#> 716 612.7240   2.837546
#> 717 505.9952   3.525611
#> 718 500.7168   3.898305
#> 719 536.9256   3.195746
#> 720 696.2949   4.846859
#> 721 500.7294   2.608861
#> 722 966.0497   2.738686
#> 723 535.7167   4.664270
#> 724 500.0577   2.599057
#> 725 500.4244   2.712691
#> 726 994.7934   2.639407
#> 727 500.4909   2.622463
#> 728 500.4705   2.644691
#> 729 500.0414   3.351097
#> 730 501.2744   2.961278
#> 731 500.0940   4.048456
#> 732 501.9171   3.144457
#> 733 864.3429   4.202029
#> 734 515.9766   2.527695
#> 735 502.3852   2.752822
#> 736 500.0889   2.547186
#> 737 509.5506   2.734344
#> 738 500.0356   2.866004
#> 739 500.1896   4.329832
#> 740 501.2962   3.274779
#> 741 518.6412   2.632582
#> 742 501.4530   3.394124
#> 743 656.8462   3.425065
#> 744 503.1297   2.612688
#> 745 500.1802   2.814271
#> 746 568.6914   2.574217
#> 747 636.9030   2.877470
#> 748 500.0750   7.303676
#> 749 567.9078   2.537084
#> 750 521.3200   2.897417
#> 751 515.8150   2.538105
#> 752 501.3649   3.417774
#> 753 505.7063   2.766221
#> 754 951.2941   3.459740
#> 755 501.3930   2.522571
#> 756 500.0284   3.056931
#> 757 825.0726   2.703473
#> 758 501.9608   2.601321
#> 759 500.5786   2.734599
#> 760 505.6980   3.473951
#> 761 500.2189   3.399538
#> 762 500.1779   2.705062
#> 763 553.8888   2.562533
#> 764 719.7568   2.696220
#> 765 547.1262   2.582691
#> 766 503.6655   2.604932
#> 767 500.0644   2.963582
#> 768 503.2792   3.018675
#> 769 500.1097   2.672748
#> 770 513.9552   2.774125
#> 771 554.6024   2.680574
#> 772 500.0493   2.685316
#> 773 500.0360   2.558577
#> 774 502.3868   2.543355
#> 775 537.5720   3.222210
#> 776 504.2488   2.546456
#> 777 568.8447   2.617952
#> 778 500.4107   2.549322
#> 779 508.8650   3.062081
#> 780 500.1135   2.614423
#> 781 539.4366   2.681427
#> 782 500.7717   2.947421
#> 783 501.5501   2.893711
#> 784 502.2172   2.645345
#> 785 500.1397   2.875598
#> 786 500.2188   2.632725
#> 787 501.1726   2.789712
#> 788 500.2999   2.656073
#> 789 501.3044   4.044475
#> 790 503.3035   2.665920
#> 791 581.6385   3.248181
#> 792 500.6364   2.940640
#> 793 516.0041   2.552061
#> 794 500.0729   2.800954
#> 795 515.9709   2.923041
#> 796 500.1345   3.297288
#> 797 511.4513   4.299590
#> 798 501.0406   2.668817
#> 799 501.3311   3.234275
#> 800 500.0036   2.623749
#> 801 500.1232   2.945233
#> 802 502.4028   2.555866
#> 803 500.4458   6.064520
#> 804 904.1812   2.509971
#> 805 518.4478   3.048316
#> 806 500.0816   2.651422
#> 807 501.2565   3.316551
#> 808 501.0000   2.512325
#> 809 500.2331   2.736546
#> 810 676.1381   2.543860
#> 811 500.1133   2.980910
#> 812 500.1206   3.089255
#> 813 500.7559   2.564449
#> 814 523.9033   4.068479
#> 815 503.2567   2.859302
#> 816 584.2592   2.732486
#> 817 506.1024   2.620162
#> 818 502.1027   2.831157
#> 819 500.1738   3.968459
#> 820 516.9665   2.586451
#> 821 500.1497   3.007396
#> 822 519.3086   2.731503
#> 823 500.8708   3.137317
#> 824 503.8650   2.975418
#> 825 500.3980   3.278996
#> 826 506.3496   2.990843
#> 827 502.4957   2.532326
#> 828 505.7698   3.391283
#> 829 509.9570   3.080809
#> 830 502.9910   2.557345
#> 831 566.1910   2.735711
#> 832 500.0983   2.689078
#> 833 515.5715   3.399036
#> 834 500.1324   2.616384
#> 835 500.0395   2.631710
#> 836 508.4642   3.787233
#> 837 504.2127   3.484634
#> 838 548.2947   3.065650
#> 839 772.1021   2.724008
#> 840 506.4425   3.730329
#> 841 500.0239   3.258438
#> 842 500.0725   2.675840
#> 843 500.1955   2.662790
#> 844 500.1629   3.076240
#> 845 505.3446   2.662401
#> 846 611.9442   2.953854
#> 847 509.3309   3.153109
#> 848 501.3468   2.517844
#> 849 500.2354   3.346738
#> 850 500.9605   2.718235
#> 851 795.2816   2.570986
#> 852 513.0326   2.613767
#> 853 502.8911   2.591293
#> 854 689.6340   3.099097
#> 855 515.2265   2.823932
#> 856 946.6255   2.989847
#> 857 515.0260   2.884799
#> 858 504.0355   2.609887
#> 859 534.2867   2.628933
#> 860 552.7416   2.850506
#> 861 501.1041   3.109049
#> 862 500.8668   3.742780
#> 863 500.2552   3.082061
#> 864 515.6474   2.655458
#> 865 501.3679   2.636109
#> 866 500.7574   3.883119
#> 867 500.3767   3.573612
#> 868 788.3574   2.616093
#> 869 500.4164   2.572935
#> 870 500.3406   3.152136
#> 871 767.2552   2.853656
#> 872 527.6254   2.916932
#> 873 500.5618   2.500982
#> 874 500.0665   3.676771
#> 875 502.5042   3.006847
#> 876 501.8404   2.506595
#> 877 501.4367   3.672323
#> 878 504.2894   3.664110
#> 879 525.1515   2.620757
#> 880 539.5189   3.193282
#> 881 506.5757   3.299420
#> 882 500.0087   2.605517
#> 883 500.0711   2.541206
#> 884 513.4945   2.790511
#> 885 500.0278   3.662658
#> 886 924.2322   2.627402
#> 887 514.8914   2.989622
#> 888 521.4782   2.745588
#> 889 500.7044   3.395031
#> 890 500.0819   2.720395
#> 891 510.3523   2.556990
#> 892 501.9581   2.792733
#> 893 564.9995   2.801555
#> 894 502.1846   3.184522
#> 895 507.4646   2.938674
#> 896 538.2238   3.848116
#> 897 667.0302   2.512653
#> 898 506.0897   3.119622
#> 899 571.6589   2.917231
#> 900 502.5521   2.935777
#> 901 646.5121   2.829118
#> 902 504.4703   2.656461
#> 903 500.1566   2.613397
#> 904 503.5753   2.593389
#> 905 508.5662   2.512947
#> 906 511.1414   4.149287
#> 907 502.4793   2.756173
#> 908 500.0966   2.570835
#> 909 500.6747   2.893708
#> 910 543.1450   3.353649
#> 911 510.2496   2.527386
#> 912 545.1873   4.115866
#> 913 500.6090   2.555501
#> 914 502.0203   3.292146
#> 915 500.3700   2.785630
#> 916 501.2118   2.744969
#> 917 808.5327   2.882524
#> 918 501.0489   2.529826
#> 919 500.0055   2.561464
#> 920 651.2835   4.094451
#> 921 530.2085   2.537442
#> 922 500.0937   2.832980
#> 923 804.6102   2.888068
#> 924 550.6928   2.745687
#> 925 504.5256   2.708267
#> 926 504.2269   2.755173
#> 927 521.3106   4.356965