Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   505.9831   2.597793
#> 2   500.2759   2.771139
#> 3   501.2309   2.500004
#> 4   502.8734   3.103010
#> 5   505.0958   2.634380
#> 6   501.5163   3.748403
#> 7   551.1098   2.587641
#> 8   500.4508   2.818616
#> 9   500.6347   2.543907
#> 10  513.5813   2.584317
#> 11  502.5164   2.569164
#> 12  507.0623   3.999429
#> 13  503.1599   2.844645
#> 14  503.8061   3.736510
#> 15  500.3419   2.631266
#> 16  500.0222   3.882347
#> 17  502.9289   2.806443
#> 18  723.3083   2.750725
#> 19  865.1887   2.579713
#> 20  500.2384   2.861404
#> 21  550.0814   2.820766
#> 22  738.9106   2.975585
#> 23  509.1960   2.604315
#> 24  500.1610   3.058448
#> 25  645.6348   2.856301
#> 26  511.4021   2.586994
#> 27  502.1812   2.896472
#> 28  635.4090   3.012102
#> 29  539.0790   2.515192
#> 30  500.8169   4.355316
#> 31  501.1206   2.567689
#> 32  507.3580   3.196012
#> 33  500.2885   2.575963
#> 34  500.7267   2.941496
#> 35  503.0184   2.984853
#> 36  501.9542   3.266751
#> 37  504.3256   2.537315
#> 38  501.1027   3.114832
#> 39  506.5956   2.525855
#> 40  507.9215   2.687100
#> 41  500.1034   2.869512
#> 42  501.3342   2.661846
#> 43  516.5088   2.550596
#> 44  500.3469   2.506832
#> 45  565.6428   3.882166
#> 46  501.8485   2.655618
#> 47  500.8048   2.505275
#> 48  500.4627   2.580200
#> 49  500.7774   2.572003
#> 50  500.2958   3.584078
#> 51  758.2126   2.523259
#> 52  500.1464   2.686901
#> 53  574.2666   2.742915
#> 54  500.0791   2.539687
#> 55  501.6328   2.681305
#> 56  542.5251   2.629707
#> 57  500.4498   2.770632
#> 58  566.9060   3.253479
#> 59  827.4942   2.769681
#> 60  506.6671   2.596736
#> 61  511.0200   3.691073
#> 62  906.5534   2.585120
#> 63  500.2381   2.650926
#> 64  501.7791   2.965785
#> 65  500.4057   2.509039
#> 66  500.1461   2.571228
#> 67  500.7004   2.511055
#> 68  500.0207   2.665116
#> 69  917.6747   2.798944
#> 70  506.6109   2.762706
#> 71  500.2178   3.071402
#> 72  500.6570   2.650475
#> 73  501.2390   3.123528
#> 74  586.9439   2.622751
#> 75  778.3813   3.110469
#> 76  739.7643   2.834781
#> 77  500.1087   2.593560
#> 78  500.0669   2.968808
#> 79  577.0344   3.362921
#> 80  500.4717   2.942065
#> 81  500.1101   2.601411
#> 82  696.4792   2.883171
#> 83  510.0355   2.512222
#> 84  532.1098   2.951546
#> 85  500.0497   2.584556
#> 86  500.1762   2.593201
#> 87  500.0113   2.972131
#> 88  500.3541   2.980381
#> 89  501.0164   2.722165
#> 90  509.1649   2.970679
#> 91  623.4072   3.324804
#> 92  512.8444   3.135900
#> 93  766.5114   2.757477
#> 94  502.5761   3.423151
#> 95  681.8945   2.765524
#> 96  500.6181   2.966387
#> 97  684.4491   2.794452
#> 98  500.6614   2.872448
#> 99  500.7785   2.914475
#> 100 585.8976   3.648165
#> 101 570.9206   3.138428
#> 102 500.0182   2.780024
#> 103 514.6981   2.792388
#> 104 500.4161   2.977113
#> 105 502.5859   3.147950
#> 106 505.0686   2.963135
#> 107 777.6690   2.772087
#> 108 711.7203   2.517488
#> 109 620.8954   2.608571
#> 110 500.1787   2.505770
#> 111 500.1774   2.740481
#> 112 500.1051   2.860352
#> 113 506.4223   2.544158
#> 114 525.2872   2.533882
#> 115 500.0108   2.623030
#> 116 567.1164   2.524958
#> 117 510.1469   2.918046
#> 118 516.6046   3.134045
#> 119 505.8238   2.838535
#> 120 524.5089   2.664375
#> 121 500.2556   3.609971
#> 122 654.1419   2.528914
#> 123 500.0839   3.075623
#> 124 502.5293   2.646233
#> 125 500.1166   2.613777
#> 126 503.8576   2.565706
#> 127 674.3512   2.854206
#> 128 631.3903   2.736651
#> 129 500.6866   2.774672
#> 130 635.6054   3.344619
#> 131 500.2395   2.864089
#> 132 502.1172   2.760415
#> 133 584.4593   2.672938
#> 134 501.3144   3.313130
#> 135 501.4453   3.242731
#> 136 582.9782   3.308445
#> 137 500.0916   3.127945
#> 138 500.4694   2.533659
#> 139 526.0074   2.799967
#> 140 500.5468   2.907969
#> 141 552.1517   3.001782
#> 142 501.4231   2.698132
#> 143 500.9732   3.232245
#> 144 507.8758   2.581023
#> 145 588.1066   2.589428
#> 146 500.7034   2.654057
#> 147 500.1626   2.790611
#> 148 500.1845   3.905482
#> 149 703.1416   2.784169
#> 150 500.4155   2.785362
#> 151 501.0364   2.962210
#> 152 568.1256   2.864694
#> 153 504.8272   2.709403
#> 154 624.6253   2.906691
#> 155 500.0637   2.537339
#> 156 502.9122   4.703297
#> 157 500.2672   2.503306
#> 158 505.6803   3.090572
#> 159 500.0567   2.565113
#> 160 570.1068   3.164836
#> 161 500.0773   2.699042
#> 162 504.1975   2.740837
#> 163 500.0104   2.915987
#> 164 586.3373   3.444200
#> 165 517.0107   2.705964
#> 166 503.8161   4.128046
#> 167 500.4571   2.945264
#> 168 811.0498   2.820183
#> 169 500.1041   2.653106
#> 170 500.1377   3.556547
#> 171 588.7361   3.012350
#> 172 520.7934   3.105404
#> 173 500.0726   2.653171
#> 174 661.7090   3.056234
#> 175 500.2075   2.796204
#> 176 500.4203   5.205144
#> 177 500.0947   2.606104
#> 178 501.7742   2.797329
#> 179 500.2144   2.890852
#> 180 500.4210   2.719527
#> 181 551.8241   2.965970
#> 182 510.3709   2.706990
#> 183 500.9188   3.174645
#> 184 503.8988   2.545646
#> 185 500.0734   3.272275
#> 186 551.1330   2.863250
#> 187 500.0894   3.303618
#> 188 504.7766   2.529109
#> 189 500.5288   3.105414
#> 190 543.4817   3.470056
#> 191 625.4487   2.858268
#> 192 519.6497   4.089129
#> 193 500.9791   3.136111
#> 194 502.2721   2.940567
#> 195 500.3831   2.664595
#> 196 500.1464   2.815099
#> 197 500.0594   2.985965
#> 198 504.4803   2.761953
#> 199 500.1450   2.517348
#> 200 500.3277   2.676511
#> 201 678.1650   2.649126
#> 202 517.6217   3.007152
#> 203 557.5718   3.578196
#> 204 560.4416   2.931823
#> 205 501.3786   2.509996
#> 206 596.3044   3.032449
#> 207 500.1444   2.680434
#> 208 575.5553   2.582370
#> 209 500.1131   3.146310
#> 210 503.0401   2.934277
#> 211 500.0146   2.767079
#> 212 636.6269   2.738272
#> 213 536.8067   2.800011
#> 214 507.5846   2.631722
#> 215 880.6213   2.900674
#> 216 507.2053   2.537240
#> 217 500.0941   2.582579
#> 218 559.6386   2.889138
#> 219 901.2263   2.803711
#> 220 503.3213   2.525968
#> 221 528.1101   2.918613
#> 222 503.5993   2.757891
#> 223 501.5188   2.581989
#> 224 501.0014   3.208553
#> 225 865.9985   2.658529
#> 226 631.7392   2.605411
#> 227 539.0245   2.668435
#> 228 500.1263   3.453728
#> 229 762.3252   3.283540
#> 230 500.8599   2.989519
#> 231 500.1918   2.866440
#> 232 500.6890   2.786224
#> 233 501.4068   2.533941
#> 234 508.5706   2.618162
#> 235 726.6008   3.594813
#> 236 501.5372   2.533579
#> 237 500.0608   3.438453
#> 238 505.1211   2.748848
#> 239 500.2736   2.556338
#> 240 501.1665   4.019284
#> 241 501.0361   2.707239
#> 242 531.3719   4.635441
#> 243 500.8008   2.575992
#> 244 519.7353   2.522602
#> 245 500.0691   2.559485
#> 246 741.1032   2.544708
#> 247 522.2255   2.819162
#> 248 504.0084   2.785938
#> 249 500.1365   3.475632
#> 250 502.6352   2.734381
#> 251 500.9659   2.801386
#> 252 501.4869   2.552018
#> 253 500.0907   3.487057
#> 254 500.1948   2.587211
#> 255 510.5620   3.566273
#> 256 507.9249   2.603008
#> 257 522.8506   2.779349
#> 258 500.1114   2.685263
#> 259 501.8011   3.449584
#> 260 500.5782   3.069618
#> 261 504.8103   3.510302
#> 262 500.6133   3.532156
#> 263 531.2930   2.560558
#> 264 505.1516   2.843390
#> 265 501.5068   2.501796
#> 266 510.3729   2.635307
#> 267 589.4136   2.728205
#> 268 500.2755   4.374234
#> 269 505.9585   2.813604
#> 270 538.8387   3.040127
#> 271 595.9750   2.636774
#> 272 515.6406   3.802408
#> 273 533.4291   2.575002
#> 274 538.7371   2.635216
#> 275 501.6926   2.903625
#> 276 508.4187   3.020034
#> 277 504.5468   3.294632
#> 278 500.0502   2.912421
#> 279 500.0119   2.955246
#> 280 503.8285   2.502905
#> 281 608.8661   2.576704
#> 282 500.0901   2.660216
#> 283 503.6321   2.526620
#> 284 503.8518   2.981388
#> 285 500.3109   3.460371
#> 286 504.2667   2.857521
#> 287 503.2050   3.520531
#> 288 500.3172   2.680557
#> 289 502.0665   2.969316
#> 290 586.0640   2.935122
#> 291 813.2624   2.690237
#> 292 500.0500   2.774098
#> 293 500.8426   2.662073
#> 294 501.1534   2.545411
#> 295 500.1600   2.509626
#> 296 515.4068   2.844468
#> 297 507.7707   4.515823
#> 298 915.1849   3.044982
#> 299 829.2597   2.738446
#> 300 500.0535   2.724332
#> 301 640.9862   2.992722
#> 302 534.5851   2.536117
#> 303 500.3488   2.719163
#> 304 500.8761   2.535757
#> 305 500.3406   4.141689
#> 306 500.3298   3.045933
#> 307 503.8527   3.087365
#> 308 653.9830   3.197723
#> 309 500.6075   2.599712
#> 310 539.0269   2.513206
#> 311 500.3461   2.727027
#> 312 611.0060   2.667805
#> 313 500.0067   2.510546
#> 314 501.4581   2.569369
#> 315 503.5372   3.370504
#> 316 500.0053   2.518165
#> 317 503.8092   3.110551
#> 318 504.8306   2.618972
#> 319 541.1274   2.728723
#> 320 500.4982   3.203657
#> 321 500.7469   2.520632
#> 322 520.0138   3.901411
#> 323 501.1903   2.654716
#> 324 500.1689   3.303190
#> 325 531.6756   2.680160
#> 326 582.5325   3.473160
#> 327 510.4311   2.609660
#> 328 500.3104   2.954948
#> 329 509.4828   2.918048
#> 330 502.2231   2.568772
#> 331 500.4745   2.561468
#> 332 552.7619   3.089660
#> 333 737.1721   2.910928
#> 334 562.7311   2.501123
#> 335 500.1713   3.330229
#> 336 502.0402   2.789257
#> 337 586.7201   4.372288
#> 338 502.5117   3.148608
#> 339 500.4138   3.593308
#> 340 506.5027   2.921312
#> 341 500.6459   2.829815
#> 342 500.0443   2.631742
#> 343 510.8626   3.394759
#> 344 569.4537   2.568759
#> 345 501.1742   3.162203
#> 346 510.3804   2.747983
#> 347 539.0318   2.562787
#> 348 515.0482   2.676226
#> 349 500.7288   2.636612
#> 350 638.9551   3.063600
#> 351 500.4058   2.926984
#> 352 500.0353   2.833455
#> 353 695.3147   2.654500
#> 354 665.1609   3.302354
#> 355 509.7679   2.535157
#> 356 500.3683   2.879356
#> 357 500.5865   2.617203
#> 358 500.0702   2.511253
#> 359 500.3630   2.866765
#> 360 629.5741   2.739002
#> 361 500.4366   3.033881
#> 362 507.2327   2.630124
#> 363 500.4256   2.594057
#> 364 528.9736   2.869654
#> 365 580.9082   3.582959
#> 366 504.6574   3.197099
#> 367 504.8640   2.587059
#> 368 500.2351   2.812672
#> 369 609.8192   2.775635
#> 370 500.0595   2.997891
#> 371 507.1287   3.240873
#> 372 508.7065   2.951899
#> 373 577.8347   3.080123
#> 374 845.8221   2.656902
#> 375 500.0804   4.167615
#> 376 500.0084   2.924226
#> 377 501.4736   2.537489
#> 378 500.3423   2.963364
#> 379 525.8047   2.816110
#> 380 751.2293   2.672823
#> 381 500.8203   3.966900
#> 382 500.2334   2.573667
#> 383 501.8088   2.614667
#> 384 500.0061   2.804280
#> 385 505.4753   2.530260
#> 386 500.0365   3.127156
#> 387 500.0024   2.799640
#> 388 552.2674   3.037781
#> 389 503.9933   2.567428
#> 390 500.0365   4.100653
#> 391 500.4465   2.542337
#> 392 500.1442   2.759570
#> 393 500.0018   2.805254
#> 394 673.4035   4.153657
#> 395 501.9770   2.527327
#> 396 500.0003   3.423324
#> 397 780.2928   3.613968
#> 398 607.2778   2.803313
#> 399 500.6239   2.692898
#> 400 505.3967   2.733164
#> 401 501.8903   2.657031
#> 402 525.9843   4.044949
#> 403 579.1096   2.616176
#> 404 500.0673   3.000227
#> 405 502.1886   2.970243
#> 406 500.8380   2.531282
#> 407 836.4723   2.725880
#> 408 527.6568   3.053929
#> 409 630.7107   2.582001
#> 410 501.5764   2.774889
#> 411 532.7732   3.158900
#> 412 504.0073   2.874975
#> 413 503.1871   2.671365
#> 414 500.0151   2.663807
#> 415 502.6519   2.745998
#> 416 500.1257   2.761171
#> 417 575.7395   2.530006
#> 418 508.5590   3.571948
#> 419 501.8589   2.531610
#> 420 500.2095   3.078563
#> 421 500.0296   2.975522
#> 422 507.0543   2.519210
#> 423 500.1556   2.851155
#> 424 500.0102   4.156526
#> 425 500.1997   2.838224
#> 426 503.9255   2.668676
#> 427 500.0693   2.654158
#> 428 500.9167   2.979039
#> 429 527.4068   3.147054
#> 430 500.0935   2.928522
#> 431 500.6416   3.786034
#> 432 500.2180   2.883232
#> 433 500.4444   3.567167
#> 434 500.0044   3.358944
#> 435 503.0232   2.581434
#> 436 501.5238   2.680471
#> 437 621.6025   2.751190
#> 438 500.0156   3.715760
#> 439 502.3103   2.504899
#> 440 511.4321   2.717378
#> 441 501.4257   2.779906
#> 442 588.3365   2.895707
#> 443 508.1167   2.520790
#> 444 500.7500   2.713607
#> 445 507.0062   2.890077
#> 446 500.2301   2.648966
#> 447 518.2663   2.863913
#> 448 667.9240   2.645039
#> 449 501.0725   3.448257
#> 450 525.6893   2.759112
#> 451 502.9778   2.743645
#> 452 500.6871   3.331696
#> 453 665.7900   2.745273
#> 454 923.3961   2.561276
#> 455 883.3591   2.581119
#> 456 500.9261   2.601704
#> 457 501.8230   4.688797
#> 458 500.2974   2.548620
#> 459 500.2737   3.883841
#> 460 500.3332   2.913800
#> 461 528.6915   2.980141
#> 462 500.0897   3.588024
#> 463 568.3647   2.800272
#> 464 500.0211   2.948989
#> 465 537.7370   3.121746
#> 466 501.0551   3.164248
#> 467 501.0033   2.809094
#> 468 511.1312   2.602940
#> 469 500.0481   2.839961
#> 470 506.1643   2.636507
#> 471 528.8645   2.918955
#> 472 528.9848   2.587997
#> 473 543.6381   2.681265
#> 474 501.7700   2.586496
#> 475 500.0622   2.824020
#> 476 500.6720   3.392962
#> 477 500.9454   2.669665
#> 478 500.7624   2.756770
#> 479 502.7618   2.609132
#> 480 507.1164   2.964789
#> 481 893.1696   2.659693
#> 482 547.0365   2.901438
#> 483 769.3104   2.797497
#> 484 814.3589   3.628345
#> 485 500.1423   3.261514
#> 486 500.4464   2.643196
#> 487 500.8005   3.186367
#> 488 500.0512   3.834689
#> 489 536.9532   3.375707
#> 490 500.1506   3.206232
#> 491 500.2686   3.047981
#> 492 766.9295   3.785744
#> 493 500.2841   2.890044
#> 494 571.0091   2.570495
#> 495 509.4675   3.721588
#> 496 670.5050   2.780170
#> 497 500.0356   2.833870
#> 498 500.0537   2.698632
#> 499 500.0139   2.762865
#> 500 500.0119   2.594677
#> 501 508.0837   2.668062
#> 502 500.0445   3.080658
#> 503 515.3402   2.811243
#> 504 501.3697   2.566451
#> 505 500.2042   3.804236
#> 506 500.0155   2.522512
#> 507 500.0228   2.767606
#> 508 500.3264   2.622074
#> 509 540.2049   2.912325
#> 510 515.7458   2.595431
#> 511 732.7818   2.515871
#> 512 701.3053   3.883064
#> 513 567.9929   2.906923
#> 514 539.0460   2.516608
#> 515 500.6841   2.881753
#> 516 500.3288   2.616468
#> 517 502.3544   2.597549
#> 518 501.2944   3.647125
#> 519 512.7049   2.839470
#> 520 610.4433   2.535437
#> 521 520.4773   3.028387
#> 522 502.2125   2.681377
#> 523 749.9748   3.291180
#> 524 500.2805   2.968464
#> 525 500.6261   2.595788
#> 526 517.2076   2.996791
#> 527 501.2144   2.902970
#> 528 659.5559   2.991236
#> 529 524.3238   4.719291
#> 530 500.5445   6.074679
#> 531 569.7027   3.021905
#> 532 505.1802   2.799830
#> 533 572.9244   2.729772
#> 534 502.6315   3.323843
#> 535 506.0688   4.058996
#> 536 500.4515   2.715803
#> 537 503.7341   2.851835
#> 538 520.0559   3.104135
#> 539 749.4506   2.986865
#> 540 526.9258   2.706351
#> 541 534.3609   2.714687
#> 542 500.2332   2.719893
#> 543 500.0353   2.679792
#> 544 503.3018   2.877873
#> 545 506.6261   2.900776
#> 546 500.2079   3.047562
#> 547 663.3223   2.532387
#> 548 655.2533   3.672454
#> 549 509.5207   3.066731
#> 550 540.3678   2.781188
#> 551 500.0027   3.070707
#> 552 745.2240   2.722768
#> 553 500.1157   2.669559
#> 554 500.0110   3.018268
#> 555 592.2560   2.925748
#> 556 500.4063   2.619845
#> 557 500.5195   2.501420
#> 558 500.5672   3.594793
#> 559 521.6420   2.554650
#> 560 656.3648   3.087843
#> 561 500.3757   2.601268
#> 562 500.1485   2.931026
#> 563 905.1617   2.559618
#> 564 501.3396   2.800146
#> 565 510.6263   3.313811
#> 566 863.6680   2.565160
#> 567 500.0116   3.498884
#> 568 778.8795   3.002196
#> 569 880.3864   2.654344
#> 570 875.1666   3.346934
#> 571 803.8671   3.982990
#> 572 508.3490   2.728625
#> 573 500.0746   2.511482
#> 574 614.4101   2.888696
#> 575 501.3059   2.619395
#> 576 509.3470   2.661400
#> 577 525.2985   2.551549
#> 578 501.4219   3.133758
#> 579 500.1342   3.379414
#> 580 500.0511   2.591832
#> 581 527.9241   3.344010
#> 582 514.3331   2.573670
#> 583 500.9678   2.521999
#> 584 540.1287   2.547527
#> 585 874.3841   2.791130
#> 586 500.8276   2.611007
#> 587 506.3958   2.677670
#> 588 502.7172   2.541845
#> 589 503.4689   2.626024
#> 590 501.4089   2.945889
#> 591 515.8821   3.685560
#> 592 500.4722   2.549862
#> 593 506.2004   2.926447
#> 594 515.1906   4.225766
#> 595 590.7528   3.139171
#> 596 998.4452   2.840104
#> 597 574.0344   3.306145
#> 598 500.0390   2.676313
#> 599 504.3978   2.654129
#> 600 500.3702   3.515414
#> 601 500.6915   2.518616
#> 602 594.2540   3.285034
#> 603 505.0098   2.990406
#> 604 501.4408   2.733632
#> 605 511.4023   2.600504
#> 606 530.6665   2.768184
#> 607 530.1018   2.527169
#> 608 507.0796   2.631204
#> 609 930.3468   2.757116
#> 610 545.7482   3.879394
#> 611 500.5392   2.937247
#> 612 505.8380   2.502817
#> 613 639.2115   2.728027
#> 614 524.6125   3.284527
#> 615 529.4576   3.383628
#> 616 930.3515   2.794690
#> 617 610.7484   3.345997
#> 618 503.9539   3.100891
#> 619 506.5052   2.965104
#> 620 562.4300   3.653419
#> 621 505.0365   3.367888
#> 622 502.8820   3.880617
#> 623 500.3074   2.645219
#> 624 614.2538   2.846788
#> 625 534.0554   4.577273
#> 626 500.0389   3.022060
#> 627 529.7314   4.290502
#> 628 838.9291   2.538205
#> 629 501.0485   2.766265
#> 630 673.1765   5.127411
#> 631 500.0919   2.557575
#> 632 500.3137   2.682153
#> 633 507.2664   2.616832
#> 634 558.7756   2.825582
#> 635 547.4157   2.738400
#> 636 513.7624   2.765090
#> 637 500.0143   2.918091
#> 638 500.0388   2.551604
#> 639 507.2534   2.590677
#> 640 819.1153   4.060134
#> 641 514.8107   2.665283
#> 642 500.0778   2.813939
#> 643 511.6185   2.984500
#> 644 500.1010   2.568049
#> 645 515.0305   2.507575
#> 646 502.5519   3.529587
#> 647 506.9334   2.540886
#> 648 500.6363   3.159246
#> 649 991.2767   2.688724
#> 650 500.0064   4.156142
#> 651 500.1776   3.159734
#> 652 500.3249   2.871391
#> 653 500.3417   4.975948
#> 654 500.9997   2.520132
#> 655 501.2807   3.101026
#> 656 500.1223   2.772574
#> 657 718.6980   2.741216
#> 658 500.1146   2.627796
#> 659 500.1507   2.881394
#> 660 512.5096   2.685551
#> 661 500.8198   2.528715
#> 662 502.4497   2.539536
#> 663 510.0264   3.117749
#> 664 509.6492   2.718342
#> 665 502.9869   2.617311
#> 666 501.7343   2.679480
#> 667 500.1277   2.791749
#> 668 501.1898   2.519669
#> 669 500.5520   4.073500
#> 670 504.7190   3.323522
#> 671 740.5145   3.075891
#> 672 500.8516   2.786773
#> 673 500.0966   3.351911
#> 674 635.5774   2.638493
#> 675 586.8557   2.934011
#> 676 500.0039   2.731312
#> 677 500.0799   2.838830
#> 678 500.7370   3.016842
#> 679 501.1399   3.138575
#> 680 505.4328   2.696794
#> 681 500.2141   2.553684
#> 682 815.5004   2.871613
#> 683 531.1487   2.581742
#> 684 500.2401   2.758060
#> 685 506.1318   3.374846
#> 686 525.0838   2.686292
#> 687 500.3889   3.252506
#> 688 506.6786   2.838295
#> 689 500.1358   3.093194
#> 690 659.0397   5.602117
#> 691 500.2926   2.555288
#> 692 501.6595   2.771648
#> 693 519.1283   2.796049
#> 694 512.1312   3.431762
#> 695 500.0255   3.332469
#> 696 546.7196   2.653847
#> 697 500.1690   3.225074
#> 698 500.4805   2.744267
#> 699 733.8382   3.697976
#> 700 518.3009   2.587906
#> 701 500.0212   3.343666
#> 702 511.1531   2.540522
#> 703 535.8212   2.911093
#> 704 507.3784   2.610916
#> 705 740.8046   2.635336
#> 706 835.9266   2.606247
#> 707 511.1667   2.727186
#> 708 510.6440   2.610385
#> 709 893.9209   2.942332
#> 710 508.0824   2.824869
#> 711 569.4882   2.771877
#> 712 503.5590   2.643747
#> 713 546.6549   2.809043
#> 714 528.3731   3.897055
#> 715 549.9063   2.527754
#> 716 500.1572   2.951507
#> 717 500.0139   2.768585
#> 718 501.8874   2.781525
#> 719 501.0436   2.908655
#> 720 515.8720   2.670546
#> 721 500.0844   2.609382
#> 722 666.3762   3.281568
#> 723 500.0162   2.644575
#> 724 509.8693   2.654927
#> 725 500.8820   2.936357
#> 726 501.5188   2.849558
#> 727 500.1774   3.363884
#> 728 539.8844   3.249508
#> 729 500.4585   3.103960
#> 730 564.0449   2.545497
#> 731 527.3455   2.557511
#> 732 502.1677   2.515283
#> 733 835.2146   2.566606
#> 734 500.0562   2.595564
#> 735 512.8648   2.632570
#> 736 500.9104   2.868486
#> 737 510.3906   2.596811
#> 738 631.3648   2.760254
#> 739 504.1650   2.828657
#> 740 500.1442   3.539433
#> 741 792.3639   3.514543
#> 742 528.3291   2.675022
#> 743 500.8758   2.659337
#> 744 548.4975   2.718997
#> 745 500.3064   2.992950
#> 746 564.9938   2.536248
#> 747 500.4044   3.395872
#> 748 509.0832   2.760513
#> 749 506.7561   2.843243
#> 750 522.7470   3.122759
#> 751 502.2470   2.568506
#> 752 509.7089   3.438554
#> 753 652.6231   3.217226
#> 754 594.6930   2.530923
#> 755 500.0937   2.888876
#> 756 504.3849   2.554968
#> 757 500.2964   2.958876
#> 758 500.6487   2.635832
#> 759 570.2929   2.789754
#> 760 500.3059   3.072926
#> 761 512.7819   2.604963
#> 762 501.7710   3.059084
#> 763 681.0798   2.647562
#> 764 500.2125   2.950350
#> 765 526.7348   2.853569
#> 766 504.4815   2.960896
#> 767 505.2111   2.803052
#> 768 501.0412   2.605991
#> 769 553.6307   2.608698
#> 770 605.3606   2.518869
#> 771 505.1697   4.393209
#> 772 570.7353   3.082691
#> 773 677.9211   3.006478
#> 774 585.3020   3.261397
#> 775 510.5248   2.585302
#> 776 502.9131   3.296132
#> 777 760.9611   3.218460
#> 778 569.9016   2.749339
#> 779 511.8309   3.041326
#> 780 501.4962   4.433400
#> 781 531.6799   3.515247
#> 782 500.2033   2.577269
#> 783 500.5258   2.904108
#> 784 500.4198   2.501970
#> 785 500.9159   2.959929
#> 786 502.9661   2.666571
#> 787 501.3565   2.571590
#> 788 670.0455   3.373689
#> 789 589.2482   3.041200
#> 790 951.1881   3.058376
#> 791 783.3515   2.540598
#> 792 503.0327   2.538025
#> 793 509.5847   3.444033
#> 794 524.6521   3.732881
#> 795 501.1398   2.840979
#> 796 502.1823   3.425977
#> 797 501.3354   3.307889
#> 798 558.3161   3.107030
#> 799 512.5577   2.522595
#> 800 554.1604   2.543258
#> 801 510.6326   3.244072
#> 802 501.0264   2.646895
#> 803 520.4858   3.055283
#> 804 518.8153   2.521322
#> 805 621.1099   3.107552
#> 806 500.2117   2.745053
#> 807 707.3491   2.750067
#> 808 500.2948   3.028791
#> 809 500.1390   2.747850
#> 810 500.4546   2.652897
#> 811 500.3575   3.314253
#> 812 505.8676   3.353458
#> 813 500.0124   3.355767
#> 814 500.3261   3.923681
#> 815 502.0404   2.780345
#> 816 500.1182   2.939621
#> 817 500.0239   2.589980
#> 818 530.1241   2.856458
#> 819 566.4500   4.121087
#> 820 500.0213   5.919222
#> 821 500.1440   2.638957
#> 822 536.6921   2.821820
#> 823 502.8052   3.036355
#> 824 500.1936   2.773155
#> 825 518.0448   3.116229
#> 826 501.6487   3.044873
#> 827 516.5181   4.490245
#> 828 543.8464   3.378598
#> 829 500.1120   2.810143
#> 830 501.7217   2.559229
#> 831 552.3138   3.338391
#> 832 563.4402   2.512662
#> 833 522.5798   3.279747
#> 834 500.2976   2.536226
#> 835 513.6294   2.663139
#> 836 968.6816   2.593258
#> 837 508.0364   3.678806
#> 838 502.4098   3.145026
#> 839 505.3506   2.652036
#> 840 501.2868   3.706947
#> 841 502.6908   2.599036
#> 842 500.1885   2.671309
#> 843 516.0350   3.077992
#> 844 549.4318   2.552845
#> 845 500.0549   3.115972
#> 846 593.1950   2.728049
#> 847 500.0091   2.641599
#> 848 688.6775   3.024215
#> 849 502.9231   2.618329
#> 850 500.4902   2.828014
#> 851 560.1208   2.671319
#> 852 503.5805   2.586079
#> 853 502.9621   2.569616
#> 854 577.0335   2.610547
#> 855 501.4225   3.154647
#> 856 581.0821   3.106179
#> 857 500.3840   2.723629
#> 858 512.6059   2.828801
#> 859 508.2801   2.733331
#> 860 520.5596   2.731452
#> 861 503.5901   3.437831
#> 862 500.0675   2.957052
#> 863 782.1115   2.651121
#> 864 520.5584   2.619355
#> 865 501.0967   2.911281
#> 866 500.2522   3.168488
#> 867 501.2880   2.950987
#> 868 500.6631   2.765244
#> 869 500.1208   2.529014
#> 870 502.4022   2.656304
#> 871 500.3907   2.567537
#> 872 504.3598   3.054711
#> 873 504.7341   2.641087
#> 874 622.2465   3.853208
#> 875 500.1678   2.983288
#> 876 500.0447   2.509884
#> 877 807.2721   2.933318
#> 878 651.9909   2.544000
#> 879 505.4883   2.588139
#> 880 500.0761   2.604581
#> 881 518.7611   3.145324
#> 882 507.4719   3.609040
#> 883 500.3411   2.864298
#> 884 697.3351   3.654991
#> 885 537.4900   3.005496
#> 886 585.0759   2.619619
#> 887 501.2434   2.682805
#> 888 504.8910   2.932243
#> 889 501.4699   2.719360
#> 890 512.4825   2.646544
#> 891 505.3088   2.960237
#> 892 514.0292   2.578959
#> 893 500.0276   2.890344
#> 894 632.8731   2.555160
#> 895 500.0658   2.994337
#> 896 538.3633   2.881291
#> 897 500.4663   2.796764
#> 898 500.1797   2.620338
#> 899 791.1488   3.424817
#> 900 500.3714   3.247473
#> 901 504.2237   2.874016
#> 902 500.2597   3.668367
#> 903 500.1593   2.596696
#> 904 500.2962   2.604488
#> 905 761.8321   2.532349
#> 906 500.0950   2.542427
#> 907 500.0029   2.703336
#> 908 500.0606   3.072077
#> 909 500.8112   3.269685
#> 910 500.0176   2.658211
#> 911 536.6447   2.563627
#> 912 555.7360   2.645387
#> 913 503.1602   2.614271
#> 914 500.1484   2.569391
#> 915 506.7815   2.530095
#> 916 502.9970   2.586333
#> 917 501.4522   2.674352
#> 918 500.2341   2.592941
#> 919 534.6980   2.811330
#> 920 500.0541   2.542552
#> 921 501.1469   2.551492
#> 922 503.3360   2.690567
#> 923 566.2141   2.500636
#> 924 500.7158   3.491352
#> 925 502.2325   2.730409
#> 926 621.5017   2.742460
#> 927 500.3276   2.598662
#> 928 501.0298   3.341753
#> 929 500.1267   2.791147
#> 930 721.3940   3.153233
#> 931 753.8659   2.870472
#> 932 835.2276   2.569247
#> 933 567.6107   2.616867
#> 934 502.3209   3.281247
#> 935 500.1635   2.560251
#> 936 500.0117   2.741070
#> 937 500.9866   2.866518
#> 938 503.2719   2.932922
#> 939 500.1314   2.820422
#> 940 510.7914   3.678388
#> 941 535.2321   3.439446
#> 942 569.2312   3.053937
#> 943 501.7796   2.997870
#> 944 500.0055   2.575549
#> 945 500.1323   3.571906
#> 946 500.4556   2.882827
#> 947 500.3905   2.576234
#> 948 530.5485   3.156083
#> 949 504.1539   2.612760
#> 950 504.4533   2.614978
#> 951 503.3080   2.614840
#> 952 500.1339   2.702961
#> 953 677.4555   4.131287
#> 954 540.7906   2.769910
#> 955 500.5820   2.513825
#> 956 899.3352   2.513717
#> 957 501.3736   3.328526