Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   500.3482   3.042891
#> 2   877.3126   3.631992
#> 3   503.5671   2.676151
#> 4   500.0794   2.869043
#> 5   506.2070   3.324791
#> 6   500.0226   3.068826
#> 7   500.0545   2.715572
#> 8   500.0946   3.457584
#> 9   629.8665   2.658269
#> 10  500.5357   3.236086
#> 11  500.8676   3.057693
#> 12  501.9525   2.602733
#> 13  502.5852   2.798612
#> 14  617.0568   2.549843
#> 15  500.1702   2.750939
#> 16  543.0700   2.878971
#> 17  500.3924   2.769289
#> 18  504.2107   2.894039
#> 19  500.1441   2.674858
#> 20  500.3281   3.130097
#> 21  500.7762   3.581570
#> 22  500.0636   3.204125
#> 23  501.7912   2.673459
#> 24  821.8531   2.950730
#> 25  548.2830   2.501376
#> 26  646.9877   2.980490
#> 27  610.0331   2.761641
#> 28  983.7925   3.287385
#> 29  500.0754   2.528454
#> 30  849.2388   2.706823
#> 31  502.0548   2.771897
#> 32  911.4233   2.668571
#> 33  504.2587   3.173473
#> 34  500.6353   2.913179
#> 35  501.5862   2.715996
#> 36  500.0474   2.515717
#> 37  505.1620   3.856486
#> 38  502.7398   2.801220
#> 39  501.2886   2.727704
#> 40  500.3523   2.984889
#> 41  501.1441   2.508995
#> 42  824.4902   3.361922
#> 43  500.2951   2.577958
#> 44  500.3360   3.166945
#> 45  550.0515   2.812888
#> 46  515.9751   2.943210
#> 47  538.6475   2.598594
#> 48  658.8473   3.046850
#> 49  500.8262   3.279408
#> 50  500.3520   2.549788
#> 51  506.3153   2.852573
#> 52  502.1905   3.376387
#> 53  500.0591   2.935372
#> 54  500.4649   2.749763
#> 55  605.7854   2.709959
#> 56  506.1652   3.338095
#> 57  500.5053   2.566640
#> 58  521.8781   2.766427
#> 59  670.9832   4.256683
#> 60  518.7160   2.657235
#> 61  503.3567   2.686936
#> 62  500.0048   2.986579
#> 63  500.0253   3.211764
#> 64  823.5954   2.816231
#> 65  505.2764   2.915587
#> 66  500.4317   4.479485
#> 67  503.3839   2.515062
#> 68  985.1460   2.510686
#> 69  514.4226   2.520022
#> 70  500.6753   2.623815
#> 71  601.9574   2.726880
#> 72  501.2349   3.014610
#> 73  579.5179   2.872113
#> 74  500.0852   2.613820
#> 75  707.5502   2.754309
#> 76  908.9713   3.113439
#> 77  520.4246   2.616276
#> 78  549.1875   3.297943
#> 79  695.8124   2.812935
#> 80  500.9992   2.749423
#> 81  505.5088   3.101700
#> 82  501.4703   2.878374
#> 83  500.3434   2.720963
#> 84  526.5568   2.527922
#> 85  503.7858   3.432634
#> 86  506.9403   2.525903
#> 87  500.2824   3.241795
#> 88  503.0467   2.616903
#> 89  509.0491   2.524031
#> 90  789.3005   2.578922
#> 91  558.5078   2.863452
#> 92  586.3957   2.670508
#> 93  509.4476   3.613784
#> 94  500.7792   2.699853
#> 95  594.6041   2.739632
#> 96  652.2772   2.727701
#> 97  632.2136   2.654813
#> 98  500.4801   2.519296
#> 99  516.5988   3.209787
#> 100 501.2407   2.579395
#> 101 510.1790   2.685080
#> 102 593.3972   3.306745
#> 103 508.4796   3.207722
#> 104 504.6335   2.645251
#> 105 500.0355   2.620768
#> 106 605.0265   2.692649
#> 107 502.9324   3.019328
#> 108 648.0889   2.687866
#> 109 547.4641   3.403817
#> 110 507.0075   3.448998
#> 111 500.3546   2.585857
#> 112 542.9898   2.703918
#> 113 528.1176   2.738143
#> 114 500.0830   2.702271
#> 115 502.4677   3.163989
#> 116 500.3042   2.919017
#> 117 500.4257   2.558903
#> 118 522.2438   3.696587
#> 119 500.1434   2.745021
#> 120 654.5121   2.565120
#> 121 500.0081   2.554674
#> 122 577.1078   2.849402
#> 123 515.5101   2.697222
#> 124 500.3356   2.754605
#> 125 538.8490   3.449798
#> 126 684.4224   3.759407
#> 127 502.7507   2.992904
#> 128 500.0873   2.505495
#> 129 519.4644   2.776256
#> 130 500.1435   3.197655
#> 131 500.1431   2.571110
#> 132 591.8965   2.996545
#> 133 500.0771   2.893942
#> 134 500.1139   2.630048
#> 135 935.2580   2.698442
#> 136 509.4309   2.797954
#> 137 503.9480   3.173084
#> 138 510.9253   2.673989
#> 139 500.6267   2.694682
#> 140 524.7402   3.392634
#> 141 501.5373   3.347937
#> 142 500.2028   2.713041
#> 143 503.3347   2.563376
#> 144 501.6932   3.887871
#> 145 500.9466   2.560085
#> 146 501.8273   2.581858
#> 147 508.9889   3.047802
#> 148 503.8359   3.007775
#> 149 500.1143   2.794852
#> 150 502.2248   3.219049
#> 151 517.2697   2.506634
#> 152 507.4879   2.652717
#> 153 501.1385   2.737095
#> 154 787.5330   2.895536
#> 155 500.0942   2.729212
#> 156 757.1448   2.574715
#> 157 567.9585   2.651797
#> 158 500.9152   2.956471
#> 159 508.2449   2.758753
#> 160 503.2128   2.752529
#> 161 500.6755   2.867313
#> 162 500.7778   2.683145
#> 163 500.0934   2.692866
#> 164 756.2132   2.738972
#> 165 518.0592   3.176618
#> 166 500.1753   2.937468
#> 167 505.6106   2.532725
#> 168 517.6805   2.593906
#> 169 502.3701   3.025508
#> 170 501.6375   2.612466
#> 171 504.6834   2.723279
#> 172 678.6946   2.985264
#> 173 590.0379   2.680940
#> 174 746.3020   3.315892
#> 175 501.0641   2.813200
#> 176 518.3558   2.581683
#> 177 502.8010   3.136988
#> 178 503.3014   3.160370
#> 179 571.4192   2.606804
#> 180 500.1413   3.012538
#> 181 500.8822   2.616847
#> 182 500.1302   2.750970
#> 183 505.0682   2.919507
#> 184 501.4089   3.029192
#> 185 500.3938   3.783519
#> 186 765.4882   2.868409
#> 187 500.5782   3.191580
#> 188 505.0928   2.560869
#> 189 500.2496   2.541954
#> 190 501.2888   2.583445
#> 191 500.0275   2.681751
#> 192 500.2333   3.776420
#> 193 500.6608   3.927070
#> 194 812.3556   2.551139
#> 195 501.1329   3.070380
#> 196 500.7978   3.004690
#> 197 503.0777   3.018644
#> 198 607.4700   2.507565
#> 199 741.7588   2.582422
#> 200 500.1885   3.120177
#> 201 500.0042   2.669118
#> 202 500.1981   2.974423
#> 203 502.4228   3.562565
#> 204 503.0294   2.575569
#> 205 995.8890   3.342270
#> 206 916.3501   2.778627
#> 207 518.9974   2.597017
#> 208 562.0735   2.595461
#> 209 500.2419   2.676955
#> 210 500.1656   2.687766
#> 211 500.1053   3.751276
#> 212 502.8000   2.716027
#> 213 500.4846   2.673225
#> 214 500.1219   2.721082
#> 215 960.4583   3.628674
#> 216 510.8817   3.184757
#> 217 617.8932   2.959707
#> 218 501.4110   4.185932
#> 219 540.4713   2.508759
#> 220 500.2695   2.813575
#> 221 500.0411   2.714701
#> 222 502.2549   3.149540
#> 223 501.7393   3.024604
#> 224 502.2551   3.012192
#> 225 505.3575   2.801196
#> 226 500.0219   3.875407
#> 227 740.2708   2.851546
#> 228 505.9316   3.942647
#> 229 500.4125   2.600192
#> 230 500.9789   3.153205
#> 231 824.2635   2.518970
#> 232 500.0747   2.928422
#> 233 500.4792   3.849005
#> 234 500.0778   2.724580
#> 235 500.0362   2.586152
#> 236 507.7952   3.583971
#> 237 500.1442   3.236562
#> 238 614.9916   2.998342
#> 239 510.5155   3.004901
#> 240 539.5664   4.563273
#> 241 591.8857   2.511396
#> 242 504.5348   3.022921
#> 243 513.1979   2.555740
#> 244 500.1470   2.575221
#> 245 501.3194   3.448658
#> 246 500.1876   2.799309
#> 247 525.9059   3.836507
#> 248 504.8541   2.650494
#> 249 549.2910   2.968540
#> 250 513.2315   3.214534
#> 251 506.1636   2.909286
#> 252 929.9772   3.219224
#> 253 749.3194   2.854103
#> 254 590.3553   2.602119
#> 255 500.0312   2.890258
#> 256 551.0477   2.516416
#> 257 506.1259   2.640307
#> 258 500.1416   2.618560
#> 259 506.3818   2.796979
#> 260 500.2062   2.637800
#> 261 500.4460   2.745237
#> 262 500.4478   2.903985
#> 263 506.3480   2.832812
#> 264 500.0403   3.394560
#> 265 500.1953   2.645281
#> 266 500.6399   2.855249
#> 267 655.3059   2.767436
#> 268 500.2395   2.648555
#> 269 590.4767   3.044839
#> 270 500.2961   3.244753
#> 271 607.6100   3.982287
#> 272 501.3829   3.659854
#> 273 508.4564   2.609338
#> 274 500.0206   2.644331
#> 275 737.4495   2.639998
#> 276 502.5013   2.978165
#> 277 500.0496   2.534975
#> 278 500.3305   2.565938
#> 279 526.7993   4.068705
#> 280 502.5504   2.510330
#> 281 500.5737   3.178406
#> 282 500.0818   2.888209
#> 283 500.4372   2.765815
#> 284 500.0821   2.766515
#> 285 518.7785   2.526237
#> 286 500.7376   3.041653
#> 287 502.2921   2.642780
#> 288 500.2714   3.364067
#> 289 510.9811   3.767616
#> 290 500.4989   3.793673
#> 291 509.4622   3.070748
#> 292 500.0435   2.853686
#> 293 624.3673   2.841880
#> 294 540.3543   2.612582
#> 295 500.4406   3.315993
#> 296 500.1241   2.796828
#> 297 601.1325   2.539966
#> 298 500.1384   2.612919
#> 299 506.5197   2.599897
#> 300 500.1573   2.715799
#> 301 502.7632   2.626068
#> 302 501.7656   2.638153
#> 303 500.2445   2.830969
#> 304 500.5162   2.883723
#> 305 519.1292   2.860394
#> 306 502.9865   3.600262
#> 307 503.9677   2.796848
#> 308 500.1456   2.510426
#> 309 510.1215   3.020722
#> 310 500.0966   2.732277
#> 311 500.1154   2.550961
#> 312 500.4409   2.761747
#> 313 500.3990   3.720890
#> 314 500.1006   2.745982
#> 315 500.2069   2.592186
#> 316 500.9903   2.987864
#> 317 989.7457   3.926883
#> 318 545.1392   2.746208
#> 319 738.1612   2.527114
#> 320 503.5969   2.953711
#> 321 599.9574   2.698835
#> 322 501.6143   3.479199
#> 323 514.9323   2.859093
#> 324 520.8853   2.856602
#> 325 502.7888   3.216347
#> 326 500.4297   2.683543
#> 327 500.1040   2.604239
#> 328 573.2813   2.766769
#> 329 500.2912   2.590853
#> 330 500.1468   2.948217
#> 331 578.7884   2.515988
#> 332 500.0196   3.462125
#> 333 501.1241   2.514952
#> 334 500.1634   2.848663
#> 335 501.1665   2.727794
#> 336 500.0137   3.123897
#> 337 501.1806   2.505153
#> 338 710.8968   2.829543
#> 339 500.2073   3.847612
#> 340 513.7108   3.376589
#> 341 675.0649   2.907827
#> 342 542.2266   2.954279
#> 343 598.8841   2.665748
#> 344 500.0426   2.988079
#> 345 500.5756   2.555060
#> 346 844.4956   2.573638
#> 347 500.1134   2.528006
#> 348 501.2058   2.527558
#> 349 505.8323   3.561379
#> 350 500.0997   2.784527
#> 351 500.0301   2.969956
#> 352 723.1210   2.571550
#> 353 535.5932   3.316877
#> 354 500.6021   2.607399
#> 355 500.3039   2.680497
#> 356 507.5658   2.987289
#> 357 500.0793   3.715520
#> 358 500.1026   2.848967
#> 359 502.4126   2.927524
#> 360 501.7105   2.581415
#> 361 517.6496   2.573766
#> 362 571.4706   3.181795
#> 363 500.0380   3.499735
#> 364 514.8630   3.014540
#> 365 500.9433   2.596753
#> 366 501.4344   2.555365
#> 367 527.9886   3.484059
#> 368 679.3236   3.571504
#> 369 501.6283   2.950288
#> 370 537.8522   2.697968
#> 371 505.4099   3.516639
#> 372 500.3774   2.520998
#> 373 500.1002   3.150749
#> 374 500.3489   2.819234
#> 375 712.4251   2.798435
#> 376 501.3747   3.285916
#> 377 501.4553   4.656510
#> 378 500.0547   2.577763
#> 379 500.1142   2.549171
#> 380 510.2794   3.044364
#> 381 521.1004   3.263889
#> 382 859.1784   2.768396
#> 383 500.1268   3.517256
#> 384 500.0531   2.562137
#> 385 500.1846   4.456441
#> 386 501.2709   2.555463
#> 387 510.5396   3.475007
#> 388 503.7420   3.035281
#> 389 500.3652   2.596002
#> 390 500.1078   3.525885
#> 391 516.9676   2.583392
#> 392 505.6523   2.527965
#> 393 505.3923   2.665871
#> 394 500.0647   3.349702
#> 395 545.6582   3.214859
#> 396 518.2509   2.617627
#> 397 509.8326   2.552655
#> 398 500.1176   2.814688
#> 399 500.6422   2.919525
#> 400 500.0689   2.520493
#> 401 503.7282   2.663775
#> 402 529.0588   2.587425
#> 403 564.3767   2.913206
#> 404 571.5403   2.727772
#> 405 500.4168   2.871891
#> 406 615.4329   2.506298
#> 407 502.4897   3.205488
#> 408 505.7060   2.972877
#> 409 706.9313   3.118848
#> 410 503.7817   2.784567
#> 411 501.4707   2.533574
#> 412 501.2894   2.630128
#> 413 661.8058   3.580332
#> 414 500.0219   2.831617
#> 415 525.0884   2.989713
#> 416 610.2116   2.575066
#> 417 500.2450   2.980246
#> 418 500.0850   3.745182
#> 419 976.2839   3.360120
#> 420 500.1156   2.564446
#> 421 500.5973   2.709460
#> 422 527.9690   2.790623
#> 423 500.6563   2.795952
#> 424 501.5167   2.623093
#> 425 563.8499   2.894867
#> 426 956.4460   3.650067
#> 427 500.5316   3.202082
#> 428 501.2935   2.910500
#> 429 500.6279   2.919316
#> 430 500.0404   4.146868
#> 431 799.9276   2.871740
#> 432 671.1002   2.715122
#> 433 500.3073   2.551273
#> 434 556.6713   2.869569
#> 435 699.3411   2.565226
#> 436 501.4148   2.577405
#> 437 505.5849   2.961410
#> 438 515.5147   2.542567
#> 439 612.7878   2.818922
#> 440 500.2448   2.584535
#> 441 518.6213   2.616838
#> 442 504.4381   3.291438
#> 443 541.3835   3.605043
#> 444 500.6605   2.758444
#> 445 500.0934   3.106705
#> 446 500.3180   3.295946
#> 447 502.5929   2.919925
#> 448 506.5146   2.772102
#> 449 503.7144   2.537751
#> 450 501.3478   3.325172
#> 451 526.8359   3.001404
#> 452 547.3669   2.830033
#> 453 513.5587   2.851839
#> 454 500.4738   2.774760
#> 455 528.5743   3.203069
#> 456 507.5257   2.786071
#> 457 500.6007   2.606962
#> 458 501.1562   3.120668
#> 459 507.2494   2.584235
#> 460 500.2165   2.791693
#> 461 534.4645   3.175596
#> 462 521.0708   2.589279
#> 463 522.0671   2.645260
#> 464 500.2201   2.567402
#> 465 500.0486   3.104187
#> 466 577.6945   3.194370
#> 467 502.0976   2.843788
#> 468 512.3971   2.816663
#> 469 753.3818   2.734126
#> 470 506.9822   2.511073
#> 471 555.9676   2.580580
#> 472 559.9841   3.273359
#> 473 500.0069   2.887804
#> 474 501.5155   3.197169
#> 475 663.8554   2.522649
#> 476 758.9806   3.795818
#> 477 504.6939   2.676097
#> 478 502.2922   2.813535
#> 479 500.0165   2.616271
#> 480 514.3734   2.575250
#> 481 502.0861   2.582998
#> 482 500.3737   3.248214
#> 483 500.3308   3.126204
#> 484 545.6671   3.027308
#> 485 501.1297   3.845993
#> 486 500.0252   3.915295
#> 487 544.4087   3.139412
#> 488 622.4693   2.505442
#> 489 825.9937   3.304109
#> 490 575.2073   2.829517
#> 491 534.4812   2.628838
#> 492 508.0688   3.283227
#> 493 500.6228   2.755554
#> 494 916.8046   2.585705
#> 495 500.3683   2.581719
#> 496 500.2113   2.513040
#> 497 583.9358   2.657599
#> 498 606.0589   2.547890
#> 499 669.0858   3.959108
#> 500 575.4104   3.799585
#> 501 503.9547   3.502818
#> 502 500.0431   3.329750
#> 503 500.0990   3.256363
#> 504 507.9311   3.163512
#> 505 500.0678   2.564280
#> 506 504.6690   2.577105
#> 507 501.3050   2.822418
#> 508 594.5033   3.788117
#> 509 502.5619   2.638740
#> 510 503.1821   2.729922
#> 511 502.9450   3.424475
#> 512 555.1151   2.510867
#> 513 500.0871   2.615414
#> 514 516.2119   2.883418
#> 515 877.3793   2.709562
#> 516 500.6393   3.190786
#> 517 984.3646   3.755186
#> 518 519.3071   3.085502
#> 519 504.5725   3.337011
#> 520 501.4342   2.636107
#> 521 500.1856   3.164292
#> 522 518.7304   2.643077
#> 523 503.7185   3.655311
#> 524 500.2578   2.506405
#> 525 501.0191   2.501254
#> 526 500.0443   2.749551
#> 527 504.0565   2.866092
#> 528 504.3953   2.748669
#> 529 532.2965   2.601429
#> 530 502.0973   2.600269
#> 531 514.0825   2.826867
#> 532 544.7038   3.107064
#> 533 572.5802   2.739442
#> 534 504.0613   2.928254
#> 535 521.9966   2.888141
#> 536 501.1121   3.264454
#> 537 845.7844   3.085266
#> 538 503.0753   3.142130
#> 539 503.9179   3.063239
#> 540 500.1072   3.070170
#> 541 500.0639   2.740499
#> 542 500.5048   2.773849
#> 543 500.4750   2.513288
#> 544 518.1459   2.872438
#> 545 506.4724   2.558377
#> 546 501.0735   3.136591
#> 547 500.3479   2.768628
#> 548 549.3352   3.361935
#> 549 523.3477   3.712915
#> 550 500.1641   3.012552
#> 551 537.6379   2.505047
#> 552 500.6517   2.793211
#> 553 500.6786   2.537110
#> 554 500.5661   2.668867
#> 555 529.4041   2.649905
#> 556 501.0093   2.534248
#> 557 500.6741   2.554450
#> 558 500.3451   2.626114
#> 559 501.8344   3.091623
#> 560 507.2909   2.562314
#> 561 508.7756   3.059001
#> 562 637.9584   3.440474
#> 563 501.5166   3.637803
#> 564 502.2401   3.014240
#> 565 512.7747   2.911114
#> 566 503.8562   2.714538
#> 567 500.1050   2.529394
#> 568 500.5212   3.325908
#> 569 520.9734   3.340374
#> 570 524.9839   2.839965
#> 571 501.5682   2.616530
#> 572 504.4297   3.626543
#> 573 504.7307   2.566642
#> 574 501.4858   2.706806
#> 575 506.5964   2.940889
#> 576 621.7479   2.866546
#> 577 500.4215   2.709426
#> 578 539.4478   2.959290
#> 579 501.1771   2.598358
#> 580 966.0331   4.243991
#> 581 504.0450   2.704135
#> 582 535.9978   2.551889
#> 583 501.7531   3.313229
#> 584 501.0971   3.696578
#> 585 518.3669   2.985102
#> 586 510.1060   2.927466
#> 587 500.9928   2.666900
#> 588 500.0338   2.790530
#> 589 776.3341   3.165739
#> 590 637.1302   2.716390
#> 591 502.2861   2.940069
#> 592 507.4194   3.113516
#> 593 501.5813   2.736175
#> 594 515.3970   2.829618
#> 595 754.5715   2.813220
#> 596 500.0515   2.617713
#> 597 501.8808   3.098751
#> 598 596.6845   2.946745
#> 599 528.6925   4.012863
#> 600 514.9754   3.017448
#> 601 500.0525   2.881322
#> 602 504.4377   3.068196
#> 603 507.7705   2.575443
#> 604 592.7532   2.533164
#> 605 721.1501   2.751381
#> 606 501.7124   2.801273
#> 607 523.2436   2.659333
#> 608 500.1428   2.538900
#> 609 577.7413   2.722097
#> 610 500.0816   2.579761
#> 611 502.1544   5.094019
#> 612 942.9788   2.803527
#> 613 509.4600   2.531027
#> 614 504.2322   2.540744
#> 615 504.2997   2.774060
#> 616 515.0615   3.214503
#> 617 632.6717   2.570609
#> 618 500.0344   2.713055
#> 619 574.7381   3.427924
#> 620 501.5414   3.130661
#> 621 523.1791   2.604168
#> 622 507.3103   2.513233
#> 623 572.8170   2.942203
#> 624 504.7903   2.593882
#> 625 500.0955   2.531041
#> 626 502.2240   2.812127
#> 627 518.4830   2.603984
#> 628 827.8872   2.884400
#> 629 500.3400   2.769238
#> 630 500.3994   2.657341
#> 631 522.4363   2.528434
#> 632 513.7052   2.541406
#> 633 502.8485   2.579065
#> 634 560.4239   2.914755
#> 635 501.1518   2.527915
#> 636 502.5967   2.583737
#> 637 910.0911   2.730111
#> 638 833.1098   4.786386
#> 639 500.1232   2.779486
#> 640 504.0378   2.567104
#> 641 524.7816   2.927386
#> 642 500.1749   2.547703
#> 643 500.0552   2.599708
#> 644 519.5334   3.198247
#> 645 500.1476   3.162091
#> 646 500.3589   3.847509
#> 647 587.7926   2.980739
#> 648 581.6727   2.650084
#> 649 500.0842   2.534885
#> 650 503.0740   3.354556
#> 651 500.8854   4.013907
#> 652 597.3001   3.201270
#> 653 705.1811   3.213807
#> 654 539.2522   3.430524
#> 655 502.6494   2.733749
#> 656 500.3953   3.460973
#> 657 527.7517   2.892169
#> 658 500.9530   2.825939
#> 659 568.1250   2.615132
#> 660 504.3164   3.208970
#> 661 500.7022   2.639369
#> 662 500.0121   2.679336
#> 663 758.1221   3.024066
#> 664 500.0363   3.166153
#> 665 507.6910   2.692283
#> 666 500.7740   3.292592
#> 667 500.1857   2.715803
#> 668 500.1502   3.004368
#> 669 926.6962   2.704366
#> 670 500.6404   3.016499
#> 671 506.2486   4.188866
#> 672 524.5792   3.039496
#> 673 503.1633   2.752528
#> 674 505.7830   2.758458
#> 675 510.9293   3.362203
#> 676 500.0188   2.795609
#> 677 536.4603   2.566325
#> 678 508.0157   2.538534
#> 679 500.0331   3.706376
#> 680 500.2848   2.818678
#> 681 500.4550   2.842659
#> 682 566.4661   2.650804
#> 683 500.0367   3.372845
#> 684 500.0674   2.523137
#> 685 509.5849   2.802197
#> 686 500.2174   2.553814
#> 687 502.1179   2.704731
#> 688 500.0126   2.842959
#> 689 840.8511   2.937888
#> 690 500.0933   3.062261
#> 691 818.0154   3.371461
#> 692 500.4408   2.893712
#> 693 621.4284   2.770973
#> 694 503.6023   2.904796
#> 695 994.4897   2.557504
#> 696 506.2946   3.052499
#> 697 500.4190   2.782376
#> 698 612.8751   2.649370
#> 699 500.0876   3.171949
#> 700 603.0424   3.097118
#> 701 502.0058   2.740419
#> 702 500.0006   3.573275
#> 703 500.1338   2.643746
#> 704 500.4125   2.614064
#> 705 502.9567   3.080110
#> 706 500.7287   3.049356
#> 707 500.1586   2.590413
#> 708 926.7299   3.095455
#> 709 521.8534   2.954900
#> 710 500.1312   2.637712
#> 711 502.4665   3.256118
#> 712 500.6037   2.627126
#> 713 629.7361   3.365890
#> 714 500.0054   2.519768
#> 715 500.0062   3.458826
#> 716 500.7840   3.536818
#> 717 500.1694   3.909401
#> 718 502.3203   4.195612
#> 719 500.2347   2.637345
#> 720 794.1660   3.431769
#> 721 559.0504   2.511601
#> 722 501.3739   2.551823
#> 723 500.1785   2.591048
#> 724 501.2315   2.993547
#> 725 503.0079   2.514630
#> 726 500.1380   2.586344
#> 727 500.2348   2.791462
#> 728 500.4745   2.673208
#> 729 500.7045   3.197269
#> 730 501.9880   2.581804
#> 731 503.3474   2.998413
#> 732 500.3042   3.131522
#> 733 507.4474   3.633391
#> 734 596.3218   2.581024
#> 735 501.6773   2.859862
#> 736 512.6970   2.788902
#> 737 505.9473   2.529382
#> 738 530.2708   2.560101
#> 739 510.2189   3.060575
#> 740 505.4621   2.557505
#> 741 580.8503   2.522536
#> 742 514.2203   2.621969
#> 743 509.6587   3.716755
#> 744 500.0306   2.903007
#> 745 601.5145   2.883308
#> 746 510.2625   2.902453
#> 747 500.8854   2.653902
#> 748 500.3482   4.313757
#> 749 500.7502   2.727140
#> 750 877.2513   2.735209
#> 751 503.5311   2.640216
#> 752 500.0706   2.585574
#> 753 683.8261   3.939546
#> 754 601.6621   3.148507
#> 755 500.0055   2.764001
#> 756 596.1085   2.574796
#> 757 503.5108   4.558517
#> 758 555.6427   2.856514
#> 759 500.0027   2.738798
#> 760 503.4838   2.576771
#> 761 500.7123   2.542655
#> 762 500.4677   3.388769
#> 763 554.3075   2.518455
#> 764 536.6970   3.398246
#> 765 507.1625   2.522022
#> 766 503.1892   3.643786
#> 767 504.5062   2.523445
#> 768 518.2548   3.301547
#> 769 500.1109   2.828483
#> 770 525.7281   2.940357
#> 771 589.5833   2.669056
#> 772 500.6911   2.678871
#> 773 511.9131   2.946712
#> 774 500.0896   2.561759
#> 775 633.7652   2.508907
#> 776 628.2112   2.567519
#> 777 500.2729   2.815414
#> 778 681.1236   3.262550
#> 779 503.2486   3.036900
#> 780 500.0944   2.570709
#> 781 501.4315   4.112431
#> 782 595.5836   2.647764
#> 783 536.8250   2.691095
#> 784 509.8675   3.365902
#> 785 500.1290   3.708026
#> 786 500.0712   2.654377
#> 787 500.7465   2.555339
#> 788 500.2900   2.598109
#> 789 500.2729   3.083748
#> 790 511.7577   2.971029
#> 791 523.3030   3.198037
#> 792 500.0829   3.365794
#> 793 504.3733   2.528712
#> 794 877.2182   2.798033
#> 795 502.6178   3.213126
#> 796 643.0872   2.629831
#> 797 500.1318   2.626305
#> 798 500.2007   2.835436
#> 799 985.3685   2.551097
#> 800 500.5124   2.517852
#> 801 500.0914   2.630206
#> 802 500.0163   2.805566
#> 803 517.6162   2.752681
#> 804 502.1111   2.820253
#> 805 500.1092   2.518648
#> 806 572.5905   3.399880
#> 807 506.3258   2.577283
#> 808 500.0411   2.513789
#> 809 500.8458   2.739971
#> 810 909.4159   2.836811
#> 811 500.0449   2.554088
#> 812 506.3592   2.790351
#> 813 503.2147   3.158519
#> 814 552.4503   2.530246
#> 815 783.7069   2.701257
#> 816 503.6024   2.898766
#> 817 803.0240   3.250556
#> 818 500.0326   2.610385
#> 819 500.0636   2.611253
#> 820 524.2445   3.220528
#> 821 500.1275   2.646088
#> 822 832.4900   2.606421
#> 823 723.5442   3.029590
#> 824 500.1977   2.690330
#> 825 515.5147   2.532786
#> 826 582.4715   2.636715
#> 827 502.0874   4.699006
#> 828 619.7549   2.571473
#> 829 500.4857   3.762472
#> 830 503.1011   2.590832
#> 831 500.2066   2.714962
#> 832 501.2940   2.967992
#> 833 501.6256   3.477438
#> 834 500.1281   4.505227
#> 835 735.1753   2.631599
#> 836 501.5276   2.717436
#> 837 634.8347   2.976752
#> 838 501.2841   2.829484
#> 839 504.2909   3.421419
#> 840 584.4505   2.518503
#> 841 502.2271   3.464274
#> 842 500.3395   2.629479
#> 843 823.4798   2.507701
#> 844 500.0413   3.209108
#> 845 602.9125   2.950615
#> 846 513.6541   2.684539
#> 847 508.1707   2.577512
#> 848 500.4428   3.268643
#> 849 500.0189   2.566905
#> 850 568.3840   3.381062
#> 851 500.1595   2.595885
#> 852 507.1026   2.640788
#> 853 511.9731   2.564915
#> 854 860.2282   2.683881
#> 855 500.0232   3.549655
#> 856 547.2222   2.559406
#> 857 503.5521   2.969273
#> 858 510.8785   2.540932
#> 859 514.2302   2.551127
#> 860 599.5819   3.414251
#> 861 538.8630   2.641444
#> 862 508.5370   2.653262
#> 863 618.4535   2.600088
#> 864 510.7827   3.098137
#> 865 574.7425   3.391068
#> 866 537.5377   3.093279
#> 867 500.1364   2.725220
#> 868 500.0022   3.109314
#> 869 503.9017   2.793840
#> 870 624.9627   2.795869
#> 871 502.2319   3.293014
#> 872 501.1866   3.010193
#> 873 504.8529   3.888426
#> 874 500.1694   3.079320
#> 875 796.2712   3.774967
#> 876 572.7554   2.760583
#> 877 502.0960   2.555574
#> 878 509.0890   3.445712
#> 879 522.4787   3.999512
#> 880 548.0351   2.930945
#> 881 642.8865   2.636078
#> 882 502.8775   2.564658
#> 883 500.2396   2.997370
#> 884 500.0667   2.927765
#> 885 506.1792   2.588286
#> 886 500.2159   3.013523
#> 887 519.2359   2.636667
#> 888 527.7993   3.383894
#> 889 500.0136   3.596985
#> 890 500.1668   2.961531
#> 891 500.9514   2.678989
#> 892 500.1381   2.877365
#> 893 502.1656   3.621524
#> 894 505.6347   2.838731
#> 895 501.9501   3.060171
#> 896 500.0538   2.655272
#> 897 502.2431   3.029114
#> 898 610.4084   2.620003
#> 899 500.0100   2.663808
#> 900 505.9297   2.916921
#> 901 500.1004   2.886499
#> 902 790.3307   2.831829
#> 903 543.7062   3.015704
#> 904 500.1096   3.202300
#> 905 635.1769   3.479782
#> 906 586.1012   3.433032
#> 907 502.8928   2.618515
#> 908 500.7806   2.764651
#> 909 506.9771   2.746659
#> 910 603.5883   2.684182
#> 911 508.1672   2.692446
#> 912 502.1566   3.127194
#> 913 500.7064   4.838254
#> 914 552.8918   3.569480
#> 915 500.2311   2.716369
#> 916 500.3092   2.922366
#> 917 582.7102   3.061600
#> 918 950.9290   2.588215
#> 919 500.0463   5.271955
#> 920 500.3656   2.594675
#> 921 501.8060   2.501178
#> 922 587.6653   2.509786
#> 923 502.3451   2.843638
#> 924 500.0062   2.674364
#> 925 500.6013   2.839347
#> 926 519.6897   3.252026
#> 927 500.3603   2.573147
#> 928 539.4272   2.979057
#> 929 502.8914   3.062059
#> 930 500.0393   3.352096
#> 931 500.9648   3.076550
#> 932 505.0169   3.006457
#> 933 571.5870   3.557194
#> 934 504.8792   2.659641
#> 935 500.6852   3.004488
#> 936 521.2090   3.072434
#> 937 588.1198   2.606643
#> 938 500.5754   2.959763
#> 939 501.6429   2.675506
#> 940 783.7761   2.999906
#> 941 500.3997   2.637066
#> 942 500.5433   2.687494
#> 943 509.2175   2.978983
#> 944 613.3750   2.585176
#> 945 505.4042   2.502400
#> 946 500.7900   3.117210
#> 947 503.5039   2.519963
#> 948 501.5163   2.510382
#> 949 500.0350   2.502615
#> 950 679.6805   4.039204
#> 951 912.1591   2.667011
#> 952 500.2813   2.781117
#> 953 500.1221   2.845915