Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   606.7829   2.660593
#> 2   507.0531   2.776350
#> 3   500.2806   2.699527
#> 4   511.9493   3.387901
#> 5   534.5323   2.556496
#> 6   503.2521   3.205622
#> 7   629.0477   4.082219
#> 8   517.9817   2.548640
#> 9   513.3256   2.778454
#> 10  504.1735   2.540565
#> 11  508.7701   2.528043
#> 12  500.2082   2.524953
#> 13  506.9296   2.713616
#> 14  701.8830   2.693402
#> 15  501.2842   2.689982
#> 16  523.3948   2.797734
#> 17  547.8956   2.561310
#> 18  500.9174   3.067755
#> 19  546.8052   2.705021
#> 20  500.6911   2.587814
#> 21  502.7444   2.754425
#> 22  533.9121   2.721480
#> 23  503.4070   2.528648
#> 24  530.4531   2.505443
#> 25  500.1471   2.733814
#> 26  500.8607   3.937733
#> 27  500.0138   2.650010
#> 28  609.8005   2.683987
#> 29  500.8493   2.502554
#> 30  603.7360   2.769582
#> 31  502.1033   2.669464
#> 32  533.5940   2.560586
#> 33  656.1689   2.556802
#> 34  559.6853   2.675550
#> 35  539.4148   2.828551
#> 36  500.2222   2.758972
#> 37  822.1708   2.979848
#> 38  501.4434   3.524936
#> 39  500.2553   2.692071
#> 40  501.2408   2.573876
#> 41  500.0269   2.512179
#> 42  501.0269   2.591255
#> 43  649.7609   2.590986
#> 44  503.6927   2.508013
#> 45  500.0178   2.861977
#> 46  500.7109   2.737112
#> 47  500.0720   2.973493
#> 48  501.0188   2.935215
#> 49  502.2447   2.819399
#> 50  500.2273   2.677667
#> 51  500.0642   2.689537
#> 52  524.0448   2.531985
#> 53  500.6168   2.701310
#> 54  508.8789   2.756904
#> 55  501.3082   2.970883
#> 56  501.1025   2.562603
#> 57  502.3418   2.601871
#> 58  520.5727   2.576595
#> 59  500.2233   2.819245
#> 60  502.0574   2.751034
#> 61  511.5514   2.625004
#> 62  507.5443   2.829212
#> 63  500.5041   2.732221
#> 64  743.3099   2.915990
#> 65  500.0546   3.312822
#> 66  500.6439   2.601764
#> 67  501.0068   2.596114
#> 68  504.3057   2.752895
#> 69  783.1934   2.510872
#> 70  501.5952   3.272091
#> 71  506.0818   2.833883
#> 72  529.8517   2.630531
#> 73  743.0044   2.607875
#> 74  553.0722   2.555610
#> 75  532.0383   3.004325
#> 76  500.0169   2.758028
#> 77  519.5014   2.600511
#> 78  507.4631   2.836761
#> 79  503.4080   2.839461
#> 80  503.6128   2.819964
#> 81  500.2197   2.811248
#> 82  519.5092   2.539173
#> 83  508.9611   2.778790
#> 84  500.7196   2.860008
#> 85  504.6372   3.085581
#> 86  555.0048   2.798110
#> 87  502.8334   2.586596
#> 88  500.3746   2.545186
#> 89  543.5664   3.314491
#> 90  500.0195   2.559244
#> 91  596.8014   2.859702
#> 92  523.4784   2.805154
#> 93  617.6668   2.901489
#> 94  502.5810   2.922358
#> 95  501.1951   3.283959
#> 96  675.6187   3.271471
#> 97  500.1418   2.530045
#> 98  800.2664   3.313883
#> 99  511.0297   2.730985
#> 100 500.0982   3.377912
#> 101 500.8435   3.788839
#> 102 506.2326   2.773518
#> 103 500.0295   2.585489
#> 104 718.5474   2.829002
#> 105 500.0496   2.813830
#> 106 502.3349   2.886201
#> 107 500.2227   2.548642
#> 108 566.2860   2.613838
#> 109 802.9672   4.899890
#> 110 505.9811   3.251967
#> 111 560.4186   2.530562
#> 112 500.0261   2.591847
#> 113 500.8785   2.764296
#> 114 500.5542   2.626643
#> 115 500.1156   2.996504
#> 116 613.0859   2.871345
#> 117 517.1094   3.365979
#> 118 905.8143   2.622535
#> 119 531.2294   2.950851
#> 120 507.3950   2.542969
#> 121 506.0793   2.639145
#> 122 535.6356   4.140423
#> 123 501.0042   4.118353
#> 124 844.1117   2.609353
#> 125 500.0187   2.523559
#> 126 501.8092   2.580694
#> 127 501.1476   2.938982
#> 128 500.3921   2.892062
#> 129 519.1956   2.823059
#> 130 506.9972   2.736696
#> 131 507.0209   4.069950
#> 132 500.2919   2.541799
#> 133 501.6216   3.038187
#> 134 939.3849   2.607462
#> 135 502.0672   2.609111
#> 136 503.5606   3.196116
#> 137 500.4066   3.451112
#> 138 507.8822   2.535394
#> 139 655.4206   2.726783
#> 140 641.8582   3.632920
#> 141 615.4279   3.127768
#> 142 500.2117   2.782619
#> 143 500.3792   2.632773
#> 144 501.5971   2.535533
#> 145 507.4394   3.207477
#> 146 500.0161   2.894659
#> 147 500.2014   2.537826
#> 148 500.0011   2.948295
#> 149 500.3243   2.848937
#> 150 500.0492   2.505537
#> 151 501.9928   4.079729
#> 152 523.3201   3.245519
#> 153 672.8293   2.503090
#> 154 535.6793   2.606315
#> 155 507.5607   2.568663
#> 156 512.7626   3.191731
#> 157 500.4348   3.488367
#> 158 505.9974   2.579727
#> 159 500.0114   2.675419
#> 160 526.7139   2.957332
#> 161 649.9111   2.867561
#> 162 500.6132   2.622769
#> 163 849.1318   3.538209
#> 164 500.7214   3.070904
#> 165 500.2559   2.563435
#> 166 918.2887   2.759894
#> 167 585.5310   3.088643
#> 168 500.1583   2.806874
#> 169 504.8954   2.578465
#> 170 900.9013   2.906610
#> 171 503.1943   2.863179
#> 172 595.5846   2.812288
#> 173 500.2577   2.516475
#> 174 500.8506   2.514503
#> 175 500.0253   2.659569
#> 176 500.0281   3.274847
#> 177 500.0042   2.527232
#> 178 500.7692   2.757401
#> 179 582.5187   2.564215
#> 180 545.0443   2.584681
#> 181 501.7758   3.454497
#> 182 500.1903   2.527433
#> 183 501.2518   4.140923
#> 184 500.2726   3.333121
#> 185 500.0603   2.763975
#> 186 622.4006   2.908183
#> 187 551.8690   2.583427
#> 188 500.2966   3.996641
#> 189 505.0692   2.734593
#> 190 519.7459   3.147426
#> 191 568.4404   2.552938
#> 192 503.4426   3.773213
#> 193 500.3215   2.783148
#> 194 500.8609   6.309707
#> 195 516.0347   2.938434
#> 196 501.1784   3.186622
#> 197 984.1126   2.792345
#> 198 500.5749   2.724186
#> 199 500.3402   3.044321
#> 200 500.0541   3.745612
#> 201 500.2665   2.627808
#> 202 507.9887   2.635454
#> 203 507.3394   3.721237
#> 204 507.7721   2.794832
#> 205 509.6562   2.593338
#> 206 627.8609   5.146476
#> 207 500.5689   3.273082
#> 208 500.1961   2.744294
#> 209 526.6987   2.679121
#> 210 500.0385   2.663857
#> 211 500.6320   2.545326
#> 212 502.1067   4.914793
#> 213 557.0660   2.547854
#> 214 509.8691   3.101516
#> 215 500.8038   2.508407
#> 216 502.6821   2.515228
#> 217 500.1018   2.670410
#> 218 502.3479   2.884366
#> 219 774.3551   3.067092
#> 220 502.5183   3.115678
#> 221 717.8409   3.862884
#> 222 502.1024   2.708476
#> 223 546.0300   3.128155
#> 224 503.6401   2.772881
#> 225 500.0242   3.658030
#> 226 500.1128   2.891261
#> 227 500.0562   4.210770
#> 228 504.9750   3.331905
#> 229 514.2147   2.639188
#> 230 500.0041   3.207472
#> 231 518.4061   2.590155
#> 232 500.0416   3.448566
#> 233 915.0503   2.549815
#> 234 505.9871   3.194109
#> 235 520.0885   2.835219
#> 236 500.8834   2.570985
#> 237 500.0004   2.546419
#> 238 592.0836   3.177650
#> 239 500.1310   2.575558
#> 240 500.1482   2.911644
#> 241 859.1172   2.790693
#> 242 517.4853   2.504354
#> 243 552.2874   2.740633
#> 244 500.0147   2.779060
#> 245 535.7907   2.848886
#> 246 506.7499   2.831011
#> 247 784.7031   2.886860
#> 248 500.6182   2.531261
#> 249 501.2430   2.647157
#> 250 500.0918   3.394261
#> 251 500.0070   3.582768
#> 252 502.1755   3.981197
#> 253 580.2602   2.923001
#> 254 506.5133   2.811298
#> 255 500.6097   2.671043
#> 256 724.7930   3.518165
#> 257 636.0785   2.544939
#> 258 500.6190   3.266508
#> 259 500.0060   2.516063
#> 260 615.1947   2.942659
#> 261 542.7807   2.634336
#> 262 598.3240   2.575005
#> 263 530.4434   2.959920
#> 264 500.5285   2.831165
#> 265 500.1220   2.851998
#> 266 500.8790   3.777527
#> 267 507.4819   2.980750
#> 268 501.0964   2.795202
#> 269 500.1424   2.990252
#> 270 500.1165   2.938850
#> 271 535.7016   3.090697
#> 272 505.6934   2.509821
#> 273 500.0261   2.634415
#> 274 525.9250   4.056259
#> 275 500.0861   2.510732
#> 276 502.0905   3.077139
#> 277 501.3154   2.650659
#> 278 500.5986   2.938875
#> 279 505.0258   2.504353
#> 280 501.2988   2.644545
#> 281 572.9141   3.132003
#> 282 559.7472   2.963985
#> 283 610.3750   2.586435
#> 284 500.4220   2.552560
#> 285 562.5543   2.535010
#> 286 501.3162   4.344764
#> 287 534.7789   2.571336
#> 288 526.4106   2.546979
#> 289 502.5579   2.502970
#> 290 558.2338   2.971170
#> 291 518.2067   3.701820
#> 292 511.5401   2.503376
#> 293 563.8504   3.977673
#> 294 501.6342   2.562510
#> 295 500.0560   2.900555
#> 296 512.9522   2.534273
#> 297 501.0912   2.725040
#> 298 504.4362   2.760631
#> 299 530.8775   2.901654
#> 300 501.9621   4.015020
#> 301 501.7050   2.837504
#> 302 500.0796   2.843713
#> 303 551.9126   3.984038
#> 304 526.4177   2.928499
#> 305 500.6634   2.753299
#> 306 504.2884   3.569638
#> 307 645.0484   2.603261
#> 308 591.2557   2.703188
#> 309 849.3108   2.516506
#> 310 500.0390   3.322885
#> 311 500.4454   2.827331
#> 312 507.1541   2.863124
#> 313 506.6872   3.068404
#> 314 544.1835   3.008142
#> 315 504.0913   2.921212
#> 316 500.6975   2.670503
#> 317 762.2385   3.414540
#> 318 544.6927   2.538695
#> 319 500.9327   3.649834
#> 320 700.8949   2.503691
#> 321 501.4573   2.605674
#> 322 562.9126   2.825976
#> 323 504.9869   2.529942
#> 324 505.4486   3.011963
#> 325 525.5299   3.006286
#> 326 521.3258   2.823248
#> 327 510.0412   2.928229
#> 328 833.0794   2.570190
#> 329 662.0950   2.704377
#> 330 502.6535   2.739649
#> 331 500.3274   3.397974
#> 332 868.7380   2.922655
#> 333 500.1693   5.402453
#> 334 500.1653   3.726305
#> 335 603.9254   2.519006
#> 336 511.0411   3.254275
#> 337 500.9063   4.843297
#> 338 588.3385   2.595149
#> 339 767.4435   3.488838
#> 340 500.0685   2.754822
#> 341 515.2498   3.566314
#> 342 500.3610   3.522419
#> 343 500.0594   2.559232
#> 344 504.0559   3.879988
#> 345 555.6034   2.675792
#> 346 501.2346   4.012047
#> 347 519.4013   2.519380
#> 348 871.9555   2.505179
#> 349 842.7665   3.414199
#> 350 905.5747   2.667274
#> 351 510.8814   2.518522
#> 352 668.5786   2.503748
#> 353 573.5929   2.783224
#> 354 501.0874   2.630439
#> 355 525.6402   3.189912
#> 356 501.7617   2.534003
#> 357 511.0229   3.049316
#> 358 600.6891   2.521532
#> 359 508.9643   3.161401
#> 360 548.1051   3.515490
#> 361 501.2539   2.831807
#> 362 501.5087   2.549532
#> 363 505.0828   2.864091
#> 364 500.5792   2.760317
#> 365 513.4636   2.838597
#> 366 501.3825   2.847063
#> 367 500.6121   2.546900
#> 368 500.2630   2.954288
#> 369 500.4689   2.508224
#> 370 502.4259   3.366092
#> 371 506.2184   2.845660
#> 372 627.7929   2.551326
#> 373 503.7972   3.164157
#> 374 502.2074   2.707323
#> 375 500.8240   3.118767
#> 376 520.4771   3.090416
#> 377 929.8624   3.707670
#> 378 501.2252   2.542445
#> 379 506.0749   2.510666
#> 380 500.7800   3.333865
#> 381 500.1426   2.501770
#> 382 503.7644   2.597899
#> 383 837.3543   2.576274
#> 384 501.2471   2.943697
#> 385 501.7384   2.823962
#> 386 501.5095   3.052576
#> 387 735.1132   2.536465
#> 388 505.5342   3.241950
#> 389 500.5227   2.718700
#> 390 519.6771   4.194339
#> 391 564.0657   2.711667
#> 392 500.3840   2.666585
#> 393 796.0755   3.445243
#> 394 501.8783   3.731887
#> 395 500.0144   2.567817
#> 396 507.3911   2.616120
#> 397 500.0313   3.162070
#> 398 576.1092   3.050179
#> 399 500.8067   2.559949
#> 400 841.9171   2.812893
#> 401 520.4488   2.520358
#> 402 500.1834   2.912626
#> 403 937.9924   2.530255
#> 404 506.1694   4.415887
#> 405 500.0621   2.837120
#> 406 500.4130   2.982569
#> 407 500.0917   3.369161
#> 408 509.1168   2.853920
#> 409 500.2649   2.567518
#> 410 507.0918   2.678608
#> 411 514.2432   3.013127
#> 412 505.1758   4.073420
#> 413 511.3330   2.856269
#> 414 507.9590   2.668392
#> 415 505.2535   2.684447
#> 416 520.4993   3.052536
#> 417 500.4058   2.644221
#> 418 501.1216   2.914189
#> 419 552.5529   3.138941
#> 420 581.7732   2.521878
#> 421 619.0382   2.704206
#> 422 500.1867   6.567916
#> 423 936.8846   2.876534
#> 424 623.5413   2.533122
#> 425 542.1378   2.629878
#> 426 517.5313   2.962802
#> 427 500.5564   3.087261
#> 428 533.1803   2.510220
#> 429 509.0187   3.341752
#> 430 711.4811   2.861154
#> 431 531.6870   3.812226
#> 432 526.7586   2.637356
#> 433 517.8669   2.981146
#> 434 500.0461   3.612148
#> 435 500.8367   2.679418
#> 436 500.0253   2.643159
#> 437 504.4018   2.832216
#> 438 503.6136   3.038115
#> 439 500.0188   3.059200
#> 440 567.8063   2.869553
#> 441 543.1513   3.066751
#> 442 621.2541   3.016008
#> 443 508.2585   2.720209
#> 444 501.5162   2.516458
#> 445 500.4186   6.080457
#> 446 505.5602   3.534326
#> 447 684.8182   2.674026
#> 448 500.0032   2.724529
#> 449 504.4080   2.862002
#> 450 501.7416   2.823930
#> 451 500.1882   2.651028
#> 452 503.1491   2.769908
#> 453 532.7245   2.574564
#> 454 500.3226   2.650528
#> 455 851.5137   3.427216
#> 456 568.6105   3.088015
#> 457 506.9392   3.462831
#> 458 500.6721   2.912950
#> 459 501.0135   3.669470
#> 460 500.1987   2.591758
#> 461 735.8300   3.488219
#> 462 503.1933   2.542045
#> 463 665.6788   2.598721
#> 464 504.2011   2.751930
#> 465 817.6255   3.347047
#> 466 514.8851   2.908892
#> 467 617.0264   2.699530
#> 468 501.3861   3.296831
#> 469 508.1639   2.534633
#> 470 504.7688   3.128890
#> 471 502.3338   3.039690
#> 472 500.2939   3.749169
#> 473 536.7385   2.641729
#> 474 503.4565   3.482727
#> 475 530.2116   2.622858
#> 476 500.4024   2.580641
#> 477 500.8774   3.067191
#> 478 501.7851   3.370460
#> 479 719.4165   2.585004
#> 480 559.2702   2.567856
#> 481 736.3586   3.281237
#> 482 558.1713   2.593258
#> 483 500.1866   2.518863
#> 484 660.5716   2.938448
#> 485 564.5660   2.998665
#> 486 500.0544   2.940061
#> 487 503.3635   3.938863
#> 488 500.6736   2.765212
#> 489 923.9408   2.972013
#> 490 500.9960   3.077447
#> 491 501.5541   2.543796
#> 492 506.9139   3.018870
#> 493 500.5972   2.579135
#> 494 510.2882   3.049473
#> 495 500.8475   2.616867
#> 496 500.0324   3.349598
#> 497 500.1309   2.723017
#> 498 503.5950   2.595199
#> 499 500.0808   2.817332
#> 500 500.2331   3.430851
#> 501 508.6539   2.822620
#> 502 500.3470   2.541100
#> 503 506.8111   2.505233
#> 504 500.3020   3.263342
#> 505 500.9478   2.565992
#> 506 502.5678   2.637110
#> 507 520.2813   3.410070
#> 508 500.0509   2.619372
#> 509 502.4664   3.037841
#> 510 555.7829   3.699865
#> 511 501.2281   2.943334
#> 512 512.5349   2.502814
#> 513 500.0357   2.578129
#> 514 606.6226   3.235254
#> 515 504.2303   3.046439
#> 516 502.5815   2.672040
#> 517 514.0871   2.501621
#> 518 522.9238   2.838032
#> 519 500.4593   2.764845
#> 520 521.9825   2.616833
#> 521 656.2319   2.814493
#> 522 513.7506   2.540920
#> 523 512.4873   3.670607
#> 524 510.8586   2.600046
#> 525 500.2665   2.547462
#> 526 500.4046   2.676501
#> 527 500.1468   2.585844
#> 528 501.1383   3.002246
#> 529 502.2684   2.952415
#> 530 519.5149   3.325714
#> 531 501.7552   2.634974
#> 532 524.5531   2.550852
#> 533 501.4316   2.698370
#> 534 513.5482   2.927644
#> 535 500.9734   2.571562
#> 536 503.9615   2.518453
#> 537 500.0909   2.509145
#> 538 637.5023   2.513052
#> 539 513.9980   3.132324
#> 540 502.0628   2.853226
#> 541 500.4295   2.729588
#> 542 500.0097   2.979909
#> 543 688.3201   2.543421
#> 544 718.6664   6.031543
#> 545 500.4630   3.633549
#> 546 831.6529   2.532380
#> 547 566.9505   2.567774
#> 548 807.5002   3.253161
#> 549 504.2628   2.590660
#> 550 500.0296   2.791653
#> 551 500.0692   2.906088
#> 552 542.7179   2.765730
#> 553 500.2114   2.595606
#> 554 500.5928   2.615124
#> 555 500.0678   2.814363
#> 556 500.0108   2.675055
#> 557 500.0502   2.622547
#> 558 502.1263   3.155701
#> 559 505.3515   2.835642
#> 560 891.2471   2.960968
#> 561 886.8599   2.570800
#> 562 514.7463   3.110221
#> 563 500.1579   2.663186
#> 564 504.6416   3.558201
#> 565 509.4879   2.581690
#> 566 504.3822   2.502835
#> 567 500.4171   2.623470
#> 568 500.2545   3.327977
#> 569 530.0717   2.679048
#> 570 500.9129   2.661905
#> 571 517.8792   2.837785
#> 572 557.7777   2.517005
#> 573 544.8872   2.543264
#> 574 501.7004   3.554274
#> 575 500.7720   2.783132
#> 576 537.7103   3.140502
#> 577 505.8512   3.058476
#> 578 527.6749   2.597142
#> 579 506.3756   2.596256
#> 580 501.3840   2.652281
#> 581 521.5323   2.802010
#> 582 532.0546   3.321532
#> 583 503.0531   2.577763
#> 584 501.8539   2.966210
#> 585 500.0005   3.004195
#> 586 504.6298   3.177199
#> 587 504.5791   3.141395
#> 588 501.7271   2.683689
#> 589 500.2340   2.997214
#> 590 500.0646   3.015301
#> 591 506.9186   3.159955
#> 592 525.4675   2.553510
#> 593 507.9874   3.562546
#> 594 500.2239   2.981167
#> 595 522.6888   3.039936
#> 596 500.5124   4.078995
#> 597 500.0098   2.573489
#> 598 562.0640   2.585890
#> 599 506.2279   3.594711
#> 600 502.2170   2.646919
#> 601 501.1186   3.109064
#> 602 537.8393   2.535064
#> 603 500.2541   3.801522
#> 604 509.6679   3.118038
#> 605 512.5574   3.025556
#> 606 502.7305   4.000668
#> 607 522.2034   3.398127
#> 608 577.2813   2.512455
#> 609 604.5889   2.546315
#> 610 508.5430   3.418629
#> 611 535.7396   2.943816
#> 612 501.2649   4.422327
#> 613 501.9292   2.790429
#> 614 530.4854   3.124290
#> 615 500.1984   2.657406
#> 616 508.4966   3.176138
#> 617 510.3108   3.192332
#> 618 503.8354   2.503606
#> 619 501.8287   2.939032
#> 620 504.9248   2.569648
#> 621 500.9127   2.684276
#> 622 528.9817   3.279911
#> 623 942.2625   2.736791
#> 624 508.2106   2.979796
#> 625 625.3358   2.537470
#> 626 500.4998   2.637192
#> 627 763.4539   2.783885
#> 628 502.0014   2.637017
#> 629 500.2732   2.608871
#> 630 503.8908   2.590174
#> 631 500.1107   3.115822
#> 632 500.3125   2.747147
#> 633 568.8031   2.608213
#> 634 501.8322   2.524832
#> 635 501.7606   2.875396
#> 636 516.6895   2.796333
#> 637 504.1584   2.720288
#> 638 501.2715   2.511116
#> 639 503.8725   3.266833
#> 640 500.3606   2.627505
#> 641 884.5268   2.659934
#> 642 500.0966   2.956364
#> 643 727.8368   2.619500
#> 644 500.1852   3.002157
#> 645 514.8738   6.181788
#> 646 573.0216   3.378369
#> 647 590.6564   4.132402
#> 648 603.9935   2.982692
#> 649 530.2960   3.244343
#> 650 500.2824   3.349322
#> 651 501.5956   2.851015
#> 652 503.2416   3.141614
#> 653 500.2616   3.095395
#> 654 502.7711   2.706280
#> 655 505.5667   2.972885
#> 656 558.1893   2.602661
#> 657 500.2312   3.515393
#> 658 500.2266   3.584490
#> 659 529.0156   3.058998
#> 660 511.6267   2.713915
#> 661 643.9634   3.112656
#> 662 500.2555   4.229898
#> 663 501.1912   4.205028
#> 664 500.0808   3.081675
#> 665 500.5515   2.533264
#> 666 500.0606   2.504632
#> 667 673.8096   3.369859
#> 668 500.8475   2.795140
#> 669 501.0379   2.626794
#> 670 522.9431   3.511633
#> 671 502.3102   2.600002
#> 672 735.0553   2.627251
#> 673 543.5581   2.503050
#> 674 501.7366   2.591906
#> 675 589.6083   2.830750
#> 676 943.8719   2.518455
#> 677 500.1305   2.943554
#> 678 506.2162   3.062083
#> 679 508.4075   3.468372
#> 680 501.0009   2.759794
#> 681 630.2185   3.216168
#> 682 508.8099   2.640442
#> 683 500.7690   2.939330
#> 684 500.5676   2.824603
#> 685 675.5234   3.034069
#> 686 609.9491   3.165870
#> 687 564.6562   3.034669
#> 688 500.0126   3.621232
#> 689 541.4755   2.510372
#> 690 500.2515   3.886391
#> 691 500.0893   3.392923
#> 692 500.5877   3.858964
#> 693 502.5011   3.019479
#> 694 500.7614   2.510218
#> 695 500.1887   2.622859
#> 696 501.2570   2.519124
#> 697 502.6669   4.029793
#> 698 559.0041   2.610768
#> 699 538.4041   3.179493
#> 700 515.4252   2.897769
#> 701 527.1010   2.862873
#> 702 602.6603   2.939349
#> 703 501.6352   2.650790
#> 704 851.7572   2.652033
#> 705 500.0525   2.796625
#> 706 500.2409   3.152855
#> 707 505.5376   3.487729
#> 708 523.5136   2.575616
#> 709 500.6120   3.545691
#> 710 908.4303   2.586984
#> 711 944.3075   2.970188
#> 712 500.0087   2.847110
#> 713 500.6235   2.576803
#> 714 500.9449   2.522619
#> 715 533.0715   3.124787
#> 716 500.0733   2.750008
#> 717 500.1502   2.681740
#> 718 500.2122   3.485361
#> 719 500.0049   3.042745
#> 720 512.7234   3.075846
#> 721 568.2682   3.218507
#> 722 500.9076   2.695539
#> 723 500.3015   3.370994
#> 724 530.7118   2.598357
#> 725 500.3174   3.096215
#> 726 500.0088   2.586244
#> 727 500.1195   2.883052
#> 728 515.9802   2.542642
#> 729 563.1117   3.559279
#> 730 681.6451   2.735827
#> 731 501.1591   2.908209
#> 732 502.3226   2.515583
#> 733 597.3607   4.178256
#> 734 500.3905   3.641675
#> 735 521.8531   2.804531
#> 736 500.8928   2.512952
#> 737 500.0822   2.603068
#> 738 542.6394   2.663512
#> 739 500.1853   2.587792
#> 740 500.2744   3.010822
#> 741 500.0932   3.071220
#> 742 500.1075   2.511619
#> 743 653.0462   2.517699
#> 744 538.3695   2.572742
#> 745 505.8921   2.779227
#> 746 500.1207   3.837121
#> 747 500.3867   2.610621
#> 748 500.3392   2.687382
#> 749 533.6706   3.709040
#> 750 500.4254   2.705199
#> 751 500.1337   2.938514
#> 752 507.9789   2.875617
#> 753 501.0401   3.293884
#> 754 500.4326   3.203171
#> 755 664.4230   3.074315
#> 756 785.2550   2.799064
#> 757 500.1249   3.053301
#> 758 511.2541   2.530479
#> 759 503.6174   2.760824
#> 760 519.8326   2.519816
#> 761 902.3956   3.028674
#> 762 500.1336   2.646230
#> 763 500.0044   2.734977
#> 764 500.0071   2.883902
#> 765 501.1900   2.888817
#> 766 559.3883   2.594110
#> 767 501.8973   3.005914
#> 768 582.6765   2.892610
#> 769 502.9060   2.668564
#> 770 579.0678   3.101429
#> 771 500.0503   2.733670
#> 772 500.9338   2.622502
#> 773 500.1331   3.497413
#> 774 500.0321   2.516343
#> 775 500.4113   2.667623
#> 776 531.7750   2.805750
#> 777 501.9169   2.940326
#> 778 622.8120   2.518058
#> 779 503.7623   2.634458
#> 780 806.7006   2.621900
#> 781 619.6460   2.694131
#> 782 518.8228   2.902334
#> 783 500.6437   3.508109
#> 784 500.7002   2.567248
#> 785 566.3024   2.962955
#> 786 500.0437   2.507989
#> 787 500.5880   2.583665
#> 788 563.9218   2.532960
#> 789 509.8167   2.525698
#> 790 500.0295   2.860561
#> 791 859.6004   2.673146
#> 792 502.0045   3.307468
#> 793 525.0168   2.562536
#> 794 500.0046   2.653517
#> 795 500.4281   2.621085
#> 796 500.1817   2.541668
#> 797 514.4448   3.910175
#> 798 524.2823   2.760671
#> 799 533.7100   2.688675
#> 800 500.2383   3.035743
#> 801 511.2457   2.588821
#> 802 502.7156   3.476457
#> 803 500.8702   3.217488
#> 804 501.4841   3.357712
#> 805 508.6266   3.057597
#> 806 500.9182   2.542182
#> 807 573.4840   3.870550
#> 808 509.3897   3.738717
#> 809 500.2545   2.543190
#> 810 500.6710   2.739861
#> 811 503.8149   2.719476
#> 812 500.0762   2.726384
#> 813 504.5118   2.666594
#> 814 505.1481   2.592855
#> 815 500.1726   2.583088
#> 816 500.3603   2.567499
#> 817 502.0797   2.520715
#> 818 504.3775   2.732631
#> 819 500.0447   2.668480
#> 820 500.1675   3.563807
#> 821 500.0304   2.795123
#> 822 500.2479   3.444830
#> 823 706.6569   2.537097
#> 824 500.4999   2.822810
#> 825 609.5797   2.652474
#> 826 519.6659   2.676644
#> 827 500.1300   2.685799
#> 828 500.3406   3.101560
#> 829 700.8099   2.501713
#> 830 500.6788   2.691822
#> 831 502.1650   2.783241
#> 832 501.0513   2.703141
#> 833 500.1367   2.743994
#> 834 629.2556   2.512359
#> 835 500.0446   2.500928
#> 836 573.7509   2.662542
#> 837 500.0167   2.545741
#> 838 500.6386   2.720563
#> 839 500.1747   3.650888
#> 840 543.1362   3.200938
#> 841 501.3240   3.531200
#> 842 529.3867   2.745266
#> 843 595.3826   2.546586
#> 844 505.0590   2.879652
#> 845 519.9601   2.944297
#> 846 500.3213   2.787528
#> 847 550.0125   3.146698
#> 848 531.2050   3.825571
#> 849 860.0219   2.845961
#> 850 650.4995   2.725588
#> 851 510.3159   2.749873
#> 852 606.0144   2.605564
#> 853 501.3982   2.538408
#> 854 500.0324   2.535638
#> 855 500.0109   2.964382
#> 856 501.7344   3.461762
#> 857 500.7894   2.662519
#> 858 503.1051   2.984994
#> 859 525.7066   2.687397
#> 860 565.1626   4.018408
#> 861 500.1113   2.726604
#> 862 500.0425   2.581173
#> 863 513.0568   3.436712
#> 864 744.7881   3.307873
#> 865 510.8620   2.989191
#> 866 501.7070   2.877179
#> 867 500.2231   2.553823
#> 868 500.2572   3.078607
#> 869 507.5109   2.825521
#> 870 500.0016   2.659905
#> 871 500.1084   3.657411
#> 872 500.4322   2.911355
#> 873 501.0728   2.544938
#> 874 500.1629   2.654414
#> 875 500.8935   2.965807
#> 876 502.1971   3.240025
#> 877 502.2084   2.925428
#> 878 500.5861   2.901950
#> 879 501.8447   2.892354
#> 880 505.1394   3.379520
#> 881 627.9756   2.514869
#> 882 775.3006   3.311113
#> 883 523.1132   3.285914
#> 884 558.4513   2.510608
#> 885 670.2464   2.924442
#> 886 500.5866   2.549537
#> 887 500.0142   4.015610
#> 888 503.0608   3.940500
#> 889 503.4099   2.995178
#> 890 506.2065   2.569945
#> 891 500.2059   2.819146
#> 892 503.9819   2.595621
#> 893 501.9338   2.576223
#> 894 500.0166   2.572121
#> 895 503.4150   3.557496
#> 896 501.9206   3.140357
#> 897 500.9077   3.830856
#> 898 501.2820   3.656968
#> 899 579.6220   3.059341
#> 900 502.9300   2.709812
#> 901 507.1670   3.109689
#> 902 502.9377   3.418077
#> 903 500.2936   3.105474