Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
sample.temporal.ETAS.generation(
  theta=list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08),
  beta.p=log(10),
  M0=2.5,
  T1=0, T2=1000,
  Ht=Ht )
#>           ts magnitudes
#> 1   500.5586   3.001231
#> 2   500.1992   3.480800
#> 3   955.4117   3.453811
#> 4   500.7163   3.204744
#> 5   519.5546   3.194679
#> 6   501.9707   2.634186
#> 7   500.4418   2.514343
#> 8   720.4328   2.971607
#> 9   544.1927   3.142603
#> 10  944.5610   2.538374
#> 11  590.3375   2.703166
#> 12  500.2040   2.793040
#> 13  513.2418   2.978561
#> 14  628.4203   2.649658
#> 15  501.4956   2.863926
#> 16  500.9548   2.632254
#> 17  500.6110   2.899178
#> 18  643.5661   2.580023
#> 19  500.7972   3.394372
#> 20  500.0211   3.120531
#> 21  508.8718   2.726104
#> 22  510.7510   2.986957
#> 23  758.0975   3.116883
#> 24  934.3092   3.198278
#> 25  502.5427   2.528252
#> 26  500.0261   3.569117
#> 27  545.1075   3.032329
#> 28  506.0260   2.812756
#> 29  500.2194   2.527309
#> 30  503.0969   2.761171
#> 31  600.1234   2.853627
#> 32  500.0294   3.379831
#> 33  503.8873   2.721403
#> 34  502.0808   2.908269
#> 35  618.0515   2.682972
#> 36  502.2384   2.738923
#> 37  522.1491   2.514752
#> 38  501.7956   3.130425
#> 39  500.8770   2.756432
#> 40  504.0125   2.915393
#> 41  564.4086   3.206839
#> 42  500.1314   2.726827
#> 43  511.3822   2.901074
#> 44  699.5049   2.800213
#> 45  500.6917   2.633993
#> 46  501.1426   2.577733
#> 47  500.2591   3.442786
#> 48  502.3235   2.523577
#> 49  506.0868   3.005797
#> 50  502.2819   2.591707
#> 51  500.6748   3.111180
#> 52  779.3509   2.934477
#> 53  500.5099   2.540719
#> 54  500.4429   2.596147
#> 55  505.2970   2.634819
#> 56  500.3501   2.614186
#> 57  677.1191   2.841234
#> 58  545.3989   2.726274
#> 59  563.6138   2.762866
#> 60  527.7411   2.725972
#> 61  502.3467   2.718915
#> 62  500.9128   3.170178
#> 63  684.2223   2.690330
#> 64  581.5556   2.548749
#> 65  500.0011   2.571356
#> 66  500.0146   3.200223
#> 67  501.8025   2.519082
#> 68  500.0716   2.894414
#> 69  504.1868   2.943168
#> 70  500.9480   2.933950
#> 71  500.2794   3.083499
#> 72  654.4857   2.801952
#> 73  502.5343   2.506641
#> 74  667.3268   2.793820
#> 75  500.1036   2.791373
#> 76  507.3320   3.129862
#> 77  540.2438   2.565063
#> 78  520.9810   2.638652
#> 79  502.5729   2.607335
#> 80  894.8036   3.960836
#> 81  500.2157   2.551187
#> 82  963.8526   2.841921
#> 83  511.1996   3.378449
#> 84  500.5348   2.914748
#> 85  716.8335   3.124303
#> 86  500.4349   3.631212
#> 87  500.5374   2.633012
#> 88  683.1431   3.016461
#> 89  500.3171   2.870624
#> 90  500.0211   2.735337
#> 91  535.5640   2.626882
#> 92  500.4255   2.893019
#> 93  500.2265   3.072342
#> 94  764.8819   2.557815
#> 95  912.5365   2.735716
#> 96  500.0498   2.966663
#> 97  569.8689   2.898008
#> 98  500.1968   2.683085
#> 99  500.0077   2.688252
#> 100 500.3511   3.492809
#> 101 502.4254   2.920580
#> 102 529.5937   2.794230
#> 103 500.2423   2.586632
#> 104 515.5294   3.770754
#> 105 513.5167   2.688888
#> 106 500.2533   4.305442
#> 107 501.1768   4.172858
#> 108 500.1260   2.653316
#> 109 531.3525   2.520693
#> 110 638.0394   2.647143
#> 111 508.6852   2.993472
#> 112 500.6937   2.595467
#> 113 500.7340   3.459839
#> 114 679.7313   2.823858
#> 115 500.0097   2.957006
#> 116 603.1156   4.162633
#> 117 500.7715   3.295413
#> 118 534.5585   3.443160
#> 119 503.0039   4.485574
#> 120 692.2174   2.592905
#> 121 500.2138   2.631371
#> 122 500.5870   2.520487
#> 123 553.6766   2.594237
#> 124 557.3465   3.359286
#> 125 500.4714   2.624490
#> 126 541.8804   2.788107
#> 127 502.1783   2.736996
#> 128 500.9714   2.646114
#> 129 505.2003   3.315404
#> 130 587.6285   2.712436
#> 131 930.2516   3.925804
#> 132 500.0453   2.506199
#> 133 500.8804   2.909604
#> 134 549.6452   2.534638
#> 135 504.1744   2.938410
#> 136 500.0073   2.692647
#> 137 500.0451   2.821425
#> 138 501.3575   3.049309
#> 139 502.5832   2.625403
#> 140 500.4001   2.567532
#> 141 505.6345   3.085963
#> 142 500.2610   2.592600
#> 143 500.1468   2.501296
#> 144 500.5155   3.076903
#> 145 507.4514   2.963993
#> 146 500.2269   2.514784
#> 147 916.2778   2.764620
#> 148 507.6534   3.774677
#> 149 500.8736   2.796942
#> 150 509.1283   2.548040
#> 151 503.8758   3.387579
#> 152 510.7250   3.018981
#> 153 500.0460   3.207797
#> 154 500.2148   3.727545
#> 155 502.9873   2.715930
#> 156 533.1800   2.703078
#> 157 782.3109   2.873653
#> 158 501.5734   2.690951
#> 159 660.0860   2.752748
#> 160 501.3894   3.060558
#> 161 503.5954   2.971804
#> 162 500.2200   2.950768
#> 163 818.8070   2.539444
#> 164 500.0377   2.990830
#> 165 542.6001   3.136690
#> 166 533.9806   3.679419
#> 167 502.5265   3.594604
#> 168 802.5938   2.770601
#> 169 556.6959   2.922673
#> 170 504.3048   2.549692
#> 171 500.2345   3.516904
#> 172 500.3397   2.517241
#> 173 508.6719   2.655833
#> 174 500.7265   3.226989
#> 175 501.6510   2.561219
#> 176 500.6483   3.149348
#> 177 638.7700   3.527956
#> 178 503.1059   2.721768
#> 179 500.3838   4.144887
#> 180 500.1169   2.829179
#> 181 656.2778   2.613129
#> 182 500.1377   3.030431
#> 183 513.6471   2.887521
#> 184 758.4754   2.858620
#> 185 685.2571   3.042323
#> 186 500.4823   3.033583
#> 187 683.1157   3.192073
#> 188 500.0050   2.500854
#> 189 500.1875   2.571964
#> 190 570.6028   2.773450
#> 191 500.1121   3.176226
#> 192 514.8165   2.543764
#> 193 500.3084   2.760610
#> 194 500.6238   2.856715
#> 195 500.2967   2.534861
#> 196 500.1697   3.096699
#> 197 500.1039   2.511880
#> 198 502.2196   2.580683
#> 199 509.3940   2.524791
#> 200 500.2518   2.794940
#> 201 511.3164   2.570480
#> 202 555.5120   2.783480
#> 203 500.9467   2.866416
#> 204 500.0648   3.502067
#> 205 501.4889   2.663613
#> 206 500.0311   3.486906
#> 207 500.9653   2.588565
#> 208 501.6766   3.012861
#> 209 653.4788   2.525453
#> 210 533.4613   3.974357
#> 211 506.2940   3.721386
#> 212 505.6364   3.532467
#> 213 517.2482   2.643009
#> 214 558.6411   2.935193
#> 215 501.6068   2.697936
#> 216 596.5665   2.766347
#> 217 513.2061   2.814736
#> 218 503.3501   2.801082
#> 219 500.1962   3.348742
#> 220 551.4992   2.509712
#> 221 502.7998   2.541645
#> 222 500.0787   3.024995
#> 223 537.0597   2.825690
#> 224 658.1801   3.129207
#> 225 671.0450   2.974950
#> 226 522.6830   3.129446
#> 227 874.5952   2.556119
#> 228 514.1934   2.811027
#> 229 500.0247   3.500931
#> 230 503.1715   2.905881
#> 231 537.2665   2.642687
#> 232 596.6109   3.017312
#> 233 500.2111   3.015104
#> 234 524.4080   4.202639
#> 235 507.7913   2.797741
#> 236 830.6405   2.908542
#> 237 536.8716   2.583764
#> 238 601.4420   2.799034
#> 239 500.1082   3.060938
#> 240 500.2796   2.589730
#> 241 531.7068   2.672000
#> 242 501.2786   2.606435
#> 243 505.4550   2.568085
#> 244 500.2468   2.788547
#> 245 500.3909   2.968105
#> 246 505.1286   2.563447
#> 247 885.1415   3.101228
#> 248 515.4417   2.612345
#> 249 500.8534   2.979958
#> 250 500.1663   3.450541
#> 251 500.0237   3.133840
#> 252 509.4239   3.176526
#> 253 507.1579   2.632669
#> 254 503.3231   2.706311
#> 255 836.5710   2.932749
#> 256 534.2844   3.483616
#> 257 543.1365   3.185115
#> 258 581.2814   2.514771
#> 259 522.9733   4.847371
#> 260 737.4043   2.867293
#> 261 500.2905   2.541933
#> 262 502.5945   3.162186
#> 263 501.3941   3.275868
#> 264 500.0797   2.596970
#> 265 500.1317   2.611741
#> 266 549.9625   2.628480
#> 267 500.8356   2.831178
#> 268 730.4018   3.856934
#> 269 512.5544   2.685060
#> 270 508.2537   3.246056
#> 271 500.4552   2.585911
#> 272 500.2679   2.800563
#> 273 616.6022   2.578164
#> 274 531.3217   2.664206
#> 275 665.8755   2.516530
#> 276 500.1463   2.931204
#> 277 500.0150   2.844698
#> 278 500.3007   3.948863
#> 279 567.4116   2.719873
#> 280 568.1248   2.697146
#> 281 505.4838   2.517756
#> 282 505.8967   3.436410
#> 283 628.3907   3.032934
#> 284 507.0345   2.712439
#> 285 504.4142   2.672918
#> 286 500.1608   2.584605
#> 287 503.5814   2.647776
#> 288 500.0140   2.510873
#> 289 536.6794   2.544030
#> 290 500.1816   2.578813
#> 291 500.0381   2.503826
#> 292 536.5776   3.777568
#> 293 500.2646   2.802415
#> 294 501.3795   2.504822
#> 295 500.5335   2.631685
#> 296 517.1560   2.592209
#> 297 546.4849   2.517674
#> 298 500.1176   3.399481
#> 299 503.5706   2.898572
#> 300 501.4948   2.643576
#> 301 501.8284   2.685533
#> 302 537.6325   3.597169
#> 303 501.1256   2.512689
#> 304 500.1300   2.812818
#> 305 501.8755   2.538395
#> 306 500.3933   2.662650
#> 307 500.1746   2.601912
#> 308 500.9067   3.472259
#> 309 500.5719   2.999274
#> 310 547.4121   2.928044
#> 311 500.7053   5.289432
#> 312 540.5136   3.323557
#> 313 506.1427   2.634879
#> 314 500.4109   3.208176
#> 315 500.8204   3.755266
#> 316 997.0292   2.879522
#> 317 500.9010   2.583021
#> 318 517.2799   3.027704
#> 319 501.1130   2.907911
#> 320 500.1000   2.546391
#> 321 505.7065   2.537427
#> 322 580.0420   2.647231
#> 323 887.0511   2.559371
#> 324 500.5850   2.639823
#> 325 500.0749   3.115867
#> 326 501.1641   2.752141
#> 327 500.3697   2.932457
#> 328 500.0878   3.655960
#> 329 506.4331   2.740792
#> 330 735.0234   2.625600
#> 331 712.3921   3.334719
#> 332 694.3948   2.959749
#> 333 510.9490   2.718425
#> 334 500.4435   2.642947
#> 335 500.5455   3.192635
#> 336 564.5889   3.037483
#> 337 500.0386   2.631810
#> 338 500.0669   2.736013
#> 339 504.9496   3.538175
#> 340 504.1953   2.663437
#> 341 632.1386   2.755585
#> 342 640.9954   3.111201
#> 343 500.1468   2.569581
#> 344 500.7013   2.737105
#> 345 500.0717   3.299834
#> 346 540.7647   2.707192
#> 347 508.3338   3.581736
#> 348 512.0290   3.232581
#> 349 504.5316   4.509033
#> 350 500.2221   2.588499
#> 351 509.5672   2.891773
#> 352 505.4649   2.582676
#> 353 515.2280   2.568932
#> 354 500.0400   2.544425
#> 355 502.1969   2.592406
#> 356 511.7136   3.009480
#> 357 502.3188   3.782688
#> 358 501.5402   3.647238
#> 359 500.0725   3.542266
#> 360 516.2355   2.571422
#> 361 500.4336   3.479499
#> 362 827.5836   2.782502
#> 363 507.9546   4.408350
#> 364 589.7948   2.500595
#> 365 500.1141   2.720217
#> 366 503.9513   2.913571
#> 367 500.0322   2.975664
#> 368 603.5432   3.855550
#> 369 846.1728   2.513310
#> 370 530.7566   2.643182
#> 371 500.6840   2.558899
#> 372 500.9270   2.868639
#> 373 502.4674   2.536330
#> 374 650.3838   2.970783
#> 375 510.7800   3.845483
#> 376 508.4222   3.002852
#> 377 527.4383   3.041192
#> 378 619.6607   3.334200
#> 379 895.8812   2.586331
#> 380 505.9313   2.789666
#> 381 929.4842   3.224423
#> 382 564.6094   2.599810
#> 383 894.7751   2.610926
#> 384 669.5524   3.226435
#> 385 500.1419   2.524937
#> 386 507.4757   2.909127
#> 387 668.1961   3.003077
#> 388 657.8468   3.427629
#> 389 506.2658   2.581839
#> 390 500.0542   2.662418
#> 391 767.5028   2.539011
#> 392 500.4069   2.859922
#> 393 566.0067   2.676938
#> 394 500.0487   3.500088
#> 395 500.3592   3.202936
#> 396 500.0440   3.590310
#> 397 558.3747   2.861267
#> 398 500.0197   3.106943
#> 399 512.6446   2.638584
#> 400 500.9661   2.515223
#> 401 736.8689   3.087576
#> 402 848.0337   2.895745
#> 403 639.0784   3.003080
#> 404 500.6761   2.623680
#> 405 500.0186   3.572010
#> 406 500.2458   2.751556
#> 407 506.4878   2.613507
#> 408 500.4003   2.848526
#> 409 500.0648   2.683170
#> 410 507.4526   3.280075
#> 411 513.2154   3.074852
#> 412 504.9803   3.506754
#> 413 660.2099   2.691880
#> 414 577.7365   2.997169
#> 415 503.5713   2.659309
#> 416 500.2623   2.668610
#> 417 574.7975   2.770537
#> 418 500.3523   2.927920
#> 419 501.2634   2.876309
#> 420 500.0125   3.224972
#> 421 536.2500   3.405742
#> 422 500.6073   2.860254
#> 423 500.3233   3.383800
#> 424 610.8299   2.854210
#> 425 500.0319   2.801299
#> 426 618.6612   3.728205
#> 427 679.8158   2.649823
#> 428 500.0145   3.557595
#> 429 500.2218   2.563152
#> 430 548.3372   3.793942
#> 431 500.8454   3.548737
#> 432 563.9249   2.709236
#> 433 531.2327   3.551518
#> 434 500.0218   3.094346
#> 435 500.2860   4.008719
#> 436 502.0651   3.236726
#> 437 500.2290   3.191913
#> 438 524.3632   2.761776
#> 439 522.9428   2.868718
#> 440 508.3567   2.589148
#> 441 502.9560   2.842565
#> 442 500.2204   2.789835
#> 443 526.2428   2.545380
#> 444 515.6579   2.809387
#> 445 510.8008   2.717594
#> 446 509.2109   2.512726
#> 447 500.0266   2.664852
#> 448 501.7718   2.550111
#> 449 542.4961   4.099910
#> 450 501.7243   2.782331
#> 451 781.6809   3.734705
#> 452 539.9992   4.150833
#> 453 500.0869   2.729425
#> 454 503.2504   2.545945
#> 455 500.8456   3.083301
#> 456 583.1115   3.412903
#> 457 504.1531   2.540748
#> 458 673.9301   2.642675
#> 459 500.0647   2.689065
#> 460 900.0141   2.844872
#> 461 504.6636   2.758279
#> 462 518.7944   2.761114
#> 463 501.7671   2.833294
#> 464 501.6127   2.674183
#> 465 500.3332   3.352946
#> 466 502.7814   2.858122
#> 467 557.1065   3.532418
#> 468 702.7037   2.579685
#> 469 500.1746   3.076102
#> 470 503.5582   2.844056
#> 471 500.0914   2.671137
#> 472 500.1373   2.591150
#> 473 522.3704   2.549573
#> 474 500.0893   2.822071
#> 475 683.5583   2.747306
#> 476 597.9595   2.805094
#> 477 500.1613   3.276786
#> 478 500.0103   3.004781
#> 479 502.6909   2.552893
#> 480 500.2893   3.028465
#> 481 508.9770   4.304073
#> 482 506.9640   2.663618
#> 483 505.6848   3.723574
#> 484 581.7280   2.571896
#> 485 500.3962   4.124884
#> 486 503.1336   2.703090
#> 487 500.0442   2.681376
#> 488 501.5368   2.509570
#> 489 500.1482   2.720506
#> 490 500.0409   2.574402
#> 491 503.4126   2.617462
#> 492 500.0832   2.754451
#> 493 967.8362   4.200159
#> 494 500.1095   2.650441
#> 495 500.5222   2.595431
#> 496 616.2590   3.209891
#> 497 500.5651   2.555863
#> 498 514.5170   2.710801
#> 499 500.4673   2.637790
#> 500 509.0010   4.483921
#> 501 510.4115   2.803959
#> 502 500.1036   2.931217
#> 503 500.2996   2.972247
#> 504 501.0548   3.583084
#> 505 501.2230   2.735908
#> 506 588.1198   2.566702
#> 507 539.4065   2.626859
#> 508 500.6603   3.406458
#> 509 501.1243   2.686188
#> 510 500.4913   2.603661
#> 511 575.3338   3.690144
#> 512 500.3131   3.108650
#> 513 513.3922   2.559053
#> 514 500.3689   2.701785
#> 515 508.1874   2.527874
#> 516 500.1292   3.012628
#> 517 500.5241   2.555850
#> 518 510.3620   2.529249
#> 519 500.0256   2.667181
#> 520 665.2348   2.872972
#> 521 500.2496   2.674881
#> 522 500.3423   2.882504
#> 523 500.1098   2.594857
#> 524 677.8664   2.715848
#> 525 560.8138   2.817409
#> 526 500.2739   2.584388
#> 527 547.7795   2.726506
#> 528 500.5080   2.577853
#> 529 502.4667   3.150031
#> 530 500.0705   2.517418
#> 531 500.3272   2.563349
#> 532 542.4953   2.749228
#> 533 500.3950   2.920443
#> 534 516.5504   2.595922
#> 535 507.6557   2.841994
#> 536 507.4106   3.367731
#> 537 505.9648   3.260063
#> 538 500.0017   3.167010
#> 539 588.9007   2.508927
#> 540 548.7336   3.131490
#> 541 732.6444   2.596129
#> 542 682.0626   2.904831
#> 543 538.6452   4.309184
#> 544 511.8465   2.532735
#> 545 552.8811   3.474766
#> 546 502.0802   2.546703
#> 547 585.4981   2.551431
#> 548 687.3169   2.547501
#> 549 500.2913   2.593249
#> 550 525.1872   2.811515
#> 551 928.5592   2.639409
#> 552 507.7589   2.745282
#> 553 501.7543   2.837414
#> 554 708.3523   2.516397
#> 555 517.9435   2.510010
#> 556 504.8135   3.984622
#> 557 500.4304   2.512973
#> 558 594.5210   4.087089
#> 559 500.0380   2.759928
#> 560 512.8787   2.578502
#> 561 500.9423   2.548897
#> 562 936.7480   2.740890
#> 563 500.0993   2.842142
#> 564 503.5278   2.623740
#> 565 520.4630   2.523329
#> 566 725.7992   2.574257
#> 567 500.1838   2.609384
#> 568 730.7843   2.662242
#> 569 517.8553   2.829032
#> 570 502.0603   2.976295
#> 571 503.0615   2.782440
#> 572 500.5592   2.593972
#> 573 500.0203   3.299979
#> 574 500.6283   3.019051
#> 575 755.3378   3.035682
#> 576 500.7562   3.264518
#> 577 551.1251   3.029409
#> 578 502.5187   2.817558
#> 579 598.9952   2.793027
#> 580 604.7856   2.646904
#> 581 500.0614   2.851911
#> 582 502.3764   2.945969
#> 583 533.6907   2.947054
#> 584 534.7615   2.685201
#> 585 606.4187   2.689674
#> 586 600.3038   2.976008
#> 587 500.1621   2.529151
#> 588 509.9386   2.588177
#> 589 500.3158   2.559780
#> 590 608.8761   3.039967
#> 591 544.4868   2.753156
#> 592 500.5891   2.660570
#> 593 501.5520   2.764600
#> 594 500.2813   3.225289
#> 595 504.9134   3.584773
#> 596 811.2507   2.701381
#> 597 500.2875   2.516780
#> 598 500.0045   3.022517
#> 599 500.3335   2.736618
#> 600 854.3620   3.256060
#> 601 503.1428   2.636127
#> 602 500.4037   2.945779
#> 603 502.8965   2.658689
#> 604 506.5770   2.579483
#> 605 503.7784   2.552262
#> 606 508.5327   2.682093
#> 607 500.9355   2.745928
#> 608 701.7354   2.748849
#> 609 578.9566   2.713899
#> 610 504.5487   2.612028
#> 611 504.2985   2.914949
#> 612 676.1249   2.749423
#> 613 500.0023   2.532770
#> 614 502.7519   2.852245
#> 615 536.1837   3.809507
#> 616 500.4332   3.289372
#> 617 500.0844   2.861361
#> 618 538.4705   2.652174
#> 619 500.1231   2.501324
#> 620 532.6883   3.129029
#> 621 509.6537   3.085962
#> 622 500.4234   2.665828
#> 623 513.7655   2.617952
#> 624 522.0529   2.615316
#> 625 522.3541   2.541515
#> 626 529.8889   2.620905
#> 627 507.4003   2.777902
#> 628 505.6265   2.568561
#> 629 500.2051   2.560539
#> 630 948.1767   3.341752
#> 631 655.4163   3.332014
#> 632 500.1592   2.586059
#> 633 500.0902   2.889873
#> 634 500.0849   2.646122
#> 635 500.0733   2.983739
#> 636 519.8961   2.930323
#> 637 503.3038   2.554833
#> 638 501.0124   3.238678
#> 639 501.0881   2.644186
#> 640 500.0947   2.539547
#> 641 514.4517   2.709636
#> 642 504.3759   2.637294
#> 643 500.1054   2.525484
#> 644 500.4869   2.599829
#> 645 500.3236   4.715969
#> 646 500.0134   3.876895
#> 647 797.3108   2.971225
#> 648 500.2258   2.821729
#> 649 557.3603   3.060179
#> 650 595.5395   2.831943
#> 651 500.1195   3.190694
#> 652 501.1620   3.024140
#> 653 500.6249   2.518320
#> 654 500.0606   2.779183
#> 655 501.5294   3.088766
#> 656 500.6359   2.653433
#> 657 500.6500   2.885778
#> 658 503.4215   2.793328
#> 659 535.7799   2.617085
#> 660 502.2263   2.514002
#> 661 500.4221   2.931011
#> 662 500.0033   3.394720
#> 663 513.1725   2.618430
#> 664 500.0517   3.655016
#> 665 644.5861   2.835906
#> 666 501.1009   3.391143
#> 667 508.7248   3.362966
#> 668 506.1063   4.570281
#> 669 517.5082   2.840080
#> 670 983.3587   2.503202
#> 671 515.8571   2.580963
#> 672 524.2566   2.565626
#> 673 500.0478   3.006367
#> 674 500.2968   2.626362
#> 675 500.1932   2.563627
#> 676 501.0023   2.656414
#> 677 500.5581   2.850755
#> 678 500.3651   3.047482
#> 679 502.7867   2.518081
#> 680 501.3912   3.155655
#> 681 723.0847   2.517662
#> 682 501.6426   2.625905
#> 683 500.7089   2.584675
#> 684 500.2941   3.226942
#> 685 859.9774   3.168826
#> 686 500.4062   3.224451
#> 687 500.0955   3.427293
#> 688 501.1536   3.253922
#> 689 500.0613   2.616624
#> 690 503.3091   3.004623
#> 691 503.4560   3.201029
#> 692 635.3270   2.528105
#> 693 501.4810   2.506899
#> 694 551.6381   3.103625
#> 695 510.6797   2.552499
#> 696 541.6471   2.754268
#> 697 500.1703   4.268798
#> 698 512.3099   3.140811
#> 699 500.8770   3.253185
#> 700 500.0172   2.569870
#> 701 518.4896   2.623051
#> 702 500.0859   3.307434
#> 703 649.3912   2.811626
#> 704 501.5808   2.614658
#> 705 510.9495   2.786513
#> 706 534.6057   2.605670
#> 707 961.3705   3.334931
#> 708 501.2382   2.524762
#> 709 971.5713   3.854951
#> 710 500.0819   2.516318
#> 711 529.3329   2.529723
#> 712 537.3759   3.228060
#> 713 826.5017   2.620158
#> 714 539.6600   3.133868
#> 715 500.4821   3.170235
#> 716 530.9650   2.752228
#> 717 500.2335   2.853591
#> 718 767.2547   2.537560
#> 719 500.5051   2.607903
#> 720 570.3038   2.809874
#> 721 500.0158   3.822045
#> 722 617.0161   2.988592
#> 723 506.5518   2.511456
#> 724 529.2432   2.762743
#> 725 570.8881   3.481960
#> 726 500.0846   2.563742
#> 727 500.0737   2.506722
#> 728 501.2096   2.569872
#> 729 578.4282   4.140875
#> 730 532.1776   2.742146
#> 731 659.6460   3.529939
#> 732 500.5183   2.719607
#> 733 590.4681   2.501243
#> 734 504.9240   3.262213
#> 735 500.0132   3.492925
#> 736 516.5090   3.001435
#> 737 521.8372   2.625160
#> 738 500.1813   2.500175
#> 739 510.1661   4.058874
#> 740 527.1048   2.759947
#> 741 500.1094   2.648259
#> 742 539.1512   2.833495
#> 743 500.3767   2.851956
#> 744 504.7477   2.722977
#> 745 500.3891   3.921773
#> 746 500.1130   3.086519
#> 747 500.0580   2.523380
#> 748 500.3097   3.139970
#> 749 584.7522   3.329177
#> 750 500.1524   2.679668
#> 751 500.2245   2.981796
#> 752 719.2785   2.512733
#> 753 500.1757   2.612079
#> 754 861.5483   3.169467
#> 755 950.6643   2.557549
#> 756 502.5483   2.817478
#> 757 518.7445   2.532401
#> 758 500.4617   2.870152
#> 759 539.9116   2.784393
#> 760 500.0550   3.011270
#> 761 706.0926   2.506135
#> 762 520.1348   3.166871
#> 763 500.1054   2.762513
#> 764 578.5459   2.817479
#> 765 500.1397   3.042802
#> 766 500.6403   2.955920
#> 767 500.1834   3.048657
#> 768 520.2887   3.730823
#> 769 500.0834   2.871292
#> 770 519.9486   2.501250
#> 771 501.0842   3.689936
#> 772 898.5052   2.565760
#> 773 502.0454   2.748210
#> 774 500.1516   2.702456
#> 775 506.4951   4.133941
#> 776 503.3511   3.773696
#> 777 500.0527   2.759829
#> 778 500.5299   2.830889
#> 779 501.0389   3.537495
#> 780 505.5360   2.703796
#> 781 500.1336   3.179898
#> 782 514.3460   2.740293
#> 783 510.4944   2.943113
#> 784 642.4558   4.909553
#> 785 982.6683   3.138967
#> 786 652.7567   3.244369
#> 787 500.1796   2.917072
#> 788 518.8066   3.246552
#> 789 503.2071   2.608875
#> 790 733.9686   2.583661
#> 791 500.4538   2.818858
#> 792 500.0027   2.787820
#> 793 501.0568   2.565404
#> 794 502.7974   2.622485
#> 795 506.7798   2.567056
#> 796 828.7319   2.542857
#> 797 613.0341   2.677880
#> 798 648.9665   2.986092
#> 799 584.6967   2.900665
#> 800 622.7222   2.661278
#> 801 516.0865   2.811819
#> 802 569.0633   2.605502
#> 803 503.6154   2.542356
#> 804 507.3677   2.531214
#> 805 500.0556   3.288187
#> 806 500.8250   2.741422
#> 807 505.4320   2.522254
#> 808 538.5474   3.198353
#> 809 505.5702   2.826506
#> 810 568.1877   2.585654
#> 811 516.3718   2.639915
#> 812 948.9244   2.589504
#> 813 558.8062   4.708849
#> 814 500.0452   2.539777
#> 815 557.4656   2.739716
#> 816 500.3440   2.649462
#> 817 524.0411   3.637446
#> 818 743.4795   3.008027
#> 819 698.2264   3.757303
#> 820 692.0891   3.479964
#> 821 521.0361   2.628355
#> 822 500.3693   3.682971
#> 823 555.3616   3.233789
#> 824 509.0012   3.021977
#> 825 511.0694   2.764571
#> 826 500.1394   2.674236
#> 827 536.6023   3.012272
#> 828 794.0399   2.579009
#> 829 650.2593   3.190394
#> 830 500.0341   2.527517
#> 831 770.9219   3.331469
#> 832 691.3788   3.751587
#> 833 547.0977   3.193175
#> 834 500.2291   2.771565
#> 835 573.3800   3.757822
#> 836 500.1595   2.958518
#> 837 500.3223   3.004297
#> 838 500.1929   2.506217
#> 839 510.8763   2.784445
#> 840 500.0424   3.444371
#> 841 500.0854   3.136216
#> 842 500.6473   4.576541
#> 843 743.8845   2.643953
#> 844 500.4016   2.585271
#> 845 513.1252   2.696421
#> 846 511.6224   2.755288
#> 847 514.0164   3.301297
#> 848 541.0041   2.847632
#> 849 500.0035   2.914175
#> 850 611.6203   2.585295
#> 851 500.3504   2.635890
#> 852 853.2825   2.620926
#> 853 500.1322   2.615570
#> 854 500.5475   2.559056
#> 855 512.4685   2.579529
#> 856 500.1005   2.933917
#> 857 650.1727   2.655207
#> 858 500.2718   2.864712
#> 859 675.3240   3.033281
#> 860 687.2753   2.952012
#> 861 540.3098   2.602894
#> 862 501.4368   2.632133
#> 863 507.0681   2.538839
#> 864 504.4121   3.999406
#> 865 513.5334   2.703098
#> 866 508.4078   3.403261
#> 867 508.0131   2.834809
#> 868 621.2961   2.894997
#> 869 655.3462   2.708945
#> 870 522.1853   3.070802
#> 871 500.8316   3.690643
#> 872 508.3231   2.500776
#> 873 500.1656   3.486396
#> 874 500.1574   2.831682
#> 875 500.0000   2.717982
#> 876 891.2983   4.142416
#> 877 517.7304   3.087765
#> 878 509.9923   3.351021
#> 879 507.8002   3.019938
#> 880 656.8589   2.796289
#> 881 500.3910   3.271361
#> 882 535.7431   2.696275
#> 883 500.3732   2.623136
#> 884 680.4145   3.288775
#> 885 513.1549   2.591916
#> 886 508.9912   2.904256
#> 887 737.3403   2.839924
#> 888 500.1399   3.005158
#> 889 500.9654   2.601709
#> 890 502.1802   2.604057
#> 891 500.1287   2.602406
#> 892 503.3711   2.520618
#> 893 500.1063   3.448052
#> 894 500.2055   2.770528
#> 895 516.1637   3.075192
#> 896 500.0456   2.747163
#> 897 518.6676   2.535370
#> 898 500.8077   2.663428
#> 899 502.3314   2.589849
#> 900 515.4893   2.508770
#> 901 507.0267   2.676873
#> 902 509.8058   2.511961
#> 903 623.4499   3.142928
#> 904 504.8865   2.728776
#> 905 533.8027   2.596880
#> 906 500.0848   2.568860
#> 907 517.2968   2.771494
#> 908 500.4789   2.639058
#> 909 501.9190   2.714017
#> 910 500.0769   3.763671
#> 911 500.1499   2.556783
#> 912 502.1583   2.572480
#> 913 500.0277   2.934058
#> 914 507.8685   3.610781
#> 915 500.4680   2.773003
#> 916 597.0498   2.687101
#> 917 540.2350   2.593575
#> 918 501.2552   2.602595
#> 919 554.0402   2.979726
#> 920 503.3428   2.543025