Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample_GR_magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample_GR_magnitudes(n = 100, beta.p = log(10), M0 = 2.5)
#>   [1] 2.565276 2.655549 2.569139 2.501348 2.621134 2.863835 3.951483 2.662309
#>   [9] 3.151434 2.852220 3.695151 3.367160 2.530887 2.875541 2.861103 5.086882
#>  [17] 2.526571 2.684223 2.879123 4.286285 2.627990 2.916647 2.628079 2.546219
#>  [25] 2.714801 2.701172 2.602952 3.073764 2.942081 3.022327 2.570992 3.044000
#>  [33] 3.293238 2.648600 2.890063 2.591994 2.674703 2.680624 2.586984 3.192044
#>  [41] 2.803385 2.727180 2.878512 3.014003 3.449880 2.755975 2.628730 3.061200
#>  [49] 2.517700 3.192936 2.666113 3.253941 3.668454 2.774410 2.656357 3.720594
#>  [57] 2.529600 2.612591 2.614557 2.904233 3.037120 2.781924 2.615917 3.278323
#>  [65] 3.460555 2.589337 2.799001 2.552430 2.591576 2.894900 3.288935 2.979786
#>  [73] 2.514374 2.507744 3.301460 3.292746 3.135457 2.538304 2.772561 2.598416
#>  [81] 2.568976 2.634589 2.658002 2.836325 3.198680 3.270320 3.104304 2.568132
#>  [89] 2.942378 2.732049 2.987930 2.561954 2.546024 3.150199 2.505936 2.822833
#>  [97] 2.630371 3.094246 2.802842 2.573921