Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample_GR_magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample_GR_magnitudes(n = 100, beta.p = log(10), M0 = 2.5)
#>   [1] 2.563625 3.044290 2.770117 2.848921 2.833841 2.527470 2.576935 3.216981
#>   [9] 3.307933 2.975841 2.913140 2.568385 3.792889 2.709000 2.818547 2.716959
#>  [17] 2.779336 2.559271 2.587460 2.523337 3.356719 4.549608 2.611074 2.708790
#>  [25] 3.256245 2.504656 2.561985 3.873829 2.765855 3.042014 2.529586 3.282631
#>  [33] 2.887395 2.750460 2.647841 2.503358 2.703423 2.763432 3.656803 2.857606
#>  [41] 2.501426 4.628313 2.545441 3.581096 2.820538 3.224139 2.737966 2.523516
#>  [49] 2.845871 3.934745 2.600856 2.624768 2.559385 3.227851 2.571346 2.543537
#>  [57] 2.688888 2.511312 2.745993 2.687224 2.749366 2.905425 2.964867 2.876985
#>  [65] 2.920979 4.115747 3.314977 4.817524 3.014441 2.880645 3.361048 2.968897
#>  [73] 3.512561 2.842806 2.614647 2.546635 2.671348 2.830045 2.890637 2.929516
#>  [81] 3.114020 4.038268 2.693748 2.506151 2.571463 2.767798 3.002019 3.195095
#>  [89] 2.613549 2.912089 2.516853 2.580288 2.890254 2.729211 2.733918 2.699733
#>  [97] 3.051525 3.583169 2.646968 3.277088