Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample_GR_magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample_GR_magnitudes(n = 100, beta.p = log(10), M0 = 2.5)
#>   [1] 4.067022 2.613451 2.671687 3.006588 3.019995 2.868644 2.779199 3.442575
#>   [9] 2.836478 2.776403 2.514141 3.428033 2.828919 2.959814 2.534677 3.473637
#>  [17] 3.900535 3.555761 3.575146 3.110088 3.331577 2.740641 4.147613 3.072818
#>  [25] 3.014210 2.675871 3.517427 2.760077 2.970676 3.848894 2.748381 2.925391
#>  [33] 3.199656 2.880251 3.841337 2.665225 2.937564 2.889720 2.549353 2.735920
#>  [41] 2.771371 2.736964 2.818843 2.524371 3.030153 2.896669 2.883661 2.909120
#>  [49] 2.533057 3.074826 2.645458 3.468829 2.848203 2.862761 2.517885 2.551384
#>  [57] 2.665437 2.678430 2.679418 2.595848 2.716276 2.751562 2.853881 2.563020
#>  [65] 2.861800 3.365279 2.678976 2.849953 3.051181 3.021748 2.807142 2.521789
#>  [73] 3.211913 2.542565 2.546314 2.553341 2.567564 2.651453 3.013612 3.114860
#>  [81] 4.025689 4.203778 3.421790 2.667030 3.106005 2.641846 2.652974 2.824852
#>  [89] 4.327524 4.239459 2.823487 2.898946 2.629584 3.290701 2.961258 2.879068
#>  [97] 2.779839 2.697757 2.985909 2.939876