Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample_GR_magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample_GR_magnitudes(n = 100, beta.p = log(10), M0 = 2.5)
#>   [1] 3.782411 2.882975 2.636240 2.635088 3.051574 3.279676 2.523098 3.178869
#>   [9] 2.797164 2.577356 2.564729 2.541284 2.698569 3.200097 2.593103 3.997489
#>  [17] 2.564679 2.535377 2.760022 3.384844 3.143720 2.724627 3.536243 4.099784
#>  [25] 2.739050 2.923180 2.912347 2.779520 2.531256 2.506728 2.846683 2.928756
#>  [33] 2.986639 3.148837 2.781374 2.613431 2.780962 2.640830 2.761699 3.417692
#>  [41] 2.509729 2.576234 2.512377 2.904925 2.917346 3.036495 2.593355 3.179065
#>  [49] 3.212861 2.692665 2.550265 2.995275 2.834230 2.539322 2.796439 2.758306
#>  [57] 3.002400 2.653811 3.037151 2.892939 3.818341 4.269291 2.856068 2.940818
#>  [65] 2.649670 3.020772 2.630095 2.932609 3.060205 2.690585 2.912065 2.825443
#>  [73] 2.857264 3.642672 3.073019 2.597036 2.629205 2.960152 3.195328 3.339618
#>  [81] 3.042801 3.200441 2.580331 2.740955 2.514472 4.817742 2.864164 3.016500
#>  [89] 2.718052 2.528462 2.850257 3.029099 2.671009 2.937640 2.555874 2.669804
#>  [97] 2.720096 3.014481 2.913222 2.652641