Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample_GR_magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample_GR_magnitudes(n = 100, beta.p = log(10), M0 = 2.5)
#>   [1] 2.722849 2.582393 3.024660 2.733507 2.529267 2.698230 3.431753 3.260953
#>   [9] 2.650463 2.691704 2.576572 2.729816 3.235823 2.631213 2.829799 2.824295
#>  [17] 2.528455 2.806634 2.798741 3.505738 2.660921 3.158023 2.884738 2.638515
#>  [25] 2.556970 3.120169 2.838795 2.968702 2.558201 4.824592 2.782388 2.640655
#>  [33] 2.822370 2.752602 2.624318 2.881316 3.006883 2.599422 2.834572 2.738514
#>  [41] 4.328084 3.258951 3.127287 2.565637 2.626546 2.593538 2.800719 3.367017
#>  [49] 2.553638 2.618945 3.648728 3.371058 2.781148 2.532806 3.708504 3.041621
#>  [57] 3.056910 2.674339 2.550865 2.942504 2.736952 2.977078 2.935630 3.209564
#>  [65] 2.617963 2.838306 2.627278 2.535282 2.700256 2.583471 2.513518 2.576704
#>  [73] 2.823953 2.868594 2.544217 4.260388 4.883589 3.171008 2.819100 2.605412
#>  [81] 2.976739 2.749235 2.823902 2.776735 2.640741 3.169122 3.706600 2.613261
#>  [89] 2.957361 2.809391 3.737236 3.012176 2.585299 2.574018 2.823521 4.432838
#>  [97] 2.598904 2.561152 2.601299 3.013012