Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample_GR_magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample_GR_magnitudes(n = 100, beta.p = log(10), M0 = 2.5)
#>   [1] 3.102806 2.815853 2.686869 2.502117 3.387766 2.904878 2.751637 2.653478
#>   [9] 2.668679 2.870284 2.774892 3.878027 2.548858 3.121241 3.329764 3.320863
#>  [17] 3.582932 2.646814 3.111182 3.307412 3.051763 4.112588 3.051066 2.799469
#>  [25] 3.324094 3.522416 2.657714 3.243261 2.551678 2.709822 3.256272 3.131483
#>  [33] 2.500657 2.663152 2.520402 2.930954 2.661265 4.245195 2.552100 2.582201
#>  [41] 2.626679 3.095550 2.533508 3.054471 3.600654 4.184001 2.889144 2.597245
#>  [49] 2.668281 3.039406 3.117756 2.578154 2.750099 2.528627 2.583438 2.629083
#>  [57] 3.286292 3.084457 2.679933 2.637991 2.625143 2.647055 3.008905 2.655883
#>  [65] 2.969773 2.747697 3.944241 2.992620 4.042300 2.720575 2.915031 2.599097
#>  [73] 2.549450 2.544295 2.633686 2.542259 2.626893 3.399966 2.646249 2.669332
#>  [81] 2.907286 3.086668 3.355827 2.778861 2.987184 2.662476 3.933590 3.162263
#>  [89] 2.784287 2.576418 2.773293 3.017981 2.506862 2.516484 2.703955 2.771309
#>  [97] 2.945656 3.031479 2.699330 2.824161