Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample_GR_magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample_GR_magnitudes(n = 100, beta.p = log(10), M0 = 2.5)
#>   [1] 4.506962 2.972540 2.833267 3.008313 3.103630 2.847599 2.609935 3.150950
#>   [9] 3.372622 2.854234 3.125952 2.834474 2.795530 2.749458 2.654685 2.529817
#>  [17] 2.965519 2.872880 3.389334 2.847592 2.599227 2.667839 2.582693 2.598374
#>  [25] 2.852208 2.686039 2.821219 2.740652 3.041428 3.645406 2.503321 2.611665
#>  [33] 2.886285 2.589367 2.574048 2.860405 2.570762 2.858171 4.356929 3.075359
#>  [41] 2.525745 2.512953 3.565917 3.072694 3.085277 2.863773 3.155253 2.559106
#>  [49] 2.742247 2.581195 2.889394 2.505397 2.741503 2.770966 2.990375 2.680197
#>  [57] 3.888183 3.088278 2.569572 4.477035 2.551085 3.072951 2.983071 2.591776
#>  [65] 2.586678 2.508933 3.252455 2.604239 2.612381 3.730887 3.050879 2.934531
#>  [73] 2.610945 2.739747 2.608602 2.512609 2.956107 2.736370 2.791143 3.159655
#>  [81] 2.526157 2.964470 3.825072 2.558447 2.627503 3.291531 3.473636 2.999799
#>  [89] 2.915961 2.766589 2.556262 3.179636 2.994710 2.510139 2.902132 2.591449
#>  [97] 2.502365 2.522312 3.177661 3.030292