Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample_GR_magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample_GR_magnitudes(n = 100, beta.p = log(10), M0 = 2.5)
#>   [1] 3.010207 2.977944 2.727194 2.656467 2.797035 3.236363 2.589347 2.557403
#>   [9] 2.848817 2.746119 2.528964 3.969873 2.714515 3.404716 2.565322 3.111796
#>  [17] 2.513325 3.126395 2.985492 3.096147 2.512656 2.689523 2.706942 2.882904
#>  [25] 2.981247 3.443782 2.825497 3.481875 2.656969 3.124641 2.586578 2.609621
#>  [33] 3.018054 3.585759 2.600237 2.500639 2.659313 3.062850 2.774295 2.741993
#>  [41] 3.131018 2.611436 2.521975 2.829056 3.112567 2.527060 2.608794 3.041858
#>  [49] 2.666134 3.636883 2.681933 3.630823 2.639567 3.642662 2.579450 2.574479
#>  [57] 4.477282 2.634228 4.815827 2.585137 3.108776 2.951415 3.437226 2.791552
#>  [65] 2.697979 2.807144 2.809329 2.700537 3.038435 3.033214 2.634267 3.837249
#>  [73] 2.738168 2.622650 3.214018 2.665314 2.648935 2.853661 2.834162 3.184234
#>  [81] 3.420943 3.044828 2.513568 2.523519 3.299773 3.070468 3.480425 2.817072
#>  [89] 2.719105 2.730526 2.767626 3.471171 2.571678 3.178097 2.843626 3.163106
#>  [97] 2.939514 2.875340 2.976611 3.496162