Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample_GR_magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample_GR_magnitudes(n = 100, beta.p = log(10), M0 = 2.5)
#>   [1] 4.306709 2.944634 2.575399 3.202266 2.550676 2.662066 2.715610 3.746075
#>   [9] 2.558650 2.892967 3.036345 3.092362 2.592120 4.407113 3.226427 2.684419
#>  [17] 2.930978 2.764932 3.020096 3.356788 3.213683 2.546561 2.514803 3.118964
#>  [25] 2.800018 3.070179 3.124049 2.563683 4.019588 2.763711 2.546588 4.010031
#>  [33] 2.713685 2.686510 2.664653 2.606061 3.409516 2.593911 2.965454 3.188315
#>  [41] 3.483759 2.595378 2.679358 2.909867 3.222622 2.598187 2.505835 2.506241
#>  [49] 2.627053 3.308714 2.765447 2.879825 4.099907 3.102728 2.606792 3.110479
#>  [57] 3.415944 2.567594 3.247983 2.517334 2.553885 3.181790 3.016186 3.001037
#>  [65] 2.647462 3.101806 2.506962 2.946314 2.532918 2.813586 2.527983 2.836814
#>  [73] 3.024076 2.944054 2.756732 2.576006 2.603905 2.639665 3.520685 2.797227
#>  [81] 3.015269 4.653122 2.557553 3.319834 2.792296 2.651941 3.213808 2.909081
#>  [89] 3.143582 2.992907 3.185303 2.537072 2.514326 2.582479 2.820094 2.785030
#>  [97] 2.805960 2.634133 5.399505 2.726272