Return a sample of magnitudes drawn from the GR distribution

Return a sample of magnitudes drawn from the GR distribution

Usage

sample_GR_magnitudes(n, beta.p, M0)

Arguments

n

Number of events in the sample.

beta.p

Related to the b-value via b ln(10).

M0

Minimum magnitude for the sample.

Value

A list of magnitudes of length n drawn from a GR distribution.

Examples

sample_GR_magnitudes(n = 100, beta.p = log(10), M0 = 2.5)
#>   [1] 3.264722 3.865149 3.718498 2.872702 2.707736 2.778885 3.005309 3.216066
#>   [9] 4.040281 2.537539 2.502875 3.074191 3.252254 3.538096 2.565686 2.899193
#>  [17] 3.455107 2.503725 2.627809 2.760555 2.975859 3.016969 2.773348 2.818284
#>  [25] 2.544535 3.120450 3.002644 2.704333 2.715076 3.026517 2.699705 3.538156
#>  [33] 3.086593 2.785647 2.739209 2.613822 3.637869 3.324269 2.940304 3.100748
#>  [41] 2.512692 2.567885 2.887344 2.638703 2.585668 4.035616 2.622009 3.125344
#>  [49] 2.663892 3.596801 2.841082 2.772537 2.558073 4.552025 3.340889 2.659250
#>  [57] 3.575852 3.193017 3.273774 3.027792 2.519397 2.696719 3.917600 2.860226
#>  [65] 2.664285 2.602687 2.638658 2.571934 3.158837 2.653939 2.509991 2.669413
#>  [73] 3.306514 2.533057 3.044911 3.039825 3.737968 3.689199 3.754584 2.980111
#>  [81] 2.648180 2.524590 2.639015 2.760987 2.843679 2.553669 2.706788 2.844941
#>  [89] 2.803958 2.764743 2.513176 2.540724 3.299376 2.644882 2.537550 2.892392
#>  [97] 2.833985 2.643590 2.872110 2.503428