Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   500.0389   2.853449
#> 2   507.0653   2.685007
#> 3   505.8198   2.529634
#> 4   500.2101   2.754932
#> 5   500.2872   3.266490
#> 6   512.0314   3.381299
#> 7   500.4784   3.203864
#> 8   506.4639   2.715585
#> 9   501.5647   2.501301
#> 10  501.8192   2.749718
#> 11  500.1240   2.530998
#> 12  514.2169   2.787684
#> 13  546.3114   2.735124
#> 14  500.0697   2.653093
#> 15  553.7723   3.069370
#> 16  500.6240   2.746632
#> 17  585.9259   3.347066
#> 18  500.4076   3.032261
#> 19  520.7798   3.658860
#> 20  500.1619   2.690359
#> 21  599.4910   3.565443
#> 22  500.0527   2.530387
#> 23  511.6093   2.620783
#> 24  500.3571   4.000086
#> 25  891.1407   2.527887
#> 26  562.8445   2.553787
#> 27  500.6839   4.486687
#> 28  660.0465   2.577167
#> 29  500.6276   2.519343
#> 30  700.9847   3.469593
#> 31  501.2471   2.663365
#> 32  507.8828   3.221034
#> 33  551.4863   2.539513
#> 34  539.2085   2.850822
#> 35  500.3280   6.130673
#> 36  736.7688   2.613405
#> 37  645.9136   2.502712
#> 38  500.1628   3.303836
#> 39  599.0133   3.219468
#> 40  503.1313   3.360807
#> 41  528.8481   2.537575
#> 42  500.4307   3.082470
#> 43  503.0621   2.592038
#> 44  500.9103   2.920823
#> 45  500.0943   2.793684
#> 46  511.2338   3.613774
#> 47  779.7173   2.568959
#> 48  519.6774   3.834910
#> 49  500.0252   2.508391
#> 50  612.3361   2.852611
#> 51  610.6878   2.653043
#> 52  500.4774   4.864589
#> 53  555.2027   2.733858
#> 54  581.5433   2.501065
#> 55  500.2542   2.971196
#> 56  509.8652   2.834468
#> 57  504.7732   2.540910
#> 58  614.1012   2.639471
#> 59  500.0286   2.922449
#> 60  527.1828   3.101326
#> 61  500.2019   3.251469
#> 62  500.3127   2.707954
#> 63  500.0091   3.036720
#> 64  500.5622   2.889065
#> 65  500.0128   2.580129
#> 66  953.6511   3.012071
#> 67  501.7451   3.067636
#> 68  502.8552   2.948802
#> 69  713.9202   3.261042
#> 70  535.2569   2.561863
#> 71  500.3894   2.764359
#> 72  500.0592   3.015197
#> 73  500.2446   2.657509
#> 74  502.1531   2.801868
#> 75  502.9170   2.992197
#> 76  500.3373   2.733744
#> 77  989.8972   2.507597
#> 78  535.5355   2.714672
#> 79  500.4932   3.035876
#> 80  526.0654   3.417534
#> 81  531.6141   3.081371
#> 82  622.0649   3.204082
#> 83  511.5936   2.510862
#> 84  500.7621   2.910612
#> 85  500.2316   3.648487
#> 86  502.7659   2.930939
#> 87  605.0837   2.500572
#> 88  579.6285   2.727829
#> 89  500.5907   2.546050
#> 90  500.2527   2.547743
#> 91  594.8061   2.520861
#> 92  556.5526   2.728816
#> 93  500.1248   2.616959
#> 94  587.4161   3.136479
#> 95  885.1016   2.693257
#> 96  504.9571   3.298945
#> 97  500.3138   2.635237
#> 98  599.3016   3.701173
#> 99  500.1889   3.122083
#> 100 501.7054   3.399939
#> 101 500.1543   3.887919
#> 102 558.4263   2.739310
#> 103 868.0363   3.294713
#> 104 568.2655   3.098521
#> 105 501.6705   3.485431
#> 106 631.4634   3.240283
#> 107 531.2179   2.662642
#> 108 980.4196   3.116054
#> 109 624.3273   2.833731
#> 110 501.5921   2.670971
#> 111 502.1367   3.448216
#> 112 500.0152   2.961632
#> 113 522.4415   3.368767
#> 114 523.9627   3.062392
#> 115 500.4203   3.404060
#> 116 500.1835   2.717619
#> 117 500.0130   2.673227
#> 118 523.9773   3.040116
#> 119 508.1346   2.856741
#> 120 562.0104   3.697092
#> 121 505.2945   2.665749
#> 122 596.8465   3.168362
#> 123 526.2696   2.731856
#> 124 526.3153   2.725398
#> 125 500.2516   2.573885
#> 126 672.2179   3.357454
#> 127 500.3189   2.594217
#> 128 663.5551   3.580489
#> 129 503.3520   2.853706
#> 130 502.5566   3.855815
#> 131 500.0377   2.561142
#> 132 500.5414   3.013105
#> 133 518.7572   2.663494
#> 134 500.1463   2.764051
#> 135 531.6556   2.792836
#> 136 515.0391   2.524198
#> 137 509.4045   2.603766
#> 138 500.5023   3.036135
#> 139 501.2260   2.837340
#> 140 500.4955   2.881695
#> 141 500.0160   2.725424
#> 142 514.8758   3.237811
#> 143 500.2176   3.430649
#> 144 621.8028   3.651006
#> 145 500.5769   3.197746
#> 146 502.3042   2.910052
#> 147 500.4687   2.638631
#> 148 517.4511   2.790338
#> 149 507.1710   3.213713
#> 150 500.1867   2.888205
#> 151 770.2843   3.275736
#> 152 500.2967   2.725420
#> 153 517.8056   2.978140
#> 154 500.2942   2.529973
#> 155 501.0554   2.618001
#> 156 500.0981   2.878269
#> 157 501.6636   3.004358
#> 158 500.5754   2.669727
#> 159 500.1688   2.501959
#> 160 500.1097   3.592170
#> 161 502.1839   2.642358
#> 162 524.5224   2.645790
#> 163 512.8741   2.619168
#> 164 500.0276   2.868328
#> 165 500.0666   2.582497
#> 166 500.7544   2.671847
#> 167 500.7964   3.521943
#> 168 513.6883   2.635454
#> 169 507.2335   2.878779
#> 170 749.2202   2.500183
#> 171 501.2052   2.610195
#> 172 503.4632   2.610788
#> 173 500.1170   2.669203
#> 174 720.1785   2.583504
#> 175 715.4601   2.615128
#> 176 508.8137   2.575982
#> 177 500.4058   2.797473
#> 178 502.3298   2.723582
#> 179 604.3456   2.653870
#> 180 807.9465   3.570259
#> 181 507.5857   2.542832
#> 182 518.5241   2.885662
#> 183 506.8297   3.381677
#> 184 500.3682   2.555858
#> 185 888.5581   2.671977
#> 186 500.0216   4.063363
#> 187 521.9829   2.780508
#> 188 508.6890   2.888069
#> 189 574.0479   2.544261
#> 190 614.8342   3.166513
#> 191 500.5393   2.718410
#> 192 501.0401   2.732252
#> 193 500.0082   2.612007
#> 194 505.0353   3.507551
#> 195 500.3542   2.619539
#> 196 500.3468   2.750758
#> 197 513.7650   2.712001
#> 198 500.5161   3.454460
#> 199 617.7669   2.597106
#> 200 502.5293   2.845765
#> 201 500.3811   2.906141
#> 202 534.5161   2.726723
#> 203 500.6640   2.794869
#> 204 551.5693   3.369579
#> 205 500.1637   2.724613
#> 206 506.8609   2.570425
#> 207 500.3890   3.147620
#> 208 500.1760   2.517786
#> 209 500.1803   3.132099
#> 210 500.1096   3.265748
#> 211 500.0933   2.776303
#> 212 694.5903   2.708135
#> 213 500.0025   3.376421
#> 214 851.0230   2.679749
#> 215 519.3454   4.604874
#> 216 501.5024   2.784573
#> 217 548.0393   2.774036
#> 218 516.7034   3.522247
#> 219 500.1238   3.683115
#> 220 504.3917   2.958904
#> 221 501.5316   4.155978
#> 222 500.0690   2.596088
#> 223 500.5272   3.125300
#> 224 500.0152   2.691250
#> 225 500.4431   2.823209
#> 226 500.0906   2.507156
#> 227 500.0582   3.441814
#> 228 523.7634   2.874349
#> 229 743.9083   2.574074
#> 230 821.9418   2.668474
#> 231 502.6079   3.124940
#> 232 557.9531   2.581118
#> 233 510.7387   2.849490
#> 234 500.2431   2.844130
#> 235 507.2243   2.689794
#> 236 599.7597   3.277365
#> 237 501.3642   2.675870
#> 238 500.9348   3.126135
#> 239 511.4053   2.806871
#> 240 518.6999   2.809029
#> 241 500.0076   3.089162
#> 242 500.2465   3.187951
#> 243 500.8049   2.564361
#> 244 505.7532   2.969163
#> 245 500.2053   2.732512
#> 246 510.3835   3.029265
#> 247 500.5453   3.233496
#> 248 501.4655   2.735431
#> 249 803.1636   3.230579
#> 250 516.3905   2.543357
#> 251 500.0352   2.705003
#> 252 500.8490   2.701372
#> 253 507.6346   2.680109
#> 254 500.0220   2.902465
#> 255 500.0233   2.521569
#> 256 500.1859   2.819194
#> 257 597.3002   2.815781
#> 258 528.3660   2.965473
#> 259 740.0809   2.578983
#> 260 500.2343   2.892536
#> 261 505.2281   2.534628
#> 262 653.1060   5.365935
#> 263 500.4029   2.812270
#> 264 501.2158   3.539135
#> 265 513.6432   2.660967
#> 266 864.7337   2.774053
#> 267 505.2097   3.120105
#> 268 502.8889   3.857785
#> 269 502.6236   3.350076
#> 270 502.0651   2.692233
#> 271 948.2190   2.848066
#> 272 501.4079   3.183683
#> 273 501.2508   2.502015
#> 274 503.5814   2.777450
#> 275 644.8004   2.535605
#> 276 503.5797   3.006573
#> 277 531.6984   2.837109
#> 278 500.6106   2.799471
#> 279 501.3038   3.305668
#> 280 500.1479   3.314291
#> 281 503.7026   3.235709
#> 282 500.3518   2.644607
#> 283 500.0477   2.982140
#> 284 500.0107   2.511035
#> 285 520.3994   3.169919
#> 286 500.0176   2.874762
#> 287 500.1153   2.536614
#> 288 501.3770   2.933464
#> 289 500.2066   2.623400
#> 290 788.8164   3.139072
#> 291 649.7014   2.523952
#> 292 500.8563   2.579702
#> 293 500.1656   2.506324
#> 294 500.9026   3.719734
#> 295 643.0446   3.101073
#> 296 500.2695   2.958953
#> 297 503.3702   2.737911
#> 298 937.4120   3.615068
#> 299 500.4467   2.741981
#> 300 553.1629   2.613846
#> 301 501.3869   2.562726
#> 302 503.3933   2.689092
#> 303 503.5062   2.502308
#> 304 520.4348   2.657515
#> 305 505.6323   2.712976
#> 306 606.5057   3.265257
#> 307 500.7083   2.759121
#> 308 500.1019   6.030034
#> 309 500.6674   2.810655
#> 310 500.0132   2.503287
#> 311 501.2079   3.461427
#> 312 506.2893   2.701728
#> 313 506.7025   3.498091
#> 314 586.4280   2.826550
#> 315 500.3654   3.564952
#> 316 500.0299   2.599620
#> 317 506.4342   3.748724
#> 318 500.0169   2.907145
#> 319 504.3826   3.245381
#> 320 583.5435   2.807206
#> 321 530.1225   2.591806
#> 322 509.7966   3.475829
#> 323 500.2757   2.632415
#> 324 500.0171   2.564519
#> 325 536.1587   2.680643
#> 326 500.0979   2.755422
#> 327 500.7772   3.365528
#> 328 500.0981   2.796430
#> 329 514.4291   2.639009
#> 330 518.7975   3.244236
#> 331 502.6628   2.683082
#> 332 579.0319   2.862275
#> 333 504.5627   3.803916
#> 334 533.5210   4.025446
#> 335 510.0114   2.511900
#> 336 508.4163   3.315076
#> 337 501.5666   2.605562
#> 338 500.0939   2.567438
#> 339 698.5569   2.756508
#> 340 500.3227   2.574111
#> 341 509.3033   4.209051
#> 342 503.0528   2.913215
#> 343 500.1447   2.866512
#> 344 501.5676   3.054304
#> 345 530.1061   3.904275
#> 346 527.5068   3.253060
#> 347 647.8039   2.980637
#> 348 501.4001   2.603810
#> 349 500.0589   2.670800
#> 350 506.6157   2.688554
#> 351 500.1977   2.993225
#> 352 810.6210   3.619400
#> 353 503.8427   2.594200
#> 354 552.6215   2.622026
#> 355 820.9929   3.355497
#> 356 511.7997   3.241201
#> 357 500.0214   2.630976
#> 358 504.0874   4.297996
#> 359 502.9635   3.728007
#> 360 549.9182   2.989466
#> 361 504.0604   2.583455
#> 362 500.0845   2.643558
#> 363 526.9086   2.820875
#> 364 502.6072   3.223690
#> 365 501.0404   3.236380
#> 366 501.7190   2.706398
#> 367 504.5118   3.967322
#> 368 500.0380   3.520651
#> 369 735.6105   2.528418
#> 370 527.8997   3.213348
#> 371 502.2573   2.676613
#> 372 519.8032   3.228509
#> 373 500.0085   2.580191
#> 374 500.9265   2.778706
#> 375 736.4788   3.809041
#> 376 500.3425   4.314524
#> 377 504.8660   3.385201
#> 378 500.3149   2.626787
#> 379 505.7576   3.053179
#> 380 501.8370   2.668024
#> 381 502.5745   2.653469
#> 382 500.0100   3.661605
#> 383 528.6752   3.344695
#> 384 799.1642   2.724006
#> 385 504.3112   2.570985
#> 386 500.0107   2.560719
#> 387 503.7757   2.709840
#> 388 503.3312   2.957641
#> 389 600.9541   2.631483
#> 390 500.2719   3.373596
#> 391 500.9259   3.306028
#> 392 504.4311   2.739426
#> 393 511.8726   2.533406
#> 394 511.3843   2.682075
#> 395 502.5961   2.570887
#> 396 544.9374   2.757572
#> 397 501.0113   3.215187
#> 398 512.8227   2.519858
#> 399 519.0924   2.687890
#> 400 502.0965   2.624445
#> 401 516.5655   2.572250
#> 402 509.6697   3.043982
#> 403 510.6620   2.530319
#> 404 500.0922   2.882414
#> 405 505.8101   2.502520
#> 406 500.1225   2.702337
#> 407 500.7965   2.628866
#> 408 501.2488   2.649558
#> 409 500.3766   3.478752
#> 410 530.9450   2.669623
#> 411 517.6878   2.666941
#> 412 507.3738   2.897574
#> 413 500.2281   2.568007
#> 414 525.2482   2.663149
#> 415 533.4293   2.613592
#> 416 500.1404   2.625494
#> 417 513.2518   2.715248
#> 418 500.1180   2.640621
#> 419 500.0192   2.525136
#> 420 519.4023   2.564997
#> 421 715.6047   3.172082
#> 422 920.6082   2.611664
#> 423 930.6043   3.637093
#> 424 500.7073   2.626307
#> 425 593.6317   2.868331
#> 426 500.1236   3.058708
#> 427 502.5669   3.704910
#> 428 500.2395   2.740456
#> 429 526.3873   2.913546
#> 430 808.6830   2.861417
#> 431 501.3145   2.712063
#> 432 500.7991   2.563201
#> 433 507.6303   2.619276
#> 434 500.5591   3.094119
#> 435 502.0729   2.529979
#> 436 500.0558   3.274949
#> 437 681.6571   2.924559
#> 438 507.3487   4.846789
#> 439 500.0928   2.642246
#> 440 500.3238   2.848556
#> 441 502.2550   2.764205
#> 442 570.2147   3.148645
#> 443 500.2010   2.845262
#> 444 500.0659   2.847937
#> 445 500.4445   4.031302
#> 446 501.3898   2.628119
#> 447 500.6715   3.062651
#> 448 512.9573   2.565344
#> 449 502.0145   3.691140
#> 450 506.6233   2.738885
#> 451 545.0447   2.647229
#> 452 502.4992   2.558469
#> 453 500.0574   2.757777
#> 454 501.3583   2.779650
#> 455 515.9619   2.582793
#> 456 500.4143   2.867789
#> 457 513.3055   4.209349
#> 458 500.6039   2.642745
#> 459 535.8023   3.536252
#> 460 552.4422   2.582765
#> 461 507.3472   3.479840
#> 462 500.0342   2.847412
#> 463 501.1054   2.759035
#> 464 746.2911   3.171571
#> 465 501.8790   2.764685
#> 466 501.4683   2.613946
#> 467 500.6370   3.200081
#> 468 500.5823   2.635416
#> 469 697.2752   3.380920
#> 470 503.5235   2.660780
#> 471 500.3603   2.633936
#> 472 511.0219   4.084314
#> 473 656.8201   2.825389
#> 474 500.4986   2.867865
#> 475 506.7521   2.742507
#> 476 501.0921   3.302146
#> 477 500.3884   3.303249
#> 478 592.0354   2.932685
#> 479 540.5489   3.473078
#> 480 533.1308   4.428020
#> 481 557.2986   2.976917
#> 482 581.7153   3.241238
#> 483 500.6974   2.933211
#> 484 501.3347   2.577518
#> 485 500.7129   2.859047
#> 486 500.1594   3.112348
#> 487 515.5567   2.802865
#> 488 894.8027   3.259299
#> 489 501.2907   2.832443
#> 490 549.9302   3.611415
#> 491 699.3883   2.596516
#> 492 728.6794   3.025080
#> 493 500.3762   3.173404
#> 494 500.3087   2.796273
#> 495 501.4892   3.313631
#> 496 500.2122   2.615607
#> 497 500.7717   2.504510
#> 498 503.1973   2.612724
#> 499 503.7634   3.163852
#> 500 759.8334   3.420501
#> 501 500.0283   2.588824
#> 502 500.0034   2.572385
#> 503 500.4675   2.663852
#> 504 501.1791   3.346540
#> 505 503.6159   2.525230
#> 506 532.6071   2.716475
#> 507 638.0258   2.867904
#> 508 500.0194   2.933862
#> 509 500.1833   4.418414
#> 510 500.3159   2.527172
#> 511 507.9216   2.680724
#> 512 559.7045   2.652581
#> 513 666.8193   3.402090
#> 514 510.6973   2.541739
#> 515 500.5952   2.536014
#> 516 500.3056   3.102427
#> 517 995.1302   3.055759
#> 518 546.5216   2.991617
#> 519 502.4832   2.507628
#> 520 699.9593   4.730748
#> 521 509.9541   2.793613
#> 522 514.7438   2.621157
#> 523 533.8403   3.328063
#> 524 502.3334   2.528871
#> 525 511.8551   2.594880
#> 526 623.9047   2.518681
#> 527 500.0018   2.669849
#> 528 502.0737   2.863199
#> 529 514.7688   2.647987
#> 530 500.3895   3.201357
#> 531 614.5713   2.689537
#> 532 501.4644   2.593705
#> 533 506.8976   3.143161
#> 534 500.0051   2.759814
#> 535 534.4197   2.627292
#> 536 511.9903   2.848858
#> 537 500.0087   2.722262
#> 538 500.0204   2.708251
#> 539 767.6663   2.721647
#> 540 503.8472   2.755614
#> 541 500.0818   3.353853
#> 542 534.7776   3.154477
#> 543 607.4350   3.476395
#> 544 754.4744   2.911326
#> 545 504.5157   2.618307
#> 546 500.2532   2.708335
#> 547 500.2702   2.937801
#> 548 500.2953   2.802623
#> 549 500.1128   3.040085
#> 550 501.1864   2.600729
#> 551 502.1645   2.562674
#> 552 500.0260   3.040206
#> 553 502.3630   3.075603
#> 554 627.1537   3.313474
#> 555 517.7613   2.820093
#> 556 695.7244   2.999768
#> 557 500.2803   2.539272
#> 558 950.6620   2.548974
#> 559 503.1372   3.187537
#> 560 504.1614   4.622355
#> 561 517.1090   2.724353
#> 562 508.9245   2.842169
#> 563 500.5494   2.682522
#> 564 507.4436   2.693230
#> 565 501.1010   2.528924
#> 566 649.1268   3.956670
#> 567 585.2137   3.081809
#> 568 500.1160   2.988802
#> 569 500.3757   2.560537
#> 570 500.0998   2.873531
#> 571 529.2571   3.095810
#> 572 500.2206   2.806943
#> 573 501.4676   2.639238
#> 574 700.1146   2.551248
#> 575 528.7058   2.539196
#> 576 955.8941   2.624875
#> 577 500.4494   2.883287
#> 578 505.0251   3.146489
#> 579 522.3165   2.677360
#> 580 810.4231   2.770922
#> 581 508.0893   3.234840
#> 582 563.2770   3.400772
#> 583 505.9713   3.391439
#> 584 501.6154   3.232440
#> 585 508.0004   2.742736
#> 586 504.4606   2.543356
#> 587 502.8984   2.603801
#> 588 523.2133   2.814891
#> 589 500.5053   2.561888
#> 590 500.5763   2.621289
#> 591 500.0225   3.213010
#> 592 500.3587   3.045587
#> 593 500.1105   2.645117
#> 594 500.6648   2.857744
#> 595 500.2233   2.536976
#> 596 500.2990   3.436431
#> 597 506.3744   2.826380
#> 598 653.7309   2.549154
#> 599 502.3387   3.574318
#> 600 500.0367   4.426441
#> 601 615.8842   2.547232
#> 602 500.2992   3.855429
#> 603 500.2687   2.735587
#> 604 503.6699   2.815843
#> 605 500.0406   2.912398
#> 606 546.9871   2.727298
#> 607 591.1051   3.080085
#> 608 502.0325   2.629765
#> 609 690.3065   2.584479
#> 610 501.4580   5.152392
#> 611 521.4346   2.846116
#> 612 522.4283   3.078973
#> 613 947.9802   2.810749
#> 614 501.1641   3.092707
#> 615 520.6959   2.951712
#> 616 501.2729   2.651042
#> 617 503.7110   2.568855
#> 618 838.1423   3.130293
#> 619 501.5809   2.577199
#> 620 646.7367   2.971855
#> 621 551.2104   2.832384
#> 622 519.1866   2.581805
#> 623 500.3727   2.504218
#> 624 502.5739   2.526894
#> 625 552.6880   2.720322
#> 626 517.4258   2.731760
#> 627 506.7289   3.012251
#> 628 500.0198   2.863104
#> 629 626.9273   2.790752
#> 630 656.8730   3.502747
#> 631 500.8960   3.483232
#> 632 606.1385   3.249410
#> 633 508.8285   2.889308
#> 634 678.3468   2.809353
#> 635 611.4585   2.574217
#> 636 878.5042   2.798070
#> 637 500.3360   2.907838
#> 638 500.6079   2.952656
#> 639 501.0126   2.525170
#> 640 501.4115   2.582943
#> 641 879.0269   3.583192
#> 642 587.2533   2.818344
#> 643 500.0780   2.610979
#> 644 504.7782   2.697811
#> 645 500.9455   3.226913
#> 646 502.1946   2.543792
#> 647 529.5251   2.697844
#> 648 502.8497   2.856248
#> 649 558.9738   4.438656
#> 650 500.0202   3.466527
#> 651 502.9966   3.087535
#> 652 500.3688   2.773254
#> 653 500.8420   2.737543
#> 654 500.1939   3.008831
#> 655 506.9534   2.839388
#> 656 500.0714   3.081099
#> 657 782.8013   2.600434
#> 658 570.7948   3.116957
#> 659 502.0798   2.558354
#> 660 968.6985   2.542315
#> 661 782.0825   3.113607
#> 662 848.4764   3.221141
#> 663 575.8307   3.366556
#> 664 500.1842   3.384649
#> 665 532.9450   2.714143
#> 666 500.1381   2.944476
#> 667 535.0341   2.586747
#> 668 505.5890   2.989545
#> 669 501.8655   2.647601
#> 670 500.0142   3.021037
#> 671 500.8602   2.818062
#> 672 504.6751   2.524047
#> 673 502.4366   2.541787
#> 674 538.4237   2.992916
#> 675 759.8590   2.756374
#> 676 503.3269   2.999047
#> 677 522.0457   3.256697
#> 678 507.4275   2.839517
#> 679 517.9368   2.585178
#> 680 537.9943   2.792848
#> 681 500.0488   2.569016
#> 682 526.0900   2.630901
#> 683 500.2709   3.047475
#> 684 617.7315   2.548756
#> 685 503.7280   2.552768
#> 686 763.2167   2.825409
#> 687 523.3367   2.741181
#> 688 500.9599   2.844125
#> 689 501.2018   2.574484
#> 690 500.1179   2.906065
#> 691 517.6154   2.904149
#> 692 511.6409   2.883396
#> 693 500.0183   2.757087
#> 694 803.3289   3.637767
#> 695 500.0355   2.874519
#> 696 525.1159   2.818635
#> 697 500.4509   2.920602
#> 698 590.4991   2.527343
#> 699 500.0077   2.635248
#> 700 500.1293   3.369262
#> 701 504.8616   2.858816
#> 702 668.3518   2.544966
#> 703 533.9554   3.460718
#> 704 500.2415   2.709366
#> 705 516.3595   2.647333
#> 706 598.9380   3.210103
#> 707 512.5446   3.287565
#> 708 500.8133   2.762881
#> 709 500.0309   2.667334
#> 710 500.2453   2.576424
#> 711 799.6693   2.770835
#> 712 517.8949   2.772328
#> 713 585.5866   2.916387
#> 714 506.9653   2.521566
#> 715 500.0515   3.323506
#> 716 502.6693   2.578693
#> 717 533.4869   2.660706
#> 718 537.5760   2.521214
#> 719 500.0984   2.837817
#> 720 501.4654   2.669613
#> 721 501.2136   2.503090
#> 722 699.4071   3.090311
#> 723 500.0018   2.816674
#> 724 512.7403   3.288443
#> 725 691.7659   3.130214
#> 726 501.2417   3.106486
#> 727 502.8802   2.636362
#> 728 504.3700   2.676241
#> 729 515.3452   3.893228
#> 730 501.9805   2.542978
#> 731 500.1107   3.225116
#> 732 500.2091   2.511613
#> 733 518.1998   4.894854
#> 734 500.5211   2.751067
#> 735 500.5059   2.823439
#> 736 515.8417   2.732525
#> 737 501.3310   3.241920
#> 738 504.7925   3.699988
#> 739 508.0123   2.674841
#> 740 649.6495   2.597924
#> 741 501.2426   2.577017
#> 742 567.3439   3.064467
#> 743 502.0558   4.732749
#> 744 504.7738   3.556647
#> 745 500.0810   2.720516
#> 746 799.3205   2.568836
#> 747 502.6522   3.150333
#> 748 547.8243   2.906825
#> 749 501.0368   3.396905
#> 750 500.0948   2.713597
#> 751 500.4117   3.056897
#> 752 530.2000   2.885231
#> 753 504.8438   2.617486
#> 754 531.2590   2.539116
#> 755 500.5244   2.814934
#> 756 508.8274   2.600723
#> 757 526.6201   2.990605
#> 758 500.3081   2.850136
#> 759 533.6932   3.561441
#> 760 500.0420   2.766003
#> 761 500.8360   3.293557
#> 762 500.7271   2.503315
#> 763 500.2057   2.880977
#> 764 501.2960   2.905554
#> 765 503.9530   2.705278
#> 766 507.7694   3.585240
#> 767 510.6625   2.601246
#> 768 500.2924   2.827364
#> 769 500.5594   2.997797
#> 770 501.2758   3.105423
#> 771 546.0086   2.716717
#> 772 500.8171   2.565926
#> 773 503.6630   3.327334
#> 774 500.1538   2.515489
#> 775 500.0951   2.674392
#> 776 502.1420   2.961003
#> 777 658.3810   2.940409
#> 778 501.1757   3.020804
#> 779 500.0690   2.932397
#> 780 500.6721   3.238333
#> 781 506.7767   2.995725
#> 782 500.4224   2.636946
#> 783 505.1526   3.193126
#> 784 539.9397   2.727992
#> 785 534.7171   2.816438
#> 786 500.5686   3.055331
#> 787 500.0635   2.539054
#> 788 570.9495   2.567290
#> 789 502.3422   3.261075
#> 790 991.2748   4.495730
#> 791 770.4474   2.556003
#> 792 524.2895   2.501344
#> 793 500.2229   3.595934
#> 794 515.2527   3.265401
#> 795 530.4537   2.619920
#> 796 511.5293   2.749365
#> 797 500.6051   3.047371
#> 798 500.3538   3.082712
#> 799 610.7205   2.720357
#> 800 500.0799   2.760362
#> 801 511.6167   3.142969
#> 802 500.2948   3.534925
#> 803 509.0612   2.955346
#> 804 659.4368   2.701667
#> 805 534.1805   2.677629
#> 806 503.5167   2.549909
#> 807 500.0851   3.088147
#> 808 664.9747   3.361958
#> 809 977.6763   2.823133
#> 810 500.1931   3.278842
#> 811 500.0014   2.678560
#> 812 519.6728   2.658423
#> 813 530.9324   3.196088
#> 814 500.7763   2.526652
#> 815 500.7823   2.866987
#> 816 500.6073   2.628492
#> 817 500.0313   2.869266
#> 818 689.5342   3.922857
#> 819 500.1275   2.579425
#> 820 721.3425   2.822336
#> 821 784.6060   3.070131
#> 822 507.1257   3.565637
#> 823 500.4046   3.643437
#> 824 755.3170   2.803775
#> 825 500.0653   2.973994
#> 826 706.3191   2.855120
#> 827 500.0294   2.756573
#> 828 503.6481   2.617821
#> 829 737.7017   3.051023
#> 830 500.0056   2.762962
#> 831 506.6681   3.153027
#> 832 503.3835   3.301855
#> 833 500.7618   2.971950
#> 834 500.1338   2.746620
#> 835 519.0533   2.928184
#> 836 500.7993   4.318129
#> 837 500.0146   2.775586
#> 838 504.9049   2.612269
#> 839 517.6103   2.841814
#> 840 514.5680   3.193831
#> 841 500.1657   2.738745
#> 842 805.6325   2.543975
#> 843 503.1729   2.594504
#> 844 500.1509   2.853219
#> 845 501.2315   2.978070
#> 846 501.4855   2.528479
#> 847 500.0449   2.695042
#> 848 645.3654   2.720901
#> 849 556.4570   3.239246
#> 850 550.7409   2.576505
#> 851 521.7947   2.784217
#> 852 500.1170   2.512670
#> 853 501.7737   2.584735
#> 854 508.3546   3.053993
#> 855 500.6930   2.590486
#> 856 500.0242   2.562957
#> 857 500.2535   2.619755
#> 858 500.0415   2.828343
#> 859 659.1714   2.660652
#> 860 500.2570   2.997314
#> 861 503.9144   2.746425
#> 862 564.5329   3.942340
#> 863 508.9287   2.720470
#> 864 500.8606   2.564689
#> 865 588.3908   2.577854
#> 866 500.2565   3.522750
#> 867 590.1738   2.909960
#> 868 501.5133   3.534077
#> 869 501.5217   2.950382
#> 870 802.8831   2.624533
#> 871 500.4392   3.562937
#> 872 522.5904   4.601657
#> 873 500.3355   3.442155
#> 874 500.1185   3.103695
#> 875 500.5086   2.903914
#> 876 500.0024   2.894110
#> 877 518.3045   3.039365
#> 878 506.0392   2.791585
#> 879 502.3433   2.544836
#> 880 522.9091   2.817307
#> 881 516.8781   2.529550
#> 882 695.3307   2.539403
#> 883 500.2245   2.716865
#> 884 500.9191   2.555625
#> 885 500.0267   3.078881
#> 886 505.0415   3.035927
#> 887 501.1024   3.007714
#> 888 503.8930   2.743236
#> 889 593.3557   2.564191
#> 890 500.2083   2.640039
#> 891 500.1587   2.616126
#> 892 523.5573   2.751298
#> 893 537.8583   2.718099
#> 894 500.0216   2.576378
#> 895 599.5114   2.562398
#> 896 521.9732   2.835485
#> 897 530.7868   3.122281
#> 898 503.0986   3.637227
#> 899 534.0506   2.565916
#> 900 701.0727   2.666371
#> 901 500.0529   2.866858
#> 902 500.3829   3.233794
#> 903 526.4763   2.572969
#> 904 510.1776   3.009666
#> 905 506.3294   3.057484
#> 906 500.0455   2.582218
#> 907 501.6709   2.610164
#> 908 593.5196   2.984556
#> 909 501.0766   2.914642
#> 910 766.6590   2.867500
#> 911 552.4871   2.672945
#> 912 528.2249   3.965379