Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   505.5789   2.754875
#> 2   988.7438   2.619158
#> 3   500.0304   3.475357
#> 4   512.9177   2.907884
#> 5   506.3365   3.323773
#> 6   500.0131   2.892468
#> 7   501.9936   2.746852
#> 8   500.6043   3.034901
#> 9   505.3167   3.163619
#> 10  500.0113   2.625334
#> 11  500.5502   4.205636
#> 12  538.6268   2.606354
#> 13  507.1246   3.190183
#> 14  500.6223   3.186322
#> 15  500.2226   3.050590
#> 16  500.0705   2.673088
#> 17  632.7242   2.916469
#> 18  573.2716   2.606810
#> 19  502.2191   3.066973
#> 20  503.9266   3.012643
#> 21  507.8722   2.836795
#> 22  503.1395   3.127172
#> 23  500.0244   2.503272
#> 24  531.1017   2.959192
#> 25  544.9581   3.327998
#> 26  500.0291   2.670757
#> 27  502.1779   2.520475
#> 28  500.5192   2.506501
#> 29  500.3204   3.010155
#> 30  609.0372   3.031038
#> 31  500.5235   2.849581
#> 32  505.9646   2.752970
#> 33  502.1777   3.974392
#> 34  503.4150   2.502056
#> 35  503.8550   3.677069
#> 36  500.9757   2.862050
#> 37  968.1102   2.925021
#> 38  522.9686   3.980111
#> 39  756.6814   3.024415
#> 40  624.8262   3.480887
#> 41  507.6815   2.606151
#> 42  501.5053   2.609126
#> 43  549.8844   2.634645
#> 44  683.2149   2.538745
#> 45  500.2072   3.875958
#> 46  516.2178   2.576734
#> 47  500.0566   2.640039
#> 48  500.1971   2.731393
#> 49  500.9924   2.575026
#> 50  500.0418   3.226221
#> 51  500.3482   2.526727
#> 52  500.5922   3.434098
#> 53  500.3260   2.655978
#> 54  501.8012   3.252370
#> 55  500.6604   2.611558
#> 56  500.6179   3.027011
#> 57  501.0338   2.692828
#> 58  501.2240   3.506478
#> 59  500.2544   3.080540
#> 60  744.3348   2.779207
#> 61  809.9002   4.170224
#> 62  500.9536   3.637518
#> 63  500.0751   2.770605
#> 64  501.5134   2.621895
#> 65  549.3104   2.507133
#> 66  514.5922   3.083392
#> 67  564.9877   2.914301
#> 68  507.6604   2.768919
#> 69  502.2050   3.315341
#> 70  777.1850   2.921197
#> 71  500.5823   2.760421
#> 72  500.5810   3.189608
#> 73  900.2897   3.187446
#> 74  616.0913   3.767600
#> 75  524.2099   2.656451
#> 76  536.5100   3.393787
#> 77  719.9322   3.150965
#> 78  500.3143   2.740109
#> 79  500.9113   2.596675
#> 80  501.1840   2.692570
#> 81  504.1198   3.061566
#> 82  875.5854   2.747334
#> 83  570.1219   3.532130
#> 84  500.4346   2.524371
#> 85  503.9514   2.627737
#> 86  522.7390   3.166332
#> 87  520.3833   3.182591
#> 88  502.4025   3.172941
#> 89  515.7228   2.822050
#> 90  531.5375   2.685652
#> 91  537.6988   2.810049
#> 92  505.5176   3.419032
#> 93  504.5340   3.239396
#> 94  500.0313   2.589987
#> 95  500.4454   3.683529
#> 96  500.8881   2.814763
#> 97  501.1663   3.655281
#> 98  558.0769   2.714555
#> 99  631.3776   2.521036
#> 100 501.0775   2.601315
#> 101 660.8778   2.832532
#> 102 500.0233   3.212460
#> 103 504.2233   2.968225
#> 104 541.6323   3.571600
#> 105 501.0416   3.129856
#> 106 504.9731   2.723612
#> 107 605.9686   2.995580
#> 108 500.1742   3.276628
#> 109 613.3355   3.137984
#> 110 501.9756   2.688813
#> 111 511.9337   3.416056
#> 112 502.7777   3.133108
#> 113 500.2985   3.518580
#> 114 502.3101   2.676295
#> 115 511.5234   2.550788
#> 116 541.1793   2.595493
#> 117 606.9095   2.990079
#> 118 594.9325   2.769995
#> 119 943.9922   3.158490
#> 120 500.4101   2.944683
#> 121 500.1134   2.763441
#> 122 674.1344   2.864951
#> 123 505.6125   3.733909
#> 124 531.0425   2.992183
#> 125 500.0871   2.704723
#> 126 677.3241   2.707592
#> 127 505.6234   2.882735
#> 128 501.4671   2.956674
#> 129 520.3596   3.258881
#> 130 500.5607   2.544078
#> 131 509.7859   2.663176
#> 132 500.8401   3.401153
#> 133 500.4304   3.339008
#> 134 506.7296   3.070533
#> 135 500.4714   2.564241
#> 136 500.1521   2.593331
#> 137 500.3143   2.786319
#> 138 526.0690   2.504444
#> 139 520.8993   2.513185
#> 140 523.1721   2.718255
#> 141 501.4997   2.760351
#> 142 601.9822   2.954314
#> 143 629.9359   2.611237
#> 144 821.8410   2.731042
#> 145 500.8445   2.513163
#> 146 501.1633   3.273099
#> 147 811.0323   2.930477
#> 148 503.8732   3.710626
#> 149 683.5018   2.671368
#> 150 507.8687   3.528530
#> 151 500.9885   2.864945
#> 152 503.4691   3.096573
#> 153 510.2762   3.505005
#> 154 520.3586   2.505446
#> 155 502.1968   2.515817
#> 156 500.1090   2.967967
#> 157 501.0982   3.275478
#> 158 505.6276   2.801542
#> 159 501.8206   2.738435
#> 160 529.6411   2.514711
#> 161 500.0173   2.860116
#> 162 502.2702   2.579473
#> 163 500.0510   2.562060
#> 164 916.2994   3.338111
#> 165 500.1044   2.552700
#> 166 529.2247   2.608164
#> 167 500.2991   3.528549
#> 168 500.0409   2.591518
#> 169 512.2833   3.208035
#> 170 505.9799   4.297555
#> 171 508.9909   2.719588
#> 172 667.1960   2.549879
#> 173 504.0178   2.833675
#> 174 501.9449   2.762787
#> 175 500.9437   2.651290
#> 176 503.5975   3.722235
#> 177 523.2889   3.735498
#> 178 550.4113   2.701352
#> 179 500.4649   2.575161
#> 180 608.2224   3.104379
#> 181 500.1025   2.629502
#> 182 500.9438   2.937672
#> 183 500.0006   2.509256
#> 184 500.1187   3.137627
#> 185 515.8101   4.033168
#> 186 527.6037   3.155270
#> 187 500.1287   2.741153
#> 188 506.9084   2.623065
#> 189 500.1239   2.980878
#> 190 508.9122   2.927416
#> 191 510.3911   3.246143
#> 192 504.0357   2.536567
#> 193 500.2129   2.877579
#> 194 500.6938   3.348194
#> 195 500.0684   2.651450
#> 196 500.5875   2.731934
#> 197 500.2673   2.847663
#> 198 504.9494   3.540255
#> 199 500.0691   2.524427
#> 200 506.4746   3.011272
#> 201 516.0453   2.820728
#> 202 562.2255   2.647804
#> 203 500.3332   2.568191
#> 204 500.2210   2.744486
#> 205 500.0356   3.132002
#> 206 962.3045   2.506841
#> 207 501.0805   3.176546
#> 208 505.4994   3.027487
#> 209 500.0038   2.555804
#> 210 721.3910   3.087230
#> 211 506.5523   2.536892
#> 212 500.7898   2.724907
#> 213 672.4826   2.924351
#> 214 500.4997   2.632912
#> 215 500.5382   2.639229
#> 216 515.2540   3.401324
#> 217 501.3900   4.545828
#> 218 653.2495   2.567551
#> 219 502.5997   2.822791
#> 220 500.1672   2.959349
#> 221 539.2122   3.536719
#> 222 501.3992   2.566263
#> 223 501.4430   3.225650
#> 224 881.3393   2.574790
#> 225 761.2520   2.632086
#> 226 500.0554   2.688947
#> 227 528.7406   2.500573
#> 228 500.1523   2.608332
#> 229 508.3834   3.203293
#> 230 525.7740   2.686484
#> 231 502.8646   2.548095
#> 232 500.0087   2.726817
#> 233 500.1035   3.587592
#> 234 683.0872   2.753596
#> 235 510.6422   2.725801
#> 236 500.5769   2.505843
#> 237 506.1483   2.587595
#> 238 500.0381   3.334034
#> 239 502.3355   3.142217
#> 240 762.0342   2.805406
#> 241 500.4376   2.815450
#> 242 514.5352   2.678666
#> 243 747.7924   2.760114
#> 244 500.1474   2.749831
#> 245 503.2512   2.500554
#> 246 500.2994   2.563660
#> 247 504.3044   3.481532
#> 248 500.2878   2.577063
#> 249 756.9707   3.098601
#> 250 500.1885   2.615352
#> 251 503.6104   2.652812
#> 252 500.0508   2.787238
#> 253 522.0860   2.805540
#> 254 501.9045   2.571134
#> 255 500.1439   3.015345
#> 256 671.6013   2.505943
#> 257 501.6581   3.764419
#> 258 511.0656   3.132586
#> 259 504.0186   3.135365
#> 260 515.0628   2.704837
#> 261 510.1556   2.684323
#> 262 509.2837   2.540028
#> 263 500.6561   3.291607
#> 264 500.2903   2.641308
#> 265 500.0857   3.056145
#> 266 502.5045   2.588629
#> 267 965.7468   2.682758
#> 268 626.9414   2.900977
#> 269 500.0798   2.822418
#> 270 504.5664   2.889515
#> 271 500.0690   2.732030
#> 272 768.4577   2.761688
#> 273 501.5917   2.545206
#> 274 500.6987   3.284576
#> 275 501.6482   2.644763
#> 276 500.0139   2.530595
#> 277 972.2906   3.810194
#> 278 768.4267   3.074615
#> 279 500.1765   3.430169
#> 280 524.5201   2.621259
#> 281 504.0112   2.709769
#> 282 500.9531   2.763186
#> 283 510.5408   2.533168
#> 284 505.5472   2.695490
#> 285 540.9636   2.737897
#> 286 500.0622   3.291496
#> 287 500.9919   2.557406
#> 288 500.1949   3.003231
#> 289 504.0843   2.667545
#> 290 500.4713   2.625107
#> 291 500.5593   2.604554
#> 292 500.0078   2.684086
#> 293 500.0273   2.687217
#> 294 500.1582   2.587752
#> 295 503.1589   2.958761
#> 296 667.4764   2.769393
#> 297 502.1578   2.529339
#> 298 500.2876   2.603571
#> 299 506.1995   3.382029
#> 300 648.7018   3.696763
#> 301 500.3457   2.945956
#> 302 500.7734   3.552862
#> 303 697.9209   2.888548
#> 304 516.8184   2.538626
#> 305 500.1217   2.566517
#> 306 501.3404   2.648534
#> 307 521.3960   3.460692
#> 308 500.1096   3.048908
#> 309 565.8619   2.599749
#> 310 509.8470   3.039676
#> 311 552.9324   3.099203
#> 312 526.4782   3.085525
#> 313 501.0944   3.012729
#> 314 500.5255   2.530630
#> 315 502.5403   3.626330
#> 316 500.2936   3.129319
#> 317 603.4409   2.566073
#> 318 556.8130   2.981841
#> 319 726.7612   2.695152
#> 320 543.4881   2.589052
#> 321 502.5724   5.906815
#> 322 567.1131   2.518053
#> 323 502.0788   2.792402
#> 324 570.8247   2.544699
#> 325 575.2933   2.937465
#> 326 532.3542   4.798589
#> 327 513.5994   2.819627
#> 328 500.3989   4.678547
#> 329 502.3107   2.943221
#> 330 533.3907   2.906079
#> 331 501.6504   2.502010
#> 332 535.6425   3.120806
#> 333 501.7299   2.540486
#> 334 502.7870   2.940282
#> 335 524.5803   2.618898
#> 336 521.0827   2.515456
#> 337 509.2917   2.823034
#> 338 800.2017   2.732420
#> 339 510.2111   2.881605
#> 340 500.4291   3.113204
#> 341 500.1509   2.627324
#> 342 507.5493   2.986011
#> 343 516.7635   2.626511
#> 344 501.2186   2.969304
#> 345 500.1169   2.754669
#> 346 804.0821   2.557056
#> 347 500.2489   2.533023
#> 348 500.4474   2.516999
#> 349 508.9327   3.006121
#> 350 501.5844   3.540314
#> 351 612.5364   2.630455
#> 352 500.0765   2.971802
#> 353 743.1800   3.128006
#> 354 500.3504   3.220493
#> 355 728.0169   2.687617
#> 356 513.2865   3.348186
#> 357 500.0148   2.915930
#> 358 510.2093   2.941831
#> 359 708.6838   3.047400
#> 360 564.3100   2.732108
#> 361 500.3825   3.436625
#> 362 596.0935   2.566228
#> 363 979.2224   3.047734
#> 364 500.0577   2.684853
#> 365 823.5700   2.508472
#> 366 501.0649   4.107703
#> 367 750.0041   2.505389
#> 368 501.5707   3.286761
#> 369 500.0427   2.668962
#> 370 935.5532   2.565663
#> 371 502.3365   2.751493
#> 372 500.0523   2.722329
#> 373 500.5695   3.028970
#> 374 502.8293   3.064611
#> 375 534.5425   2.901177
#> 376 509.4276   3.083583
#> 377 527.7343   2.882690
#> 378 504.5188   2.612701
#> 379 500.1441   2.522141
#> 380 500.2593   2.651905
#> 381 564.5557   2.865023
#> 382 502.4376   2.561511
#> 383 741.5332   3.206630
#> 384 500.1147   3.329159
#> 385 513.1117   2.504183
#> 386 661.5930   2.542926
#> 387 502.0309   2.515243
#> 388 921.6441   2.859927
#> 389 500.5175   2.851348
#> 390 500.3599   2.982826
#> 391 525.3424   3.067148
#> 392 735.5754   2.608877
#> 393 500.7143   2.985561
#> 394 500.3108   2.826755
#> 395 509.0895   2.942463
#> 396 829.4299   2.535753
#> 397 512.5639   2.661971
#> 398 500.5803   4.071858
#> 399 500.0924   2.841428
#> 400 535.0947   2.653862
#> 401 711.4112   2.877825
#> 402 743.8028   2.529517
#> 403 500.2385   2.995771
#> 404 500.2068   2.768369
#> 405 500.0111   2.545865
#> 406 535.0628   2.870679
#> 407 522.2462   4.178796
#> 408 500.2907   2.574315
#> 409 500.0921   2.797544
#> 410 507.1388   2.713048
#> 411 655.4335   2.963006
#> 412 916.7024   3.196403
#> 413 500.0568   2.658842
#> 414 697.5087   3.532085
#> 415 501.8724   2.962135
#> 416 501.1521   2.633332
#> 417 500.1917   2.589879
#> 418 500.8480   2.547030
#> 419 500.8887   2.844113
#> 420 500.0771   3.094685
#> 421 800.2501   2.582713
#> 422 514.2647   2.941298
#> 423 501.6505   3.597674
#> 424 983.0009   2.538466
#> 425 699.0448   2.959537
#> 426 500.3588   2.680711
#> 427 818.7329   3.043354
#> 428 505.2434   3.179191
#> 429 504.7705   2.660326
#> 430 500.0385   3.131713
#> 431 500.0026   3.225844
#> 432 558.3090   3.643389
#> 433 502.8825   2.792291
#> 434 503.5728   3.110025
#> 435 615.9955   2.874696
#> 436 502.5113   2.773798
#> 437 513.9273   3.184095
#> 438 502.6384   2.664476
#> 439 500.8093   2.668966
#> 440 502.3180   2.836230
#> 441 507.4684   2.519642
#> 442 512.5290   3.055215
#> 443 501.8447   2.617620
#> 444 500.2140   2.682531
#> 445 669.8134   2.918552
#> 446 508.5624   3.872865
#> 447 522.6680   3.628295
#> 448 663.3097   2.843711
#> 449 512.8878   4.241959
#> 450 586.3419   3.127027
#> 451 509.4827   3.247436
#> 452 500.6184   3.644608
#> 453 901.2554   3.521205
#> 454 564.7182   3.161128
#> 455 693.4585   2.553196
#> 456 653.1976   2.643243
#> 457 502.0317   2.645234
#> 458 500.7719   3.263587
#> 459 501.9404   2.576569
#> 460 642.1976   3.082587
#> 461 578.3126   2.576022
#> 462 500.1490   2.916307
#> 463 512.0734   3.032972
#> 464 500.2955   2.638122
#> 465 500.1430   2.550437
#> 466 972.9849   2.845035
#> 467 506.8029   2.574928
#> 468 500.6013   2.759842
#> 469 517.7133   2.621506
#> 470 532.1897   2.553583
#> 471 500.1287   2.854271
#> 472 500.5531   2.688599
#> 473 528.1224   2.920046
#> 474 503.2024   2.700108
#> 475 548.6881   2.586903
#> 476 506.1603   2.838551
#> 477 500.0785   2.699418
#> 478 500.8040   2.596005
#> 479 500.6984   3.064938
#> 480 500.1180   3.167700
#> 481 500.0172   2.640815
#> 482 520.1288   2.846119
#> 483 500.9233   2.586253
#> 484 504.3173   2.731671
#> 485 531.6452   3.046650
#> 486 501.0080   3.267966
#> 487 508.3816   2.610404
#> 488 500.0155   2.889494
#> 489 734.1710   2.592819
#> 490 500.7618   2.859657
#> 491 952.3743   2.557108
#> 492 500.4222   3.596558
#> 493 897.4654   2.874793
#> 494 500.0353   2.857234
#> 495 500.1898   3.513755
#> 496 500.7100   2.674246
#> 497 500.2874   2.827542
#> 498 821.5665   2.882362
#> 499 500.5360   2.503804
#> 500 506.7250   2.974402
#> 501 500.5267   3.073162
#> 502 500.1462   3.604463
#> 503 501.5312   3.219382
#> 504 581.7488   2.690003
#> 505 500.0311   2.902976
#> 506 507.5668   2.529505
#> 507 663.8407   2.635951
#> 508 502.8417   2.811613
#> 509 501.2528   2.609583
#> 510 500.1345   2.721329
#> 511 500.4237   2.776397
#> 512 500.0262   2.750372
#> 513 607.8168   3.466519
#> 514 500.3315   3.123085
#> 515 661.6556   2.632561
#> 516 503.0386   2.626215
#> 517 755.4511   2.612265
#> 518 500.2110   2.866600
#> 519 913.6995   2.748611
#> 520 500.0344   3.676994
#> 521 500.3092   3.194090
#> 522 500.6868   2.560061
#> 523 593.3215   2.557623
#> 524 540.1192   2.637269
#> 525 500.1246   3.053803
#> 526 500.2713   2.828813
#> 527 520.9545   2.740111
#> 528 500.0398   3.104144
#> 529 504.7837   3.527663
#> 530 503.9331   3.250954
#> 531 501.0646   3.188093
#> 532 504.5952   2.807480
#> 533 515.5399   3.006432
#> 534 500.4156   2.710090
#> 535 502.6513   2.818577
#> 536 500.1427   2.599980
#> 537 500.1576   2.659530
#> 538 506.4593   3.021879
#> 539 570.9428   2.536410
#> 540 500.3508   2.864341
#> 541 511.5045   2.656902
#> 542 500.7237   2.662510
#> 543 500.1161   3.251382
#> 544 545.4057   3.893296
#> 545 500.8714   2.560980
#> 546 501.3417   3.229595
#> 547 500.5505   4.409966
#> 548 690.0063   2.937751
#> 549 555.3039   2.563254
#> 550 500.1485   2.553134
#> 551 500.6708   2.635748
#> 552 924.8465   2.537044
#> 553 513.2661   2.764479
#> 554 500.0032   2.672880
#> 555 503.1162   2.825910
#> 556 509.5248   2.599324
#> 557 500.1001   2.564622
#> 558 500.8543   2.757779
#> 559 500.1440   2.526983
#> 560 546.9575   2.867137
#> 561 587.4733   3.565602
#> 562 527.2032   2.792985
#> 563 500.5513   2.666405
#> 564 670.6373   2.600632
#> 565 540.2428   3.094950
#> 566 527.9128   2.847982
#> 567 500.0481   2.591321
#> 568 521.0250   2.500978
#> 569 500.0402   4.366183
#> 570 534.6909   3.234752
#> 571 507.8461   2.906642
#> 572 657.4308   3.117351
#> 573 500.3399   2.570110
#> 574 782.2888   2.573307
#> 575 554.6626   2.659611
#> 576 501.1471   2.574152
#> 577 500.2674   2.709044
#> 578 502.3299   2.501516
#> 579 500.1057   2.957454
#> 580 723.4964   2.524966
#> 581 501.3435   4.126023
#> 582 509.4094   2.807265
#> 583 573.4677   2.866166
#> 584 500.2071   2.686391
#> 585 500.3180   3.293554
#> 586 905.3649   2.590223
#> 587 501.5046   2.519624
#> 588 510.2177   2.631691
#> 589 500.2882   2.761977
#> 590 697.8428   3.210664
#> 591 666.0798   2.711419
#> 592 501.5274   4.239104
#> 593 501.6076   2.699143
#> 594 539.6883   2.500011
#> 595 611.8916   2.609289
#> 596 555.7409   2.539394
#> 597 917.5083   3.141323
#> 598 500.2408   2.540645
#> 599 501.8116   3.170035
#> 600 547.8043   4.153454
#> 601 690.7416   2.535823
#> 602 713.8558   2.752049
#> 603 505.6484   3.646253
#> 604 501.4136   2.704566
#> 605 500.5640   4.223560
#> 606 803.8662   2.617941
#> 607 501.5424   3.459877
#> 608 543.6400   2.594828
#> 609 676.4367   2.876535
#> 610 504.2548   2.726027
#> 611 556.4375   2.587641
#> 612 501.0167   3.145299
#> 613 500.8801   4.595628
#> 614 500.2771   3.185020
#> 615 502.7535   2.903354
#> 616 501.0135   3.455793
#> 617 500.2332   2.559140
#> 618 502.3995   2.612510
#> 619 500.1238   2.521448
#> 620 501.1948   2.596887
#> 621 500.3238   3.205371
#> 622 500.0602   2.836769
#> 623 536.0628   2.896881
#> 624 803.9596   2.526953
#> 625 500.0440   2.841646
#> 626 500.0030   2.787178
#> 627 502.5299   2.679106
#> 628 705.1154   2.700173
#> 629 500.1373   2.689418
#> 630 500.0669   2.908035
#> 631 511.2667   2.719670
#> 632 642.9262   2.621898
#> 633 500.1404   2.559933
#> 634 500.6244   2.575383
#> 635 520.1470   2.647493
#> 636 500.8547   2.516391
#> 637 501.3501   2.624120
#> 638 500.3536   3.259694
#> 639 527.2779   2.937054
#> 640 517.3096   2.547607
#> 641 516.1254   2.502776
#> 642 500.0787   2.879035
#> 643 500.0224   3.164276
#> 644 500.1481   2.519653
#> 645 690.3819   2.719218
#> 646 500.0820   2.954276
#> 647 502.5412   2.549334
#> 648 606.0367   2.541225
#> 649 519.2278   2.745783
#> 650 510.1812   2.513761
#> 651 500.0143   3.191166
#> 652 971.3042   2.942608
#> 653 500.7066   3.966039
#> 654 500.6694   2.539588
#> 655 500.5097   3.369378
#> 656 918.2243   2.640302
#> 657 501.2050   2.586594
#> 658 508.8138   3.165456
#> 659 502.6623   2.542460
#> 660 506.4522   2.609718
#> 661 500.1969   2.659318
#> 662 501.7717   2.938432
#> 663 500.0240   2.734548
#> 664 717.6621   3.016556
#> 665 500.5889   3.169243
#> 666 500.2831   2.509736
#> 667 504.6585   2.908222
#> 668 500.1117   2.826382
#> 669 520.2858   3.136599
#> 670 942.5422   2.759661
#> 671 500.0413   3.024504
#> 672 672.0985   3.945603
#> 673 977.3935   2.706353
#> 674 500.0154   3.441592
#> 675 737.0516   2.570083
#> 676 508.9011   2.735755
#> 677 526.8856   2.820687
#> 678 500.5229   2.832507
#> 679 715.6719   2.839827
#> 680 500.0344   2.921606
#> 681 500.1797   2.800493
#> 682 512.0106   2.671012
#> 683 500.1087   3.210133
#> 684 555.6886   2.554239
#> 685 645.4836   2.550613
#> 686 504.1529   2.804890
#> 687 509.1073   3.176405
#> 688 500.0640   2.637257
#> 689 500.3760   2.874432
#> 690 500.4469   3.050629
#> 691 524.5354   3.296431
#> 692 500.8963   3.130081
#> 693 501.5875   4.415895
#> 694 508.2939   2.985024
#> 695 500.2047   2.782097
#> 696 502.4273   2.977654
#> 697 500.1200   3.186229
#> 698 500.6188   2.770402
#> 699 500.0918   3.037458
#> 700 500.0010   2.560461
#> 701 931.2317   2.871499
#> 702 529.9218   3.104132
#> 703 500.0993   2.582840
#> 704 501.5389   2.578450
#> 705 519.5484   3.387874
#> 706 550.6660   2.627574
#> 707 518.1726   2.782132
#> 708 504.1626   2.504858
#> 709 500.0387   2.864660
#> 710 530.1000   3.445510
#> 711 514.3125   2.783381
#> 712 500.1528   3.593583
#> 713 869.0856   2.538372
#> 714 524.3312   3.119655
#> 715 610.3151   2.740518
#> 716 500.2422   3.153803
#> 717 502.2228   2.525915
#> 718 504.8836   2.944503
#> 719 502.6946   2.767992
#> 720 509.7802   2.734900
#> 721 500.0188   2.688031
#> 722 500.0207   2.710695
#> 723 500.1253   3.210592
#> 724 500.1491   2.745262
#> 725 790.1034   2.690358
#> 726 500.2045   2.509506
#> 727 504.1069   2.891496
#> 728 501.0002   4.149121
#> 729 504.8438   2.689761
#> 730 502.5818   2.940321
#> 731 500.0476   3.346374
#> 732 501.9039   3.745960
#> 733 570.7509   3.516188
#> 734 534.1547   3.424013
#> 735 500.2997   3.111041
#> 736 533.4127   2.958608
#> 737 504.0226   2.594800
#> 738 501.3504   3.202577
#> 739 500.2075   2.724883
#> 740 552.6776   3.419777
#> 741 611.4315   3.107739
#> 742 556.5311   2.520931
#> 743 512.7526   2.838974
#> 744 639.7401   2.974847
#> 745 504.8226   2.781225
#> 746 500.3026   2.884706
#> 747 559.9710   3.035585
#> 748 517.9195   2.776724
#> 749 517.8164   2.565356
#> 750 597.5738   2.877484
#> 751 521.9724   2.751911
#> 752 506.5306   2.915472
#> 753 504.7686   2.769088
#> 754 514.6487   2.627541
#> 755 500.0638   2.565224
#> 756 506.1201   2.977207
#> 757 988.0497   2.573942
#> 758 501.5826   2.638578
#> 759 504.3365   2.842625
#> 760 862.6221   2.680900
#> 761 508.8437   2.546139
#> 762 500.6696   2.512305
#> 763 501.8773   3.088706
#> 764 566.1114   3.240465
#> 765 664.4527   2.616901
#> 766 551.1947   3.597764
#> 767 500.5139   3.336093
#> 768 523.5061   2.710436
#> 769 504.6540   3.298069
#> 770 500.1127   3.262231
#> 771 567.3575   3.046163
#> 772 540.5833   3.384310
#> 773 504.3683   2.723704
#> 774 500.3995   3.933615
#> 775 656.7331   2.908596
#> 776 501.8514   2.539342
#> 777 626.5266   2.760079
#> 778 501.5216   2.507741
#> 779 500.7745   2.761734
#> 780 501.2984   3.025611
#> 781 513.6824   3.291989
#> 782 500.9908   2.685012
#> 783 500.1222   3.090577
#> 784 894.9946   3.169278
#> 785 511.9581   2.912082
#> 786 534.4297   2.982561
#> 787 862.5818   2.582431
#> 788 506.0221   2.593428
#> 789 558.1448   2.538515
#> 790 502.4155   2.560921
#> 791 500.0449   3.752559
#> 792 556.7502   2.764445
#> 793 502.7878   3.830910
#> 794 500.1240   2.828751
#> 795 515.9427   2.597133
#> 796 502.7472   3.501277
#> 797 500.5581   2.811521
#> 798 505.1369   2.854764
#> 799 509.9912   2.596760
#> 800 502.2850   2.580712
#> 801 505.8961   2.505559
#> 802 509.4686   2.964685
#> 803 573.1978   2.847543
#> 804 502.0944   3.521612
#> 805 587.1580   3.049958
#> 806 503.5974   2.611829
#> 807 501.5803   2.894672
#> 808 503.1488   2.536781
#> 809 501.8615   2.691767
#> 810 505.1078   2.790409
#> 811 501.4163   2.801762
#> 812 500.0176   4.033728
#> 813 503.2339   2.689712
#> 814 501.7573   2.913853
#> 815 501.6902   3.054398
#> 816 500.1473   2.825268
#> 817 503.9346   3.615834
#> 818 578.4207   3.187031
#> 819 500.3415   3.610616
#> 820 501.5153   3.807373
#> 821 500.2261   2.726142
#> 822 500.2035   2.593669
#> 823 933.6165   3.003087
#> 824 515.1068   2.674615
#> 825 504.1129   2.960478
#> 826 611.7106   2.924082
#> 827 501.3312   3.125059
#> 828 525.3309   2.601410
#> 829 500.1168   2.840812
#> 830 598.2100   2.664777
#> 831 548.4067   3.172200
#> 832 502.0608   2.691302
#> 833 500.3044   3.365818
#> 834 500.5700   3.254912
#> 835 571.6978   3.392151
#> 836 500.0393   2.822522
#> 837 643.7961   3.744504
#> 838 500.3306   2.518499
#> 839 634.0375   2.598788
#> 840 500.3291   2.520771
#> 841 500.0153   2.502127
#> 842 500.0002   3.199285
#> 843 500.4226   2.542629
#> 844 501.4498   2.748560
#> 845 670.2582   2.978067
#> 846 504.8103   2.630030
#> 847 507.4411   3.971300
#> 848 569.1616   2.649478
#> 849 502.4320   2.621912
#> 850 634.6652   3.933849
#> 851 593.2056   3.807041
#> 852 503.9478   3.005235
#> 853 727.9944   4.666413
#> 854 500.3803   3.349603
#> 855 500.5777   3.062365
#> 856 672.0306   2.618349
#> 857 507.1787   2.603983
#> 858 525.5183   2.581847
#> 859 501.7312   2.891638
#> 860 504.0480   2.530890
#> 861 571.9365   2.941017
#> 862 500.2572   3.559293
#> 863 500.7291   2.746064
#> 864 575.8866   2.737101
#> 865 510.2501   2.992432
#> 866 502.5139   2.732480
#> 867 500.0425   2.519143
#> 868 504.0947   2.623899
#> 869 500.0443   2.915138
#> 870 522.4903   2.620887
#> 871 501.0109   2.515213
#> 872 927.9793   2.739662
#> 873 500.2894   3.496267
#> 874 652.8214   2.947681
#> 875 502.2774   2.884950
#> 876 517.0569   2.744509
#> 877 541.8148   2.656677
#> 878 500.1418   2.567297
#> 879 500.3200   3.228959
#> 880 506.6104   2.550771
#> 881 521.6998   2.657589
#> 882 503.3318   3.084272
#> 883 505.5759   2.507153
#> 884 508.5472   2.873496
#> 885 505.4851   2.680918
#> 886 500.2263   2.639372
#> 887 500.3305   2.621501