Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   586.1351   3.078901
#> 2   507.2616   2.615871
#> 3   509.7462   3.331023
#> 4   501.3778   2.707263
#> 5   508.7203   3.097143
#> 6   815.4582   2.960785
#> 7   500.0385   2.604602
#> 8   500.2156   2.677202
#> 9   500.0492   2.901954
#> 10  837.7634   2.556174
#> 11  508.6822   2.671069
#> 12  500.0427   2.890564
#> 13  558.3574   2.618547
#> 14  709.1590   3.783011
#> 15  524.7821   2.501311
#> 16  511.1372   2.772285
#> 17  740.1210   2.819158
#> 18  500.8008   2.595965
#> 19  509.7031   3.263621
#> 20  501.3130   2.909702
#> 21  500.6205   2.674216
#> 22  504.8053   4.052732
#> 23  500.6917   2.754023
#> 24  504.6834   2.572486
#> 25  502.5891   3.188070
#> 26  503.5360   2.916042
#> 27  500.2752   3.254216
#> 28  512.8145   2.854787
#> 29  500.2168   2.518873
#> 30  611.2079   2.915786
#> 31  500.0894   2.674046
#> 32  588.8077   2.743511
#> 33  500.3875   3.632632
#> 34  500.9772   2.938008
#> 35  803.7002   3.253150
#> 36  500.3809   4.538229
#> 37  500.1903   2.510424
#> 38  503.2132   4.493583
#> 39  507.3494   2.523728
#> 40  500.2061   2.575163
#> 41  506.6551   3.138012
#> 42  521.7540   2.751502
#> 43  500.0348   2.510080
#> 44  501.0459   2.753868
#> 45  500.2229   2.757955
#> 46  502.2716   2.856183
#> 47  523.6030   2.700888
#> 48  514.0793   3.213601
#> 49  500.1680   3.014806
#> 50  896.4646   2.841937
#> 51  502.4420   4.945330
#> 52  505.2141   3.195618
#> 53  534.3341   2.822876
#> 54  620.4372   2.701922
#> 55  501.6320   3.189115
#> 56  502.4885   2.839424
#> 57  505.8563   2.776398
#> 58  501.0196   2.608881
#> 59  604.8043   2.940078
#> 60  514.0961   2.507659
#> 61  818.9745   2.670651
#> 62  502.7960   3.265420
#> 63  500.0965   2.635730
#> 64  528.2159   2.661177
#> 65  782.8671   2.707487
#> 66  728.5844   3.181470
#> 67  870.8442   3.528840
#> 68  834.2972   2.627080
#> 69  503.3368   2.700919
#> 70  501.0949   2.538871
#> 71  500.3718   2.689769
#> 72  504.4252   2.734027
#> 73  510.8201   2.768086
#> 74  502.5329   2.696385
#> 75  965.1315   2.669819
#> 76  528.4922   2.523410
#> 77  500.0773   4.038089
#> 78  968.2071   2.948810
#> 79  501.1301   2.564383
#> 80  503.2984   2.523133
#> 81  500.1966   2.659899
#> 82  500.5504   3.192049
#> 83  504.5127   2.548582
#> 84  500.2481   2.504608
#> 85  510.3728   2.507143
#> 86  501.2986   2.685779
#> 87  509.6012   2.531102
#> 88  501.0389   2.501688
#> 89  503.7633   2.683400
#> 90  612.5575   2.789310
#> 91  504.0612   2.831217
#> 92  500.5258   3.043476
#> 93  524.0180   2.564474
#> 94  501.8275   2.619413
#> 95  500.2372   2.678989
#> 96  705.4326   3.515311
#> 97  908.1433   3.399977
#> 98  561.2399   2.746886
#> 99  500.3727   2.642668
#> 100 500.3619   3.449827
#> 101 500.2790   2.930025
#> 102 501.7170   2.574876
#> 103 564.5456   2.854160
#> 104 500.0803   2.711520
#> 105 501.8545   2.778438
#> 106 549.2612   2.507860
#> 107 513.5929   2.549434
#> 108 500.1823   2.541342
#> 109 568.1828   2.533452
#> 110 502.4159   3.558643
#> 111 509.1280   2.547491
#> 112 505.6657   2.612313
#> 113 504.6154   2.949322
#> 114 500.5459   3.471939
#> 115 505.4803   2.729903
#> 116 898.2303   2.552555
#> 117 501.0139   3.109185
#> 118 759.0251   2.781416
#> 119 502.1739   2.548852
#> 120 500.1333   2.711017
#> 121 501.6792   2.805836
#> 122 500.1460   3.492388
#> 123 504.4871   2.607767
#> 124 528.6402   2.871887
#> 125 500.4095   2.779328
#> 126 635.8264   3.066203
#> 127 505.2183   2.955782
#> 128 776.6528   2.529611
#> 129 512.9158   2.844257
#> 130 500.9649   2.892653
#> 131 507.4356   2.715198
#> 132 851.3208   2.958952
#> 133 502.0771   3.923381
#> 134 501.2512   2.576206
#> 135 500.1759   2.854757
#> 136 500.0701   2.691423
#> 137 967.9712   3.161608
#> 138 874.2685   3.701168
#> 139 502.4649   3.741000
#> 140 514.4839   3.153662
#> 141 597.5864   3.173798
#> 142 500.1436   2.640044
#> 143 500.3664   3.141578
#> 144 500.3608   2.976153
#> 145 500.1547   3.692891
#> 146 514.2354   2.886792
#> 147 500.0085   2.503270
#> 148 504.5489   2.692823
#> 149 576.6927   2.574643
#> 150 501.4736   2.962701
#> 151 516.0242   3.122641
#> 152 500.0385   2.761333
#> 153 500.0601   2.606284
#> 154 524.3420   2.514719
#> 155 501.1186   4.169050
#> 156 552.4934   2.789978
#> 157 501.3679   2.868103
#> 158 501.9790   2.685246
#> 159 507.6031   3.461806
#> 160 501.7871   3.793439
#> 161 531.1755   2.841641
#> 162 507.5559   4.458394
#> 163 500.2572   2.507263
#> 164 515.6517   2.939060
#> 165 874.1047   2.652616
#> 166 513.6478   4.812835
#> 167 500.6004   3.188957
#> 168 500.1432   2.854746
#> 169 500.2183   2.544705
#> 170 500.4139   3.194200
#> 171 500.4730   2.765679
#> 172 501.7466   2.545200
#> 173 503.4599   2.971887
#> 174 591.5395   2.648562
#> 175 512.2621   2.735361
#> 176 500.0433   2.973293
#> 177 500.8171   2.934626
#> 178 543.1535   2.999438
#> 179 504.8577   2.582166
#> 180 500.2384   2.511274
#> 181 502.1068   2.876764
#> 182 504.4952   3.187292
#> 183 500.0759   3.150005
#> 184 573.3414   2.889390
#> 185 523.6784   3.967107
#> 186 588.9437   2.557320
#> 187 631.0617   2.853826
#> 188 500.0831   3.154965
#> 189 717.7514   2.835334
#> 190 510.9089   2.961108
#> 191 553.2941   2.864654
#> 192 801.7231   3.318722
#> 193 502.6572   2.960668
#> 194 503.9993   2.653338
#> 195 559.7277   2.693925
#> 196 502.7263   2.519785
#> 197 500.0082   2.612376
#> 198 501.1648   3.778900
#> 199 636.7719   2.713460
#> 200 500.2430   2.679054
#> 201 500.0360   2.701369
#> 202 500.1165   2.752842
#> 203 837.4525   2.955151
#> 204 502.0379   2.575378
#> 205 500.3700   2.770193
#> 206 501.5925   2.851283
#> 207 500.6606   3.623615
#> 208 500.9117   2.839863
#> 209 505.8823   3.237495
#> 210 576.2332   2.916852
#> 211 502.3835   3.964946
#> 212 501.0424   2.521431
#> 213 500.2221   2.767010
#> 214 526.4303   2.579638
#> 215 779.9414   2.731122
#> 216 501.6365   3.043245
#> 217 507.3030   3.676758
#> 218 502.3423   4.033321
#> 219 501.0421   2.539175
#> 220 635.2112   2.589874
#> 221 784.4016   3.106902
#> 222 605.7727   2.608157
#> 223 579.3194   2.614727
#> 224 500.9461   2.562385
#> 225 506.9632   2.824140
#> 226 517.5744   2.616750
#> 227 514.8495   3.007114
#> 228 500.1446   3.082550
#> 229 500.1898   3.086836
#> 230 500.0092   3.287598
#> 231 581.6940   2.632486
#> 232 528.3345   2.757971
#> 233 500.0483   2.623767
#> 234 500.3763   3.322669
#> 235 500.2122   2.981598
#> 236 500.3339   2.536786
#> 237 514.1294   3.003538
#> 238 503.9975   3.689754
#> 239 506.4056   6.184881
#> 240 500.7298   2.663216
#> 241 501.8289   2.583679
#> 242 502.4344   2.607008
#> 243 503.8084   3.444204
#> 244 543.0748   2.745765
#> 245 505.2561   2.711331
#> 246 500.1634   2.528982
#> 247 559.2727   2.551282
#> 248 522.8150   2.619162
#> 249 523.8775   2.945246
#> 250 500.0271   2.860167
#> 251 523.8775   2.922515
#> 252 500.2703   4.087689
#> 253 885.4249   4.431211
#> 254 869.9390   2.716959
#> 255 501.3618   2.512658
#> 256 500.2245   2.638880
#> 257 514.4570   2.796307
#> 258 500.1342   3.673625
#> 259 500.0199   2.769200
#> 260 506.8406   3.629346
#> 261 502.5350   2.876277
#> 262 506.2884   3.411338
#> 263 501.9622   2.828575
#> 264 824.4892   3.315745
#> 265 512.0883   2.605536
#> 266 733.0170   3.121500
#> 267 509.2430   3.086689
#> 268 504.4700   2.841703
#> 269 500.6359   3.684543
#> 270 500.5675   2.760228
#> 271 500.2860   3.010814
#> 272 572.4085   2.733230
#> 273 500.1852   2.985794
#> 274 500.1594   4.073459
#> 275 506.2508   2.507495
#> 276 503.7825   2.520600
#> 277 500.1920   3.078919
#> 278 502.5237   3.351211
#> 279 623.1092   2.503874
#> 280 500.6641   3.077725
#> 281 552.4717   2.504898
#> 282 503.8849   2.511182
#> 283 501.3998   3.119366
#> 284 500.1475   2.721222
#> 285 503.9743   2.560450
#> 286 501.3316   2.893289
#> 287 500.8108   2.516182
#> 288 635.1923   2.533762
#> 289 500.6892   3.285066
#> 290 503.0517   3.142417
#> 291 529.4653   2.679392
#> 292 500.0452   3.233068
#> 293 819.1336   2.539445
#> 294 505.7048   2.765264
#> 295 502.3001   3.828508
#> 296 540.8600   2.597962
#> 297 500.5107   2.532320
#> 298 516.8822   3.353365
#> 299 500.1383   2.790797
#> 300 500.0716   2.535376
#> 301 500.8742   2.762406
#> 302 506.5054   2.984417
#> 303 500.6974   3.159139
#> 304 519.5698   2.849105
#> 305 500.0283   2.551110
#> 306 505.8574   3.505297
#> 307 506.7406   2.815208
#> 308 539.0977   4.364995
#> 309 501.7063   3.206036
#> 310 676.4879   3.713534
#> 311 501.7335   3.094106
#> 312 966.9166   2.720742
#> 313 593.9955   2.649891
#> 314 757.5463   4.006770
#> 315 504.2652   3.189584
#> 316 500.4503   2.790263
#> 317 511.3742   2.747765
#> 318 599.1658   2.968429
#> 319 528.3818   2.717956
#> 320 524.8261   2.760314
#> 321 643.4634   2.865596
#> 322 500.3793   3.331900
#> 323 500.1824   2.755366
#> 324 500.4420   3.082148
#> 325 507.0522   2.890323
#> 326 531.3182   4.151661
#> 327 739.5947   2.569308
#> 328 503.4894   3.263042
#> 329 502.7478   2.537692
#> 330 505.7155   2.813973
#> 331 500.0337   2.763591
#> 332 500.7907   3.170550
#> 333 501.3507   2.902715
#> 334 500.0251   2.525925
#> 335 500.0664   3.137096
#> 336 500.0542   4.859494
#> 337 511.4822   2.861377
#> 338 500.2825   3.789197
#> 339 503.0968   2.908245
#> 340 624.9090   2.621119
#> 341 501.0949   3.058022
#> 342 501.5343   2.689385
#> 343 583.5050   2.569098
#> 344 524.8348   2.843992
#> 345 506.9017   2.683952
#> 346 502.8422   4.271822
#> 347 695.7592   2.748278
#> 348 500.3094   2.978832
#> 349 507.7408   2.763584
#> 350 503.2678   2.505433
#> 351 500.1014   2.662685
#> 352 500.0730   2.930204
#> 353 502.7124   2.837335
#> 354 502.0790   3.104821
#> 355 516.5933   3.044072
#> 356 500.1489   2.514595
#> 357 906.7698   3.104471
#> 358 505.4508   2.655699
#> 359 501.2100   2.966017
#> 360 506.1342   2.892394
#> 361 516.2767   3.081191
#> 362 507.6500   2.806647
#> 363 501.2829   2.730453
#> 364 515.1740   3.046665
#> 365 838.9236   3.490801
#> 366 501.4252   2.813555
#> 367 501.5325   2.566508
#> 368 504.2748   2.785606
#> 369 500.2185   2.587618
#> 370 500.1715   2.661906
#> 371 500.7366   2.897446
#> 372 536.3278   2.707663
#> 373 502.1008   3.079759
#> 374 621.2845   2.885255
#> 375 856.8290   3.012604
#> 376 636.8527   3.110698
#> 377 500.4636   2.974466
#> 378 586.1764   3.224538
#> 379 501.0968   2.670217
#> 380 502.4838   2.836722
#> 381 500.0337   3.585023
#> 382 500.1799   2.800153
#> 383 500.0016   3.081635
#> 384 500.0220   2.698249
#> 385 509.1504   2.862813
#> 386 527.7779   2.504062
#> 387 505.2776   2.938520
#> 388 500.0739   2.606164
#> 389 502.2849   3.550274
#> 390 618.5534   2.740987
#> 391 503.9535   3.706773
#> 392 510.3715   2.737540
#> 393 528.9495   2.519417
#> 394 665.8859   3.568551
#> 395 512.3825   3.501750
#> 396 501.0877   2.835824
#> 397 501.0332   2.754975
#> 398 515.3705   2.629550
#> 399 500.4999   2.528819
#> 400 536.3544   4.748109
#> 401 529.5905   3.231069
#> 402 500.5823   3.676416
#> 403 500.4820   3.210867
#> 404 501.3854   3.225282
#> 405 529.5556   4.760155
#> 406 543.3416   3.086899
#> 407 500.0272   2.891829
#> 408 585.7273   2.570477
#> 409 500.5119   2.895285
#> 410 738.3245   2.661500
#> 411 503.0296   2.756130
#> 412 548.7249   2.666520
#> 413 500.0866   2.531613
#> 414 979.6910   2.633581
#> 415 501.4771   2.507399
#> 416 505.3379   2.677934
#> 417 520.2688   3.568818
#> 418 549.2443   2.560083
#> 419 581.7643   2.882143
#> 420 500.3195   2.915690
#> 421 513.8416   2.654466
#> 422 516.3802   2.604588
#> 423 540.1130   2.544536
#> 424 559.3334   3.919789
#> 425 772.2233   2.663699
#> 426 500.0214   2.723607
#> 427 534.8359   2.677405
#> 428 500.2519   2.520874
#> 429 504.2510   2.935161
#> 430 500.0197   2.574762
#> 431 500.7233   2.868484
#> 432 514.0994   2.728366
#> 433 500.0562   3.061836
#> 434 510.1452   2.562576
#> 435 500.9594   4.034282
#> 436 500.2354   2.779139
#> 437 508.9581   2.735607
#> 438 500.0313   2.828505
#> 439 500.3809   2.531315
#> 440 525.7342   2.519954
#> 441 505.0565   3.407844
#> 442 553.0141   3.344699
#> 443 509.2743   2.835213
#> 444 617.1423   3.189016
#> 445 500.2678   2.666473
#> 446 500.5261   2.579396
#> 447 546.1251   2.674686
#> 448 501.2215   2.693592
#> 449 502.4297   2.788543
#> 450 500.7376   2.895394
#> 451 501.4565   2.613080
#> 452 523.1941   2.694959
#> 453 510.5310   3.245052
#> 454 500.8249   3.877887
#> 455 503.1145   2.664623
#> 456 500.0114   3.399481
#> 457 793.1992   2.505184
#> 458 635.4767   3.103384
#> 459 600.6205   2.653727
#> 460 527.5265   2.622294
#> 461 501.4334   2.537388
#> 462 500.1793   3.806350
#> 463 718.8296   4.015525
#> 464 500.0301   2.861549
#> 465 530.3787   3.053940
#> 466 777.5203   3.156221
#> 467 500.2171   2.575441
#> 468 515.4579   2.880036
#> 469 552.2971   2.532402
#> 470 500.0008   3.631506
#> 471 500.6433   2.515892
#> 472 504.2352   2.576233
#> 473 501.1988   3.254773
#> 474 877.6419   2.565684
#> 475 500.2608   2.644273
#> 476 505.2783   3.481024
#> 477 504.4579   2.895756
#> 478 553.6632   3.255011
#> 479 501.3913   2.873025
#> 480 500.3048   4.462433
#> 481 526.7596   2.816914
#> 482 515.6591   2.635698
#> 483 501.5416   2.559423
#> 484 500.0278   2.747633
#> 485 500.4644   2.793087
#> 486 500.3581   2.648137
#> 487 500.7878   2.887850
#> 488 585.7443   2.636682
#> 489 739.3140   2.569768
#> 490 503.9125   2.844696
#> 491 506.2450   2.615213
#> 492 510.0200   3.036975
#> 493 500.0757   2.699002
#> 494 613.2151   2.932556
#> 495 501.5226   3.262960
#> 496 509.4631   3.338639
#> 497 500.0356   3.815689
#> 498 711.5320   3.150059
#> 499 502.3666   3.029371
#> 500 596.9607   3.262122
#> 501 507.8217   2.612088
#> 502 500.0633   2.516077
#> 503 863.8432   2.978844
#> 504 500.1058   3.340310
#> 505 705.1505   2.649992
#> 506 524.9546   2.507068
#> 507 502.1069   2.521275
#> 508 500.2238   2.873757
#> 509 522.6722   2.770075
#> 510 504.1663   2.911124
#> 511 500.2025   2.616079
#> 512 500.9723   3.398347
#> 513 501.1923   2.680358
#> 514 696.1640   2.694619
#> 515 717.2723   2.981205
#> 516 501.2136   3.465604
#> 517 514.5159   2.755871
#> 518 576.2393   3.750235
#> 519 500.9980   3.050939
#> 520 998.0218   3.122455
#> 521 521.9554   2.690788
#> 522 505.5566   3.564821
#> 523 501.0141   2.518444
#> 524 500.4961   2.623958
#> 525 504.6041   3.096108
#> 526 500.9218   3.471863
#> 527 507.8334   2.757377
#> 528 500.7123   3.095067
#> 529 500.9144   2.678667
#> 530 513.2953   2.888550
#> 531 593.8642   2.618606
#> 532 537.5211   2.762363
#> 533 539.8652   2.998957
#> 534 500.9952   3.558435
#> 535 500.4668   2.955582
#> 536 534.7581   3.671146
#> 537 637.9989   2.559854
#> 538 500.0541   3.456892
#> 539 650.8823   2.844736
#> 540 527.3849   2.666039
#> 541 500.8946   2.508434
#> 542 532.4072   3.132076
#> 543 572.7850   2.633600
#> 544 731.4611   3.325905
#> 545 721.1484   3.062461
#> 546 500.1103   2.684892
#> 547 661.1474   2.855537
#> 548 500.8196   2.518731
#> 549 501.0748   2.601104
#> 550 500.1499   3.220238
#> 551 805.6277   3.419967
#> 552 562.9555   2.814769
#> 553 729.9178   2.993588
#> 554 501.7442   3.387982
#> 555 529.2534   3.239370
#> 556 511.2204   2.967553
#> 557 501.6934   2.998616
#> 558 503.8847   3.573016
#> 559 500.1993   3.337830
#> 560 500.2890   2.960587
#> 561 500.8024   2.831410
#> 562 992.3681   2.826727
#> 563 957.8908   2.686925
#> 564 844.9428   2.859264
#> 565 503.8731   2.670911
#> 566 524.3903   2.529678
#> 567 555.1334   2.635288
#> 568 501.2349   2.767328
#> 569 500.0032   3.325734
#> 570 504.2753   2.591504
#> 571 505.9374   2.682777
#> 572 500.4171   3.523620
#> 573 500.1229   3.460054
#> 574 500.4868   2.875837
#> 575 554.4251   3.202457
#> 576 500.1219   2.742470
#> 577 505.2044   2.818283
#> 578 515.9905   2.931172
#> 579 500.4919   3.514161
#> 580 500.3103   2.823192
#> 581 500.0546   3.692496
#> 582 504.0417   2.512192
#> 583 500.0032   3.764118
#> 584 500.3316   2.684418
#> 585 501.0463   2.533572
#> 586 501.7624   2.548650
#> 587 542.0871   2.908874
#> 588 501.8065   2.593948
#> 589 500.2380   3.909248
#> 590 501.2151   2.954679
#> 591 554.6214   2.622187
#> 592 500.0852   3.044756
#> 593 500.5973   3.002793
#> 594 744.3285   3.078785
#> 595 500.0290   2.607471
#> 596 502.6030   2.943533
#> 597 546.2431   3.506011
#> 598 502.4985   3.180361
#> 599 502.6426   3.215116
#> 600 506.4887   3.106843
#> 601 500.4160   2.998058
#> 602 591.5786   2.687047
#> 603 763.4328   2.782533
#> 604 508.6343   2.819611
#> 605 500.0502   3.508127
#> 606 502.8752   2.626425
#> 607 511.1640   2.790358
#> 608 500.0158   2.737327
#> 609 957.4665   3.072148
#> 610 502.8723   2.662494
#> 611 673.3174   2.594520
#> 612 500.0238   2.810642
#> 613 501.2965   2.656565
#> 614 582.6752   4.173765
#> 615 500.1106   2.964082
#> 616 500.0285   3.065064
#> 617 501.3025   2.945697
#> 618 500.1418   2.559953
#> 619 522.5887   2.887796
#> 620 503.6936   2.707769
#> 621 500.4415   2.583448
#> 622 500.0452   2.838546
#> 623 503.8739   2.679007
#> 624 530.0389   2.683305
#> 625 500.1553   2.551945
#> 626 559.7593   2.537564
#> 627 501.1437   2.944603
#> 628 572.0699   3.026287
#> 629 534.1708   2.508050
#> 630 514.7887   2.635935
#> 631 508.9842   2.667377
#> 632 731.6679   5.675496
#> 633 540.5717   2.581173
#> 634 535.7822   2.753308
#> 635 524.0961   3.191878
#> 636 539.6764   2.972677
#> 637 500.1212   3.370685
#> 638 517.6712   4.682648
#> 639 508.4446   2.992854
#> 640 502.5481   3.113493
#> 641 551.9496   3.297909
#> 642 507.8576   3.075010
#> 643 500.0065   2.942468
#> 644 501.1301   2.930228
#> 645 544.4870   3.603322
#> 646 531.2369   3.224366
#> 647 621.5976   2.672597
#> 648 501.9657   3.059638
#> 649 500.2739   2.586699
#> 650 500.5288   2.504971
#> 651 501.5522   2.804096
#> 652 500.0308   2.512362
#> 653 507.9065   2.831854
#> 654 503.7744   2.648704
#> 655 506.5448   2.535669
#> 656 500.0256   2.583978
#> 657 508.8925   3.166351
#> 658 648.8743   2.604699
#> 659 664.8098   2.741017
#> 660 555.6436   2.898049
#> 661 501.2732   2.749839
#> 662 503.2555   2.544813
#> 663 507.3170   2.872068
#> 664 501.2722   2.708230
#> 665 503.5767   2.884012
#> 666 510.0509   2.771107
#> 667 500.0910   2.631172
#> 668 509.1037   3.162899
#> 669 505.7335   2.669790
#> 670 650.4199   3.036751
#> 671 500.0953   2.700502
#> 672 500.3684   3.834993
#> 673 516.3798   2.541429
#> 674 503.5033   3.270144
#> 675 602.4796   2.943206
#> 676 559.7706   3.073207
#> 677 500.0096   3.157934
#> 678 658.0510   2.725448
#> 679 530.7105   2.657305
#> 680 500.2404   2.701291
#> 681 500.2655   2.765680
#> 682 502.6161   3.243793
#> 683 500.7143   2.537118
#> 684 500.9089   2.817084
#> 685 503.2297   3.019803
#> 686 537.8136   3.495466
#> 687 628.6830   2.617505
#> 688 640.8525   2.834810
#> 689 500.0064   2.524151
#> 690 528.0558   2.902727
#> 691 500.2627   4.317616
#> 692 881.9011   2.653606
#> 693 516.6295   3.243257
#> 694 500.3183   2.737139
#> 695 500.6121   2.826954
#> 696 505.1216   2.723525
#> 697 500.0885   2.615452
#> 698 500.0703   2.845312
#> 699 500.3951   2.723714
#> 700 500.1539   3.435241
#> 701 613.3345   2.868811
#> 702 500.1356   2.514196
#> 703 551.8996   2.513239
#> 704 505.6148   3.239121
#> 705 501.9395   2.821416
#> 706 500.2570   2.797868
#> 707 742.9916   2.984013
#> 708 807.9693   2.694254
#> 709 530.7514   3.208529
#> 710 553.3011   2.957496
#> 711 504.7058   2.582346
#> 712 501.3914   2.540370
#> 713 500.1562   3.652393
#> 714 500.2040   2.876011
#> 715 502.3469   3.356525
#> 716 540.8144   2.842235
#> 717 606.7612   2.949626
#> 718 509.0236   2.984259
#> 719 506.9035   2.508212
#> 720 500.6705   2.613648
#> 721 501.3802   2.831938
#> 722 500.1999   2.701559
#> 723 530.4225   2.640676
#> 724 584.0856   2.549911
#> 725 500.5399   2.784451
#> 726 547.2419   4.180711
#> 727 544.3272   4.478390
#> 728 510.1821   2.532583
#> 729 500.4818   3.207213
#> 730 506.6930   3.868518
#> 731 504.7675   3.106658
#> 732 792.6584   2.523471
#> 733 500.5565   2.612104
#> 734 572.2280   3.022363
#> 735 500.1059   2.512404
#> 736 501.9838   2.877640
#> 737 506.9867   3.936984
#> 738 500.1069   3.826417
#> 739 500.0894   2.722569
#> 740 501.8842   2.837053
#> 741 836.2196   2.934024
#> 742 503.9574   2.834200
#> 743 603.6467   3.199242
#> 744 615.3519   2.646108
#> 745 556.1701   3.036946
#> 746 500.0007   2.517294
#> 747 506.9832   2.651971
#> 748 507.0098   2.760767
#> 749 534.1281   3.197123
#> 750 592.4102   4.917398
#> 751 501.6296   2.827584
#> 752 500.0930   2.519426
#> 753 500.1026   3.441956
#> 754 500.2415   2.950339
#> 755 511.7907   2.643941
#> 756 500.8147   2.943228
#> 757 500.0730   2.784290
#> 758 544.5228   2.692335
#> 759 520.0332   2.504026
#> 760 501.2233   2.751402
#> 761 524.8576   3.040127
#> 762 651.8324   2.750887
#> 763 502.9099   2.562186
#> 764 500.5836   2.971569
#> 765 535.1030   2.575395
#> 766 500.9621   2.524692
#> 767 521.0377   3.504492
#> 768 503.5673   3.283102
#> 769 510.9389   2.520153
#> 770 514.1559   3.280576
#> 771 500.6554   2.859617
#> 772 502.3221   3.305257
#> 773 500.3078   2.853347
#> 774 500.3544   2.791091
#> 775 502.8661   3.415361
#> 776 502.8949   2.727731
#> 777 651.0680   2.604059
#> 778 506.3168   2.820740
#> 779 509.6622   2.507341
#> 780 503.3693   2.794535
#> 781 500.7832   2.989337
#> 782 576.3730   2.660255
#> 783 501.7159   4.119804
#> 784 500.0069   3.191587
#> 785 500.1032   2.571923
#> 786 500.2565   2.919512
#> 787 501.1365   2.707916
#> 788 625.7768   2.671206
#> 789 501.4101   2.634164
#> 790 500.0562   2.787429
#> 791 669.2551   2.716095
#> 792 539.0196   2.578113
#> 793 500.4355   2.809377
#> 794 501.1185   2.746008
#> 795 500.0902   2.644007
#> 796 500.2050   2.994645
#> 797 581.6604   2.973452
#> 798 500.2754   3.168738
#> 799 512.8829   3.104304
#> 800 504.4696   2.776610
#> 801 548.9797   2.642837
#> 802 500.2538   2.790195
#> 803 582.0011   3.254674
#> 804 577.0659   2.539966
#> 805 500.1092   2.547168
#> 806 565.8045   3.072539
#> 807 504.1123   2.643690
#> 808 501.2791   2.905614
#> 809 500.1227   3.177521
#> 810 500.8197   2.727045
#> 811 500.0130   2.934942
#> 812 535.9617   3.007961
#> 813 502.2242   2.666431
#> 814 500.6936   2.852721
#> 815 502.7162   3.417218
#> 816 506.2064   3.029166
#> 817 551.3434   3.197525
#> 818 501.2588   2.885028
#> 819 500.0586   3.662817
#> 820 577.0059   2.857567
#> 821 502.3513   2.652410
#> 822 536.4824   2.778500
#> 823 502.0210   2.854226
#> 824 500.8919   2.735191
#> 825 500.1406   2.593833
#> 826 553.3157   2.860747
#> 827 502.1010   2.994420
#> 828 500.3033   3.377101
#> 829 527.4300   2.678331
#> 830 642.9375   2.537138
#> 831 510.9435   3.024477
#> 832 524.0591   2.738537
#> 833 543.8436   3.219853
#> 834 514.5366   2.750718
#> 835 503.3930   2.787576
#> 836 500.1739   2.683848
#> 837 500.3352   2.972153
#> 838 506.2991   3.017788
#> 839 663.1686   3.033759
#> 840 500.2180   2.510269
#> 841 720.8149   2.748601
#> 842 501.1520   3.628980
#> 843 506.6626   2.570789
#> 844 513.9279   2.638268
#> 845 739.3958   2.738186
#> 846 541.0498   3.543326
#> 847 615.7133   2.546435
#> 848 679.0203   2.584491
#> 849 526.8359   2.651934
#> 850 500.1837   2.823620
#> 851 527.3545   2.624969
#> 852 511.0762   2.856883
#> 853 500.0750   2.931684
#> 854 612.6838   2.506299
#> 855 504.6517   2.561539
#> 856 647.9659   2.888428
#> 857 510.0038   2.644051
#> 858 517.5133   4.477417
#> 859 537.1250   2.555003
#> 860 504.0440   2.525064
#> 861 681.4535   2.977099
#> 862 503.5683   2.739668
#> 863 501.6396   2.582605
#> 864 500.0336   2.812903
#> 865 601.6688   2.629676
#> 866 515.2400   2.605154
#> 867 502.4130   2.526782
#> 868 508.9827   2.551761
#> 869 509.8450   2.787412
#> 870 899.7265   2.726504
#> 871 500.7160   2.936787
#> 872 516.4142   3.154768
#> 873 505.7583   2.735609
#> 874 500.2222   2.836394
#> 875 521.2001   2.582727
#> 876 649.3987   2.568136
#> 877 500.1918   2.908333
#> 878 500.1383   2.780672
#> 879 575.1785   3.550765
#> 880 500.1402   2.821811
#> 881 500.3397   2.803274
#> 882 500.0026   3.425670
#> 883 500.3242   3.563092
#> 884 523.5534   2.535192
#> 885 505.6933   2.763624
#> 886 561.8768   2.531797
#> 887 500.3742   2.950658
#> 888 560.5531   3.117741
#> 889 500.5511   2.505881
#> 890 540.9351   2.618269
#> 891 500.1104   2.525512
#> 892 718.7101   2.600786
#> 893 504.9270   2.839842
#> 894 500.1022   2.988184
#> 895 500.4554   2.957465
#> 896 602.1638   2.622042
#> 897 500.2283   2.849006
#> 898 500.1580   3.238423
#> 899 502.4766   2.659269
#> 900 511.2500   2.698495
#> 901 500.6482   2.997799
#> 902 551.5212   2.654168
#> 903 501.7361   2.516934
#> 904 552.6891   3.142052
#> 905 500.0494   2.644714
#> 906 557.1905   3.257202
#> 907 504.1956   2.762158
#> 908 501.6901   2.898656
#> 909 500.2182   3.587502
#> 910 502.7590   2.512951
#> 911 530.7604   3.312605
#> 912 590.2878   3.952176
#> 913 500.6673   3.036323
#> 914 582.5550   3.768443
#> 915 500.2326   2.904037
#> 916 505.4099   4.110807
#> 917 536.6280   2.545463
#> 918 500.2999   3.363499
#> 919 671.9555   2.500642
#> 920 500.4935   2.550269
#> 921 502.0380   2.859413
#> 922 852.0364   3.556803
#> 923 970.2911   2.556951
#> 924 502.1546   2.799390
#> 925 565.4297   2.593751
#> 926 728.4145   2.661575
#> 927 500.1335   2.966857
#> 928 511.9256   2.596994
#> 929 500.0657   2.798268
#> 930 502.7099   2.687058
#> 931 500.5446   3.435747
#> 932 510.6030   3.310288
#> 933 526.1229   2.584176
#> 934 500.1468   2.724041
#> 935 509.7360   4.342711
#> 936 517.0520   3.102693
#> 937 500.5572   2.593470
#> 938 790.9336   3.838468
#> 939 519.2657   2.625406
#> 940 598.8120   3.368318
#> 941 500.5985   2.657540
#> 942 510.7788   2.998053
#> 943 500.0658   3.001612
#> 944 506.4942   2.683855
#> 945 900.3384   2.515552
#> 946 558.6645   2.605547
#> 947 501.5630   2.572179
#> 948 503.7813   2.558038
#> 949 649.1670   3.185589
#> 950 502.1502   2.621146
#> 951 585.1674   2.531820
#> 952 509.9630   2.553212
#> 953 529.8497   2.539575
#> 954 504.1631   3.692440
#> 955 502.0740   3.802294
#> 956 768.3355   3.144241
#> 957 500.5918   3.392825
#> 958 915.2133   2.943628
#> 959 500.7929   2.782619
#> 960 500.8251   2.634378
#> 961 511.1533   3.933235
#> 962 500.9768   2.819551
#> 963 505.0039   2.606379
#> 964 500.2538   2.910028
#> 965 502.6102   3.107857
#> 966 502.3248   2.512604
#> 967 500.0007   2.979069
#> 968 501.0611   3.817173
#> 969 500.6666   2.575534
#> 970 500.5803   2.973179
#> 971 501.5152   2.830490