Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   500.1015   2.750246
#> 2   500.0281   2.584738
#> 3   504.9414   3.443449
#> 4   500.2880   3.669596
#> 5   509.9570   2.585078
#> 6   500.3548   2.916276
#> 7   809.9623   2.767364
#> 8   542.1763   2.576170
#> 9   500.7494   3.385011
#> 10  971.2221   2.891319
#> 11  620.6130   3.327565
#> 12  501.1130   3.001727
#> 13  504.0457   3.074665
#> 14  500.6810   2.512749
#> 15  500.2466   2.619697
#> 16  555.1395   2.741263
#> 17  503.0386   3.189324
#> 18  500.4488   2.714058
#> 19  510.0144   2.739719
#> 20  525.7740   3.117974
#> 21  812.8784   2.759114
#> 22  917.7037   2.866731
#> 23  500.1858   2.719556
#> 24  507.2087   2.662111
#> 25  819.2437   2.905725
#> 26  689.5248   3.995237
#> 27  566.1242   2.691618
#> 28  720.2489   3.101058
#> 29  500.1246   2.741180
#> 30  571.0419   3.239979
#> 31  984.3579   2.656850
#> 32  547.6523   3.378622
#> 33  501.1448   2.631052
#> 34  612.5281   2.750608
#> 35  500.0256   2.697625
#> 36  522.2394   3.267267
#> 37  500.2342   2.567884
#> 38  796.0734   2.517406
#> 39  500.7300   3.742129
#> 40  500.0757   2.526443
#> 41  500.0291   2.908427
#> 42  500.7305   3.060176
#> 43  504.2055   2.649164
#> 44  500.6709   2.536099
#> 45  500.1907   2.726141
#> 46  500.2273   2.577117
#> 47  534.1098   2.544288
#> 48  513.8655   2.921214
#> 49  511.6555   2.655544
#> 50  500.0434   2.615207
#> 51  582.1097   2.691556
#> 52  507.1576   2.815088
#> 53  502.7685   2.521885
#> 54  500.5350   2.592492
#> 55  569.4163   3.221517
#> 56  507.1853   3.056150
#> 57  899.9334   3.256647
#> 58  500.0097   2.845636
#> 59  707.2295   3.154556
#> 60  511.7118   2.694007
#> 61  505.2846   2.940886
#> 62  500.0958   2.924882
#> 63  621.0772   2.893547
#> 64  611.5725   2.542582
#> 65  500.8311   2.549217
#> 66  548.4560   4.016520
#> 67  500.2889   4.061423
#> 68  519.4462   2.741049
#> 69  500.2375   2.562865
#> 70  502.3501   2.969610
#> 71  504.6558   2.728465
#> 72  934.7663   2.813222
#> 73  500.3375   2.618493
#> 74  553.7785   2.601782
#> 75  500.0712   2.623197
#> 76  502.7076   2.524778
#> 77  512.1314   2.702087
#> 78  500.1180   3.121342
#> 79  500.8435   2.911489
#> 80  502.9034   2.535221
#> 81  500.5905   3.488460
#> 82  568.7095   2.570312
#> 83  501.3573   2.774717
#> 84  502.6788   2.705019
#> 85  500.6858   2.643744
#> 86  764.5836   3.402051
#> 87  500.0108   3.164827
#> 88  510.1281   2.600395
#> 89  502.0456   2.589726
#> 90  622.2179   2.653619
#> 91  528.6920   2.635555
#> 92  501.9898   2.641729
#> 93  501.9831   2.911146
#> 94  500.0056   3.168967
#> 95  501.0912   2.783080
#> 96  715.5636   2.503121
#> 97  500.2195   2.652337
#> 98  546.2571   2.680546
#> 99  526.3649   2.945767
#> 100 500.5466   3.353490
#> 101 501.2936   3.108510
#> 102 500.4274   2.749166
#> 103 500.1754   2.656698
#> 104 532.2808   6.036101
#> 105 518.6508   2.719481
#> 106 503.0301   2.744544
#> 107 778.9656   2.918477
#> 108 805.1416   2.786283
#> 109 500.0321   2.751496
#> 110 504.3599   2.764571
#> 111 538.2423   2.937310
#> 112 502.4536   2.881805
#> 113 500.2623   2.559812
#> 114 506.6057   2.534132
#> 115 543.5908   2.632480
#> 116 512.3140   2.678885
#> 117 543.5055   3.461100
#> 118 500.3005   2.988210
#> 119 500.3507   2.511544
#> 120 500.0003   2.527252
#> 121 631.0861   3.055447
#> 122 500.0132   2.582417
#> 123 500.2172   2.805912
#> 124 503.1307   3.185648
#> 125 515.6333   2.728950
#> 126 517.8086   2.837072
#> 127 502.1292   2.633295
#> 128 507.4725   2.507238
#> 129 500.2236   2.586979
#> 130 503.4364   2.500297
#> 131 500.5084   2.880424
#> 132 682.6339   3.216074
#> 133 500.0736   3.002705
#> 134 558.7322   3.400874
#> 135 500.1057   2.551661
#> 136 500.1953   2.699902
#> 137 501.2590   2.517856
#> 138 510.2948   2.505017
#> 139 526.1478   4.346970
#> 140 500.2137   2.547937
#> 141 529.5672   2.601343
#> 142 787.1049   4.211495
#> 143 648.5572   2.778176
#> 144 501.0564   3.200496
#> 145 515.1209   2.688285
#> 146 501.1473   2.557858
#> 147 732.0353   2.600420
#> 148 500.2855   2.687449
#> 149 500.1528   2.578459
#> 150 500.0887   2.798536
#> 151 500.3897   2.612731
#> 152 502.0126   2.921924
#> 153 503.9133   2.890494
#> 154 500.1789   2.887483
#> 155 991.0204   2.743293
#> 156 500.0126   2.834017
#> 157 631.9365   2.974102
#> 158 524.5238   2.561476
#> 159 502.6473   2.654338
#> 160 502.0390   2.628689
#> 161 500.0444   2.586071
#> 162 504.0465   3.159208
#> 163 500.8125   3.447712
#> 164 591.7819   2.541301
#> 165 500.0383   3.087391
#> 166 500.0048   3.506161
#> 167 568.8941   3.798392
#> 168 873.1703   2.542983
#> 169 500.0567   3.062306
#> 170 678.9523   2.625211
#> 171 518.3107   2.655674
#> 172 500.6747   3.317800
#> 173 578.1218   3.104696
#> 174 505.4250   2.807487
#> 175 514.4499   2.643197
#> 176 500.2599   2.759917
#> 177 504.4213   2.695745
#> 178 500.1564   2.854137
#> 179 530.1384   3.243892
#> 180 500.7433   2.623813
#> 181 596.7280   2.585697
#> 182 500.0499   2.790728
#> 183 506.6826   2.520503
#> 184 500.1113   4.251736
#> 185 502.2506   3.301079
#> 186 508.7383   2.724347
#> 187 509.7489   2.726424
#> 188 500.1859   3.826505
#> 189 583.7496   2.807185
#> 190 873.0630   2.851209
#> 191 500.0045   2.714532
#> 192 500.2244   3.714064
#> 193 504.7677   2.527236
#> 194 500.0008   2.500313
#> 195 500.0890   2.770111
#> 196 500.4064   2.791976
#> 197 517.6211   3.918643
#> 198 509.0966   2.719440
#> 199 500.0173   2.572749
#> 200 622.0067   3.124243
#> 201 504.8242   2.685992
#> 202 500.3427   2.788576
#> 203 663.4903   2.552921
#> 204 508.2861   2.752119
#> 205 509.3134   2.537514
#> 206 500.0570   3.354063
#> 207 500.0559   2.520902
#> 208 548.3798   2.746468
#> 209 500.0390   2.580597
#> 210 503.0985   3.054992
#> 211 500.0517   2.682407
#> 212 603.5043   2.550675
#> 213 519.8736   2.817505
#> 214 500.0654   2.802911
#> 215 642.6058   4.343701
#> 216 500.5406   2.908293
#> 217 509.9259   2.542963
#> 218 969.1606   2.555862
#> 219 512.9304   2.737804
#> 220 656.6122   2.776013
#> 221 580.6057   3.101004
#> 222 501.2565   3.560043
#> 223 520.4295   2.708677
#> 224 500.1804   2.644359
#> 225 502.0848   3.048162
#> 226 500.0960   4.016051
#> 227 506.8509   2.570315
#> 228 500.4822   2.679599
#> 229 502.1991   2.934709
#> 230 509.0934   2.757199
#> 231 508.3737   3.060757
#> 232 506.2844   2.785513
#> 233 610.3285   3.172793
#> 234 500.0960   2.824196
#> 235 500.2499   4.561786
#> 236 598.3304   3.558659
#> 237 504.1093   3.514753
#> 238 500.4219   2.734894
#> 239 502.6741   3.506721
#> 240 500.0412   2.677067
#> 241 504.5144   2.716316
#> 242 500.4093   3.262625
#> 243 502.8366   2.927829
#> 244 500.5908   2.668108
#> 245 525.0617   2.543936
#> 246 500.1634   3.060274
#> 247 500.1693   3.061516
#> 248 501.4130   2.837507
#> 249 513.9688   2.760817
#> 250 501.3772   2.502423
#> 251 510.5908   2.783993
#> 252 503.1070   3.428513
#> 253 500.3030   2.604067
#> 254 614.2983   3.090714
#> 255 512.8197   2.786624
#> 256 501.0037   2.694679
#> 257 500.3217   2.566509
#> 258 500.3432   2.521349
#> 259 500.2118   2.547626
#> 260 509.1174   2.881838
#> 261 500.0184   2.692319
#> 262 623.6802   3.735415
#> 263 508.4031   2.678957
#> 264 501.0413   2.794304
#> 265 500.4381   2.679991
#> 266 500.9392   3.704813
#> 267 500.1753   2.634790
#> 268 506.4549   3.228371
#> 269 693.3128   3.647369
#> 270 500.4995   2.579126
#> 271 590.7393   2.875121
#> 272 980.9726   2.654133
#> 273 507.8658   2.887064
#> 274 500.5325   2.739591
#> 275 514.8083   2.500703
#> 276 500.2768   2.500333
#> 277 500.5645   2.913499
#> 278 509.5455   2.950411
#> 279 507.9477   2.693292
#> 280 500.0454   2.858860
#> 281 500.3371   2.691564
#> 282 541.7363   2.729136
#> 283 930.7826   2.800708
#> 284 505.1952   2.637856
#> 285 507.4901   2.622129
#> 286 500.0630   3.133236
#> 287 590.1887   2.867184
#> 288 570.7517   2.533742
#> 289 574.6124   2.595857
#> 290 506.5703   4.929288
#> 291 525.6326   2.690129
#> 292 518.5080   2.506044
#> 293 511.4040   3.184985
#> 294 500.1380   2.997039
#> 295 513.5726   2.971580
#> 296 516.6847   2.672892
#> 297 502.0985   2.921126
#> 298 506.9738   3.114500
#> 299 827.7484   2.545294
#> 300 503.8701   2.663250
#> 301 501.5207   2.538533
#> 302 502.6828   2.622666
#> 303 655.3763   2.661931
#> 304 724.9602   2.523300
#> 305 500.6113   3.574160
#> 306 500.1639   3.427164
#> 307 546.4744   2.576914
#> 308 722.8450   2.908421
#> 309 751.9061   2.763896
#> 310 588.4911   3.029235
#> 311 500.1921   2.521389
#> 312 524.3524   2.513015
#> 313 501.8045   2.544541
#> 314 554.2235   4.274627
#> 315 529.8649   3.554307
#> 316 522.6898   3.254260
#> 317 528.2416   2.889533
#> 318 879.0357   3.908654
#> 319 645.5917   2.588273
#> 320 500.7787   3.402876
#> 321 536.1104   2.952578
#> 322 553.3524   3.074472
#> 323 505.8712   2.501272
#> 324 900.7159   3.438427
#> 325 555.9118   4.080074
#> 326 504.3791   2.987125
#> 327 500.6696   2.744281
#> 328 501.3355   2.710217
#> 329 502.3113   3.185123
#> 330 503.1076   2.750165
#> 331 500.4311   2.722119
#> 332 501.5201   2.588474
#> 333 500.0258   2.829164
#> 334 500.0805   2.653995
#> 335 525.4251   3.125669
#> 336 561.2837   2.816486
#> 337 509.0854   2.729244
#> 338 532.9322   3.092257
#> 339 500.0601   2.534393
#> 340 505.2820   2.949771
#> 341 500.0256   2.824240
#> 342 502.3124   2.607187
#> 343 501.3254   2.530604
#> 344 528.7380   2.806511
#> 345 865.3573   2.713369
#> 346 501.3949   2.530504
#> 347 899.3890   2.861396
#> 348 500.4544   2.627833
#> 349 553.1159   3.179558
#> 350 500.7064   2.902594
#> 351 500.9495   2.535281
#> 352 502.4806   2.614171
#> 353 500.0427   3.242488
#> 354 500.9833   2.645871
#> 355 504.2617   2.654490
#> 356 606.7514   2.744194
#> 357 500.0276   2.941406
#> 358 730.0930   2.588055
#> 359 842.5394   2.884493
#> 360 501.9560   3.454133
#> 361 500.1832   3.229403
#> 362 516.3356   3.481275
#> 363 618.5805   2.588302
#> 364 728.9274   3.362547
#> 365 500.9163   2.760878
#> 366 500.2264   2.506266
#> 367 500.6165   2.685499
#> 368 502.6625   2.615395
#> 369 500.1354   2.501938
#> 370 742.9723   2.825732
#> 371 500.2078   3.346502
#> 372 519.5534   2.593437
#> 373 502.5173   2.959554
#> 374 513.4563   2.707161
#> 375 500.5664   3.015237
#> 376 663.3984   2.744668
#> 377 833.6447   2.686472
#> 378 500.0377   2.718827
#> 379 501.1246   2.764180
#> 380 506.3571   2.654793
#> 381 517.0098   2.660665
#> 382 501.9802   3.070940
#> 383 500.1529   3.452910
#> 384 575.9979   3.167518
#> 385 501.3644   3.339886
#> 386 503.2058   2.830000
#> 387 505.4663   2.519561
#> 388 500.1196   3.375297
#> 389 501.1383   2.823522
#> 390 506.2434   3.078337
#> 391 500.0744   3.032511
#> 392 972.9116   2.607340
#> 393 801.6632   3.172402
#> 394 508.3338   2.800230
#> 395 500.0564   2.525426
#> 396 512.4814   2.535891
#> 397 500.3089   2.985477
#> 398 651.7114   2.668550
#> 399 500.1211   3.078217
#> 400 850.8948   2.643739
#> 401 501.3576   2.754596
#> 402 500.5868   2.922503
#> 403 713.8365   2.624314
#> 404 583.9425   2.915189
#> 405 540.3615   2.501065
#> 406 520.7667   3.354212
#> 407 500.1942   2.550108
#> 408 564.5666   2.666021
#> 409 500.0570   2.969415
#> 410 595.4244   4.345995
#> 411 502.5155   2.574454
#> 412 500.0540   2.805314
#> 413 529.6242   3.043759
#> 414 500.1341   2.631361
#> 415 500.0183   3.035374
#> 416 530.0282   3.090748
#> 417 500.3006   3.142368
#> 418 500.3668   3.161614
#> 419 537.5752   2.648027
#> 420 501.4637   2.842271
#> 421 535.2752   3.444935
#> 422 500.2773   2.522180
#> 423 502.8647   3.298091
#> 424 521.4026   2.770636
#> 425 500.2034   3.343203
#> 426 512.5545   3.051633
#> 427 500.1717   3.524355
#> 428 500.2408   2.579794
#> 429 510.1327   3.609929
#> 430 500.7731   3.327765
#> 431 630.3120   2.989186
#> 432 504.4548   2.575764
#> 433 535.3678   3.568219
#> 434 500.9077   2.503268
#> 435 501.2451   2.761666
#> 436 670.6755   2.558039
#> 437 508.8368   2.770640
#> 438 530.2180   2.503006
#> 439 512.9331   3.369625
#> 440 509.3514   2.988169
#> 441 557.6919   3.929912
#> 442 501.6957   3.287441
#> 443 500.0184   3.387608
#> 444 502.2567   3.449858
#> 445 500.3538   2.608713
#> 446 503.7071   2.863496
#> 447 502.9398   2.608580
#> 448 507.2189   4.018596
#> 449 500.3286   2.752688
#> 450 588.7453   3.532290
#> 451 500.3221   3.206124
#> 452 918.4584   3.132772
#> 453 500.1406   2.979867
#> 454 507.7438   2.729587
#> 455 503.8822   3.408567
#> 456 500.0594   2.559221
#> 457 514.8779   2.759671
#> 458 545.8278   2.507570
#> 459 513.4121   2.785653
#> 460 512.5543   3.403199
#> 461 500.2611   2.839412
#> 462 503.7906   2.642603
#> 463 500.1485   2.657388
#> 464 526.9700   2.622485
#> 465 727.8507   2.524086
#> 466 502.6687   2.630088
#> 467 500.1888   2.846944
#> 468 501.3839   2.918530
#> 469 943.2428   3.782708
#> 470 500.8900   2.693026
#> 471 501.2110   2.667161
#> 472 500.0253   3.466914
#> 473 533.5646   3.175124
#> 474 501.7599   2.647953
#> 475 500.0405   2.947429
#> 476 500.0038   3.368347
#> 477 503.8847   2.503199
#> 478 500.0429   3.355887
#> 479 500.0723   2.912849
#> 480 537.8243   2.698781
#> 481 500.1417   2.716572
#> 482 500.1420   3.334787
#> 483 506.5561   2.651376
#> 484 514.4066   2.653071
#> 485 500.0829   3.002973
#> 486 516.4208   3.156788
#> 487 500.9040   2.893003
#> 488 536.7279   2.897655
#> 489 500.0088   2.749386
#> 490 509.7076   3.203470
#> 491 501.4261   3.076029
#> 492 506.9345   2.652134
#> 493 723.7732   2.729916
#> 494 500.8085   2.892154
#> 495 502.3086   2.695688
#> 496 504.3111   3.280373
#> 497 500.0919   2.967158
#> 498 502.9244   2.561133
#> 499 556.1115   2.944726
#> 500 533.4734   2.571294
#> 501 504.3438   2.667021
#> 502 500.0036   2.745481
#> 503 598.6639   2.733059
#> 504 718.5151   2.584740
#> 505 511.1602   2.977257
#> 506 602.6424   3.440290
#> 507 507.2797   2.732657
#> 508 580.2812   3.341959
#> 509 878.5377   2.990954
#> 510 500.6644   2.767570
#> 511 540.8097   2.920005
#> 512 655.3600   2.580470
#> 513 820.5870   2.818846
#> 514 574.1664   3.247093
#> 515 595.1988   2.620969
#> 516 500.4115   2.585971
#> 517 566.8779   2.807885
#> 518 500.1869   2.941398
#> 519 502.5759   3.072586
#> 520 500.7022   2.547877
#> 521 535.6365   3.493509
#> 522 600.4456   3.097760
#> 523 513.5284   2.881946
#> 524 500.0071   2.672325
#> 525 500.0181   2.930203
#> 526 502.2287   4.163326
#> 527 617.6224   2.644562
#> 528 927.6883   2.536479
#> 529 500.1623   2.669069
#> 530 500.6675   2.871055
#> 531 503.5228   3.070659
#> 532 502.2391   2.581134
#> 533 501.1069   2.720924
#> 534 676.9297   2.896842
#> 535 500.0977   3.009809
#> 536 501.3504   2.624906
#> 537 509.3685   2.906904
#> 538 500.0863   2.533554
#> 539 507.5999   4.435461
#> 540 500.4368   3.038973
#> 541 500.5463   2.787337
#> 542 500.2034   2.793383
#> 543 528.1744   2.679380
#> 544 500.6509   3.247311
#> 545 513.5055   2.788170
#> 546 501.1290   2.672998
#> 547 501.1131   2.779975
#> 548 529.3555   3.600674
#> 549 591.2866   2.617940
#> 550 501.1992   2.569734
#> 551 500.3247   3.652068
#> 552 500.0552   2.746182
#> 553 500.2569   3.936921
#> 554 500.1945   3.067838
#> 555 504.2464   3.398402
#> 556 992.4507   2.520336
#> 557 533.2395   2.857104
#> 558 506.1646   2.896013
#> 559 500.2259   3.003411
#> 560 500.1387   3.150223
#> 561 500.5216   2.544863
#> 562 500.0992   2.874824
#> 563 500.7238   2.984053
#> 564 503.1999   2.523370
#> 565 548.8095   3.029391
#> 566 559.3386   4.236732
#> 567 696.5249   2.603156
#> 568 500.1149   2.594162
#> 569 508.0531   4.370255
#> 570 923.1941   2.809199
#> 571 514.4093   3.598297
#> 572 500.3705   2.712971
#> 573 642.6284   2.528022
#> 574 500.0052   3.333756
#> 575 601.0929   2.768391
#> 576 502.1095   3.006716
#> 577 500.0286   2.682076
#> 578 500.4637   2.561799
#> 579 566.6704   2.522127
#> 580 505.7797   2.933187
#> 581 576.3840   2.978170
#> 582 500.0458   2.803484
#> 583 538.0750   3.381179
#> 584 523.6622   2.799253
#> 585 507.3969   2.596609
#> 586 500.7464   2.566740
#> 587 501.2238   2.778466
#> 588 501.0263   2.512667
#> 589 503.2092   2.591730
#> 590 501.6959   2.586523
#> 591 500.0590   2.521609
#> 592 503.2099   3.011835
#> 593 504.6929   2.902227
#> 594 632.7436   3.324343
#> 595 500.1163   2.620431
#> 596 595.0556   2.655824
#> 597 528.4282   2.830881
#> 598 500.0145   3.668858
#> 599 572.6639   2.501926
#> 600 500.4036   3.135585
#> 601 500.9242   3.640143
#> 602 510.4534   2.665422
#> 603 512.7289   2.711038
#> 604 510.2145   3.382064
#> 605 568.8720   2.840252
#> 606 509.0637   2.887069
#> 607 530.0629   2.938896
#> 608 500.4179   2.857457
#> 609 502.2600   2.694421
#> 610 682.0328   4.066328
#> 611 500.1216   2.575721
#> 612 964.8056   3.830990
#> 613 553.8599   3.031648
#> 614 662.2219   2.942862
#> 615 521.0171   2.728645
#> 616 501.4003   2.986621
#> 617 572.3356   2.940003
#> 618 500.4572   2.527699
#> 619 500.1829   2.974342
#> 620 501.7776   2.876064
#> 621 501.8462   3.693014
#> 622 551.9988   2.550636
#> 623 658.7194   2.961475
#> 624 994.3146   2.806462
#> 625 512.5224   2.701851
#> 626 511.8334   2.531781
#> 627 504.8757   3.548275
#> 628 503.4871   2.699542
#> 629 500.3930   3.183356
#> 630 500.3633   3.109655
#> 631 521.3541   2.559694
#> 632 502.4443   3.067399
#> 633 554.3020   3.413741
#> 634 500.0336   3.089718
#> 635 517.5010   4.381985
#> 636 506.0977   2.532211
#> 637 666.6334   2.863055
#> 638 500.4193   3.684381
#> 639 500.5154   2.541003
#> 640 544.5285   2.697772
#> 641 524.2326   3.067578
#> 642 520.0931   2.574508
#> 643 501.3829   3.105528
#> 644 500.1766   2.520267
#> 645 503.8317   3.188060
#> 646 565.0208   2.712217
#> 647 657.9539   3.129943
#> 648 502.6698   2.748337
#> 649 988.7555   2.557963
#> 650 500.6355   3.107316
#> 651 501.0324   2.973494
#> 652 500.8071   2.502921
#> 653 502.5489   4.629403
#> 654 509.0949   2.813634
#> 655 627.4430   2.798820
#> 656 500.0787   2.504649
#> 657 506.7383   2.587706
#> 658 500.5201   2.862731
#> 659 500.4153   2.671019
#> 660 573.4416   2.552229
#> 661 501.5082   2.982566
#> 662 510.4975   2.560359
#> 663 504.8127   2.507365
#> 664 590.9271   3.223165
#> 665 500.3749   2.892463
#> 666 503.7051   2.874337
#> 667 576.9294   2.910628
#> 668 503.1965   2.824307
#> 669 501.1135   3.264220
#> 670 500.8831   3.098487
#> 671 501.3186   3.880584
#> 672 512.5237   3.650517
#> 673 508.2332   2.642728
#> 674 504.6132   3.151442
#> 675 500.7798   2.661151
#> 676 500.0369   3.267876
#> 677 512.4310   4.548061
#> 678 502.1370   3.360017
#> 679 510.7692   2.810272
#> 680 500.1033   3.322431
#> 681 500.5689   2.741473
#> 682 504.8400   3.643147
#> 683 522.2039   3.742063
#> 684 690.0676   2.504004
#> 685 500.0008   3.652810
#> 686 540.0968   2.594550
#> 687 503.0388   2.605144
#> 688 501.9734   2.694036
#> 689 721.4908   2.560537
#> 690 500.9970   2.671471
#> 691 500.7067   3.522632
#> 692 500.5765   2.514552
#> 693 500.3015   3.920757
#> 694 501.3936   3.641575
#> 695 505.4762   2.580464
#> 696 500.1891   3.145627
#> 697 510.0988   2.613289
#> 698 578.1917   2.825905
#> 699 500.3860   3.277086
#> 700 502.8266   2.872614
#> 701 500.2012   2.982668
#> 702 522.6961   2.578204
#> 703 631.7898   3.262493
#> 704 502.0923   2.929935
#> 705 501.1157   2.509729
#> 706 539.2592   3.231477
#> 707 527.0769   2.545965
#> 708 500.2490   2.613351
#> 709 513.9185   3.220933
#> 710 503.2803   2.651320
#> 711 515.8384   3.271913
#> 712 545.8267   3.082657
#> 713 501.0176   3.105427
#> 714 507.5317   3.480096
#> 715 500.0611   3.494172
#> 716 500.7322   2.650152
#> 717 505.8289   2.799698
#> 718 500.7628   2.555969
#> 719 537.6191   3.011695
#> 720 510.7066   2.776906
#> 721 500.0419   2.839030
#> 722 962.8286   2.835098
#> 723 503.1393   3.042504
#> 724 500.1848   2.752813
#> 725 500.1744   2.528183
#> 726 583.2176   3.158260
#> 727 500.3813   2.852583
#> 728 501.5829   2.791527
#> 729 501.3219   3.651138
#> 730 514.2605   2.527446
#> 731 532.6172   3.469434
#> 732 500.0384   3.242890
#> 733 507.4489   2.568814
#> 734 500.5769   2.996896
#> 735 500.0248   2.595376
#> 736 501.6838   2.813786
#> 737 507.5481   2.612345
#> 738 501.3378   2.535792
#> 739 511.8409   3.328970
#> 740 524.5597   2.803458
#> 741 504.8921   2.563286
#> 742 500.0256   2.751902
#> 743 503.8190   3.165762
#> 744 656.0433   2.548575
#> 745 507.1809   2.973811
#> 746 500.1649   3.224932
#> 747 622.1152   2.885038
#> 748 501.0433   2.840154
#> 749 531.4058   3.621186
#> 750 666.2828   3.506248
#> 751 521.1034   4.221211
#> 752 500.2278   2.804183
#> 753 541.7039   2.516587
#> 754 500.1094   2.726304
#> 755 500.3149   2.895815
#> 756 631.6740   2.745141
#> 757 525.1878   2.507920
#> 758 528.4333   3.122694
#> 759 500.3483   4.707256
#> 760 500.0650   2.618239
#> 761 524.4358   2.715548
#> 762 500.0556   3.041346
#> 763 512.3772   2.622049
#> 764 540.8822   2.631229
#> 765 512.8048   3.629492
#> 766 573.5988   4.432379
#> 767 500.2199   2.504910
#> 768 501.3642   3.055788
#> 769 500.0283   2.633693
#> 770 557.0259   2.569707
#> 771 500.1374   2.836885
#> 772 608.1144   2.548778
#> 773 503.2165   3.106638
#> 774 500.4509   3.588692
#> 775 630.9791   2.572592
#> 776 500.1964   2.701700
#> 777 521.0582   2.614788
#> 778 568.1367   2.543613
#> 779 677.0612   2.678919
#> 780 500.1105   2.798338
#> 781 500.5645   2.710393
#> 782 524.9210   2.743681
#> 783 774.5150   2.643786
#> 784 500.5601   3.137519
#> 785 500.0240   2.985443
#> 786 503.9682   2.848009
#> 787 895.2767   2.676736
#> 788 500.0380   2.568381
#> 789 989.3625   3.106844
#> 790 556.4811   2.678475
#> 791 500.9085   3.741209
#> 792 500.6345   3.194345
#> 793 630.6082   2.533194
#> 794 508.6348   2.792179
#> 795 900.6378   2.814191
#> 796 572.1462   2.774215
#> 797 526.4586   2.603289
#> 798 500.7418   2.664173
#> 799 672.9711   2.782652
#> 800 533.0551   2.705955
#> 801 503.7143   3.371005
#> 802 632.0244   3.415164
#> 803 500.2990   3.440737
#> 804 501.5813   3.028573
#> 805 538.1505   3.033350
#> 806 582.9043   2.517886
#> 807 502.8931   2.898975
#> 808 779.1741   2.755867
#> 809 502.4202   2.890306
#> 810 504.9111   2.921254
#> 811 582.3552   2.575340
#> 812 500.0246   2.919924
#> 813 500.0405   2.566002
#> 814 506.7649   3.067070
#> 815 736.9907   2.586017
#> 816 634.7308   3.050938
#> 817 500.3797   3.041703
#> 818 579.3371   2.559583
#> 819 609.2400   2.504006
#> 820 605.5242   3.107318
#> 821 500.9991   2.784769
#> 822 500.0244   2.687066
#> 823 501.4745   2.637732
#> 824 501.0781   2.676805
#> 825 500.3599   2.931208
#> 826 501.4180   2.581096
#> 827 533.4579   2.918138
#> 828 500.9557   2.881234
#> 829 502.5939   2.884652
#> 830 502.1160   3.410603
#> 831 500.0067   2.975078
#> 832 510.3214   2.643155
#> 833 500.7136   3.627866
#> 834 743.4996   2.715359
#> 835 500.2154   3.308888
#> 836 500.0284   2.551464
#> 837 832.7441   2.838040
#> 838 511.6467   2.772953
#> 839 504.2109   3.042304
#> 840 500.0122   2.782495
#> 841 514.5742   2.776916
#> 842 508.5653   2.613346
#> 843 500.0706   3.042150
#> 844 500.4564   4.160332
#> 845 500.5232   2.563552
#> 846 923.3280   2.561336
#> 847 500.8079   4.527496
#> 848 508.4103   2.771826
#> 849 511.6998   2.584721
#> 850 500.6664   2.880297
#> 851 500.1509   2.576782
#> 852 502.7049   2.869281
#> 853 500.6524   2.865830
#> 854 500.3413   3.801164
#> 855 502.1860   2.521266
#> 856 502.4792   2.679289
#> 857 501.6984   2.890152