Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   514.3224   2.765592
#> 2   525.1838   2.599248
#> 3   519.1376   2.606305
#> 4   502.4994   2.787694
#> 5   692.0008   2.791528
#> 6   711.5982   2.867001
#> 7   979.5727   2.812908
#> 8   500.2158   2.700614
#> 9   500.1061   2.510669
#> 10  524.6675   2.673983
#> 11  500.0224   2.701554
#> 12  508.8208   3.678204
#> 13  555.7244   2.887637
#> 14  502.3171   3.280395
#> 15  500.5046   2.968910
#> 16  500.8791   2.886355
#> 17  731.2694   3.120866
#> 18  506.4794   2.648108
#> 19  504.6473   2.607824
#> 20  500.4795   2.812084
#> 21  501.3874   2.562711
#> 22  796.7600   2.543651
#> 23  500.0775   2.804980
#> 24  501.6723   3.098197
#> 25  500.0460   2.599626
#> 26  500.6707   2.858923
#> 27  512.3939   3.825766
#> 28  510.1351   2.775709
#> 29  515.5166   2.680026
#> 30  500.2285   3.539282
#> 31  500.2352   3.002049
#> 32  875.0152   3.375060
#> 33  538.3452   2.513399
#> 34  509.6346   3.224435
#> 35  502.6315   2.634010
#> 36  500.4234   3.189601
#> 37  500.8425   2.597844
#> 38  500.3278   2.543998
#> 39  502.3565   2.868564
#> 40  576.4208   2.952439
#> 41  574.4798   2.909740
#> 42  598.2596   2.747778
#> 43  613.8358   2.932381
#> 44  500.5680   2.899704
#> 45  504.4924   3.518454
#> 46  544.2928   2.778296
#> 47  500.6969   2.702307
#> 48  582.4779   3.221966
#> 49  504.6413   3.212198
#> 50  500.5052   2.525261
#> 51  500.0443   2.943790
#> 52  556.3990   2.754048
#> 53  501.4882   2.582270
#> 54  500.9360   2.688690
#> 55  674.5437   2.525295
#> 56  500.4814   2.706741
#> 57  623.2214   2.897642
#> 58  500.0410   2.511718
#> 59  506.1546   2.513808
#> 60  546.4950   2.730548
#> 61  654.9349   3.700887
#> 62  500.2426   3.648873
#> 63  500.1295   2.832067
#> 64  502.7396   2.690550
#> 65  824.3157   2.937977
#> 66  500.5851   2.733586
#> 67  500.1219   2.630907
#> 68  708.8005   2.703504
#> 69  504.4418   2.842862
#> 70  500.1364   2.931998
#> 71  500.1430   3.169742
#> 72  501.8241   3.331299
#> 73  617.7029   2.521571
#> 74  568.4118   2.567759
#> 75  500.1809   2.865639
#> 76  528.2643   3.114822
#> 77  500.1274   2.530508
#> 78  500.4380   2.753021
#> 79  505.8453   2.632717
#> 80  640.6919   3.104867
#> 81  500.0505   3.703750
#> 82  516.0440   2.767663
#> 83  509.3472   2.551664
#> 84  500.5290   3.409405
#> 85  500.0056   3.076136
#> 86  502.3155   3.278442
#> 87  500.0677   2.521824
#> 88  826.2577   2.513467
#> 89  930.8255   2.913215
#> 90  502.1736   3.124766
#> 91  529.3080   2.897974
#> 92  500.8913   3.154593
#> 93  506.7171   4.348796
#> 94  501.1795   2.606342
#> 95  768.9591   2.502826
#> 96  500.2787   2.630088
#> 97  500.3238   2.558534
#> 98  500.4799   3.120897
#> 99  524.3119   2.606548
#> 100 544.5153   3.336629
#> 101 670.5855   2.669902
#> 102 506.8010   2.512330
#> 103 503.1508   3.018959
#> 104 500.6032   4.691631
#> 105 503.6313   3.941570
#> 106 590.0619   2.717787
#> 107 500.2783   2.658169
#> 108 500.2941   3.858030
#> 109 500.9724   2.701256
#> 110 511.1930   2.507108
#> 111 502.1983   2.514578
#> 112 504.9199   4.023873
#> 113 500.0225   2.514070
#> 114 622.3778   2.724954
#> 115 616.9128   2.604765
#> 116 511.5760   2.741862
#> 117 502.2408   2.535850
#> 118 878.8210   2.618034
#> 119 500.8167   2.682210
#> 120 510.9933   4.673183
#> 121 685.1977   2.846589
#> 122 500.4215   3.742949
#> 123 500.2527   2.733069
#> 124 513.4691   2.544687
#> 125 503.0952   2.856589
#> 126 500.3444   2.690730
#> 127 500.0688   3.260266
#> 128 558.9964   3.213647
#> 129 501.3013   2.725228
#> 130 718.3563   2.696280
#> 131 500.5635   2.697468
#> 132 546.5107   3.371741
#> 133 500.0714   2.525176
#> 134 513.5507   2.714184
#> 135 525.0947   2.564214
#> 136 660.5172   2.678757
#> 137 503.9249   2.671226
#> 138 500.0494   2.660528
#> 139 500.5431   2.578844
#> 140 629.1948   3.193420
#> 141 869.6236   4.178372
#> 142 559.4525   3.940290
#> 143 500.0969   2.888951
#> 144 501.4826   3.198055
#> 145 500.1368   2.949861
#> 146 500.0464   2.638440
#> 147 514.7775   2.651028
#> 148 500.1655   2.878598
#> 149 537.4059   2.912846
#> 150 500.0784   2.602454
#> 151 517.3109   2.501103
#> 152 615.7296   2.873198
#> 153 500.5980   2.947575
#> 154 532.0947   2.605548
#> 155 501.0336   2.708950
#> 156 914.0635   2.672636
#> 157 619.2045   3.885486
#> 158 500.0119   3.437921
#> 159 507.7428   2.642917
#> 160 500.1660   3.257129
#> 161 585.3903   2.867907
#> 162 503.6739   2.715687
#> 163 500.1855   2.556924
#> 164 500.0526   3.369395
#> 165 500.6083   2.732835
#> 166 536.5050   2.539738
#> 167 502.8064   2.919102
#> 168 525.1915   3.789070
#> 169 543.7691   2.803570
#> 170 500.1850   3.181911
#> 171 680.7870   2.679272
#> 172 500.5690   2.510900
#> 173 500.1710   2.621984
#> 174 500.4358   2.522074
#> 175 535.8245   2.517722
#> 176 500.3522   2.747483
#> 177 500.6246   3.176677
#> 178 503.0391   2.630167
#> 179 500.3027   3.294025
#> 180 500.1494   2.620691
#> 181 501.1625   3.106057
#> 182 560.0349   2.930307
#> 183 500.1291   2.823339
#> 184 611.8407   2.596677
#> 185 500.0192   2.675958
#> 186 500.1991   2.661073
#> 187 932.4641   2.616562
#> 188 500.9455   2.669215
#> 189 502.8672   2.548628
#> 190 500.0018   2.564214
#> 191 500.1759   2.504736
#> 192 500.8242   2.648562
#> 193 500.6448   2.542222
#> 194 530.2120   2.644567
#> 195 502.4741   2.969020
#> 196 500.0480   2.708637
#> 197 500.0952   3.168823
#> 198 500.3278   2.613738
#> 199 519.1253   2.658842
#> 200 500.0632   3.265880
#> 201 685.7702   2.752375
#> 202 568.6355   3.146385
#> 203 521.5540   2.678978
#> 204 507.1724   2.928996
#> 205 672.8927   2.751260
#> 206 659.7202   3.720349
#> 207 502.0010   2.668819
#> 208 531.2109   2.916449
#> 209 502.9077   3.739205
#> 210 504.7424   3.540254
#> 211 501.7697   2.622458
#> 212 520.0302   2.999478
#> 213 500.3072   3.589679
#> 214 521.9052   2.576871
#> 215 838.9967   2.605866
#> 216 522.5059   2.543452
#> 217 500.2858   2.685343
#> 218 500.0279   3.027163
#> 219 500.9465   2.983760
#> 220 500.0833   2.711137
#> 221 500.9931   2.937371
#> 222 504.7270   2.715394
#> 223 501.6419   3.509510
#> 224 505.6645   2.612265
#> 225 500.1154   3.530380
#> 226 610.3716   3.330889
#> 227 517.7455   3.037111
#> 228 500.0756   3.147311
#> 229 500.0543   3.767650
#> 230 605.0893   2.546999
#> 231 502.5815   2.531662
#> 232 522.2819   2.552536
#> 233 500.1330   2.778518
#> 234 566.2623   2.551601
#> 235 592.3309   2.545835
#> 236 550.6603   2.612087
#> 237 500.0212   2.670645
#> 238 501.3321   2.822515
#> 239 505.6953   2.596428
#> 240 504.2716   2.721698
#> 241 502.2220   2.565917
#> 242 500.9507   2.762157
#> 243 500.3470   2.834618
#> 244 668.4845   2.520279
#> 245 519.7776   2.746749
#> 246 500.1159   4.666021
#> 247 500.9869   2.683018
#> 248 512.0993   2.675758
#> 249 501.2422   2.532684
#> 250 605.1131   2.659223
#> 251 506.9922   3.328528
#> 252 506.0392   2.930191
#> 253 500.7350   2.638933
#> 254 501.2999   2.562803
#> 255 500.3474   3.432573
#> 256 514.9523   2.756299
#> 257 576.8595   3.365608
#> 258 510.6001   2.796729
#> 259 555.6183   2.684110
#> 260 508.8268   2.838749
#> 261 506.7138   2.843117
#> 262 500.0005   3.189007
#> 263 500.1875   2.740364
#> 264 531.7974   2.659727
#> 265 500.5106   2.954311
#> 266 506.7978   2.738374
#> 267 500.0729   3.296419
#> 268 522.0296   2.618307
#> 269 500.1590   3.364452
#> 270 501.5725   3.175030
#> 271 500.3718   2.674285
#> 272 673.9723   2.795087
#> 273 501.2640   2.702525
#> 274 500.8040   2.701518
#> 275 505.6074   2.567199
#> 276 526.2376   2.526960
#> 277 501.1893   3.108229
#> 278 500.0366   3.059350
#> 279 500.2168   3.239276
#> 280 500.7472   3.213597
#> 281 500.4399   2.539498
#> 282 508.0797   2.610845
#> 283 846.9884   2.960499
#> 284 500.4651   3.501648
#> 285 540.9450   3.371744
#> 286 500.8619   2.622271
#> 287 507.2972   2.726685
#> 288 500.6518   3.412760
#> 289 500.0122   3.268672
#> 290 500.8618   2.570520
#> 291 742.7778   2.662638
#> 292 500.6679   2.778658
#> 293 500.4247   3.176027
#> 294 500.2809   2.582300
#> 295 688.2588   2.595873
#> 296 507.2205   2.713573
#> 297 500.1637   3.068038
#> 298 503.3954   3.079318
#> 299 678.1355   2.649724
#> 300 503.1990   4.067798
#> 301 507.9981   2.862446
#> 302 508.4592   2.543536
#> 303 667.0502   3.163975
#> 304 568.5365   3.306054
#> 305 500.0524   2.703058
#> 306 500.0366   2.614952
#> 307 501.1221   2.931003
#> 308 500.1482   2.674613
#> 309 500.2327   3.963063
#> 310 819.9139   3.014565
#> 311 500.5884   2.549370
#> 312 508.4875   2.505743
#> 313 503.2187   2.790304
#> 314 517.2687   2.993050
#> 315 510.9393   3.410583
#> 316 545.6432   2.793236
#> 317 500.7318   2.512100
#> 318 502.5453   3.893729
#> 319 502.3765   2.692541
#> 320 511.3393   2.745082
#> 321 542.7635   2.740683
#> 322 502.1734   2.709674
#> 323 500.0220   2.677945
#> 324 500.3107   2.714217
#> 325 500.1280   3.618436
#> 326 664.2883   2.721599
#> 327 503.8263   2.712952
#> 328 610.8393   2.619038
#> 329 550.9260   3.296015
#> 330 507.8326   3.684963
#> 331 810.7244   2.614887
#> 332 509.9076   2.712614
#> 333 765.5121   2.694322
#> 334 500.0379   2.850068
#> 335 505.4059   2.917337
#> 336 500.1488   2.808830
#> 337 500.8791   3.005451
#> 338 520.1547   2.738013
#> 339 590.9310   3.172120
#> 340 500.1560   2.561106
#> 341 503.4507   2.609212
#> 342 585.7934   2.821683
#> 343 501.0256   2.637570
#> 344 501.3227   4.692739
#> 345 507.9232   2.948895
#> 346 513.2259   3.602886
#> 347 860.9470   3.374636
#> 348 504.3932   3.081905
#> 349 500.1005   2.822980
#> 350 644.9744   2.721841
#> 351 501.1946   2.998384
#> 352 655.4004   2.982780
#> 353 500.4988   3.413556
#> 354 521.2953   2.577073
#> 355 521.1253   3.024355
#> 356 501.0568   2.564941
#> 357 500.5951   2.567729
#> 358 775.1218   3.356369
#> 359 500.3649   2.726459
#> 360 500.0225   2.500284
#> 361 516.5691   2.618085
#> 362 504.7314   2.513550
#> 363 840.1272   2.570299
#> 364 500.2709   3.374512
#> 365 736.2728   2.809261
#> 366 624.9783   2.811651
#> 367 500.0743   3.006871
#> 368 500.0895   2.576722
#> 369 500.0020   2.714911
#> 370 503.7579   3.133469
#> 371 500.2217   2.524234
#> 372 508.1019   2.688126
#> 373 500.0572   3.033944
#> 374 500.1431   3.055896
#> 375 500.1825   3.330704
#> 376 560.4249   2.517313
#> 377 500.0196   2.526632
#> 378 512.1008   2.525454
#> 379 524.6469   2.628234
#> 380 516.3164   2.715187
#> 381 507.2301   3.112422
#> 382 504.8144   2.952779
#> 383 500.0969   2.673178
#> 384 509.0673   2.669367
#> 385 510.8245   2.642351
#> 386 521.5871   3.234628
#> 387 501.1915   3.328678
#> 388 500.5331   2.757784
#> 389 631.3295   3.031273
#> 390 500.6313   2.904663
#> 391 509.0511   2.932400
#> 392 500.7403   2.992198
#> 393 501.0744   2.569765
#> 394 500.2594   3.669543
#> 395 500.9708   2.637508
#> 396 512.2287   2.557433
#> 397 916.8472   3.039262
#> 398 501.8857   2.508900
#> 399 500.1436   2.709226
#> 400 500.5036   2.526190
#> 401 849.9058   3.839824
#> 402 501.4050   3.480141
#> 403 500.3608   2.758480
#> 404 519.0934   2.892581
#> 405 519.8456   2.886693
#> 406 507.8637   3.149919
#> 407 500.0085   2.532452
#> 408 500.8542   3.244156
#> 409 504.4293   2.835972
#> 410 536.9244   2.876962
#> 411 503.5180   2.614782
#> 412 505.5064   2.565818
#> 413 504.4905   2.786118
#> 414 500.2618   2.634809
#> 415 506.6460   2.758655
#> 416 500.1797   2.815636
#> 417 501.1802   2.699212
#> 418 500.0391   2.828938
#> 419 500.3307   3.399162
#> 420 921.8741   3.046330
#> 421 502.4001   2.794954
#> 422 556.7047   2.545175
#> 423 714.5209   3.560748
#> 424 561.0441   2.713755
#> 425 500.0489   3.386687
#> 426 918.5449   2.674500
#> 427 525.7910   2.572914
#> 428 838.3183   3.368990
#> 429 584.0591   2.572895
#> 430 515.0398   2.769393
#> 431 503.9812   3.012482
#> 432 503.7711   2.541933
#> 433 637.3663   3.842129
#> 434 534.1229   2.576503
#> 435 500.0024   2.602506
#> 436 504.2479   2.588716
#> 437 500.0246   2.643489
#> 438 508.0913   3.516958
#> 439 500.0876   2.549162
#> 440 540.0794   2.612438
#> 441 500.1291   3.577313
#> 442 502.2298   2.581221
#> 443 500.6738   3.321220
#> 444 567.4547   2.761623
#> 445 509.5672   2.755200
#> 446 502.6486   2.816776
#> 447 545.4541   2.840509
#> 448 500.7687   2.984471
#> 449 625.4036   2.521850
#> 450 594.8989   3.091919
#> 451 947.0540   2.943268
#> 452 519.3419   2.655486
#> 453 500.2870   2.581117
#> 454 695.0643   2.782478
#> 455 991.9901   2.502032
#> 456 500.0064   2.584569
#> 457 510.7031   3.194291
#> 458 500.0354   2.692803
#> 459 500.3112   3.214561
#> 460 514.3409   3.051233
#> 461 500.1354   3.249583
#> 462 501.0781   3.436331
#> 463 501.6889   2.942318
#> 464 548.6546   2.803225
#> 465 541.6177   2.925871
#> 466 575.6670   2.551347
#> 467 501.7475   2.826583
#> 468 547.7993   2.513631
#> 469 500.2568   3.076824
#> 470 504.4332   2.932265
#> 471 500.2686   3.605857
#> 472 500.1231   2.675146
#> 473 500.0510   2.560682
#> 474 500.2776   3.485228
#> 475 500.4365   3.609350
#> 476 560.9198   2.724788
#> 477 648.0506   2.668249
#> 478 545.2897   2.596743
#> 479 516.1690   3.003749
#> 480 500.1693   2.911305
#> 481 521.9902   3.013023
#> 482 504.0405   2.608160
#> 483 502.0555   3.003177
#> 484 500.1138   3.124256
#> 485 501.2886   2.662526
#> 486 503.5568   2.864035
#> 487 977.4227   2.691904
#> 488 504.0525   2.537397
#> 489 507.5708   2.851039
#> 490 500.1320   2.567318
#> 491 627.0558   2.666833
#> 492 500.0173   2.991739
#> 493 500.0024   2.703620
#> 494 500.1304   2.732133
#> 495 530.1193   3.025724
#> 496 500.3928   2.737931
#> 497 631.9746   2.681788
#> 498 502.9804   2.552974
#> 499 551.7646   3.026590
#> 500 500.0724   2.623052
#> 501 500.0910   3.602163
#> 502 569.9425   3.155555
#> 503 524.3976   2.688713
#> 504 500.3075   2.721078
#> 505 501.4100   2.971938
#> 506 589.3612   2.939341
#> 507 501.8867   3.418296
#> 508 674.0999   3.100224
#> 509 501.2008   3.253251
#> 510 513.6479   2.625054
#> 511 502.7497   2.787087
#> 512 942.1752   3.587844
#> 513 735.5081   2.704668
#> 514 780.3487   2.561621
#> 515 612.7899   2.620209
#> 516 511.1001   3.029723
#> 517 500.3296   2.720775
#> 518 549.6809   2.503918
#> 519 500.3177   3.222810
#> 520 504.4242   3.171849
#> 521 508.2004   3.020669
#> 522 500.0714   2.543983
#> 523 500.1839   2.770842
#> 524 935.7049   2.893484
#> 525 589.2592   2.643369
#> 526 500.0578   2.559773
#> 527 501.8992   2.804544
#> 528 501.3508   2.593843
#> 529 501.0570   2.722148
#> 530 505.0815   2.562345
#> 531 834.2707   2.635716
#> 532 572.8573   3.300048
#> 533 502.9284   2.916723
#> 534 501.1639   3.009334
#> 535 500.0889   2.557056
#> 536 512.2152   3.934361
#> 537 500.4556   3.278801
#> 538 500.1938   2.745329
#> 539 571.8257   3.961055
#> 540 581.7464   2.742965
#> 541 504.0457   3.507447
#> 542 500.9206   2.642234
#> 543 888.6118   2.796151
#> 544 500.3438   3.348127
#> 545 500.5748   2.577761
#> 546 500.7231   3.168238
#> 547 500.4789   3.049176
#> 548 629.4875   2.970680
#> 549 593.3558   4.434072
#> 550 500.0208   2.923856
#> 551 500.0684   3.544398
#> 552 687.8997   2.709934
#> 553 532.7706   2.598793
#> 554 803.6134   2.845699
#> 555 500.2467   2.768013
#> 556 500.7083   2.907295
#> 557 711.1211   2.622928
#> 558 500.1908   2.877669
#> 559 519.2825   2.681591
#> 560 500.8541   4.991997
#> 561 502.6124   2.619692
#> 562 500.0316   2.798438
#> 563 500.9660   2.933168
#> 564 503.7466   2.500315
#> 565 500.1909   2.973986
#> 566 520.0537   2.920014
#> 567 500.2694   2.995929
#> 568 500.0108   2.588838
#> 569 500.1731   3.738818
#> 570 602.9910   2.824972
#> 571 769.7468   3.553677
#> 572 503.0685   2.567044
#> 573 830.4122   2.560216
#> 574 500.0731   3.240967
#> 575 510.0606   2.745069
#> 576 500.2667   2.716105
#> 577 504.4164   3.097455
#> 578 512.5206   2.683891
#> 579 539.6558   3.481984
#> 580 515.1969   2.870174
#> 581 614.5233   2.685789
#> 582 512.7537   3.511753
#> 583 500.2036   2.549458
#> 584 644.1551   2.504698
#> 585 500.7519   3.324606
#> 586 572.5738   2.517529
#> 587 500.0761   2.529374
#> 588 501.9732   2.795844
#> 589 520.2561   2.500245
#> 590 500.4369   2.502707
#> 591 500.0201   3.076126
#> 592 500.3074   2.934902
#> 593 531.4335   2.791475
#> 594 501.7978   3.189535
#> 595 575.4103   3.230835
#> 596 514.5582   3.071934
#> 597 652.4539   2.835232
#> 598 508.0410   2.873781
#> 599 590.6842   2.511723
#> 600 961.4160   2.523534
#> 601 812.3070   2.684572
#> 602 502.6191   2.545518
#> 603 500.1167   2.518518
#> 604 507.2236   3.003225
#> 605 500.0472   2.728440
#> 606 501.4312   3.094423
#> 607 562.8631   2.523673
#> 608 520.8855   3.015763
#> 609 538.9401   2.547113
#> 610 511.0393   2.546979
#> 611 508.3239   2.639059
#> 612 531.5453   3.318947
#> 613 500.1107   3.603129
#> 614 648.1597   3.588885
#> 615 502.2110   2.764655
#> 616 500.5283   2.775374
#> 617 549.7640   2.702230
#> 618 500.2576   2.807809
#> 619 510.6543   2.723777
#> 620 529.4603   2.833636
#> 621 517.3967   2.742437
#> 622 500.0948   2.535990
#> 623 511.7623   2.508314
#> 624 521.7277   2.615390
#> 625 500.9105   4.211171
#> 626 524.9925   2.919718
#> 627 557.9393   2.573555
#> 628 500.1546   2.551414
#> 629 501.7101   4.326955
#> 630 500.1482   2.737284
#> 631 501.0674   2.511316
#> 632 503.2290   2.737265
#> 633 505.7834   3.144768
#> 634 500.1585   3.457812
#> 635 500.0282   2.884663
#> 636 510.3915   3.802476
#> 637 500.2681   2.728890
#> 638 505.6500   2.911576
#> 639 501.7962   2.983669
#> 640 502.0042   3.266352
#> 641 504.3982   2.514162
#> 642 569.5683   2.619402
#> 643 501.6863   2.510742
#> 644 501.9372   3.247640
#> 645 507.1895   2.588006
#> 646 500.3516   2.920191
#> 647 564.9892   4.334212
#> 648 549.9742   2.799543
#> 649 503.9895   2.639411
#> 650 500.6647   2.835012
#> 651 500.0600   2.611327
#> 652 520.3414   3.362105
#> 653 501.7064   3.910046
#> 654 515.3292   2.848878
#> 655 500.0546   2.745549
#> 656 512.8173   3.255803
#> 657 639.3391   2.782147
#> 658 503.0958   2.690692
#> 659 999.8893   2.540392
#> 660 500.9983   2.656348
#> 661 500.1666   3.057438
#> 662 559.1925   2.574220
#> 663 510.5231   3.582658
#> 664 829.1829   2.792048
#> 665 572.8570   2.535554
#> 666 500.8005   2.680595
#> 667 500.4084   2.605846
#> 668 540.5011   2.512414
#> 669 547.2452   2.694161
#> 670 500.0212   2.613095
#> 671 510.4663   3.295475
#> 672 513.9215   2.901794
#> 673 514.2372   2.666390
#> 674 500.7691   2.657745
#> 675 574.8629   2.590050
#> 676 513.0746   2.588815
#> 677 500.0242   3.032337
#> 678 502.7787   2.598695
#> 679 500.5224   2.865053
#> 680 500.8422   2.672402
#> 681 503.6425   2.624265
#> 682 500.1231   2.788406
#> 683 500.0512   3.177337
#> 684 501.4925   3.068433
#> 685 500.2054   2.915188
#> 686 504.4858   3.053703
#> 687 501.3746   2.532164
#> 688 500.5463   2.636219
#> 689 500.0606   2.527061
#> 690 502.3594   3.024067
#> 691 541.3047   2.558331
#> 692 882.0988   2.763156
#> 693 500.0884   2.962931
#> 694 500.1675   2.665600
#> 695 923.1012   2.642015
#> 696 500.0506   2.806591
#> 697 731.3926   2.989119
#> 698 595.7182   2.622654
#> 699 501.9993   2.715356
#> 700 501.3483   2.654532
#> 701 534.3492   4.579617
#> 702 566.5052   2.515100
#> 703 511.0157   3.007619
#> 704 801.4225   2.893313
#> 705 603.1827   3.846057
#> 706 956.7463   3.158252
#> 707 500.0279   2.627209
#> 708 500.0128   3.181084
#> 709 500.2492   2.828306
#> 710 513.6235   2.652396
#> 711 535.8013   3.030562
#> 712 522.1347   3.153549
#> 713 552.4740   3.006481
#> 714 500.1896   2.733861
#> 715 715.1665   2.570396
#> 716 504.6581   2.568000
#> 717 776.9812   3.162980
#> 718 502.0992   3.223784
#> 719 500.2750   2.635306
#> 720 503.5776   3.289608
#> 721 611.6631   2.727471
#> 722 516.2209   2.718587
#> 723 500.4195   2.686786
#> 724 518.7938   3.247389
#> 725 500.5935   3.616102
#> 726 518.3997   2.521160
#> 727 500.4326   2.672690
#> 728 500.2052   2.524411
#> 729 500.1710   2.587471
#> 730 500.1308   2.603873
#> 731 500.2384   3.225111
#> 732 513.4759   3.035316
#> 733 561.8475   2.948801
#> 734 501.6770   2.910494
#> 735 518.9015   2.923638
#> 736 500.4073   2.712640
#> 737 531.8192   2.950976
#> 738 500.0239   2.736698
#> 739 537.3092   3.091060
#> 740 775.5965   2.770115
#> 741 505.4206   2.859692
#> 742 500.2414   2.828607
#> 743 760.1031   2.745942
#> 744 507.7612   2.521735
#> 745 530.4933   2.670743
#> 746 631.7947   3.510613
#> 747 503.0224   2.542350
#> 748 539.8941   3.120324
#> 749 503.2629   2.827749
#> 750 500.6856   3.015851
#> 751 644.1854   2.864796
#> 752 500.1103   3.190035
#> 753 542.6824   2.707868
#> 754 506.5730   2.662059
#> 755 501.7003   2.715302
#> 756 617.8296   3.958513
#> 757 517.0186   2.555541
#> 758 500.1681   2.744905
#> 759 500.2192   2.653078
#> 760 500.0147   2.724386
#> 761 500.1007   3.077653
#> 762 503.5099   2.947754
#> 763 505.6565   3.047015
#> 764 523.6642   2.760093
#> 765 551.0900   2.673602
#> 766 508.2046   4.067879
#> 767 500.0370   2.729717
#> 768 560.8941   2.923076
#> 769 544.2041   3.323118
#> 770 503.4615   3.036810
#> 771 523.8512   2.916740
#> 772 547.2170   2.947337
#> 773 900.9303   2.734650
#> 774 500.2669   2.710077
#> 775 533.6893   2.953723
#> 776 500.0095   2.693105
#> 777 504.7385   2.658669
#> 778 500.3023   2.549084
#> 779 500.1219   2.722105
#> 780 500.1106   3.483832
#> 781 501.8059   3.971789
#> 782 605.9687   2.699123
#> 783 814.4818   2.523122
#> 784 538.1184   4.657824
#> 785 501.5864   2.662104
#> 786 500.2154   2.909758
#> 787 545.4913   2.693916
#> 788 558.2854   3.093611
#> 789 500.0525   2.504170
#> 790 508.2598   3.771586
#> 791 503.3561   2.775155
#> 792 500.8990   3.609196
#> 793 501.4204   3.726766
#> 794 538.5751   2.801784
#> 795 956.1642   2.989973
#> 796 500.3632   3.248139
#> 797 523.7182   2.907081
#> 798 505.6045   2.673619
#> 799 503.6185   3.301656
#> 800 990.5648   4.753817
#> 801 510.2230   3.141825
#> 802 508.7068   2.532453
#> 803 803.9142   2.597791
#> 804 839.4420   3.021671
#> 805 500.0947   2.868845
#> 806 614.2682   2.935379
#> 807 506.8704   4.040710
#> 808 500.0185   2.648738
#> 809 508.3269   2.903025
#> 810 508.9093   3.433680
#> 811 516.7127   2.542069
#> 812 500.0597   2.628489
#> 813 500.0437   2.611545
#> 814 500.4912   2.697892
#> 815 951.4493   2.627950
#> 816 501.7691   2.819881
#> 817 587.1695   2.662109
#> 818 501.3095   3.105792
#> 819 793.6246   3.094548
#> 820 501.6665   2.525410
#> 821 593.0539   2.518087
#> 822 563.6794   2.790428
#> 823 517.0131   2.668638
#> 824 681.4722   2.683762
#> 825 500.0211   3.896339
#> 826 500.0100   2.921341
#> 827 504.7201   4.662230
#> 828 500.4671   2.647412
#> 829 501.7632   2.669094
#> 830 782.7399   3.068443
#> 831 505.7632   3.168754
#> 832 500.0965   2.837763
#> 833 848.8359   5.043000
#> 834 507.3662   2.503149
#> 835 902.4247   3.114674
#> 836 944.8199   3.771875
#> 837 525.4123   2.667630
#> 838 512.3376   2.704547
#> 839 528.2573   3.247371
#> 840 678.5715   3.040423
#> 841 500.1053   3.083439
#> 842 500.0858   2.771550
#> 843 502.2395   3.460271
#> 844 501.4459   2.695813
#> 845 500.0142   2.992073
#> 846 805.8749   2.918459
#> 847 500.2806   3.350092
#> 848 500.4086   2.910083
#> 849 511.5393   2.555433
#> 850 543.1141   2.673120
#> 851 500.1270   2.564683
#> 852 500.7792   2.520450
#> 853 558.9683   2.611652
#> 854 500.1739   2.717749
#> 855 552.4270   3.076162
#> 856 503.2310   4.319065
#> 857 500.0164   2.612457
#> 858 500.1602   2.525577
#> 859 500.1430   3.073440
#> 860 500.0872   2.584571
#> 861 500.1965   2.621550
#> 862 500.0447   2.510497
#> 863 500.0527   2.727007
#> 864 500.0728   2.616028
#> 865 609.1691   2.723367
#> 866 524.4153   3.853764
#> 867 502.6024   2.871293
#> 868 503.4047   3.578800
#> 869 509.3405   2.734741
#> 870 539.2038   2.780580
#> 871 521.2701   2.721760
#> 872 501.3285   2.798529
#> 873 500.2621   2.558269
#> 874 517.5293   3.489522
#> 875 505.8194   2.562299
#> 876 518.7168   3.176600
#> 877 500.1890   2.964208
#> 878 500.1768   2.645924
#> 879 796.4899   3.738351
#> 880 547.9277   2.614625
#> 881 716.6648   3.699512
#> 882 502.3289   2.645936
#> 883 537.7283   2.661736
#> 884 747.2634   2.672090
#> 885 508.0633   2.503354
#> 886 506.1790   2.996189
#> 887 500.0038   2.578764
#> 888 500.3266   3.255318
#> 889 506.9474   2.689604
#> 890 512.9018   2.582930
#> 891 509.8741   2.715024
#> 892 503.0262   2.785427
#> 893 500.5655   2.838918
#> 894 500.0181   2.530297
#> 895 500.7588   2.629255
#> 896 502.2914   2.838929
#> 897 500.0668   2.527918
#> 898 501.6986   3.200533
#> 899 740.4964   3.240717
#> 900 508.9174   2.897445
#> 901 550.5466   2.861378
#> 902 525.1289   3.305105
#> 903 534.6968   2.909695
#> 904 502.9002   2.984861
#> 905 555.4117   2.704230
#> 906 502.8020   2.553295
#> 907 507.6055   2.884372
#> 908 501.1878   3.068138
#> 909 500.1606   2.537450
#> 910 537.1316   3.021375
#> 911 500.6979   2.808017
#> 912 500.3015   2.616520
#> 913 500.0964   2.526291
#> 914 643.6930   2.992071
#> 915 950.0081   2.518713
#> 916 502.4768   2.739265
#> 917 500.2624   2.629589
#> 918 611.8132   2.678494
#> 919 500.0991   2.662521
#> 920 503.8936   3.164173