Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   824.4296   2.904637
#> 2   500.0044   2.503628
#> 3   500.2048   3.562306
#> 4   611.6499   2.607387
#> 5   534.9798   2.798465
#> 6   500.0037   2.997920
#> 7   531.9458   2.689464
#> 8   615.1042   2.932604
#> 9   501.0753   3.047092
#> 10  574.4797   2.542990
#> 11  506.8380   2.934863
#> 12  500.4225   3.328572
#> 13  616.0030   3.086416
#> 14  528.2508   3.185515
#> 15  506.7256   2.940384
#> 16  500.6198   2.607485
#> 17  502.3842   3.095644
#> 18  500.4632   3.040764
#> 19  502.4261   2.941064
#> 20  579.0497   2.736512
#> 21  500.2437   2.938768
#> 22  511.7648   2.555693
#> 23  500.7929   2.908122
#> 24  501.3857   2.829680
#> 25  500.0444   2.919386
#> 26  500.8602   3.222591
#> 27  502.4585   2.729275
#> 28  500.0355   3.028881
#> 29  534.4866   2.849612
#> 30  500.5476   2.528320
#> 31  536.5222   2.792691
#> 32  501.2514   3.187916
#> 33  516.3211   2.620714
#> 34  610.2782   2.775212
#> 35  732.0010   2.752253
#> 36  507.8344   2.974433
#> 37  512.3274   3.062010
#> 38  594.5989   2.869021
#> 39  500.0656   3.223694
#> 40  500.6874   2.797696
#> 41  510.0233   3.181488
#> 42  500.3173   4.969958
#> 43  749.7797   3.375901
#> 44  500.0474   3.869970
#> 45  501.1146   2.793059
#> 46  996.0863   2.508792
#> 47  502.5156   3.184327
#> 48  507.2519   2.531396
#> 49  500.3418   3.061137
#> 50  646.3317   2.765206
#> 51  599.1182   3.256753
#> 52  576.4921   2.578649
#> 53  500.0558   2.919077
#> 54  500.0538   4.006331
#> 55  502.1953   3.794840
#> 56  526.3133   2.756533
#> 57  501.8486   2.597020
#> 58  510.2120   2.840159
#> 59  503.0516   2.573542
#> 60  501.7379   3.086274
#> 61  510.7124   2.590051
#> 62  733.1850   2.625916
#> 63  500.0383   2.637245
#> 64  806.1191   3.599807
#> 65  627.5088   2.887452
#> 66  500.5351   2.805650
#> 67  503.5637   2.617039
#> 68  503.1982   2.578723
#> 69  613.8385   2.570253
#> 70  500.1650   3.080249
#> 71  502.3658   3.022732
#> 72  591.5481   2.836748
#> 73  590.7910   2.823095
#> 74  500.0298   2.638857
#> 75  689.4882   3.426415
#> 76  501.0780   2.694317
#> 77  500.2353   3.167549
#> 78  501.0897   2.505195
#> 79  500.1655   2.504656
#> 80  633.0200   2.776394
#> 81  500.0769   2.531469
#> 82  500.2838   2.542847
#> 83  572.2482   2.836334
#> 84  725.6549   2.944653
#> 85  500.7796   2.673310
#> 86  681.8487   3.656985
#> 87  554.0120   2.506799
#> 88  500.4379   2.628597
#> 89  505.7127   2.705813
#> 90  502.2837   2.696151
#> 91  597.7102   3.018359
#> 92  618.3649   2.829709
#> 93  529.2954   2.716883
#> 94  550.5542   3.584554
#> 95  500.0401   3.013984
#> 96  500.9977   2.500982
#> 97  501.7052   4.217087
#> 98  509.6487   3.746033
#> 99  530.9766   2.612223
#> 100 503.0758   3.986786
#> 101 500.0774   2.528960
#> 102 500.0758   2.674039
#> 103 500.6729   2.500675
#> 104 505.0218   2.595732
#> 105 503.7066   2.807549
#> 106 502.9556   2.635551
#> 107 575.6162   2.586629
#> 108 500.6262   2.528446
#> 109 502.5967   2.850458
#> 110 552.6283   2.839010
#> 111 501.9482   3.336478
#> 112 516.9146   3.201654
#> 113 500.2300   2.515280
#> 114 518.6590   2.516212
#> 115 500.1643   2.521105
#> 116 559.9414   2.794493
#> 117 503.0613   2.690794
#> 118 500.6134   2.529641
#> 119 583.0832   2.663509
#> 120 502.9130   4.997936
#> 121 500.4139   2.502357
#> 122 521.4812   2.639320
#> 123 503.9010   3.468081
#> 124 503.9858   2.866505
#> 125 500.0880   2.761424
#> 126 505.1419   2.741248
#> 127 546.4291   2.649822
#> 128 569.2387   2.666600
#> 129 500.0092   3.991193
#> 130 500.1087   2.579695
#> 131 500.8691   3.208640
#> 132 573.5161   2.564487
#> 133 506.9980   2.585093
#> 134 500.5092   3.377655
#> 135 500.0388   3.028382
#> 136 500.0167   3.097897
#> 137 516.9636   2.507704
#> 138 521.5376   2.547528
#> 139 500.2040   5.390264
#> 140 502.2469   3.042264
#> 141 834.3386   2.695599
#> 142 503.6126   3.591990
#> 143 515.8617   2.831046
#> 144 516.1131   2.674923
#> 145 517.6164   2.889726
#> 146 500.0225   2.538903
#> 147 841.9674   2.897317
#> 148 508.6524   2.879085
#> 149 512.6439   2.865836
#> 150 502.5337   2.649745
#> 151 500.8379   2.744516
#> 152 503.8589   3.189360
#> 153 504.0210   6.026063
#> 154 507.9932   2.982496
#> 155 500.0440   2.573297
#> 156 500.6432   2.854414
#> 157 500.1743   2.847444
#> 158 501.4964   2.973918
#> 159 500.0065   2.624726
#> 160 500.4599   2.735149
#> 161 502.6578   3.688571
#> 162 523.6830   3.464266
#> 163 501.7162   3.600215
#> 164 834.4670   3.400002
#> 165 500.7123   2.947809
#> 166 500.5751   2.818184
#> 167 650.4674   3.001597
#> 168 949.8135   2.554575
#> 169 500.0072   4.145222
#> 170 500.2405   2.569550
#> 171 500.7795   3.827577
#> 172 538.4536   2.504288
#> 173 918.0576   3.650719
#> 174 501.7553   3.195315
#> 175 573.4048   2.593757
#> 176 539.5236   2.520174
#> 177 500.6065   3.571523
#> 178 500.1964   2.691678
#> 179 500.1312   3.009435
#> 180 500.1919   2.633282
#> 181 500.4617   2.564265
#> 182 506.1671   2.602873
#> 183 500.0964   2.665532
#> 184 500.2579   2.638328
#> 185 526.7997   2.888404
#> 186 957.1753   4.535124
#> 187 500.0385   2.605925
#> 188 738.5285   2.680028
#> 189 504.7192   2.752458
#> 190 559.2870   2.889310
#> 191 520.2111   2.554953
#> 192 514.2411   2.849433
#> 193 501.7213   2.690208
#> 194 515.1866   2.512444
#> 195 895.6510   2.668474
#> 196 500.1155   3.544616
#> 197 500.9658   2.906489
#> 198 500.6433   3.015906
#> 199 601.0194   2.570364
#> 200 513.1650   2.581438
#> 201 500.0724   2.836419
#> 202 535.1096   3.006887
#> 203 643.8909   3.291954
#> 204 576.8062   2.605364
#> 205 502.8035   3.095388
#> 206 500.1391   3.494897
#> 207 501.2606   3.293267
#> 208 500.0504   3.220709
#> 209 500.0467   2.529234
#> 210 500.0076   2.574837
#> 211 560.1524   2.595337
#> 212 506.5761   4.397024
#> 213 502.1996   2.678615
#> 214 716.7088   2.691994
#> 215 500.5352   2.583764
#> 216 503.5222   2.805145
#> 217 500.1793   2.596993
#> 218 502.1807   2.854050
#> 219 500.2112   2.714524
#> 220 549.6277   2.697312
#> 221 500.5487   2.817767
#> 222 848.0571   2.774184
#> 223 505.2292   2.799026
#> 224 501.3245   2.636614
#> 225 500.7932   3.010863
#> 226 504.9643   2.544332
#> 227 500.2496   2.543229
#> 228 575.2539   2.515905
#> 229 549.7474   3.143640
#> 230 500.5512   2.598451
#> 231 534.9000   3.006140
#> 232 500.0141   2.533254
#> 233 500.1225   2.794583
#> 234 560.3697   3.471887
#> 235 500.3129   3.043054
#> 236 500.0091   2.671337
#> 237 508.0578   2.772004
#> 238 500.0461   3.150330
#> 239 500.7409   2.533118
#> 240 500.6920   2.650804
#> 241 500.2056   2.738119
#> 242 500.0215   2.734856
#> 243 504.9111   2.633854
#> 244 617.1223   3.376562
#> 245 539.8637   2.844092
#> 246 710.2770   3.052102
#> 247 519.3694   2.740764
#> 248 502.4250   2.766497
#> 249 504.1399   3.725207
#> 250 507.7901   2.732859
#> 251 753.4573   2.660427
#> 252 500.8078   2.566791
#> 253 594.2954   2.696445
#> 254 515.7145   2.880189
#> 255 500.0058   2.718727
#> 256 500.3436   2.944460
#> 257 500.1212   2.801308
#> 258 560.6524   3.017181
#> 259 501.4942   3.499802
#> 260 500.0494   2.727924
#> 261 764.6693   3.285050
#> 262 501.5109   2.749403
#> 263 820.1157   2.913248
#> 264 518.1861   2.962571
#> 265 502.4265   2.599203
#> 266 500.4742   2.556926
#> 267 505.7151   2.657214
#> 268 500.2281   2.560443
#> 269 509.9868   3.339461
#> 270 501.1276   2.909430
#> 271 500.3637   2.566822
#> 272 524.3030   3.398156
#> 273 620.8783   2.617385
#> 274 654.0402   3.464276
#> 275 940.0874   2.540446
#> 276 500.4112   4.176546
#> 277 503.3217   2.781612
#> 278 559.2902   2.587073
#> 279 504.0439   2.508646
#> 280 609.2632   2.707750
#> 281 799.4103   2.562009
#> 282 500.2620   2.737869
#> 283 500.3241   3.827455
#> 284 509.5441   2.507064
#> 285 687.2828   2.728914
#> 286 513.8693   2.790651
#> 287 501.4089   3.087128
#> 288 502.9353   3.040889
#> 289 503.8251   3.231109
#> 290 500.0331   3.458690
#> 291 524.7971   4.845859
#> 292 918.8725   2.756486
#> 293 500.6388   3.117376
#> 294 500.4805   2.585429
#> 295 510.0175   2.987663
#> 296 517.0502   2.575909
#> 297 526.7825   3.136574
#> 298 502.6731   2.506613
#> 299 500.9815   2.538740
#> 300 500.7121   3.627205
#> 301 501.9276   2.715825
#> 302 500.6054   2.629958
#> 303 602.8693   3.837152
#> 304 510.3182   2.851432
#> 305 507.4686   2.833051
#> 306 500.0748   3.749660
#> 307 563.3444   3.638029
#> 308 845.0316   2.510796
#> 309 500.3139   2.534743
#> 310 500.1700   3.958436
#> 311 500.2350   2.534250
#> 312 668.3567   4.403258
#> 313 543.7584   3.934958
#> 314 505.0502   2.887903
#> 315 502.8518   2.627528
#> 316 586.8585   2.930558
#> 317 503.4739   3.619118
#> 318 553.2088   2.998839
#> 319 500.2173   2.633300
#> 320 535.9850   2.866851
#> 321 505.7009   3.645679
#> 322 512.9644   2.566343
#> 323 503.7717   2.567335
#> 324 500.2749   2.970704
#> 325 500.6951   3.002390
#> 326 648.9742   2.584209
#> 327 500.0797   2.879322
#> 328 500.1166   2.889186
#> 329 500.5184   3.036878
#> 330 544.8170   3.938308
#> 331 500.0783   3.603182
#> 332 500.2384   2.663702
#> 333 500.9798   2.510353
#> 334 598.4827   3.732479
#> 335 500.0365   3.212856
#> 336 500.3173   2.609232
#> 337 502.2583   4.810690
#> 338 666.6294   3.718605
#> 339 518.2590   3.094968
#> 340 500.6148   2.876952
#> 341 502.1301   3.523187
#> 342 500.0692   2.529250
#> 343 503.3211   2.629775
#> 344 568.4826   2.612428
#> 345 532.5226   3.112414
#> 346 507.0263   3.343606
#> 347 512.6736   4.224579
#> 348 672.5243   4.389062
#> 349 501.9912   2.666465
#> 350 500.4395   2.717836
#> 351 501.4741   2.527498
#> 352 500.0946   2.587600
#> 353 500.0061   2.746001
#> 354 674.4228   3.985852
#> 355 526.1835   2.682776
#> 356 547.1088   2.810685
#> 357 559.0858   2.965068
#> 358 528.1859   2.535787
#> 359 500.3956   2.671945
#> 360 505.3590   3.179100
#> 361 500.1339   2.658357
#> 362 502.9915   2.612836
#> 363 567.4506   2.594709
#> 364 503.9136   2.572506
#> 365 527.2388   3.080817
#> 366 509.3120   2.727321
#> 367 505.5412   2.510385
#> 368 500.6428   2.818688
#> 369 500.0055   3.091043
#> 370 549.9152   3.502069
#> 371 501.1244   3.085439
#> 372 503.7075   2.673668
#> 373 517.5840   2.766986
#> 374 501.2474   2.967376
#> 375 614.9570   2.676426
#> 376 527.9371   2.842599
#> 377 506.2119   3.738176
#> 378 500.2594   3.141547
#> 379 501.0500   2.729002
#> 380 501.0466   3.927445
#> 381 502.9221   2.934090
#> 382 500.1324   2.697263
#> 383 564.2833   2.609624
#> 384 504.5784   2.540768
#> 385 519.0577   3.944275
#> 386 530.6528   2.565194
#> 387 558.0860   2.584993
#> 388 519.3830   3.618114
#> 389 500.1162   2.841108
#> 390 542.7116   2.829247
#> 391 824.6890   2.657951
#> 392 501.8093   5.534364
#> 393 683.6589   3.548375
#> 394 909.5032   4.558709
#> 395 500.0075   2.953684
#> 396 500.1365   2.650993
#> 397 859.7496   2.744210
#> 398 575.8939   3.734488
#> 399 500.5581   2.798551
#> 400 501.5609   2.510146
#> 401 500.4823   3.583527
#> 402 500.5971   2.681826
#> 403 501.1262   3.695302
#> 404 576.7425   2.585474
#> 405 500.2223   2.521895
#> 406 500.0973   3.040335
#> 407 500.0173   2.734205
#> 408 500.0511   2.653063
#> 409 505.9067   3.012623
#> 410 535.9477   3.168136
#> 411 501.2141   2.890024
#> 412 501.5403   2.653545
#> 413 523.5208   3.484656
#> 414 500.1076   2.942546
#> 415 500.4789   2.703980
#> 416 503.7064   2.553682
#> 417 500.1107   2.570753
#> 418 500.2335   2.574692
#> 419 755.6511   3.492642
#> 420 500.0467   2.614133
#> 421 550.6117   2.660394
#> 422 528.1210   2.846776
#> 423 500.0493   2.710357
#> 424 502.5875   2.995435
#> 425 505.2547   3.349462
#> 426 505.0071   3.255278
#> 427 500.6835   3.802839
#> 428 500.0688   2.622278
#> 429 501.4410   2.514316
#> 430 510.7543   2.867480
#> 431 518.1271   3.415241
#> 432 506.9072   2.569598
#> 433 912.0009   3.024328
#> 434 529.7164   2.609702
#> 435 500.0595   2.858909
#> 436 501.8637   2.747871
#> 437 501.1930   2.748198
#> 438 500.4929   2.731314
#> 439 512.6348   2.724831
#> 440 501.1852   3.114133
#> 441 574.2013   3.962086
#> 442 557.1139   3.126797
#> 443 500.3252   2.515660
#> 444 504.0866   2.856375
#> 445 508.1129   2.852823
#> 446 500.9497   2.611675
#> 447 878.7247   4.528434
#> 448 500.8169   2.525437
#> 449 502.7792   3.068813
#> 450 593.7066   3.026699
#> 451 500.0618   3.221949
#> 452 500.3136   2.888623
#> 453 501.1139   2.631318
#> 454 500.4683   2.816858
#> 455 556.6537   3.676350
#> 456 500.1463   2.651252
#> 457 658.0541   2.500283
#> 458 511.0016   4.072084
#> 459 510.5586   2.584329
#> 460 500.1256   3.239673
#> 461 535.5941   2.703185
#> 462 502.7300   2.714176
#> 463 500.0170   3.062900
#> 464 758.8396   2.592774
#> 465 538.5597   3.649583
#> 466 504.1737   2.578151
#> 467 510.7124   2.602931
#> 468 507.4009   3.456753
#> 469 519.8055   3.025635
#> 470 701.7981   2.732691
#> 471 509.7941   2.769976
#> 472 517.4700   2.606473
#> 473 500.0991   2.779824
#> 474 505.6702   3.465329
#> 475 501.2697   2.682856
#> 476 501.1627   2.523052
#> 477 936.9558   2.947301
#> 478 503.0822   2.854686
#> 479 516.7761   2.550266
#> 480 502.1481   3.115227
#> 481 500.0891   2.628497
#> 482 504.7302   2.808910
#> 483 505.4984   2.633107
#> 484 500.5661   2.994285
#> 485 500.0912   2.804205
#> 486 633.7378   2.633337
#> 487 500.7172   2.548813
#> 488 500.4298   2.874947
#> 489 509.3834   2.736223
#> 490 500.3820   4.477557
#> 491 850.8021   2.510527
#> 492 535.2182   2.987420
#> 493 508.2258   3.595019
#> 494 500.6423   3.050725
#> 495 527.0060   2.869434
#> 496 500.1866   2.600502
#> 497 500.0137   3.218102
#> 498 507.4219   2.650011
#> 499 602.7822   3.009006
#> 500 926.9435   2.690844
#> 501 501.2883   2.726516
#> 502 514.3721   2.536910
#> 503 638.9192   2.703138
#> 504 505.1121   2.504207
#> 505 512.1709   2.641368
#> 506 527.7036   3.178762
#> 507 504.0563   2.937374
#> 508 531.1832   2.520262
#> 509 621.7445   3.435882
#> 510 501.9804   2.680011
#> 511 500.6277   2.771770
#> 512 501.4134   3.017852
#> 513 500.9852   2.903280
#> 514 500.4984   3.327989
#> 515 624.1832   3.306320
#> 516 500.9520   2.945034
#> 517 532.3412   2.751897
#> 518 517.5464   2.547468
#> 519 507.9598   2.907586
#> 520 501.1830   2.955978
#> 521 500.3544   2.536104
#> 522 500.5318   2.862303
#> 523 500.1883   3.850963
#> 524 500.9027   2.523640
#> 525 590.8694   2.849613
#> 526 570.0417   3.034797
#> 527 500.8558   3.655397
#> 528 541.1405   4.319814
#> 529 500.0653   2.514236
#> 530 514.5844   2.520600
#> 531 501.7555   3.149631
#> 532 524.5292   2.631552
#> 533 518.2843   2.715653
#> 534 590.1419   2.709309
#> 535 500.8511   3.129081
#> 536 504.0344   2.652516
#> 537 501.2934   2.528960
#> 538 561.6299   2.539697
#> 539 500.1828   2.553671
#> 540 500.5037   2.864873
#> 541 590.1988   2.631014
#> 542 515.4884   3.690027
#> 543 500.0103   2.585823
#> 544 554.7981   2.521570
#> 545 500.2819   3.748891
#> 546 500.3776   4.413794
#> 547 512.9267   3.008420
#> 548 884.7166   3.092588
#> 549 500.0599   2.528974
#> 550 814.1289   2.860943
#> 551 500.4852   2.569369
#> 552 680.0028   2.731692
#> 553 505.4277   4.042529
#> 554 500.2693   3.017205
#> 555 500.7573   3.001054
#> 556 510.0574   2.849942
#> 557 500.0150   2.848358
#> 558 501.1682   2.558415
#> 559 500.2707   2.526076
#> 560 501.1972   2.828629
#> 561 500.0191   2.972712
#> 562 500.7465   2.529273
#> 563 500.4452   2.680056
#> 564 526.5037   2.823229
#> 565 528.1686   2.899254
#> 566 507.7324   2.558356
#> 567 501.0377   3.599453
#> 568 502.1953   2.524280
#> 569 500.4232   3.007684
#> 570 715.2793   2.615678
#> 571 501.5625   2.793077
#> 572 500.2046   3.298954
#> 573 925.6805   2.679331
#> 574 656.6875   3.025036
#> 575 555.9157   2.597362
#> 576 534.7760   2.816305
#> 577 500.2040   3.965478
#> 578 659.1674   2.974767
#> 579 648.6646   2.992182
#> 580 500.4297   2.676542
#> 581 505.0444   2.616110
#> 582 529.3336   2.672942
#> 583 571.7930   3.380596
#> 584 500.1568   3.087031
#> 585 500.0995   2.754877
#> 586 690.7793   3.124989
#> 587 516.9268   2.803621
#> 588 635.5186   3.209711
#> 589 500.0674   2.540109
#> 590 611.5078   2.814966
#> 591 505.4102   2.833607
#> 592 521.6001   3.804093
#> 593 676.0842   2.508151
#> 594 510.2874   2.591041
#> 595 501.3405   2.630470
#> 596 501.5973   2.733991
#> 597 874.3648   2.666817
#> 598 504.7641   3.986693
#> 599 501.7188   2.555890
#> 600 506.6351   2.579509
#> 601 504.2569   2.762361
#> 602 597.1763   2.794855
#> 603 501.1305   2.555351
#> 604 585.9843   3.516571
#> 605 501.9653   2.911250
#> 606 830.5563   2.654890
#> 607 985.6020   2.801263
#> 608 500.1395   2.861272
#> 609 500.0651   2.957665
#> 610 503.9266   2.748597
#> 611 500.1055   2.851891
#> 612 505.8632   2.718407
#> 613 586.3397   2.509411
#> 614 501.1087   2.538113
#> 615 528.3707   2.598527
#> 616 989.7110   2.809273
#> 617 500.0380   3.408376
#> 618 604.4274   2.688647
#> 619 513.0225   2.652344
#> 620 519.4142   2.909906
#> 621 505.3063   3.225537
#> 622 542.9263   2.795550
#> 623 500.8283   3.553440
#> 624 500.2355   2.643225
#> 625 948.3053   2.522742
#> 626 503.4538   2.694223
#> 627 551.1980   2.825260
#> 628 512.0229   2.728367
#> 629 591.8821   2.757878
#> 630 627.4405   2.877726
#> 631 500.3296   2.729168
#> 632 500.4171   2.724906
#> 633 504.1089   2.504643
#> 634 500.8695   3.526588
#> 635 547.1204   2.512471
#> 636 541.1811   4.037319
#> 637 500.3340   2.526496
#> 638 500.4336   3.115463
#> 639 500.0792   3.286342
#> 640 500.4810   2.540604
#> 641 504.6983   3.913501
#> 642 572.5735   2.943173
#> 643 676.5162   2.830444
#> 644 519.8628   2.769368
#> 645 516.9042   3.038317
#> 646 501.3975   2.583715
#> 647 513.9658   2.779191
#> 648 610.6955   3.126611
#> 649 501.0495   2.849617
#> 650 639.8291   2.571180
#> 651 500.0104   2.625540
#> 652 500.0921   2.525275
#> 653 553.2683   2.581491
#> 654 946.8503   2.549061
#> 655 500.6723   3.364098
#> 656 500.1404   3.312360
#> 657 779.8345   2.546954
#> 658 516.3018   2.520254
#> 659 500.1140   2.682008
#> 660 501.4612   2.874934
#> 661 576.1437   2.678410
#> 662 500.0421   2.536650
#> 663 502.2554   2.579787
#> 664 508.9367   2.842047
#> 665 500.9933   2.555593
#> 666 501.4086   2.520998
#> 667 500.5999   3.104276
#> 668 501.5524   4.905193
#> 669 500.0216   2.852259
#> 670 505.3811   2.692761
#> 671 500.5432   2.613637
#> 672 501.2938   2.508818
#> 673 636.4693   3.634259
#> 674 500.2026   2.681481
#> 675 503.5783   2.791880
#> 676 500.5412   2.878080
#> 677 500.0304   2.588260
#> 678 506.9759   2.656019
#> 679 501.2553   3.394150
#> 680 677.7449   2.599568
#> 681 652.5325   2.662619
#> 682 500.8369   2.832681
#> 683 529.8879   2.995772
#> 684 574.6410   2.576517
#> 685 540.1223   3.319400
#> 686 512.1856   2.768431
#> 687 850.7864   2.964961
#> 688 501.6250   2.517670
#> 689 500.5843   2.701593
#> 690 502.4613   3.237101
#> 691 500.2651   2.614115
#> 692 500.0294   2.793993
#> 693 527.7095   2.688694
#> 694 514.2410   3.249544
#> 695 501.1485   2.758623
#> 696 500.2531   2.599016
#> 697 505.8154   3.009661
#> 698 500.5443   2.787204
#> 699 526.3947   3.337115
#> 700 500.0651   2.620416
#> 701 768.3517   2.698039
#> 702 794.6463   2.555307
#> 703 501.4585   2.824158
#> 704 736.9375   3.149488
#> 705 500.7900   2.649539
#> 706 500.0789   3.215791
#> 707 626.7284   2.847953
#> 708 505.4348   3.292124
#> 709 514.3551   2.980342
#> 710 500.5377   2.514669
#> 711 503.5652   3.132857
#> 712 500.2649   2.976392
#> 713 600.2804   2.779591
#> 714 500.0379   2.852308
#> 715 500.2295   2.719981
#> 716 520.7353   2.681594
#> 717 502.2227   2.835516
#> 718 515.1539   2.786304
#> 719 511.1116   2.582743
#> 720 501.6506   2.964699
#> 721 502.7936   7.222986
#> 722 501.8574   2.527927
#> 723 563.0944   2.948700
#> 724 514.4486   2.913589
#> 725 503.9833   3.538407
#> 726 500.0194   2.802606
#> 727 897.4698   4.163847
#> 728 652.3954   2.649528
#> 729 500.6756   3.675008
#> 730 511.4973   2.601174
#> 731 500.0180   2.764482
#> 732 517.0430   2.659647
#> 733 500.0307   2.558077
#> 734 748.9550   3.113456
#> 735 500.5770   2.586582
#> 736 504.5731   2.658875
#> 737 542.1214   2.687179
#> 738 509.0817   3.018431
#> 739 500.3615   2.716657
#> 740 517.0296   2.579133
#> 741 503.2998   3.034042
#> 742 500.0397   2.801990
#> 743 501.3858   2.562065
#> 744 501.5828   3.172430
#> 745 565.9607   4.002366
#> 746 524.1276   3.367847
#> 747 500.1485   3.148275
#> 748 500.3723   2.774075
#> 749 508.0369   3.051593
#> 750 918.9603   2.802660
#> 751 507.1382   2.803264
#> 752 505.0021   2.856614
#> 753 500.9321   2.769260
#> 754 686.9968   3.330766
#> 755 501.0775   2.649601
#> 756 500.3831   3.046403
#> 757 500.8497   2.533377
#> 758 502.5113   2.680920
#> 759 500.3274   2.848718
#> 760 501.4285   2.756648
#> 761 500.2024   2.511871
#> 762 590.1364   3.292884
#> 763 502.0811   2.804426
#> 764 537.5504   2.934660
#> 765 866.2205   3.034279
#> 766 524.4274   3.417022
#> 767 500.0084   2.923624
#> 768 500.8085   3.464259
#> 769 501.6496   2.563605
#> 770 500.3402   2.704654
#> 771 500.0140   2.722807
#> 772 540.3227   2.675427
#> 773 500.3379   3.171330
#> 774 897.1486   4.012605
#> 775 841.3687   3.736737
#> 776 527.3043   2.636880
#> 777 500.1232   4.437972
#> 778 501.4668   2.600512
#> 779 505.0232   2.569034
#> 780 500.0452   2.578375
#> 781 508.3264   3.860424
#> 782 620.3113   2.791193
#> 783 550.2819   2.511812
#> 784 500.0388   2.545130
#> 785 500.2031   2.798283
#> 786 500.6718   2.719143
#> 787 881.8512   2.533321
#> 788 500.1996   2.606354
#> 789 545.5047   2.708744
#> 790 550.7995   3.313938
#> 791 500.3145   2.737922
#> 792 531.1331   2.729920
#> 793 518.6789   5.122923
#> 794 547.5393   2.651714
#> 795 504.5128   4.067501
#> 796 743.9594   2.618129
#> 797 503.7429   2.749147
#> 798 500.0578   2.844747
#> 799 500.2809   2.551194
#> 800 500.4380   2.722233
#> 801 576.2975   2.675405
#> 802 514.1009   2.541079
#> 803 501.4925   2.811121
#> 804 600.1017   2.724561
#> 805 500.4938   2.718365
#> 806 507.8105   2.548214
#> 807 557.3947   3.861163
#> 808 601.4173   2.553035
#> 809 564.5104   3.420954
#> 810 549.8945   2.740696
#> 811 500.1690   2.918493
#> 812 501.4929   3.292739
#> 813 500.0633   2.665097
#> 814 582.0045   2.541728
#> 815 505.8407   2.603371
#> 816 501.0493   2.526708
#> 817 502.9144   3.237254
#> 818 506.4728   3.196678
#> 819 500.7162   3.136145
#> 820 559.2081   2.602162
#> 821 514.0412   2.934428
#> 822 700.5137   2.628887
#> 823 500.3557   2.672022
#> 824 504.3813   2.597235
#> 825 500.2663   2.966246
#> 826 502.8149   3.311398
#> 827 501.7081   2.835854
#> 828 552.8401   2.634408
#> 829 504.3634   3.826157
#> 830 532.6482   3.221479
#> 831 565.2445   2.618579
#> 832 501.0441   2.624174
#> 833 510.5599   2.676041
#> 834 500.1111   2.805417
#> 835 500.2576   2.528595
#> 836 500.2171   2.991787
#> 837 500.7502   2.639046
#> 838 508.3820   2.556454
#> 839 723.6078   2.695996
#> 840 572.5499   2.859076
#> 841 500.1383   3.507091
#> 842 500.8709   3.438990
#> 843 540.7489   2.662232
#> 844 500.0349   2.897096
#> 845 526.3725   2.612879
#> 846 554.4499   2.881823
#> 847 962.9955   2.626334
#> 848 506.5421   3.447950
#> 849 585.8880   2.778952
#> 850 525.2438   2.555480
#> 851 501.5371   3.084815
#> 852 705.3597   2.639119
#> 853 984.9357   2.703293
#> 854 609.8377   2.578313
#> 855 674.0899   2.518190
#> 856 501.6828   2.575319
#> 857 500.1537   4.291300
#> 858 546.3710   3.668140
#> 859 500.5318   3.149582
#> 860 502.7172   2.666535
#> 861 500.4217   2.553361
#> 862 502.2581   2.782043
#> 863 571.8114   3.009670
#> 864 500.4908   2.691605
#> 865 500.2092   2.677313
#> 866 500.1353   3.301702
#> 867 746.9184   2.917686
#> 868 514.6405   2.676420
#> 869 508.6638   2.559894
#> 870 503.7013   2.808807
#> 871 500.1501   2.971782
#> 872 653.7341   3.000899
#> 873 530.6330   2.547904
#> 874 520.3469   3.018430
#> 875 501.0519   2.754940
#> 876 501.8825   2.531086
#> 877 500.2558   3.038099
#> 878 502.8363   3.234223
#> 879 500.0228   3.920177
#> 880 505.1956   3.325659
#> 881 505.9361   3.733778
#> 882 681.7125   3.515259
#> 883 527.9436   3.388809
#> 884 500.0014   2.877260
#> 885 502.0865   2.996803
#> 886 840.7030   3.025462
#> 887 502.6566   3.144881
#> 888 500.1574   2.520699
#> 889 510.0041   3.073603
#> 890 500.0012   2.585873
#> 891 500.0170   2.632997
#> 892 500.1971   3.366667
#> 893 500.9641   2.718618
#> 894 530.0226   3.165944
#> 895 503.5989   2.741465
#> 896 500.7978   3.583914
#> 897 509.6230   3.332962
#> 898 501.3613   2.762210
#> 899 501.2704   3.190970
#> 900 672.4974   2.860232
#> 901 502.7433   6.852627
#> 902 504.6425   3.081329
#> 903 572.2925   2.569388
#> 904 604.2970   2.597638
#> 905 503.5677   2.617001
#> 906 502.6949   3.285519
#> 907 563.3260   2.794154
#> 908 538.5002   2.947571
#> 909 510.0750   2.813526
#> 910 500.7267   2.723732
#> 911 500.7932   2.660402
#> 912 507.9184   3.063705
#> 913 503.1148   3.037844
#> 914 505.3743   2.709545