Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   500.0717   3.453707
#> 2   506.0749   2.529286
#> 3   516.3410   3.426739
#> 4   502.1214   3.174202
#> 5   502.3496   3.230961
#> 6   512.5852   2.750375
#> 7   501.1835   2.965594
#> 8   512.8956   3.878631
#> 9   500.1337   2.880985
#> 10  500.2200   2.591981
#> 11  500.7164   3.245264
#> 12  561.8029   2.675096
#> 13  508.1163   2.675352
#> 14  500.0602   3.999008
#> 15  568.1226   3.575782
#> 16  501.2717   2.542212
#> 17  505.1563   2.503378
#> 18  549.7401   3.866972
#> 19  503.2262   3.512208
#> 20  501.8759   3.003222
#> 21  500.1004   2.576449
#> 22  505.5560   2.822620
#> 23  500.0215   2.660315
#> 24  529.5737   2.791504
#> 25  560.4576   6.276101
#> 26  501.5590   3.126955
#> 27  507.8161   2.922698
#> 28  500.0929   2.645258
#> 29  604.0835   3.362029
#> 30  547.2289   4.273454
#> 31  622.9816   2.757150
#> 32  518.2629   2.925576
#> 33  506.3781   2.636429
#> 34  500.1045   3.840160
#> 35  505.1267   3.922096
#> 36  501.8180   2.924844
#> 37  500.3884   2.584050
#> 38  500.0287   2.805132
#> 39  501.1986   2.749444
#> 40  506.4577   3.296246
#> 41  500.2678   3.982940
#> 42  500.1125   3.014698
#> 43  500.6836   3.060919
#> 44  500.1441   3.113571
#> 45  587.4116   2.649692
#> 46  502.8802   2.852479
#> 47  698.6924   2.867545
#> 48  500.1344   2.681538
#> 49  500.0254   2.748629
#> 50  516.8511   3.014539
#> 51  501.6246   2.928749
#> 52  500.5201   2.612579
#> 53  624.8629   2.875905
#> 54  845.5875   2.767620
#> 55  552.3102   3.678264
#> 56  510.0117   2.877688
#> 57  596.9706   2.530618
#> 58  650.8277   4.225801
#> 59  664.2730   2.516287
#> 60  500.7296   2.546108
#> 61  500.2559   2.552475
#> 62  526.8880   2.980740
#> 63  536.6431   3.322838
#> 64  739.2400   2.892341
#> 65  524.3895   2.514960
#> 66  500.1525   2.627163
#> 67  643.6469   3.022966
#> 68  500.0497   2.771565
#> 69  502.3029   3.057833
#> 70  509.7369   2.777163
#> 71  500.4821   2.543549
#> 72  500.2002   2.871168
#> 73  505.2733   3.737428
#> 74  500.0650   2.843195
#> 75  500.0314   2.694367
#> 76  500.0675   2.698897
#> 77  500.3830   2.700967
#> 78  504.9436   2.936576
#> 79  500.0601   2.504675
#> 80  687.4828   4.013515
#> 81  509.9752   2.529638
#> 82  500.1253   2.795661
#> 83  500.0056   3.824551
#> 84  660.2210   3.293338
#> 85  504.3134   2.713807
#> 86  503.2392   2.812662
#> 87  500.1015   2.586029
#> 88  501.1067   2.626274
#> 89  500.8063   2.590968
#> 90  500.6235   2.839161
#> 91  500.0316   4.919135
#> 92  500.9585   3.238806
#> 93  735.6464   3.200827
#> 94  534.7870   2.621557
#> 95  500.7507   4.584412
#> 96  500.4322   2.667790
#> 97  500.0662   2.706199
#> 98  502.7013   2.566986
#> 99  500.2593   2.577214
#> 100 504.6771   3.078698
#> 101 500.0868   2.562237
#> 102 522.7759   2.757345
#> 103 518.9310   2.624212
#> 104 500.9289   2.825999
#> 105 500.1442   2.972981
#> 106 500.0646   2.527181
#> 107 605.1972   3.565323
#> 108 501.9452   2.568555
#> 109 510.1841   2.602218
#> 110 500.0035   2.520814
#> 111 500.1260   2.820689
#> 112 500.2142   3.338204
#> 113 500.1429   3.203350
#> 114 500.4042   2.811102
#> 115 503.9594   2.787906
#> 116 501.7121   2.504941
#> 117 500.0075   2.732289
#> 118 500.3199   2.759778
#> 119 581.3335   2.555747
#> 120 505.7954   3.276966
#> 121 500.4134   2.898418
#> 122 500.2395   3.332119
#> 123 500.2586   2.676535
#> 124 952.8005   3.876072
#> 125 507.7374   2.852140
#> 126 500.0759   3.494795
#> 127 501.5792   2.522097
#> 128 500.1031   2.525891
#> 129 549.6548   2.617949
#> 130 500.4104   3.480905
#> 131 505.6270   2.980113
#> 132 849.3889   4.078952
#> 133 557.0826   2.971443
#> 134 503.8468   3.219703
#> 135 513.3607   2.535134
#> 136 504.2148   3.101790
#> 137 939.4436   3.258852
#> 138 500.1580   2.654896
#> 139 500.0079   2.858392
#> 140 515.6164   3.015800
#> 141 500.4393   3.177124
#> 142 657.2280   3.235162
#> 143 500.0130   2.867578
#> 144 899.2475   2.597263
#> 145 501.0858   2.788444
#> 146 500.1109   2.537485
#> 147 697.3846   3.647115
#> 148 507.5607   2.986524
#> 149 780.0773   3.261970
#> 150 527.2769   3.123584
#> 151 501.2794   2.603042
#> 152 500.2294   2.772446
#> 153 508.1802   2.957318
#> 154 602.5042   3.710434
#> 155 530.6086   2.545729
#> 156 528.8240   2.771289
#> 157 999.3802   3.749975
#> 158 520.5470   2.568214
#> 159 511.2973   2.695098
#> 160 502.0136   3.112070
#> 161 533.5898   3.010525
#> 162 537.8894   2.902829
#> 163 500.2855   3.462181
#> 164 716.5296   2.601028
#> 165 502.7853   2.524135
#> 166 584.4195   4.828081
#> 167 500.6677   4.126749
#> 168 511.3998   2.677130
#> 169 539.3545   2.511768
#> 170 500.1595   3.118983
#> 171 500.6125   2.605345
#> 172 743.9896   2.806542
#> 173 502.8122   3.239850
#> 174 511.6211   2.754167
#> 175 500.0210   2.650452
#> 176 504.2870   3.481934
#> 177 749.3790   3.413402
#> 178 502.3012   2.561737
#> 179 500.2200   2.774719
#> 180 500.8433   3.084428
#> 181 509.1451   3.404427
#> 182 570.2851   3.250004
#> 183 909.8775   3.119626
#> 184 501.3461   3.359405
#> 185 506.7520   2.940531
#> 186 957.4850   4.034864
#> 187 582.7395   2.516748
#> 188 502.4455   2.595839
#> 189 581.0442   4.220864
#> 190 500.0069   2.533714
#> 191 690.7149   2.898934
#> 192 502.3406   3.628957
#> 193 654.3754   4.127757
#> 194 500.3752   2.592071
#> 195 691.2102   3.255963
#> 196 535.0081   2.525262
#> 197 505.3451   2.536402
#> 198 502.8826   4.280910
#> 199 746.9444   2.833670
#> 200 607.7496   2.743920
#> 201 500.2311   3.066688
#> 202 512.4035   2.509388
#> 203 500.1731   3.092312
#> 204 506.7850   2.861804
#> 205 502.0690   3.213544
#> 206 500.6755   3.256082
#> 207 812.9303   2.857160
#> 208 500.3234   2.739586
#> 209 542.6052   3.079441
#> 210 503.6086   2.602474
#> 211 500.3426   2.990737
#> 212 524.7825   2.828104
#> 213 500.2685   2.524985
#> 214 504.3884   2.584638
#> 215 500.5970   2.677392
#> 216 504.3492   2.662315
#> 217 500.4797   3.289944
#> 218 524.7800   2.798315
#> 219 500.8802   2.501258
#> 220 501.8621   2.689807
#> 221 500.6754   2.506029
#> 222 593.9593   2.985496
#> 223 598.4565   2.632531
#> 224 500.0209   2.690942
#> 225 503.6857   2.801385
#> 226 501.3840   2.717143
#> 227 500.0765   3.104896
#> 228 510.4952   2.997495
#> 229 500.7852   2.581962
#> 230 697.3344   2.739533
#> 231 501.0571   3.139646
#> 232 538.1203   2.988208
#> 233 673.7881   2.784061
#> 234 526.7652   2.544372
#> 235 500.0474   2.975046
#> 236 591.9784   2.695030
#> 237 608.1392   2.521250
#> 238 503.8255   2.819657
#> 239 517.2993   3.771731
#> 240 500.0396   3.354404
#> 241 500.0620   3.181469
#> 242 500.8694   2.575961
#> 243 564.8961   2.937534
#> 244 502.0319   2.745673
#> 245 504.2723   3.010350
#> 246 530.6165   3.744146
#> 247 500.5113   2.600277
#> 248 500.4337   3.447394
#> 249 537.0884   3.009539
#> 250 500.1300   4.366351
#> 251 500.6442   2.619873
#> 252 838.5042   3.296875
#> 253 502.2900   2.865756
#> 254 502.6407   3.327178
#> 255 511.3193   3.029974
#> 256 500.3676   3.134517
#> 257 501.9215   4.395136
#> 258 500.0957   2.574446
#> 259 500.4351   2.580949
#> 260 502.4115   2.578077
#> 261 502.9002   3.566210
#> 262 651.9349   3.012166
#> 263 503.4724   2.706484
#> 264 500.6696   2.819966
#> 265 500.6843   3.489902
#> 266 500.2333   2.678603
#> 267 500.0578   4.424906
#> 268 500.7005   3.221507
#> 269 500.9025   2.558645
#> 270 509.9863   2.550340
#> 271 785.3392   2.669033
#> 272 507.9010   3.000490
#> 273 500.0500   3.771966
#> 274 935.8343   2.706808
#> 275 529.5627   3.322572
#> 276 964.2172   2.634403
#> 277 560.8036   3.087899
#> 278 613.7336   2.530096
#> 279 500.7987   3.141701
#> 280 500.7463   2.561893
#> 281 500.5408   3.612086
#> 282 507.0553   2.839386
#> 283 500.1188   2.589585
#> 284 671.7984   3.029164
#> 285 504.0983   2.651355
#> 286 500.2218   2.873037
#> 287 584.9894   2.600956
#> 288 578.4510   4.401286
#> 289 500.3605   2.536749
#> 290 503.8051   2.592987
#> 291 736.2800   2.625949
#> 292 715.6761   2.620462
#> 293 509.8882   2.696933
#> 294 500.6773   3.372236
#> 295 501.2511   3.032029
#> 296 500.1285   2.685849
#> 297 500.0946   2.718117
#> 298 569.3822   3.330695
#> 299 519.7821   2.909514
#> 300 522.1809   2.848821
#> 301 519.0032   2.589482
#> 302 500.4026   2.995296
#> 303 500.3379   2.599294
#> 304 529.6715   2.742419
#> 305 500.1728   3.441325
#> 306 500.0952   2.521750
#> 307 500.2107   3.499099
#> 308 537.5172   2.519626
#> 309 617.7911   2.514584
#> 310 500.2390   3.035725
#> 311 597.9621   2.647188
#> 312 812.3976   2.822488
#> 313 503.7035   2.607946
#> 314 522.5498   2.702323
#> 315 500.0242   2.790227
#> 316 501.6468   2.739210
#> 317 820.0041   2.619937
#> 318 506.8198   3.971992
#> 319 504.4688   2.721397
#> 320 500.9794   3.232102
#> 321 638.8177   3.459139
#> 322 952.3991   2.901361
#> 323 503.8397   3.281490
#> 324 500.5540   3.031240
#> 325 510.4341   2.914478
#> 326 500.2409   2.559534
#> 327 500.0708   2.621474
#> 328 500.1781   2.672474
#> 329 503.9213   2.859091
#> 330 500.0395   2.796704
#> 331 503.3500   2.835858
#> 332 500.5857   2.732817
#> 333 500.3170   4.148115
#> 334 517.1171   2.561682
#> 335 635.7574   2.604901
#> 336 507.4556   2.601981
#> 337 519.7470   2.845813
#> 338 740.1841   3.340809
#> 339 500.0725   2.792592
#> 340 503.3541   2.744470
#> 341 631.6661   2.571209
#> 342 504.1528   2.500621
#> 343 502.8619   2.662177
#> 344 500.0214   2.937944
#> 345 501.3220   3.082927
#> 346 513.1323   2.804631
#> 347 500.1719   3.201421
#> 348 500.9110   2.544485
#> 349 500.0167   3.432632
#> 350 513.6375   3.645755
#> 351 762.9941   2.929590
#> 352 501.6672   2.869965
#> 353 519.7294   3.459364
#> 354 517.8919   2.626446
#> 355 513.4228   2.604451
#> 356 500.1553   2.609077
#> 357 502.0958   3.061470
#> 358 565.3216   2.828962
#> 359 591.7238   2.710911
#> 360 517.7610   2.804905
#> 361 501.0458   2.875400
#> 362 500.3934   2.557672
#> 363 500.7865   3.011663
#> 364 500.1256   2.928465
#> 365 502.5870   3.371744
#> 366 504.1514   2.721888
#> 367 501.7055   2.508914
#> 368 505.1841   2.697247
#> 369 500.5957   2.983005
#> 370 516.8893   3.365529
#> 371 502.9785   3.254286
#> 372 501.0465   2.720839
#> 373 500.2652   2.604519
#> 374 512.6087   2.946172
#> 375 747.7676   2.784033
#> 376 502.9212   2.829413
#> 377 510.2947   3.166143
#> 378 517.7753   3.096165
#> 379 507.1268   2.908039
#> 380 501.7242   2.727349
#> 381 500.1332   2.630867
#> 382 500.0755   2.680610
#> 383 500.8264   6.105941
#> 384 500.3735   2.944838
#> 385 764.8490   2.668506
#> 386 501.3693   2.647305
#> 387 500.0196   2.556291
#> 388 500.8614   2.886664
#> 389 530.3864   2.559336
#> 390 525.9736   4.168153
#> 391 696.9971   2.811017
#> 392 501.0771   2.814390
#> 393 500.0091   3.231032
#> 394 500.0136   2.539618
#> 395 515.0360   3.450079
#> 396 747.2252   2.502153
#> 397 510.5733   2.764726
#> 398 610.7412   2.880409
#> 399 500.2232   3.184637
#> 400 500.0274   3.424069
#> 401 555.6243   2.656881
#> 402 507.3192   2.585540
#> 403 915.9260   3.142708
#> 404 547.0321   3.720351
#> 405 541.6295   3.014297
#> 406 500.6276   2.747268
#> 407 993.2813   2.607915
#> 408 500.1665   2.744274
#> 409 582.1392   3.190746
#> 410 500.0112   2.509671
#> 411 500.0126   3.010292
#> 412 500.2049   2.814569
#> 413 521.3971   3.731056
#> 414 504.6374   2.561887
#> 415 657.7165   2.539676
#> 416 553.2551   2.567428
#> 417 942.6823   2.582530
#> 418 552.6477   2.875706
#> 419 503.3626   2.546187
#> 420 571.6431   2.569597
#> 421 583.5381   3.719871
#> 422 501.1510   2.530868
#> 423 500.8382   2.941573
#> 424 502.2185   4.231973
#> 425 500.0025   2.915349
#> 426 544.7895   2.580764
#> 427 551.0074   2.787058
#> 428 965.5944   3.059514
#> 429 546.2589   3.105389
#> 430 506.4902   2.653913
#> 431 500.3570   2.745562
#> 432 511.3728   2.762814
#> 433 501.2083   2.595544
#> 434 500.4913   4.397412
#> 435 679.5535   2.510461
#> 436 683.0947   3.225581
#> 437 500.5198   2.630009
#> 438 500.0965   3.678189
#> 439 502.0442   2.936710
#> 440 500.3324   3.049316
#> 441 506.0229   2.575060
#> 442 512.6805   2.549186
#> 443 500.6214   3.508419
#> 444 500.0560   2.544235
#> 445 529.0074   2.660304
#> 446 688.5242   2.516122
#> 447 501.5657   3.044979
#> 448 500.1900   2.711620
#> 449 507.9947   2.908489
#> 450 500.0707   2.620181
#> 451 536.6590   2.888422
#> 452 514.7259   3.057380
#> 453 511.2936   2.600885
#> 454 500.2869   2.520475
#> 455 511.8090   2.715919
#> 456 536.7047   3.008004
#> 457 500.0480   2.980669
#> 458 622.4769   2.639082
#> 459 500.0559   3.079627
#> 460 500.0560   2.561305
#> 461 540.1189   3.042211
#> 462 502.6783   3.243754
#> 463 500.0740   2.845753
#> 464 869.6736   2.574489
#> 465 676.8067   2.501293
#> 466 500.1144   3.069313
#> 467 500.0182   2.859768
#> 468 511.7678   2.873943
#> 469 501.1140   2.510570
#> 470 501.9854   2.771747
#> 471 501.0033   2.635835
#> 472 506.7687   3.050306
#> 473 506.7578   3.600629
#> 474 539.8697   3.127426
#> 475 515.4804   2.556169
#> 476 567.1111   2.597622
#> 477 500.8136   4.081328
#> 478 897.8623   2.593751
#> 479 502.9617   2.622018
#> 480 770.5566   2.541044
#> 481 573.3299   2.703000
#> 482 501.2274   2.842335
#> 483 500.3259   2.547604
#> 484 502.9003   3.404123
#> 485 500.2517   4.805275
#> 486 507.2176   3.720035
#> 487 757.4920   2.790087
#> 488 500.1295   2.887646
#> 489 571.5333   2.625211
#> 490 601.5823   2.997998
#> 491 545.3659   2.674529
#> 492 500.2673   2.668516
#> 493 501.0005   2.814781
#> 494 500.0647   2.561945
#> 495 502.1544   2.615332
#> 496 660.1273   2.624816
#> 497 500.0344   2.784005
#> 498 501.8510   3.811851
#> 499 516.9570   2.617718
#> 500 508.2324   2.665957
#> 501 954.0769   2.507258
#> 502 513.9400   2.740342
#> 503 529.1898   2.716385
#> 504 504.1967   2.501456
#> 505 503.3231   2.677035
#> 506 500.7206   3.075138
#> 507 511.8188   2.898027
#> 508 500.0284   2.525671
#> 509 503.5587   2.777250
#> 510 502.0927   2.982671
#> 511 561.0042   2.593883
#> 512 501.9426   2.947500
#> 513 500.4233   2.663855
#> 514 536.2922   3.090624
#> 515 500.0652   2.524999
#> 516 513.2498   2.579275
#> 517 589.2059   2.742504
#> 518 500.2142   3.707108
#> 519 500.5537   2.956937
#> 520 543.6508   4.041690
#> 521 500.0734   2.860012
#> 522 720.0034   2.762549
#> 523 524.6512   2.978998
#> 524 500.1300   3.014113
#> 525 503.7365   3.058424
#> 526 500.0049   2.544180
#> 527 500.3420   2.530377
#> 528 524.6887   3.106659
#> 529 504.5635   2.529333
#> 530 523.3616   2.579977
#> 531 500.4888   2.502998
#> 532 500.0943   4.280882
#> 533 503.4324   2.502671
#> 534 500.0015   2.683904
#> 535 503.0455   3.848446
#> 536 500.2797   2.606296
#> 537 500.0738   2.933434
#> 538 500.5714   2.519041
#> 539 500.0344   2.942002
#> 540 500.2181   3.660511
#> 541 559.0907   3.685203
#> 542 501.0611   2.621483
#> 543 503.7590   3.563917
#> 544 506.7534   2.599769
#> 545 500.0302   2.586741
#> 546 502.0168   2.548679
#> 547 500.1746   2.681149
#> 548 538.2979   2.663542
#> 549 501.3471   4.012534
#> 550 502.7609   2.752077
#> 551 501.1757   3.059744
#> 552 501.3783   2.629219
#> 553 566.2829   3.475180
#> 554 500.0687   3.286313
#> 555 500.0848   3.100491
#> 556 540.7575   2.513239
#> 557 501.7644   2.930260
#> 558 528.9146   3.759917
#> 559 500.1871   2.783167
#> 560 665.6989   3.676397
#> 561 500.7615   2.553795
#> 562 501.4023   3.145077
#> 563 505.5156   3.073462
#> 564 507.2500   2.963874
#> 565 776.9231   2.726919
#> 566 501.3781   3.677090
#> 567 642.3671   3.440731
#> 568 500.2491   2.682176
#> 569 518.9457   3.782098
#> 570 500.1350   3.129065
#> 571 650.2866   3.045144
#> 572 500.5180   3.485397
#> 573 906.1386   2.760381
#> 574 501.3841   3.400607
#> 575 500.7773   2.920135
#> 576 501.6060   2.504360
#> 577 512.0912   2.544316
#> 578 503.3101   2.745467
#> 579 500.0938   3.057334
#> 580 500.0412   2.534340
#> 581 510.4309   3.339657
#> 582 500.5547   3.811197
#> 583 500.0964   3.708994
#> 584 500.1928   2.807200
#> 585 504.7545   2.609763
#> 586 500.1039   2.974238
#> 587 556.7624   3.154712
#> 588 501.0184   2.760818
#> 589 507.2936   2.550319
#> 590 883.5328   3.013458
#> 591 802.0053   3.142708
#> 592 500.2212   2.557466
#> 593 503.2342   2.879537
#> 594 500.2809   3.049874
#> 595 502.0200   2.556798
#> 596 506.4429   2.509604
#> 597 500.2470   2.931914
#> 598 500.3395   4.012424
#> 599 597.3235   2.547287
#> 600 512.3549   2.792215
#> 601 500.9189   3.119101
#> 602 500.0365   2.560755
#> 603 500.7202   4.571852
#> 604 550.4353   3.382226
#> 605 613.7590   2.601595
#> 606 500.4669   3.698840
#> 607 501.4535   2.838393
#> 608 612.0715   3.376918
#> 609 655.3036   2.748688
#> 610 500.4664   2.779788
#> 611 500.0165   2.785110
#> 612 500.0125   2.943453
#> 613 509.1615   2.555709
#> 614 501.1767   3.860483
#> 615 500.2136   2.798458
#> 616 823.3242   2.966284
#> 617 502.9640   2.638465
#> 618 500.0152   3.479158
#> 619 500.0688   2.688062
#> 620 610.7388   2.868644
#> 621 501.7318   2.648853
#> 622 500.0124   2.897859
#> 623 539.3736   3.195622
#> 624 500.0918   2.724907
#> 625 500.0512   2.927225
#> 626 508.1905   2.694677
#> 627 500.0872   3.027947
#> 628 500.1018   3.716529
#> 629 583.8582   2.520257
#> 630 504.5051   2.565250
#> 631 500.5108   2.733877
#> 632 532.5557   2.714936
#> 633 501.8359   2.747939
#> 634 931.4182   2.700237
#> 635 501.6075   3.096748
#> 636 579.9381   2.754347
#> 637 500.6911   2.788636
#> 638 528.6816   2.821029
#> 639 500.0399   2.613585
#> 640 501.2827   2.602591
#> 641 503.2717   3.506729
#> 642 596.1308   2.749590
#> 643 531.3118   2.973482
#> 644 515.7566   2.539658
#> 645 500.1676   3.401226
#> 646 500.1373   2.558884
#> 647 500.0019   3.092394
#> 648 529.8376   3.287831
#> 649 500.1245   2.678568
#> 650 500.2171   2.959003
#> 651 500.2316   2.587883
#> 652 500.2829   2.551997
#> 653 572.7987   3.259273
#> 654 500.2338   2.569926
#> 655 515.8346   2.536424
#> 656 602.8563   2.585897
#> 657 500.0604   2.903670
#> 658 506.0232   3.074128
#> 659 500.0140   2.655820
#> 660 500.4106   2.591458
#> 661 500.1296   3.365149
#> 662 500.2472   3.312883
#> 663 567.3727   2.742127
#> 664 503.6998   3.333127
#> 665 500.0482   2.980054
#> 666 530.9380   2.616799
#> 667 501.0876   2.718865
#> 668 530.2364   3.155751
#> 669 500.3778   2.723314
#> 670 503.8917   2.840968
#> 671 520.2524   2.506291
#> 672 501.6387   3.140282
#> 673 736.0325   2.873692
#> 674 506.5673   2.983976
#> 675 503.1729   2.634152
#> 676 500.0221   2.589039
#> 677 504.6011   2.610869
#> 678 500.0389   2.543183
#> 679 500.1156   2.688541
#> 680 503.8659   2.534351
#> 681 500.0192   3.127004
#> 682 562.9950   2.774551
#> 683 500.2888   3.104758
#> 684 533.7616   2.879585
#> 685 502.1357   2.917518
#> 686 529.9418   2.665866
#> 687 583.9725   3.547036
#> 688 611.8608   2.779262
#> 689 644.7689   2.854240
#> 690 508.6773   2.675383
#> 691 501.0753   2.993197
#> 692 500.5699   3.848593
#> 693 500.2481   2.678592
#> 694 500.0869   2.812222
#> 695 723.9309   3.189055
#> 696 503.7772   3.623157
#> 697 500.2436   3.274780
#> 698 502.9553   2.926759
#> 699 500.4418   2.563086
#> 700 506.2924   2.605155
#> 701 501.0407   2.799053
#> 702 500.2322   3.234876
#> 703 503.5155   2.502666
#> 704 500.9245   3.142456
#> 705 533.7754   2.617737
#> 706 507.7769   2.509777
#> 707 513.7162   2.973536
#> 708 508.2645   2.902927
#> 709 500.1084   2.587188
#> 710 677.2767   2.706905
#> 711 509.6885   2.612232
#> 712 774.1737   3.043685
#> 713 500.9180   2.697102
#> 714 500.0006   3.356873
#> 715 506.0727   2.811473
#> 716 501.3380   2.708866
#> 717 508.6573   2.585006
#> 718 500.1590   3.103977
#> 719 529.1279   2.686771
#> 720 500.1440   2.521418
#> 721 500.0132   2.561385
#> 722 500.0033   3.155222
#> 723 520.8238   2.517078
#> 724 518.6678   2.924216
#> 725 500.0005   2.723714
#> 726 899.9996   3.283917
#> 727 501.8544   2.685869
#> 728 500.7231   2.802546
#> 729 527.8105   2.593590
#> 730 683.1216   3.835831
#> 731 634.0061   2.564039
#> 732 552.8904   2.596883
#> 733 501.1528   3.359592
#> 734 500.0692   3.173365
#> 735 500.1573   3.695331
#> 736 517.3997   3.167914
#> 737 505.4388   2.603285
#> 738 500.0821   2.565384
#> 739 502.5748   3.139286
#> 740 500.4843   2.958031
#> 741 500.2422   2.570667
#> 742 526.0185   2.724569
#> 743 562.3205   2.633884
#> 744 537.0152   2.653863
#> 745 505.1705   3.069749
#> 746 500.2738   2.893216
#> 747 511.4717   3.444438
#> 748 500.3561   2.631724
#> 749 501.4090   2.581941
#> 750 500.0247   3.034164
#> 751 606.5525   2.751947
#> 752 750.0074   3.710610
#> 753 502.3032   2.551911
#> 754 501.4489   2.551249
#> 755 501.1921   2.789268
#> 756 500.0480   3.567637
#> 757 580.8782   2.735397
#> 758 503.0604   2.936493
#> 759 527.9928   3.095277
#> 760 522.1866   2.880557
#> 761 510.7037   2.711044
#> 762 502.5470   2.899384
#> 763 519.3954   3.154207
#> 764 500.4069   2.707337
#> 765 513.9206   2.667278
#> 766 500.0151   2.668443
#> 767 555.8430   3.173660
#> 768 509.9646   3.446033
#> 769 500.3104   2.600194
#> 770 500.1726   2.505101
#> 771 539.7504   2.819618
#> 772 500.0249   2.861929
#> 773 500.0990   3.880381
#> 774 509.0718   2.902040
#> 775 500.6074   2.668026
#> 776 500.0064   2.721211
#> 777 711.0241   2.774896
#> 778 513.7860   4.226827
#> 779 526.3135   4.782584
#> 780 513.6965   2.590145
#> 781 500.5506   2.527592
#> 782 500.4356   2.610328
#> 783 508.2793   3.336443
#> 784 500.3369   2.920256
#> 785 505.5538   2.804495
#> 786 501.4598   3.267751
#> 787 502.8225   3.051214
#> 788 504.1958   2.796318
#> 789 520.4978   3.289438
#> 790 505.3604   2.983716
#> 791 571.8665   3.008115
#> 792 502.1707   2.529509
#> 793 504.7307   2.539463
#> 794 580.3164   2.541388
#> 795 500.1020   3.138889
#> 796 510.1676   3.531525
#> 797 504.7302   2.561862
#> 798 504.9687   3.383316
#> 799 512.2004   2.968434
#> 800 517.2237   2.663686
#> 801 501.6535   4.986151
#> 802 568.3209   2.955461
#> 803 509.5339   2.754980
#> 804 501.3509   2.607603
#> 805 730.3498   3.205933
#> 806 506.3059   2.969172
#> 807 500.6028   2.534880
#> 808 512.5698   3.031277
#> 809 500.0817   2.695582
#> 810 549.4227   2.568756
#> 811 500.0308   2.511031
#> 812 588.9783   3.288928
#> 813 552.9446   3.135243
#> 814 650.6820   4.537243
#> 815 501.1637   2.537460
#> 816 500.4967   2.873500
#> 817 500.8100   2.655872
#> 818 572.1226   3.097525
#> 819 500.4707   2.794882
#> 820 500.7296   3.037802
#> 821 573.9647   4.120137
#> 822 500.0020   2.690350
#> 823 501.8083   2.771189
#> 824 627.0301   3.019266
#> 825 500.5264   3.574944
#> 826 536.1121   2.539839
#> 827 500.2320   3.382797
#> 828 504.4685   2.899507
#> 829 500.2971   2.631316
#> 830 531.0579   2.624980
#> 831 500.3778   3.219746
#> 832 501.3052   3.693502
#> 833 500.1086   2.609276
#> 834 602.2894   3.063102
#> 835 980.4866   3.450848
#> 836 502.1971   2.635313
#> 837 500.1218   3.056664
#> 838 509.3655   3.094109
#> 839 500.0728   2.567077
#> 840 500.1320   2.541121
#> 841 524.6731   2.752522
#> 842 599.5166   3.028552
#> 843 558.0934   2.750220
#> 844 623.9578   2.808681
#> 845 500.0500   3.519305
#> 846 508.4367   2.712108
#> 847 500.0084   4.989126
#> 848 500.5116   2.521905
#> 849 501.1624   3.036200
#> 850 500.0100   2.660185
#> 851 508.4807   3.508302
#> 852 502.6030   2.711405
#> 853 777.0160   2.903972
#> 854 502.5080   3.408150
#> 855 500.5941   2.529898
#> 856 694.4080   3.654366
#> 857 500.1032   2.572516
#> 858 521.3080   2.777297
#> 859 500.0258   2.686750
#> 860 502.2609   3.267591
#> 861 511.6940   2.595862
#> 862 500.7239   2.528410
#> 863 632.6069   3.807510
#> 864 504.1548   2.976873
#> 865 663.7112   2.552453
#> 866 508.9192   2.736892
#> 867 500.0730   3.279986
#> 868 501.0473   2.679459
#> 869 504.0591   2.626460
#> 870 510.6884   3.527181
#> 871 500.1620   2.529851
#> 872 582.5208   2.865481
#> 873 500.1698   2.662057
#> 874 500.5948   3.123325
#> 875 502.0405   3.151910
#> 876 500.1882   2.800239
#> 877 503.2367   3.089753
#> 878 501.8408   3.032616
#> 879 500.0335   2.596116
#> 880 501.0093   2.735498
#> 881 508.9585   2.886702
#> 882 500.7314   2.801083
#> 883 504.4081   2.779041
#> 884 500.5288   2.823209
#> 885 500.0162   2.634351
#> 886 500.2003   2.676653
#> 887 989.4304   2.842688
#> 888 747.5208   2.623019
#> 889 500.2161   2.577480
#> 890 501.9546   2.628454
#> 891 560.5758   2.723309
#> 892 500.9177   2.662173
#> 893 500.0630   2.793867
#> 894 670.1112   2.545689
#> 895 501.9035   4.381606
#> 896 501.3676   3.776732
#> 897 501.0206   3.382211
#> 898 500.0007   2.744994
#> 899 500.8685   4.158614
#> 900 513.4735   2.645749