Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = NULL)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

Deprecated argument for controlling parallelism. Use future::plan(future::multisession, workers = ncore) (or similar) to configure parallelism in your code instead.

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   500.0032   2.854616
#> 2   500.0655   2.780891
#> 3   713.8028   3.261780
#> 4   500.5522   2.623776
#> 5   501.1390   3.970416
#> 6   542.1219   2.850916
#> 7   500.1280   2.950413
#> 8   866.2634   2.504654
#> 9   500.0231   2.620150
#> 10  500.4643   3.257538
#> 11  532.0721   3.410215
#> 12  501.1492   2.736483
#> 13  550.6206   3.253832
#> 14  873.2340   2.881762
#> 15  500.3545   2.788701
#> 16  500.2155   3.629569
#> 17  500.8710   2.825316
#> 18  501.3197   4.472085
#> 19  787.7554   2.835625
#> 20  500.5086   2.540215
#> 21  500.0443   2.631805
#> 22  500.7057   2.505770
#> 23  508.5884   2.970961
#> 24  533.6891   2.795836
#> 25  553.1657   3.217741
#> 26  839.4900   3.340627
#> 27  500.0144   2.785300
#> 28  504.5322   3.206107
#> 29  500.3140   2.646065
#> 30  500.0308   3.967302
#> 31  500.8546   3.074422
#> 32  502.1909   2.785456
#> 33  503.8887   3.076674
#> 34  855.1528   2.654004
#> 35  501.4763   2.737228
#> 36  500.1098   3.038190
#> 37  502.7603   2.508212
#> 38  500.3024   2.535980
#> 39  582.8725   4.031959
#> 40  500.6329   2.624804
#> 41  500.3811   2.521872
#> 42  508.4905   2.678007
#> 43  568.8361   2.807217
#> 44  575.3960   2.545399
#> 45  568.5297   2.637165
#> 46  500.1665   3.316059
#> 47  504.0558   2.893835
#> 48  564.5399   2.763253
#> 49  502.0673   3.510807
#> 50  500.4579   3.024627
#> 51  500.5994   3.709169
#> 52  502.1751   3.182956
#> 53  517.5227   2.824998
#> 54  524.2643   2.883920
#> 55  500.4475   2.539232
#> 56  500.9387   2.538403
#> 57  742.8428   3.101129
#> 58  547.8998   3.090444
#> 59  870.3823   2.640910
#> 60  589.0198   3.636705
#> 61  500.0933   2.506779
#> 62  539.7389   3.211839
#> 63  500.9460   5.255884
#> 64  567.9297   2.505857
#> 65  507.8302   2.532196
#> 66  508.8259   2.554455
#> 67  503.2162   2.551838
#> 68  529.6975   2.533715
#> 69  500.0528   2.537830
#> 70  500.1183   2.801075
#> 71  886.8631   3.158906
#> 72  835.7394   3.400991
#> 73  501.4694   2.921424
#> 74  510.1971   2.717168
#> 75  502.5630   2.608587
#> 76  502.5964   2.729410
#> 77  508.3499   2.847450
#> 78  501.3239   2.803648
#> 79  542.1850   2.862290
#> 80  500.6167   3.861527
#> 81  544.6078   2.705999
#> 82  500.6747   2.851835
#> 83  500.0425   2.557596
#> 84  500.0817   2.683586
#> 85  502.0122   2.804447
#> 86  500.1303   2.517477
#> 87  504.2234   2.511644
#> 88  500.0070   2.881117
#> 89  522.9585   2.560843
#> 90  500.6196   2.840689
#> 91  503.3062   2.599428
#> 92  505.4624   3.483968
#> 93  500.0542   2.552335
#> 94  521.3262   2.788851
#> 95  510.9712   3.195284
#> 96  509.0260   2.851096
#> 97  501.5004   2.609149
#> 98  500.0686   2.863939
#> 99  500.2396   3.401764
#> 100 518.5011   2.821746
#> 101 751.0976   2.780488
#> 102 501.1338   2.775511
#> 103 513.3572   2.625854
#> 104 500.0728   2.526289
#> 105 500.0115   2.629557
#> 106 500.2932   3.056035
#> 107 530.0139   2.895211
#> 108 504.3143   2.979233
#> 109 510.2640   2.553038
#> 110 502.1753   2.935308
#> 111 610.3992   2.756230
#> 112 504.3626   2.998366
#> 113 531.0271   3.751637
#> 114 500.3029   3.267037
#> 115 500.7009   2.568039
#> 116 522.7515   2.681736
#> 117 552.1375   2.503798
#> 118 598.8892   2.666463
#> 119 507.7827   4.645108
#> 120 511.6021   2.771235
#> 121 501.6558   2.753408
#> 122 502.1336   2.959659
#> 123 528.3673   2.692991
#> 124 545.1249   2.523389
#> 125 503.4852   2.913642
#> 126 928.0368   2.609337
#> 127 506.1168   2.824679
#> 128 503.0776   2.964848
#> 129 500.8660   2.882263
#> 130 516.1435   2.541522
#> 131 500.5290   2.705078
#> 132 514.3679   2.709000
#> 133 501.1523   2.561178
#> 134 525.4111   2.719497
#> 135 585.1283   2.814874
#> 136 500.0773   2.586706
#> 137 518.7717   2.621077
#> 138 500.6757   2.723735
#> 139 500.0033   2.518941
#> 140 539.1291   2.537739
#> 141 503.6076   2.693861
#> 142 678.7806   2.731122
#> 143 500.0466   3.436384
#> 144 503.4793   2.631310
#> 145 510.3107   3.434426
#> 146 500.2444   3.199720
#> 147 507.3831   3.092660
#> 148 502.0313   2.547327
#> 149 501.1083   3.365604
#> 150 507.9164   2.710538
#> 151 500.3303   2.622885
#> 152 500.0163   2.841978
#> 153 515.4334   3.320814
#> 154 500.1260   2.902781
#> 155 500.8232   2.688602
#> 156 605.1385   2.580925
#> 157 500.0244   2.752767
#> 158 500.5522   3.400283
#> 159 521.0548   3.133270
#> 160 500.1857   3.089420
#> 161 507.6017   2.507152
#> 162 505.4870   3.046560
#> 163 524.7744   2.963047
#> 164 542.4416   2.879388
#> 165 500.3415   2.687982
#> 166 502.5181   2.682534
#> 167 500.0780   3.096202
#> 168 501.2471   2.623289
#> 169 574.7156   3.012360
#> 170 500.0083   2.723776
#> 171 500.9009   2.894095
#> 172 693.8177   3.151315
#> 173 501.6855   2.730307
#> 174 500.1250   4.031900
#> 175 501.1865   3.069321
#> 176 500.0599   3.389396
#> 177 500.0101   2.993986
#> 178 500.0573   3.096394
#> 179 506.3039   2.694912
#> 180 503.4941   3.079831
#> 181 551.2476   2.948259
#> 182 500.1751   3.081314
#> 183 501.8812   3.285649
#> 184 500.0040   2.521936
#> 185 639.9234   2.627090
#> 186 522.7442   2.871408
#> 187 807.7381   3.113380
#> 188 500.0143   2.724664
#> 189 500.0963   2.622568
#> 190 500.0003   2.969777
#> 191 523.4298   3.742565
#> 192 500.0004   2.938693
#> 193 500.3580   3.145363
#> 194 500.2085   3.316677
#> 195 501.1673   2.525906
#> 196 500.0809   2.643767
#> 197 500.7559   2.514858
#> 198 569.0511   3.314865
#> 199 541.2401   2.830343
#> 200 500.3676   2.545303
#> 201 532.0550   3.091448
#> 202 569.1797   2.918655
#> 203 590.3978   3.536539
#> 204 500.1957   2.712415
#> 205 501.3546   3.014729
#> 206 500.5008   2.651808
#> 207 547.3229   2.587205
#> 208 514.3307   2.724585
#> 209 774.3424   2.879256
#> 210 500.0525   2.644805
#> 211 501.0921   2.504739
#> 212 500.0587   3.060931
#> 213 539.5130   3.945342
#> 214 500.1221   2.746927
#> 215 687.2272   2.712561
#> 216 501.4394   2.580104
#> 217 500.2225   2.640882
#> 218 500.8665   3.040860
#> 219 506.8335   3.393350
#> 220 526.0242   2.741535
#> 221 500.0850   3.680920
#> 222 560.2019   3.008387
#> 223 821.5655   2.835977
#> 224 541.1469   2.791549
#> 225 515.6912   2.563234
#> 226 500.2249   3.270426
#> 227 500.0802   2.584837
#> 228 501.3057   2.571205
#> 229 500.7500   2.905633
#> 230 500.3187   2.592972
#> 231 500.0056   2.987409
#> 232 566.2306   2.697251
#> 233 511.6338   2.669155
#> 234 592.7353   2.970152
#> 235 502.3577   2.555644
#> 236 500.2092   2.814883
#> 237 506.8741   2.740397
#> 238 531.5993   2.808546
#> 239 501.5690   2.584216
#> 240 500.4671   2.846822
#> 241 500.1938   3.968872
#> 242 506.7649   3.517805
#> 243 518.4082   2.556841
#> 244 714.9027   3.580454
#> 245 504.4019   2.521357
#> 246 500.0621   3.389379
#> 247 504.7302   2.805492
#> 248 846.3918   2.723147
#> 249 502.5146   2.573402
#> 250 575.4934   2.891027
#> 251 500.3525   3.753055
#> 252 504.2268   3.307176
#> 253 500.9884   2.520054
#> 254 500.4787   2.634015
#> 255 517.5471   2.785688
#> 256 504.7092   2.572263
#> 257 500.1652   3.326087
#> 258 500.1444   2.946442
#> 259 562.6667   2.816536
#> 260 500.8489   3.967220
#> 261 503.4568   2.575424
#> 262 500.5408   2.712843
#> 263 535.7478   2.722871
#> 264 503.0440   2.505847
#> 265 501.6469   2.978926
#> 266 500.1914   2.549277
#> 267 500.2420   2.795765
#> 268 706.0738   2.975787
#> 269 500.2023   2.520266
#> 270 558.2244   3.270560
#> 271 502.3445   2.551288
#> 272 500.3966   4.525233
#> 273 502.3570   3.254618
#> 274 500.0570   2.958732
#> 275 500.8002   3.288178
#> 276 502.3600   3.237213
#> 277 513.4484   2.716582
#> 278 827.3388   2.514280
#> 279 514.9905   2.621376
#> 280 500.1004   2.775034
#> 281 504.8311   2.684060
#> 282 500.0851   3.316631
#> 283 530.3534   3.368967
#> 284 503.1264   2.822525
#> 285 501.6829   3.032493
#> 286 500.4978   2.753019
#> 287 500.1760   3.001477
#> 288 500.1089   3.083631
#> 289 500.4063   3.185870
#> 290 502.1292   3.501435
#> 291 500.2566   2.855497
#> 292 500.2782   2.615833
#> 293 506.2043   2.540743
#> 294 500.5786   2.532977
#> 295 835.7779   3.064024
#> 296 598.6749   2.636469
#> 297 500.1816   2.625643
#> 298 507.6586   2.630242
#> 299 500.8252   3.780653
#> 300 515.5917   3.801302
#> 301 500.8573   3.333991
#> 302 504.7317   3.494254
#> 303 525.4431   2.871331
#> 304 501.4974   3.320907
#> 305 500.1258   2.816408
#> 306 500.2580   2.607881
#> 307 517.7392   2.562927
#> 308 660.6362   2.575170
#> 309 500.2047   3.526625
#> 310 500.3326   2.505184
#> 311 500.0178   2.849576
#> 312 778.0157   2.544787
#> 313 507.2650   2.741524
#> 314 500.0245   2.574325
#> 315 500.1179   2.673075
#> 316 906.4233   2.982453
#> 317 506.4204   3.021488
#> 318 500.5300   2.560837
#> 319 544.2420   2.832285
#> 320 502.0393   2.991681
#> 321 500.1378   2.662548
#> 322 500.3980   3.049663
#> 323 547.0170   2.643111
#> 324 500.1473   3.285748
#> 325 505.9810   2.515426
#> 326 625.7603   2.535860
#> 327 545.1848   3.059854
#> 328 500.7428   2.545931
#> 329 500.1899   3.542073
#> 330 518.4442   3.182581
#> 331 500.2313   2.941554
#> 332 572.9843   2.758831
#> 333 672.9870   2.507547
#> 334 500.3483   3.723052
#> 335 500.0813   3.386005
#> 336 500.4971   2.759742
#> 337 505.1548   3.675592
#> 338 500.3093   2.560727
#> 339 500.1288   2.774823
#> 340 644.5258   2.956106
#> 341 500.0197   2.788384
#> 342 500.2406   3.308453
#> 343 506.8874   2.590092
#> 344 500.5382   2.570411
#> 345 588.2299   4.161593
#> 346 736.8495   2.650334
#> 347 502.6355   2.541604
#> 348 503.9450   2.985048
#> 349 500.0024   2.835761
#> 350 774.1793   4.111953
#> 351 533.1647   2.898557
#> 352 504.9935   2.715260
#> 353 510.2065   3.064376
#> 354 883.8320   2.876293
#> 355 500.1242   2.593084
#> 356 504.2244   2.728136
#> 357 500.5839   2.617154
#> 358 501.2064   2.537476
#> 359 520.9743   2.787368
#> 360 731.3676   2.793519
#> 361 500.2916   2.970921
#> 362 971.0284   2.635810
#> 363 513.4053   2.708095
#> 364 505.2553   2.737343
#> 365 904.2985   2.889601
#> 366 500.4550   2.523457
#> 367 507.9834   2.589059
#> 368 500.2312   2.935482
#> 369 700.1910   2.683677
#> 370 500.0885   3.273096
#> 371 786.7696   2.579414
#> 372 500.6296   2.882173
#> 373 527.7238   2.582970
#> 374 502.5844   3.111534
#> 375 500.3610   3.121150
#> 376 500.1038   3.404763
#> 377 502.5334   2.546980
#> 378 500.6668   2.779393
#> 379 508.7654   2.653602
#> 380 500.2560   2.555971
#> 381 500.1912   2.713771
#> 382 503.3527   2.926802
#> 383 524.6648   3.013091
#> 384 505.4511   4.144621
#> 385 522.7855   3.006071
#> 386 500.8417   2.558667
#> 387 531.1482   2.823112
#> 388 502.5038   2.958388
#> 389 500.0588   2.637816
#> 390 500.1674   2.752724
#> 391 500.0040   2.574439
#> 392 500.6861   3.443827
#> 393 500.7917   3.280453
#> 394 505.7334   2.641461
#> 395 500.0751   2.631777
#> 396 545.8965   3.118381
#> 397 500.7060   2.664610
#> 398 500.3191   2.522009
#> 399 501.2366   2.782657
#> 400 515.5600   2.668987
#> 401 618.1721   3.045645
#> 402 503.0257   2.648029
#> 403 526.2337   2.977131
#> 404 502.6628   4.146811
#> 405 501.5845   2.573459
#> 406 501.9862   2.562958
#> 407 602.3266   2.678375
#> 408 500.8756   2.629410
#> 409 503.9047   4.234980
#> 410 501.2380   2.581463
#> 411 520.3920   3.944090
#> 412 500.4574   4.040931
#> 413 781.8102   2.619839
#> 414 500.0963   2.787441
#> 415 503.4733   2.531377
#> 416 588.7487   2.603305
#> 417 525.5632   3.987369
#> 418 500.2115   2.649748
#> 419 501.5396   2.807262
#> 420 500.0364   2.634961
#> 421 511.4235   2.922698
#> 422 500.7583   2.801128
#> 423 868.0775   2.679985
#> 424 711.3771   2.736686
#> 425 526.4490   2.512064
#> 426 502.7284   3.876422
#> 427 649.4820   2.660135
#> 428 506.1078   2.504449
#> 429 500.3422   3.235509
#> 430 500.7444   2.788946
#> 431 500.0876   2.716372
#> 432 566.3811   2.557619
#> 433 500.1835   2.566723
#> 434 500.2916   4.312868
#> 435 506.5065   2.648122
#> 436 500.3761   2.502128
#> 437 502.8863   2.772280
#> 438 500.6096   2.904977
#> 439 500.6741   2.595527
#> 440 505.7552   2.636752
#> 441 500.5774   3.561520
#> 442 500.2155   2.510721
#> 443 504.4378   3.067478
#> 444 599.4094   2.520178
#> 445 502.2532   2.630268
#> 446 504.4861   2.588868
#> 447 500.0079   3.422162
#> 448 504.7077   2.722198
#> 449 500.0969   2.695744
#> 450 500.0109   2.560670
#> 451 500.0039   2.944633
#> 452 510.9847   2.572931
#> 453 501.0709   2.846642
#> 454 729.4269   2.733749
#> 455 500.0234   2.832548
#> 456 500.9715   2.798412
#> 457 506.0281   2.522133
#> 458 572.8562   2.844347
#> 459 723.9454   2.748720
#> 460 500.9135   2.539834
#> 461 500.0496   2.624625
#> 462 512.0231   4.042528
#> 463 500.2473   2.789424
#> 464 500.4589   2.520571
#> 465 502.6013   2.634161
#> 466 750.1773   3.052061
#> 467 504.1495   2.601230
#> 468 507.5543   2.931324
#> 469 506.3050   3.029320
#> 470 501.0762   3.369792
#> 471 511.3551   3.073782
#> 472 500.0241   3.149804
#> 473 500.2587   3.701575
#> 474 502.7998   2.908714
#> 475 600.8766   3.368129
#> 476 500.1475   2.797261
#> 477 525.1240   2.878675
#> 478 500.1271   2.544271
#> 479 649.2093   2.534993
#> 480 716.0344   3.748154
#> 481 500.0300   3.372554
#> 482 503.0773   2.538874
#> 483 501.9196   2.786577
#> 484 501.8664   2.898917
#> 485 500.2238   3.009619
#> 486 503.0977   2.768283
#> 487 500.3335   4.268717
#> 488 585.2559   3.566736
#> 489 601.2010   2.677338
#> 490 650.2500   2.651784
#> 491 500.0476   2.859774
#> 492 501.8089   3.054433
#> 493 751.1815   2.877828
#> 494 501.0314   3.076577
#> 495 500.0096   2.584713
#> 496 690.2768   2.891467
#> 497 510.5593   2.785120
#> 498 500.8462   2.871940
#> 499 530.2181   2.908518
#> 500 500.6660   3.210408
#> 501 996.7465   4.067119
#> 502 500.0222   2.963707
#> 503 500.4675   2.707036
#> 504 500.1625   2.603662
#> 505 503.8181   2.646990
#> 506 501.3357   2.859348
#> 507 500.8091   3.663410
#> 508 503.5260   2.876056
#> 509 500.1446   2.972119
#> 510 505.7593   3.034408
#> 511 628.5861   2.874796
#> 512 538.0633   3.521141
#> 513 500.2112   2.681502
#> 514 755.9096   2.907605
#> 515 506.8462   3.641870
#> 516 503.5766   2.797739
#> 517 753.8884   2.657400
#> 518 500.0065   3.176783
#> 519 500.0106   2.576056
#> 520 518.4545   2.676269
#> 521 503.9589   2.516162
#> 522 503.7785   3.415677
#> 523 501.4646   3.630600
#> 524 510.6930   2.902971
#> 525 623.2636   2.811544
#> 526 500.0529   3.000625
#> 527 976.7484   2.997646
#> 528 507.9394   3.039558
#> 529 500.8255   2.564227
#> 530 501.7190   2.646185
#> 531 716.8143   3.295284
#> 532 513.2840   2.757030
#> 533 572.3665   2.788170
#> 534 500.1018   2.655480
#> 535 501.6944   2.777410
#> 536 950.9570   3.407911
#> 537 601.1610   3.689464
#> 538 612.7866   2.627034
#> 539 520.5008   2.533506
#> 540 651.6988   2.944021
#> 541 500.1105   2.923706
#> 542 500.2902   2.770376
#> 543 500.0082   3.107920
#> 544 532.6221   2.637402
#> 545 500.2604   3.521356
#> 546 500.1159   3.143783
#> 547 523.5270   3.178059
#> 548 500.2675   2.518689
#> 549 506.2262   3.352898
#> 550 502.1750   3.490072
#> 551 501.3713   2.598800
#> 552 545.7519   3.568065
#> 553 512.3169   3.825235
#> 554 506.6720   2.596798
#> 555 500.9075   2.925320
#> 556 500.2355   2.961098
#> 557 506.6278   2.894179
#> 558 500.0412   2.806695
#> 559 500.9048   3.845740
#> 560 500.9044   2.694457
#> 561 505.8327   2.534051
#> 562 501.1962   2.680240
#> 563 704.6959   3.061817
#> 564 506.4741   2.520801
#> 565 500.6093   3.066036
#> 566 652.7297   2.655037
#> 567 600.2649   3.087987
#> 568 500.0536   3.062451
#> 569 509.3992   2.533834
#> 570 502.6996   2.699721
#> 571 660.4749   3.935060
#> 572 518.4177   2.820025
#> 573 500.3237   2.790047
#> 574 500.8445   3.106468
#> 575 508.0271   2.508937
#> 576 600.3687   2.675135
#> 577 508.7596   3.108235
#> 578 500.1698   3.626835
#> 579 717.9482   2.753670
#> 580 553.6009   3.051811
#> 581 504.0711   2.811102
#> 582 505.5888   2.767194
#> 583 500.1568   2.633738
#> 584 500.0719   3.917051
#> 585 577.0694   4.002186
#> 586 508.6204   2.700731
#> 587 500.1350   2.553956
#> 588 507.9960   2.553231
#> 589 502.2794   2.565343
#> 590 503.6889   2.509461
#> 591 511.1627   2.929874
#> 592 501.0000   2.568898
#> 593 501.0657   2.980586
#> 594 734.4215   3.316751
#> 595 500.6863   2.561413
#> 596 501.1407   2.824148
#> 597 559.4668   2.736350
#> 598 509.1547   2.520288
#> 599 500.1341   2.607661
#> 600 500.1167   2.814871
#> 601 525.4035   2.591765
#> 602 500.6402   4.280072
#> 603 502.6204   3.167679
#> 604 509.1047   3.082147
#> 605 626.0118   2.592645
#> 606 506.0022   2.530674
#> 607 500.6308   2.929611
#> 608 505.8541   3.218676
#> 609 510.7171   2.985856
#> 610 548.7607   3.375890
#> 611 500.1506   2.665792
#> 612 581.4071   3.008436
#> 613 500.5356   2.543176
#> 614 525.3829   4.140854
#> 615 500.2544   2.642114
#> 616 503.6210   2.560671
#> 617 506.9758   3.510445
#> 618 533.9881   2.984538
#> 619 500.0304   2.671789
#> 620 500.3801   2.701216
#> 621 507.0502   2.642615
#> 622 500.0018   2.716123
#> 623 500.6243   2.708173
#> 624 500.2035   3.459359
#> 625 618.7029   3.514609
#> 626 725.4935   2.660589
#> 627 534.0475   2.754634
#> 628 512.9562   2.538097
#> 629 992.9082   3.283782
#> 630 622.0021   3.301076
#> 631 500.9642   2.649195
#> 632 501.9976   2.521641
#> 633 513.8273   2.558872
#> 634 614.7213   3.030786
#> 635 500.1334   2.826492
#> 636 501.6852   3.089221
#> 637 659.5956   3.091833
#> 638 520.9635   2.784057
#> 639 548.8307   2.887638
#> 640 501.3838   2.510656
#> 641 513.0951   2.839148
#> 642 528.4752   3.428897
#> 643 508.4032   2.582323
#> 644 502.4472   2.936311
#> 645 500.0165   2.968369
#> 646 500.0338   2.724923
#> 647 805.6340   2.841571
#> 648 500.3195   4.050075
#> 649 530.4774   2.598628
#> 650 514.6941   2.732511
#> 651 934.0398   2.668700
#> 652 505.8425   3.262281
#> 653 504.1511   4.207413
#> 654 560.1548   2.866062
#> 655 500.1606   2.784904
#> 656 500.2151   3.027271
#> 657 581.8588   3.063009
#> 658 500.7235   2.752696
#> 659 526.8225   2.668467
#> 660 500.1815   2.549631
#> 661 500.2998   2.572842
#> 662 500.0626   3.149113
#> 663 568.0528   3.301383
#> 664 604.4616   3.035733
#> 665 548.8987   2.599669
#> 666 500.3209   3.703771
#> 667 500.3077   2.576963
#> 668 500.0627   2.750840
#> 669 501.3508   2.669431
#> 670 500.2140   2.724355
#> 671 511.4711   2.679164
#> 672 675.4002   2.617028
#> 673 568.0370   2.534115
#> 674 613.9328   3.316708
#> 675 500.4507   3.389093
#> 676 513.8117   3.081739
#> 677 517.9335   2.697290
#> 678 510.6642   2.508669
#> 679 775.1044   2.860631
#> 680 503.0883   2.544779
#> 681 501.7954   3.869064
#> 682 505.0770   3.690492
#> 683 502.3693   2.624973
#> 684 500.7322   2.952325
#> 685 512.9551   2.611285
#> 686 810.1535   3.419570
#> 687 519.0342   2.931639
#> 688 500.2767   3.324537
#> 689 500.2639   2.671271
#> 690 502.2894   3.632033
#> 691 515.7089   2.848431
#> 692 500.0845   3.329881
#> 693 506.1709   2.856023
#> 694 534.0601   2.560410
#> 695 574.1295   2.559997
#> 696 501.0102   2.768814
#> 697 502.0373   2.725628
#> 698 609.1318   2.541084
#> 699 502.2182   3.085702
#> 700 500.1798   2.963157
#> 701 615.6003   2.731082
#> 702 500.6219   2.742752
#> 703 714.6782   3.078958
#> 704 503.6265   2.771870
#> 705 507.1408   2.588602
#> 706 500.7895   3.229644
#> 707 536.5297   3.367354
#> 708 504.6259   2.913819
#> 709 500.7691   2.697419
#> 710 530.9963   2.939535
#> 711 500.0400   2.887545
#> 712 502.6196   2.506818
#> 713 701.9562   3.507099
#> 714 500.0497   2.637648
#> 715 501.8200   2.928218
#> 716 503.4178   2.553831
#> 717 514.5645   2.586479
#> 718 586.9798   4.038188
#> 719 501.8240   3.145555
#> 720 511.2383   3.745246
#> 721 757.3789   2.799526
#> 722 500.7862   2.863700
#> 723 500.0300   3.079229
#> 724 500.0615   2.534690
#> 725 500.6209   2.755129
#> 726 799.6679   2.519130
#> 727 500.1771   3.211573
#> 728 583.9223   4.412476
#> 729 868.5574   2.840100
#> 730 540.1110   2.847393
#> 731 962.7131   3.003989
#> 732 500.0078   3.320241
#> 733 500.0302   2.520846
#> 734 917.3508   2.933075
#> 735 515.6119   2.836764
#> 736 989.0182   3.360170
#> 737 500.0002   2.569358
#> 738 500.3080   2.729593
#> 739 500.1889   2.566850
#> 740 505.0714   3.717516
#> 741 505.5342   3.349465
#> 742 500.5668   2.813429
#> 743 697.6649   3.202997
#> 744 503.4639   3.690863
#> 745 771.8729   2.796435
#> 746 500.0060   2.585312
#> 747 551.3229   3.021414
#> 748 500.0673   3.794868
#> 749 578.5401   2.512098
#> 750 500.8097   3.082262
#> 751 500.6262   3.088649
#> 752 500.0327   3.777717
#> 753 533.0367   2.511328
#> 754 505.0205   2.566340
#> 755 531.6265   2.877161
#> 756 504.1328   3.527190
#> 757 505.4638   2.548777
#> 758 500.0172   3.114307
#> 759 507.5942   2.757571
#> 760 501.1456   3.206730
#> 761 895.0670   2.650912
#> 762 794.1680   2.932289
#> 763 500.5064   2.975327
#> 764 510.3667   2.871404
#> 765 521.4598   2.606399
#> 766 500.3359   2.807704
#> 767 518.7028   2.930600
#> 768 706.8875   2.790983
#> 769 501.7676   2.872467
#> 770 630.0094   3.135365
#> 771 500.0347   2.564334
#> 772 500.1112   3.129276
#> 773 500.9432   2.811783
#> 774 500.4939   2.573933
#> 775 507.9006   2.514985
#> 776 500.0615   3.080214
#> 777 500.2490   2.838000
#> 778 500.8469   2.842268
#> 779 502.6342   3.170063
#> 780 504.6794   2.547118
#> 781 500.0511   2.639230
#> 782 500.4415   4.127833
#> 783 503.3205   2.505293
#> 784 515.1511   2.740565
#> 785 511.4692   2.517850
#> 786 518.2649   2.626042
#> 787 511.8871   2.735897
#> 788 594.9205   2.655240
#> 789 500.0061   2.664033
#> 790 500.3927   3.991874
#> 791 929.4452   3.275436
#> 792 500.0218   3.109405
#> 793 522.3055   2.530137
#> 794 506.7493   2.531952
#> 795 500.0600   2.979972
#> 796 500.0650   2.556088
#> 797 533.9373   2.643910
#> 798 500.4323   2.506870
#> 799 579.1744   3.150980
#> 800 500.0700   2.898745
#> 801 505.6726   3.399938
#> 802 597.8628   2.727484
#> 803 610.9439   3.234368
#> 804 500.2917   2.999056
#> 805 923.9522   2.630346
#> 806 507.9634   3.298027
#> 807 752.9236   3.064728
#> 808 508.4706   3.147998
#> 809 686.0899   2.802141
#> 810 500.0917   2.779540
#> 811 576.1309   2.892352
#> 812 993.8163   3.404703
#> 813 500.1418   3.979286
#> 814 500.0592   3.125898
#> 815 504.1548   2.982973
#> 816 500.0808   2.691097
#> 817 500.2321   2.551959
#> 818 515.2581   2.797350
#> 819 500.0526   2.754944
#> 820 500.3589   2.919185
#> 821 502.0887   2.738639
#> 822 695.1368   4.355848
#> 823 930.5577   2.995662
#> 824 553.8196   2.889074
#> 825 500.1323   2.779311
#> 826 607.4561   2.650818
#> 827 500.1475   2.980337
#> 828 500.1465   2.505707
#> 829 500.1650   2.690147
#> 830 513.5959   3.117430
#> 831 516.9845   2.881953
#> 832 500.0806   2.711062
#> 833 500.5436   3.318960
#> 834 732.1927   2.880644
#> 835 500.3272   2.745749
#> 836 523.0116   2.585858
#> 837 501.1593   3.320282
#> 838 500.0381   2.701062
#> 839 500.0622   4.004593
#> 840 501.0328   2.628190
#> 841 974.3578   2.673795
#> 842 853.3719   2.903474
#> 843 505.8506   3.001119
#> 844 507.5956   3.029950
#> 845 554.6184   2.654751
#> 846 500.0577   2.656851
#> 847 625.0030   2.565871
#> 848 500.2060   3.291920
#> 849 537.2368   2.561590
#> 850 501.7311   2.671216
#> 851 500.0829   2.774002
#> 852 718.6848   2.591608
#> 853 500.7050   3.590896
#> 854 500.3628   2.736408
#> 855 500.6927   2.798212
#> 856 500.9769   2.613391
#> 857 555.3276   5.502145
#> 858 949.1582   2.514399
#> 859 500.5937   2.726254
#> 860 863.2190   2.842812
#> 861 504.3886   3.247477
#> 862 503.9278   2.710497
#> 863 773.2358   2.572321
#> 864 521.2877   3.745299
#> 865 534.2966   2.837219
#> 866 500.1008   2.504077
#> 867 500.0008   2.645580
#> 868 500.5236   2.815893
#> 869 535.1523   3.092800
#> 870 500.2529   2.734118
#> 871 511.2485   2.647994
#> 872 644.6988   2.966438
#> 873 500.9951   3.214449
#> 874 501.9475   3.288887
#> 875 507.1174   2.536887
#> 876 538.3617   3.180600
#> 877 501.3575   3.291352
#> 878 510.7562   3.102778
#> 879 500.0204   2.811362
#> 880 500.1817   2.572271
#> 881 582.5702   3.298225
#> 882 709.8207   2.716923
#> 883 500.0456   3.092432
#> 884 607.2586   3.953882
#> 885 500.2078   2.606570
#> 886 502.2577   2.520314
#> 887 500.0416   2.812530
#> 888 516.6734   4.226668
#> 889 500.0993   2.623263
#> 890 808.4391   2.919688
#> 891 505.8671   2.816136
#> 892 501.8809   2.676515
#> 893 845.7195   2.973565
#> 894 500.3507   2.712030
#> 895 500.0888   2.653761
#> 896 505.2021   2.596615
#> 897 640.7561   2.602183
#> 898 513.4276   3.312654
#> 899 573.3992   2.618855
#> 900 522.7658   2.895593