Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   852.3964   3.601945
#> 2   513.9143   3.720689
#> 3   579.7055   2.617637
#> 4   501.4170   2.781826
#> 5   501.7428   3.438569
#> 6   722.3113   4.147920
#> 7   500.0882   3.038822
#> 8   503.2041   2.534994
#> 9   503.1304   2.655637
#> 10  501.5548   3.303641
#> 11  552.2386   4.294261
#> 12  500.0701   3.438943
#> 13  508.1087   2.925990
#> 14  500.1638   3.068619
#> 15  500.6066   2.542978
#> 16  510.4213   2.550369
#> 17  507.8759   2.877821
#> 18  501.0964   2.965930
#> 19  504.2322   2.634304
#> 20  544.9808   2.807937
#> 21  500.6337   3.180194
#> 22  500.4596   2.751305
#> 23  545.0030   3.011049
#> 24  501.6034   3.223168
#> 25  500.5932   3.151597
#> 26  500.3744   2.822924
#> 27  503.0862   3.236178
#> 28  500.0886   2.644051
#> 29  504.0339   4.612148
#> 30  500.0938   2.512155
#> 31  596.9565   2.703973
#> 32  506.2129   2.890519
#> 33  780.5791   2.619409
#> 34  500.1697   2.645805
#> 35  507.8599   2.979382
#> 36  557.5480   2.524972
#> 37  500.0914   3.439413
#> 38  500.8799   2.726064
#> 39  518.1909   2.742907
#> 40  510.1144   3.158150
#> 41  531.2186   3.077146
#> 42  500.1867   2.837447
#> 43  500.2098   2.545438
#> 44  500.0685   3.875692
#> 45  883.8174   2.765977
#> 46  627.1505   2.758054
#> 47  500.4749   3.500865
#> 48  500.4508   2.555359
#> 49  577.7258   2.592707
#> 50  736.8489   2.803516
#> 51  500.2248   3.406187
#> 52  500.1133   2.954535
#> 53  501.0961   2.978069
#> 54  500.3433   3.711525
#> 55  500.0184   2.910267
#> 56  565.5578   2.531984
#> 57  603.3783   2.845247
#> 58  516.7489   2.856207
#> 59  624.5083   3.061054
#> 60  500.1065   2.520127
#> 61  540.9657   2.567984
#> 62  532.0886   3.921229
#> 63  563.3064   4.571778
#> 64  501.7219   3.353099
#> 65  500.1239   2.584468
#> 66  511.7526   2.540876
#> 67  501.8654   2.865320
#> 68  522.5969   2.878795
#> 69  500.2031   3.474055
#> 70  503.6869   2.607290
#> 71  500.5364   2.850532
#> 72  504.8879   2.544774
#> 73  857.5538   2.586561
#> 74  500.2383   2.664546
#> 75  500.3611   2.653331
#> 76  500.0954   2.508420
#> 77  501.0217   2.829291
#> 78  500.5972   2.945778
#> 79  505.3562   2.865677
#> 80  500.0191   2.805001
#> 81  504.7179   3.181513
#> 82  501.2854   2.624741
#> 83  547.1124   3.789774
#> 84  500.1273   2.743985
#> 85  509.3558   3.243940
#> 86  500.0981   2.552576
#> 87  611.5110   3.170896
#> 88  502.3182   2.713668
#> 89  504.1554   2.731684
#> 90  500.1920   2.666213
#> 91  500.9190   2.619774
#> 92  520.3340   3.047352
#> 93  503.0208   3.920251
#> 94  506.5734   3.137088
#> 95  502.5672   3.159450
#> 96  501.5058   3.087729
#> 97  501.1199   4.077170
#> 98  608.3546   2.739470
#> 99  585.9152   2.509697
#> 100 503.4886   2.763043
#> 101 527.4813   3.200337
#> 102 530.9801   2.969858
#> 103 502.8556   2.643742
#> 104 503.7353   3.130339
#> 105 790.0144   3.083687
#> 106 500.2774   2.634344
#> 107 507.6947   2.598879
#> 108 500.0520   3.107239
#> 109 519.5624   4.007118
#> 110 504.6243   2.651778
#> 111 500.1665   3.669579
#> 112 504.1404   2.692630
#> 113 618.9051   2.508861
#> 114 500.1337   3.134731
#> 115 509.7870   4.378730
#> 116 503.5001   2.523608
#> 117 589.0522   2.546110
#> 118 500.2499   2.658001
#> 119 500.0486   2.515626
#> 120 500.2004   4.630935
#> 121 503.1503   2.511601
#> 122 513.8929   3.845389
#> 123 500.6839   2.656639
#> 124 562.1092   3.121878
#> 125 758.6017   2.964225
#> 126 505.4913   2.681541
#> 127 500.6592   2.814165
#> 128 500.1929   2.595577
#> 129 504.6933   2.632958
#> 130 519.8364   3.610487
#> 131 568.8135   2.599808
#> 132 500.1154   3.380897
#> 133 522.1420   2.533968
#> 134 500.0283   3.032198
#> 135 564.6250   3.137861
#> 136 796.3238   5.592323
#> 137 681.5781   2.853221
#> 138 500.0784   2.723104
#> 139 500.2897   2.627341
#> 140 790.0575   2.534627
#> 141 501.1422   2.799382
#> 142 500.0595   2.566002
#> 143 847.6287   2.722119
#> 144 547.2594   2.539551
#> 145 502.5855   3.036407
#> 146 524.4020   2.788628
#> 147 500.0505   2.550964
#> 148 500.4617   2.867772
#> 149 579.0092   2.691851
#> 150 504.9144   2.627473
#> 151 501.7668   2.611309
#> 152 509.1611   3.786206
#> 153 507.2629   2.564001
#> 154 965.8038   2.582069
#> 155 500.5372   2.666691
#> 156 501.4541   2.645641
#> 157 502.3750   4.241924
#> 158 617.5922   5.443784
#> 159 500.8378   2.994769
#> 160 501.3638   2.817527
#> 161 603.5071   2.540514
#> 162 522.8924   2.735568
#> 163 550.0427   2.693821
#> 164 500.8678   3.004313
#> 165 508.6419   2.767516
#> 166 516.1623   2.637058
#> 167 502.2932   2.910907
#> 168 672.8908   2.609937
#> 169 500.6859   2.589789
#> 170 503.1115   2.675069
#> 171 564.6786   3.181554
#> 172 500.0152   3.220520
#> 173 537.8585   2.577617
#> 174 509.2148   3.013790
#> 175 508.5258   3.716578
#> 176 502.9547   2.643667
#> 177 500.2357   3.000169
#> 178 500.4329   3.155543
#> 179 539.9421   2.630361
#> 180 500.7269   2.707624
#> 181 507.0212   2.972679
#> 182 502.5019   2.576934
#> 183 518.8804   2.569010
#> 184 500.0180   2.770853
#> 185 520.1231   2.552092
#> 186 556.6166   2.821541
#> 187 500.0662   2.583032
#> 188 519.0562   3.359321
#> 189 500.0359   3.069081
#> 190 500.6543   2.500891
#> 191 500.0034   3.378393
#> 192 620.0472   2.771585
#> 193 500.0418   2.870645
#> 194 516.6192   2.925069
#> 195 500.6026   4.250078
#> 196 532.5941   3.473367
#> 197 594.8212   2.593340
#> 198 500.0739   2.544140
#> 199 500.5887   3.116184
#> 200 507.2535   2.679108
#> 201 531.8186   2.590683
#> 202 506.4619   2.573770
#> 203 501.2525   5.521910
#> 204 506.3944   2.853665
#> 205 516.5638   3.388937
#> 206 635.8033   2.576884
#> 207 635.5636   2.737834
#> 208 500.2983   2.596521
#> 209 914.9315   3.354924
#> 210 546.7276   2.571628
#> 211 500.5032   2.568762
#> 212 505.6921   2.663684
#> 213 500.6427   2.709563
#> 214 500.1313   2.562872
#> 215 724.7985   3.392747
#> 216 715.4063   2.537864
#> 217 500.9682   2.615057
#> 218 500.1111   2.812776
#> 219 514.9568   2.582786
#> 220 502.0846   3.743520
#> 221 501.8041   3.759037
#> 222 500.1044   3.752517
#> 223 500.8140   2.617927
#> 224 500.1260   4.582288
#> 225 500.0023   3.851454
#> 226 888.8460   3.110322
#> 227 501.0629   2.539123
#> 228 552.0412   3.272656
#> 229 577.4928   2.548459
#> 230 635.1520   2.792031
#> 231 857.6255   2.601561
#> 232 526.8843   2.553200
#> 233 762.7462   3.153124
#> 234 500.0280   3.413721
#> 235 556.4893   2.870712
#> 236 500.0221   3.423831
#> 237 500.1107   2.641107
#> 238 510.3278   3.004586
#> 239 500.7064   3.001398
#> 240 500.0417   3.810862
#> 241 500.3462   2.817322
#> 242 500.8661   3.057241
#> 243 501.4879   2.518270
#> 244 534.1054   2.691393
#> 245 505.1321   3.165868
#> 246 512.5051   2.949605
#> 247 500.2222   3.845392
#> 248 529.6726   5.023994
#> 249 506.9085   2.622700
#> 250 500.1367   2.791842
#> 251 500.1003   3.244930
#> 252 548.6824   2.761600
#> 253 505.5904   2.926065
#> 254 500.2423   2.624472
#> 255 501.9502   3.028722
#> 256 595.0278   3.137426
#> 257 501.2489   2.548391
#> 258 503.0724   2.714590
#> 259 503.1002   2.785698
#> 260 594.8360   2.627904
#> 261 500.0938   2.580396
#> 262 500.2675   3.015479
#> 263 500.1039   4.832279
#> 264 580.8629   3.410076
#> 265 500.7290   2.731111
#> 266 500.0770   3.154580
#> 267 500.5701   2.534018
#> 268 542.8341   2.845508
#> 269 674.4051   2.527284
#> 270 503.3859   2.740144
#> 271 715.8296   3.229459
#> 272 500.0193   2.619462
#> 273 502.3305   3.344955
#> 274 501.9100   2.578360
#> 275 500.0705   2.524982
#> 276 717.3780   2.569832
#> 277 509.9104   2.722711
#> 278 500.0150   2.788503
#> 279 500.1233   2.640903
#> 280 501.1404   2.686881
#> 281 505.5178   4.292003
#> 282 504.3181   2.528038
#> 283 500.0587   2.558093
#> 284 562.0584   2.857366
#> 285 500.4604   2.862530
#> 286 520.5568   3.798268
#> 287 508.8872   2.942108
#> 288 500.9945   3.681737
#> 289 500.3761   2.653334
#> 290 502.1282   2.528058
#> 291 502.8809   2.561786
#> 292 500.2067   2.817574
#> 293 942.0947   2.751662
#> 294 500.5430   2.951543
#> 295 885.0090   2.604737
#> 296 500.2429   3.451210
#> 297 500.8758   3.099591
#> 298 500.0057   3.384025
#> 299 500.1461   4.006951
#> 300 503.0081   3.752542
#> 301 500.0125   3.076939
#> 302 500.0894   3.235705
#> 303 505.7307   2.674392
#> 304 501.7219   2.700493
#> 305 536.8272   3.489242
#> 306 500.9198   2.527879
#> 307 563.4002   2.781582
#> 308 507.7427   2.643405
#> 309 534.1670   3.201635
#> 310 547.9479   3.358695
#> 311 505.8318   3.253468
#> 312 502.1784   3.245147
#> 313 510.0069   2.724619
#> 314 640.8585   2.603077
#> 315 500.0325   2.541793
#> 316 572.2964   2.768640
#> 317 508.2093   2.570907
#> 318 500.1156   2.703072
#> 319 535.7867   2.514940
#> 320 505.9742   3.040495
#> 321 519.9318   2.837263
#> 322 500.0636   2.673491
#> 323 512.8238   2.740404
#> 324 511.9561   2.830717
#> 325 500.1940   2.973940
#> 326 500.3039   2.581674
#> 327 500.3075   2.674186
#> 328 500.0957   2.949862
#> 329 502.0146   3.843391
#> 330 508.1288   3.371215
#> 331 508.6013   2.645918
#> 332 501.0768   3.412840
#> 333 507.0625   3.496122
#> 334 539.4694   2.749434
#> 335 500.5016   2.806980
#> 336 500.5856   2.973359
#> 337 501.4148   3.869857
#> 338 503.4610   2.682801
#> 339 508.7785   3.313318
#> 340 503.8535   2.534562
#> 341 505.8032   4.006100
#> 342 519.1531   2.990643
#> 343 500.1517   2.822576
#> 344 500.8172   2.931268
#> 345 581.8971   2.529218
#> 346 533.4640   2.878130
#> 347 500.3798   2.835577
#> 348 502.8835   2.708151
#> 349 500.0967   2.554291
#> 350 648.0626   2.807228
#> 351 506.5806   3.038270
#> 352 500.0217   3.079826
#> 353 533.8172   2.568936
#> 354 500.0087   2.567902
#> 355 583.5672   2.744199
#> 356 502.2929   2.530637
#> 357 502.5264   2.572180
#> 358 704.8935   2.591019
#> 359 500.0221   2.838807
#> 360 531.5662   2.817294
#> 361 500.2546   3.881167
#> 362 500.0572   2.832559
#> 363 500.3312   2.723505
#> 364 500.2412   3.130125
#> 365 517.6869   2.569612
#> 366 563.8116   2.905973
#> 367 542.6092   3.012512
#> 368 500.0293   2.565814
#> 369 500.0485   2.751979
#> 370 501.5425   2.563498
#> 371 532.1014   2.757148
#> 372 536.9087   2.652409
#> 373 885.8798   2.595138
#> 374 503.2585   4.715031
#> 375 554.3749   2.615529
#> 376 501.7195   2.550535
#> 377 502.0338   2.692909
#> 378 515.3174   3.456674
#> 379 500.0696   2.791854
#> 380 515.2734   2.777349
#> 381 500.2892   2.710299
#> 382 500.7219   2.562930
#> 383 743.7139   3.109727
#> 384 502.9547   2.520426
#> 385 500.1045   2.714816
#> 386 513.7976   2.809589
#> 387 521.7378   3.677828
#> 388 737.0332   3.502452
#> 389 514.6294   2.649936
#> 390 649.0694   2.785621
#> 391 500.1353   4.343726
#> 392 516.9589   3.203151
#> 393 506.8119   2.866988
#> 394 511.8788   2.807073
#> 395 503.3214   2.888351
#> 396 793.9884   2.740038
#> 397 561.9222   2.586016
#> 398 541.3136   3.160390
#> 399 500.1228   2.657836
#> 400 501.9987   3.703049
#> 401 500.8731   2.899075
#> 402 500.4988   3.552108
#> 403 500.8024   3.466089
#> 404 500.0608   2.719633
#> 405 506.2113   3.177712
#> 406 513.9202   3.004349
#> 407 521.8265   2.876260
#> 408 834.9981   2.973484
#> 409 527.3213   3.732860
#> 410 620.9150   3.030378
#> 411 654.7785   2.888775
#> 412 510.1210   4.204997
#> 413 500.9782   3.190805
#> 414 500.8437   2.537458
#> 415 515.5821   2.864915
#> 416 500.5071   3.194851
#> 417 524.9138   2.844357
#> 418 501.5177   2.556118
#> 419 500.2220   3.594365
#> 420 500.0477   4.471878
#> 421 527.8556   2.937098
#> 422 500.1808   2.655979
#> 423 502.2597   2.529490
#> 424 510.2780   2.719143
#> 425 504.3973   2.888049
#> 426 511.8121   2.801963
#> 427 500.6340   2.976850
#> 428 515.9558   3.356030
#> 429 812.1742   2.647447
#> 430 500.0721   3.001630
#> 431 501.8356   2.568639
#> 432 502.4781   3.972903
#> 433 592.8831   2.587049
#> 434 505.5218   3.126057
#> 435 500.3205   3.066286
#> 436 566.1914   2.550342
#> 437 500.9717   2.589564
#> 438 500.0162   2.519247
#> 439 501.5209   3.639525
#> 440 506.3752   3.749410
#> 441 502.0603   2.774169
#> 442 999.9373   2.836465
#> 443 509.9198   2.639682
#> 444 521.4917   2.708192
#> 445 500.6029   2.567299
#> 446 502.8555   3.505192
#> 447 500.7339   3.465537
#> 448 512.9844   2.825930
#> 449 544.4070   3.152364
#> 450 643.8994   3.738118
#> 451 501.5838   2.923706
#> 452 568.1559   2.729796
#> 453 519.3453   2.677423
#> 454 506.2970   2.590391
#> 455 596.3427   2.676734
#> 456 500.6691   2.664193
#> 457 502.4696   3.115634
#> 458 511.1267   2.829051
#> 459 743.1985   2.826005
#> 460 893.0336   2.838799
#> 461 500.6048   2.702514
#> 462 500.2285   2.793071
#> 463 577.9418   2.763467
#> 464 524.0218   2.779337
#> 465 500.1050   2.792481
#> 466 504.2173   2.697633
#> 467 649.1741   3.051391
#> 468 505.8932   4.585541
#> 469 500.0110   2.814184
#> 470 515.2410   2.628557
#> 471 500.2562   2.638345
#> 472 510.7171   3.197228
#> 473 526.1855   3.110923
#> 474 563.8972   2.663295
#> 475 504.9606   2.676934
#> 476 544.6875   2.773099
#> 477 521.1676   3.299871
#> 478 530.9146   2.869229
#> 479 500.1880   2.621134
#> 480 506.8840   3.838800
#> 481 561.9444   3.193571
#> 482 570.9614   2.504532
#> 483 501.1288   2.508346
#> 484 500.3272   2.512693
#> 485 763.3934   2.591084
#> 486 500.3918   2.890083
#> 487 605.1869   2.722807
#> 488 512.7707   2.739290
#> 489 500.2497   2.697503
#> 490 500.7865   2.597757
#> 491 527.1053   2.646384
#> 492 505.2985   3.111716
#> 493 823.6797   2.729089
#> 494 500.4542   2.772166
#> 495 503.2481   2.825078
#> 496 529.6615   2.569750
#> 497 500.0481   2.531179
#> 498 965.1996   3.099854
#> 499 504.0454   3.120464
#> 500 541.0848   2.964359
#> 501 500.0614   2.563491
#> 502 511.3662   3.030733
#> 503 530.9490   2.510867
#> 504 509.0014   2.523090
#> 505 566.7655   2.997245
#> 506 507.0676   2.546672
#> 507 568.0774   2.694881
#> 508 603.1628   2.524714
#> 509 660.8475   2.527580
#> 510 500.3678   3.335998
#> 511 500.2305   2.673582
#> 512 500.2190   2.831880
#> 513 508.8352   2.605052
#> 514 513.8883   4.355526
#> 515 500.2089   2.645947
#> 516 502.3727   2.758388
#> 517 520.8808   2.895138
#> 518 573.2044   2.545912
#> 519 500.0880   2.597346
#> 520 500.9781   2.917565
#> 521 505.3360   2.596854
#> 522 500.3547   2.589966
#> 523 501.0302   2.968847
#> 524 500.1633   2.692311
#> 525 500.3943   2.879763
#> 526 500.0646   2.735169
#> 527 500.2866   3.039276
#> 528 500.1703   2.585253
#> 529 500.1984   4.300918
#> 530 501.9972   2.605471
#> 531 587.8758   2.571457
#> 532 500.0863   3.062774
#> 533 589.8324   2.500275
#> 534 500.6836   2.598988
#> 535 500.6705   2.908720
#> 536 521.9271   3.482943
#> 537 500.2815   2.530387
#> 538 839.3616   2.985038
#> 539 500.7176   2.506175
#> 540 501.2102   2.745769
#> 541 501.0656   2.781744
#> 542 810.5362   2.706082
#> 543 500.6000   2.936106
#> 544 506.9086   2.509758
#> 545 508.9144   2.534559
#> 546 503.0213   3.317566
#> 547 520.3094   2.571642
#> 548 502.4634   5.246131
#> 549 500.2125   2.889263
#> 550 518.6242   2.714398
#> 551 509.7986   2.589430
#> 552 500.0428   2.788582
#> 553 500.0502   2.823071
#> 554 501.0540   2.815606
#> 555 500.1261   3.074078
#> 556 662.1775   3.016338
#> 557 500.7179   4.050636
#> 558 500.0876   3.082490
#> 559 507.0320   2.787154
#> 560 500.8312   3.201233
#> 561 501.1565   2.611550
#> 562 502.7547   3.413740
#> 563 503.6562   3.226229
#> 564 500.3468   2.777479
#> 565 517.9272   3.715319
#> 566 787.7690   2.875521
#> 567 526.9486   2.926165
#> 568 622.3258   2.501322
#> 569 500.3373   2.907354
#> 570 594.8633   3.462355
#> 571 504.0320   4.149499
#> 572 682.7059   2.588061
#> 573 500.0828   3.942563
#> 574 598.6348   3.294329
#> 575 500.8042   2.677227
#> 576 526.6558   2.651262
#> 577 500.1551   2.939056
#> 578 566.3563   2.819293
#> 579 500.0026   2.558204
#> 580 502.7828   2.950432
#> 581 500.9466   2.970460
#> 582 667.0537   2.995274
#> 583 562.2493   2.592886
#> 584 518.3011   4.700250
#> 585 500.0485   2.608102
#> 586 508.6940   2.619424
#> 587 500.2048   2.592485
#> 588 598.1865   2.629032
#> 589 500.6222   2.524657
#> 590 500.0153   2.635939
#> 591 536.6034   2.935924
#> 592 500.2919   2.745737
#> 593 626.7564   3.745026
#> 594 511.2917   2.785906
#> 595 510.6812   2.636946
#> 596 502.0792   2.812088
#> 597 501.9423   2.717660
#> 598 506.2453   5.054258
#> 599 502.0291   2.991664
#> 600 503.9494   2.858756
#> 601 501.8890   2.567013
#> 602 503.1315   2.765147
#> 603 505.0487   2.748650
#> 604 524.2307   2.997040
#> 605 501.6011   2.572454
#> 606 629.7865   2.748745
#> 607 899.3413   3.396435
#> 608 500.0267   2.527311
#> 609 500.5198   2.522492
#> 610 667.2469   2.600835
#> 611 523.0043   2.567833
#> 612 502.4752   4.592587
#> 613 500.6716   4.372168
#> 614 819.3608   2.788887
#> 615 548.3365   2.652992
#> 616 501.8180   2.636168
#> 617 860.5060   2.933012
#> 618 500.0355   2.989627
#> 619 502.0872   3.319534
#> 620 517.0373   3.464927
#> 621 500.0031   2.508346
#> 622 826.1269   2.527957
#> 623 501.0894   3.000726
#> 624 502.4512   3.251705
#> 625 529.8672   2.710718
#> 626 501.1475   3.920122
#> 627 500.0153   2.637458
#> 628 500.0385   3.456163
#> 629 504.2198   2.511476
#> 630 501.5957   2.576901
#> 631 501.3362   3.176332
#> 632 516.7014   3.052861
#> 633 505.7807   2.972763
#> 634 511.3795   3.406911
#> 635 500.1279   3.675378
#> 636 561.6109   2.748878
#> 637 500.0106   2.953827
#> 638 504.4070   3.872333
#> 639 500.2368   2.593628
#> 640 609.5850   3.080422
#> 641 501.7690   2.844876
#> 642 501.9722   3.080779
#> 643 758.3280   2.515179
#> 644 503.1904   2.523502
#> 645 519.5134   2.745303
#> 646 504.0979   3.117492
#> 647 500.1031   3.184082
#> 648 500.4126   2.783666
#> 649 500.3135   3.427849
#> 650 820.7139   3.005179
#> 651 734.9565   3.077151
#> 652 500.5972   2.692658
#> 653 510.4616   2.644040
#> 654 500.0675   2.541893
#> 655 896.8003   2.536790
#> 656 529.1895   2.643680
#> 657 501.8822   2.506562
#> 658 918.0030   3.436908
#> 659 500.5404   2.514404
#> 660 500.3935   2.907228
#> 661 567.0528   3.394061
#> 662 500.6389   2.820973
#> 663 508.3290   3.166372
#> 664 557.1795   2.761505
#> 665 509.0794   2.988676
#> 666 500.2609   2.916587
#> 667 508.0307   2.855758
#> 668 503.8500   3.165031
#> 669 500.9053   2.550464
#> 670 765.1878   2.781957
#> 671 502.0353   2.525033
#> 672 500.0825   2.705221
#> 673 576.5142   2.807652
#> 674 513.8471   2.652658
#> 675 500.1787   2.899840
#> 676 517.1751   2.868560
#> 677 500.9948   2.531247
#> 678 504.0299   2.862075
#> 679 706.0090   2.676366
#> 680 588.3340   2.515559
#> 681 654.0396   4.782134
#> 682 503.3601   3.543706
#> 683 500.3260   2.785832
#> 684 514.6352   2.510686
#> 685 500.0786   3.245789
#> 686 516.6962   2.896470
#> 687 670.4213   3.342723
#> 688 572.3926   2.699202
#> 689 508.5171   2.914369
#> 690 500.0452   3.055118
#> 691 663.5928   2.960081
#> 692 500.0541   2.604482
#> 693 585.1485   3.181508
#> 694 513.5512   3.848556
#> 695 500.1445   2.540801
#> 696 524.3972   2.654317
#> 697 500.4154   3.006576
#> 698 502.4233   2.978207
#> 699 535.6231   2.555947
#> 700 502.2935   2.562482
#> 701 500.1638   3.486787
#> 702 500.0458   2.880021
#> 703 500.3032   2.795407
#> 704 603.1476   2.935019
#> 705 500.2694   3.298845
#> 706 642.4584   2.803176
#> 707 742.8298   2.573085
#> 708 502.7060   2.687782
#> 709 500.0047   4.292284
#> 710 500.4175   3.721698
#> 711 500.5037   2.679524
#> 712 500.8342   2.619660
#> 713 502.6843   3.060374
#> 714 500.0696   2.596107
#> 715 502.5341   2.710680
#> 716 502.4544   2.551718
#> 717 521.8617   3.059134
#> 718 504.3928   3.265300
#> 719 500.0176   2.579398
#> 720 501.0927   3.453203
#> 721 500.4762   2.506671
#> 722 504.0484   2.545331
#> 723 537.3886   2.637592
#> 724 500.0253   2.917111
#> 725 500.3471   2.944350
#> 726 667.7409   2.652624
#> 727 546.1754   2.775565
#> 728 500.0533   3.177384
#> 729 500.3069   3.650140
#> 730 565.3905   2.792381
#> 731 500.0066   2.801181
#> 732 902.9817   3.097795
#> 733 501.2250   2.556421
#> 734 579.2579   3.148958
#> 735 522.4762   2.578837
#> 736 882.7920   2.665298
#> 737 531.6182   2.790957
#> 738 500.1846   2.804050
#> 739 542.6950   3.344652
#> 740 500.7746   2.777039
#> 741 500.1622   2.882064
#> 742 500.2021   2.539850
#> 743 547.5520   2.850927
#> 744 521.8884   2.539433
#> 745 501.8848   2.705712
#> 746 500.1917   2.580550
#> 747 501.1542   2.717675
#> 748 500.0201   2.768760
#> 749 535.8906   2.821124
#> 750 500.1448   2.503489
#> 751 507.7674   2.625940
#> 752 710.9050   2.516702
#> 753 500.1567   2.769373
#> 754 500.1191   2.867356
#> 755 501.9325   2.593736
#> 756 561.9280   2.729073
#> 757 501.1256   2.521661
#> 758 500.1807   2.673386
#> 759 500.1179   2.981761
#> 760 548.8113   3.225133
#> 761 563.7926   3.007587
#> 762 501.0446   2.534933
#> 763 500.1418   2.596903
#> 764 500.1143   2.983742
#> 765 518.6885   3.524250
#> 766 504.7508   3.931687
#> 767 504.7002   2.956849
#> 768 510.9478   3.049611
#> 769 824.5693   2.810534
#> 770 929.8079   2.500741
#> 771 506.5492   2.822120
#> 772 511.8768   2.765116
#> 773 500.4225   3.146280
#> 774 500.3125   2.594385
#> 775 500.1316   2.815806
#> 776 801.6926   2.512708
#> 777 522.7609   2.842194
#> 778 501.0462   2.643453
#> 779 559.4819   2.717710
#> 780 500.1800   2.511715
#> 781 505.1783   2.557036
#> 782 527.4528   3.037337
#> 783 884.8803   3.046605
#> 784 501.6259   3.105599
#> 785 500.2042   4.432593
#> 786 516.9067   2.866716
#> 787 678.1871   2.699436
#> 788 501.2734   2.690150
#> 789 500.1175   3.029123
#> 790 501.1250   3.321507
#> 791 501.8299   3.169020
#> 792 500.5361   2.594802
#> 793 500.1201   2.558583
#> 794 555.5949   3.237070
#> 795 508.2343   2.524519
#> 796 501.4712   2.545775
#> 797 502.8323   3.321141
#> 798 500.1122   4.005305
#> 799 501.9703   2.551422
#> 800 501.8506   3.406938
#> 801 502.0330   2.779697
#> 802 714.5259   3.291591
#> 803 500.0674   2.702249
#> 804 616.5372   3.143398
#> 805 507.7165   2.811907
#> 806 501.3696   2.572750
#> 807 784.2941   2.892802
#> 808 500.1414   2.965114
#> 809 809.8335   2.795329
#> 810 506.0346   2.551116
#> 811 541.3282   3.328973
#> 812 551.2357   3.948259
#> 813 504.1348   2.609845
#> 814 511.8375   4.417843
#> 815 525.7711   2.812849
#> 816 571.6579   2.508866
#> 817 501.4282   2.534379
#> 818 500.0514   2.809579
#> 819 500.1251   2.588204
#> 820 995.0311   3.822145
#> 821 500.0146   2.895453
#> 822 500.0943   2.670011
#> 823 506.2282   3.351986
#> 824 609.0743   2.649217
#> 825 524.4506   2.757138
#> 826 548.1895   2.628855
#> 827 522.1725   2.584542
#> 828 509.1520   2.703652
#> 829 682.8664   5.730833
#> 830 500.1778   2.718832
#> 831 513.6080   2.645970
#> 832 505.9841   2.521083
#> 833 548.5208   2.557096
#> 834 673.4845   2.681041
#> 835 500.3025   2.858483
#> 836 600.9613   2.872348
#> 837 500.7564   2.924533
#> 838 500.0558   2.972890
#> 839 625.4239   2.546237
#> 840 799.4256   3.270191
#> 841 500.3446   2.531006
#> 842 501.5535   2.567527
#> 843 500.0390   2.648288
#> 844 589.0885   3.254276
#> 845 505.7197   2.554520
#> 846 500.2620   2.644186
#> 847 501.0514   2.516978
#> 848 500.0547   3.067934
#> 849 731.3757   2.742825
#> 850 500.1009   2.998471
#> 851 500.0295   2.808382
#> 852 500.3011   2.754203
#> 853 603.3437   2.886197
#> 854 718.6032   2.893303
#> 855 540.0048   2.511534
#> 856 816.8975   3.418433
#> 857 500.0863   2.843941
#> 858 500.9326   2.533197
#> 859 501.0292   3.345556
#> 860 501.3287   2.712574
#> 861 879.2404   4.153649
#> 862 510.6926   2.660378
#> 863 500.0087   2.663708
#> 864 509.9504   2.678376
#> 865 603.2896   3.088685
#> 866 500.0911   2.713230
#> 867 503.3685   3.288722
#> 868 514.7342   2.738868
#> 869 526.0733   2.656857
#> 870 946.3079   2.606376
#> 871 831.5613   2.835021
#> 872 500.4316   3.112753
#> 873 500.2241   2.915607
#> 874 532.2841   3.062612
#> 875 521.4342   3.472836
#> 876 500.1010   2.747559
#> 877 500.5286   2.508471
#> 878 500.4491   2.947694
#> 879 502.5168   2.640630
#> 880 628.5550   2.677722
#> 881 518.5163   2.894035
#> 882 879.7695   3.739294
#> 883 501.6543   2.907303
#> 884 578.8366   2.630082
#> 885 500.3205   2.982348
#> 886 503.4116   2.655839
#> 887 500.1631   2.895088
#> 888 501.1545   2.923472
#> 889 507.9808   2.975311
#> 890 500.4067   2.995587
#> 891 500.1670   3.088300
#> 892 500.0549   2.575112
#> 893 500.6213   2.684021
#> 894 623.3744   2.730925
#> 895 500.4929   2.739611
#> 896 500.2339   3.274551
#> 897 500.8305   2.857666
#> 898 507.2076   2.589457
#> 899 501.0247   3.194475
#> 900 504.6926   2.593332
#> 901 500.2514   3.237475
#> 902 500.2389   3.224055
#> 903 586.1973   2.619675