Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   500.0799   3.069963
#> 2   501.8686   2.568093
#> 3   518.6628   2.659556
#> 4   531.5754   2.574316
#> 5   500.2132   2.519218
#> 6   539.9867   3.690318
#> 7   638.4927   4.299014
#> 8   510.6262   2.521603
#> 9   500.1531   2.573876
#> 10  501.5937   2.590356
#> 11  500.9860   2.656973
#> 12  500.0778   2.518227
#> 13  656.4156   2.843291
#> 14  501.6785   3.003386
#> 15  500.0753   2.532397
#> 16  503.0449   2.684075
#> 17  517.8610   2.974439
#> 18  519.8132   3.449438
#> 19  542.3888   2.691417
#> 20  500.4154   3.094746
#> 21  937.6286   2.526500
#> 22  738.7756   3.080944
#> 23  502.9696   2.827339
#> 24  500.9494   3.007377
#> 25  500.0745   2.790658
#> 26  502.3221   2.558911
#> 27  502.7987   3.340848
#> 28  611.9643   2.869650
#> 29  502.7091   3.108157
#> 30  501.1895   2.597875
#> 31  996.6708   3.001033
#> 32  500.5432   2.746072
#> 33  501.8637   2.739142
#> 34  500.4367   2.530652
#> 35  500.9159   3.548269
#> 36  730.6214   2.682134
#> 37  501.4981   2.538276
#> 38  781.8431   2.871633
#> 39  500.0376   3.593369
#> 40  500.7752   3.389519
#> 41  500.3333   3.455301
#> 42  502.3976   2.615554
#> 43  505.6151   2.625124
#> 44  500.9185   2.962927
#> 45  500.0637   3.033883
#> 46  500.1930   2.596594
#> 47  513.8370   2.940840
#> 48  500.0276   2.968735
#> 49  505.2054   2.559020
#> 50  911.7166   2.961229
#> 51  822.8891   2.985317
#> 52  501.6094   2.525270
#> 53  502.1900   2.823731
#> 54  512.5593   2.863641
#> 55  500.0047   2.509127
#> 56  533.1183   2.823798
#> 57  754.5095   2.703077
#> 58  501.3022   2.532728
#> 59  502.2568   3.020978
#> 60  522.1150   2.545102
#> 61  502.0709   2.826356
#> 62  525.2116   3.134530
#> 63  814.6001   2.661462
#> 64  502.0196   3.572318
#> 65  503.6158   3.052836
#> 66  500.5319   3.517857
#> 67  543.4457   2.559363
#> 68  501.8343   3.437552
#> 69  502.3287   3.177386
#> 70  500.5009   3.019857
#> 71  501.2066   2.749615
#> 72  502.6531   2.548215
#> 73  567.8756   2.719902
#> 74  503.6458   3.594270
#> 75  500.2003   3.971217
#> 76  515.2789   2.518095
#> 77  690.1905   3.459213
#> 78  500.8859   2.675505
#> 79  934.2094   3.276009
#> 80  641.9730   2.639492
#> 81  500.1835   3.299884
#> 82  500.0498   2.725905
#> 83  501.2537   2.957260
#> 84  524.0954   3.174379
#> 85  532.2242   2.518700
#> 86  500.4293   2.650947
#> 87  501.0233   3.143583
#> 88  941.4584   3.351498
#> 89  549.0752   2.599892
#> 90  517.3376   2.753328
#> 91  500.5174   2.927311
#> 92  676.0750   2.998368
#> 93  563.6500   3.034489
#> 94  502.1467   2.570053
#> 95  505.8350   2.738560
#> 96  500.0758   2.561469
#> 97  589.7514   2.640753
#> 98  500.1199   2.871215
#> 99  502.8216   2.744851
#> 100 500.0315   2.847304
#> 101 580.0571   2.523511
#> 102 500.1559   2.536364
#> 103 500.1843   2.661096
#> 104 509.5146   2.510395
#> 105 500.7205   2.628840
#> 106 501.8504   2.664411
#> 107 607.8917   3.866933
#> 108 933.2666   3.782442
#> 109 500.0753   2.705741
#> 110 501.6009   2.743503
#> 111 590.6652   2.908994
#> 112 593.6747   2.955286
#> 113 566.9326   2.916094
#> 114 504.5806   3.501687
#> 115 512.1989   3.366687
#> 116 521.3417   2.537005
#> 117 500.3787   3.339281
#> 118 556.7172   2.708539
#> 119 513.9331   3.017813
#> 120 500.6825   3.254621
#> 121 537.4082   3.911418
#> 122 543.7304   2.892472
#> 123 500.6274   2.752954
#> 124 618.8923   2.816771
#> 125 535.6203   2.540577
#> 126 526.3267   2.614935
#> 127 500.2893   2.642741
#> 128 500.3942   2.685273
#> 129 500.1582   2.561608
#> 130 682.5460   2.517575
#> 131 500.8576   3.125526
#> 132 887.9050   2.791715
#> 133 521.0961   2.694171
#> 134 938.7208   2.681375
#> 135 517.9898   3.157866
#> 136 730.2880   2.854988
#> 137 500.6620   2.616619
#> 138 500.4838   2.631229
#> 139 505.4630   2.849546
#> 140 500.3176   3.244453
#> 141 531.9516   2.956684
#> 142 696.3977   2.954240
#> 143 504.9107   2.715447
#> 144 500.0354   2.677116
#> 145 500.1007   3.848819
#> 146 504.9260   2.583689
#> 147 500.0852   2.735265
#> 148 500.0376   2.773577
#> 149 501.1598   2.638553
#> 150 506.8382   3.052377
#> 151 587.3095   2.887762
#> 152 500.0429   2.506034
#> 153 672.7104   2.677631
#> 154 788.8607   3.773552
#> 155 500.3195   4.698850
#> 156 506.5457   3.956564
#> 157 515.2693   2.745273
#> 158 500.8464   2.827815
#> 159 500.0731   2.887499
#> 160 502.5187   2.751592
#> 161 542.4034   2.635452
#> 162 579.2691   3.327961
#> 163 530.7331   2.566307
#> 164 501.8164   2.745502
#> 165 500.8604   2.553254
#> 166 504.7545   2.850347
#> 167 574.0504   3.233053
#> 168 500.0660   2.740826
#> 169 507.8876   2.944225
#> 170 504.1676   2.600221
#> 171 500.1465   2.752643
#> 172 500.5199   4.331165
#> 173 627.7996   2.925774
#> 174 505.0245   4.224129
#> 175 511.8922   2.518131
#> 176 513.1060   2.541908
#> 177 536.5784   3.786379
#> 178 500.1213   2.696174
#> 179 519.7187   2.757009
#> 180 782.9542   2.818931
#> 181 501.0019   4.338360
#> 182 537.7887   2.785187
#> 183 501.1921   2.533689
#> 184 500.2336   3.308783
#> 185 834.7400   2.548741
#> 186 500.3417   2.848920
#> 187 500.0540   2.579812
#> 188 530.2081   3.029535
#> 189 500.3047   2.733016
#> 190 502.7626   2.500334
#> 191 537.9670   2.872816
#> 192 500.1181   2.561678
#> 193 509.3525   2.655142
#> 194 500.0441   2.785504
#> 195 505.9487   2.658156
#> 196 500.1995   2.712470
#> 197 586.3725   2.944246
#> 198 513.5555   3.044199
#> 199 507.7787   2.941747
#> 200 618.3299   2.531861
#> 201 532.5481   2.625547
#> 202 500.7576   2.596952
#> 203 501.2280   2.523901
#> 204 500.2808   2.875184
#> 205 500.2244   3.668764
#> 206 544.5548   3.450987
#> 207 506.0453   2.505210
#> 208 501.4562   2.619121
#> 209 500.0835   2.844916
#> 210 500.3068   2.566856
#> 211 524.2655   2.578967
#> 212 500.4105   2.960778
#> 213 511.2824   2.557316
#> 214 500.2832   2.998104
#> 215 500.6655   2.799100
#> 216 501.0050   2.921549
#> 217 593.3473   3.006024
#> 218 610.6962   2.822650
#> 219 502.4086   2.538475
#> 220 563.4343   3.399591
#> 221 506.3971   2.659581
#> 222 500.7947   3.031434
#> 223 625.0542   2.812052
#> 224 500.0218   2.587432
#> 225 520.2555   3.644262
#> 226 501.3693   2.591427
#> 227 688.7428   3.306325
#> 228 500.2228   2.709718
#> 229 615.1392   3.032379
#> 230 500.1837   2.708772
#> 231 502.3846   2.926077
#> 232 500.1543   2.872572
#> 233 501.4966   2.710980
#> 234 500.1119   2.530797
#> 235 501.4972   3.120721
#> 236 500.2519   2.968051
#> 237 500.2135   2.580361
#> 238 777.9205   2.963112
#> 239 501.8517   2.815193
#> 240 500.8909   3.460420
#> 241 565.2355   2.702066
#> 242 665.7531   2.758247
#> 243 983.6997   2.782586
#> 244 500.3913   2.979482
#> 245 501.7219   2.641491
#> 246 500.4827   2.731984
#> 247 500.0830   2.678435
#> 248 500.4453   2.547156
#> 249 500.0132   3.390842
#> 250 500.7288   2.638935
#> 251 500.1323   2.551589
#> 252 554.1508   2.844811
#> 253 510.6190   3.119582
#> 254 501.5443   2.538275
#> 255 572.8156   2.698970
#> 256 500.4060   2.838371
#> 257 500.0258   2.561657
#> 258 509.9392   3.510287
#> 259 501.4522   2.969513
#> 260 500.1956   3.833587
#> 261 500.2960   3.179464
#> 262 693.0259   2.689603
#> 263 561.8534   3.096420
#> 264 502.5407   2.635760
#> 265 500.1071   7.064527
#> 266 501.8475   2.872879
#> 267 501.7824   3.195580
#> 268 601.4262   2.677841
#> 269 500.6024   2.654147
#> 270 500.9165   2.757192
#> 271 501.0027   3.889108
#> 272 500.2405   2.788200
#> 273 500.0327   2.548134
#> 274 665.0193   2.538701
#> 275 501.7958   4.148786
#> 276 500.3576   2.535674
#> 277 500.3505   2.774672
#> 278 504.5216   2.924125
#> 279 501.3852   3.793337
#> 280 500.0937   3.464661
#> 281 500.1838   2.927389
#> 282 514.8644   2.866128
#> 283 661.8868   2.841943
#> 284 505.5044   2.718382
#> 285 500.3502   3.708968
#> 286 771.4178   2.534366
#> 287 722.1025   2.704907
#> 288 500.0218   2.661830
#> 289 513.9005   2.680022
#> 290 650.5488   2.530697
#> 291 502.0394   2.504826
#> 292 500.4176   3.253272
#> 293 507.9129   2.525688
#> 294 500.3226   3.203999
#> 295 500.2686   2.706462
#> 296 500.1281   2.952599
#> 297 500.3558   2.701134
#> 298 502.4522   2.749937
#> 299 500.0521   3.044417
#> 300 500.0158   2.751731
#> 301 540.3635   2.946569
#> 302 502.2653   2.854263
#> 303 503.9204   2.700549
#> 304 500.3096   2.972588
#> 305 504.0615   2.955277
#> 306 500.5360   2.584571
#> 307 716.4828   2.646562
#> 308 500.0089   3.160826
#> 309 789.0687   2.519819
#> 310 500.2629   2.588423
#> 311 500.1323   2.828268
#> 312 509.3493   2.887329
#> 313 519.3686   3.059658
#> 314 504.4321   3.091437
#> 315 710.3327   2.540073
#> 316 500.2981   2.500454
#> 317 512.9896   2.761302
#> 318 660.4598   3.830434
#> 319 500.0114   2.569604
#> 320 500.2061   3.585650
#> 321 501.2044   2.914454
#> 322 500.0514   2.783540
#> 323 500.5082   2.613590
#> 324 781.6083   3.153784
#> 325 943.3962   3.124761
#> 326 607.7694   2.900557
#> 327 526.7230   3.159447
#> 328 870.4742   2.821276
#> 329 500.3581   3.019909
#> 330 502.2781   2.638939
#> 331 513.1568   2.769455
#> 332 808.2509   2.973038
#> 333 524.8387   3.255703
#> 334 842.0961   4.659429
#> 335 503.2523   3.868294
#> 336 895.9028   3.031334
#> 337 501.0261   2.679378
#> 338 701.6040   2.752379
#> 339 501.1679   3.530706
#> 340 518.4760   3.312926
#> 341 572.4404   2.589307
#> 342 503.8093   2.792103
#> 343 505.5694   2.527822
#> 344 500.1755   3.439694
#> 345 500.0132   2.576453
#> 346 517.6311   2.764471
#> 347 538.8924   2.670254
#> 348 500.5435   2.614334
#> 349 501.0557   2.727223
#> 350 502.7203   2.904978
#> 351 500.0406   3.814841
#> 352 500.8875   2.669017
#> 353 500.5737   2.653810
#> 354 500.1370   2.789283
#> 355 500.2642   2.508129
#> 356 504.0761   2.510634
#> 357 500.1570   2.817579
#> 358 665.8711   2.608621
#> 359 597.9271   2.800084
#> 360 500.3871   2.549632
#> 361 531.5012   3.051690
#> 362 526.7686   2.583967
#> 363 500.0319   3.735167
#> 364 899.9855   3.667208
#> 365 566.9124   2.623305
#> 366 744.2492   2.812321
#> 367 500.2625   4.207021
#> 368 506.1309   3.091991
#> 369 846.4705   3.360881
#> 370 500.2809   3.038012
#> 371 509.4374   2.617573
#> 372 501.5084   3.378164
#> 373 500.3249   3.530372
#> 374 500.2778   2.770176
#> 375 500.2411   2.861242
#> 376 500.0743   2.703986
#> 377 500.0539   2.989994
#> 378 501.2530   3.401941
#> 379 601.8088   2.628995
#> 380 508.8403   2.601253
#> 381 696.0050   2.516473
#> 382 502.4179   2.799850
#> 383 501.9920   2.755059
#> 384 727.9726   3.769582
#> 385 510.5972   2.862590
#> 386 500.0075   3.185217
#> 387 510.3804   3.832449
#> 388 500.0556   3.368306
#> 389 544.2065   3.088642
#> 390 500.4087   4.341765
#> 391 509.9807   2.518011
#> 392 502.8112   2.560502
#> 393 623.8992   2.667025
#> 394 501.7927   3.214928
#> 395 500.0741   5.189046
#> 396 500.1767   2.752781
#> 397 500.0668   2.970397
#> 398 500.0059   2.630585
#> 399 598.0289   2.878419
#> 400 554.0364   2.574285
#> 401 500.8067   2.896194
#> 402 512.0397   2.622312
#> 403 580.4797   2.767008
#> 404 506.8414   3.026421
#> 405 501.8825   2.644763
#> 406 554.9324   2.974659
#> 407 702.1478   2.615535
#> 408 500.6725   2.624683
#> 409 523.1167   2.575538
#> 410 503.5723   3.139127
#> 411 522.1599   2.592067
#> 412 523.7303   2.703927
#> 413 507.8591   3.087800
#> 414 500.2360   2.672079
#> 415 533.4152   3.317890
#> 416 501.4909   3.710217
#> 417 525.1802   2.554838
#> 418 500.1808   2.875607
#> 419 542.7412   3.467327
#> 420 500.0065   2.505160
#> 421 507.5341   2.559299
#> 422 574.1373   3.071119
#> 423 595.9903   2.572836
#> 424 949.8478   3.042195
#> 425 501.1362   3.919155
#> 426 518.2100   2.538707
#> 427 500.0745   2.781134
#> 428 505.1638   2.706890
#> 429 500.0355   2.646653
#> 430 500.4859   3.155162
#> 431 504.7456   2.758470
#> 432 510.8830   3.902944
#> 433 680.2905   3.179143
#> 434 500.7836   2.596249
#> 435 526.7316   4.010808
#> 436 517.7982   2.551942
#> 437 500.7925   3.261364
#> 438 501.4072   4.007261
#> 439 512.7941   2.818230
#> 440 500.7994   2.621480
#> 441 518.7922   2.581100
#> 442 585.6856   3.079749
#> 443 878.3401   2.926749
#> 444 538.7000   3.448351
#> 445 500.7846   2.625142
#> 446 548.2229   2.692903
#> 447 500.2478   2.749172
#> 448 501.0005   2.930253
#> 449 505.4366   2.667531
#> 450 500.2096   2.987798
#> 451 500.0290   4.150152
#> 452 500.4257   2.679369
#> 453 989.0480   3.813381
#> 454 500.0352   2.811702
#> 455 556.7583   4.018440
#> 456 500.1090   3.117241
#> 457 500.3563   2.707308
#> 458 500.0577   2.690186
#> 459 501.0336   3.917997
#> 460 501.7633   2.889372
#> 461 998.9485   3.041764
#> 462 612.4287   3.216968
#> 463 509.9761   2.608287
#> 464 536.4430   3.056220
#> 465 531.3930   3.088141
#> 466 500.0445   2.899480
#> 467 751.3790   2.810121
#> 468 545.8507   3.779158
#> 469 500.0502   2.914723
#> 470 658.2520   2.664646
#> 471 500.3242   3.081890
#> 472 506.7691   2.860560
#> 473 500.1125   2.541612
#> 474 594.1173   2.625518
#> 475 570.8434   3.048272
#> 476 500.0812   2.914085
#> 477 500.9567   2.726117
#> 478 500.4977   2.622975
#> 479 500.2075   3.225610
#> 480 500.1348   2.784969
#> 481 500.4819   4.817875
#> 482 500.2282   3.392302
#> 483 501.7187   3.359770
#> 484 500.0234   2.851121
#> 485 500.2923   2.933199
#> 486 500.8797   3.039947
#> 487 501.9962   4.219222
#> 488 507.3253   3.414414
#> 489 501.4146   2.597418
#> 490 500.0335   3.142483
#> 491 660.9449   2.789219
#> 492 549.5594   2.674151
#> 493 502.2814   3.652696
#> 494 501.8869   3.368271
#> 495 510.6244   2.776072
#> 496 509.8032   2.920946
#> 497 506.7009   2.573431
#> 498 501.3977   2.805251
#> 499 500.0755   2.718482
#> 500 629.1118   2.972493
#> 501 506.7816   2.688310
#> 502 690.9396   2.533355
#> 503 501.9876   3.291598
#> 504 500.2646   2.792820
#> 505 500.4979   2.575635
#> 506 517.5131   2.677646
#> 507 504.6360   2.806607
#> 508 500.7374   2.745981
#> 509 747.0180   3.204413
#> 510 505.5767   3.240445
#> 511 562.2262   2.930885
#> 512 503.7063   3.288375
#> 513 500.1753   2.647612
#> 514 500.0165   2.603319
#> 515 908.2955   2.570981
#> 516 501.1519   2.637736
#> 517 500.1513   3.183601
#> 518 611.3939   2.801830
#> 519 500.0255   2.553220
#> 520 536.9782   2.681617
#> 521 501.4916   3.099002
#> 522 500.4882   2.584290
#> 523 501.4589   3.278927
#> 524 759.2234   3.287377
#> 525 503.8226   2.611905
#> 526 502.7924   2.783425
#> 527 503.6486   3.536961
#> 528 541.3882   2.501405
#> 529 501.1517   3.703960
#> 530 500.6230   2.865770
#> 531 571.6126   3.506047
#> 532 964.9534   2.821087
#> 533 505.0075   2.600085
#> 534 501.8830   2.551453
#> 535 518.0977   2.671472
#> 536 500.2602   3.825529
#> 537 500.6470   2.860477
#> 538 570.5312   2.846846
#> 539 500.0275   3.001471
#> 540 500.2123   3.072488
#> 541 500.0106   2.763384
#> 542 500.0460   2.746193
#> 543 500.3866   3.130178
#> 544 503.2276   2.544026
#> 545 500.0840   2.814791
#> 546 534.2859   3.841294
#> 547 542.2266   2.654449
#> 548 646.7205   2.540206
#> 549 501.6866   2.740933
#> 550 522.9843   2.727205
#> 551 500.4239   2.580521
#> 552 500.1492   3.127376
#> 553 521.7289   3.876499
#> 554 500.1349   2.809431
#> 555 508.4381   3.423034
#> 556 522.0203   2.647499
#> 557 504.8666   3.184091
#> 558 502.0413   2.561369
#> 559 512.0470   2.978343
#> 560 500.9539   2.860654
#> 561 504.8911   2.534408
#> 562 501.2101   3.246508
#> 563 506.1702   2.669625
#> 564 501.1592   2.675579
#> 565 509.4795   2.705007
#> 566 500.9064   2.510656
#> 567 500.0804   2.506010
#> 568 511.9469   2.747538
#> 569 719.2846   2.721683
#> 570 502.4138   3.216061
#> 571 536.9271   2.540844
#> 572 501.6010   3.079703
#> 573 593.5693   3.026423
#> 574 500.1214   2.565936
#> 575 649.6370   3.540468
#> 576 528.9325   2.855599
#> 577 770.7629   3.021876
#> 578 660.5612   2.755330
#> 579 501.0285   2.678314
#> 580 500.6169   3.120808
#> 581 585.8774   3.177028
#> 582 501.6903   2.767530
#> 583 928.9910   2.772082
#> 584 500.0195   2.704833
#> 585 773.9797   2.719885
#> 586 500.4160   4.131872
#> 587 500.9949   2.790416
#> 588 500.0714   3.111401
#> 589 663.1399   2.570358
#> 590 578.5810   2.898646
#> 591 500.1188   2.738641
#> 592 509.7124   2.549280
#> 593 507.1396   3.117044
#> 594 500.3892   2.906937
#> 595 664.4233   3.062947
#> 596 500.0219   2.967051
#> 597 786.2734   2.536536
#> 598 500.1700   2.615148
#> 599 609.8165   2.896568
#> 600 500.1364   2.561168
#> 601 501.4825   2.822372
#> 602 500.1166   2.974060
#> 603 502.3713   3.181908
#> 604 506.2513   3.202214
#> 605 501.2528   2.599756
#> 606 502.0715   2.702257
#> 607 515.7084   2.744388
#> 608 505.5564   2.530607
#> 609 510.5045   3.205371
#> 610 500.0543   3.019541
#> 611 500.0864   2.642787
#> 612 623.5184   3.464011
#> 613 551.2494   2.648581
#> 614 500.3820   2.804087
#> 615 501.1839   3.179238
#> 616 580.4070   5.578855
#> 617 502.6280   3.231052
#> 618 500.0542   2.659126
#> 619 500.0628   2.736561
#> 620 500.3158   2.754005
#> 621 506.7389   3.140819
#> 622 500.2188   2.613546
#> 623 500.4361   2.629537
#> 624 509.7700   2.571146
#> 625 561.3279   3.972647
#> 626 712.4720   3.199235
#> 627 500.1097   2.713061
#> 628 500.5282   2.854998
#> 629 501.3549   4.702292
#> 630 513.2741   2.801738
#> 631 500.2049   3.126500
#> 632 503.2741   2.638272
#> 633 636.2031   2.779517
#> 634 505.9446   4.086524
#> 635 504.9859   2.921215
#> 636 501.3316   2.572136
#> 637 500.4263   3.642677
#> 638 570.7194   2.525833
#> 639 500.3931   2.698603
#> 640 501.3341   3.195495
#> 641 500.1312   2.774830
#> 642 504.9016   2.726582
#> 643 605.1893   2.560151
#> 644 720.4238   3.484351
#> 645 500.6463   2.914469
#> 646 502.7847   2.544920
#> 647 674.2100   2.832303
#> 648 504.3565   2.579360
#> 649 505.5827   2.884589
#> 650 500.0954   2.824026
#> 651 501.8825   2.831935
#> 652 500.1778   2.629828
#> 653 718.9285   2.689375
#> 654 507.9355   3.209554
#> 655 500.3255   2.521422
#> 656 527.5261   2.882742
#> 657 500.3868   2.636234
#> 658 500.0051   2.592443
#> 659 552.8983   3.259424
#> 660 500.4600   2.569751
#> 661 502.3117   3.075080
#> 662 540.1925   2.524763
#> 663 500.0762   2.845698
#> 664 500.7176   2.651028
#> 665 500.6142   3.544852
#> 666 500.0391   3.112201
#> 667 502.5187   2.621882
#> 668 501.6412   2.568387
#> 669 583.6903   2.516397
#> 670 500.1569   3.835432
#> 671 500.1356   3.471301
#> 672 550.9219   2.719195
#> 673 500.0686   3.578176
#> 674 517.8100   2.871404
#> 675 505.3801   2.727914
#> 676 646.6827   2.605209
#> 677 500.4661   3.002522
#> 678 574.3866   3.032522
#> 679 501.4920   3.074728
#> 680 677.3220   2.652497
#> 681 802.8040   3.535169
#> 682 500.5624   2.977146
#> 683 501.3771   2.881833
#> 684 633.9755   2.773748
#> 685 501.7648   2.750713
#> 686 523.4124   2.817624
#> 687 517.1506   3.427469
#> 688 500.1157   3.076896
#> 689 500.0221   3.152736
#> 690 500.2942   2.855757
#> 691 500.5312   2.711138
#> 692 996.5872   3.166950
#> 693 619.4088   3.508046
#> 694 500.1194   2.598709
#> 695 500.4498   3.400862
#> 696 869.2188   2.949872
#> 697 631.4239   2.942744
#> 698 754.8418   2.585367
#> 699 500.1054   2.626673
#> 700 500.0731   2.729119
#> 701 508.3899   2.595797
#> 702 504.0758   2.994139
#> 703 503.2838   2.574188
#> 704 518.8639   2.509900
#> 705 500.1453   2.762230
#> 706 500.0573   2.719833
#> 707 517.2600   2.868227
#> 708 573.6654   3.706147
#> 709 537.5875   2.817506
#> 710 501.4685   3.227363
#> 711 521.8678   2.554758
#> 712 796.1874   2.550614
#> 713 501.5222   2.774379
#> 714 501.2768   2.879745
#> 715 500.0878   2.905218
#> 716 515.2553   2.599328
#> 717 500.0364   3.066326
#> 718 633.0836   2.565648
#> 719 500.0734   2.512846
#> 720 500.6116   2.520753
#> 721 785.9122   2.561649
#> 722 502.9600   2.567500
#> 723 500.2245   2.646813
#> 724 500.0223   2.684337
#> 725 500.9357   3.995414
#> 726 500.1579   3.204387
#> 727 590.9442   2.799134
#> 728 505.5786   2.982841
#> 729 517.0915   2.502372
#> 730 500.5567   3.806171
#> 731 506.3796   3.831788
#> 732 500.5364   3.478244
#> 733 506.0096   2.511917
#> 734 500.2005   2.940567
#> 735 500.1079   3.368700
#> 736 506.6041   2.541217
#> 737 520.2837   2.940936
#> 738 501.2062   2.526885
#> 739 535.8296   3.195814
#> 740 558.6252   2.596716
#> 741 501.5160   2.760438
#> 742 501.6124   4.112129
#> 743 500.4097   3.937907
#> 744 969.6430   2.777410
#> 745 506.5455   3.725373
#> 746 500.4018   2.589758
#> 747 667.7337   2.791158
#> 748 500.0729   3.399758
#> 749 501.9307   2.899920
#> 750 515.9300   2.949437
#> 751 503.3805   2.591351
#> 752 500.7899   2.932832
#> 753 500.1832   2.517303
#> 754 601.1606   2.635524
#> 755 500.0655   3.567583
#> 756 501.2379   3.117337
#> 757 506.2836   3.190344
#> 758 500.7519   3.562617
#> 759 500.0002   2.716838
#> 760 502.8072   2.813244
#> 761 511.2377   3.450586
#> 762 500.2544   2.846319
#> 763 500.0386   3.055572
#> 764 504.8288   2.602282
#> 765 500.0901   2.760134
#> 766 503.8535   2.579275
#> 767 617.6498   2.664660
#> 768 512.7227   3.207250
#> 769 512.1473   2.722242
#> 770 500.0419   2.525036
#> 771 500.2514   3.133561
#> 772 643.2709   2.861831
#> 773 761.5360   3.538978
#> 774 500.2697   2.638556
#> 775 500.4576   3.184966
#> 776 500.3051   2.553758
#> 777 500.0831   2.748604
#> 778 514.2624   3.530677
#> 779 501.7029   2.708793
#> 780 500.8135   2.556408
#> 781 597.0072   3.435334
#> 782 500.5347   3.245615
#> 783 500.1241   2.928085
#> 784 594.2369   2.860413
#> 785 500.5644   2.503465
#> 786 513.5990   2.721449
#> 787 501.0640   2.770531
#> 788 533.2540   3.090839
#> 789 500.2976   2.830803
#> 790 504.8816   2.797561
#> 791 916.2878   3.077673
#> 792 524.1485   3.206380
#> 793 512.0129   2.601658
#> 794 671.3581   2.669819
#> 795 500.8077   2.666475
#> 796 582.9017   3.318522
#> 797 779.4459   2.531227
#> 798 500.4602   4.495840
#> 799 562.2668   2.595924
#> 800 507.8792   3.007525
#> 801 517.2100   2.560047
#> 802 719.9261   2.819587
#> 803 634.0256   2.643145
#> 804 500.2195   2.979932
#> 805 505.0486   2.723848
#> 806 500.8663   2.612883
#> 807 500.0079   2.878268
#> 808 551.0633   2.900641
#> 809 511.1001   2.570488
#> 810 508.9623   2.732691
#> 811 502.3250   2.683403
#> 812 599.5146   2.576203
#> 813 508.5001   4.150260
#> 814 500.4615   2.754298
#> 815 500.2504   2.948023
#> 816 500.6124   2.571689
#> 817 528.8946   3.150595
#> 818 693.2262   2.536584
#> 819 540.4571   2.672310
#> 820 500.8188   3.027256
#> 821 506.1080   2.887245
#> 822 500.2494   3.215044
#> 823 500.1170   2.992858
#> 824 684.2126   2.622317
#> 825 507.0201   2.513955
#> 826 500.1754   3.086263
#> 827 986.6746   2.969614
#> 828 500.0498   2.949435
#> 829 521.6441   2.541026
#> 830 500.5143   3.405711
#> 831 807.3142   2.726459
#> 832 500.4846   2.822969
#> 833 500.1171   2.594083
#> 834 570.3315   3.333600
#> 835 500.1027   3.232378
#> 836 654.0256   2.595302
#> 837 500.0558   2.804013
#> 838 502.0989   3.142338
#> 839 500.5712   3.353051
#> 840 761.0686   2.757679
#> 841 567.7263   3.479345
#> 842 500.3832   2.504800
#> 843 502.8363   2.946613
#> 844 514.5026   2.620923
#> 845 504.2837   2.615736
#> 846 500.0367   2.793210
#> 847 500.0142   3.213193
#> 848 500.0867   3.285444
#> 849 500.2463   2.967111
#> 850 502.1958   3.227931
#> 851 524.2368   4.410768
#> 852 504.4786   3.349385
#> 853 502.5332   2.511870
#> 854 501.2481   2.981225