Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = NULL)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

Deprecated argument for controlling parallelism. Use future::plan(future::multisession, workers = ncore) (or similar) to configure parallelism in your code instead.

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   596.4649   3.087894
#> 2   500.1122   2.774164
#> 3   500.0057   2.507601
#> 4   597.7804   3.339860
#> 5   566.0529   2.628501
#> 6   511.0387   3.321288
#> 7   500.6373   3.095547
#> 8   501.9968   2.813735
#> 9   500.1709   3.008952
#> 10  502.3830   2.604203
#> 11  520.8047   2.609678
#> 12  500.4991   3.429913
#> 13  500.3099   2.869055
#> 14  519.6111   2.663976
#> 15  500.1418   2.651358
#> 16  538.9235   2.534357
#> 17  502.2413   2.808633
#> 18  500.6455   2.998234
#> 19  500.0348   2.524617
#> 20  501.6893   2.870990
#> 21  500.0966   2.933973
#> 22  671.7110   2.575323
#> 23  500.0548   2.560560
#> 24  500.3422   3.502583
#> 25  502.0122   2.989396
#> 26  573.2066   3.499021
#> 27  500.0392   2.590948
#> 28  699.6414   3.427095
#> 29  500.2599   2.612848
#> 30  503.6165   2.594710
#> 31  530.6867   2.519976
#> 32  500.6218   3.037130
#> 33  501.8455   2.545104
#> 34  502.7070   2.514679
#> 35  762.3669   2.930360
#> 36  501.4216   2.516497
#> 37  500.7291   2.838297
#> 38  500.2337   2.699663
#> 39  588.9177   2.538560
#> 40  539.7377   2.539059
#> 41  899.2145   3.238032
#> 42  505.2645   2.585394
#> 43  500.0373   2.535459
#> 44  505.4941   2.892198
#> 45  506.5943   2.889895
#> 46  687.1065   2.785666
#> 47  501.2727   3.680375
#> 48  645.7387   3.064866
#> 49  518.0455   2.573875
#> 50  500.0653   2.778417
#> 51  698.5009   2.641699
#> 52  500.4072   2.607259
#> 53  501.8497   2.533237
#> 54  502.6657   3.092731
#> 55  500.7450   2.550104
#> 56  500.1066   2.600979
#> 57  510.3389   2.509237
#> 58  543.4225   3.213187
#> 59  500.3129   3.205261
#> 60  654.3582   2.524976
#> 61  500.3396   2.597174
#> 62  501.5882   3.495466
#> 63  504.1981   2.518748
#> 64  502.4294   3.395701
#> 65  505.5290   2.598675
#> 66  500.2631   2.723253
#> 67  500.0226   2.829234
#> 68  505.7506   3.279853
#> 69  768.0779   2.568042
#> 70  519.2648   3.180196
#> 71  516.7468   2.778039
#> 72  592.6559   2.967478
#> 73  526.8675   2.821520
#> 74  505.3326   2.547752
#> 75  508.9162   2.890121
#> 76  507.0857   2.905206
#> 77  579.2997   2.680198
#> 78  500.2734   2.866006
#> 79  501.3228   3.033441
#> 80  506.0067   2.727696
#> 81  533.1987   2.905498
#> 82  505.5320   3.057805
#> 83  500.6603   2.804213
#> 84  606.7061   2.680988
#> 85  502.2507   2.822436
#> 86  503.1849   2.514392
#> 87  503.0504   2.503776
#> 88  502.8445   2.567008
#> 89  501.7222   2.977395
#> 90  533.4924   3.006555
#> 91  528.3208   2.875529
#> 92  504.3841   2.924824
#> 93  501.2925   3.603988
#> 94  500.1010   4.791082
#> 95  580.4055   3.044527
#> 96  524.9660   2.670150
#> 97  669.8671   3.167400
#> 98  500.9421   2.654365
#> 99  505.8858   3.303232
#> 100 501.3956   2.601893
#> 101 501.6016   2.651582
#> 102 512.1526   2.520301
#> 103 500.9763   4.196031
#> 104 505.7114   4.197916
#> 105 528.8461   2.635613
#> 106 573.4069   2.683864
#> 107 510.8320   3.397021
#> 108 530.3099   2.675145
#> 109 500.0010   2.746946
#> 110 543.8900   3.140545
#> 111 500.8727   2.889245
#> 112 500.1842   3.414997
#> 113 514.3902   2.807710
#> 114 500.8351   2.673905
#> 115 555.9788   2.608215
#> 116 501.4646   2.774772
#> 117 500.2629   4.176530
#> 118 558.8153   2.584383
#> 119 535.1741   3.496126
#> 120 501.1682   2.840646
#> 121 533.8421   3.892938
#> 122 500.0191   2.593898
#> 123 782.5474   2.894869
#> 124 500.1222   2.715529
#> 125 502.6459   3.493569
#> 126 825.5930   2.869149
#> 127 500.0287   2.728331
#> 128 503.7624   2.792686
#> 129 502.2095   3.080870
#> 130 501.0516   3.042641
#> 131 504.0101   2.709883
#> 132 502.2675   2.722684
#> 133 500.6130   2.512779
#> 134 500.8976   2.946005
#> 135 505.9821   3.263055
#> 136 500.3394   3.147175
#> 137 559.3625   2.650182
#> 138 500.1463   3.166759
#> 139 548.5670   2.629021
#> 140 500.6381   3.279833
#> 141 500.0398   3.151425
#> 142 767.4532   2.791021
#> 143 507.4927   3.134437
#> 144 500.0216   3.246074
#> 145 501.0394   3.375827
#> 146 512.8507   2.524060
#> 147 515.2535   2.513463
#> 148 520.4532   2.653886
#> 149 506.2582   3.290489
#> 150 500.3349   3.170470
#> 151 500.0278   2.505752
#> 152 517.4877   3.720003
#> 153 500.1348   2.598471
#> 154 500.0276   2.685946
#> 155 659.4841   2.531415
#> 156 510.1261   4.370764
#> 157 500.9683   3.000387
#> 158 500.0754   2.885585
#> 159 500.1052   2.522819
#> 160 500.0301   2.641143
#> 161 505.6429   2.596978
#> 162 501.3058   2.640083
#> 163 500.3188   3.579338
#> 164 505.0214   2.927421
#> 165 501.6409   2.642125
#> 166 545.2093   2.790072
#> 167 501.1129   3.678486
#> 168 504.4594   2.659637
#> 169 614.8665   2.535890
#> 170 505.1409   3.381379
#> 171 500.4680   2.821042
#> 172 537.3873   2.789968
#> 173 663.9946   2.783401
#> 174 500.9842   3.074521
#> 175 555.2262   2.552242
#> 176 500.3567   2.727314
#> 177 631.4402   2.744438
#> 178 500.5700   2.558039
#> 179 652.3525   2.602019
#> 180 557.9316   3.265485
#> 181 511.6310   2.513590
#> 182 500.7668   2.761609
#> 183 500.1234   2.668711
#> 184 694.6683   2.949394
#> 185 502.3932   2.805321
#> 186 501.1148   4.086533
#> 187 500.0001   2.878556
#> 188 508.3628   3.664765
#> 189 501.1923   2.584580
#> 190 500.7803   2.963428
#> 191 508.4099   2.619216
#> 192 518.8617   2.828792
#> 193 500.5105   2.794378
#> 194 500.2339   2.805105
#> 195 506.6417   3.284164
#> 196 732.0894   2.516009
#> 197 679.5328   2.894585
#> 198 501.9870   3.044577
#> 199 542.2158   3.121209
#> 200 500.0147   2.542853
#> 201 500.1006   2.854054
#> 202 500.6864   2.537691
#> 203 501.3760   2.639113
#> 204 500.1167   2.786361
#> 205 616.0025   3.246045
#> 206 513.0467   2.968199
#> 207 532.5268   2.668465
#> 208 583.6977   3.130468
#> 209 507.4676   3.305213
#> 210 662.3330   3.184969
#> 211 500.1281   2.639910
#> 212 500.1383   3.309817
#> 213 500.2843   3.122183
#> 214 867.1509   3.599819
#> 215 514.1796   2.603193
#> 216 508.0307   4.027096
#> 217 500.1276   2.647361
#> 218 500.1060   2.768695
#> 219 502.5898   2.914156
#> 220 500.9456   2.930728
#> 221 500.3141   2.644927
#> 222 505.2008   2.565071
#> 223 500.3312   2.621679
#> 224 503.1160   3.016931
#> 225 530.2751   2.536279
#> 226 500.2403   2.523579
#> 227 500.1116   2.577780
#> 228 500.1160   2.726454
#> 229 500.0116   3.400227
#> 230 500.1627   2.755501
#> 231 581.6490   2.596734
#> 232 500.0678   2.820357
#> 233 538.3553   3.012101
#> 234 508.9952   2.937329
#> 235 507.8466   2.975355
#> 236 501.5547   3.176948
#> 237 594.6662   2.749666
#> 238 500.4036   3.107014
#> 239 511.8025   2.530009
#> 240 500.1839   2.760344
#> 241 524.9408   2.662000
#> 242 506.9871   4.227344
#> 243 513.8021   3.095380
#> 244 514.1277   3.762344
#> 245 500.2237   2.716281
#> 246 500.5048   2.790784
#> 247 540.5176   2.619761
#> 248 640.5157   2.560956
#> 249 500.0665   2.602712
#> 250 531.6856   3.013276
#> 251 500.0991   2.699419
#> 252 533.8199   3.765065
#> 253 527.6165   3.089674
#> 254 888.0654   3.386988
#> 255 539.9495   3.368944
#> 256 500.0195   2.679099
#> 257 500.0018   3.083573
#> 258 500.1302   3.865274
#> 259 555.1231   2.647853
#> 260 802.9440   3.164762
#> 261 503.6286   2.558232
#> 262 606.0508   3.163143
#> 263 504.2155   2.628419
#> 264 500.1206   2.767258
#> 265 949.5767   2.679647
#> 266 726.0462   2.766087
#> 267 518.3396   3.840322
#> 268 504.7549   2.759200
#> 269 507.5137   3.511770
#> 270 500.1345   2.650178
#> 271 507.3002   2.744110
#> 272 502.5986   2.632813
#> 273 505.5308   3.128970
#> 274 502.6318   2.517600
#> 275 732.8586   3.015013
#> 276 500.7826   3.326632
#> 277 500.5652   3.372412
#> 278 501.7177   2.947805
#> 279 500.1469   2.703141
#> 280 695.4488   3.107942
#> 281 500.3090   3.396710
#> 282 698.8069   2.799241
#> 283 501.2404   2.761338
#> 284 502.8884   2.869440
#> 285 510.1561   2.505110
#> 286 506.3158   3.182351
#> 287 500.4680   2.541541
#> 288 500.2593   2.632121
#> 289 504.4700   2.936612
#> 290 501.8308   2.601354
#> 291 501.4164   2.585346
#> 292 502.9582   2.777931
#> 293 500.2893   2.838890
#> 294 503.7332   2.791753
#> 295 511.8622   2.591959
#> 296 963.1151   2.576940
#> 297 500.2531   2.582560
#> 298 595.5520   3.157028
#> 299 503.3424   2.988656
#> 300 504.6097   2.976655
#> 301 503.6909   3.600015
#> 302 503.9336   2.598602
#> 303 504.3387   2.817803
#> 304 503.3340   2.846331
#> 305 803.3033   2.826712
#> 306 501.8575   2.627245
#> 307 500.0042   2.703872
#> 308 507.5597   2.592906
#> 309 516.1782   2.596849
#> 310 500.7939   2.628481
#> 311 507.0946   2.769410
#> 312 583.9980   2.664413
#> 313 505.8687   3.045729
#> 314 501.8818   2.655375
#> 315 785.3047   2.952406
#> 316 572.9796   3.264323
#> 317 540.6866   3.184339
#> 318 515.7583   3.054490
#> 319 501.9228   2.976469
#> 320 500.0135   3.012040
#> 321 500.0223   2.645545
#> 322 546.4193   2.910187
#> 323 504.8444   3.461883
#> 324 500.2098   2.761439
#> 325 500.0118   2.680528
#> 326 503.7570   2.519878
#> 327 500.3528   3.297780
#> 328 552.2678   3.155608
#> 329 500.1436   3.178658
#> 330 500.9061   2.581194
#> 331 587.3994   2.980006
#> 332 880.4687   3.038115
#> 333 500.5817   2.945109
#> 334 527.6854   3.528961
#> 335 500.2611   3.150962
#> 336 501.1418   2.648290
#> 337 738.2275   2.733115
#> 338 541.9030   2.716398
#> 339 500.1713   3.871324
#> 340 535.2238   3.254169
#> 341 500.0979   2.732839
#> 342 500.3793   2.673921
#> 343 500.6732   2.620740
#> 344 501.1347   2.663817
#> 345 500.3796   3.178437
#> 346 592.8101   2.508466
#> 347 538.6488   2.534120
#> 348 583.7033   2.833635
#> 349 531.5231   3.490721
#> 350 515.6520   2.552739
#> 351 502.1885   2.751140
#> 352 500.0279   4.323621
#> 353 500.0026   2.975935
#> 354 934.2964   2.766744
#> 355 500.4268   2.846398
#> 356 555.0519   3.082082
#> 357 500.3243   2.800984
#> 358 530.4473   4.408055
#> 359 505.3853   3.152621
#> 360 500.1188   2.571141
#> 361 500.1154   3.622945
#> 362 528.7638   3.089532
#> 363 500.0867   2.798642
#> 364 511.5259   3.326186
#> 365 500.3396   3.741614
#> 366 500.0213   3.649909
#> 367 502.1628   3.095456
#> 368 500.0206   2.739276
#> 369 500.4470   2.897931
#> 370 666.8820   2.504089
#> 371 507.9905   2.754783
#> 372 525.9681   2.523094
#> 373 549.8705   2.586731
#> 374 510.0147   3.585681
#> 375 500.5485   3.147825
#> 376 503.1018   3.397384
#> 377 500.9259   3.037802
#> 378 500.1101   2.990091
#> 379 502.0820   3.218463
#> 380 654.5072   2.875460
#> 381 504.4091   2.882664
#> 382 500.3613   3.427371
#> 383 500.8902   3.559972
#> 384 500.3444   2.690354
#> 385 655.6504   2.887086
#> 386 500.2445   2.544727
#> 387 633.6763   2.552166
#> 388 862.4937   2.916333
#> 389 508.7291   3.109837
#> 390 500.8583   3.539269
#> 391 584.3548   2.659092
#> 392 504.6556   4.179165
#> 393 503.7848   3.418620
#> 394 520.7864   2.624526
#> 395 500.4998   3.707625
#> 396 500.4122   3.564046
#> 397 501.0082   2.605641
#> 398 500.9680   2.606498
#> 399 500.1087   3.094812
#> 400 579.0286   2.679257
#> 401 928.1081   2.697914
#> 402 501.2475   3.250301
#> 403 519.8117   2.520757
#> 404 500.2492   2.534026
#> 405 603.8546   3.723206
#> 406 507.4420   2.934069
#> 407 500.1815   2.620468
#> 408 509.9276   2.625726
#> 409 538.4399   3.640175
#> 410 757.9420   3.594331
#> 411 969.9739   2.761951
#> 412 516.1873   3.005229
#> 413 500.1162   2.686954
#> 414 500.6124   2.538494
#> 415 500.2050   4.096858
#> 416 500.1629   2.907793
#> 417 519.8313   3.411856
#> 418 764.3559   2.654646
#> 419 500.5416   2.588943
#> 420 692.6279   3.354279
#> 421 524.3457   3.401308
#> 422 500.0341   4.059656
#> 423 550.4019   3.766727
#> 424 500.3845   2.699088
#> 425 501.7305   4.041322
#> 426 542.4648   2.936539
#> 427 873.1460   2.503969
#> 428 506.9848   3.057042
#> 429 650.1168   2.989072
#> 430 610.2312   3.002232
#> 431 500.1051   2.786408
#> 432 876.7762   3.491288
#> 433 501.2583   2.980290
#> 434 500.3056   2.734048
#> 435 502.6876   2.622339
#> 436 500.0494   3.067777
#> 437 607.9004   3.140177
#> 438 500.1816   2.540115
#> 439 503.1432   2.910691
#> 440 500.3492   3.306428
#> 441 612.1323   4.103632
#> 442 555.9065   2.869525
#> 443 530.8659   3.345599
#> 444 500.6878   5.282994
#> 445 534.0830   3.936352
#> 446 705.2577   2.609792
#> 447 647.2301   2.893402
#> 448 539.4027   3.133764
#> 449 745.0879   2.993854
#> 450 500.3303   2.917637
#> 451 501.1225   3.424057
#> 452 503.0827   2.588013
#> 453 503.0672   2.629422
#> 454 515.0722   3.111340
#> 455 500.4035   3.736845
#> 456 522.1213   4.003621
#> 457 534.9748   2.898950
#> 458 503.6380   3.534760
#> 459 500.9448   2.514280
#> 460 502.5467   3.705587
#> 461 500.1753   2.674826
#> 462 500.1079   2.581986
#> 463 547.3596   2.768827
#> 464 643.9122   2.772602
#> 465 558.3030   2.867324
#> 466 616.1332   2.798592
#> 467 531.9635   2.552693
#> 468 504.7490   2.535741
#> 469 500.7594   2.568383
#> 470 500.3657   2.887234
#> 471 506.3021   2.695665
#> 472 500.5984   2.866881
#> 473 509.3282   3.016340
#> 474 517.2140   2.541861
#> 475 522.0276   2.784430
#> 476 502.0938   3.014359
#> 477 917.6555   2.920129
#> 478 500.0876   2.971867
#> 479 680.3209   3.298326
#> 480 606.3122   3.969613
#> 481 820.3379   2.623701
#> 482 530.8988   2.626710
#> 483 500.7143   2.608055
#> 484 501.2603   2.621129
#> 485 500.3368   2.861634
#> 486 500.5605   2.676442
#> 487 500.0196   2.534365
#> 488 500.3810   3.085214
#> 489 500.3196   2.895523
#> 490 500.6381   3.303415
#> 491 502.3920   4.657420
#> 492 500.0472   2.756040
#> 493 502.4659   2.842945
#> 494 535.9189   2.945517
#> 495 500.3949   3.294762
#> 496 512.8305   3.038300
#> 497 512.9246   2.650169
#> 498 502.1076   2.548780
#> 499 674.4364   2.748622
#> 500 500.0281   3.363525
#> 501 600.3503   2.601068
#> 502 530.2529   3.501664
#> 503 526.7396   2.733735
#> 504 500.1265   2.621623
#> 505 510.2523   3.397736
#> 506 500.1378   2.538371
#> 507 500.5431   2.778184
#> 508 539.7864   3.638974
#> 509 501.8369   2.659806
#> 510 520.6117   3.101051
#> 511 870.2732   2.786511
#> 512 548.0833   2.728831
#> 513 513.8830   2.672879
#> 514 500.2041   3.126429
#> 515 695.0709   5.211174
#> 516 504.6287   2.503688
#> 517 538.9578   2.741310
#> 518 582.6581   2.970252
#> 519 597.1899   2.620153
#> 520 500.0568   2.563017
#> 521 850.2885   3.943892
#> 522 558.6863   2.610049
#> 523 697.8093   2.580396
#> 524 500.9625   2.736955
#> 525 524.9306   2.880725
#> 526 503.4867   2.861115
#> 527 982.4685   3.236294
#> 528 506.3339   3.678375
#> 529 597.0546   2.657477
#> 530 501.2991   2.557778
#> 531 505.7691   3.649706
#> 532 500.0049   2.937715
#> 533 500.3089   2.671515
#> 534 561.9694   3.766382
#> 535 501.1181   2.624737
#> 536 500.4914   3.149307
#> 537 652.5073   2.662295
#> 538 995.5638   3.031786
#> 539 512.3572   2.596111
#> 540 538.4069   2.734139
#> 541 503.5327   2.588724
#> 542 500.0019   2.561410
#> 543 503.3913   2.931656
#> 544 505.1461   2.602717
#> 545 500.6941   4.084295
#> 546 615.9461   3.075430
#> 547 501.6322   2.635808
#> 548 502.0716   3.436618
#> 549 563.9437   2.599429
#> 550 501.6632   3.857406
#> 551 720.9226   2.607752
#> 552 501.3673   2.554593
#> 553 500.8930   3.026814
#> 554 594.5844   2.890212
#> 555 503.8123   3.047261
#> 556 511.3456   3.313940
#> 557 500.4081   2.660785
#> 558 500.0885   2.604995
#> 559 500.3309   2.618692
#> 560 507.3951   2.550600
#> 561 538.0483   3.445396
#> 562 606.2194   2.860146
#> 563 500.3201   2.540016
#> 564 500.0746   2.567626
#> 565 548.3018   2.575399
#> 566 525.7684   2.794895
#> 567 500.4187   2.588980
#> 568 502.6018   2.647218
#> 569 501.7081   2.863857
#> 570 502.4379   4.695988
#> 571 500.0134   3.085257
#> 572 505.1057   2.862886
#> 573 528.0709   3.476493
#> 574 500.0481   2.617676
#> 575 511.6119   2.809243
#> 576 502.2848   2.979811
#> 577 627.6345   2.833526
#> 578 891.2378   2.652970
#> 579 511.2910   3.757773
#> 580 500.0543   2.934933
#> 581 506.0375   2.507074
#> 582 686.3092   2.850629
#> 583 500.0698   2.961039
#> 584 500.0244   2.782163
#> 585 500.9629   3.465373
#> 586 500.3828   2.630902
#> 587 578.2526   2.890620
#> 588 517.9899   2.546541
#> 589 503.3219   2.522906
#> 590 500.0004   3.001307
#> 591 513.0208   3.131312
#> 592 502.1558   2.859335
#> 593 505.5556   3.096552
#> 594 874.9684   2.608092
#> 595 500.2159   3.759111
#> 596 500.1481   3.163679
#> 597 537.8589   2.705028
#> 598 500.1477   3.039094
#> 599 516.1383   2.651983
#> 600 907.2241   2.837245
#> 601 504.1234   3.226149
#> 602 578.5148   3.614208
#> 603 606.6386   2.647808
#> 604 533.1046   2.963244
#> 605 507.0988   2.744611
#> 606 504.4151   2.531877
#> 607 530.1666   2.664619
#> 608 535.3872   2.751196
#> 609 500.0822   2.606186
#> 610 500.1224   2.590513
#> 611 500.7204   2.654261
#> 612 503.2186   2.825423
#> 613 500.6981   3.128750
#> 614 755.4148   2.776039
#> 615 673.0606   2.503357
#> 616 500.7375   2.643966
#> 617 505.9268   3.280609
#> 618 500.1346   3.355409
#> 619 537.1037   2.840845
#> 620 508.8296   2.971661
#> 621 544.0067   3.488411
#> 622 647.3818   3.138299
#> 623 535.8080   2.559749
#> 624 500.3441   3.065885
#> 625 507.0695   2.917640
#> 626 500.1022   2.513565
#> 627 500.1293   2.687272
#> 628 500.1353   2.604283
#> 629 500.2340   3.901918
#> 630 566.3967   2.562556
#> 631 500.3019   2.829258
#> 632 501.0789   2.881474
#> 633 504.1072   2.638876
#> 634 539.5289   2.716677
#> 635 500.5202   2.630675
#> 636 500.0583   2.726349
#> 637 528.4451   2.595013
#> 638 501.0228   2.538874
#> 639 500.1519   3.182280
#> 640 501.4947   3.105828
#> 641 504.6037   2.929073
#> 642 505.4265   2.626581
#> 643 504.2485   2.865224
#> 644 501.4475   2.613915
#> 645 500.0611   2.932367
#> 646 500.0465   3.575477
#> 647 554.9136   3.000912
#> 648 500.3356   2.844345
#> 649 500.4487   2.576778
#> 650 509.0235   2.720048
#> 651 500.1684   2.550638
#> 652 515.6968   2.504803
#> 653 500.0297   2.729021
#> 654 506.1752   3.470166
#> 655 500.5279   2.539181
#> 656 529.7742   2.501021
#> 657 521.6174   3.228098
#> 658 611.1240   2.665662
#> 659 500.1715   2.810783
#> 660 538.2740   3.732215
#> 661 506.0254   3.118053
#> 662 520.0510   3.728664
#> 663 500.0250   2.717288
#> 664 500.1514   2.991986
#> 665 500.1650   2.511238
#> 666 500.4905   2.649148
#> 667 502.1738   2.650287
#> 668 603.2117   2.620843
#> 669 501.4841   2.705914
#> 670 591.1716   2.822965
#> 671 500.0736   2.555945
#> 672 506.7783   2.513313
#> 673 530.0415   2.664527
#> 674 514.3774   2.551503
#> 675 518.0125   2.823702
#> 676 510.7622   2.934058
#> 677 502.1858   2.679520
#> 678 528.4959   2.853016
#> 679 500.5175   2.758450
#> 680 591.6375   2.548514
#> 681 712.9127   4.181946
#> 682 945.4645   3.198603
#> 683 549.7570   2.966053
#> 684 500.0494   3.155769
#> 685 510.5721   3.015856
#> 686 541.6681   2.625247
#> 687 511.4906   3.626027
#> 688 513.6002   2.535237
#> 689 500.3219   2.518281
#> 690 555.1768   2.577339
#> 691 500.5837   2.710796
#> 692 502.7865   2.647907
#> 693 729.9631   2.511596
#> 694 500.1989   2.950930
#> 695 507.6087   2.869325
#> 696 500.0627   2.828196
#> 697 508.1248   3.150948
#> 698 506.6144   2.749849
#> 699 500.9682   2.617797
#> 700 505.2704   2.506178
#> 701 544.8648   2.572445
#> 702 576.8676   2.505186
#> 703 500.1751   3.728442
#> 704 500.7966   3.293741
#> 705 555.2748   3.157260
#> 706 502.4557   3.031397
#> 707 515.3823   3.513929
#> 708 500.4869   2.924747
#> 709 520.9400   2.771137
#> 710 500.3766   2.826280
#> 711 505.6148   3.179054
#> 712 500.2024   3.493720
#> 713 567.6659   3.199465
#> 714 500.0607   2.956441
#> 715 500.0667   2.640671
#> 716 500.0737   2.520954
#> 717 513.4760   2.748502
#> 718 500.0495   3.641890
#> 719 508.3492   2.582479
#> 720 568.5324   2.637924
#> 721 500.4863   2.537041
#> 722 500.0175   2.751290
#> 723 500.2126   2.713648
#> 724 500.3489   2.715006
#> 725 500.3689   2.622125
#> 726 500.4970   2.652009
#> 727 526.7662   2.659480
#> 728 500.3018   3.086690
#> 729 503.3219   2.742898
#> 730 500.0439   3.011666
#> 731 506.8647   2.797901
#> 732 500.0065   2.712328
#> 733 520.0132   3.042857
#> 734 500.4733   3.383701
#> 735 500.5448   3.217261
#> 736 500.2673   2.735011
#> 737 500.3853   2.896032
#> 738 500.0778   2.829508
#> 739 623.9981   2.527128
#> 740 504.5873   3.545092
#> 741 500.8674   2.628710
#> 742 503.9011   3.237272
#> 743 501.5898   2.559720
#> 744 579.5384   2.869584
#> 745 796.2732   2.774638
#> 746 502.6202   2.639581
#> 747 500.0119   3.043066
#> 748 548.8524   2.868129
#> 749 505.2801   3.461214
#> 750 511.9638   4.789858
#> 751 665.3883   2.628637
#> 752 556.4269   2.799166
#> 753 576.4849   3.623141
#> 754 593.2198   3.062482
#> 755 501.2825   3.285197
#> 756 716.0711   3.072827
#> 757 500.6678   3.680627
#> 758 645.6230   3.050568
#> 759 500.1608   3.078394
#> 760 502.5870   2.585404
#> 761 637.7164   2.918033
#> 762 500.5957   3.068224
#> 763 507.1343   3.341561
#> 764 500.2114   4.168368
#> 765 521.0233   2.701546
#> 766 500.0489   2.604976
#> 767 500.0943   3.496193
#> 768 500.4273   2.836766
#> 769 501.5403   2.624963
#> 770 506.0789   2.796027
#> 771 511.1944   3.267172
#> 772 500.0562   2.655638
#> 773 787.9619   2.541928
#> 774 516.8227   2.520218
#> 775 554.5706   3.557927
#> 776 519.0924   3.616056
#> 777 501.6209   2.584854
#> 778 502.8556   2.749540
#> 779 559.2037   2.685375
#> 780 500.7685   2.624080
#> 781 503.5811   3.007081
#> 782 500.1711   2.541488
#> 783 500.0308   2.838463
#> 784 536.8817   2.646962
#> 785 501.7432   2.732087
#> 786 514.1478   2.571192
#> 787 538.7831   2.798714
#> 788 501.3224   2.663088
#> 789 885.7997   3.361121
#> 790 500.1328   2.884585
#> 791 500.8299   2.655567
#> 792 502.2477   2.557986
#> 793 500.2689   3.107975
#> 794 501.2860   2.647216
#> 795 518.8274   2.757526
#> 796 529.8612   3.125903
#> 797 500.7262   3.070013
#> 798 818.9531   2.638259
#> 799 500.0755   3.370972
#> 800 505.1171   2.524256
#> 801 503.6481   2.772028
#> 802 500.0239   2.570545
#> 803 500.0960   2.640972
#> 804 500.2199   2.507196
#> 805 500.1369   3.315238
#> 806 501.6772   2.734982
#> 807 500.1810   2.751084
#> 808 515.3344   2.559057
#> 809 638.0778   2.549115
#> 810 510.8389   3.785391
#> 811 545.0315   2.501070
#> 812 500.6271   2.968507
#> 813 500.2596   2.537154
#> 814 501.3743   3.097236
#> 815 580.3676   2.543581
#> 816 500.3092   2.824999
#> 817 976.3236   2.943207
#> 818 503.1430   3.399693
#> 819 717.2518   2.660582
#> 820 586.6330   2.699776
#> 821 524.7588   2.544519
#> 822 599.7704   2.555466
#> 823 501.5796   2.634364
#> 824 500.0022   2.535048
#> 825 500.4256   2.904515
#> 826 500.6097   2.639702
#> 827 501.1758   2.794688
#> 828 500.6262   2.688776
#> 829 502.6539   4.388122
#> 830 696.5186   2.668667
#> 831 500.8811   3.074217
#> 832 514.7731   3.400116
#> 833 500.0234   2.904099
#> 834 500.4543   3.705684
#> 835 503.3027   2.629094
#> 836 519.1437   2.642216
#> 837 507.8031   2.567916
#> 838 501.3794   3.068411
#> 839 533.3347   2.863669
#> 840 501.3810   2.603115
#> 841 571.4108   3.171449
#> 842 502.4788   2.972106
#> 843 503.5330   3.353725
#> 844 627.4519   2.533370
#> 845 503.3736   3.025433
#> 846 503.9416   3.581930
#> 847 559.4154   2.617115
#> 848 500.5115   3.064371
#> 849 625.5522   3.682527
#> 850 548.4455   2.753394
#> 851 500.5794   4.436886
#> 852 503.1259   2.633595
#> 853 772.1561   3.462915
#> 854 501.6972   3.561278
#> 855 872.2638   2.808111
#> 856 503.7585   2.917457
#> 857 561.9965   2.529643
#> 858 504.8757   2.553790
#> 859 501.7298   2.556520
#> 860 504.9760   2.500528
#> 861 522.4040   3.046721
#> 862 500.5764   2.988526
#> 863 837.5225   2.736191
#> 864 570.9879   2.695359
#> 865 500.6485   2.914261
#> 866 500.4944   2.516587
#> 867 502.4134   2.534862
#> 868 515.1105   2.980395
#> 869 508.6322   2.915990
#> 870 506.8419   3.294901
#> 871 639.8123   3.106050
#> 872 506.8597   3.407828
#> 873 877.9478   4.690096
#> 874 500.4063   2.552943
#> 875 921.4935   3.388646
#> 876 500.0300   3.073168
#> 877 538.8338   3.184559
#> 878 505.0445   2.624276
#> 879 508.9598   2.556684
#> 880 545.6867   2.623950
#> 881 500.3888   3.000901
#> 882 590.3528   3.237873
#> 883 960.3187   3.049291
#> 884 500.2559   2.971223
#> 885 500.2306   2.844812
#> 886 500.4578   2.611880
#> 887 504.1108   2.652025
#> 888 502.6882   2.961015
#> 889 517.0276   2.819104
#> 890 500.0897   3.134451
#> 891 501.1063   2.781537
#> 892 500.0792   2.780238
#> 893 585.9263   3.134408
#> 894 500.6278   2.766034
#> 895 872.0421   2.900411
#> 896 548.7080   2.558343
#> 897 500.0738   3.174522
#> 898 500.1456   3.124321
#> 899 500.0123   3.706657
#> 900 500.0908   3.673728
#> 901 501.5270   3.748762
#> 902 504.3317   2.961976
#> 903 501.1218   2.520277
#> 904 591.0058   3.764267
#> 905 501.8025   2.642467
#> 906 517.2138   2.574932
#> 907 791.6841   3.102214
#> 908 501.5089   3.303793
#> 909 501.1450   3.538431
#> 910 541.5477   2.521536
#> 911 507.9541   3.994062
#> 912 597.5355   2.521915
#> 913 507.0113   2.709653
#> 914 517.4669   2.890307
#> 915 502.2274   2.898651
#> 916 500.5330   2.512348
#> 917 516.2405   2.724455
#> 918 500.8633   2.719997
#> 919 500.3081   3.975422
#> 920 548.8964   3.038367
#> 921 577.1389   2.602207
#> 922 871.7002   3.166825
#> 923 527.3452   2.527119
#> 924 500.0879   2.782241
#> 925 512.3181   5.056815
#> 926 501.5476   2.807158
#> 927 503.8953   3.160033
#> 928 505.5960   2.597385