Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   500.6815   2.921474
#> 2   514.7280   2.501826
#> 3   534.4467   2.590735
#> 4   521.0534   3.496716
#> 5   538.0613   2.552830
#> 6   502.6904   2.502596
#> 7   898.4928   3.312281
#> 8   515.1760   4.462513
#> 9   500.8534   2.753916
#> 10  655.1387   2.833775
#> 11  500.4203   3.148489
#> 12  500.0142   3.982232
#> 13  500.0115   2.581859
#> 14  506.0868   3.046967
#> 15  501.5921   3.229441
#> 16  502.1336   2.716967
#> 17  500.4359   3.019114
#> 18  518.1930   3.471941
#> 19  500.1168   2.663429
#> 20  521.5899   2.704248
#> 21  536.4378   3.203770
#> 22  502.9234   2.996004
#> 23  500.2446   3.098100
#> 24  523.7160   2.645879
#> 25  502.7438   2.564433
#> 26  503.3003   2.537321
#> 27  500.1176   2.614245
#> 28  771.2084   2.729828
#> 29  545.9912   2.793946
#> 30  500.8003   2.832832
#> 31  622.7021   3.353592
#> 32  500.0052   2.977206
#> 33  677.1351   2.761463
#> 34  500.5611   2.691829
#> 35  500.6714   2.906088
#> 36  560.3380   4.014843
#> 37  504.2709   2.594906
#> 38  563.0470   2.705540
#> 39  501.6370   3.267572
#> 40  538.4254   2.827220
#> 41  500.7236   3.425038
#> 42  536.3187   2.731827
#> 43  515.7701   2.607337
#> 44  524.8009   2.500763
#> 45  501.0724   3.542033
#> 46  500.4005   2.558087
#> 47  500.5444   2.502825
#> 48  503.9211   3.431772
#> 49  500.9847   2.519700
#> 50  668.0389   2.536216
#> 51  503.0427   2.849532
#> 52  501.0521   2.703832
#> 53  608.2209   2.645896
#> 54  501.2428   3.103260
#> 55  500.0403   2.673741
#> 56  500.3698   2.509600
#> 57  502.4372   3.565769
#> 58  502.1205   3.694791
#> 59  500.9789   3.658313
#> 60  511.1665   3.525189
#> 61  500.4379   3.016955
#> 62  500.0103   3.054233
#> 63  500.0922   2.764953
#> 64  500.2790   2.603165
#> 65  828.3634   2.783765
#> 66  503.2009   2.688007
#> 67  516.1793   2.501359
#> 68  504.0891   2.795993
#> 69  500.2082   2.829348
#> 70  502.5891   2.804603
#> 71  502.7946   2.874665
#> 72  653.4600   3.104989
#> 73  500.4506   2.975621
#> 74  713.1125   2.887349
#> 75  501.1132   2.852784
#> 76  561.7052   2.583526
#> 77  500.0249   2.620956
#> 78  555.4298   2.565926
#> 79  503.1404   2.690295
#> 80  500.7103   2.534350
#> 81  500.3815   2.675409
#> 82  536.0221   3.685873
#> 83  537.1483   2.620293
#> 84  500.6632   3.482123
#> 85  500.2331   2.640815
#> 86  575.3068   3.639952
#> 87  500.3485   3.178555
#> 88  500.1322   3.069563
#> 89  504.5022   2.879631
#> 90  500.7628   3.147392
#> 91  535.0693   3.229414
#> 92  500.3254   2.713429
#> 93  500.0320   2.721440
#> 94  500.7650   2.810659
#> 95  637.4107   3.039809
#> 96  782.5138   3.082015
#> 97  500.0149   2.635602
#> 98  500.2939   2.634537
#> 99  501.1625   2.914389
#> 100 500.0658   3.526511
#> 101 507.4710   3.093360
#> 102 751.4487   3.030807
#> 103 501.0336   2.709919
#> 104 500.0757   2.638795
#> 105 559.9542   3.039330
#> 106 521.4491   3.006095
#> 107 502.3141   3.104896
#> 108 846.7674   2.622222
#> 109 501.6804   3.053238
#> 110 522.8720   3.298920
#> 111 502.5685   3.738651
#> 112 502.9157   2.890755
#> 113 515.5078   2.547809
#> 114 500.0142   2.879600
#> 115 500.7718   2.888969
#> 116 500.2015   3.132746
#> 117 502.2312   2.696508
#> 118 503.7029   3.935497
#> 119 500.2025   4.311541
#> 120 585.7973   3.070901
#> 121 500.2271   3.215256
#> 122 540.0272   2.767602
#> 123 510.2964   2.650764
#> 124 836.6378   3.206856
#> 125 884.5846   2.878449
#> 126 586.8884   2.517006
#> 127 569.0081   3.184169
#> 128 501.5405   2.712600
#> 129 530.0971   4.186403
#> 130 500.5209   3.315630
#> 131 500.8014   2.504785
#> 132 805.1998   2.866220
#> 133 501.0003   3.633097
#> 134 570.6142   3.064930
#> 135 500.0225   2.509176
#> 136 500.0164   3.330341
#> 137 500.3116   2.582873
#> 138 500.6398   2.641386
#> 139 643.6776   2.543537
#> 140 504.3301   3.475729
#> 141 500.5898   2.516479
#> 142 505.1651   3.071877
#> 143 501.5265   2.596342
#> 144 503.0838   4.589230
#> 145 501.1585   2.670363
#> 146 509.5947   2.686351
#> 147 533.8567   2.732791
#> 148 501.7292   2.631093
#> 149 500.9639   2.545389
#> 150 501.5075   3.015435
#> 151 500.3469   3.905692
#> 152 500.6020   3.584794
#> 153 541.8376   3.317952
#> 154 511.2669   2.629487
#> 155 500.0847   3.326489
#> 156 501.7298   2.591680
#> 157 500.3124   2.545759
#> 158 505.5957   2.665303
#> 159 500.9483   2.984886
#> 160 500.2257   2.594963
#> 161 500.2948   4.545400
#> 162 500.3427   2.745385
#> 163 500.6227   3.518063
#> 164 501.3773   2.694065
#> 165 540.5933   3.100106
#> 166 504.6006   2.779658
#> 167 501.1229   3.208299
#> 168 559.2974   2.541475
#> 169 500.2368   2.553048
#> 170 505.3906   3.134213
#> 171 500.0282   3.600208
#> 172 764.1974   2.544073
#> 173 503.6855   2.570412
#> 174 510.2718   2.669327
#> 175 500.2470   2.640981
#> 176 522.0902   2.632670
#> 177 502.6634   2.813817
#> 178 500.5619   3.250888
#> 179 517.5180   2.822997
#> 180 503.1283   2.592281
#> 181 579.9588   2.739661
#> 182 501.1170   3.179278
#> 183 500.0314   2.822590
#> 184 502.9624   3.723353
#> 185 500.1375   2.519545
#> 186 531.6250   2.561783
#> 187 501.0224   2.604433
#> 188 500.1158   3.079196
#> 189 531.3997   2.589724
#> 190 500.1551   2.715243
#> 191 677.7505   2.511877
#> 192 500.6717   2.701257
#> 193 600.2906   3.056371
#> 194 539.3918   2.964822
#> 195 503.4144   2.625267
#> 196 500.1403   2.896362
#> 197 559.0184   3.481495
#> 198 500.0086   2.585255
#> 199 501.0789   3.427137
#> 200 647.3129   2.615669
#> 201 583.9105   2.909274
#> 202 879.4629   2.532475
#> 203 500.0423   2.833695
#> 204 500.0206   2.548082
#> 205 501.0856   2.527087
#> 206 503.6021   2.583219
#> 207 500.4443   2.912744
#> 208 509.5023   2.814739
#> 209 500.0190   2.988996
#> 210 500.1056   2.847988
#> 211 500.0514   2.699795
#> 212 754.2124   3.029775
#> 213 536.4095   3.541945
#> 214 500.0257   2.778990
#> 215 500.0178   2.721106
#> 216 885.7425   2.650298
#> 217 500.1951   2.544796
#> 218 577.1195   3.229606
#> 219 587.6650   2.610871
#> 220 503.3326   2.922156
#> 221 500.3186   4.027898
#> 222 967.1651   2.876243
#> 223 704.2354   2.726179
#> 224 624.4241   2.631181
#> 225 539.6552   2.509361
#> 226 614.2507   4.307649
#> 227 558.8269   3.705460
#> 228 500.0262   4.581959
#> 229 500.1444   2.614211
#> 230 500.3688   3.098993
#> 231 506.2129   4.120681
#> 232 662.4010   3.469492
#> 233 500.4740   3.514622
#> 234 501.3266   2.695676
#> 235 502.1570   2.745894
#> 236 508.2048   2.703079
#> 237 696.2121   2.504329
#> 238 655.3025   3.662605
#> 239 500.0056   2.554170
#> 240 680.7778   2.667507
#> 241 500.1342   2.644506
#> 242 677.2904   2.679563
#> 243 502.1934   2.609359
#> 244 500.0203   2.860726
#> 245 546.1614   2.995289
#> 246 518.8884   2.710805
#> 247 500.1873   2.592385
#> 248 500.2570   2.557698
#> 249 882.5787   2.782384
#> 250 500.2916   2.789128
#> 251 565.4549   2.544908
#> 252 502.9840   2.650241
#> 253 506.3909   3.187637
#> 254 505.3909   3.267777
#> 255 501.3473   2.841704
#> 256 630.1625   2.713354
#> 257 501.3826   2.905831
#> 258 507.2719   2.508611
#> 259 597.6465   2.642292
#> 260 500.6723   2.974361
#> 261 503.2923   2.555597
#> 262 672.9516   2.586897
#> 263 500.0213   2.693291
#> 264 507.8387   2.632669
#> 265 500.1369   2.531065
#> 266 502.5348   4.265240
#> 267 689.8562   3.528450
#> 268 500.0322   3.367197
#> 269 501.5448   2.633562
#> 270 815.4179   2.914242
#> 271 646.4991   2.971769
#> 272 506.3485   3.150917
#> 273 546.1793   2.822518
#> 274 500.5791   3.328841
#> 275 500.0006   2.551673
#> 276 500.5768   3.529843
#> 277 501.2998   2.602738
#> 278 526.0545   3.343013
#> 279 501.9574   3.057770
#> 280 650.4892   2.709876
#> 281 503.8841   2.564157
#> 282 502.4309   2.516678
#> 283 907.4377   3.368649
#> 284 505.3051   2.723061
#> 285 500.0440   3.571903
#> 286 500.2043   2.961465
#> 287 503.0392   2.816416
#> 288 525.0224   2.712500
#> 289 503.2788   2.844355
#> 290 500.0706   2.675394
#> 291 500.4048   2.762537
#> 292 500.1191   3.115388
#> 293 500.8143   3.405020
#> 294 511.2897   2.527414
#> 295 500.2495   2.876334
#> 296 500.0594   3.276980
#> 297 647.2696   3.167954
#> 298 502.4725   2.636064
#> 299 500.6979   2.684773
#> 300 500.1513   2.807564
#> 301 510.8182   2.578000
#> 302 696.6372   2.723620
#> 303 504.3411   2.516187
#> 304 501.4041   2.682421
#> 305 906.6034   2.685244
#> 306 500.0496   3.098638
#> 307 500.0844   2.735105
#> 308 500.2480   3.086204
#> 309 523.9081   2.565900
#> 310 508.8048   2.682469
#> 311 683.8439   2.554710
#> 312 500.0059   2.896010
#> 313 532.9238   2.535512
#> 314 667.6950   3.040515
#> 315 509.3633   2.882644
#> 316 500.0121   2.978968
#> 317 504.2004   2.524069
#> 318 503.7599   3.111105
#> 319 500.0151   3.369083
#> 320 500.2492   2.708185
#> 321 500.5227   2.657720
#> 322 501.2781   2.529907
#> 323 500.0164   2.634735
#> 324 705.4421   2.718307
#> 325 598.3121   2.879965
#> 326 518.4708   2.750401
#> 327 500.0409   2.760525
#> 328 531.9650   2.507865
#> 329 896.8133   2.566771
#> 330 500.2816   2.614642
#> 331 500.6887   2.536900
#> 332 556.7138   2.515512
#> 333 500.3328   2.986161
#> 334 747.6446   2.987316
#> 335 576.5884   2.566221
#> 336 501.2385   2.515094
#> 337 510.1869   2.850136
#> 338 500.1024   2.520754
#> 339 521.5799   2.547836
#> 340 536.9088   2.863650
#> 341 510.7413   3.661857
#> 342 500.8137   2.509242
#> 343 659.2392   2.819469
#> 344 507.7779   2.909450
#> 345 546.2750   3.566313
#> 346 501.3251   2.794982
#> 347 500.2062   2.554285
#> 348 504.1223   3.153091
#> 349 500.4118   2.514687
#> 350 529.1346   2.531306
#> 351 507.2322   2.626266
#> 352 535.8935   2.676662
#> 353 501.8903   2.850166
#> 354 500.5181   2.968284
#> 355 500.1626   3.767828
#> 356 528.5282   4.209585
#> 357 586.6867   2.687987
#> 358 500.1450   2.894808
#> 359 536.5182   3.076448
#> 360 548.4997   3.308654
#> 361 500.3293   2.551557
#> 362 570.2201   2.727774
#> 363 500.2932   4.189227
#> 364 500.1320   2.864979
#> 365 647.5530   2.714676
#> 366 501.2561   2.786352
#> 367 639.5813   2.860284
#> 368 500.5166   2.667392
#> 369 568.7459   2.664232
#> 370 504.6247   2.801208
#> 371 500.0738   2.782080
#> 372 508.6022   2.935242
#> 373 500.0929   2.977518
#> 374 884.2949   4.082337
#> 375 500.0233   4.063983
#> 376 982.9187   2.544630
#> 377 502.9394   3.158340
#> 378 508.6122   3.022515
#> 379 500.4088   6.407124
#> 380 500.0656   2.766395
#> 381 500.0086   2.739667
#> 382 541.9766   2.861018
#> 383 885.7372   3.566320
#> 384 503.8778   3.182208
#> 385 880.6256   2.549922
#> 386 500.4531   2.706443
#> 387 500.1437   3.627108
#> 388 501.9960   2.504349
#> 389 535.9160   3.004127
#> 390 563.0141   2.589384
#> 391 515.0442   2.877976
#> 392 590.4227   2.752207
#> 393 500.6847   3.017281
#> 394 538.2864   2.739519
#> 395 500.0091   2.588880
#> 396 509.7963   2.593585
#> 397 507.7223   3.352639
#> 398 820.5065   2.756810
#> 399 522.9391   2.549358
#> 400 520.4602   2.769126
#> 401 671.9885   2.534753
#> 402 500.4933   3.208510
#> 403 501.6832   2.704624
#> 404 509.3120   2.584280
#> 405 500.3779   4.396727
#> 406 501.0673   3.227536
#> 407 501.2092   2.796249
#> 408 526.6004   4.031074
#> 409 529.0396   3.266141
#> 410 500.1574   3.064598
#> 411 522.2897   2.727457
#> 412 500.6891   3.074205
#> 413 508.5144   3.054008
#> 414 500.2714   3.445746
#> 415 500.0229   2.557307
#> 416 504.2033   3.234194
#> 417 691.2174   2.622720
#> 418 523.1036   2.662070
#> 419 500.3921   2.708965
#> 420 500.5305   2.539700
#> 421 535.9706   3.021939
#> 422 728.9232   2.866384
#> 423 505.5721   2.537754
#> 424 675.3854   3.502144
#> 425 504.4675   2.506510
#> 426 526.5520   2.556706
#> 427 508.0108   2.532045
#> 428 500.5556   4.048772
#> 429 500.1415   2.535391
#> 430 501.9951   2.940325
#> 431 500.0458   2.718163
#> 432 500.4171   2.818280
#> 433 501.4875   2.882431
#> 434 510.7982   3.161749
#> 435 504.5113   3.102837
#> 436 579.9463   2.675512
#> 437 550.5171   4.268309
#> 438 500.0250   2.721533
#> 439 539.9365   2.526355
#> 440 508.4662   2.509325
#> 441 500.5593   2.966946
#> 442 500.2014   3.040396
#> 443 504.5034   2.663703
#> 444 561.9013   3.247340
#> 445 500.8785   2.541240
#> 446 582.8244   2.958420
#> 447 741.2623   2.897942
#> 448 513.0507   3.009636
#> 449 503.0564   3.735706
#> 450 538.2556   2.683657
#> 451 547.4150   3.040796
#> 452 502.7343   3.106129
#> 453 500.0722   2.594186
#> 454 542.1772   3.318613
#> 455 522.3104   2.629176
#> 456 500.0176   2.871420
#> 457 501.6098   2.596086
#> 458 500.5326   3.228339
#> 459 506.9134   3.098199
#> 460 500.2452   2.783973
#> 461 500.3881   3.512311
#> 462 501.0714   4.026255
#> 463 532.3333   2.832634
#> 464 500.0958   2.667036
#> 465 662.7790   3.089476
#> 466 509.7555   2.883621
#> 467 797.6303   2.803572
#> 468 508.5514   2.676861
#> 469 513.1354   2.746638
#> 470 500.2545   2.597983
#> 471 643.6257   2.500838
#> 472 894.1115   2.884842
#> 473 503.9932   3.066673
#> 474 521.2167   3.113621
#> 475 501.0020   2.977955
#> 476 501.5206   2.739961
#> 477 500.4916   2.791074
#> 478 543.4933   2.500372
#> 479 503.5354   3.062373
#> 480 501.3854   3.546318
#> 481 503.6821   2.505492
#> 482 501.3695   2.622825
#> 483 500.3040   3.669996
#> 484 567.4974   2.792044
#> 485 513.7372   2.593763
#> 486 509.2961   4.752530
#> 487 500.8490   2.812564
#> 488 876.9377   2.592775
#> 489 500.9633   2.587087
#> 490 677.2884   2.773300
#> 491 810.5014   2.957841
#> 492 500.3703   2.554920
#> 493 500.6772   2.781313
#> 494 531.1165   2.736990
#> 495 502.1957   2.801763
#> 496 500.0075   3.052222
#> 497 509.4767   2.736233
#> 498 546.0015   3.814956
#> 499 524.9258   3.013505
#> 500 511.9490   2.854272
#> 501 500.3366   3.020474
#> 502 920.7515   2.660418
#> 503 509.2588   2.580681
#> 504 507.9113   3.233162
#> 505 500.0040   4.859557
#> 506 764.5310   2.599485
#> 507 502.5664   3.047350
#> 508 509.5866   3.780633
#> 509 526.3249   3.898007
#> 510 508.1539   4.112351
#> 511 500.0742   2.751630
#> 512 500.0205   2.982671
#> 513 503.1401   3.806370
#> 514 502.9250   2.689811
#> 515 500.0788   2.867813
#> 516 583.6207   2.732010
#> 517 501.3326   2.551768
#> 518 518.8858   2.530035
#> 519 532.9022   2.749564
#> 520 508.3183   2.715783
#> 521 504.8948   2.720036
#> 522 657.9168   3.908637
#> 523 501.4622   2.545343
#> 524 555.2574   3.602887
#> 525 516.5469   2.506940
#> 526 500.2675   2.659394
#> 527 524.4975   2.764186
#> 528 589.0034   2.561285
#> 529 545.3979   2.991394
#> 530 770.1378   2.785849
#> 531 502.9541   2.504410
#> 532 500.0850   3.342885
#> 533 500.2007   4.117333
#> 534 705.6954   2.773940
#> 535 500.9166   3.488933
#> 536 515.1794   2.696944
#> 537 501.0620   2.517779
#> 538 500.5344   3.104579
#> 539 611.1638   3.937359
#> 540 595.4227   2.821227
#> 541 500.9578   2.714472
#> 542 510.9425   2.558683
#> 543 500.2024   3.657548
#> 544 500.3647   3.276326
#> 545 540.5218   2.867964
#> 546 544.4847   2.632197
#> 547 503.8647   2.589872
#> 548 500.2145   2.931679
#> 549 500.0132   2.914613
#> 550 502.0630   2.616496
#> 551 512.2340   2.759979
#> 552 534.3416   3.481875
#> 553 544.5017   2.623415
#> 554 500.2428   3.319626
#> 555 500.0393   2.819679
#> 556 500.2136   2.596968
#> 557 503.9399   3.464861
#> 558 500.0121   2.782182
#> 559 502.8309   4.348380
#> 560 501.3793   3.273608
#> 561 500.4836   3.025255
#> 562 502.0269   2.803564
#> 563 500.4534   2.660502
#> 564 500.0213   3.833355
#> 565 500.0529   2.906117
#> 566 500.7175   2.607223
#> 567 501.0155   2.751859
#> 568 579.0235   2.633161
#> 569 500.1113   3.091352
#> 570 501.5962   2.604867
#> 571 516.6633   2.792854
#> 572 500.2107   2.501877
#> 573 500.9854   2.838455
#> 574 500.5357   3.374616
#> 575 507.0403   2.500354
#> 576 938.9211   2.696918
#> 577 505.0226   2.851393
#> 578 500.8081   2.897479
#> 579 500.1352   3.101298
#> 580 697.0752   2.917948
#> 581 530.2008   3.027365
#> 582 527.8374   2.728039
#> 583 622.9851   3.114453
#> 584 510.4043   2.750902
#> 585 505.1900   2.881034
#> 586 501.0135   2.543644
#> 587 502.1533   3.805883
#> 588 511.5707   3.180775
#> 589 502.4217   2.790908
#> 590 503.5258   3.331578
#> 591 500.4166   2.793309
#> 592 501.5442   2.697461
#> 593 500.0054   2.577959
#> 594 500.0220   3.097571
#> 595 500.0673   2.543109
#> 596 504.6083   2.961099
#> 597 500.3480   3.177900
#> 598 500.0125   4.017028
#> 599 606.0674   2.992009
#> 600 540.9439   3.176524
#> 601 505.5796   2.681407
#> 602 517.8103   2.719132
#> 603 508.5893   2.883809
#> 604 500.3068   3.814135
#> 605 503.3440   2.601232
#> 606 505.7128   2.833406
#> 607 501.8262   2.540299
#> 608 558.4718   2.946756
#> 609 500.1796   3.992019
#> 610 519.1677   2.677363
#> 611 500.8627   2.709107
#> 612 513.6543   2.666667
#> 613 504.2890   3.292830
#> 614 824.7491   2.818601
#> 615 500.1250   2.731777
#> 616 897.0359   2.663851
#> 617 500.2704   3.445462
#> 618 532.4992   2.766224
#> 619 723.3869   2.614892
#> 620 505.6368   2.651000
#> 621 509.4128   2.729588
#> 622 507.1030   2.522049
#> 623 504.4987   2.519514
#> 624 519.3580   3.292397
#> 625 500.4916   2.709861
#> 626 500.2166   2.578002
#> 627 500.3605   2.636054
#> 628 500.4377   2.780327
#> 629 502.3629   2.602583
#> 630 500.3136   2.803169
#> 631 500.8364   2.524266
#> 632 501.6961   2.573766
#> 633 500.3170   3.037998
#> 634 502.6031   2.801521
#> 635 509.2897   2.500471
#> 636 777.0879   2.530329
#> 637 840.3221   2.594822
#> 638 524.4772   2.745136
#> 639 500.0197   2.907226
#> 640 500.0718   2.592943
#> 641 502.5676   2.598555
#> 642 989.1249   2.987265
#> 643 514.2915   2.599107
#> 644 623.3053   2.778576
#> 645 522.3359   2.792173
#> 646 503.9171   2.920666
#> 647 528.1427   3.078347
#> 648 507.2935   2.852438
#> 649 506.5114   2.738430
#> 650 525.0267   3.558977
#> 651 595.9809   2.523366
#> 652 507.7532   3.487611
#> 653 500.6924   2.733066
#> 654 500.0107   2.506193
#> 655 828.5272   3.290323
#> 656 815.6785   2.525641
#> 657 501.3730   3.752445
#> 658 548.8907   2.506620
#> 659 500.4198   2.920752
#> 660 589.9524   2.578569
#> 661 501.7774   3.384503
#> 662 698.1934   3.152163
#> 663 504.3135   3.032013
#> 664 501.2626   2.539797
#> 665 600.3870   2.915548
#> 666 500.2674   2.660204
#> 667 532.3674   2.703167
#> 668 500.0724   2.533211
#> 669 793.5196   2.678039
#> 670 507.1008   2.995872
#> 671 723.3029   2.541989
#> 672 732.9262   2.969254
#> 673 500.2749   2.693794
#> 674 500.2778   2.643919
#> 675 504.2671   2.967050
#> 676 500.1990   2.556194
#> 677 511.9812   2.587016
#> 678 625.3903   5.262507
#> 679 500.4035   2.669141
#> 680 500.3886   3.662836
#> 681 500.0427   2.661233
#> 682 500.8235   2.984606
#> 683 508.1700   3.261409
#> 684 551.5775   2.906421
#> 685 501.0505   2.797714
#> 686 500.2567   2.704359
#> 687 500.0260   3.287380
#> 688 505.3011   2.702811
#> 689 871.4061   3.922756
#> 690 519.4253   2.788528
#> 691 505.3498   3.377708
#> 692 500.2863   2.638726
#> 693 501.9645   2.523563
#> 694 514.9445   2.664053
#> 695 512.5044   2.597448
#> 696 500.4285   3.016434
#> 697 501.4112   2.616134
#> 698 505.9377   2.807819
#> 699 516.4581   3.642873
#> 700 500.8341   2.896723
#> 701 500.9813   2.579831
#> 702 500.2121   3.173885
#> 703 503.5135   2.959308
#> 704 501.5096   4.656340
#> 705 521.3010   2.740944
#> 706 521.4594   2.504836
#> 707 501.1719   4.303717
#> 708 500.0092   3.063759
#> 709 500.4464   2.724537
#> 710 501.2734   2.656332
#> 711 513.5319   2.804024
#> 712 862.6804   2.619614
#> 713 500.0446   2.554850
#> 714 744.7298   2.909690
#> 715 500.2575   2.954399
#> 716 536.7448   3.324787
#> 717 574.5497   2.598709
#> 718 500.0024   2.619691
#> 719 500.7226   3.225452
#> 720 501.6393   2.798339
#> 721 521.6105   2.516920
#> 722 513.6829   2.756598
#> 723 501.0515   2.526299
#> 724 615.7172   2.591654
#> 725 500.6462   3.044013
#> 726 500.6104   2.795978
#> 727 834.1640   2.635888
#> 728 583.9257   2.975740
#> 729 797.0606   2.811561
#> 730 519.4723   2.962358
#> 731 503.8135   2.637425
#> 732 556.6323   3.040003
#> 733 500.1993   2.549898
#> 734 500.3495   2.813742
#> 735 500.0745   2.770313
#> 736 521.1440   2.705822
#> 737 505.1602   3.167232
#> 738 502.4097   3.479939
#> 739 500.3120   2.629364
#> 740 612.2997   3.095766
#> 741 575.6512   2.699455
#> 742 502.3588   2.732241
#> 743 500.1970   2.810612
#> 744 513.9211   3.664859
#> 745 504.0050   2.543816
#> 746 504.2829   3.070292
#> 747 517.3262   3.005065
#> 748 500.1061   2.563484
#> 749 500.9074   2.542473
#> 750 501.5873   2.647096
#> 751 508.4211   4.138860
#> 752 683.1295   2.735097
#> 753 631.1736   2.628219
#> 754 500.5503   2.567429
#> 755 596.3217   2.805037
#> 756 500.1034   2.549062
#> 757 500.0675   2.793720
#> 758 500.0278   3.032880
#> 759 502.6623   2.524838
#> 760 501.7815   2.783921
#> 761 500.8003   2.992951
#> 762 662.4116   2.791183
#> 763 722.8231   2.707270
#> 764 525.8826   3.163450
#> 765 633.3237   2.749479
#> 766 502.4446   2.716734
#> 767 500.9611   2.761115
#> 768 501.7228   3.118981
#> 769 505.8755   3.003973
#> 770 647.0687   2.769083
#> 771 538.9106   2.827206
#> 772 502.1892   2.671317
#> 773 514.3559   2.519186
#> 774 684.9765   2.797813
#> 775 500.2914   2.620438
#> 776 503.8064   2.680459
#> 777 812.4751   2.683507
#> 778 500.0103   2.754294
#> 779 500.5498   2.511526
#> 780 521.3312   3.365148
#> 781 704.9577   2.865640
#> 782 502.2213   3.001452
#> 783 821.7307   2.654739
#> 784 500.7414   3.235767
#> 785 500.1135   4.152979
#> 786 531.3124   2.782437
#> 787 500.3552   2.739369
#> 788 689.0804   2.594561
#> 789 501.4744   2.504261
#> 790 502.6212   2.778516
#> 791 501.1025   3.074841
#> 792 504.1361   2.777848
#> 793 501.3045   2.618838
#> 794 500.1240   2.507051
#> 795 500.0179   2.931611
#> 796 543.8635   2.504138
#> 797 500.3859   2.623096
#> 798 500.0278   2.509725
#> 799 501.6354   3.525821
#> 800 505.9483   3.311313
#> 801 709.1310   2.577349
#> 802 941.6546   4.386597
#> 803 543.6980   2.734309
#> 804 504.7898   3.185942
#> 805 588.1292   2.833561
#> 806 685.4816   2.884233
#> 807 505.9898   2.593588
#> 808 503.3795   3.629459
#> 809 500.0964   3.036387
#> 810 963.1463   2.790202
#> 811 501.0474   2.509231
#> 812 500.3298   3.534956
#> 813 500.2851   2.956798
#> 814 504.9829   2.768180
#> 815 546.1181   3.603122
#> 816 501.9877   2.502387
#> 817 515.3976   3.013806
#> 818 501.2512   3.198253
#> 819 613.8407   4.011163
#> 820 572.8095   2.511507
#> 821 537.0395   2.944911
#> 822 500.5113   2.539259
#> 823 659.0363   2.916555
#> 824 501.7443   2.547175
#> 825 582.8288   2.636024
#> 826 500.4600   2.795452
#> 827 518.4650   3.816609
#> 828 500.0964   2.539220
#> 829 528.2654   2.657523
#> 830 521.0394   2.596787
#> 831 500.5121   2.507026
#> 832 500.0625   2.580378
#> 833 540.6686   3.863431
#> 834 500.3078   2.879377
#> 835 502.6832   2.708688
#> 836 500.3094   2.990180
#> 837 508.4077   2.655536
#> 838 515.1993   2.752934
#> 839 538.2338   4.390298
#> 840 648.4376   2.943614
#> 841 500.2354   2.646890
#> 842 541.9823   2.831603
#> 843 507.2693   2.643305
#> 844 500.0211   2.566277
#> 845 626.7766   3.139088
#> 846 503.7847   2.548564
#> 847 502.2056   2.926159
#> 848 500.9620   2.635731
#> 849 501.3178   2.505339
#> 850 502.8878   2.668682
#> 851 521.0993   2.730821
#> 852 883.4330   2.529878
#> 853 500.4819   3.470170
#> 854 502.4002   2.989677
#> 855 596.1047   2.695646
#> 856 500.4883   2.815261
#> 857 500.0598   2.849397
#> 858 503.4964   3.288945
#> 859 500.4725   3.394218
#> 860 500.3317   2.803567
#> 861 501.0454   2.682441
#> 862 513.4129   2.905877
#> 863 506.1414   2.721641
#> 864 502.8008   2.697203
#> 865 500.2197   3.257713
#> 866 503.9301   2.563200
#> 867 505.7468   2.531009
#> 868 522.1427   2.529724
#> 869 500.0869   3.376125
#> 870 500.3135   2.779308
#> 871 500.0263   3.682991
#> 872 500.4582   2.737815
#> 873 647.6850   2.513744
#> 874 500.2929   3.359791
#> 875 500.0606   2.593913
#> 876 725.8158   2.714771
#> 877 528.5011   2.628142
#> 878 714.8813   2.696795
#> 879 500.0303   2.985981
#> 880 500.5012   3.972657
#> 881 500.3190   3.385520
#> 882 779.9610   2.844320
#> 883 500.0580   3.423148
#> 884 500.9155   2.977307
#> 885 502.0677   2.926353
#> 886 613.1620   2.576626
#> 887 500.3759   2.523926
#> 888 502.3372   3.082699
#> 889 500.5502   3.125629
#> 890 500.0209   2.754154
#> 891 517.2588   3.458853
#> 892 531.6585   2.520458
#> 893 500.3789   2.554165
#> 894 790.8629   2.697702
#> 895 503.0146   3.112148
#> 896 507.4852   2.796223
#> 897 500.1289   3.421127
#> 898 959.2616   2.568989
#> 899 500.8925   2.505132
#> 900 504.0159   2.536314
#> 901 500.0007   2.646323
#> 902 514.5017   2.533903
#> 903 500.1316   2.584848
#> 904 500.1025   2.611018
#> 905 500.1823   3.331289
#> 906 527.6574   2.502258
#> 907 502.2690   2.526196
#> 908 524.2783   3.703332
#> 909 500.4818   2.544896
#> 910 500.8591   3.153065
#> 911 505.9764   3.260041
#> 912 506.7878   3.884094
#> 913 504.0571   2.520761
#> 914 639.9223   2.919756
#> 915 500.1852   2.582376
#> 916 501.2262   2.628772
#> 917 502.6683   2.788113
#> 918 500.0790   2.883945
#> 919 500.4997   3.222794
#> 920 500.0163   2.756772
#> 921 845.1128   2.943561
#> 922 500.8901   2.576908
#> 923 500.2819   3.484085
#> 924 531.5078   2.605445
#> 925 500.0686   2.547929
#> 926 500.0236   2.576108