Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = NULL)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

Deprecated argument for controlling parallelism. Use future::plan(future::multisession, workers = ncore) (or similar) to configure parallelism in your code instead.

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   673.9341   2.532985
#> 2   515.9706   2.547470
#> 3   625.4076   2.584586
#> 4   828.7094   2.786686
#> 5   921.8883   4.360517
#> 6   500.5331   2.521237
#> 7   539.0931   2.806188
#> 8   500.0026   3.522394
#> 9   500.2553   3.065051
#> 10  506.4746   3.202831
#> 11  527.6555   2.554095
#> 12  500.1372   2.828518
#> 13  500.9616   2.827710
#> 14  747.2843   2.696378
#> 15  500.4959   4.080056
#> 16  500.3166   3.124068
#> 17  502.6676   2.505862
#> 18  501.2085   2.558536
#> 19  504.6805   2.517142
#> 20  528.3704   4.385899
#> 21  502.0983   2.655960
#> 22  512.7747   2.532853
#> 23  503.4205   2.829646
#> 24  712.7995   2.628618
#> 25  518.1258   2.694929
#> 26  500.3719   3.020393
#> 27  500.0424   3.131797
#> 28  733.9833   3.119108
#> 29  500.7223   3.791902
#> 30  500.4740   2.942798
#> 31  503.3026   2.544673
#> 32  500.5079   2.888056
#> 33  500.3897   2.520975
#> 34  500.6122   2.824147
#> 35  500.0708   3.063983
#> 36  509.5867   4.713047
#> 37  500.4951   3.309042
#> 38  500.0913   3.142026
#> 39  629.3599   2.952373
#> 40  512.1945   3.605951
#> 41  501.6068   3.186116
#> 42  511.4815   4.123174
#> 43  500.0087   2.861956
#> 44  500.1381   2.829289
#> 45  693.6341   3.097285
#> 46  527.5255   3.170553
#> 47  578.1838   2.670518
#> 48  501.1987   2.617851
#> 49  517.5450   2.683891
#> 50  518.2611   2.537413
#> 51  501.1770   3.058596
#> 52  500.0504   2.565274
#> 53  500.2127   2.598288
#> 54  723.5392   2.586056
#> 55  501.2298   2.726991
#> 56  500.1515   2.850502
#> 57  510.2579   2.512757
#> 58  588.8770   2.645740
#> 59  500.1363   2.916838
#> 60  540.6403   2.921529
#> 61  500.2902   2.811306
#> 62  501.4373   2.546444
#> 63  525.2619   3.285211
#> 64  532.2829   2.579727
#> 65  503.5620   3.023307
#> 66  500.3887   2.635861
#> 67  516.5327   2.518907
#> 68  508.7678   2.656027
#> 69  600.6504   2.614788
#> 70  501.2001   2.512641
#> 71  519.3195   3.318022
#> 72  603.7736   2.645849
#> 73  503.4985   2.950387
#> 74  500.0201   3.751989
#> 75  502.8955   2.746404
#> 76  506.8037   2.690038
#> 77  555.2722   2.574621
#> 78  504.6532   3.163392
#> 79  580.4184   3.088780
#> 80  830.5686   2.724100
#> 81  939.6951   2.829356
#> 82  568.4348   3.162683
#> 83  500.0343   2.654604
#> 84  840.9291   2.672101
#> 85  500.1161   2.647223
#> 86  500.5800   2.596591
#> 87  536.8588   2.535249
#> 88  589.8184   2.870890
#> 89  503.7039   3.134519
#> 90  500.0649   2.597083
#> 91  607.7097   3.160692
#> 92  507.5661   2.507288
#> 93  501.8536   2.565408
#> 94  503.1420   2.699700
#> 95  510.8555   2.834821
#> 96  501.3164   2.926968
#> 97  501.0931   2.696451
#> 98  587.7833   2.895883
#> 99  501.7312   2.595778
#> 100 809.2298   2.777361
#> 101 514.5404   2.566821
#> 102 501.0798   2.919406
#> 103 520.7801   2.860714
#> 104 502.1084   2.871580
#> 105 532.0530   3.193304
#> 106 542.1409   3.208041
#> 107 507.0524   3.129609
#> 108 511.8053   2.946758
#> 109 504.7230   2.775846
#> 110 502.1032   4.282471
#> 111 542.7769   2.645893
#> 112 521.6796   2.758710
#> 113 500.0352   2.556124
#> 114 501.6744   3.493084
#> 115 516.1205   3.738540
#> 116 500.1626   2.760884
#> 117 503.4424   3.108949
#> 118 513.9760   3.091153
#> 119 501.0657   2.556964
#> 120 500.0930   2.500541
#> 121 500.0057   2.727088
#> 122 500.8850   3.044424
#> 123 520.9433   2.649292
#> 124 500.6271   2.642371
#> 125 506.7894   2.856079
#> 126 520.0044   2.563104
#> 127 808.5357   2.682746
#> 128 500.2036   3.512157
#> 129 514.2532   3.975372
#> 130 500.6284   2.769430
#> 131 518.2677   2.729688
#> 132 505.1459   2.568541
#> 133 511.1627   3.656041
#> 134 500.0032   2.528052
#> 135 506.7654   2.809629
#> 136 500.0409   3.105535
#> 137 500.0140   2.584206
#> 138 507.4484   3.756080
#> 139 501.1019   2.705884
#> 140 529.2985   2.810673
#> 141 522.8812   2.903541
#> 142 501.2001   2.680356
#> 143 691.3927   2.966899
#> 144 500.0044   2.652931
#> 145 523.4838   2.753973
#> 146 500.6197   2.569079
#> 147 500.2462   3.654103
#> 148 564.0678   3.072100
#> 149 500.0286   2.558535
#> 150 528.9938   3.230541
#> 151 515.7321   2.653732
#> 152 500.1969   2.604799
#> 153 547.9442   2.692012
#> 154 597.6385   2.594727
#> 155 611.0571   2.721923
#> 156 510.5447   2.516355
#> 157 529.6131   2.847547
#> 158 562.7408   3.498635
#> 159 544.8893   2.583782
#> 160 521.9210   2.727715
#> 161 519.1132   2.569812
#> 162 510.8366   2.967166
#> 163 500.0434   2.520032
#> 164 500.6076   2.606997
#> 165 500.8659   2.529399
#> 166 502.1330   2.796214
#> 167 536.8937   2.659398
#> 168 595.2779   2.946992
#> 169 511.3023   3.073852
#> 170 666.5766   2.887526
#> 171 510.4889   2.604800
#> 172 501.8029   3.227736
#> 173 500.2530   3.067306
#> 174 503.7596   2.532573
#> 175 507.2585   3.501188
#> 176 509.7686   3.057648
#> 177 502.3964   2.933647
#> 178 633.9291   2.575690
#> 179 500.3102   2.774327
#> 180 500.0049   3.029400
#> 181 764.3851   2.500246
#> 182 500.1381   2.529548
#> 183 506.5657   2.925359
#> 184 500.3133   2.645735
#> 185 500.8096   2.520963
#> 186 504.4719   2.893748
#> 187 588.7704   2.764370
#> 188 518.1589   2.566433
#> 189 501.9602   3.003087
#> 190 500.4392   3.032154
#> 191 503.0658   3.043005
#> 192 536.2538   3.348108
#> 193 500.1251   2.930408
#> 194 501.9244   2.652591
#> 195 975.9878   3.852783
#> 196 501.3887   2.577933
#> 197 501.5487   2.768175
#> 198 583.5553   2.588757
#> 199 501.6186   2.506236
#> 200 501.5054   3.964391
#> 201 500.3742   2.943236
#> 202 500.7022   3.502366
#> 203 558.4387   3.728636
#> 204 616.8717   3.056760
#> 205 500.2723   3.742140
#> 206 507.4006   2.582894
#> 207 517.2139   3.386885
#> 208 556.0655   2.673554
#> 209 500.1490   2.505999
#> 210 516.3675   3.324487
#> 211 588.9596   2.583145
#> 212 526.0415   3.551160
#> 213 867.0335   3.156709
#> 214 535.1561   2.561573
#> 215 500.7636   2.984338
#> 216 509.6344   2.653351
#> 217 502.0317   2.817469
#> 218 509.1389   2.532235
#> 219 509.8126   2.907734
#> 220 501.8598   2.771025
#> 221 507.4478   3.606707
#> 222 500.3170   2.506271
#> 223 520.8009   2.824071
#> 224 760.5616   3.350486
#> 225 500.5481   2.865154
#> 226 503.9456   3.084364
#> 227 500.2971   3.355634
#> 228 503.5729   2.832398
#> 229 502.6073   3.697310
#> 230 500.0008   3.069993
#> 231 500.0982   2.575635
#> 232 500.0295   2.662158
#> 233 537.1817   2.689566
#> 234 503.9201   3.183242
#> 235 501.1455   2.515373
#> 236 500.7058   2.747756
#> 237 500.4532   2.515844
#> 238 500.0556   2.793349
#> 239 521.5504   2.588035
#> 240 528.1038   2.926943
#> 241 501.2110   2.684830
#> 242 672.1890   3.114752
#> 243 500.0649   3.776028
#> 244 500.4287   4.069197
#> 245 510.0681   2.605472
#> 246 505.9089   2.608506
#> 247 507.9114   3.103191
#> 248 545.6590   3.083324
#> 249 502.1999   2.918407
#> 250 500.6650   2.531026
#> 251 500.0579   3.039943
#> 252 509.3737   2.504817
#> 253 504.1312   2.827624
#> 254 516.8049   3.070553
#> 255 504.2122   4.035015
#> 256 500.2234   3.208764
#> 257 512.8387   3.267753
#> 258 505.6538   2.638627
#> 259 973.4670   3.231197
#> 260 538.6581   3.064945
#> 261 521.0785   2.516389
#> 262 500.0018   2.548184
#> 263 747.8123   2.955899
#> 264 501.9335   3.638686
#> 265 500.7080   2.973969
#> 266 500.1652   2.586738
#> 267 500.6781   2.712713
#> 268 877.6126   2.507195
#> 269 500.2246   2.891016
#> 270 513.3835   3.095736
#> 271 500.5775   3.127775
#> 272 556.5068   2.556935
#> 273 555.0179   3.393105
#> 274 500.0635   3.034140
#> 275 513.7627   2.546738
#> 276 525.9062   2.972897
#> 277 502.6716   2.565924
#> 278 508.8938   2.549403
#> 279 502.9247   2.578182
#> 280 500.3141   2.581058
#> 281 789.1751   2.717509
#> 282 500.0348   2.504651
#> 283 504.8661   2.704862
#> 284 503.5524   2.578107
#> 285 503.1456   3.010822
#> 286 500.4339   2.558936
#> 287 500.2245   3.553809
#> 288 573.1022   2.791382
#> 289 500.3142   2.578206
#> 290 503.8917   2.880076
#> 291 500.6050   2.792909
#> 292 607.1686   3.459342
#> 293 718.8926   3.114143
#> 294 562.8924   2.995875
#> 295 500.8557   2.912915
#> 296 625.7256   2.744616
#> 297 500.0093   2.906098
#> 298 500.6781   2.699843
#> 299 640.0886   2.713174
#> 300 665.8287   3.055913
#> 301 500.0017   2.957022
#> 302 500.1861   3.487153
#> 303 501.9758   2.948923
#> 304 530.6156   2.778265
#> 305 531.4612   2.573444
#> 306 500.9540   2.503673
#> 307 734.0468   2.600636
#> 308 545.9307   2.786640
#> 309 500.2574   2.915293
#> 310 508.6207   2.836542
#> 311 542.1543   3.046446
#> 312 770.5909   3.756222
#> 313 509.0098   2.573167
#> 314 519.8942   2.520021
#> 315 727.8014   2.770124
#> 316 541.1050   2.692614
#> 317 521.2065   2.963042
#> 318 500.3422   3.165894
#> 319 505.7598   2.752421
#> 320 501.4527   2.588169
#> 321 504.0215   2.633746
#> 322 501.8877   2.741305
#> 323 518.2938   2.991158
#> 324 512.0703   2.862527
#> 325 501.6771   2.732489
#> 326 508.8826   2.665370
#> 327 623.7599   2.752774
#> 328 500.1419   2.720513
#> 329 500.0339   2.762533
#> 330 784.8932   3.442654
#> 331 500.1984   2.537400
#> 332 587.1375   3.726462
#> 333 500.3886   2.541791
#> 334 500.6357   2.824824
#> 335 500.9276   2.518705
#> 336 510.6919   2.548936
#> 337 501.9129   2.938601
#> 338 500.0631   2.580726
#> 339 500.6240   3.024652
#> 340 502.8746   2.527818
#> 341 509.5596   3.626675
#> 342 500.2701   3.102471
#> 343 500.3647   2.553273
#> 344 501.3375   3.004682
#> 345 975.2951   4.088700
#> 346 506.5352   3.631845
#> 347 514.4936   2.668136
#> 348 727.4096   2.958047
#> 349 635.9644   2.642834
#> 350 500.1341   2.596576
#> 351 502.6420   2.779348
#> 352 518.7423   3.095602
#> 353 501.9951   3.144359
#> 354 505.8334   4.507245
#> 355 728.9806   3.272743
#> 356 520.2913   2.939665
#> 357 508.4927   2.567725
#> 358 500.6049   2.659380
#> 359 511.9502   2.542744
#> 360 535.3303   2.918547
#> 361 500.2407   2.528927
#> 362 500.0526   2.701677
#> 363 500.0180   2.848592
#> 364 504.4825   3.422727
#> 365 645.1621   3.519050
#> 366 665.7341   2.714024
#> 367 504.0842   2.749670
#> 368 501.5023   2.858603
#> 369 501.5059   2.758639
#> 370 516.9024   2.600926
#> 371 505.9484   2.594860
#> 372 598.9987   2.944021
#> 373 508.0862   3.553207
#> 374 500.2266   3.686305
#> 375 506.1324   2.787098
#> 376 501.3490   2.951895
#> 377 500.0731   3.463592
#> 378 500.5402   2.624056
#> 379 501.6986   2.502896
#> 380 500.4668   2.734691
#> 381 503.2628   2.773693
#> 382 503.5647   2.981414
#> 383 936.9667   2.587904
#> 384 500.2417   3.029319
#> 385 500.0420   3.113653
#> 386 502.5004   3.392501
#> 387 502.5281   2.680726
#> 388 501.1572   2.534231
#> 389 501.1747   3.130771
#> 390 771.5556   2.648101
#> 391 511.4917   2.620662
#> 392 506.6701   2.701887
#> 393 500.0144   3.400351
#> 394 519.3423   2.769682
#> 395 500.0070   3.192209
#> 396 502.3067   2.541562
#> 397 500.8555   3.242296
#> 398 504.0118   2.899536
#> 399 501.0675   3.252117
#> 400 502.9175   2.618234
#> 401 500.0184   2.737505
#> 402 805.7597   3.292995
#> 403 500.0021   2.598836
#> 404 732.8222   2.564523
#> 405 507.4553   2.671335
#> 406 500.0340   3.181256
#> 407 534.8842   3.933579
#> 408 522.6104   2.675273
#> 409 500.4141   2.620430
#> 410 506.7895   2.828854
#> 411 500.1069   2.678891
#> 412 508.5656   2.883707
#> 413 500.1509   2.738507
#> 414 501.0285   2.630658
#> 415 500.7298   2.634853
#> 416 500.9716   2.856186
#> 417 500.2131   3.226596
#> 418 500.7637   2.802308
#> 419 510.6904   3.032701
#> 420 503.5428   3.679693
#> 421 500.3907   2.745002
#> 422 515.3264   2.825178
#> 423 866.1333   3.451496
#> 424 500.0219   2.628494
#> 425 500.6331   3.117199
#> 426 502.7464   3.170526
#> 427 501.4412   3.281533
#> 428 502.1297   3.418858
#> 429 597.6078   2.661645
#> 430 804.3634   3.283374
#> 431 517.1410   2.880582
#> 432 663.9693   2.592237
#> 433 501.1451   2.533854
#> 434 500.0080   2.660417
#> 435 503.0684   2.610476
#> 436 712.6287   2.728613
#> 437 500.0035   2.602524
#> 438 589.8115   2.698495
#> 439 500.6538   2.650770
#> 440 671.3197   3.291747
#> 441 501.0240   2.719190
#> 442 546.0253   3.127690
#> 443 500.1185   3.235581
#> 444 501.5026   2.735214
#> 445 511.8294   2.648404
#> 446 500.0402   2.630047
#> 447 501.0408   3.125630
#> 448 500.4985   3.653287
#> 449 500.4407   2.902880
#> 450 501.6228   2.908704
#> 451 500.0195   3.749237
#> 452 869.3658   2.878165
#> 453 500.0450   2.504639
#> 454 502.5140   2.513980
#> 455 662.0088   2.799471
#> 456 500.5005   2.699898
#> 457 500.4638   2.675039
#> 458 504.2645   3.795221
#> 459 500.6837   2.561345
#> 460 502.6784   3.622397
#> 461 622.5949   2.532831
#> 462 501.3618   3.039458
#> 463 564.7012   4.026258
#> 464 500.0007   2.590345
#> 465 528.0439   2.542281
#> 466 500.6598   4.655270
#> 467 506.7126   2.790510
#> 468 506.4268   2.702921
#> 469 508.3891   3.220921
#> 470 502.5108   3.023827
#> 471 543.4592   2.661375
#> 472 501.4035   2.600943
#> 473 537.6054   3.463366
#> 474 506.0273   2.601597
#> 475 641.4596   2.560757
#> 476 506.9107   2.500813
#> 477 500.0750   2.742794
#> 478 596.6607   2.551722
#> 479 500.0909   3.939973
#> 480 500.3729   2.915077
#> 481 503.8135   2.896779
#> 482 500.0621   3.489822
#> 483 774.5791   2.954159
#> 484 500.5809   3.260436
#> 485 500.0463   2.968493
#> 486 500.1312   2.607011
#> 487 500.2261   2.740053
#> 488 500.0066   2.835520
#> 489 504.0422   2.642081
#> 490 500.3249   4.776085
#> 491 500.0286   2.941254
#> 492 500.8464   2.701516
#> 493 500.0372   2.937702
#> 494 708.7520   2.545419
#> 495 733.3049   3.010468
#> 496 506.5330   2.818773
#> 497 531.8936   2.986069
#> 498 500.0227   2.997504
#> 499 502.2667   3.717880
#> 500 502.7636   2.582815
#> 501 500.0822   2.540139
#> 502 502.4636   3.142032
#> 503 500.2081   2.565913
#> 504 512.5330   2.578000
#> 505 502.2737   2.593792
#> 506 506.3438   4.105671
#> 507 538.2815   3.303303
#> 508 500.0327   2.888087
#> 509 502.3101   3.311869
#> 510 501.3831   2.679926
#> 511 765.4247   3.558857
#> 512 524.2810   2.746969
#> 513 502.2470   2.826648
#> 514 500.6338   3.022230
#> 515 596.5616   2.954110
#> 516 770.4078   2.703121
#> 517 500.7497   2.605519
#> 518 500.0484   2.765897
#> 519 500.0977   2.738681
#> 520 503.0021   2.796780
#> 521 615.6478   2.565023
#> 522 524.1374   2.532240
#> 523 506.7577   2.527252
#> 524 531.3658   3.065696
#> 525 570.7568   2.525143
#> 526 501.9387   2.665665
#> 527 667.2526   2.600224
#> 528 501.0206   2.805144
#> 529 500.1493   3.138570
#> 530 500.1471   2.707937
#> 531 535.6605   4.402411
#> 532 897.6998   3.116385
#> 533 501.4713   2.552352
#> 534 772.9926   2.650196
#> 535 510.7751   2.692575
#> 536 502.1472   2.511543
#> 537 506.1597   2.740779
#> 538 501.2734   3.430064
#> 539 501.8949   2.768359
#> 540 500.8951   3.341775
#> 541 500.3939   3.735098
#> 542 500.1053   2.762075
#> 543 771.8607   2.523504
#> 544 500.1480   3.652049
#> 545 500.1614   2.666828
#> 546 500.4873   2.973550
#> 547 500.0421   2.745638
#> 548 549.2346   2.869548
#> 549 504.2462   2.755438
#> 550 646.5685   2.687226
#> 551 507.7148   2.642837
#> 552 505.7435   2.730998
#> 553 569.7763   2.590819
#> 554 500.1590   3.130961
#> 555 545.1500   2.923355
#> 556 784.1484   2.594636
#> 557 500.0174   2.540327
#> 558 603.8310   2.725667
#> 559 500.3480   2.644327
#> 560 507.8633   2.624551
#> 561 500.2840   2.515009
#> 562 504.3081   2.556722
#> 563 617.2813   2.551082
#> 564 515.4475   2.584441
#> 565 500.0523   2.638950
#> 566 500.0218   2.533160
#> 567 787.6986   3.333286
#> 568 510.8660   2.791564
#> 569 500.0859   2.933705
#> 570 844.1246   2.794254
#> 571 557.3125   2.631146
#> 572 504.1614   3.009722
#> 573 561.0359   3.180372
#> 574 571.4146   3.086648
#> 575 657.7247   2.601753
#> 576 500.0239   2.805376
#> 577 604.2390   3.195014
#> 578 590.1194   2.517654
#> 579 726.6392   2.782126
#> 580 500.0505   4.094643
#> 581 799.6385   2.992220
#> 582 506.5970   3.544605
#> 583 500.2496   2.850093
#> 584 619.6385   2.653735
#> 585 500.2994   4.006026
#> 586 500.0330   2.637743
#> 587 515.1692   3.089759
#> 588 594.0735   2.694187
#> 589 500.1034   2.859430
#> 590 500.6801   3.771197
#> 591 500.4491   2.705756
#> 592 500.9905   2.635322
#> 593 500.0039   2.582155
#> 594 582.1083   3.160984
#> 595 500.1141   3.384513
#> 596 571.2088   3.113058
#> 597 579.9838   2.893866
#> 598 500.3902   4.273866
#> 599 500.0312   2.547067
#> 600 927.2160   2.671458
#> 601 500.9147   3.337039
#> 602 514.6074   2.540614
#> 603 500.2567   2.827865
#> 604 506.3453   2.973268
#> 605 974.2276   2.642957
#> 606 785.2984   3.113965
#> 607 502.8396   3.018300
#> 608 500.1333   3.303795
#> 609 606.1684   2.653701
#> 610 527.0517   2.794364
#> 611 500.1958   2.666101
#> 612 521.3920   2.885521
#> 613 922.3378   2.638051
#> 614 506.8893   3.007043
#> 615 500.1347   2.874135
#> 616 506.9311   2.553520
#> 617 503.9802   2.985638
#> 618 501.8154   3.967599
#> 619 562.2774   3.708576
#> 620 500.4115   3.213534
#> 621 533.2881   2.652635
#> 622 505.9224   3.231609
#> 623 502.7955   3.107154
#> 624 500.5388   2.622121
#> 625 500.3115   2.613236
#> 626 500.0289   4.518324
#> 627 500.0207   2.984128
#> 628 502.5688   2.802321
#> 629 508.9876   2.551194
#> 630 500.0149   2.829846
#> 631 608.8438   2.837637
#> 632 519.7632   2.677217
#> 633 535.0571   3.179954
#> 634 500.0906   3.197362
#> 635 765.2964   2.873683
#> 636 500.6858   2.765550
#> 637 508.1562   2.844129
#> 638 739.7153   2.573422
#> 639 500.3522   2.800736
#> 640 768.3246   3.576356
#> 641 508.1990   2.533908
#> 642 500.1400   2.849473
#> 643 516.6438   3.936722
#> 644 500.0166   4.107444
#> 645 513.5430   3.616746
#> 646 509.5025   2.852226
#> 647 503.1750   3.460279
#> 648 503.0123   3.066231
#> 649 502.0635   2.560514
#> 650 507.1203   3.325876
#> 651 536.7710   2.747148
#> 652 533.6755   2.668761
#> 653 511.0797   2.631969
#> 654 501.0115   3.197883
#> 655 500.6498   2.528495
#> 656 500.6518   2.842139
#> 657 500.0192   2.623748
#> 658 892.4208   2.667285
#> 659 512.8850   2.783606
#> 660 778.7208   3.183733
#> 661 500.0892   2.666184
#> 662 500.1744   2.570260
#> 663 503.3924   3.601904
#> 664 501.1209   3.025149
#> 665 503.5141   3.040034
#> 666 619.4427   2.595637
#> 667 577.9906   2.771203
#> 668 502.5338   2.815815
#> 669 503.9614   2.801536
#> 670 717.9385   2.657667
#> 671 813.3025   3.021201
#> 672 500.3027   3.005498
#> 673 512.0072   2.588308
#> 674 502.3842   2.561439
#> 675 709.3880   2.775077
#> 676 562.6563   2.910912
#> 677 501.9526   2.975431
#> 678 711.7542   2.606119
#> 679 502.0546   2.996441
#> 680 518.7948   2.574281
#> 681 505.9154   2.562939
#> 682 501.0413   2.725388
#> 683 842.1024   2.561634
#> 684 500.3733   2.605663
#> 685 677.0970   2.714728
#> 686 519.6072   3.822635
#> 687 500.5329   2.724243
#> 688 529.2578   2.944088
#> 689 501.8148   2.882045
#> 690 501.1243   3.047400
#> 691 958.4759   3.264515
#> 692 500.6379   2.660178
#> 693 501.9801   3.123317
#> 694 500.6502   2.720824
#> 695 503.6605   2.704428
#> 696 521.0771   3.573264
#> 697 509.1809   2.927688
#> 698 503.2438   2.669186
#> 699 500.2924   3.467737
#> 700 635.2392   2.631804
#> 701 775.7097   3.822229
#> 702 712.4926   2.685219
#> 703 508.9481   2.611427
#> 704 504.7701   3.268046
#> 705 500.2447   3.054849
#> 706 500.3632   2.543121
#> 707 582.8331   3.293541
#> 708 500.9575   2.508497
#> 709 635.5049   4.199445
#> 710 500.0938   2.883660
#> 711 500.0626   2.601981
#> 712 504.8698   2.527625
#> 713 514.0938   3.282263
#> 714 560.7814   3.537716
#> 715 501.1707   2.709275
#> 716 507.2993   3.024734
#> 717 501.6422   2.643256
#> 718 504.1775   2.718198
#> 719 500.2026   3.057983
#> 720 511.6570   3.106068
#> 721 501.5762   2.968713
#> 722 501.0231   2.543627
#> 723 976.9193   3.099170
#> 724 501.0260   2.735565
#> 725 500.0556   2.806324
#> 726 502.2133   3.363153
#> 727 520.5879   2.527012
#> 728 500.0268   2.768005
#> 729 506.0545   2.997135
#> 730 500.0085   2.723396
#> 731 517.2713   2.630003
#> 732 549.7908   4.451880
#> 733 502.1511   3.082422
#> 734 500.0856   2.892815
#> 735 500.3635   2.587155
#> 736 664.4536   2.685648
#> 737 750.3821   2.955886
#> 738 506.9466   2.650640
#> 739 500.2003   2.681250
#> 740 500.1068   2.547473
#> 741 500.2800   3.071301
#> 742 500.5256   2.812320
#> 743 500.8279   4.308915
#> 744 503.9244   3.224412
#> 745 500.2291   2.861664
#> 746 500.1085   2.725994
#> 747 500.0205   2.735372
#> 748 543.0900   3.133769
#> 749 503.0374   2.836553
#> 750 505.4422   2.564165
#> 751 520.5696   2.892436
#> 752 504.1000   2.758254
#> 753 510.8177   3.009588
#> 754 507.0171   2.656635
#> 755 501.3245   4.345258
#> 756 500.0083   2.666067
#> 757 543.5705   2.871301
#> 758 688.3957   3.145096
#> 759 500.1815   2.579658
#> 760 541.2927   2.668212
#> 761 503.5566   3.095420
#> 762 651.5412   2.624514
#> 763 532.8692   2.951286
#> 764 500.0771   3.245829
#> 765 500.0010   3.642114
#> 766 500.4109   3.299348
#> 767 500.2433   2.894494
#> 768 767.1466   2.727275
#> 769 500.3681   2.502443
#> 770 500.2121   2.957067
#> 771 501.2428   2.687833
#> 772 502.0653   3.380722
#> 773 500.6897   3.161818
#> 774 688.4786   3.015270
#> 775 516.8307   3.009063
#> 776 503.0092   2.520592
#> 777 642.5487   2.628907
#> 778 682.0030   3.343184
#> 779 500.7533   2.775100
#> 780 500.2711   2.807185
#> 781 764.0114   3.075651
#> 782 502.0043   2.952789
#> 783 622.5469   2.740539
#> 784 616.5728   4.159490
#> 785 577.8260   2.944850
#> 786 508.7547   3.268246
#> 787 543.7404   2.677670
#> 788 500.5342   2.609786
#> 789 501.0268   2.685158
#> 790 654.1733   2.851694
#> 791 558.7994   2.609559
#> 792 560.6802   2.773690
#> 793 500.7523   2.541074
#> 794 501.4415   2.523081
#> 795 500.0420   2.790101
#> 796 500.0978   2.502224
#> 797 501.7666   2.589973
#> 798 522.4738   2.817204
#> 799 502.6931   3.237552
#> 800 510.9096   2.859765
#> 801 500.0448   2.659697
#> 802 536.6376   2.661090
#> 803 504.4263   2.782751
#> 804 532.5368   2.844923
#> 805 500.0992   4.380040
#> 806 500.2205   2.930563
#> 807 500.2703   2.549239
#> 808 500.1824   2.577824
#> 809 547.7903   3.165408
#> 810 500.1730   2.913095
#> 811 500.3790   2.541694
#> 812 596.7121   2.733274
#> 813 500.0222   3.306818
#> 814 814.2754   3.169836
#> 815 548.5976   3.091762
#> 816 688.2362   2.645725
#> 817 507.8073   2.813724
#> 818 500.7525   2.884594
#> 819 500.3439   2.646729
#> 820 558.9946   3.559449
#> 821 827.8874   2.812324
#> 822 500.0058   2.591378
#> 823 502.7857   3.240057
#> 824 500.1112   3.488943
#> 825 500.2704   3.033686
#> 826 500.6073   2.596513
#> 827 500.0403   3.087630
#> 828 505.7281   2.991510
#> 829 500.3491   2.667585
#> 830 500.0138   2.661560
#> 831 587.5231   2.641752
#> 832 500.1684   3.155708
#> 833 538.9954   2.690689
#> 834 500.5432   2.597516
#> 835 500.0274   2.658868
#> 836 500.1706   2.522932
#> 837 500.1223   2.809690
#> 838 541.3138   2.865501
#> 839 502.5734   3.004694
#> 840 559.9231   2.685807
#> 841 974.8250   2.808649
#> 842 500.0528   2.545561
#> 843 508.2425   2.624117
#> 844 501.2889   2.848927
#> 845 515.2270   2.749473
#> 846 500.1682   2.529605
#> 847 501.4165   2.688626
#> 848 502.9178   2.656571
#> 849 503.6281   2.504949
#> 850 501.8754   2.667882
#> 851 700.3322   2.565807
#> 852 570.4814   2.640570
#> 853 508.3829   2.884302
#> 854 500.1754   2.682325
#> 855 501.5132   2.921163
#> 856 500.0422   3.374057
#> 857 500.0121   2.713336
#> 858 506.7397   2.668613
#> 859 960.2985   2.893313
#> 860 500.0344   3.302614
#> 861 500.2517   3.402825
#> 862 500.0902   3.131520
#> 863 524.7150   2.827518
#> 864 500.4565   3.071402
#> 865 501.6717   3.203716
#> 866 527.9116   3.558227
#> 867 500.7194   2.979293
#> 868 507.0695   3.094420
#> 869 513.0547   3.435768
#> 870 505.6156   2.678750
#> 871 709.8587   2.904659
#> 872 502.5800   3.002505
#> 873 501.6260   3.311059
#> 874 501.5017   2.904730
#> 875 502.2730   3.026590
#> 876 500.3709   2.694892
#> 877 502.1892   2.633305
#> 878 512.4684   3.727054
#> 879 500.9097   2.592492
#> 880 546.5114   3.047711
#> 881 503.3085   2.898089
#> 882 755.9935   2.915155
#> 883 664.0465   2.792697
#> 884 926.7247   2.990441
#> 885 500.3990   2.847820
#> 886 505.4188   2.724827
#> 887 500.0744   2.892752
#> 888 508.8359   2.873043
#> 889 524.7615   2.651789
#> 890 514.9442   3.020411
#> 891 500.2858   3.072791
#> 892 510.7697   2.500977
#> 893 505.1194   2.742564
#> 894 500.0738   3.439969
#> 895 566.9339   4.016969
#> 896 503.5071   2.604969
#> 897 501.9806   2.897828
#> 898 977.7606   2.560418
#> 899 500.2884   2.695262
#> 900 502.7879   3.394185
#> 901 500.0897   2.891147
#> 902 513.6902   2.801921
#> 903 500.2958   4.498706
#> 904 501.2745   2.560855
#> 905 500.5633   2.987553
#> 906 500.5563   2.887550
#> 907 902.4275   2.909056
#> 908 502.1181   2.654649
#> 909 500.5415   3.102712
#> 910 507.8295   2.695403
#> 911 521.1700   3.287180
#> 912 500.7489   2.572168
#> 913 500.3589   2.609168
#> 914 989.3764   2.555725
#> 915 521.4036   2.656105
#> 916 504.5039   3.970853
#> 917 500.4422   2.829448
#> 918 504.2810   2.662252
#> 919 500.1673   2.573519
#> 920 500.2113   2.827510
#> 921 500.0570   2.794076
#> 922 563.1461   2.846213
#> 923 709.8056   2.853352
#> 924 500.6373   3.163633