Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   630.4690   2.512088
#> 2   501.0224   2.549541
#> 3   505.4614   2.535187
#> 4   500.4487   2.565072
#> 5   500.0406   2.789289
#> 6   505.3188   3.048815
#> 7   500.2931   2.758995
#> 8   500.0701   3.031363
#> 9   534.8090   3.461677
#> 10  835.2060   2.858857
#> 11  626.2082   2.954338
#> 12  500.0093   3.057340
#> 13  507.3143   2.642296
#> 14  503.4067   3.320089
#> 15  501.6056   2.625800
#> 16  532.8296   3.509653
#> 17  501.9474   2.863371
#> 18  537.0580   2.690019
#> 19  551.3906   2.636152
#> 20  575.9632   3.226919
#> 21  500.0525   2.699660
#> 22  500.4233   3.217864
#> 23  512.0125   6.690949
#> 24  686.2775   2.772774
#> 25  505.9252   2.691396
#> 26  520.0438   2.516684
#> 27  603.6267   2.603994
#> 28  687.4467   3.313018
#> 29  500.0285   2.614631
#> 30  513.3657   2.746677
#> 31  919.1480   2.563205
#> 32  503.1590   2.863693
#> 33  503.8287   2.762031
#> 34  505.8828   2.836445
#> 35  567.1868   3.038684
#> 36  763.6678   2.867870
#> 37  500.2404   3.034203
#> 38  514.7175   2.698537
#> 39  683.1220   2.911602
#> 40  732.0129   2.645149
#> 41  590.3439   2.956485
#> 42  501.3212   3.893976
#> 43  500.1957   3.127984
#> 44  500.0188   2.709532
#> 45  500.0612   4.249737
#> 46  507.8732   3.048863
#> 47  538.9492   3.065276
#> 48  500.2517   2.934053
#> 49  500.4137   2.622532
#> 50  518.0750   2.958760
#> 51  500.4527   2.672090
#> 52  503.4901   3.181338
#> 53  506.2085   2.590899
#> 54  500.6041   2.706322
#> 55  518.0191   2.642635
#> 56  505.5071   3.660026
#> 57  503.6667   2.522486
#> 58  500.0814   3.181508
#> 59  642.4359   3.421716
#> 60  547.2856   2.657613
#> 61  786.4007   2.881964
#> 62  570.5974   2.763799
#> 63  539.2802   2.767182
#> 64  712.5478   2.662851
#> 65  922.1016   3.311695
#> 66  503.6109   2.677197
#> 67  500.2127   2.613370
#> 68  535.9705   4.079100
#> 69  500.0739   2.597790
#> 70  500.1018   3.247206
#> 71  512.8217   2.969821
#> 72  504.7437   2.549767
#> 73  522.2146   3.435924
#> 74  528.6937   2.900380
#> 75  500.5670   3.009428
#> 76  508.7365   2.577071
#> 77  500.2471   2.630331
#> 78  925.9028   2.643912
#> 79  506.0578   3.700745
#> 80  500.6031   2.528672
#> 81  505.4745   2.631937
#> 82  500.0021   2.596046
#> 83  500.1515   3.767358
#> 84  503.4339   2.579334
#> 85  640.6029   3.098570
#> 86  559.2533   2.985560
#> 87  555.7198   2.505664
#> 88  500.0086   2.739458
#> 89  974.5646   3.014882
#> 90  507.0909   2.678339
#> 91  519.5211   2.505309
#> 92  500.0219   2.566285
#> 93  595.1618   2.646588
#> 94  643.3153   2.526024
#> 95  512.2490   3.538953
#> 96  500.0067   2.648813
#> 97  501.1946   2.661704
#> 98  515.2396   2.842082
#> 99  501.0373   2.730655
#> 100 501.3931   2.565441
#> 101 500.1373   2.631986
#> 102 501.1920   2.752142
#> 103 703.1736   2.688609
#> 104 501.2815   3.274366
#> 105 500.0459   2.553204
#> 106 685.3594   3.313015
#> 107 619.5162   2.625832
#> 108 507.2260   2.924892
#> 109 500.0193   2.929998
#> 110 564.0412   2.842044
#> 111 500.6672   3.168588
#> 112 505.0528   3.048434
#> 113 500.0576   2.741484
#> 114 500.9534   2.618134
#> 115 526.4886   2.678979
#> 116 781.8150   2.886855
#> 117 500.0744   2.621478
#> 118 690.0236   3.382954
#> 119 502.3491   3.397984
#> 120 604.4752   2.845175
#> 121 500.7973   3.040415
#> 122 500.3445   3.214057
#> 123 500.0008   3.865844
#> 124 500.1816   2.715285
#> 125 512.4554   3.329423
#> 126 503.3426   2.551680
#> 127 500.3863   2.513789
#> 128 857.7814   3.171552
#> 129 500.0507   2.547111
#> 130 500.4703   2.877786
#> 131 500.2199   3.128287
#> 132 501.2756   3.054138
#> 133 500.0755   2.881227
#> 134 757.3363   2.525432
#> 135 740.1672   2.714984
#> 136 501.3703   3.204344
#> 137 778.1022   3.212343
#> 138 911.8866   2.687803
#> 139 500.5892   2.845061
#> 140 712.8081   2.914745
#> 141 772.0329   2.534591
#> 142 500.3543   2.602749
#> 143 500.3346   2.536805
#> 144 683.0889   2.641730
#> 145 501.4532   2.805001
#> 146 500.3635   2.946174
#> 147 535.9187   3.473275
#> 148 512.5494   2.830045
#> 149 500.0954   2.837582
#> 150 876.5989   3.090690
#> 151 525.0101   2.756395
#> 152 501.1864   2.849239
#> 153 500.0827   3.006372
#> 154 500.4683   3.809920
#> 155 501.8411   3.403289
#> 156 500.4666   2.727512
#> 157 513.2712   3.142550
#> 158 500.1763   2.725714
#> 159 504.2166   3.572012
#> 160 500.0135   2.532655
#> 161 500.3190   2.594526
#> 162 645.6419   2.546574
#> 163 503.3681   3.393535
#> 164 509.6553   4.591546
#> 165 500.4136   3.139898
#> 166 500.2568   2.880642
#> 167 622.8624   2.797597
#> 168 500.2215   2.899836
#> 169 502.7447   2.587871
#> 170 504.9023   3.165897
#> 171 928.3968   2.796699
#> 172 500.2855   2.667015
#> 173 523.0292   3.144094
#> 174 501.0564   3.323743
#> 175 894.9526   3.069305
#> 176 513.7160   3.064763
#> 177 577.0112   2.571667
#> 178 528.4613   2.562664
#> 179 500.2154   2.820552
#> 180 505.6963   3.662190
#> 181 506.5742   2.996110
#> 182 500.2376   2.601209
#> 183 500.0502   2.932379
#> 184 501.3711   2.820880
#> 185 502.8506   3.042728
#> 186 562.0577   2.870453
#> 187 500.8217   2.708198
#> 188 500.0471   2.812868
#> 189 542.5300   2.706063
#> 190 501.2824   3.171282
#> 191 540.4113   2.639886
#> 192 514.2939   3.339493
#> 193 523.2796   2.527243
#> 194 502.2124   3.000066
#> 195 501.4783   2.523839
#> 196 571.8807   3.209309
#> 197 565.0188   3.046998
#> 198 501.8140   3.685227
#> 199 505.6463   2.566301
#> 200 508.7170   2.929135
#> 201 537.5213   2.702595
#> 202 523.0591   3.308279
#> 203 545.0313   2.540716
#> 204 500.0396   3.075619
#> 205 828.0180   3.244160
#> 206 500.0137   2.911769
#> 207 500.0172   3.138330
#> 208 831.1655   2.643369
#> 209 735.6690   3.311393
#> 210 515.4732   2.747833
#> 211 500.8871   2.712494
#> 212 500.4412   2.718363
#> 213 502.1103   2.786371
#> 214 807.1462   2.749440
#> 215 500.2614   2.791432
#> 216 501.5007   3.874211
#> 217 500.3004   2.633019
#> 218 500.0752   2.584385
#> 219 500.1653   2.786717
#> 220 500.0282   2.879812
#> 221 503.1954   2.990836
#> 222 501.1436   2.885949
#> 223 500.0659   2.874155
#> 224 623.2061   2.767736
#> 225 501.7390   2.527640
#> 226 504.2969   2.748292
#> 227 501.9761   3.481641
#> 228 502.9146   3.045666
#> 229 500.5977   3.517034
#> 230 531.5984   2.623401
#> 231 500.0114   3.337043
#> 232 505.0737   2.775216
#> 233 500.3933   3.182582
#> 234 531.2380   2.732608
#> 235 501.9623   2.779236
#> 236 572.4093   2.561492
#> 237 500.3707   3.335092
#> 238 500.1553   3.256882
#> 239 501.1566   2.724883
#> 240 500.0666   2.597827
#> 241 500.1036   2.776385
#> 242 502.5444   2.834200
#> 243 501.8276   2.566625
#> 244 520.9298   2.698759
#> 245 500.4005   4.147337
#> 246 510.0576   3.023472
#> 247 500.1349   2.815549
#> 248 501.0253   3.888150
#> 249 557.4865   2.635094
#> 250 500.0367   2.936672
#> 251 520.4657   2.740402
#> 252 500.3096   2.561509
#> 253 500.7144   2.587384
#> 254 570.6330   2.722888
#> 255 501.9197   2.511968
#> 256 500.0227   2.729572
#> 257 500.0021   2.533122
#> 258 587.7803   3.204174
#> 259 500.3291   2.681996
#> 260 550.8401   2.732541
#> 261 500.4468   2.677253
#> 262 500.5123   2.668132
#> 263 500.7948   3.092416
#> 264 500.0792   2.551011
#> 265 635.6681   3.008545
#> 266 500.0067   2.778241
#> 267 508.7001   3.131604
#> 268 503.7659   2.881043
#> 269 509.3671   2.561616
#> 270 500.1154   3.345410
#> 271 502.2292   2.691532
#> 272 508.2077   2.735302
#> 273 698.1217   2.690314
#> 274 624.0530   2.868832
#> 275 500.7988   2.889486
#> 276 697.2742   2.949813
#> 277 503.8477   3.696728
#> 278 504.2262   2.891387
#> 279 500.0026   3.477895
#> 280 500.0541   3.103322
#> 281 500.4171   3.035055
#> 282 500.8777   3.088843
#> 283 545.1296   4.394409
#> 284 523.8497   2.534435
#> 285 549.8346   3.399862
#> 286 500.6164   3.105760
#> 287 599.5853   3.363975
#> 288 670.3397   2.905422
#> 289 500.8610   3.215342
#> 290 555.5947   3.214077
#> 291 500.3429   2.850507
#> 292 513.5660   2.529586
#> 293 600.9305   2.660873
#> 294 500.1125   2.598342
#> 295 554.8556   2.715280
#> 296 511.2787   2.694852
#> 297 500.6395   2.756935
#> 298 500.4437   2.891523
#> 299 500.0181   2.804036
#> 300 738.7731   2.684310
#> 301 510.1834   2.919049
#> 302 503.4723   2.912865
#> 303 512.2576   2.786427
#> 304 500.5440   3.019271
#> 305 500.0512   3.784213
#> 306 518.7908   2.677606
#> 307 508.6675   4.628683
#> 308 603.1642   2.713785
#> 309 501.0656   3.304042
#> 310 500.3142   3.063145
#> 311 504.5635   2.747556
#> 312 501.9087   2.633954
#> 313 506.7909   2.550713
#> 314 501.9457   2.561882
#> 315 611.9358   2.706321
#> 316 501.9513   2.957823
#> 317 500.1093   2.613599
#> 318 500.0627   2.543030
#> 319 515.9548   3.417047
#> 320 502.1971   2.997714
#> 321 503.0458   2.977918
#> 322 501.1207   2.650712
#> 323 537.5360   2.607853
#> 324 500.4941   3.766284
#> 325 525.1085   2.651764
#> 326 501.8775   3.604962
#> 327 706.1453   2.529350
#> 328 500.2120   2.921837
#> 329 504.5697   2.951324
#> 330 602.2321   4.381961
#> 331 500.0293   2.724218
#> 332 500.0424   2.602948
#> 333 500.3234   2.533520
#> 334 691.2251   2.767752
#> 335 508.5455   3.254335
#> 336 508.3846   2.784523
#> 337 501.6770   3.391565
#> 338 500.0077   2.634583
#> 339 564.4671   3.138910
#> 340 501.7718   3.372873
#> 341 502.7027   2.979798
#> 342 500.0365   2.535605
#> 343 719.7627   3.493571
#> 344 505.3423   2.931581
#> 345 502.1088   2.929258
#> 346 525.5726   3.067201
#> 347 500.0627   2.557267
#> 348 500.0255   2.850054
#> 349 500.3784   2.541155
#> 350 507.2101   2.710169
#> 351 500.1327   2.529974
#> 352 501.7451   2.573911
#> 353 500.0036   3.207441
#> 354 500.0286   2.716682
#> 355 500.3592   2.705305
#> 356 500.1389   2.562436
#> 357 503.3413   3.123936
#> 358 534.1040   2.640394
#> 359 501.9178   2.969450
#> 360 500.0411   2.707244
#> 361 738.3160   2.606317
#> 362 505.4734   2.513303
#> 363 500.9098   2.890397
#> 364 500.1988   2.871741
#> 365 500.1763   2.587500
#> 366 500.4015   3.181926
#> 367 500.0641   3.746636
#> 368 500.7582   4.569699
#> 369 531.9274   2.653711
#> 370 536.6097   3.452494
#> 371 500.5379   2.765056
#> 372 500.0486   2.555130
#> 373 609.2066   2.505174
#> 374 623.1074   4.377420
#> 375 500.0648   2.954050
#> 376 501.5740   2.639853
#> 377 513.7991   2.892601
#> 378 513.0445   2.830482
#> 379 925.8246   3.662875
#> 380 511.9733   2.547468
#> 381 500.0671   3.206128
#> 382 500.1778   2.833856
#> 383 500.1761   2.811835
#> 384 604.5823   3.196772
#> 385 541.0926   2.915952
#> 386 603.7970   2.516972
#> 387 524.7831   3.296781
#> 388 501.8381   2.597122
#> 389 500.0605   2.896930
#> 390 500.6320   2.634969
#> 391 501.1548   2.607599
#> 392 501.1136   2.503881
#> 393 777.4466   2.776612
#> 394 678.1517   3.075612
#> 395 502.3328   2.896285
#> 396 500.0767   2.910299
#> 397 979.2158   2.604169
#> 398 501.1418   2.941695
#> 399 502.1417   2.523381
#> 400 501.3639   3.259202
#> 401 562.7748   3.193977
#> 402 518.5912   2.716625
#> 403 500.0736   2.940512
#> 404 506.5816   2.696038
#> 405 982.4397   2.914037
#> 406 569.9834   2.669432
#> 407 500.0401   2.680262
#> 408 500.3597   3.169209
#> 409 501.1427   3.354163
#> 410 654.5831   2.868404
#> 411 566.0695   3.302925
#> 412 500.0341   2.747733
#> 413 948.1666   2.628351
#> 414 501.2986   2.724954
#> 415 500.2892   2.755887
#> 416 500.6647   2.771117
#> 417 515.3138   4.115066
#> 418 502.2292   2.664261
#> 419 501.2736   3.478291
#> 420 503.9922   3.152327
#> 421 502.5086   3.090652
#> 422 521.7965   2.525264
#> 423 500.0517   3.267899
#> 424 500.1220   2.569219
#> 425 500.7065   3.851800
#> 426 503.1209   2.504839
#> 427 500.1556   3.081687
#> 428 500.1616   3.010530
#> 429 501.6991   2.622083
#> 430 500.4495   2.535352
#> 431 516.4338   2.945143
#> 432 525.6391   2.651956
#> 433 515.1963   3.783095
#> 434 500.0406   2.597144
#> 435 643.1935   2.655410
#> 436 500.1854   2.760350
#> 437 515.0132   3.952989
#> 438 504.8937   3.010971
#> 439 507.8100   3.261879
#> 440 694.0786   3.579999
#> 441 501.3907   3.015297
#> 442 512.5544   2.886347
#> 443 878.5421   2.639900
#> 444 500.3142   2.798369
#> 445 500.3277   3.211519
#> 446 666.8068   3.201256
#> 447 517.5925   2.574067
#> 448 504.7794   2.918054
#> 449 500.0902   3.022739
#> 450 618.7714   2.940970
#> 451 503.8277   2.978145
#> 452 880.7054   3.232545
#> 453 506.0749   3.025742
#> 454 500.1040   2.591971
#> 455 580.1313   2.603595
#> 456 501.1794   2.907309
#> 457 536.1041   2.680753
#> 458 521.0124   2.624799
#> 459 648.0041   3.134058
#> 460 500.6324   2.544305
#> 461 500.1855   3.259614
#> 462 857.7218   3.173500
#> 463 500.2941   3.329105
#> 464 500.3205   3.467414
#> 465 522.5284   2.637741
#> 466 500.2738   2.544243
#> 467 513.6638   2.540311
#> 468 500.4378   2.730972
#> 469 500.2636   2.643769
#> 470 642.4259   2.616493
#> 471 500.5106   2.898550
#> 472 501.1924   2.775355
#> 473 500.2380   3.624234
#> 474 984.1075   2.662866
#> 475 505.3160   2.646938
#> 476 500.1046   2.635652
#> 477 511.6897   3.022238
#> 478 500.7490   2.835369
#> 479 509.8919   2.741463
#> 480 504.1267   2.691422
#> 481 504.6760   2.687196
#> 482 502.3153   2.532395
#> 483 500.3064   2.537984
#> 484 530.9849   3.471330
#> 485 500.1494   2.801967
#> 486 500.7753   2.507483
#> 487 501.0890   2.668983
#> 488 568.3895   2.936361
#> 489 758.9474   3.228573
#> 490 792.9699   2.677848
#> 491 797.2293   3.096071
#> 492 500.2210   2.557203
#> 493 500.0226   2.625056
#> 494 500.0700   2.852934
#> 495 502.8530   3.337877
#> 496 502.4576   3.199574
#> 497 501.0645   2.555032
#> 498 637.8559   2.787735
#> 499 500.0520   2.592361
#> 500 500.9205   2.523871
#> 501 501.1368   3.338704
#> 502 558.2056   3.488980
#> 503 508.9455   2.873104
#> 504 500.9060   3.086532
#> 505 553.2003   2.588514
#> 506 933.9995   2.807724
#> 507 500.1685   2.874839
#> 508 508.3428   3.147525
#> 509 545.9388   2.921879
#> 510 500.0348   2.690967
#> 511 508.0706   3.935573
#> 512 500.0314   2.605572
#> 513 774.9796   2.645916
#> 514 500.0428   2.784734
#> 515 507.9699   2.649227
#> 516 506.4242   3.357699
#> 517 539.1115   2.697183
#> 518 500.1247   2.607882
#> 519 527.6673   2.588789
#> 520 503.5970   2.628541
#> 521 506.7323   2.562205
#> 522 572.8831   3.582313
#> 523 527.9958   2.793947
#> 524 500.1435   2.688874
#> 525 515.6691   2.778228
#> 526 500.1084   2.806222
#> 527 502.7079   2.613560
#> 528 537.1445   3.233093
#> 529 616.5075   2.501683
#> 530 500.0146   3.344238
#> 531 501.1048   3.309352
#> 532 512.5657   3.362552
#> 533 505.3851   2.678808
#> 534 920.8541   3.273372
#> 535 526.3610   3.216602
#> 536 501.4442   3.268655
#> 537 501.4659   2.705878
#> 538 500.1878   2.513631
#> 539 501.1032   2.507341
#> 540 500.3168   3.024960
#> 541 501.9304   3.013425
#> 542 502.2232   2.643825
#> 543 500.0483   2.613898
#> 544 501.7102   3.080510
#> 545 500.3628   2.711886
#> 546 500.0796   2.578155
#> 547 547.6690   2.932016
#> 548 919.4448   2.515781
#> 549 522.3106   2.597713
#> 550 500.0457   2.732622
#> 551 500.7512   2.558031
#> 552 500.3401   2.665510
#> 553 507.0675   3.212410
#> 554 510.2510   3.254527
#> 555 522.9129   2.830592
#> 556 572.3649   2.695137
#> 557 522.5633   3.002230
#> 558 501.2618   2.638930
#> 559 718.3351   2.561339
#> 560 500.0166   2.580174
#> 561 500.2387   2.933879
#> 562 616.0289   2.759192
#> 563 514.3117   3.191441
#> 564 500.2825   2.609954
#> 565 501.1099   2.683275
#> 566 501.2739   3.192490
#> 567 501.0912   2.749843
#> 568 557.5589   2.847712
#> 569 501.2196   2.650905
#> 570 500.2545   2.678824
#> 571 500.0658   2.578768
#> 572 503.1899   3.207671
#> 573 640.8641   2.630796
#> 574 500.1996   3.452619
#> 575 516.1671   2.517384
#> 576 517.2663   2.715446
#> 577 503.9922   3.303422
#> 578 500.1562   3.297427
#> 579 501.1188   2.517746
#> 580 952.3828   3.479143
#> 581 500.2912   2.926392
#> 582 574.2879   3.451840
#> 583 521.3357   2.615223
#> 584 507.9092   2.599977
#> 585 500.0151   2.893542
#> 586 565.0262   2.545316
#> 587 523.5448   3.060718
#> 588 917.5947   2.602399
#> 589 500.6536   2.552447
#> 590 623.2269   2.524070
#> 591 820.2619   2.544682
#> 592 500.1495   2.989729
#> 593 784.4668   2.936056
#> 594 500.4729   2.686241
#> 595 876.8026   2.612712
#> 596 500.0118   3.118602
#> 597 566.9397   3.203783
#> 598 504.0580   2.681187
#> 599 507.0907   3.241695
#> 600 822.3393   2.598167
#> 601 601.3932   2.903705
#> 602 803.1871   2.994543
#> 603 503.0880   2.849740
#> 604 500.1334   2.602514
#> 605 512.2751   2.803106
#> 606 500.0298   2.723171
#> 607 675.1655   2.629358
#> 608 567.3164   3.004597
#> 609 575.7638   2.640145
#> 610 500.5547   2.751477
#> 611 500.0627   3.006893
#> 612 961.9139   2.865398
#> 613 502.8606   2.595574
#> 614 522.4431   2.853706
#> 615 998.2040   2.610724
#> 616 567.6021   3.681370
#> 617 500.3375   2.833830
#> 618 500.2611   2.530808
#> 619 500.0416   2.792668
#> 620 500.1647   3.384243
#> 621 550.5859   2.537920
#> 622 518.2121   3.168882
#> 623 500.5025   2.602377
#> 624 500.0336   3.631545
#> 625 834.8181   3.077730
#> 626 504.6227   2.894911
#> 627 500.0549   3.180690
#> 628 506.4626   3.072878
#> 629 500.5869   2.816798
#> 630 507.8480   2.767660
#> 631 566.2768   2.689426
#> 632 500.5978   2.600256
#> 633 500.7877   2.892150
#> 634 500.0834   2.535082
#> 635 523.7395   3.751394
#> 636 519.9330   2.616235
#> 637 553.6874   3.948961
#> 638 704.8852   3.362029
#> 639 500.1520   2.611654
#> 640 619.2100   2.594962
#> 641 685.0691   3.283512
#> 642 740.5325   2.704353
#> 643 502.3897   2.702723
#> 644 509.6406   2.610732
#> 645 570.7258   3.321710
#> 646 502.8185   3.040366
#> 647 500.2758   4.370758
#> 648 505.0202   3.098710
#> 649 511.1240   2.960447
#> 650 502.8313   3.880834
#> 651 500.2602   3.033203
#> 652 500.2372   2.813091
#> 653 500.6951   4.211871
#> 654 520.3717   2.687639
#> 655 518.2328   3.082993
#> 656 500.2537   3.102654
#> 657 500.5583   3.574745
#> 658 500.0130   2.695259
#> 659 500.2308   2.727212
#> 660 501.0039   2.596289
#> 661 529.0402   2.649950
#> 662 517.3846   3.727510
#> 663 505.5463   3.025046
#> 664 519.1497   3.926522
#> 665 534.5962   2.518389
#> 666 618.6984   3.033551
#> 667 501.6416   2.944238
#> 668 502.6105   2.857755
#> 669 677.3048   3.029900
#> 670 502.0191   2.660777
#> 671 500.2101   2.568147
#> 672 725.7329   2.534532
#> 673 502.8023   2.774919
#> 674 500.0299   2.811368
#> 675 501.3061   2.859078
#> 676 500.3095   2.833495
#> 677 500.0657   2.600783
#> 678 507.5200   3.265753
#> 679 508.7748   2.519316
#> 680 500.0660   3.478486
#> 681 503.0810   3.470369
#> 682 514.5035   2.740144
#> 683 500.2982   3.151639
#> 684 500.2897   3.033190
#> 685 853.2865   2.683298
#> 686 500.0051   2.550210
#> 687 504.8085   2.723128
#> 688 503.6364   3.103593
#> 689 503.0505   2.671866
#> 690 517.0288   3.699051
#> 691 500.8945   2.601986
#> 692 515.9813   2.554190
#> 693 500.4135   2.593286
#> 694 500.3196   2.907758
#> 695 500.1398   3.141080
#> 696 915.0686   2.575833
#> 697 500.4983   2.598394
#> 698 500.0020   3.537025
#> 699 500.1675   3.043353
#> 700 501.0446   2.524328
#> 701 500.0278   2.722055
#> 702 916.3778   2.601776
#> 703 502.5715   3.143231
#> 704 512.5273   2.753445
#> 705 500.2416   3.337291
#> 706 503.3464   2.543545
#> 707 599.2942   2.924159
#> 708 802.3115   2.701884
#> 709 505.6395   2.898326
#> 710 500.5216   3.238068
#> 711 522.5264   2.934315
#> 712 500.7285   2.899387
#> 713 500.0793   3.003935
#> 714 500.0486   3.275248
#> 715 621.6692   2.658028
#> 716 503.9188   2.577780
#> 717 500.5322   2.564153
#> 718 538.1902   2.512655
#> 719 508.9943   2.646769
#> 720 524.1425   2.517403
#> 721 569.8527   3.732561
#> 722 500.1702   2.630186
#> 723 546.0300   3.121983
#> 724 598.5621   2.962900
#> 725 553.3504   2.908307
#> 726 539.1512   4.306672
#> 727 510.0071   3.323968
#> 728 508.9157   3.303796
#> 729 567.7737   2.699238
#> 730 523.0494   2.546565
#> 731 532.7589   2.608403
#> 732 506.1900   2.711047
#> 733 500.4491   3.413418
#> 734 500.0769   2.503791
#> 735 503.9587   2.922558
#> 736 500.2484   2.622569
#> 737 500.0386   3.063532
#> 738 508.6466   3.003051
#> 739 601.7142   3.096823
#> 740 740.3507   2.517806
#> 741 503.5261   2.660391
#> 742 500.0049   3.035995
#> 743 501.8852   2.749212
#> 744 500.1189   2.954631
#> 745 507.2802   2.548271
#> 746 501.0128   3.630013
#> 747 500.2866   2.618053
#> 748 501.0353   3.179149
#> 749 500.8840   2.843723
#> 750 501.9857   3.460634
#> 751 682.9691   2.504129
#> 752 500.6789   2.639715
#> 753 583.1148   2.802316
#> 754 503.6031   2.784934
#> 755 500.4793   2.802934
#> 756 500.0541   3.239983
#> 757 500.3607   3.050840
#> 758 500.1238   3.273844
#> 759 506.8519   3.412552
#> 760 501.7678   2.770489
#> 761 614.8215   2.763883
#> 762 608.2717   2.588840
#> 763 516.8498   2.941441
#> 764 505.2582   2.762334
#> 765 505.9345   3.278142
#> 766 502.3792   3.057400
#> 767 500.0918   3.370416
#> 768 500.1945   3.085567
#> 769 784.7472   2.575916
#> 770 528.7832   2.541976
#> 771 500.0749   3.151744
#> 772 561.6757   2.562152
#> 773 500.7424   2.679408
#> 774 504.9947   2.551348
#> 775 500.1236   3.069222
#> 776 500.1131   3.094919
#> 777 501.0967   3.276450
#> 778 500.7426   2.734312
#> 779 527.1260   2.939016
#> 780 500.0318   3.715752
#> 781 500.5943   2.883730
#> 782 588.4038   2.693556
#> 783 549.2217   2.518834
#> 784 720.7627   2.921364
#> 785 501.8648   3.056498
#> 786 500.0081   3.026328
#> 787 583.0913   3.842182
#> 788 530.7371   2.713770
#> 789 521.8542   2.610008
#> 790 500.8623   2.633551
#> 791 500.2476   3.074675
#> 792 500.0321   2.804910
#> 793 545.4501   2.734058
#> 794 510.8319   5.276414
#> 795 500.0042   2.722700
#> 796 514.2362   3.104388
#> 797 502.5638   2.779874
#> 798 510.2036   3.282377
#> 799 515.6643   3.013907
#> 800 505.3317   3.099953
#> 801 708.1818   3.084934
#> 802 500.9111   2.647533
#> 803 560.5921   2.520900
#> 804 500.1266   2.611054
#> 805 500.4437   3.019761
#> 806 500.1467   4.706133
#> 807 501.1181   2.542392
#> 808 501.2044   2.566139
#> 809 503.2854   3.081917
#> 810 500.4926   2.660907
#> 811 916.5295   3.375725
#> 812 503.0860   2.559035
#> 813 826.5471   2.524520
#> 814 878.7122   2.740069
#> 815 536.2969   2.797870
#> 816 718.5037   3.052622
#> 817 500.3823   3.836785
#> 818 507.9619   2.834499
#> 819 548.9099   2.719926
#> 820 500.4462   3.601737
#> 821 502.3091   2.517540
#> 822 500.0022   2.683215
#> 823 500.1771   2.685123
#> 824 508.4629   2.730455
#> 825 500.4213   3.653103
#> 826 500.0197   2.801152
#> 827 500.6463   3.463858
#> 828 503.0337   4.140630
#> 829 506.8588   2.547070
#> 830 598.4849   2.681327
#> 831 910.0176   2.577649
#> 832 502.5218   2.757583
#> 833 570.0752   3.929945
#> 834 500.0630   2.828046
#> 835 507.9547   2.800741
#> 836 500.4829   2.545014
#> 837 501.2113   3.761297
#> 838 550.6101   2.593445
#> 839 500.1338   2.578400
#> 840 505.2875   2.731312
#> 841 687.9267   3.059384
#> 842 500.2015   2.644781
#> 843 505.0173   2.961620
#> 844 986.2770   2.960666
#> 845 935.0163   2.553575
#> 846 623.0268   2.556644
#> 847 523.0235   3.057592
#> 848 745.1510   2.507717
#> 849 506.5207   2.514859
#> 850 503.4028   3.061720
#> 851 544.0360   2.802686
#> 852 948.9872   3.020060
#> 853 739.9732   2.613749
#> 854 522.1689   3.079355
#> 855 534.9733   3.051247
#> 856 504.3265   2.631567
#> 857 502.4480   2.605749
#> 858 500.4831   2.687853
#> 859 536.1130   2.661425
#> 860 500.0285   2.711133
#> 861 938.6672   3.414812
#> 862 505.2341   4.316272
#> 863 500.1199   2.561320
#> 864 515.4340   2.683410
#> 865 510.2831   4.067668
#> 866 508.7206   2.808335
#> 867 962.6761   2.860230
#> 868 500.2034   3.116020
#> 869 500.8982   2.862771
#> 870 835.9822   2.532445
#> 871 500.0536   2.806987
#> 872 503.0086   3.252561
#> 873 505.4539   2.858692
#> 874 528.9122   2.948362
#> 875 504.2800   2.806626
#> 876 500.4079   2.818190
#> 877 605.9664   2.556309
#> 878 500.2891   2.716908
#> 879 534.1290   2.622543
#> 880 995.6008   2.546352
#> 881 567.5062   2.633357
#> 882 500.3283   2.503231
#> 883 519.9158   3.215479
#> 884 500.0030   3.245885
#> 885 520.1588   2.955168
#> 886 502.2854   2.633956
#> 887 502.8258   3.093260
#> 888 500.0991   2.539817
#> 889 500.2937   3.098203
#> 890 535.5519   2.694905