Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   692.8560   3.482771
#> 2   502.2555   2.566848
#> 3   502.4908   2.501906
#> 4   517.7249   3.252201
#> 5   500.2381   2.545595
#> 6   500.5759   2.578657
#> 7   503.8291   5.519194
#> 8   500.1425   3.166032
#> 9   522.0822   3.069194
#> 10  509.3959   3.265808
#> 11  500.8630   2.798576
#> 12  504.2969   2.709072
#> 13  507.1477   3.036430
#> 14  500.3339   3.083426
#> 15  500.4539   2.608861
#> 16  505.5930   2.623573
#> 17  500.6838   2.703386
#> 18  501.8578   3.115171
#> 19  501.1822   3.526336
#> 20  502.5906   2.836263
#> 21  502.3147   2.531143
#> 22  507.4834   3.121733
#> 23  501.2326   2.855500
#> 24  514.1918   2.673037
#> 25  506.3097   2.500007
#> 26  501.4800   2.886556
#> 27  503.2396   2.745458
#> 28  858.2083   3.089061
#> 29  500.0471   2.668329
#> 30  500.6264   2.573790
#> 31  500.4571   2.936252
#> 32  500.9436   2.511815
#> 33  501.3619   2.769035
#> 34  562.5850   2.572419
#> 35  503.8209   2.768739
#> 36  505.7708   2.704667
#> 37  500.1679   2.764843
#> 38  502.9357   3.379791
#> 39  506.7423   3.519595
#> 40  501.9115   3.030203
#> 41  502.3018   2.560034
#> 42  500.3032   2.857203
#> 43  569.7277   3.217348
#> 44  506.3696   2.601908
#> 45  728.6717   2.676617
#> 46  567.9068   2.504363
#> 47  500.2921   3.039739
#> 48  500.5236   2.515782
#> 49  500.0341   3.269454
#> 50  500.0600   2.909320
#> 51  547.1800   3.016548
#> 52  530.4593   3.029947
#> 53  511.0174   2.742025
#> 54  514.1566   2.665546
#> 55  521.7865   2.561561
#> 56  500.1758   2.787011
#> 57  500.2602   3.593593
#> 58  500.3188   3.040111
#> 59  503.2150   3.891982
#> 60  500.1528   2.503909
#> 61  500.9893   2.694544
#> 62  503.7059   3.152491
#> 63  500.0776   2.746154
#> 64  500.0393   2.836830
#> 65  500.1010   2.627204
#> 66  500.1791   3.054908
#> 67  500.7851   2.877966
#> 68  500.0489   2.970695
#> 69  500.0588   2.537394
#> 70  951.0335   2.816227
#> 71  500.2938   2.709089
#> 72  500.0102   2.597669
#> 73  501.0756   2.549027
#> 74  636.6902   3.125596
#> 75  576.3447   2.890419
#> 76  557.8079   2.748262
#> 77  500.9475   2.673728
#> 78  500.7420   2.940591
#> 79  921.5726   2.946826
#> 80  500.3274   2.553870
#> 81  520.6854   2.921831
#> 82  535.8908   2.679494
#> 83  682.7936   4.232277
#> 84  500.1077   3.728587
#> 85  500.0684   2.879015
#> 86  503.9315   2.502177
#> 87  695.5918   2.918406
#> 88  500.6447   2.553826
#> 89  506.5529   3.833781
#> 90  501.1195   2.543889
#> 91  533.1160   3.264493
#> 92  500.0361   2.659634
#> 93  910.2045   2.729268
#> 94  503.0514   2.694879
#> 95  501.2305   3.238583
#> 96  501.0678   2.564967
#> 97  505.3940   2.634885
#> 98  501.0978   2.505992
#> 99  758.6241   3.060226
#> 100 523.2507   2.891149
#> 101 502.7178   2.562758
#> 102 506.4990   2.786307
#> 103 502.2557   2.640363
#> 104 500.0314   2.796917
#> 105 964.2571   2.982961
#> 106 503.4061   2.768885
#> 107 500.2757   2.676386
#> 108 577.1427   2.618483
#> 109 500.4316   4.321098
#> 110 507.9159   3.089289
#> 111 519.1907   2.736422
#> 112 524.9584   2.722245
#> 113 509.3453   3.393927
#> 114 504.3871   2.938572
#> 115 526.8874   2.511020
#> 116 501.0045   2.950443
#> 117 500.7407   3.097849
#> 118 500.0741   3.353661
#> 119 513.4268   3.137542
#> 120 639.2124   3.528671
#> 121 647.8499   4.070070
#> 122 924.9259   3.588024
#> 123 504.0027   2.697656
#> 124 500.4449   2.904016
#> 125 500.4157   2.919688
#> 126 503.6251   3.426968
#> 127 500.0851   2.916891
#> 128 505.4975   3.219314
#> 129 518.9284   2.635908
#> 130 501.0826   2.770273
#> 131 500.7411   2.503507
#> 132 502.0095   3.130132
#> 133 500.2139   3.151934
#> 134 500.5512   2.618891
#> 135 524.6846   2.719179
#> 136 500.3944   3.004311
#> 137 500.0467   3.005091
#> 138 500.1934   2.994690
#> 139 602.7355   2.702833
#> 140 504.8486   2.525363
#> 141 871.8208   3.410621
#> 142 500.1323   2.586554
#> 143 500.0554   2.898236
#> 144 500.1850   2.717327
#> 145 511.4363   3.133795
#> 146 500.2924   3.858314
#> 147 507.4052   2.635633
#> 148 508.9262   2.653308
#> 149 569.7942   2.967897
#> 150 628.3178   2.949155
#> 151 506.1152   3.903109
#> 152 500.2981   3.111597
#> 153 560.6673   2.671587
#> 154 626.0240   3.331953
#> 155 501.0088   2.949326
#> 156 500.2144   3.236383
#> 157 918.9744   3.007267
#> 158 506.6617   2.630470
#> 159 500.1121   3.413757
#> 160 501.3608   2.851831
#> 161 533.8236   2.529252
#> 162 500.0616   2.512838
#> 163 695.0882   2.663714
#> 164 511.9422   2.602375
#> 165 914.1538   3.040777
#> 166 514.9893   2.717905
#> 167 550.1912   3.195784
#> 168 511.6781   2.719154
#> 169 500.3733   2.583484
#> 170 500.0444   2.907675
#> 171 503.6180   3.077336
#> 172 500.4321   2.564484
#> 173 500.0446   3.337056
#> 174 624.7769   3.457494
#> 175 504.1570   3.213153
#> 176 660.6904   2.563054
#> 177 500.1301   2.533755
#> 178 578.5340   3.067180
#> 179 626.4132   2.630979
#> 180 845.7151   3.926776
#> 181 500.0248   2.610444
#> 182 569.1274   5.824595
#> 183 502.7571   2.624438
#> 184 578.3537   2.779024
#> 185 517.1277   2.682575
#> 186 754.7485   2.664658
#> 187 503.3778   2.615516
#> 188 544.7251   2.622962
#> 189 500.0139   2.989446
#> 190 501.0154   3.380274
#> 191 500.2555   2.711076
#> 192 500.1311   2.708708
#> 193 555.7283   3.083254
#> 194 723.4117   2.886931
#> 195 544.9516   2.910138
#> 196 554.2469   2.907896
#> 197 500.0847   3.335669
#> 198 570.4263   2.518471
#> 199 500.0493   2.604341
#> 200 586.4245   3.337580
#> 201 507.7959   3.765589
#> 202 504.3704   2.605794
#> 203 525.0246   2.649953
#> 204 511.5678   2.526164
#> 205 500.0294   2.826523
#> 206 506.8035   2.575897
#> 207 501.2379   2.680436
#> 208 510.5852   2.504900
#> 209 500.1149   2.913858
#> 210 500.7577   3.487673
#> 211 500.0406   2.921997
#> 212 500.1032   3.057525
#> 213 928.8302   2.615990
#> 214 511.5404   3.082937
#> 215 669.7956   2.507755
#> 216 500.6005   2.527385
#> 217 507.1348   3.285009
#> 218 573.1736   2.552392
#> 219 500.2137   2.542358
#> 220 662.2210   2.631639
#> 221 511.1868   2.808630
#> 222 500.0813   2.610377
#> 223 752.8823   2.835569
#> 224 503.0907   2.620202
#> 225 504.6986   2.715844
#> 226 500.0370   3.478991
#> 227 507.2403   2.512117
#> 228 756.2760   2.784137
#> 229 890.6879   3.350075
#> 230 524.4409   2.552842
#> 231 500.0366   2.689212
#> 232 500.0613   2.641399
#> 233 500.1726   2.521803
#> 234 512.0720   3.131415
#> 235 511.1198   2.837960
#> 236 608.7793   2.812080
#> 237 509.5631   2.611826
#> 238 523.0039   2.604408
#> 239 502.7388   3.563701
#> 240 805.8272   2.916350
#> 241 807.1623   2.771302
#> 242 503.5624   2.888519
#> 243 651.6511   3.449870
#> 244 516.8373   3.120475
#> 245 953.0255   2.698090
#> 246 500.4308   2.604270
#> 247 501.3352   2.651287
#> 248 501.3767   2.766991
#> 249 501.0248   2.901175
#> 250 511.5617   2.945708
#> 251 515.7229   2.665176
#> 252 500.8090   2.598070
#> 253 509.1226   2.609887
#> 254 501.8028   3.327944
#> 255 500.1696   2.702437
#> 256 509.9161   3.332456
#> 257 534.4082   2.959252
#> 258 520.7244   2.534804
#> 259 958.6905   3.516550
#> 260 590.8749   2.519329
#> 261 513.1005   2.856974
#> 262 923.7993   2.740939
#> 263 517.1737   2.786935
#> 264 569.9033   3.685333
#> 265 500.0931   2.835464
#> 266 500.2486   2.639305
#> 267 510.1320   2.662677
#> 268 503.6650   3.045930
#> 269 501.0218   2.678634
#> 270 541.1244   3.050429
#> 271 501.8427   3.075798
#> 272 508.2428   2.645156
#> 273 500.3232   3.155624
#> 274 500.9436   2.653137
#> 275 543.0185   3.114962
#> 276 606.3889   2.642529
#> 277 540.5670   2.583674
#> 278 500.0021   2.750523
#> 279 703.4498   2.516806
#> 280 500.2392   3.329784
#> 281 500.0403   2.825394
#> 282 501.0798   2.789180
#> 283 500.0316   2.717398
#> 284 503.5236   2.534044
#> 285 508.1793   2.638917
#> 286 850.3201   2.709175
#> 287 501.7454   2.510982
#> 288 500.8884   2.805053
#> 289 506.9539   2.618416
#> 290 500.4901   2.606786
#> 291 916.3924   3.226676
#> 292 554.8365   2.998931
#> 293 500.2197   2.820455
#> 294 502.2337   2.639115
#> 295 513.3800   2.518516
#> 296 522.7928   3.518657
#> 297 502.5499   2.922682
#> 298 747.8913   2.905204
#> 299 546.8017   2.627530
#> 300 501.1678   2.518989
#> 301 906.4881   2.615909
#> 302 500.7473   2.645135
#> 303 509.5948   3.273806
#> 304 502.7474   2.749878
#> 305 844.0514   2.874209
#> 306 501.0963   2.932727
#> 307 500.0599   3.771183
#> 308 607.9135   2.592591
#> 309 635.3655   2.896911
#> 310 500.1982   3.305490
#> 311 500.1826   2.795284
#> 312 729.7313   3.044589
#> 313 519.4076   3.329279
#> 314 542.0776   3.189193
#> 315 504.3225   2.541419
#> 316 509.3680   2.546853
#> 317 506.9046   2.906226
#> 318 506.5593   2.709274
#> 319 500.2488   2.782488
#> 320 603.0082   2.879849
#> 321 799.3141   3.422379
#> 322 535.8773   2.637239
#> 323 506.2375   2.814927
#> 324 500.1749   2.550457
#> 325 500.8225   2.999173
#> 326 573.4863   2.703070
#> 327 502.0013   2.761429
#> 328 503.4886   2.728865
#> 329 638.8936   3.040453
#> 330 500.2307   3.720377
#> 331 512.0544   3.256119
#> 332 500.4710   2.662399
#> 333 509.7262   3.148660
#> 334 500.3737   2.687873
#> 335 512.7467   2.752973
#> 336 510.1892   2.815185
#> 337 500.0335   3.914495
#> 338 501.0396   3.164191
#> 339 500.4271   3.205084
#> 340 504.3612   2.864392
#> 341 529.1381   2.606583
#> 342 500.1804   3.033537
#> 343 502.2877   3.025186
#> 344 500.7540   2.653833
#> 345 502.8126   3.111842
#> 346 500.6559   2.793581
#> 347 779.1073   3.477234
#> 348 501.0189   2.684993
#> 349 500.1561   2.684832
#> 350 528.0172   3.039386
#> 351 500.2994   2.825357
#> 352 524.0242   2.979019
#> 353 504.8111   2.597937
#> 354 617.1454   3.265621
#> 355 500.1203   2.614331
#> 356 500.5309   2.660081
#> 357 598.0417   2.567236
#> 358 507.9107   2.814578
#> 359 500.6402   2.625631
#> 360 500.0233   2.920562
#> 361 586.7744   2.539134
#> 362 540.1947   3.276720
#> 363 501.7845   3.685973
#> 364 547.9815   2.686357
#> 365 502.8426   3.406734
#> 366 500.0349   2.606855
#> 367 501.2638   2.607150
#> 368 548.0047   2.694434
#> 369 543.1153   2.642237
#> 370 501.0585   2.679851
#> 371 500.3768   4.890732
#> 372 505.8036   4.074873
#> 373 510.8672   3.141101
#> 374 512.0641   2.676140
#> 375 512.0839   3.153628
#> 376 553.9221   2.659655
#> 377 504.0839   2.952114
#> 378 507.6637   2.575321
#> 379 500.3411   3.016683
#> 380 535.6026   2.503982
#> 381 597.8580   4.099629
#> 382 500.9435   2.624922
#> 383 500.7437   2.681376
#> 384 501.8491   3.463960
#> 385 500.1628   2.746486
#> 386 953.4001   2.653001
#> 387 502.0908   2.751692
#> 388 880.2521   2.898311
#> 389 516.5202   3.039224
#> 390 907.2995   2.832480
#> 391 500.2048   2.544710
#> 392 540.2656   3.152967
#> 393 531.4897   2.924695
#> 394 502.2063   2.795418
#> 395 925.1635   2.510516
#> 396 505.2051   2.513095
#> 397 550.4077   4.460278
#> 398 569.5329   2.675579
#> 399 500.9116   2.603346
#> 400 543.9344   3.017353
#> 401 511.5413   4.061486
#> 402 508.3472   3.125744
#> 403 583.4423   2.919641
#> 404 515.8201   2.549774
#> 405 519.6707   3.009535
#> 406 525.4416   3.047211
#> 407 501.2150   3.828650
#> 408 504.1477   4.864275
#> 409 589.8417   2.560807
#> 410 510.5240   3.243196
#> 411 500.1581   2.788927
#> 412 500.0861   2.528479
#> 413 529.8734   2.746237
#> 414 504.6246   2.517790
#> 415 502.5992   2.634877
#> 416 506.4815   3.492077
#> 417 525.7373   2.673934
#> 418 503.2031   2.518509
#> 419 500.3129   3.300223
#> 420 572.3348   2.734861
#> 421 500.2984   2.732254
#> 422 500.2017   2.699773
#> 423 516.2716   2.778621
#> 424 642.4790   2.531789
#> 425 502.4055   2.645837
#> 426 500.9115   2.591673
#> 427 500.1530   3.090348
#> 428 500.2551   2.697127
#> 429 504.4665   2.574255
#> 430 503.2390   3.028452
#> 431 516.3852   2.671493
#> 432 501.0303   2.973151
#> 433 505.7946   2.746094
#> 434 502.5785   2.776387
#> 435 509.4982   3.001465
#> 436 500.0306   3.118238
#> 437 500.3119   2.657586
#> 438 501.0078   2.547555
#> 439 588.5339   2.602254
#> 440 520.3085   3.112857
#> 441 510.8746   2.807125
#> 442 500.1811   3.668628
#> 443 503.3871   3.612919
#> 444 500.0526   4.213278
#> 445 500.0278   2.722196
#> 446 505.3635   2.688817
#> 447 674.4976   2.512405
#> 448 592.1259   3.566967
#> 449 518.0596   2.935565
#> 450 500.0531   2.714018
#> 451 500.6928   2.693365
#> 452 505.7328   2.520442
#> 453 500.6034   3.088711
#> 454 916.5901   2.935921
#> 455 500.6553   3.140643
#> 456 500.0906   2.660543
#> 457 537.3736   3.019404
#> 458 500.1621   2.580453
#> 459 502.2098   2.678950
#> 460 504.3295   2.533865
#> 461 514.2909   2.791379
#> 462 501.8633   3.372171
#> 463 501.0696   2.930444
#> 464 567.3128   2.832956
#> 465 510.8220   3.205171
#> 466 500.0845   3.369057
#> 467 507.8636   3.527264
#> 468 500.0840   2.764378
#> 469 503.1623   3.314352
#> 470 504.9424   2.975000
#> 471 500.1132   3.064775
#> 472 500.0257   2.673992
#> 473 500.5414   2.753780
#> 474 501.2137   3.433969
#> 475 528.6287   2.630682
#> 476 751.7028   3.488996
#> 477 500.3197   3.346962
#> 478 509.9868   2.630861
#> 479 511.6711   3.204323
#> 480 632.4466   3.046489
#> 481 500.3570   2.799976
#> 482 501.4381   2.959448
#> 483 500.4424   2.556150
#> 484 502.4682   2.708017
#> 485 506.9996   4.062020
#> 486 501.7420   3.353001
#> 487 500.0190   2.532097
#> 488 503.8767   2.996706
#> 489 527.7727   2.510382
#> 490 500.0887   2.638459
#> 491 837.6264   2.565602
#> 492 526.8094   2.819082
#> 493 500.1291   2.746813
#> 494 500.9260   2.641322
#> 495 796.8435   2.728839
#> 496 501.7603   2.960108
#> 497 520.8734   2.581239
#> 498 500.0064   3.088225
#> 499 513.4157   2.631205
#> 500 586.2235   3.087622
#> 501 500.4253   2.557170
#> 502 521.5448   2.543131
#> 503 502.2962   3.063406
#> 504 508.4876   2.590745
#> 505 500.0516   3.563740
#> 506 638.9434   2.649115
#> 507 501.3108   5.089313
#> 508 534.0155   3.080311
#> 509 500.5107   2.764858
#> 510 500.2196   2.659885
#> 511 504.4377   2.617775
#> 512 574.7309   2.787573
#> 513 604.5841   2.518097
#> 514 530.7815   2.805320
#> 515 754.7198   3.319232
#> 516 504.2025   2.804211
#> 517 520.5894   2.775850
#> 518 500.1150   3.092998
#> 519 510.4187   2.953107
#> 520 560.9234   2.662537
#> 521 679.4584   2.731176
#> 522 517.8153   3.061270
#> 523 500.2219   2.710781
#> 524 504.6053   2.851259
#> 525 500.0316   3.616399
#> 526 518.6802   2.680101
#> 527 523.1889   2.577989
#> 528 732.9005   3.951006
#> 529 515.8105   3.422537
#> 530 500.4284   3.838292
#> 531 537.8594   2.679064
#> 532 643.7868   2.614478
#> 533 502.7134   2.652905
#> 534 786.8822   2.903542
#> 535 500.2758   2.831889
#> 536 500.3365   2.663787
#> 537 500.0979   2.760257
#> 538 501.6741   2.929782
#> 539 527.8842   2.673426
#> 540 691.6115   3.304967
#> 541 500.0478   3.326553
#> 542 502.7514   2.619670
#> 543 538.7389   2.615789
#> 544 920.8773   2.995804
#> 545 500.9014   2.635079
#> 546 500.2517   2.598725
#> 547 500.0825   2.549596
#> 548 510.6870   2.692318
#> 549 500.4697   3.078166
#> 550 534.0163   3.017111
#> 551 501.0301   2.722936
#> 552 730.2283   2.598529
#> 553 500.1799   2.948459
#> 554 788.2780   2.699125
#> 555 503.0246   2.679005
#> 556 500.0111   2.878124
#> 557 500.2757   2.924783
#> 558 500.9441   3.653707
#> 559 507.1032   2.738473
#> 560 500.1463   2.619132
#> 561 507.2141   2.724412
#> 562 500.0069   2.887429
#> 563 500.5193   2.557675
#> 564 500.1140   2.983909
#> 565 507.9522   3.271160
#> 566 503.2599   2.916418
#> 567 545.7057   2.906357
#> 568 520.6998   3.184343
#> 569 579.2232   2.549835
#> 570 500.8618   2.700701
#> 571 512.8864   2.786471
#> 572 500.4797   3.535939
#> 573 512.9809   2.813363
#> 574 687.7197   2.911644
#> 575 501.5975   2.575604
#> 576 669.0566   3.393294
#> 577 500.0283   2.686069
#> 578 500.3456   2.500334
#> 579 526.7778   2.576450
#> 580 500.4373   2.686925
#> 581 941.4738   2.730409
#> 582 500.0215   2.932162
#> 583 517.8581   2.853926
#> 584 500.0712   2.534839
#> 585 646.3552   3.695450
#> 586 500.2957   2.651478
#> 587 500.2374   2.545509
#> 588 500.4400   3.014739
#> 589 501.1782   2.626486
#> 590 500.0292   2.575709
#> 591 704.3980   2.605000
#> 592 501.1011   2.860594
#> 593 508.3146   2.627709
#> 594 526.1157   2.654823
#> 595 502.2530   2.910845
#> 596 519.7879   3.075460
#> 597 500.9972   3.173691
#> 598 506.2954   2.895656
#> 599 502.0148   3.052941
#> 600 500.9749   3.045389
#> 601 537.8453   3.095324
#> 602 522.0441   2.725106
#> 603 632.4058   2.995132
#> 604 501.4653   3.269774
#> 605 519.1109   2.864999
#> 606 537.3017   2.713615
#> 607 500.6213   2.572821
#> 608 523.7875   2.864617
#> 609 500.0380   2.598160
#> 610 500.2686   3.211746
#> 611 557.7704   2.532187
#> 612 500.2383   2.624555
#> 613 500.2528   2.701358
#> 614 501.8479   2.650332
#> 615 500.6370   3.925584
#> 616 501.4284   3.377564
#> 617 501.1619   2.542242
#> 618 500.1366   3.074449
#> 619 561.1529   2.849338
#> 620 500.1705   2.698854
#> 621 554.9106   2.572126
#> 622 851.6985   3.289969
#> 623 504.2251   2.602935
#> 624 501.7851   2.638204
#> 625 500.0236   2.650767
#> 626 500.0704   2.906673
#> 627 500.1881   2.542377
#> 628 559.9736   2.890034
#> 629 500.2549   2.814797
#> 630 524.3961   2.703344
#> 631 619.0673   3.056457
#> 632 500.0410   2.502145
#> 633 501.2603   2.682258
#> 634 521.5958   2.680610
#> 635 503.7591   2.518104
#> 636 520.6542   2.760257
#> 637 597.6752   3.078969
#> 638 784.9123   2.656261
#> 639 577.5564   2.533826
#> 640 544.1425   3.420116
#> 641 500.5792   3.148602
#> 642 500.3463   2.802708
#> 643 501.3158   3.213243
#> 644 506.0369   3.146532
#> 645 501.0352   3.034741
#> 646 510.3147   3.410558
#> 647 502.2232   2.518651
#> 648 500.4995   3.197075
#> 649 500.0332   2.700809
#> 650 717.6679   3.087027
#> 651 500.0095   3.533585
#> 652 527.9356   2.809449
#> 653 500.0007   3.017500
#> 654 507.1892   2.596062
#> 655 542.9172   3.501079
#> 656 507.3776   3.286357
#> 657 500.7118   2.720648
#> 658 511.1998   2.517441
#> 659 517.9269   2.677606
#> 660 501.4556   3.589368
#> 661 590.9914   2.778520
#> 662 500.3535   3.143078
#> 663 509.1682   2.843308
#> 664 500.1591   2.592992
#> 665 503.3314   3.223196
#> 666 500.9437   3.507120
#> 667 500.0521   2.838546
#> 668 503.3335   3.051867
#> 669 500.1460   2.738981
#> 670 500.0879   2.725105
#> 671 544.0992   2.650544
#> 672 500.3132   3.069607
#> 673 500.0622   2.634907
#> 674 500.4787   2.545603
#> 675 505.1837   2.559541
#> 676 695.1933   3.101866
#> 677 507.9159   2.761424
#> 678 500.1310   3.161967
#> 679 500.9448   2.723901
#> 680 500.0016   4.029879
#> 681 500.5246   2.508323
#> 682 989.9057   3.489498
#> 683 500.5718   2.693084
#> 684 500.0616   2.527655
#> 685 501.4692   2.629357
#> 686 503.9649   3.089075
#> 687 500.3104   2.606380
#> 688 501.9901   3.000402
#> 689 522.4622   2.523537
#> 690 503.5499   3.084052
#> 691 500.1709   2.710649
#> 692 500.9930   3.052125
#> 693 500.0161   3.043344
#> 694 500.1018   2.536193
#> 695 500.1310   2.727398
#> 696 500.5375   3.236484
#> 697 500.3325   3.263869
#> 698 596.6621   2.507911
#> 699 509.9066   3.097720
#> 700 500.1551   3.130484
#> 701 721.4631   3.260529
#> 702 501.7500   2.742473
#> 703 901.7866   4.338461
#> 704 501.7225   3.803666
#> 705 513.5681   2.960282
#> 706 505.7557   2.852551
#> 707 503.9339   3.578964
#> 708 501.6728   3.830872
#> 709 500.1168   2.747082
#> 710 827.4067   2.533993
#> 711 500.1837   2.869441
#> 712 511.1972   2.566009
#> 713 514.3711   2.585529
#> 714 500.0376   2.958041
#> 715 500.6289   2.939800
#> 716 500.0241   2.689158
#> 717 564.1232   2.684734
#> 718 500.0465   3.023083
#> 719 500.0557   2.561807
#> 720 500.5041   2.831865
#> 721 500.2987   3.068023
#> 722 505.2345   2.500392
#> 723 500.6591   2.704776
#> 724 523.4496   2.568742
#> 725 500.3990   2.580053
#> 726 500.3697   2.665589
#> 727 500.4123   2.830116
#> 728 500.1176   2.914672
#> 729 500.3337   2.566352
#> 730 510.8614   2.784882
#> 731 510.6953   2.950991
#> 732 501.8707   3.209882
#> 733 500.0855   2.685147
#> 734 503.6035   3.286959
#> 735 519.1224   2.894698
#> 736 516.3545   3.174071
#> 737 500.4574   2.657270
#> 738 505.6254   2.584351
#> 739 723.8259   2.718360
#> 740 701.7241   2.588807
#> 741 500.7540   2.547738
#> 742 500.0954   2.500640
#> 743 502.1224   2.562076
#> 744 500.0623   2.707439
#> 745 558.7339   2.631729
#> 746 500.0777   2.651571
#> 747 889.1727   3.462106
#> 748 551.5508   3.240556
#> 749 516.1679   2.632764
#> 750 500.3552   3.112997
#> 751 509.0579   2.726164
#> 752 500.0067   2.598200
#> 753 501.2819   2.734215
#> 754 505.0680   2.808375
#> 755 525.1934   2.587646
#> 756 950.7062   3.196431
#> 757 882.6316   2.626138
#> 758 500.1623   2.626728
#> 759 500.0474   2.500469
#> 760 500.0257   2.611976
#> 761 548.4637   3.205333
#> 762 694.7750   2.544600
#> 763 500.1075   2.945270
#> 764 858.2900   2.530389
#> 765 500.1259   2.701308
#> 766 500.1071   3.107207
#> 767 692.9106   2.837632
#> 768 507.8151   2.519908
#> 769 809.0791   2.558697
#> 770 681.3448   4.438245
#> 771 500.2844   3.350937
#> 772 500.0674   3.817688
#> 773 681.1775   4.221455
#> 774 500.1657   2.542746
#> 775 500.1170   2.930874
#> 776 528.9897   2.581294
#> 777 502.4069   2.909960
#> 778 549.0765   2.514905
#> 779 553.0187   2.789524
#> 780 500.1229   2.992038
#> 781 500.2703   2.563946
#> 782 675.8671   3.106749
#> 783 500.0816   2.703766
#> 784 507.4130   2.510701
#> 785 518.1070   3.034670
#> 786 526.8028   2.709199
#> 787 576.2516   2.530472
#> 788 526.3690   2.837350
#> 789 512.9450   2.746185
#> 790 658.0347   2.672481
#> 791 575.1904   2.661626
#> 792 501.6334   2.536877
#> 793 788.0999   2.581151
#> 794 501.6694   2.660322
#> 795 507.4286   2.917616
#> 796 836.3574   2.911468
#> 797 511.0186   3.384254
#> 798 508.4967   2.719330
#> 799 501.1658   2.860647
#> 800 505.5816   2.506595
#> 801 501.0783   3.662467
#> 802 669.5002   2.557270
#> 803 589.5469   3.037822
#> 804 945.0598   2.661102
#> 805 500.3734   2.969553
#> 806 524.7464   4.373810
#> 807 501.4814   3.092125
#> 808 500.0017   2.695321
#> 809 521.0570   2.518343
#> 810 654.2278   2.507416
#> 811 500.0723   4.138625
#> 812 757.1260   2.917830
#> 813 501.9166   2.682802
#> 814 504.5106   2.681485
#> 815 509.7009   2.594272
#> 816 546.5738   2.926578
#> 817 500.0727   2.640086
#> 818 507.0843   2.539111
#> 819 500.2041   2.656052
#> 820 502.0182   4.186475
#> 821 832.5494   2.627351
#> 822 518.7969   2.658089
#> 823 521.9352   2.894946
#> 824 500.8213   2.670092
#> 825 505.0829   2.569043
#> 826 501.4806   2.586507
#> 827 503.6830   2.633928
#> 828 500.1851   3.011138
#> 829 532.9936   3.178925
#> 830 500.5084   3.322475
#> 831 500.2008   2.554193
#> 832 514.9557   2.544245
#> 833 500.1361   2.591177
#> 834 531.6003   2.608621
#> 835 531.9139   2.515045
#> 836 500.0120   2.519339
#> 837 500.2100   2.964787
#> 838 506.5653   3.147595
#> 839 561.1871   3.329701
#> 840 501.0439   2.651381
#> 841 541.0155   3.171856
#> 842 525.8836   2.978801
#> 843 714.0430   2.764734
#> 844 501.0762   2.544568
#> 845 507.4415   2.652590
#> 846 501.2051   3.106410
#> 847 504.6161   2.818339
#> 848 682.2844   2.600889
#> 849 589.6426   2.917812
#> 850 500.0441   3.115090
#> 851 689.8923   2.810681
#> 852 500.4790   2.642648
#> 853 510.0667   2.627031
#> 854 500.7981   2.514572
#> 855 504.2573   3.219730
#> 856 515.9446   2.811090
#> 857 516.3931   2.850436
#> 858 500.6004   2.598841
#> 859 879.0326   2.922325
#> 860 500.9601   2.540281
#> 861 502.6669   2.511344
#> 862 505.8720   3.041680
#> 863 500.1988   2.560125
#> 864 500.0750   3.134116
#> 865 766.0828   3.334966
#> 866 515.3621   2.952817
#> 867 500.0244   2.693886
#> 868 513.6397   2.802618
#> 869 500.1149   2.616233
#> 870 500.3167   2.731590
#> 871 500.9993   4.577026
#> 872 500.2599   2.537189
#> 873 500.7864   2.599912
#> 874 535.1310   3.034922
#> 875 708.0734   2.770404
#> 876 517.1365   2.509635
#> 877 504.5654   3.052435
#> 878 502.0389   2.541651
#> 879 670.5507   3.087490
#> 880 500.4244   2.984707
#> 881 501.5256   3.267580
#> 882 544.6724   2.705431
#> 883 521.8466   2.688140
#> 884 506.0869   3.444738
#> 885 500.8169   2.755285