Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   512.4664   2.736108
#> 2   519.0201   2.537164
#> 3   500.0333   2.586238
#> 4   500.1273   2.802353
#> 5   507.2709   2.932178
#> 6   862.2705   2.538528
#> 7   500.5284   3.045166
#> 8   503.5005   2.945514
#> 9   500.3837   3.134809
#> 10  503.5151   3.229354
#> 11  500.4652   2.534537
#> 12  524.3894   2.506639
#> 13  503.2587   2.606511
#> 14  789.2343   3.565005
#> 15  500.0364   2.555299
#> 16  500.1437   3.527823
#> 17  500.2259   3.328395
#> 18  500.0142   4.613453
#> 19  500.8157   2.740250
#> 20  501.8670   4.180919
#> 21  500.0124   3.514225
#> 22  513.8797   2.532119
#> 23  500.3098   2.756242
#> 24  501.4530   4.366907
#> 25  648.9013   3.547956
#> 26  506.8813   3.506939
#> 27  500.1904   3.210551
#> 28  594.2964   2.552428
#> 29  500.7026   3.316676
#> 30  508.9273   2.583995
#> 31  500.0357   2.557956
#> 32  511.7013   2.619722
#> 33  871.6978   2.757683
#> 34  507.6162   2.845684
#> 35  915.4850   3.818218
#> 36  500.6043   2.714240
#> 37  501.0153   2.530974
#> 38  500.7433   2.958893
#> 39  500.2109   2.678240
#> 40  503.3688   3.066431
#> 41  509.5137   5.247074
#> 42  788.3738   2.586563
#> 43  500.1605   2.536146
#> 44  502.2047   2.913241
#> 45  500.8800   2.712814
#> 46  507.8817   2.520527
#> 47  521.0201   2.966329
#> 48  550.9221   2.836408
#> 49  520.5979   2.925351
#> 50  530.8933   2.730229
#> 51  980.4359   2.582319
#> 52  720.0205   4.322473
#> 53  505.0956   3.203895
#> 54  578.6996   3.069359
#> 55  501.3255   2.812698
#> 56  503.8007   2.552214
#> 57  500.0606   2.724421
#> 58  500.1157   2.601860
#> 59  500.0234   2.917448
#> 60  609.3861   2.919694
#> 61  912.2003   2.523824
#> 62  500.1083   2.596802
#> 63  505.4934   2.546306
#> 64  500.0628   3.313009
#> 65  566.2461   2.512096
#> 66  500.7481   2.600667
#> 67  616.0516   3.295588
#> 68  501.1015   2.618491
#> 69  500.4742   3.253606
#> 70  500.2447   2.529097
#> 71  613.5494   2.519782
#> 72  504.5752   3.203336
#> 73  500.0984   2.806602
#> 74  865.2280   3.886713
#> 75  500.2498   2.894718
#> 76  500.0318   3.116101
#> 77  511.8096   2.906374
#> 78  500.1302   2.861667
#> 79  510.2537   2.595497
#> 80  606.9907   2.573292
#> 81  953.3600   3.759893
#> 82  500.2870   3.074577
#> 83  543.4551   2.811965
#> 84  541.8083   2.603487
#> 85  541.1411   2.616353
#> 86  500.9135   2.525597
#> 87  500.1747   3.080554
#> 88  500.1849   2.607987
#> 89  500.0439   2.559544
#> 90  501.6325   3.248197
#> 91  507.0968   2.589954
#> 92  520.0516   3.028580
#> 93  518.8452   2.794310
#> 94  788.1171   3.140771
#> 95  501.4649   2.549122
#> 96  653.3795   2.675049
#> 97  544.6592   2.542289
#> 98  506.3425   2.771379
#> 99  537.3022   2.787334
#> 100 500.0204   3.372229
#> 101 511.6596   3.703490
#> 102 500.1044   2.710583
#> 103 500.3614   3.141982
#> 104 501.9952   2.754490
#> 105 557.9490   4.133289
#> 106 746.8908   2.876921
#> 107 582.6355   3.915478
#> 108 731.1006   2.908321
#> 109 500.9617   2.604800
#> 110 537.0700   2.795841
#> 111 504.0582   2.803570
#> 112 516.9171   2.538555
#> 113 968.9757   2.823156
#> 114 508.4893   3.183861
#> 115 519.7403   3.044914
#> 116 539.0367   2.625652
#> 117 551.8499   3.238495
#> 118 660.1368   3.523908
#> 119 500.1136   2.598102
#> 120 500.1971   3.502174
#> 121 500.0575   2.648505
#> 122 740.1280   2.695607
#> 123 533.3701   3.826006
#> 124 500.5607   2.703244
#> 125 511.4402   2.836802
#> 126 516.9145   4.025205
#> 127 500.3116   3.180004
#> 128 843.8067   2.608336
#> 129 502.3235   2.769079
#> 130 781.8403   3.250788
#> 131 503.8787   2.671851
#> 132 501.0688   2.717919
#> 133 621.5648   3.028316
#> 134 500.7821   2.555334
#> 135 500.1782   2.600069
#> 136 502.4769   2.664058
#> 137 570.3775   3.722462
#> 138 508.0974   2.836965
#> 139 500.0405   2.518090
#> 140 525.0614   2.611823
#> 141 504.8758   2.711191
#> 142 503.0636   2.556013
#> 143 503.1631   3.748528
#> 144 519.8405   2.524838
#> 145 500.0174   3.086064
#> 146 500.0106   2.828870
#> 147 599.4036   2.693691
#> 148 500.5134   2.553636
#> 149 605.0623   3.037473
#> 150 501.4825   2.531854
#> 151 516.4059   3.006147
#> 152 575.0234   3.016363
#> 153 541.9896   2.683024
#> 154 529.0804   2.768129
#> 155 501.3363   3.295356
#> 156 500.3983   2.552901
#> 157 500.1365   2.659912
#> 158 707.3292   3.053364
#> 159 501.4026   3.587372
#> 160 504.0882   2.976885
#> 161 507.7132   2.773418
#> 162 502.2573   2.701595
#> 163 500.5026   2.805253
#> 164 519.8179   3.516533
#> 165 500.6129   2.879646
#> 166 500.8797   3.407411
#> 167 500.3064   2.652785
#> 168 501.9402   3.039205
#> 169 500.1223   3.896113
#> 170 504.0452   3.480836
#> 171 500.1597   2.726559
#> 172 572.7643   2.522131
#> 173 561.8074   2.660041
#> 174 500.1542   2.671674
#> 175 509.5170   3.355210
#> 176 500.0829   2.604964
#> 177 500.0569   3.423941
#> 178 563.9651   3.672615
#> 179 501.6351   2.737142
#> 180 508.0185   2.588739
#> 181 574.3871   2.626873
#> 182 500.3907   2.857168
#> 183 517.0358   2.983534
#> 184 507.9553   2.837023
#> 185 500.7040   4.232086
#> 186 500.1517   3.427563
#> 187 500.1684   4.477247
#> 188 501.7630   2.818463
#> 189 506.0230   2.865865
#> 190 500.2445   2.718680
#> 191 824.5354   2.554942
#> 192 531.1020   2.746570
#> 193 517.5198   2.880081
#> 194 505.6368   4.297034
#> 195 633.7449   2.714773
#> 196 569.9690   2.889152
#> 197 500.0321   2.690819
#> 198 504.7523   2.742953
#> 199 504.9182   2.938578
#> 200 500.0319   2.560771
#> 201 500.0611   2.520577
#> 202 502.1553   2.639472
#> 203 500.1168   2.863782
#> 204 500.5594   2.597156
#> 205 500.2427   2.552978
#> 206 512.5702   3.879353
#> 207 515.3167   3.254510
#> 208 500.5294   2.784650
#> 209 521.6415   2.869478
#> 210 502.0221   2.639069
#> 211 630.3519   2.989350
#> 212 574.7839   2.756823
#> 213 507.2516   2.827576
#> 214 500.0817   2.638504
#> 215 500.5362   2.659431
#> 216 501.8586   2.799986
#> 217 500.3147   3.137240
#> 218 509.2732   3.229642
#> 219 535.1566   3.138552
#> 220 576.2166   2.715011
#> 221 500.5865   2.694622
#> 222 706.3756   3.074128
#> 223 500.1002   2.978617
#> 224 501.6513   2.523531
#> 225 727.4954   2.865139
#> 226 507.4403   2.902929
#> 227 942.4222   3.013931
#> 228 500.0306   3.395403
#> 229 808.9084   2.575349
#> 230 624.9066   3.006451
#> 231 651.2684   2.503187
#> 232 500.0918   2.652162
#> 233 517.9084   2.597580
#> 234 577.4853   3.575180
#> 235 500.1121   2.742390
#> 236 502.7515   2.507052
#> 237 500.5308   3.533887
#> 238 500.2481   2.807747
#> 239 526.0367   2.572972
#> 240 501.9476   2.777871
#> 241 501.9880   2.633545
#> 242 500.3868   3.151613
#> 243 500.1838   2.829063
#> 244 522.1853   2.948156
#> 245 512.9344   2.784557
#> 246 613.9305   3.433163
#> 247 705.4565   3.401830
#> 248 500.1665   2.885871
#> 249 506.0123   3.708619
#> 250 532.9380   3.057248
#> 251 860.3463   2.893322
#> 252 501.0364   3.350117
#> 253 500.1648   3.241794
#> 254 512.9332   2.680492
#> 255 517.3950   2.620502
#> 256 501.3112   2.694222
#> 257 785.0831   3.018463
#> 258 649.0130   3.006517
#> 259 971.6353   2.741672
#> 260 500.0891   3.214263
#> 261 513.2397   2.532119
#> 262 514.3947   2.508991
#> 263 642.6657   3.792862
#> 264 524.0205   2.667701
#> 265 572.3089   2.958511
#> 266 660.6095   2.975720
#> 267 505.5871   2.550864
#> 268 500.0206   2.800511
#> 269 712.5104   3.436701
#> 270 500.0673   2.622435
#> 271 500.2566   2.879902
#> 272 574.9414   2.968202
#> 273 513.6303   2.890529
#> 274 622.6526   2.513867
#> 275 500.1007   2.953050
#> 276 512.9242   2.554722
#> 277 501.6992   2.727348
#> 278 512.1857   2.530287
#> 279 532.2323   2.898205
#> 280 667.4105   2.611434
#> 281 500.1574   2.577562
#> 282 501.0752   2.913966
#> 283 506.3069   2.500462
#> 284 500.2084   3.602728
#> 285 501.0919   3.084240
#> 286 572.3260   2.758853
#> 287 500.1119   2.599052
#> 288 533.0329   2.548688
#> 289 500.8248   2.963853
#> 290 501.6221   2.565784
#> 291 501.1654   2.612562
#> 292 540.5650   3.002250
#> 293 506.1579   2.990491
#> 294 504.1438   3.059550
#> 295 507.5651   2.531607
#> 296 500.0375   3.544437
#> 297 509.0974   2.607041
#> 298 500.2321   2.994497
#> 299 501.3925   2.656060
#> 300 505.8596   3.626285
#> 301 500.1330   2.753358
#> 302 500.0151   2.703004
#> 303 500.1217   3.222598
#> 304 500.1511   3.304090
#> 305 518.8319   3.278396
#> 306 503.3552   3.389983
#> 307 616.1672   3.448857
#> 308 500.3056   3.201902
#> 309 501.6358   3.284188
#> 310 500.1159   3.504542
#> 311 500.0197   2.508001
#> 312 500.1013   3.378681
#> 313 503.2310   2.609105
#> 314 500.0757   2.756305
#> 315 550.8219   2.511227
#> 316 818.7868   2.706089
#> 317 501.0815   2.908730
#> 318 505.7620   2.944650
#> 319 500.4063   2.644601
#> 320 500.0758   2.949790
#> 321 523.1655   2.675815
#> 322 501.3563   4.228357
#> 323 511.9859   3.307593
#> 324 500.0148   2.543280
#> 325 541.8192   2.663997
#> 326 500.7359   2.808877
#> 327 500.2892   2.586770
#> 328 500.9830   2.579845
#> 329 714.2553   3.041894
#> 330 501.6050   2.705277
#> 331 578.0095   2.618527
#> 332 688.1156   3.485148
#> 333 512.4274   2.578424
#> 334 510.3639   3.074890
#> 335 500.2718   3.165280
#> 336 522.1609   2.570173
#> 337 500.0010   3.100503
#> 338 500.0306   2.595227
#> 339 512.0236   4.246684
#> 340 543.0375   2.957524
#> 341 512.7825   2.576700
#> 342 500.0149   2.896123
#> 343 518.9608   2.993428
#> 344 501.9747   2.501915
#> 345 501.8413   2.701405
#> 346 593.1291   2.620368
#> 347 507.6066   2.722238
#> 348 514.9323   2.644876
#> 349 621.6375   2.554038
#> 350 505.9858   2.932006
#> 351 565.2265   3.023863
#> 352 500.7837   2.756513
#> 353 656.9345   3.199042
#> 354 500.0159   2.933421
#> 355 500.5263   2.770723
#> 356 500.0145   3.136145
#> 357 502.9749   3.782749
#> 358 613.9427   2.991759
#> 359 500.2487   2.734719
#> 360 979.6486   2.629258
#> 361 500.1740   2.594041
#> 362 600.5657   3.013766
#> 363 686.6675   2.539080
#> 364 539.8916   2.501725
#> 365 500.6999   2.990992
#> 366 501.4902   2.511131
#> 367 501.2516   2.552362
#> 368 505.1014   2.941072
#> 369 559.0307   3.890488
#> 370 630.9989   2.819448
#> 371 500.0748   2.919893
#> 372 559.5247   2.589435
#> 373 501.1285   3.247848
#> 374 500.4845   3.270495
#> 375 571.2390   3.004106
#> 376 500.2810   3.820962
#> 377 505.6184   2.659606
#> 378 501.1308   3.569581
#> 379 512.5066   3.083899
#> 380 715.7282   5.108705
#> 381 597.1098   3.207600
#> 382 711.9674   3.465537
#> 383 553.0262   4.460385
#> 384 500.4418   2.816961
#> 385 500.1557   2.833843
#> 386 500.2188   2.591511
#> 387 500.9014   2.515539
#> 388 621.9102   3.234738
#> 389 500.6806   2.920005
#> 390 546.7539   2.991036
#> 391 502.1741   2.651830
#> 392 518.6261   2.960205
#> 393 626.0366   2.529191
#> 394 500.1974   2.642087
#> 395 521.6742   3.370134
#> 396 503.5782   3.209062
#> 397 567.4630   2.634004
#> 398 500.0574   2.572130
#> 399 500.0810   3.427900
#> 400 501.5662   4.050275
#> 401 513.0464   2.628506
#> 402 522.6785   2.798078
#> 403 547.6862   2.714207
#> 404 500.3612   3.023092
#> 405 572.9022   2.738041
#> 406 502.0258   2.541431
#> 407 500.6526   3.165867
#> 408 502.7374   2.572975
#> 409 515.4304   2.888572
#> 410 721.0990   2.833870
#> 411 500.0762   2.548074
#> 412 500.5854   2.907311
#> 413 500.4545   3.651246
#> 414 502.5909   3.106568
#> 415 503.5893   3.640991
#> 416 500.8840   2.757121
#> 417 501.1377   2.955167
#> 418 514.3634   2.687649
#> 419 506.1972   2.822728
#> 420 720.0382   2.612269
#> 421 500.8994   2.552651
#> 422 500.2758   3.048000
#> 423 501.0162   2.597964
#> 424 500.7976   2.670001
#> 425 500.0398   2.551217
#> 426 500.4812   2.888736
#> 427 501.9661   2.502226
#> 428 682.8489   2.546613
#> 429 534.9124   2.720600
#> 430 502.9776   2.670476
#> 431 500.0245   2.579417
#> 432 500.8141   3.154003
#> 433 500.2683   2.561656
#> 434 505.6180   2.665073
#> 435 502.1372   2.609189
#> 436 500.0464   3.020572
#> 437 501.4846   2.804965
#> 438 500.0202   2.623549
#> 439 501.9473   2.597148
#> 440 535.0451   2.868786
#> 441 510.9957   2.618456
#> 442 543.3004   2.755913
#> 443 500.1063   2.804890
#> 444 546.3783   3.003154
#> 445 723.9613   3.289438
#> 446 500.0174   2.606833
#> 447 500.4197   2.600838
#> 448 530.5386   3.073358
#> 449 536.4592   2.528233
#> 450 500.0629   2.756696
#> 451 500.0527   2.650048
#> 452 509.2985   2.791916
#> 453 503.0599   3.928423
#> 454 593.7310   3.554786
#> 455 501.1885   3.147413
#> 456 507.7933   2.969936
#> 457 500.6085   2.649274
#> 458 792.8320   2.667242
#> 459 558.4019   2.749521
#> 460 501.6308   3.565703
#> 461 500.6616   2.514297
#> 462 515.6765   2.981202
#> 463 500.6296   2.767467
#> 464 794.1493   2.739743
#> 465 896.1648   4.172725
#> 466 505.3507   2.755098
#> 467 581.6043   2.606342
#> 468 500.4068   3.130918
#> 469 765.4973   2.837503
#> 470 501.8484   2.538326
#> 471 500.0568   2.975070
#> 472 527.4201   2.542529
#> 473 501.0460   2.737163
#> 474 503.0286   2.828089
#> 475 548.5576   2.604619
#> 476 509.2552   2.825613
#> 477 500.0511   2.973319
#> 478 508.2520   2.960743
#> 479 500.8879   2.828986
#> 480 694.2138   2.656775
#> 481 501.6240   2.732368
#> 482 511.1003   2.891196
#> 483 500.0853   2.522359
#> 484 506.3078   2.567659
#> 485 500.0829   3.186647
#> 486 500.4276   2.563303
#> 487 501.9236   2.545389
#> 488 525.9977   3.325618
#> 489 503.5271   2.662671
#> 490 505.6082   2.605274
#> 491 500.0030   2.617233
#> 492 539.1837   2.692684
#> 493 502.9461   2.753981
#> 494 500.0146   2.638819
#> 495 525.0705   2.760269
#> 496 924.2820   2.537658
#> 497 538.9678   3.048776
#> 498 500.2306   3.201822
#> 499 500.4117   3.014582
#> 500 501.5259   3.669235
#> 501 500.0624   2.778688
#> 502 500.1850   3.666527
#> 503 611.0639   2.766994
#> 504 501.2892   3.375273
#> 505 747.9044   2.898179
#> 506 500.1995   3.132452
#> 507 531.8875   2.681866
#> 508 500.2734   2.591760
#> 509 501.4946   3.111453
#> 510 761.7912   2.728533
#> 511 500.7633   2.526987
#> 512 501.3026   2.558724
#> 513 509.8215   2.747414
#> 514 500.1288   2.701109
#> 515 692.2576   2.775226
#> 516 500.2311   3.030726
#> 517 507.9098   3.908071
#> 518 500.9844   3.197624
#> 519 500.1181   2.752304
#> 520 500.9437   2.914865
#> 521 500.2728   3.335071
#> 522 501.1848   2.816177
#> 523 503.7968   3.246350
#> 524 900.3847   2.742151
#> 525 520.0300   2.703149
#> 526 500.8527   2.612963
#> 527 504.7936   2.549245
#> 528 500.5920   2.593338
#> 529 509.5223   3.368631
#> 530 500.2993   4.822635
#> 531 522.7301   2.733827
#> 532 508.2871   2.635419
#> 533 500.5516   3.808392
#> 534 500.6195   3.289034
#> 535 523.3387   2.504182
#> 536 500.5593   3.760306
#> 537 500.1190   3.533501
#> 538 502.1890   2.531554
#> 539 785.9675   2.556441
#> 540 533.1507   3.202190
#> 541 500.0473   2.843244
#> 542 622.4617   2.579415
#> 543 506.5306   2.524441
#> 544 500.5143   2.601970
#> 545 505.2507   2.609435
#> 546 510.7006   2.658453
#> 547 502.9268   2.503132
#> 548 548.0220   2.670831
#> 549 512.7094   2.657327
#> 550 500.0142   2.611138
#> 551 505.1314   3.076391
#> 552 910.1767   2.628284
#> 553 500.7897   4.104140
#> 554 575.2547   2.625023
#> 555 882.3397   2.877527
#> 556 501.2056   3.350412
#> 557 526.2771   2.603715
#> 558 500.2581   2.510382
#> 559 500.9753   3.002475
#> 560 500.5562   3.000706
#> 561 500.1629   2.964358
#> 562 505.0714   2.773290
#> 563 572.9537   2.674329
#> 564 504.5172   2.949789
#> 565 506.4326   4.227139
#> 566 500.0743   2.583519
#> 567 536.1500   3.546945
#> 568 503.8567   3.113637
#> 569 500.9651   2.518700
#> 570 517.3755   2.682976
#> 571 502.2584   3.870606
#> 572 503.3551   3.442884
#> 573 500.0749   2.579354
#> 574 500.0298   2.952149
#> 575 501.2924   2.543260
#> 576 504.3923   2.502365
#> 577 500.5369   2.509790
#> 578 500.7428   2.515068
#> 579 515.9237   3.018309
#> 580 500.8077   2.790004
#> 581 500.5382   3.802308
#> 582 704.2920   2.715247
#> 583 500.1714   2.832816
#> 584 632.1806   3.048253
#> 585 500.4055   3.064871
#> 586 520.0778   5.066352
#> 587 500.2647   2.583777
#> 588 557.9691   5.368975
#> 589 648.1744   2.886658
#> 590 600.6474   2.563895
#> 591 518.2943   3.716338
#> 592 509.5483   3.324517
#> 593 572.8341   2.867675
#> 594 500.6037   4.026440
#> 595 501.3465   4.929598
#> 596 500.3075   2.638325
#> 597 578.9180   3.017585
#> 598 500.8024   2.758725
#> 599 500.3856   3.307501
#> 600 585.6681   2.586445
#> 601 502.9191   3.352147
#> 602 500.1743   2.737785
#> 603 504.3969   2.734605
#> 604 518.3723   3.141795
#> 605 502.8367   2.868433
#> 606 501.1609   3.846199
#> 607 598.3708   3.405685
#> 608 501.1561   3.363415
#> 609 665.7708   3.126041
#> 610 504.8415   3.093783
#> 611 564.5569   3.204672
#> 612 501.7416   2.582991
#> 613 502.3500   2.503342
#> 614 500.0524   2.938184
#> 615 508.7287   2.587978
#> 616 867.5463   2.509238
#> 617 714.0511   2.706367
#> 618 509.4200   2.629082
#> 619 500.7618   2.608992
#> 620 515.1489   3.120800
#> 621 500.0169   2.562823
#> 622 526.8061   2.925804
#> 623 500.8984   3.983938
#> 624 500.9665   2.811684
#> 625 500.0150   3.050506
#> 626 501.7869   3.188608
#> 627 500.0558   2.529505
#> 628 500.6427   2.615537
#> 629 501.0347   3.044033
#> 630 500.0933   3.691909
#> 631 500.7566   2.632453
#> 632 500.8064   2.714900
#> 633 500.1245   2.579832
#> 634 500.1098   2.831338
#> 635 508.1013   2.788590
#> 636 519.0376   2.582122
#> 637 508.9451   3.824564
#> 638 501.4007   3.625779
#> 639 500.3363   2.834550
#> 640 501.9180   5.842433
#> 641 514.8565   3.070215
#> 642 754.5032   2.663200
#> 643 519.1375   2.823828
#> 644 502.1975   2.892200
#> 645 528.5499   2.756945
#> 646 779.3487   2.825377
#> 647 875.3380   2.746676
#> 648 570.6948   2.834761
#> 649 838.7121   2.990906
#> 650 505.4924   4.014225
#> 651 500.7286   2.731518
#> 652 599.8385   2.945149
#> 653 654.3589   2.636334
#> 654 500.0706   3.886817
#> 655 500.3239   2.762785
#> 656 620.3896   2.737606
#> 657 500.0401   2.568791
#> 658 500.1899   3.993723
#> 659 500.4122   2.711390
#> 660 500.0769   2.559504
#> 661 591.8526   2.566550
#> 662 500.0835   2.992034
#> 663 500.0972   2.609694
#> 664 864.0778   2.550323
#> 665 509.3509   4.472847
#> 666 500.1187   2.673873
#> 667 660.2796   3.465315
#> 668 539.4351   3.274780
#> 669 501.6742   3.554501
#> 670 800.8651   2.590463
#> 671 500.0311   2.828674
#> 672 619.8305   2.612884
#> 673 565.3819   2.651718
#> 674 727.7690   2.664265
#> 675 555.6169   2.684396
#> 676 502.4290   3.182585
#> 677 501.0423   3.571923
#> 678 505.4529   3.079410
#> 679 646.1788   3.504456
#> 680 500.0833   2.673436
#> 681 500.5905   2.543015
#> 682 500.0597   2.678123
#> 683 500.0177   2.787341
#> 684 664.3360   2.976267
#> 685 500.4987   2.748591
#> 686 715.6905   2.749645
#> 687 500.0167   2.603267
#> 688 500.7702   2.507357
#> 689 508.4220   2.637364
#> 690 500.1245   2.508996
#> 691 500.1382   3.186746
#> 692 504.8923   2.587312
#> 693 500.7239   2.835966
#> 694 500.3252   2.624446
#> 695 511.2745   2.905554
#> 696 545.5313   2.531929
#> 697 646.1588   2.662086
#> 698 537.4258   3.266310
#> 699 523.4864   3.043986
#> 700 502.3385   2.526264
#> 701 500.0089   2.642166
#> 702 500.0500   2.632634
#> 703 517.6572   4.415591
#> 704 510.8715   2.891162
#> 705 500.6365   2.738991
#> 706 746.6796   2.657073
#> 707 500.0253   2.672290
#> 708 502.9518   3.172151
#> 709 500.1213   2.779453
#> 710 517.9404   3.764309
#> 711 504.6955   2.934921
#> 712 500.2553   2.706057
#> 713 501.5838   2.613680
#> 714 500.9487   2.512537
#> 715 501.5979   2.564141
#> 716 500.0876   2.511009
#> 717 528.1556   2.546484
#> 718 993.7698   2.967900
#> 719 684.6397   2.793039
#> 720 555.2305   2.695665
#> 721 500.1051   3.145916
#> 722 550.4130   2.782820
#> 723 576.4507   2.855969
#> 724 500.7413   2.956585
#> 725 500.0358   2.857422
#> 726 500.7169   2.530041
#> 727 501.9122   3.399422
#> 728 500.0767   2.534960
#> 729 519.5434   2.905881
#> 730 756.0026   2.665994
#> 731 500.6023   2.602970
#> 732 501.0074   3.581642
#> 733 500.9388   3.481990
#> 734 722.4701   2.559306
#> 735 501.0037   2.946759
#> 736 501.2744   3.765121
#> 737 590.6464   3.098465
#> 738 500.6207   2.714499
#> 739 501.7108   4.768174
#> 740 502.2536   4.688339
#> 741 525.6890   3.633680
#> 742 500.1003   2.533768
#> 743 503.1178   2.631298
#> 744 502.2909   2.557755
#> 745 500.1300   2.916350
#> 746 586.6972   2.725086
#> 747 565.0068   2.564834
#> 748 501.1864   2.702766
#> 749 505.7476   3.044285
#> 750 501.5405   2.746268
#> 751 513.0682   2.838110
#> 752 542.7698   2.632180
#> 753 500.5603   2.703258
#> 754 515.8075   2.769646
#> 755 509.7816   3.007668
#> 756 507.2539   3.196909
#> 757 500.3239   2.835366
#> 758 575.2044   2.624791
#> 759 500.0709   2.618266
#> 760 517.2373   2.543636
#> 761 500.6577   4.763762
#> 762 500.4808   2.839715
#> 763 517.1041   3.059226
#> 764 519.5801   4.278566
#> 765 501.8364   2.557860
#> 766 502.6085   2.956188
#> 767 500.2428   3.722670
#> 768 500.7377   2.653374
#> 769 718.1523   2.959458
#> 770 500.3003   2.728008
#> 771 602.5650   2.897728
#> 772 802.1868   2.803790
#> 773 515.5558   2.650102
#> 774 537.1186   2.596262
#> 775 529.9449   2.735653
#> 776 500.0819   2.549978
#> 777 500.0604   2.684810
#> 778 500.0405   2.541732
#> 779 501.9277   3.298043
#> 780 500.0138   2.981585
#> 781 615.9270   2.697959
#> 782 501.7002   2.613652
#> 783 500.1813   2.676425
#> 784 918.2275   4.162156
#> 785 500.0205   2.743173
#> 786 528.8133   3.380265
#> 787 505.6735   2.510078
#> 788 500.0564   2.631980
#> 789 540.5384   3.026904
#> 790 500.1878   2.618331
#> 791 500.0933   2.824916
#> 792 508.6817   2.846213
#> 793 501.8495   2.542443
#> 794 501.4896   2.719337
#> 795 532.1797   2.806260
#> 796 707.5229   2.644411
#> 797 534.9480   2.803052
#> 798 886.0662   2.538901
#> 799 500.4240   2.618657
#> 800 502.6346   3.295753
#> 801 534.6487   2.632338
#> 802 500.5228   2.580590
#> 803 500.7563   2.609733
#> 804 520.4355   5.956484
#> 805 500.6261   2.526067
#> 806 511.0246   2.652246
#> 807 537.4471   2.560974
#> 808 527.3649   2.681963
#> 809 500.1110   4.362855
#> 810 926.4974   2.808410
#> 811 651.6454   3.498770
#> 812 501.9490   3.849936
#> 813 680.6592   3.145104
#> 814 516.5547   2.646387
#> 815 615.2663   2.817524
#> 816 500.0412   2.772216
#> 817 500.2776   2.805680
#> 818 500.1057   3.528381
#> 819 827.7625   2.808543
#> 820 500.8165   2.576204
#> 821 503.7930   2.678385
#> 822 501.2383   4.764630
#> 823 500.3693   2.736036
#> 824 512.6750   3.752311
#> 825 508.5954   2.619030
#> 826 974.1317   2.708478
#> 827 503.1654   2.607817
#> 828 500.5776   2.654711
#> 829 500.1169   2.803970
#> 830 500.0375   2.898673
#> 831 692.4290   2.902953
#> 832 500.1206   3.637945
#> 833 504.6391   3.438256
#> 834 510.3589   2.648139
#> 835 507.3994   2.769957
#> 836 584.9765   2.740623
#> 837 500.6796   3.785688
#> 838 560.8715   2.566439
#> 839 596.9096   2.573934
#> 840 546.8073   3.575936
#> 841 502.8616   2.629014
#> 842 549.4423   3.262640
#> 843 589.4949   2.837336
#> 844 500.4806   4.096277
#> 845 501.9054   2.906437
#> 846 512.9207   2.770001
#> 847 509.0066   2.519570
#> 848 500.0915   3.904235
#> 849 506.4474   2.586903
#> 850 500.1630   2.799421
#> 851 503.0082   2.658836
#> 852 500.2081   3.121781
#> 853 504.9016   3.520260
#> 854 535.3710   3.226800
#> 855 501.7489   2.872826
#> 856 730.5472   2.815823
#> 857 534.3967   2.518858
#> 858 502.5330   3.125183
#> 859 500.0795   2.727275
#> 860 500.2575   2.555570
#> 861 714.4266   2.519227
#> 862 790.6630   2.652990
#> 863 506.3745   2.618023
#> 864 503.2411   2.927525
#> 865 500.4903   2.580260
#> 866 500.5366   2.760031
#> 867 502.0654   2.582551
#> 868 502.0415   3.043710
#> 869 508.0791   2.981308
#> 870 615.2497   2.551152
#> 871 538.9312   3.359156
#> 872 500.4673   2.908325
#> 873 500.2540   2.731738
#> 874 500.0444   3.169877
#> 875 740.7829   3.085884
#> 876 508.9890   3.801980