Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   501.0954   2.941901
#> 2   501.1629   2.614360
#> 3   586.3468   3.357990
#> 4   731.9907   2.592459
#> 5   500.2459   2.709705
#> 6   500.3233   2.578515
#> 7   597.5976   2.827089
#> 8   504.0862   3.680147
#> 9   500.0255   3.579831
#> 10  766.2502   3.811617
#> 11  500.6597   3.479139
#> 12  686.7865   2.954357
#> 13  578.3320   3.066362
#> 14  665.4296   3.617791
#> 15  501.1363   3.217106
#> 16  500.0007   2.740150
#> 17  500.8555   3.994678
#> 18  504.3663   2.613502
#> 19  501.2237   2.569787
#> 20  516.7319   2.552179
#> 21  500.0274   2.545674
#> 22  500.2714   2.693630
#> 23  500.8171   3.379972
#> 24  501.5015   3.053680
#> 25  500.0306   2.934548
#> 26  500.1699   2.779512
#> 27  500.0603   3.070877
#> 28  500.7348   3.202838
#> 29  513.1486   3.153452
#> 30  764.8820   2.662377
#> 31  500.1707   3.071768
#> 32  500.3514   2.987578
#> 33  529.1003   2.644129
#> 34  503.3749   2.834321
#> 35  555.2911   2.562503
#> 36  729.2388   2.574093
#> 37  507.8431   3.685577
#> 38  583.6614   3.507044
#> 39  500.6882   2.799994
#> 40  500.6512   2.887942
#> 41  507.0914   2.789375
#> 42  502.4269   3.378712
#> 43  507.3348   3.053109
#> 44  629.3944   2.650055
#> 45  908.7782   2.873917
#> 46  827.7431   3.501905
#> 47  501.2617   2.988453
#> 48  500.0785   2.631255
#> 49  500.7196   2.582641
#> 50  557.0930   3.847658
#> 51  800.5995   2.542381
#> 52  607.3575   2.724969
#> 53  501.9549   2.757154
#> 54  500.0239   2.964165
#> 55  500.0220   3.313802
#> 56  511.0436   2.759052
#> 57  500.1144   2.847103
#> 58  504.7562   3.460901
#> 59  571.3574   3.309173
#> 60  613.6362   3.331485
#> 61  501.0853   2.732569
#> 62  527.1402   2.512932
#> 63  500.1881   2.521137
#> 64  702.3910   3.030201
#> 65  500.4570   2.840431
#> 66  500.0763   3.055707
#> 67  503.0746   2.505344
#> 68  500.5601   3.807337
#> 69  503.1415   2.517526
#> 70  609.1734   2.672206
#> 71  787.6886   3.075619
#> 72  666.4314   2.830697
#> 73  502.3555   3.133057
#> 74  500.1486   2.852889
#> 75  500.7381   2.656683
#> 76  510.7565   3.045208
#> 77  667.8713   2.757880
#> 78  501.9725   2.833124
#> 79  732.0183   2.628216
#> 80  500.0673   3.263876
#> 81  618.4624   3.059756
#> 82  576.8336   2.511065
#> 83  552.3976   2.923387
#> 84  500.5791   4.168357
#> 85  501.5073   2.614673
#> 86  654.9968   2.946756
#> 87  505.5542   2.651409
#> 88  500.0092   2.707409
#> 89  500.0101   4.819962
#> 90  666.6188   4.203277
#> 91  626.0766   3.125559
#> 92  500.0173   2.776142
#> 93  504.1112   2.513504
#> 94  508.0119   3.047745
#> 95  501.7624   3.038846
#> 96  770.7865   3.010153
#> 97  501.0802   3.362438
#> 98  511.9121   3.121018
#> 99  501.5370   2.628272
#> 100 534.0340   3.278565
#> 101 501.1853   3.365337
#> 102 500.3198   2.531117
#> 103 501.1167   3.349312
#> 104 500.0241   2.606029
#> 105 742.5181   2.575482
#> 106 500.3649   3.510819
#> 107 512.5539   2.783561
#> 108 504.6867   3.078425
#> 109 508.3178   2.875983
#> 110 515.8850   3.324522
#> 111 500.0065   2.845197
#> 112 511.8146   2.965453
#> 113 884.8877   2.784554
#> 114 501.0601   3.279381
#> 115 501.8721   2.677584
#> 116 501.3155   2.528315
#> 117 506.0628   2.762993
#> 118 524.0920   2.514062
#> 119 500.0666   2.736922
#> 120 579.0324   2.657761
#> 121 500.2942   3.088537
#> 122 506.0125   3.259276
#> 123 919.7575   2.598103
#> 124 535.5576   2.892208
#> 125 500.0054   2.638794
#> 126 500.2084   2.962750
#> 127 956.9092   2.717954
#> 128 500.6901   2.770730
#> 129 593.3203   2.913094
#> 130 591.2091   2.663614
#> 131 500.1904   2.529197
#> 132 500.0824   2.853125
#> 133 505.0380   3.477132
#> 134 500.0871   2.590439
#> 135 537.2983   2.639028
#> 136 500.3164   2.595864
#> 137 500.4433   2.809926
#> 138 500.1572   3.227413
#> 139 525.2135   3.034300
#> 140 500.7090   2.597257
#> 141 506.0874   3.141961
#> 142 517.5407   2.712551
#> 143 580.2842   3.398227
#> 144 646.7180   3.079003
#> 145 500.1914   2.569971
#> 146 502.6737   2.999189
#> 147 500.2361   2.585862
#> 148 547.2984   2.804005
#> 149 513.4107   2.631084
#> 150 703.1218   3.208773
#> 151 517.4755   2.565070
#> 152 500.4941   2.776600
#> 153 500.2537   2.550929
#> 154 509.0991   2.702158
#> 155 500.9364   3.142433
#> 156 500.2158   4.001340
#> 157 501.6671   3.059551
#> 158 500.0426   3.049618
#> 159 509.2569   3.242239
#> 160 627.1463   4.230069
#> 161 500.2014   3.464529
#> 162 507.5406   2.999180
#> 163 500.4380   2.526190
#> 164 542.5080   3.516066
#> 165 513.2669   2.524292
#> 166 502.7718   3.476633
#> 167 596.3679   4.252851
#> 168 500.2222   2.609597
#> 169 516.6370   3.681022
#> 170 500.5157   2.625958
#> 171 500.6491   2.614570
#> 172 501.5650   2.700306
#> 173 796.5359   2.746993
#> 174 525.9808   2.621729
#> 175 501.3076   3.032883
#> 176 534.2203   3.215686
#> 177 500.9374   2.746172
#> 178 511.2921   2.507918
#> 179 859.5890   3.083730
#> 180 503.3386   3.162582
#> 181 500.1067   2.772337
#> 182 501.0327   2.531323
#> 183 506.2683   3.408568
#> 184 528.2145   2.877928
#> 185 500.1133   2.613136
#> 186 507.8653   2.806889
#> 187 500.7730   2.523546
#> 188 576.1798   3.357192
#> 189 500.3734   2.957388
#> 190 500.1064   3.028132
#> 191 532.1911   2.933409
#> 192 603.5191   2.605676
#> 193 897.1840   2.762218
#> 194 672.5515   2.662019
#> 195 500.7785   3.189402
#> 196 500.0371   2.633889
#> 197 500.7215   2.527908
#> 198 516.8604   2.671241
#> 199 502.6854   2.689753
#> 200 514.0553   3.302490
#> 201 500.8101   2.699940
#> 202 518.9986   3.040322
#> 203 545.2687   3.042487
#> 204 502.9372   2.853864
#> 205 501.5612   2.869793
#> 206 542.2203   2.531240
#> 207 662.8927   2.824453
#> 208 502.5414   2.637963
#> 209 520.0825   2.909965
#> 210 504.4390   3.560809
#> 211 649.8252   2.643747
#> 212 501.3551   2.612466
#> 213 500.0164   4.531246
#> 214 500.3756   2.666188
#> 215 944.1780   3.935399
#> 216 512.0366   2.515889
#> 217 500.0566   4.038839
#> 218 500.0082   3.157966
#> 219 952.1664   2.724846
#> 220 532.2524   3.325457
#> 221 500.4207   3.214774
#> 222 500.0454   3.041540
#> 223 500.0031   2.573660
#> 224 525.2469   2.584922
#> 225 521.7098   2.865075
#> 226 501.8088   2.669202
#> 227 517.8436   2.853854
#> 228 538.1256   2.541029
#> 229 501.2201   2.991164
#> 230 511.5111   2.538195
#> 231 500.0681   2.509888
#> 232 579.7180   2.712708
#> 233 500.0216   2.569302
#> 234 637.6028   2.928028
#> 235 517.5123   2.674807
#> 236 503.0972   3.012784
#> 237 511.4942   2.673861
#> 238 503.8705   2.665351
#> 239 500.0848   2.598078
#> 240 501.0809   2.729788
#> 241 554.5119   2.749098
#> 242 504.2430   3.033730
#> 243 515.1410   3.184220
#> 244 500.0461   2.547481
#> 245 515.3119   2.943152
#> 246 792.1535   3.852647
#> 247 522.1538   2.603191
#> 248 508.9814   2.995118
#> 249 515.5663   2.807258
#> 250 500.7767   2.562418
#> 251 503.7202   2.798749
#> 252 514.4114   2.529591
#> 253 667.8904   2.625892
#> 254 501.9258   2.608134
#> 255 501.2361   3.565050
#> 256 889.8701   2.566889
#> 257 501.8143   3.263495
#> 258 695.2629   4.041538
#> 259 718.2229   3.220468
#> 260 500.2206   2.543240
#> 261 500.1189   2.941607
#> 262 500.2379   2.788247
#> 263 530.1543   3.477977
#> 264 500.3203   2.831308
#> 265 519.4571   2.875146
#> 266 539.8420   2.729656
#> 267 903.3809   3.046271
#> 268 520.2978   3.191011
#> 269 504.3344   2.728326
#> 270 500.7035   2.769398
#> 271 503.7795   2.771187
#> 272 523.0458   2.581543
#> 273 500.3157   2.743939
#> 274 505.3045   2.732327
#> 275 503.8047   5.159547
#> 276 500.0995   3.113329
#> 277 586.2284   2.538656
#> 278 503.9579   2.952081
#> 279 553.8300   2.807046
#> 280 533.8275   2.695129
#> 281 502.0668   2.537807
#> 282 558.8132   2.614440
#> 283 520.2742   2.653783
#> 284 500.8783   2.628602
#> 285 594.7855   2.780891
#> 286 503.6182   2.594265
#> 287 602.2671   3.352267
#> 288 522.6648   2.582153
#> 289 629.1419   2.721683
#> 290 514.0489   2.812014
#> 291 504.2235   2.789494
#> 292 500.1609   2.863325
#> 293 500.4482   3.272538
#> 294 515.1240   2.643037
#> 295 500.1569   3.221022
#> 296 500.1652   2.866171
#> 297 724.3414   2.719509
#> 298 520.1769   3.089991
#> 299 540.1218   2.974444
#> 300 513.4342   2.719351
#> 301 545.5434   2.595390
#> 302 640.5833   2.909549
#> 303 501.2953   2.835307
#> 304 501.0545   2.577909
#> 305 500.0044   2.762051
#> 306 505.7627   3.695646
#> 307 540.7042   3.053039
#> 308 500.1225   2.934540
#> 309 503.0277   2.731467
#> 310 770.6130   2.665298
#> 311 557.6175   2.750923
#> 312 502.9591   3.094532
#> 313 507.0389   2.565031
#> 314 849.7605   2.690532
#> 315 757.6636   2.686916
#> 316 531.7261   2.789643
#> 317 500.3289   2.858551
#> 318 501.0830   2.775747
#> 319 501.5360   2.740079
#> 320 536.3470   2.534954
#> 321 500.0261   3.199373
#> 322 505.4190   2.775289
#> 323 521.3093   2.727356
#> 324 501.6597   2.838325
#> 325 505.0539   3.662159
#> 326 500.0513   2.838538
#> 327 800.1410   3.194735
#> 328 505.2703   2.657653
#> 329 659.1378   2.523215
#> 330 500.7067   3.443677
#> 331 500.0869   2.664704
#> 332 500.6663   3.486188
#> 333 500.0443   3.610651
#> 334 511.1723   2.553528
#> 335 711.4017   2.693246
#> 336 500.0092   3.495748
#> 337 504.5115   4.462624
#> 338 503.7660   3.016121
#> 339 510.4998   2.868701
#> 340 500.7845   3.776591
#> 341 500.1008   3.407649
#> 342 518.9401   3.217085
#> 343 513.7825   3.616432
#> 344 834.4031   2.538911
#> 345 564.9741   2.755111
#> 346 519.6125   2.585319
#> 347 728.5443   2.598665
#> 348 504.3698   2.522467
#> 349 500.6780   2.754460
#> 350 500.0398   2.611846
#> 351 509.1446   3.073265
#> 352 542.2854   3.168767
#> 353 501.0885   2.606457
#> 354 500.2995   2.510524
#> 355 513.8747   3.598249
#> 356 500.2794   2.557926
#> 357 522.3773   2.571750
#> 358 515.3899   3.101113
#> 359 511.6104   2.630781
#> 360 677.3818   3.570069
#> 361 513.1829   3.182714
#> 362 500.2730   2.797199
#> 363 805.3173   2.900320
#> 364 503.1942   3.574304
#> 365 507.9723   2.982998
#> 366 500.1468   2.585679
#> 367 500.2366   3.456352
#> 368 500.0504   2.922894
#> 369 500.0050   2.655301
#> 370 530.1286   2.657189
#> 371 507.4736   2.739032
#> 372 501.7097   2.508921
#> 373 502.2515   2.530921
#> 374 503.8506   2.514428
#> 375 803.4583   2.625060
#> 376 500.1074   2.660073
#> 377 500.2171   3.724160
#> 378 559.7361   2.590985
#> 379 516.3705   3.085541
#> 380 500.3926   3.067732
#> 381 789.2086   2.875475
#> 382 500.2143   2.745139
#> 383 526.4758   2.533917
#> 384 562.7016   2.978874
#> 385 501.3634   2.683728
#> 386 500.0580   3.673898
#> 387 573.4907   2.761321
#> 388 501.0628   2.571134
#> 389 520.5142   2.699743
#> 390 500.1557   4.950154
#> 391 504.3846   2.990820
#> 392 511.3365   2.557603
#> 393 525.2206   2.502710
#> 394 500.1996   2.615498
#> 395 500.1252   2.926782
#> 396 500.0503   2.529881
#> 397 507.7010   2.776543
#> 398 500.2682   2.855281
#> 399 525.4402   2.643378
#> 400 500.0371   3.442795
#> 401 507.9107   2.817590
#> 402 503.2864   3.475708
#> 403 738.5539   2.611549
#> 404 500.2689   3.750354
#> 405 500.3136   2.979256
#> 406 501.7836   2.733680
#> 407 508.5662   2.615947
#> 408 549.9013   2.559195
#> 409 507.6215   3.178262
#> 410 523.0999   2.502921
#> 411 501.1486   2.567450
#> 412 500.2073   2.720818
#> 413 522.7821   2.888121
#> 414 517.6717   3.194987
#> 415 504.2629   2.632547
#> 416 507.3463   2.650769
#> 417 500.1453   3.010353
#> 418 508.8028   2.584661
#> 419 508.7744   2.547843
#> 420 500.6607   3.624080
#> 421 534.3917   3.249541
#> 422 502.5811   2.659529
#> 423 500.6272   2.630718
#> 424 500.1172   3.225822
#> 425 624.4193   3.269143
#> 426 523.9426   2.594592
#> 427 500.0666   2.515257
#> 428 511.6849   2.545796
#> 429 500.3239   2.873775
#> 430 521.6907   2.799598
#> 431 500.0829   3.167412
#> 432 500.4167   4.606777
#> 433 942.7340   2.989305
#> 434 511.5033   2.575394
#> 435 997.1620   2.742283
#> 436 501.9113   3.589390
#> 437 500.1092   2.506884
#> 438 747.8434   2.645291
#> 439 535.4784   2.805572
#> 440 502.4175   2.544362
#> 441 649.8958   2.552705
#> 442 500.7115   2.634988
#> 443 514.7062   2.964119
#> 444 537.6707   2.730359
#> 445 534.8017   2.755020
#> 446 507.4545   3.037698
#> 447 500.6518   2.707903
#> 448 691.4797   2.513267
#> 449 500.3248   2.532578
#> 450 513.1296   3.780885
#> 451 501.0239   2.604551
#> 452 500.7991   2.549325
#> 453 536.3176   2.771620
#> 454 500.5734   2.660475
#> 455 500.5659   2.553619
#> 456 500.4009   2.790385
#> 457 512.0928   3.293436
#> 458 524.0765   2.526576
#> 459 500.2917   4.090307
#> 460 541.0512   2.579180
#> 461 573.9340   2.854347
#> 462 503.4014   2.596746
#> 463 500.4145   2.626664
#> 464 571.8748   5.839715
#> 465 549.2496   2.834041
#> 466 501.9769   3.340230
#> 467 509.7486   2.597117
#> 468 833.8156   2.956672
#> 469 500.0397   2.595690
#> 470 512.0443   2.554102
#> 471 526.0289   3.012530
#> 472 500.0918   3.843830
#> 473 500.4233   3.007534
#> 474 500.2149   3.145833
#> 475 550.7304   2.956106
#> 476 501.0541   3.519006
#> 477 616.8321   3.177959
#> 478 544.8328   2.829937
#> 479 500.0436   2.644349
#> 480 500.0365   2.745260
#> 481 500.1770   2.665059
#> 482 500.0764   2.537975
#> 483 502.2662   2.787215
#> 484 542.9740   2.506917
#> 485 617.8950   2.570041
#> 486 502.8457   2.788613
#> 487 500.2214   4.272305
#> 488 783.5833   2.809722
#> 489 533.7703   2.517390
#> 490 500.1326   2.858453
#> 491 502.1738   4.342534
#> 492 501.6122   2.638107
#> 493 523.7658   2.780795
#> 494 615.7830   2.805618
#> 495 502.6628   2.502604
#> 496 501.2560   2.676699
#> 497 808.9870   3.770732
#> 498 536.4063   2.998220
#> 499 572.0181   2.629354
#> 500 511.7189   2.798715
#> 501 500.6548   3.267139
#> 502 505.4435   3.025789
#> 503 672.3944   3.588966
#> 504 500.0052   2.658428
#> 505 500.8065   4.643497
#> 506 584.9861   4.512815
#> 507 502.9580   3.332964
#> 508 520.8477   2.944810
#> 509 508.1732   3.962189
#> 510 502.3121   3.279369
#> 511 500.3430   3.252020
#> 512 500.0046   2.627457
#> 513 540.3914   3.246167
#> 514 505.3017   2.587439
#> 515 500.0348   2.754879
#> 516 500.3817   2.769086
#> 517 500.0239   2.629257
#> 518 519.8971   2.695921
#> 519 600.3287   2.581278
#> 520 522.2901   3.107496
#> 521 574.1238   2.755090
#> 522 500.0342   2.502402
#> 523 501.4536   3.213630
#> 524 524.5101   3.077998
#> 525 513.2292   2.636171
#> 526 503.5751   2.567867
#> 527 500.4464   2.649645
#> 528 625.0312   2.729288
#> 529 596.9554   2.631274
#> 530 500.0087   2.546641
#> 531 503.2421   2.697689
#> 532 500.2956   3.700783
#> 533 502.1144   2.797510
#> 534 535.7591   2.711278
#> 535 964.2516   2.592405
#> 536 500.5512   2.873206
#> 537 500.7139   4.290894
#> 538 500.4072   3.660929
#> 539 500.0413   3.190104
#> 540 504.1313   2.551723
#> 541 500.0340   2.633734
#> 542 509.9899   2.826922
#> 543 501.8470   2.864009
#> 544 501.1923   2.536660
#> 545 502.8116   2.642407
#> 546 500.5011   3.273755
#> 547 522.2306   4.257414
#> 548 503.0793   2.519158
#> 549 501.6351   2.948750
#> 550 523.8636   2.611462
#> 551 506.3628   2.972137
#> 552 500.6152   2.658932
#> 553 512.3376   2.839064
#> 554 534.9352   4.050388
#> 555 939.7420   3.704652
#> 556 502.1801   3.833577
#> 557 512.4477   2.578239
#> 558 529.2116   2.602962
#> 559 500.0851   2.793570
#> 560 500.3348   2.750942
#> 561 502.5847   2.639127
#> 562 500.5786   3.959411
#> 563 983.9231   2.538607
#> 564 560.8022   2.713709
#> 565 502.0050   2.642884
#> 566 505.5318   3.235572
#> 567 754.3740   2.551494
#> 568 970.0356   2.516952
#> 569 513.0690   3.614561
#> 570 500.0220   2.605696
#> 571 506.2684   2.740726
#> 572 503.9487   2.785230
#> 573 529.3069   3.120282
#> 574 531.4553   2.518714
#> 575 708.3756   2.597623
#> 576 501.6725   2.592817
#> 577 501.9243   2.537075
#> 578 510.9672   2.609368
#> 579 501.1158   2.695439
#> 580 500.0400   2.603022
#> 581 501.8679   2.548803
#> 582 500.7903   2.554544
#> 583 503.6582   3.107648
#> 584 500.0052   2.515219
#> 585 500.1912   3.170415
#> 586 818.0661   2.593039
#> 587 512.2449   3.089443
#> 588 501.0008   2.825267
#> 589 786.9011   2.575847
#> 590 500.1873   2.557654
#> 591 502.0695   2.545122
#> 592 984.8233   2.681434
#> 593 500.1644   3.158899
#> 594 506.5672   2.566395
#> 595 500.0951   3.866079
#> 596 500.8830   3.026538
#> 597 501.8410   2.767514
#> 598 500.0361   2.511253
#> 599 662.2843   2.678077
#> 600 500.4011   2.686552
#> 601 504.1571   2.659334
#> 602 736.8038   2.961014
#> 603 500.3418   2.799139
#> 604 504.8129   2.509824
#> 605 500.0230   3.264252
#> 606 500.2171   2.755515
#> 607 865.8746   2.982961
#> 608 502.4942   2.522581
#> 609 500.7696   2.537015
#> 610 500.1246   2.944614
#> 611 515.8036   2.614834
#> 612 500.7698   2.799996
#> 613 500.1327   3.186059
#> 614 500.1825   2.670132
#> 615 501.9691   3.604958
#> 616 516.1297   2.535324
#> 617 556.5172   3.420927
#> 618 506.5242   3.268385
#> 619 634.3159   2.940473
#> 620 537.1997   3.051155
#> 621 503.0137   2.910894
#> 622 500.2176   2.828838
#> 623 633.0851   2.719116
#> 624 500.0477   2.654035
#> 625 521.8639   2.559226
#> 626 500.1272   2.939377
#> 627 503.4128   2.722276
#> 628 684.5782   2.814050
#> 629 500.5179   3.046201
#> 630 605.1702   2.900398
#> 631 531.8137   2.735727
#> 632 625.7870   2.784310
#> 633 501.0161   2.711388
#> 634 730.6230   2.643914
#> 635 500.1396   2.794154
#> 636 603.0864   3.485655
#> 637 500.5609   2.832539
#> 638 500.0222   2.988821
#> 639 503.7507   2.810167
#> 640 511.8582   4.179970
#> 641 500.2477   2.639235
#> 642 501.4482   3.153738
#> 643 501.0493   2.754793
#> 644 686.7596   2.852204
#> 645 549.4525   3.177364
#> 646 501.9090   2.720634
#> 647 501.3379   2.939389
#> 648 556.6428   2.501410
#> 649 500.6162   3.123683
#> 650 551.2958   2.919265
#> 651 500.1067   2.828363
#> 652 504.0013   2.554145
#> 653 642.0234   3.398719
#> 654 501.1111   2.901410
#> 655 740.9517   2.570838
#> 656 895.1534   2.640759
#> 657 500.0613   2.606904
#> 658 654.6479   2.935527
#> 659 575.1041   2.563798
#> 660 511.6468   2.588316
#> 661 500.0662   3.061574
#> 662 807.8029   3.580873
#> 663 948.9571   2.641476
#> 664 519.3824   2.729935
#> 665 500.0098   4.497078
#> 666 500.1930   2.844739
#> 667 500.0241   2.616511
#> 668 502.4449   2.515302
#> 669 529.6525   2.742712
#> 670 500.0392   2.680648
#> 671 500.5210   2.789026
#> 672 513.6336   2.543588
#> 673 502.6121   2.880905
#> 674 500.3641   2.627639
#> 675 500.0704   2.607805
#> 676 501.5912   3.606825
#> 677 513.6252   3.914269
#> 678 500.6144   3.171058
#> 679 608.9825   2.520870
#> 680 505.1717   2.670156
#> 681 500.0441   2.899576
#> 682 500.0339   3.061200
#> 683 959.6434   2.662539
#> 684 500.1316   2.775686
#> 685 500.5865   2.508864
#> 686 501.2878   2.782513
#> 687 518.7440   2.631921
#> 688 693.8415   3.128627
#> 689 501.0158   3.464729
#> 690 795.4087   2.605070
#> 691 906.7942   2.600283
#> 692 531.5073   2.758047
#> 693 505.8201   2.579150
#> 694 517.2645   2.864657
#> 695 501.7838   3.734831
#> 696 500.0907   2.530850
#> 697 622.8286   3.043870
#> 698 509.4792   3.151027
#> 699 527.6195   2.508921
#> 700 501.9826   3.369545
#> 701 501.0898   2.549569
#> 702 571.6617   2.952556
#> 703 918.0598   2.634112
#> 704 500.0588   2.781085
#> 705 511.4434   2.662241
#> 706 500.1229   3.886759
#> 707 501.4933   2.907794
#> 708 500.5159   2.971619
#> 709 506.6722   3.642731
#> 710 500.2585   3.092547
#> 711 529.8814   3.473544
#> 712 827.0522   2.769719
#> 713 553.0006   2.517141
#> 714 773.4893   2.756119
#> 715 796.7921   3.073343
#> 716 530.9957   3.833818
#> 717 500.7443   3.349154
#> 718 504.0466   3.544375
#> 719 500.0046   2.740882
#> 720 502.7807   3.499722
#> 721 500.2890   2.915459
#> 722 500.6433   2.892639
#> 723 981.6538   3.267167
#> 724 915.0504   3.127433
#> 725 500.7435   4.427660
#> 726 502.5678   2.902512
#> 727 501.7013   2.931347
#> 728 512.2939   2.511611
#> 729 791.7864   2.532468
#> 730 585.1486   3.465805
#> 731 500.3255   2.684952
#> 732 500.3609   2.538375
#> 733 500.0498   3.885123
#> 734 533.5316   3.736634
#> 735 637.3167   2.544966
#> 736 500.0512   2.761258
#> 737 503.9404   2.755258
#> 738 500.8543   2.518459
#> 739 500.0090   3.322448
#> 740 501.1168   3.151084
#> 741 512.9375   3.273531
#> 742 500.3228   3.675877
#> 743 500.4093   2.622160
#> 744 522.9903   3.129077
#> 745 645.1555   3.195241
#> 746 500.0029   2.795279
#> 747 828.4159   2.516043
#> 748 709.4801   2.605605
#> 749 504.0246   3.078664
#> 750 500.8158   3.455726
#> 751 500.0196   2.531180
#> 752 901.3054   2.546368
#> 753 537.0673   3.013995
#> 754 500.0972   4.001872
#> 755 500.2652   2.802906
#> 756 500.2367   3.413460
#> 757 500.8181   3.280357
#> 758 621.6611   2.609118
#> 759 535.6012   3.275009
#> 760 501.0446   3.201361
#> 761 916.1552   2.627171
#> 762 550.9348   2.987492
#> 763 500.0240   3.033229
#> 764 528.5905   2.579987
#> 765 501.2313   2.513605
#> 766 500.0482   3.013143
#> 767 500.0198   3.046231
#> 768 500.6899   3.477557
#> 769 503.8452   2.860904
#> 770 502.1221   2.627754
#> 771 501.4857   2.742343
#> 772 868.0360   2.782296
#> 773 603.2089   2.956200
#> 774 502.3171   2.517009
#> 775 563.4236   2.505139
#> 776 501.1348   2.837201
#> 777 658.5538   2.990918
#> 778 690.1034   2.649503
#> 779 590.1160   4.530634
#> 780 500.9638   3.423400
#> 781 500.2121   2.563950
#> 782 542.0388   2.550823
#> 783 500.9436   3.566397
#> 784 500.0050   2.541505
#> 785 501.4224   2.732163
#> 786 544.6613   2.573297
#> 787 500.4822   3.278463
#> 788 712.6746   2.578202
#> 789 594.3933   2.615241
#> 790 509.8902   3.907859
#> 791 500.2529   2.913804
#> 792 500.2071   2.981062
#> 793 506.1069   3.188952
#> 794 500.0943   2.898793
#> 795 505.3706   2.640554
#> 796 505.1930   2.522314
#> 797 500.0632   2.587987
#> 798 521.4640   2.551277
#> 799 516.1461   2.900586
#> 800 505.9981   3.156714
#> 801 502.5551   3.115487
#> 802 506.7283   5.201339
#> 803 560.1022   2.512725
#> 804 534.5980   2.589398
#> 805 500.8662   4.044164
#> 806 500.8881   2.557869
#> 807 500.5261   2.733050
#> 808 541.4935   3.491913
#> 809 500.0140   2.566596
#> 810 500.4389   3.491350
#> 811 903.3897   2.749354
#> 812 681.4510   2.788360
#> 813 563.4948   2.928294
#> 814 512.5254   3.294844
#> 815 500.1749   2.717636
#> 816 501.0121   3.093721
#> 817 500.4794   2.992563
#> 818 504.2337   2.705461
#> 819 502.3613   3.045762
#> 820 502.6424   3.077901
#> 821 500.2016   3.218689
#> 822 501.9013   2.523885
#> 823 500.3473   2.536883
#> 824 500.0253   2.766867
#> 825 500.1265   2.632793
#> 826 500.1188   2.532018
#> 827 517.3061   2.605255
#> 828 501.4189   2.506134
#> 829 500.6079   2.649590
#> 830 501.6657   2.771790
#> 831 501.0613   4.074415
#> 832 502.2382   3.419458
#> 833 576.5835   2.744813
#> 834 564.0001   3.173923
#> 835 501.7213   2.722787
#> 836 503.9868   2.585023
#> 837 524.3305   3.276631
#> 838 500.0428   2.842580
#> 839 535.1618   2.522617
#> 840 501.4748   2.574572
#> 841 515.4542   2.542206
#> 842 968.1405   2.658168
#> 843 501.9302   2.779374
#> 844 500.1505   3.180410
#> 845 500.1568   2.990783
#> 846 501.1047   3.377200
#> 847 500.7665   2.676360
#> 848 500.1150   3.380793
#> 849 705.2425   3.024038
#> 850 522.5872   2.517482
#> 851 831.1114   2.639061
#> 852 504.5466   2.621212
#> 853 510.6949   3.210019
#> 854 500.1905   3.046268
#> 855 539.2440   3.429285
#> 856 521.5803   2.540206
#> 857 580.1363   2.836182
#> 858 501.4162   3.391909
#> 859 500.9514   2.649679
#> 860 502.2056   2.528348
#> 861 703.2524   2.538691
#> 862 502.2890   3.210588
#> 863 500.3057   2.548442
#> 864 502.0471   2.883628
#> 865 500.5612   2.943432
#> 866 525.2929   2.856913
#> 867 503.7575   2.603871
#> 868 579.9925   2.865015
#> 869 657.6172   2.921770
#> 870 500.0677   2.538391
#> 871 543.3805   2.804388
#> 872 847.1326   2.561157
#> 873 501.1938   3.138930
#> 874 544.6851   2.716030
#> 875 514.0432   4.663220
#> 876 500.4527   2.673888
#> 877 500.0268   2.555742
#> 878 501.5937   2.798250
#> 879 502.3625   3.173935
#> 880 538.6752   2.701106
#> 881 500.0571   3.323037
#> 882 500.1005   2.551227
#> 883 501.2620   2.512284
#> 884 540.0114   2.873207
#> 885 501.6622   3.224613
#> 886 517.4631   2.577713
#> 887 502.9913   2.943496
#> 888 500.4488   2.993774
#> 889 500.0541   2.651431
#> 890 589.3109   2.506682
#> 891 519.8676   3.282537
#> 892 500.8195   2.680199
#> 893 502.5082   2.582278
#> 894 506.1069   2.579997
#> 895 500.9540   2.590878
#> 896 500.0948   2.710693
#> 897 827.4772   3.084279
#> 898 522.6238   3.724067
#> 899 500.0305   2.559326
#> 900 500.2585   3.611946
#> 901 500.7904   2.640221
#> 902 500.1272   2.735627
#> 903 500.0612   2.874734
#> 904 503.1873   2.633597
#> 905 500.2048   2.816867
#> 906 536.7704   2.874025
#> 907 501.4192   2.725425
#> 908 568.5267   2.866026
#> 909 500.0137   2.777162
#> 910 500.7954   3.020532
#> 911 535.2689   2.620455
#> 912 526.4801   2.746299
#> 913 508.2828   2.511783
#> 914 500.3890   2.756111
#> 915 531.1863   2.787509
#> 916 501.8084   2.793699
#> 917 735.2254   2.704204
#> 918 962.4723   2.861366
#> 919 500.0452   4.095060
#> 920 502.7041   2.549660
#> 921 548.0507   3.470413
#> 922 500.6155   2.640647
#> 923 501.2505   2.551737
#> 924 639.4222   2.564480
#> 925 525.2752   2.712010
#> 926 691.0709   2.937495
#> 927 544.2013   2.793770
#> 928 500.4466   2.906661
#> 929 510.2536   3.698248
#> 930 500.4055   2.708981
#> 931 500.7357   3.180771
#> 932 500.0400   2.854538
#> 933 500.2491   2.625321
#> 934 500.6615   3.128898
#> 935 508.4238   2.565351
#> 936 501.9529   3.196098
#> 937 731.9535   2.675773
#> 938 521.1914   2.677543
#> 939 502.3129   2.832795
#> 940 500.2556   2.528196
#> 941 603.3051   3.156454
#> 942 860.3606   3.586432
#> 943 501.2166   2.653513
#> 944 515.8902   2.689469
#> 945 531.4094   3.045702
#> 946 966.6537   2.528885
#> 947 501.8234   2.594261
#> 948 714.0057   2.675833