Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   500.0111   2.621645
#> 2   704.5235   2.637526
#> 3   500.9716   3.707240
#> 4   509.1964   3.165245
#> 5   934.6719   2.689988
#> 6   501.7139   2.760670
#> 7   500.0929   2.835704
#> 8   506.0118   3.233887
#> 9   557.0049   3.757304
#> 10  520.0638   2.866597
#> 11  500.7420   2.573305
#> 12  501.2201   2.918963
#> 13  513.3058   2.744810
#> 14  512.7516   3.193962
#> 15  660.3638   3.417000
#> 16  500.5509   2.834107
#> 17  562.7925   2.906245
#> 18  500.3621   2.730886
#> 19  740.5272   2.886641
#> 20  535.9506   2.954281
#> 21  527.4064   2.775450
#> 22  515.5900   2.695573
#> 23  500.4018   2.545922
#> 24  862.6141   2.544753
#> 25  507.1463   2.546829
#> 26  503.7825   3.263843
#> 27  901.5115   2.837253
#> 28  685.7244   3.679367
#> 29  500.0000   2.872387
#> 30  516.1627   2.568874
#> 31  500.0327   3.345284
#> 32  503.2402   4.425562
#> 33  500.2723   4.166880
#> 34  939.5017   2.650613
#> 35  660.5626   3.281186
#> 36  500.7353   2.846162
#> 37  546.4466   2.898732
#> 38  517.2678   3.286182
#> 39  500.3624   2.525676
#> 40  501.4443   3.332360
#> 41  502.9903   3.070424
#> 42  549.7680   2.695418
#> 43  514.1138   2.552131
#> 44  636.3825   3.162209
#> 45  540.9431   2.830364
#> 46  500.0309   4.005774
#> 47  500.0805   3.452793
#> 48  504.5544   3.530993
#> 49  505.2843   4.481474
#> 50  500.1518   3.466018
#> 51  500.6546   3.425811
#> 52  527.1221   2.537700
#> 53  538.3177   3.437800
#> 54  799.7287   3.050979
#> 55  612.8420   2.544730
#> 56  503.7567   2.815290
#> 57  500.5978   3.949571
#> 58  519.4623   4.494856
#> 59  500.0623   2.833179
#> 60  610.9756   2.741511
#> 61  502.4979   2.669574
#> 62  501.4402   3.593107
#> 63  502.1758   2.911460
#> 64  691.6104   3.109938
#> 65  501.0615   2.674806
#> 66  500.8894   4.234378
#> 67  500.1200   2.704997
#> 68  504.7146   3.180369
#> 69  514.7092   2.804592
#> 70  594.1396   2.864861
#> 71  512.4305   2.736297
#> 72  571.4938   2.998021
#> 73  519.0408   3.944683
#> 74  587.5071   2.537825
#> 75  501.2344   2.589657
#> 76  500.7914   3.311143
#> 77  500.2052   3.202847
#> 78  500.3162   2.892619
#> 79  632.4738   3.851848
#> 80  500.1224   2.740725
#> 81  500.1248   3.037575
#> 82  502.8695   4.174658
#> 83  539.3878   2.518313
#> 84  500.1087   3.734293
#> 85  500.7564   2.698893
#> 86  500.2135   2.661639
#> 87  500.3339   3.062151
#> 88  500.0084   2.691906
#> 89  531.8697   2.595031
#> 90  514.8367   3.112318
#> 91  500.2512   2.901019
#> 92  500.1661   2.953006
#> 93  615.3782   2.997176
#> 94  500.2281   2.523673
#> 95  502.4062   3.110575
#> 96  632.6410   2.886517
#> 97  501.3807   3.007994
#> 98  500.9834   2.688533
#> 99  503.9617   3.317851
#> 100 567.3453   3.060404
#> 101 500.4934   3.356245
#> 102 500.3182   3.727276
#> 103 502.2383   3.385767
#> 104 508.4794   3.076000
#> 105 514.5989   3.535747
#> 106 599.7140   3.154065
#> 107 500.0893   3.160612
#> 108 501.8769   2.574013
#> 109 518.2086   2.904715
#> 110 770.1740   3.577107
#> 111 505.6589   2.828332
#> 112 501.7776   3.839200
#> 113 501.3019   2.518590
#> 114 500.3756   3.069097
#> 115 513.4646   2.852935
#> 116 500.0149   2.847871
#> 117 585.5830   2.644488
#> 118 501.7193   3.054379
#> 119 502.2434   2.644359
#> 120 505.4975   2.556601
#> 121 541.6219   3.969299
#> 122 769.5443   2.980415
#> 123 500.1748   2.630938
#> 124 573.5033   3.008946
#> 125 501.7611   2.587323
#> 126 976.9360   2.603259
#> 127 519.6189   2.677676
#> 128 522.7165   4.152083
#> 129 566.5775   2.769541
#> 130 500.8821   2.930772
#> 131 500.0158   4.193552
#> 132 501.9494   2.531499
#> 133 500.0214   3.009485
#> 134 535.3387   2.785740
#> 135 946.5660   2.731981
#> 136 707.5749   2.533028
#> 137 500.0008   3.932571
#> 138 633.7810   2.787370
#> 139 606.8474   2.637832
#> 140 500.4317   2.918222
#> 141 500.2026   2.683102
#> 142 500.1224   2.573487
#> 143 545.9430   3.063979
#> 144 668.6947   2.668928
#> 145 502.6363   3.524319
#> 146 501.4702   3.987661
#> 147 552.9164   2.952887
#> 148 506.4668   2.636289
#> 149 539.1309   2.960171
#> 150 566.1875   2.571473
#> 151 501.9312   2.606955
#> 152 808.6103   2.920439
#> 153 500.4256   2.604239
#> 154 500.1315   2.661006
#> 155 509.6362   3.416370
#> 156 500.1492   2.858970
#> 157 500.1779   3.346906
#> 158 525.6628   3.071298
#> 159 500.2599   2.590478
#> 160 508.8824   2.548229
#> 161 500.3274   2.584971
#> 162 502.5925   2.694031
#> 163 835.5767   2.916822
#> 164 503.3034   2.767347
#> 165 504.0202   2.546466
#> 166 500.6657   3.379020
#> 167 505.7461   2.691651
#> 168 508.6801   2.533489
#> 169 591.4898   3.026216
#> 170 500.4317   2.534189
#> 171 658.5203   3.237075
#> 172 500.3326   3.683297
#> 173 500.0486   3.386398
#> 174 502.5945   2.671002
#> 175 500.0218   3.416121
#> 176 535.8287   3.500780
#> 177 500.8827   2.981947
#> 178 500.3901   2.719414
#> 179 508.5150   3.567241
#> 180 506.1983   3.013589
#> 181 500.0088   2.574079
#> 182 500.4438   2.930984
#> 183 517.9931   2.989473
#> 184 500.2217   2.861183
#> 185 502.1656   2.572092
#> 186 651.2458   2.727196
#> 187 500.0545   2.582295
#> 188 657.2582   2.583163
#> 189 654.3620   2.656330
#> 190 500.0753   2.825271
#> 191 536.7392   2.669916
#> 192 542.0145   2.798619
#> 193 527.7479   2.764410
#> 194 501.5988   3.424840
#> 195 511.4258   3.228571
#> 196 550.8296   2.669046
#> 197 506.5093   2.668872
#> 198 505.0789   3.126430
#> 199 515.2820   3.068578
#> 200 500.0470   2.915057
#> 201 500.8546   2.572645
#> 202 598.7456   2.582139
#> 203 504.8981   2.538879
#> 204 500.7613   2.514361
#> 205 530.3575   3.312779
#> 206 596.3890   2.847666
#> 207 529.6594   3.092740
#> 208 501.3681   4.822059
#> 209 500.0626   2.552490
#> 210 500.1006   2.523153
#> 211 500.2118   3.013293
#> 212 508.5820   2.764174
#> 213 500.2721   2.516493
#> 214 500.8003   3.307154
#> 215 658.0401   2.738883
#> 216 776.3871   4.588196
#> 217 544.2461   2.829231
#> 218 500.2322   3.174207
#> 219 501.1854   2.613667
#> 220 500.5463   3.109125
#> 221 500.3063   2.967819
#> 222 512.1007   2.646155
#> 223 510.5933   2.717227
#> 224 501.3416   3.063300
#> 225 502.9615   2.646160
#> 226 501.4237   2.504274
#> 227 576.3845   2.749915
#> 228 579.5028   2.852629
#> 229 525.5889   2.642047
#> 230 500.0388   3.488385
#> 231 501.9170   3.455083
#> 232 501.1003   3.296607
#> 233 500.0223   2.934851
#> 234 599.0739   3.075856
#> 235 617.9710   2.596129
#> 236 500.3004   3.107639
#> 237 500.5841   3.221062
#> 238 500.2822   3.609144
#> 239 501.2084   2.648479
#> 240 509.8777   2.644508
#> 241 561.2768   3.600407
#> 242 502.9318   3.121496
#> 243 500.6836   2.798874
#> 244 513.7763   4.259509
#> 245 500.0737   2.740968
#> 246 508.7331   2.810241
#> 247 505.5816   2.521443
#> 248 564.2488   3.743535
#> 249 532.6003   2.559146
#> 250 500.6394   3.271869
#> 251 500.0313   2.831617
#> 252 927.2285   2.561559
#> 253 506.6214   3.531489
#> 254 502.0528   3.932756
#> 255 500.0171   2.555262
#> 256 635.1758   2.829326
#> 257 541.7724   3.188129
#> 258 528.4072   2.761014
#> 259 500.2302   2.729598
#> 260 500.1781   3.490226
#> 261 500.4998   2.577707
#> 262 500.3170   3.220168
#> 263 573.1941   2.720547
#> 264 547.6197   2.672402
#> 265 500.0952   4.170028
#> 266 503.9500   3.127091
#> 267 521.4140   5.122153
#> 268 500.0987   2.613524
#> 269 500.4405   2.591108
#> 270 500.0916   2.885398
#> 271 500.2894   2.530837
#> 272 512.0918   2.597301
#> 273 501.0182   3.204644
#> 274 512.3277   3.140230
#> 275 571.3862   4.773842
#> 276 500.3964   2.866517
#> 277 519.5119   3.157067
#> 278 504.9464   2.591146
#> 279 500.3028   2.658823
#> 280 500.1004   3.097299
#> 281 501.2151   2.993350
#> 282 502.6413   2.802491
#> 283 699.6742   2.785669
#> 284 536.3334   2.705239
#> 285 884.0921   3.365018
#> 286 506.5645   2.696876
#> 287 500.0081   3.216788
#> 288 500.2131   2.615480
#> 289 500.8460   3.376588
#> 290 511.2548   2.539320
#> 291 505.2895   3.017600
#> 292 501.5693   2.708344
#> 293 500.0167   2.852458
#> 294 587.8234   2.544671
#> 295 507.9616   2.614270
#> 296 500.9491   2.732703
#> 297 510.4191   2.854046
#> 298 505.8773   2.904358
#> 299 515.4776   3.520532
#> 300 683.9151   2.527106
#> 301 500.0673   3.043463
#> 302 500.0002   2.824465
#> 303 609.2650   2.718939
#> 304 520.2164   2.769790
#> 305 500.0024   2.521362
#> 306 504.6516   2.547550
#> 307 500.0241   2.579243
#> 308 502.0958   2.671932
#> 309 500.4388   2.609119
#> 310 500.0228   4.704378
#> 311 501.5876   2.583966
#> 312 501.0924   4.336966
#> 313 501.7575   2.571776
#> 314 500.4581   6.598884
#> 315 503.9187   2.795340
#> 316 502.5593   2.578185
#> 317 500.4569   3.374704
#> 318 511.0901   2.545231
#> 319 501.9798   3.300823
#> 320 501.5033   2.624213
#> 321 502.3108   2.528598
#> 322 602.7005   2.788124
#> 323 503.5063   2.694180
#> 324 500.3236   2.528019
#> 325 541.2207   3.051464
#> 326 513.9302   2.712940
#> 327 651.4192   2.776701
#> 328 500.3021   2.955145
#> 329 500.5844   2.621467
#> 330 518.7938   2.514638
#> 331 669.6191   2.728518
#> 332 501.1608   3.180576
#> 333 513.4307   3.116391
#> 334 645.5349   3.400674
#> 335 500.6231   2.504704
#> 336 500.2406   3.501301
#> 337 603.8549   2.956954
#> 338 500.0271   3.137866
#> 339 567.0654   2.572304
#> 340 515.7865   2.861571
#> 341 500.7415   4.026957
#> 342 504.0086   2.965641
#> 343 503.3027   2.638954
#> 344 500.3159   2.613581
#> 345 501.2533   2.824048
#> 346 506.5833   3.210263
#> 347 510.4714   2.608571
#> 348 504.5773   2.665073
#> 349 500.2970   2.700718
#> 350 500.2324   3.019923
#> 351 500.1709   2.597501
#> 352 515.0312   2.515390
#> 353 545.6090   3.043437
#> 354 500.1188   3.499331
#> 355 591.6869   2.568514
#> 356 533.7812   2.692320
#> 357 500.3420   2.841432
#> 358 500.0920   2.536429
#> 359 500.4630   3.859513
#> 360 502.5213   2.837185
#> 361 510.1333   2.997324
#> 362 501.2795   2.575534
#> 363 504.7263   3.208650
#> 364 700.0799   2.613338
#> 365 826.3802   2.635839
#> 366 532.0888   4.608955
#> 367 500.0687   2.980615
#> 368 892.9209   2.573227
#> 369 500.4181   2.676648
#> 370 531.3995   3.298196
#> 371 500.1335   3.383414
#> 372 500.7043   3.085834
#> 373 500.1280   2.700080
#> 374 541.8623   2.601412
#> 375 550.6782   2.531801
#> 376 507.4942   2.618108
#> 377 543.9026   3.338037
#> 378 500.1138   3.480721
#> 379 500.3411   2.697311
#> 380 500.3687   3.211570
#> 381 599.1608   3.816250
#> 382 523.7035   2.532195
#> 383 521.6827   3.015647
#> 384 511.3916   2.906445
#> 385 500.8731   2.612940
#> 386 508.0418   2.508441
#> 387 515.8806   2.508651
#> 388 500.1934   2.517294
#> 389 500.2688   2.862848
#> 390 503.2222   2.751437
#> 391 927.0634   2.704310
#> 392 877.3307   2.633344
#> 393 504.0801   2.546177
#> 394 502.5672   3.753560
#> 395 783.9823   2.807521
#> 396 509.6583   2.507956
#> 397 500.7186   2.506990
#> 398 501.4011   3.637974
#> 399 676.1504   3.188473
#> 400 667.1061   3.383758
#> 401 500.0256   2.948072
#> 402 529.8186   3.115948
#> 403 503.0552   2.820014
#> 404 521.8446   3.019257
#> 405 500.0691   2.501444
#> 406 501.5033   2.535119
#> 407 520.9071   2.735579
#> 408 566.4233   2.781115
#> 409 522.9220   2.535229
#> 410 500.8526   2.545043
#> 411 503.5570   2.680816
#> 412 500.6247   3.482033
#> 413 508.8143   2.602390
#> 414 812.0519   2.521573
#> 415 500.0860   2.707705
#> 416 500.3008   3.204883
#> 417 525.1069   2.575896
#> 418 512.5944   3.904915
#> 419 547.5103   2.548609
#> 420 500.1071   2.763127
#> 421 547.0960   2.634807
#> 422 543.1508   2.561664
#> 423 500.7516   2.874591
#> 424 502.2652   3.401430
#> 425 703.6623   2.847712
#> 426 500.7807   2.562736
#> 427 501.9466   3.585413
#> 428 500.5879   2.853800
#> 429 500.1011   3.011133
#> 430 763.7451   2.811734
#> 431 500.2121   2.842257
#> 432 500.1083   2.663475
#> 433 500.0721   3.087628
#> 434 500.0967   3.516189
#> 435 516.1319   2.922635
#> 436 501.1815   3.223365
#> 437 567.7070   2.853067
#> 438 502.3238   2.656460
#> 439 699.2849   3.739880
#> 440 517.9757   3.613907
#> 441 500.2657   2.955451
#> 442 539.4703   2.702190
#> 443 523.1160   3.800750
#> 444 507.2097   3.211093
#> 445 697.1081   2.532416
#> 446 530.3183   2.719895
#> 447 500.7627   3.814634
#> 448 536.9751   3.329881
#> 449 517.1831   2.917938
#> 450 500.1172   2.873351
#> 451 883.5108   3.645426
#> 452 992.1066   2.609315
#> 453 500.0299   2.569765
#> 454 501.1648   2.600015
#> 455 500.1690   2.942277
#> 456 551.0556   3.177459
#> 457 524.7375   2.664843
#> 458 589.2142   3.136001
#> 459 530.7446   3.184473
#> 460 501.7507   3.078267
#> 461 502.1562   2.825800
#> 462 500.9240   2.514084
#> 463 506.9489   3.766312
#> 464 500.2119   2.896327
#> 465 502.1622   2.557711
#> 466 532.9624   2.782474
#> 467 500.0927   3.401716
#> 468 500.0575   2.672401
#> 469 500.0582   3.374115
#> 470 561.9112   2.575384
#> 471 504.3893   2.795850
#> 472 670.1540   2.548638
#> 473 500.0728   2.607762
#> 474 504.2803   2.671922
#> 475 809.3277   2.586502
#> 476 860.1930   3.166284
#> 477 500.0678   3.083260
#> 478 692.5923   2.948583
#> 479 500.5663   3.493839
#> 480 539.1178   2.924055
#> 481 754.0233   2.734953
#> 482 521.3852   3.830100
#> 483 501.4025   2.638901
#> 484 560.7238   3.331937
#> 485 500.3083   2.975113
#> 486 501.0101   2.746865
#> 487 606.1698   2.698149
#> 488 500.7167   2.681031
#> 489 500.0178   3.054945
#> 490 500.1197   2.747672
#> 491 500.0938   2.792504
#> 492 592.0187   2.615896
#> 493 501.6792   2.714868
#> 494 500.2629   2.782738
#> 495 504.4731   2.581358
#> 496 524.9967   2.578520
#> 497 659.3204   3.274027
#> 498 506.6707   2.515731
#> 499 507.1853   3.016334
#> 500 500.0509   2.928085
#> 501 544.7666   2.590927
#> 502 500.1897   2.548835
#> 503 500.1808   3.806930
#> 504 742.2073   2.721543
#> 505 525.3642   2.729610
#> 506 936.7371   2.634794
#> 507 504.5135   3.037016
#> 508 503.2554   2.911550
#> 509 500.2427   2.520378
#> 510 518.4118   2.624190
#> 511 503.0824   2.619438
#> 512 500.0505   3.026316
#> 513 504.7048   3.343751
#> 514 512.5472   2.685582
#> 515 519.4452   3.189178
#> 516 500.1440   3.179768
#> 517 500.0217   3.165787
#> 518 500.9947   2.582863
#> 519 500.2380   2.534705
#> 520 500.4960   2.832683
#> 521 500.2002   2.718563
#> 522 502.5005   3.306406
#> 523 576.6281   2.714912
#> 524 518.9635   2.947987
#> 525 592.1949   2.819881
#> 526 506.5692   2.672482
#> 527 502.2313   3.232699
#> 528 567.2312   3.269781
#> 529 686.4628   3.090667
#> 530 507.3328   2.875291
#> 531 707.1272   3.008262
#> 532 512.5247   2.636607
#> 533 532.4059   3.748345
#> 534 501.8798   2.558698
#> 535 500.4068   3.873982
#> 536 500.0485   3.292631
#> 537 986.5284   2.615133
#> 538 530.1142   2.832648
#> 539 504.0463   3.306352
#> 540 500.7690   2.518313
#> 541 500.4653   2.822970
#> 542 511.9267   3.426038
#> 543 657.5896   2.567171
#> 544 500.0840   2.953313
#> 545 642.4928   3.044324
#> 546 517.9165   2.666488
#> 547 601.8252   2.779465
#> 548 500.1797   2.906150
#> 549 559.5060   2.553116
#> 550 500.6751   2.837580
#> 551 551.8331   2.598532
#> 552 511.1079   2.801822
#> 553 500.4142   3.941527
#> 554 509.1651   2.729038
#> 555 500.4872   2.540892
#> 556 500.0151   2.669535
#> 557 500.5867   3.541399
#> 558 969.2958   2.597425
#> 559 508.8493   2.766360
#> 560 500.8021   3.408517
#> 561 500.2455   3.564940
#> 562 500.0958   2.623516
#> 563 518.0236   2.573624
#> 564 509.9955   2.696636
#> 565 503.4984   2.680883
#> 566 503.0193   2.578037
#> 567 524.3843   2.790692
#> 568 500.3133   2.811532
#> 569 500.9882   2.733244
#> 570 544.1775   2.519899
#> 571 504.0746   4.222432
#> 572 601.8433   2.537894
#> 573 503.7346   3.776233
#> 574 535.5810   2.802334
#> 575 507.6152   2.502392
#> 576 981.1106   2.861326
#> 577 500.5928   2.537655
#> 578 500.2446   3.178179
#> 579 503.8920   2.693668
#> 580 517.8394   2.727994
#> 581 502.8449   2.641302
#> 582 527.9883   2.985543
#> 583 519.4090   2.643155
#> 584 513.2379   3.136888
#> 585 500.0453   3.934508
#> 586 500.3128   2.736208
#> 587 564.1312   2.729920
#> 588 587.1281   2.509459
#> 589 613.9403   2.934970
#> 590 662.5887   2.801424
#> 591 504.5407   3.446026
#> 592 544.3456   3.186364
#> 593 500.1083   2.589228
#> 594 564.5958   2.563350
#> 595 505.8770   3.790284
#> 596 507.0250   2.577061
#> 597 500.2610   2.839120
#> 598 500.1828   2.609581
#> 599 527.3601   3.656961
#> 600 500.1285   3.656374
#> 601 514.9990   2.504240
#> 602 513.3937   3.323756
#> 603 500.9675   2.655451
#> 604 500.0785   2.549561
#> 605 500.0524   3.521738
#> 606 501.0963   3.062581
#> 607 500.1098   3.286208
#> 608 997.2007   4.133330
#> 609 513.1251   2.627442
#> 610 501.9414   3.032210
#> 611 500.4867   3.127902
#> 612 500.0101   2.709215
#> 613 500.1199   3.387539
#> 614 917.7312   2.825192
#> 615 536.2221   3.219747
#> 616 583.2505   3.023619
#> 617 500.6733   2.692364
#> 618 529.6669   2.504561
#> 619 500.3375   3.444250
#> 620 500.2833   2.844434
#> 621 501.9070   3.301625
#> 622 504.7721   2.536619
#> 623 500.2166   2.805862
#> 624 500.1513   2.779508
#> 625 500.2990   2.541173
#> 626 500.3085   3.532408
#> 627 500.1259   2.611174
#> 628 500.4400   2.813230
#> 629 500.0122   3.061907
#> 630 503.9803   2.653786
#> 631 500.0308   2.687311
#> 632 501.3039   2.797752
#> 633 500.4692   2.548325
#> 634 505.5654   2.506465
#> 635 500.1820   2.570620
#> 636 500.4575   3.153677
#> 637 500.8890   2.698136
#> 638 500.1827   3.402713
#> 639 555.9157   2.584814
#> 640 500.0649   3.128033
#> 641 501.5553   2.891161
#> 642 658.8397   2.596669
#> 643 501.4610   3.305968
#> 644 507.1205   2.793892
#> 645 500.2303   2.545846
#> 646 525.8540   2.687061
#> 647 500.0300   2.967893
#> 648 513.6683   4.334578
#> 649 502.3233   2.550821
#> 650 501.4756   2.637174
#> 651 507.3650   2.545417
#> 652 500.7333   2.539442
#> 653 508.7319   3.330254
#> 654 971.0044   3.497913
#> 655 500.2033   2.547937
#> 656 769.0723   2.650901
#> 657 596.6080   2.661098
#> 658 500.0977   2.755416
#> 659 500.6878   2.698730
#> 660 500.0438   2.552735
#> 661 503.3630   2.847607
#> 662 697.8960   2.538944
#> 663 500.8561   2.556760
#> 664 513.4211   2.654007
#> 665 500.4799   2.508140
#> 666 629.4654   2.792814
#> 667 500.7054   2.791258
#> 668 501.4820   2.928806
#> 669 500.1331   2.575070
#> 670 502.2913   2.948508
#> 671 516.7567   2.668982
#> 672 500.9901   2.809122
#> 673 639.0355   2.812405
#> 674 754.5516   2.511640
#> 675 500.9220   2.730438
#> 676 507.4152   2.566027
#> 677 692.0844   2.753524
#> 678 674.8327   3.481293
#> 679 500.8548   3.226214
#> 680 516.4853   2.585419
#> 681 537.3185   3.127909
#> 682 500.1510   3.569230
#> 683 560.6011   2.546082
#> 684 635.3266   2.675987
#> 685 772.5458   3.594353
#> 686 503.1649   3.283136
#> 687 501.5380   2.599156
#> 688 615.1762   2.561118
#> 689 502.8558   3.837180
#> 690 610.7707   2.959979
#> 691 500.1416   2.504666
#> 692 867.7479   2.741805
#> 693 500.9693   3.046448
#> 694 501.1198   2.884854
#> 695 501.1460   3.629778
#> 696 521.2642   2.817289
#> 697 532.7827   2.945333
#> 698 503.1379   2.746795
#> 699 500.5445   3.073251
#> 700 500.6744   2.507026
#> 701 501.1666   3.116670
#> 702 538.0842   3.058376
#> 703 520.8552   2.997455
#> 704 500.0260   2.679273
#> 705 700.9126   2.724443
#> 706 500.0148   3.169020
#> 707 804.2843   2.896133
#> 708 500.2033   2.804651
#> 709 513.3412   4.395262
#> 710 500.2241   2.646253
#> 711 503.2367   4.003725
#> 712 589.2578   2.510124
#> 713 516.2870   3.575109
#> 714 889.1236   2.864398
#> 715 513.1062   2.907571
#> 716 675.0791   2.829262
#> 717 500.2682   2.999267
#> 718 501.9174   3.285296
#> 719 516.6417   3.410143
#> 720 500.0590   2.656310
#> 721 503.2397   2.548241
#> 722 511.0288   3.575890
#> 723 500.6668   3.114669
#> 724 516.4568   2.538062
#> 725 515.3133   2.969569
#> 726 503.8167   2.651730
#> 727 500.1269   2.657249
#> 728 500.1160   2.919386
#> 729 500.0785   2.969261
#> 730 500.0276   2.678230
#> 731 500.6760   3.237388
#> 732 500.0218   3.055058
#> 733 536.8569   3.689641
#> 734 500.1850   2.856135
#> 735 503.1424   2.950860
#> 736 500.2726   2.523761
#> 737 501.0652   3.202700
#> 738 502.9529   2.806080
#> 739 500.1744   2.772621
#> 740 509.3749   2.502018
#> 741 513.6308   2.539767
#> 742 502.8906   2.551309
#> 743 500.0606   3.309357
#> 744 502.2609   4.886998
#> 745 500.1950   3.270232
#> 746 500.9732   2.557265
#> 747 500.4228   2.703373
#> 748 500.1810   3.122085
#> 749 821.1074   4.041310
#> 750 501.4269   2.998855
#> 751 548.1549   2.700244
#> 752 957.6660   3.371686
#> 753 508.6607   3.452464
#> 754 512.7513   3.121316
#> 755 500.4989   3.074000
#> 756 725.9078   3.377284
#> 757 500.9221   3.031794
#> 758 803.0469   3.181099
#> 759 500.7076   3.010174
#> 760 548.5171   2.936761
#> 761 526.6232   2.941630
#> 762 642.3364   2.623666
#> 763 609.9025   2.826956
#> 764 513.9797   3.311540
#> 765 573.7883   3.872556
#> 766 504.9667   4.149923
#> 767 920.8127   2.572931
#> 768 500.0437   3.636866
#> 769 530.0222   2.854613
#> 770 500.9711   2.612179
#> 771 574.4416   2.834190
#> 772 500.2090   3.665111
#> 773 502.8669   2.927613
#> 774 502.1753   3.492650
#> 775 500.1752   2.940458
#> 776 564.2598   2.600005
#> 777 502.3767   2.501123
#> 778 506.6726   3.488735
#> 779 501.5402   2.776689
#> 780 555.4050   2.635441
#> 781 501.3450   2.819345
#> 782 516.5650   2.949444
#> 783 500.1206   2.653043
#> 784 714.4045   3.198803
#> 785 502.8662   2.639790
#> 786 501.2555   2.830854
#> 787 500.2072   3.024720
#> 788 557.6113   2.635889
#> 789 558.1717   3.594191
#> 790 500.4675   2.785097
#> 791 508.9140   2.773504
#> 792 500.0880   2.502099
#> 793 503.7828   3.161672
#> 794 501.4480   2.945729
#> 795 500.0687   2.758589
#> 796 500.0770   3.101257
#> 797 509.9518   2.576871
#> 798 571.3218   2.832460
#> 799 515.4757   3.089703
#> 800 741.3984   2.732818
#> 801 500.1028   2.547843
#> 802 505.7729   2.517924
#> 803 766.2346   2.513988
#> 804 629.7019   3.146750
#> 805 516.7040   2.517982
#> 806 500.9957   3.259429
#> 807 535.8432   3.089871
#> 808 500.6726   2.583561
#> 809 530.2557   2.606353
#> 810 502.1710   2.729541
#> 811 508.8746   2.750006
#> 812 502.4055   2.699797
#> 813 500.9590   2.755690
#> 814 500.0619   3.197252
#> 815 721.2447   2.605071
#> 816 504.3698   4.010712
#> 817 500.0405   2.586091
#> 818 681.0213   2.560078
#> 819 507.5573   3.074001
#> 820 670.2993   2.525702
#> 821 742.4053   2.642011
#> 822 501.6701   3.091719
#> 823 517.4431   2.545629
#> 824 675.1404   3.045605
#> 825 500.0363   2.775845
#> 826 500.3049   4.159562
#> 827 521.7667   2.734103
#> 828 500.3917   2.532360
#> 829 505.9591   2.823984
#> 830 502.3773   2.925319
#> 831 500.2738   2.561133
#> 832 570.2195   2.692475
#> 833 500.0712   2.638505
#> 834 500.8240   2.841953
#> 835 501.0560   3.035766
#> 836 705.3904   3.434430
#> 837 500.4946   3.174309
#> 838 500.1076   3.135134
#> 839 500.4570   3.140807
#> 840 500.0671   2.954760
#> 841 504.9323   2.736868
#> 842 500.8446   2.794288
#> 843 596.0563   5.073456
#> 844 520.1781   2.988427
#> 845 500.0423   3.057879
#> 846 500.1565   3.802281
#> 847 500.6637   2.664939
#> 848 500.6784   2.567403
#> 849 531.5505   2.504153
#> 850 531.8297   2.655828
#> 851 522.8113   3.109451
#> 852 509.4817   3.059741
#> 853 501.2852   2.663092
#> 854 500.0209   3.146273
#> 855 545.1607   2.906167
#> 856 501.6149   2.961468
#> 857 501.5620   3.105913
#> 858 579.0932   2.601073
#> 859 500.3843   2.611192
#> 860 680.3361   2.708763
#> 861 502.0884   2.540698
#> 862 502.0977   2.826615
#> 863 500.3931   2.585770
#> 864 501.3829   3.429908
#> 865 501.1490   4.374374
#> 866 500.0911   3.083733
#> 867 500.0322   2.696942
#> 868 500.1301   2.924310
#> 869 501.8775   2.699934
#> 870 553.1184   2.973755
#> 871 501.1488   2.636804
#> 872 630.6316   2.702785
#> 873 500.5049   2.950867
#> 874 714.8332   2.640953
#> 875 506.8530   2.550736
#> 876 500.6924   2.622868
#> 877 524.1881   2.857017
#> 878 500.2189   3.003510
#> 879 511.1173   3.790823
#> 880 502.1958   3.058526
#> 881 520.9758   2.602818
#> 882 501.2318   3.079610
#> 883 533.5181   2.568873
#> 884 512.5934   2.748927
#> 885 571.3598   2.695680
#> 886 500.0499   3.051939
#> 887 500.4803   2.613397
#> 888 500.0539   2.823283
#> 889 895.4841   2.611313
#> 890 518.8785   3.013268
#> 891 500.2250   2.760846
#> 892 532.8876   3.152033
#> 893 500.0644   3.016381
#> 894 500.1064   2.573056
#> 895 505.3036   3.002741
#> 896 671.0027   2.819690
#> 897 522.0359   3.035656
#> 898 500.1838   2.818265
#> 899 500.1971   2.749690
#> 900 500.0189   2.853055
#> 901 500.4083   3.457284
#> 902 544.1689   2.888661
#> 903 501.1826   2.573790
#> 904 504.5212   2.999124
#> 905 500.1323   2.780354
#> 906 646.1950   2.865867
#> 907 503.3359   3.119441
#> 908 519.4527   3.089855
#> 909 590.9792   2.508270