Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   507.3170   2.781356
#> 2   571.0522   3.109726
#> 3   509.8247   2.548565
#> 4   500.0968   2.675873
#> 5   861.8449   2.521069
#> 6   629.6302   2.732990
#> 7   994.1104   2.867401
#> 8   599.1645   2.778428
#> 9   500.0506   2.550071
#> 10  503.2698   3.430548
#> 11  528.7241   3.569271
#> 12  755.1785   3.230391
#> 13  500.4972   2.827427
#> 14  500.0085   2.603251
#> 15  910.5461   2.583863
#> 16  500.0784   2.810226
#> 17  500.2374   2.633995
#> 18  502.2188   2.547208
#> 19  545.2182   2.503819
#> 20  500.2669   3.741725
#> 21  518.0731   2.557853
#> 22  584.2694   2.514966
#> 23  746.8535   4.064982
#> 24  501.3293   2.843214
#> 25  511.3665   2.504172
#> 26  500.1234   3.015569
#> 27  500.0845   2.560212
#> 28  500.2305   3.061662
#> 29  610.1056   2.578116
#> 30  500.1463   2.648359
#> 31  500.0088   3.300689
#> 32  517.9586   3.139447
#> 33  500.5730   2.925778
#> 34  507.2585   2.509645
#> 35  500.2293   3.003299
#> 36  500.2746   3.659869
#> 37  501.3903   2.754633
#> 38  500.2778   2.527722
#> 39  500.1875   2.821890
#> 40  744.4890   3.128855
#> 41  502.5363   2.930502
#> 42  541.6918   2.613169
#> 43  501.6809   2.534909
#> 44  500.0106   3.003959
#> 45  556.1411   3.618662
#> 46  500.1090   2.831942
#> 47  500.1701   2.915786
#> 48  500.1530   4.143416
#> 49  500.1399   3.531101
#> 50  500.3782   2.514213
#> 51  678.7690   2.967397
#> 52  627.0449   2.607276
#> 53  945.5992   3.853472
#> 54  500.7116   3.412529
#> 55  504.1803   2.823249
#> 56  500.1386   2.834832
#> 57  508.1700   3.357812
#> 58  623.8624   2.900314
#> 59  502.3475   3.412115
#> 60  500.1276   2.551595
#> 61  501.2418   3.900593
#> 62  506.9931   3.200258
#> 63  512.7458   2.687066
#> 64  685.0859   2.945309
#> 65  500.0634   2.599431
#> 66  500.0958   2.674783
#> 67  536.3026   2.677108
#> 68  622.5719   2.666447
#> 69  500.3250   3.006179
#> 70  500.1301   2.645710
#> 71  503.8358   2.881335
#> 72  500.4054   2.574332
#> 73  543.3774   3.176205
#> 74  502.8470   4.943760
#> 75  500.0819   2.583797
#> 76  533.2533   2.688967
#> 77  541.2007   3.113623
#> 78  650.4969   2.724864
#> 79  500.0821   2.676933
#> 80  505.7612   2.832460
#> 81  504.6897   2.521292
#> 82  672.4964   2.751289
#> 83  543.4384   2.845038
#> 84  534.0422   2.698093
#> 85  500.3951   2.702799
#> 86  544.8037   2.706408
#> 87  500.1424   2.683018
#> 88  500.1820   2.683219
#> 89  500.1566   2.705461
#> 90  500.9217   3.121414
#> 91  500.9296   2.658429
#> 92  502.2439   2.635334
#> 93  538.2711   2.580819
#> 94  503.5877   2.983773
#> 95  500.4324   2.668465
#> 96  510.8807   2.618202
#> 97  624.0509   2.650384
#> 98  500.0371   2.519808
#> 99  678.4543   2.502619
#> 100 505.9162   3.714665
#> 101 683.1665   2.774026
#> 102 519.8779   3.204322
#> 103 506.1486   2.599476
#> 104 502.8794   2.518574
#> 105 500.3420   3.110825
#> 106 500.7870   2.973747
#> 107 701.8415   2.502789
#> 108 500.3772   2.673581
#> 109 500.1802   2.940816
#> 110 501.1162   2.748369
#> 111 515.0964   2.620323
#> 112 523.5366   2.887491
#> 113 517.3149   2.605137
#> 114 512.2640   2.683359
#> 115 505.7685   2.735453
#> 116 500.2936   2.591906
#> 117 500.3942   2.624381
#> 118 503.0723   3.191785
#> 119 500.0733   2.836307
#> 120 500.2309   3.224308
#> 121 501.3985   2.514707
#> 122 500.5013   2.525241
#> 123 504.3972   2.567162
#> 124 509.8975   2.740020
#> 125 689.7856   2.765140
#> 126 500.1328   2.635851
#> 127 505.9338   3.377916
#> 128 525.0959   2.534800
#> 129 500.0331   5.173894
#> 130 500.1557   2.651533
#> 131 502.9530   2.906858
#> 132 500.0671   2.501644
#> 133 801.8431   2.947246
#> 134 694.0636   2.723304
#> 135 584.4788   2.697181
#> 136 503.1589   3.081346
#> 137 503.9071   2.503654
#> 138 500.2929   2.725001
#> 139 506.5564   2.773441
#> 140 500.3140   2.609066
#> 141 500.0234   2.773314
#> 142 593.3630   2.558951
#> 143 579.0348   2.665495
#> 144 515.9322   2.608584
#> 145 534.3683   2.780873
#> 146 526.2384   2.697768
#> 147 519.9918   2.520473
#> 148 552.1674   3.474017
#> 149 513.0979   2.690878
#> 150 579.2005   2.680302
#> 151 505.3194   4.091778
#> 152 500.2275   2.986574
#> 153 681.1109   2.934087
#> 154 526.5300   3.202968
#> 155 501.9433   2.518757
#> 156 963.5209   2.635824
#> 157 500.2810   2.844781
#> 158 500.2404   2.829970
#> 159 911.9453   3.065864
#> 160 507.2722   2.906275
#> 161 526.0449   2.720784
#> 162 500.0345   2.584036
#> 163 500.1575   2.511506
#> 164 500.5619   3.476266
#> 165 501.0518   2.843520
#> 166 500.5076   2.566283
#> 167 589.8389   3.459128
#> 168 501.8695   3.698669
#> 169 534.5956   2.971232
#> 170 502.6004   3.732008
#> 171 500.4909   2.945990
#> 172 955.0767   2.544450
#> 173 502.4062   3.137090
#> 174 500.0077   3.243769
#> 175 644.4349   3.468645
#> 176 700.6679   2.538228
#> 177 688.5462   3.012241
#> 178 506.8697   2.737405
#> 179 500.1312   2.554944
#> 180 500.9680   2.594422
#> 181 500.5857   2.575967
#> 182 500.0289   2.904789
#> 183 523.4022   2.585617
#> 184 500.8977   2.528153
#> 185 612.1346   2.735252
#> 186 500.8387   2.511329
#> 187 589.8789   2.689120
#> 188 502.5765   2.666898
#> 189 500.3386   2.504610
#> 190 650.9253   3.535660
#> 191 645.2333   2.768287
#> 192 738.7627   3.092763
#> 193 502.1011   3.011477
#> 194 565.6188   3.591914
#> 195 521.5885   2.730627
#> 196 500.4761   3.009949
#> 197 647.2704   2.965510
#> 198 503.9945   2.783032
#> 199 500.0615   2.750633
#> 200 502.7279   2.641960
#> 201 500.0958   3.015367
#> 202 536.7622   2.910116
#> 203 500.5578   2.552421
#> 204 500.1334   2.990144
#> 205 500.0478   2.822614
#> 206 500.0178   2.685526
#> 207 500.0954   3.545054
#> 208 500.8817   2.723715
#> 209 500.4685   2.734701
#> 210 520.0366   2.890839
#> 211 501.3742   2.584353
#> 212 631.1020   2.681362
#> 213 500.1625   2.500756
#> 214 706.9669   2.669810
#> 215 765.7190   3.368228
#> 216 510.6981   2.644405
#> 217 500.1483   2.687973
#> 218 500.1353   2.805093
#> 219 500.1236   2.560081
#> 220 500.1375   2.542739
#> 221 617.3054   2.868408
#> 222 557.1342   2.914465
#> 223 500.4494   2.671432
#> 224 506.7215   2.773246
#> 225 720.1352   2.568924
#> 226 503.9908   3.005269
#> 227 780.0107   2.866461
#> 228 508.8123   3.437391
#> 229 619.8726   3.188762
#> 230 530.2556   3.041059
#> 231 501.2485   2.899981
#> 232 502.9924   2.665937
#> 233 567.1965   2.646887
#> 234 538.0927   2.609266
#> 235 512.9737   2.590977
#> 236 502.2796   2.741849
#> 237 501.4708   2.824492
#> 238 517.4661   2.929021
#> 239 509.9984   2.823313
#> 240 524.0142   2.885425
#> 241 665.7491   2.687435
#> 242 502.6811   2.871896
#> 243 513.0859   2.606798
#> 244 500.3054   2.504243
#> 245 500.5249   3.626628
#> 246 585.9446   2.521651
#> 247 500.0077   3.136144
#> 248 503.0319   2.804547
#> 249 501.1147   2.731262
#> 250 500.0994   2.587299
#> 251 531.2047   2.514316
#> 252 531.8294   2.722499
#> 253 500.0491   3.499974
#> 254 500.4065   2.616624
#> 255 500.2649   3.198254
#> 256 500.0053   2.567458
#> 257 583.0167   2.730252
#> 258 500.3631   3.405787
#> 259 514.7824   2.744390
#> 260 501.5709   2.524962
#> 261 500.2345   2.731297
#> 262 500.6079   3.404664
#> 263 620.8581   3.947700
#> 264 557.3314   2.604092
#> 265 500.7873   2.542442
#> 266 567.7974   3.126476
#> 267 722.5716   3.509386
#> 268 524.9728   2.734292
#> 269 500.0666   2.663768
#> 270 525.8179   2.559349
#> 271 500.1515   2.641912
#> 272 738.1572   2.588859
#> 273 501.6930   2.601562
#> 274 518.0122   2.822067
#> 275 998.5630   2.771386
#> 276 506.5068   3.029364
#> 277 504.4586   3.429193
#> 278 542.3071   4.229611
#> 279 500.4315   2.855810
#> 280 553.1507   2.514085
#> 281 515.0202   2.609255
#> 282 533.5077   2.527384
#> 283 500.3795   2.822105
#> 284 531.2317   2.682522
#> 285 506.2905   2.714177
#> 286 502.7556   2.593031
#> 287 505.0493   2.752643
#> 288 549.5152   3.232842
#> 289 500.6084   2.691211
#> 290 723.7430   2.694890
#> 291 701.9170   2.672203
#> 292 504.0135   4.280709
#> 293 500.5799   3.017615
#> 294 520.4976   3.591410
#> 295 513.0661   3.300641
#> 296 507.8473   3.723494
#> 297 567.1645   3.041230
#> 298 501.2491   2.738594
#> 299 932.1225   3.461892
#> 300 517.4490   2.641148
#> 301 520.6060   3.905666
#> 302 545.5283   2.599914
#> 303 509.5524   2.989071
#> 304 501.6369   2.858468
#> 305 500.4261   2.589180
#> 306 500.6776   3.571122
#> 307 500.6612   2.524600
#> 308 500.4763   2.565941
#> 309 500.8904   2.691588
#> 310 506.1292   2.687031
#> 311 501.5185   2.854311
#> 312 500.4341   2.755809
#> 313 501.4132   3.531017
#> 314 629.3663   2.877656
#> 315 500.8084   2.936765
#> 316 500.1087   2.986296
#> 317 500.8278   2.750655
#> 318 510.6932   2.782919
#> 319 518.6100   2.626675
#> 320 513.4707   2.902549
#> 321 714.6478   3.225151
#> 322 524.7816   2.600574
#> 323 500.1188   2.914095
#> 324 500.2360   2.529466
#> 325 500.3413   2.521333
#> 326 501.1573   2.801889
#> 327 552.3317   4.917691
#> 328 626.8418   2.654000
#> 329 502.8379   3.168619
#> 330 637.6784   2.651215
#> 331 500.0783   2.659441
#> 332 530.4598   3.910049
#> 333 500.0288   3.076064
#> 334 500.0036   3.065855
#> 335 511.4972   3.159938
#> 336 857.0969   2.611209
#> 337 509.3960   3.103512
#> 338 500.9944   2.789188
#> 339 554.2933   2.828030
#> 340 501.3814   3.790151
#> 341 591.8715   2.723073
#> 342 500.6302   2.778307
#> 343 500.0747   3.194227
#> 344 501.1855   3.382035
#> 345 512.3761   2.923727
#> 346 500.0740   3.032218
#> 347 500.3109   2.700720
#> 348 500.4643   3.374813
#> 349 501.9980   2.938763
#> 350 556.7005   3.137325
#> 351 504.6511   2.515122
#> 352 510.0134   2.541221
#> 353 514.5777   2.697252
#> 354 500.5301   3.308826
#> 355 655.4985   2.843396
#> 356 500.0755   2.519688
#> 357 581.5581   2.708067
#> 358 503.4681   2.824175
#> 359 635.6143   3.404921
#> 360 502.7108   2.553285
#> 361 505.7866   2.970143
#> 362 500.3035   2.658485
#> 363 507.6608   3.987916
#> 364 607.5892   2.579734
#> 365 515.6633   2.522434
#> 366 511.2461   2.838995
#> 367 542.4606   2.538782
#> 368 767.5692   2.572081
#> 369 525.3595   3.756736
#> 370 518.4498   2.857786
#> 371 501.9622   3.101433
#> 372 521.5872   2.831281
#> 373 501.2056   2.668406
#> 374 500.1482   3.023245
#> 375 503.9672   2.668693
#> 376 500.1045   2.978073
#> 377 602.4478   2.541033
#> 378 555.4829   2.692430
#> 379 500.0735   2.542694
#> 380 530.0692   3.198096
#> 381 500.2627   2.775910
#> 382 500.1041   2.992764
#> 383 645.2918   3.109580
#> 384 577.4049   2.643958
#> 385 512.3742   3.234726
#> 386 516.3305   2.694443
#> 387 501.7494   2.844052
#> 388 619.0113   2.819174
#> 389 500.1646   2.635444
#> 390 501.9289   3.364386
#> 391 508.1274   2.660669
#> 392 501.4090   2.707004
#> 393 500.0531   2.713165
#> 394 810.7388   4.372512
#> 395 500.0478   2.575383
#> 396 533.5474   2.696146
#> 397 517.1982   2.889572
#> 398 507.6089   2.565302
#> 399 665.5283   2.764358
#> 400 836.9209   2.801238
#> 401 575.5830   2.565942
#> 402 500.4509   2.574809
#> 403 500.0145   2.728492
#> 404 509.2975   2.911951
#> 405 502.9992   2.522976
#> 406 673.3429   2.662274
#> 407 516.3013   4.680826
#> 408 500.0319   2.577039
#> 409 500.7648   2.785404
#> 410 864.4950   2.548614
#> 411 525.7849   3.089674
#> 412 500.1935   2.923615
#> 413 535.5131   2.740870
#> 414 500.0396   3.631436
#> 415 500.1394   2.574569
#> 416 500.1026   2.793736
#> 417 501.2382   2.597980
#> 418 500.1527   2.567459
#> 419 600.9101   2.793874
#> 420 583.2966   2.833637
#> 421 506.9930   3.023269
#> 422 510.5587   2.597096
#> 423 500.6192   2.626569
#> 424 504.2592   2.836114
#> 425 682.9355   3.376354
#> 426 848.2935   2.605702
#> 427 500.2655   2.830375
#> 428 501.0745   2.641989
#> 429 500.2155   3.002204
#> 430 500.0769   2.580745
#> 431 500.0033   2.674054
#> 432 549.6635   3.524500
#> 433 953.3471   3.499530
#> 434 500.4136   3.862112
#> 435 505.1466   2.527637
#> 436 500.0467   2.863576
#> 437 500.2116   2.573699
#> 438 550.9806   3.042099
#> 439 627.0598   3.169056
#> 440 500.3500   3.012022
#> 441 501.7049   2.731844
#> 442 502.4559   2.520825
#> 443 653.0221   2.575553
#> 444 500.0123   3.194517
#> 445 512.0955   2.698278
#> 446 601.8636   2.984186
#> 447 502.2122   2.828580
#> 448 906.6419   2.536783
#> 449 556.3028   3.086391
#> 450 504.6710   3.270206
#> 451 500.0131   3.446905
#> 452 500.0827   2.836268
#> 453 500.3085   2.546424
#> 454 500.7291   3.089482
#> 455 500.9302   3.386107
#> 456 500.5252   3.728746
#> 457 509.9432   2.720106
#> 458 500.2079   2.530831
#> 459 623.1423   3.034235
#> 460 540.6450   3.300649
#> 461 500.3379   2.728536
#> 462 504.0972   2.652729
#> 463 708.6069   3.663649
#> 464 509.6419   2.536657
#> 465 500.0328   2.514408
#> 466 500.3133   2.635365
#> 467 513.2012   2.512799
#> 468 519.0227   2.880523
#> 469 532.4107   2.932966
#> 470 535.4774   2.588718
#> 471 627.2042   2.517866
#> 472 500.1074   2.995508
#> 473 501.5894   2.818335
#> 474 500.0928   3.118256
#> 475 500.3888   2.560144
#> 476 521.9523   2.790775
#> 477 500.0034   2.582120
#> 478 500.1353   2.764781
#> 479 500.6663   2.506144
#> 480 517.1741   2.521244
#> 481 501.6293   2.803293
#> 482 503.6861   3.005616
#> 483 502.3367   2.569225
#> 484 987.5902   3.003147
#> 485 674.3931   3.692620
#> 486 501.5494   2.505819
#> 487 511.3535   2.819912
#> 488 500.5521   3.218377
#> 489 526.7598   3.094834
#> 490 537.0635   3.085069
#> 491 508.5061   2.507481
#> 492 500.8819   2.681694
#> 493 500.8803   2.548115
#> 494 505.1064   2.592201
#> 495 504.9577   2.997148
#> 496 500.8338   2.519763
#> 497 678.7757   2.993053
#> 498 505.2697   2.894109
#> 499 500.6610   2.834623
#> 500 500.1332   2.701382
#> 501 511.1783   2.871519
#> 502 787.6096   3.017069
#> 503 541.6216   2.747156
#> 504 502.8604   3.401971
#> 505 500.1686   2.514676
#> 506 517.8468   2.866950
#> 507 632.8999   3.351135
#> 508 501.9501   2.752177
#> 509 500.4469   2.858254
#> 510 500.0616   2.612557
#> 511 823.6677   2.820850
#> 512 727.6619   4.880667
#> 513 511.0457   2.776186
#> 514 586.2840   3.526887
#> 515 626.0762   2.775513
#> 516 529.8720   2.997060
#> 517 513.5146   2.970188
#> 518 506.4792   3.868166
#> 519 509.0937   2.953274
#> 520 821.5594   2.782777
#> 521 507.7902   2.505621
#> 522 521.3759   2.661350
#> 523 501.5320   2.510395
#> 524 500.9838   2.728382
#> 525 616.5618   3.281218
#> 526 501.2066   2.594605
#> 527 504.3676   4.837819
#> 528 688.1701   2.547282
#> 529 532.3443   2.873568
#> 530 755.1504   2.779124
#> 531 505.1927   2.612574
#> 532 500.7171   3.028597
#> 533 502.5802   2.791514
#> 534 534.2364   2.784142
#> 535 596.0384   2.535287
#> 536 508.7171   2.705104
#> 537 500.0868   2.547179
#> 538 523.1826   2.560886
#> 539 521.1404   2.621175
#> 540 501.6956   2.561349
#> 541 517.9692   3.544849
#> 542 500.3112   3.041118
#> 543 500.0363   3.014837
#> 544 500.3935   2.859046
#> 545 500.0013   3.091106
#> 546 500.5928   2.744674
#> 547 648.4130   2.651658
#> 548 500.3047   2.654797
#> 549 535.6273   2.950049
#> 550 500.5320   2.799853
#> 551 500.3408   2.946783
#> 552 500.4015   3.226598
#> 553 527.7232   2.663209
#> 554 500.7524   2.652551
#> 555 500.1201   2.537930
#> 556 500.0120   2.622297
#> 557 656.2110   2.747519
#> 558 544.2026   2.944991
#> 559 582.2878   2.997611
#> 560 500.4641   2.761626
#> 561 538.0674   2.579225
#> 562 500.1730   2.900066
#> 563 500.2854   2.775647
#> 564 851.9682   2.947816
#> 565 500.0565   2.676929
#> 566 504.9533   2.608610
#> 567 600.7730   2.580879
#> 568 500.9430   2.712567
#> 569 577.4235   2.875694
#> 570 500.9377   3.837119
#> 571 504.9134   2.906257
#> 572 500.7710   2.503348
#> 573 773.8912   2.564867
#> 574 500.0032   2.923659
#> 575 500.9799   2.537359
#> 576 538.2451   3.066373
#> 577 523.0168   2.713751
#> 578 500.4771   2.781403
#> 579 500.8575   2.533941
#> 580 500.6351   3.173006
#> 581 508.5651   3.373985
#> 582 775.4492   2.620717
#> 583 729.8037   3.137671
#> 584 972.5771   2.725274
#> 585 500.0835   2.570579
#> 586 512.1431   3.142333
#> 587 500.8856   2.667266
#> 588 500.7877   2.582860
#> 589 500.2894   3.307719
#> 590 500.8130   2.668208
#> 591 531.4683   2.517529
#> 592 503.6622   2.891153
#> 593 500.7153   3.004595
#> 594 500.2856   2.719705
#> 595 500.1287   2.525254
#> 596 506.9026   3.094567
#> 597 559.0017   3.439955
#> 598 526.4510   2.965653
#> 599 503.3282   2.625187
#> 600 508.4682   2.835263
#> 601 500.8776   3.373617
#> 602 511.9395   3.704719
#> 603 524.4483   2.726458
#> 604 501.2874   2.639061
#> 605 500.0465   2.628429
#> 606 517.2739   2.547971
#> 607 574.1326   2.835854
#> 608 500.0899   2.618603
#> 609 502.2593   3.061833
#> 610 548.6289   2.863943
#> 611 524.0889   2.984788
#> 612 823.3603   2.742840
#> 613 566.0618   2.636431
#> 614 516.6628   5.274735
#> 615 500.4599   2.591609
#> 616 779.8520   2.622814
#> 617 635.5095   3.267201
#> 618 532.8827   2.532965
#> 619 500.1974   3.025476
#> 620 887.9574   3.000856
#> 621 788.3894   2.591315
#> 622 516.8601   3.044425
#> 623 500.0534   3.181787
#> 624 516.8159   2.936261
#> 625 500.0169   2.700168
#> 626 501.0273   2.638038
#> 627 518.8853   2.820931
#> 628 500.1656   2.706734
#> 629 500.1364   2.699467
#> 630 504.3070   2.504177
#> 631 638.6945   2.945642
#> 632 607.9974   2.696408
#> 633 501.7696   3.360252
#> 634 556.5767   3.299906
#> 635 510.2543   2.646130
#> 636 500.1528   2.759901
#> 637 506.5089   3.232681
#> 638 519.7331   2.691490
#> 639 797.0383   2.610997
#> 640 670.9978   2.890311
#> 641 500.0195   3.122796
#> 642 822.6454   2.708775
#> 643 500.8867   3.184013
#> 644 501.8598   3.116484
#> 645 741.3818   3.671426
#> 646 513.1445   2.939259
#> 647 501.8070   3.356501
#> 648 509.3118   2.794391
#> 649 500.0114   2.931158
#> 650 809.6935   2.643788
#> 651 522.3984   2.608962
#> 652 577.8033   2.628590
#> 653 500.1396   2.959772
#> 654 503.9145   2.858269
#> 655 593.8757   2.604016
#> 656 510.4319   2.598890
#> 657 500.1649   2.509017
#> 658 554.3345   2.741889
#> 659 501.4523   2.878521
#> 660 541.0893   2.525795
#> 661 503.5378   2.747598
#> 662 501.1879   3.175059
#> 663 550.0569   2.635453
#> 664 500.1106   2.606477
#> 665 503.8915   2.591079
#> 666 501.2527   2.990578
#> 667 523.7754   2.949729
#> 668 516.4291   2.508760
#> 669 547.5865   3.126690
#> 670 500.1161   2.896710
#> 671 506.3646   2.764553
#> 672 532.9487   3.299004
#> 673 501.6843   2.819612
#> 674 500.1962   2.741579
#> 675 508.4817   3.665907
#> 676 500.6039   2.759127
#> 677 552.0658   2.851337
#> 678 502.4572   3.350934
#> 679 509.6189   2.921535
#> 680 502.0881   2.769155
#> 681 502.0069   2.524545
#> 682 516.5215   2.526976
#> 683 522.4050   2.881880
#> 684 500.1081   2.929223
#> 685 500.4230   2.578902
#> 686 913.5252   2.546610
#> 687 500.1528   3.302500
#> 688 507.2233   2.517377
#> 689 512.7686   2.549680
#> 690 520.3614   2.536746
#> 691 500.7457   3.117029
#> 692 500.3922   2.849548
#> 693 500.0873   2.697932
#> 694 500.6676   3.029945
#> 695 555.0929   2.791653
#> 696 500.9430   2.940084
#> 697 672.7560   2.874826
#> 698 597.1351   3.072987
#> 699 500.0284   2.732116
#> 700 561.6746   3.365842
#> 701 502.3046   3.003630
#> 702 501.7701   2.741848
#> 703 656.4601   2.958476
#> 704 626.4371   2.738097
#> 705 590.9347   3.267908
#> 706 631.5846   2.657152
#> 707 500.0975   2.780986
#> 708 503.8651   2.665420
#> 709 512.5564   2.579786
#> 710 607.4348   2.508839
#> 711 501.5399   2.624666
#> 712 500.2173   2.651859
#> 713 515.2975   3.541138
#> 714 535.1527   2.929844
#> 715 500.1401   3.063731
#> 716 502.1605   2.503759
#> 717 547.4408   3.299738
#> 718 500.0394   2.823733
#> 719 501.5718   2.846132
#> 720 502.5781   2.648275
#> 721 500.0699   2.617672
#> 722 500.1451   2.818271
#> 723 533.9426   2.799044
#> 724 500.6584   2.616021
#> 725 500.0392   4.155150
#> 726 500.1801   3.539346
#> 727 500.1301   2.521477
#> 728 507.1191   2.848346
#> 729 501.5547   2.581049
#> 730 500.9542   3.040958
#> 731 503.1633   2.715268
#> 732 501.1408   2.574755
#> 733 508.1762   2.868831
#> 734 500.0094   2.987696
#> 735 580.3090   2.919234
#> 736 500.0329   2.793001
#> 737 505.1224   3.497164
#> 738 500.6814   2.843752
#> 739 500.3637   2.907665
#> 740 500.0959   3.329424
#> 741 500.0090   2.843187
#> 742 501.4098   2.533164
#> 743 568.9563   3.383567
#> 744 500.7147   2.543323
#> 745 938.8479   2.670839
#> 746 681.4227   2.638358
#> 747 500.3369   3.271799
#> 748 876.9547   2.740590
#> 749 501.4687   2.840951
#> 750 765.0092   2.890265
#> 751 500.2897   2.702247
#> 752 500.7449   2.500456
#> 753 581.6263   2.599100
#> 754 517.3585   2.877088
#> 755 511.3773   2.904242
#> 756 500.4730   3.299511
#> 757 510.9544   2.941222
#> 758 508.2174   2.970366
#> 759 500.4162   2.540288
#> 760 806.0499   3.019841
#> 761 503.7985   2.932951
#> 762 505.3620   2.700847
#> 763 741.5352   2.564995
#> 764 800.0824   2.830453
#> 765 500.0648   2.562172
#> 766 500.8057   2.641460
#> 767 508.5283   2.528931
#> 768 503.3791   2.644357
#> 769 500.1276   2.542689
#> 770 505.2159   2.964107
#> 771 501.0187   2.963194
#> 772 535.3796   2.985989
#> 773 501.9877   2.544228
#> 774 504.1527   3.099513
#> 775 504.2352   3.130416
#> 776 500.0308   3.008303
#> 777 502.6039   2.874971
#> 778 511.6900   2.583247
#> 779 500.7254   2.947215
#> 780 624.0118   3.293187
#> 781 501.5371   2.746790
#> 782 708.0110   3.672822
#> 783 500.5255   2.878672
#> 784 518.2397   3.060858
#> 785 500.1448   2.500801
#> 786 794.0845   2.635568
#> 787 543.2096   2.578571
#> 788 902.6938   2.643374
#> 789 504.0544   2.595001
#> 790 510.3090   4.015715
#> 791 500.0094   2.844720
#> 792 597.0037   3.139047
#> 793 515.3742   4.026034
#> 794 500.6182   3.070732
#> 795 501.3084   2.818700
#> 796 515.8777   2.837536
#> 797 500.4722   2.655038
#> 798 501.0086   3.942570
#> 799 532.9555   2.835371
#> 800 500.5601   2.660781
#> 801 509.7594   3.408745
#> 802 500.0231   2.803322
#> 803 511.0332   2.780954
#> 804 500.2134   3.467485
#> 805 707.6648   2.512195
#> 806 500.0312   2.600839
#> 807 500.2450   3.302591
#> 808 501.6038   2.633132
#> 809 557.7352   2.816070
#> 810 509.0571   2.596197
#> 811 521.4641   2.657722
#> 812 508.7961   3.140225
#> 813 513.1598   2.509420
#> 814 500.7344   2.673315
#> 815 503.5761   2.822456
#> 816 500.0051   2.555306
#> 817 500.3340   3.235999
#> 818 501.4131   2.855295
#> 819 500.0249   4.184967
#> 820 507.9912   2.602558
#> 821 504.1625   3.031749
#> 822 545.1769   2.662123
#> 823 849.7261   2.695021
#> 824 535.8650   2.597274
#> 825 662.9430   2.639836
#> 826 501.8631   2.742224
#> 827 507.9474   2.685033
#> 828 501.9805   2.596722
#> 829 520.6228   2.684115
#> 830 501.4983   3.667823
#> 831 962.2819   3.058951
#> 832 500.1402   3.057520
#> 833 500.1985   3.474594
#> 834 500.0664   2.709052
#> 835 509.5837   2.540132
#> 836 500.2857   2.524408
#> 837 500.1053   2.833783
#> 838 584.7238   2.813126
#> 839 500.2772   2.563946
#> 840 501.0293   2.876950
#> 841 518.8393   3.294515
#> 842 613.2421   2.731954
#> 843 500.0797   3.078337
#> 844 501.3676   2.991726
#> 845 500.3873   2.898804
#> 846 528.8252   2.735887
#> 847 500.0839   2.627282
#> 848 500.0179   2.527962
#> 849 500.4876   2.584383
#> 850 500.6733   2.591962
#> 851 503.0957   3.345698
#> 852 500.0400   2.702264
#> 853 500.1645   2.716247
#> 854 500.1643   2.714952
#> 855 501.1677   2.708532
#> 856 502.0497   2.680372
#> 857 501.0188   3.358977
#> 858 857.4219   2.507081
#> 859 550.6075   2.927934
#> 860 507.2281   2.535665
#> 861 500.0986   2.564708
#> 862 500.0752   2.578180
#> 863 501.0749   2.793596
#> 864 500.0327   2.862781
#> 865 501.9593   3.037493
#> 866 924.3719   2.520402
#> 867 507.3502   2.756210
#> 868 500.3010   2.783109
#> 869 500.6795   2.735855
#> 870 606.6740   4.322912
#> 871 500.0012   2.967185
#> 872 577.9969   2.980197
#> 873 502.2385   2.996180
#> 874 974.9148   3.017951
#> 875 500.5399   2.500493
#> 876 503.5052   2.658047
#> 877 501.3354   2.700972
#> 878 500.0295   2.594249
#> 879 503.3216   3.192388
#> 880 501.1362   2.805106
#> 881 500.0691   3.449335
#> 882 500.4055   2.928902
#> 883 524.2031   2.862905
#> 884 517.6018   3.257088
#> 885 522.2597   2.580502
#> 886 510.9569   3.090339
#> 887 531.7314   3.135937
#> 888 503.8564   2.535849
#> 889 500.1205   2.549506
#> 890 507.5432   2.769541
#> 891 532.1523   2.722753
#> 892 911.2579   3.175838
#> 893 500.0139   2.963014
#> 894 503.9521   2.512431
#> 895 500.3814   2.684462
#> 896 509.0630   2.556118
#> 897 878.6773   2.536010
#> 898 773.6179   2.861991
#> 899 501.3057   3.210623
#> 900 500.2323   2.819860
#> 901 512.0054   3.199758
#> 902 513.5040   2.598893
#> 903 512.2907   2.879884
#> 904 556.4035   2.918985
#> 905 523.2654   2.558878
#> 906 898.5036   3.372314
#> 907 665.9320   2.530646
#> 908 501.4435   2.655602
#> 909 500.1448   3.718406
#> 910 501.2412   2.765938
#> 911 505.0413   2.687229
#> 912 910.5807   2.655851