Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   839.0681   2.812835
#> 2   500.5533   2.714137
#> 3   958.3803   2.872477
#> 4   500.0136   2.558556
#> 5   541.7242   2.776819
#> 6   500.5180   2.537129
#> 7   535.1701   3.076146
#> 8   661.9486   4.759154
#> 9   500.7291   3.215745
#> 10  890.8926   3.371824
#> 11  502.3720   2.513841
#> 12  515.2502   3.356352
#> 13  504.0117   2.561305
#> 14  506.5314   2.910175
#> 15  500.0790   4.307694
#> 16  606.6711   2.733636
#> 17  701.8820   3.023972
#> 18  522.5263   3.184122
#> 19  546.1727   2.737596
#> 20  500.2117   2.645815
#> 21  515.5849   3.111867
#> 22  500.0751   2.654500
#> 23  519.4195   3.161846
#> 24  823.2890   2.797239
#> 25  551.4958   2.502570
#> 26  503.9951   2.593062
#> 27  500.1335   2.695040
#> 28  512.2801   3.014496
#> 29  500.4385   2.688798
#> 30  510.4479   3.515325
#> 31  500.0990   2.998512
#> 32  501.8654   3.205241
#> 33  513.2155   2.652422
#> 34  507.8258   3.420068
#> 35  671.2432   2.948245
#> 36  515.2102   3.355789
#> 37  551.9762   2.798930
#> 38  502.7940   3.443293
#> 39  595.7261   2.508189
#> 40  507.8215   2.556806
#> 41  502.1362   2.576110
#> 42  500.1053   3.060981
#> 43  523.1482   2.570838
#> 44  500.2733   2.879379
#> 45  500.2313   2.512498
#> 46  518.5138   2.623724
#> 47  612.4543   2.801895
#> 48  500.6153   2.699098
#> 49  537.8468   2.703354
#> 50  795.2682   2.651932
#> 51  511.6533   2.537605
#> 52  501.6863   2.633665
#> 53  501.0966   2.557407
#> 54  500.4857   3.675483
#> 55  501.8179   2.791076
#> 56  502.2854   2.775113
#> 57  884.4914   2.605569
#> 58  500.0939   4.815822
#> 59  500.4923   3.281854
#> 60  500.6439   3.494023
#> 61  550.5978   3.131095
#> 62  500.2086   3.854162
#> 63  538.5428   3.502432
#> 64  500.1691   2.843941
#> 65  556.7082   2.545973
#> 66  589.2777   3.418445
#> 67  530.9726   2.888221
#> 68  830.8831   2.982318
#> 69  674.1657   2.570009
#> 70  502.2238   2.567024
#> 71  502.8445   2.658181
#> 72  613.4268   3.116297
#> 73  521.1870   2.572410
#> 74  504.1229   2.511311
#> 75  500.7498   2.967855
#> 76  500.0574   2.756125
#> 77  501.1839   3.241884
#> 78  500.0009   2.976162
#> 79  532.2852   2.658761
#> 80  541.6086   2.505700
#> 81  500.5380   2.771763
#> 82  502.4962   5.216434
#> 83  503.5304   4.010157
#> 84  504.4955   2.542197
#> 85  570.0713   3.232120
#> 86  530.5653   3.863754
#> 87  500.0540   2.521821
#> 88  500.1336   3.053713
#> 89  787.6844   2.588806
#> 90  545.0640   2.708688
#> 91  500.0381   3.161324
#> 92  501.1396   2.823387
#> 93  789.6950   2.953365
#> 94  506.5618   2.559043
#> 95  700.9336   2.536298
#> 96  576.3303   2.975223
#> 97  590.8100   2.839495
#> 98  500.2429   2.772969
#> 99  502.6114   3.506393
#> 100 725.9230   4.861784
#> 101 500.0187   2.835882
#> 102 502.0948   3.099414
#> 103 500.0493   2.623869
#> 104 500.5982   3.553767
#> 105 500.2329   2.986071
#> 106 552.4536   3.015847
#> 107 531.0392   2.673123
#> 108 500.4130   2.924027
#> 109 503.3059   3.104567
#> 110 500.6102   2.560152
#> 111 500.1798   2.690284
#> 112 813.9418   2.711128
#> 113 555.6941   2.528193
#> 114 502.6496   2.739551
#> 115 500.0206   2.751411
#> 116 500.1351   4.290747
#> 117 500.0096   2.505740
#> 118 501.4077   3.288808
#> 119 500.1197   2.517428
#> 120 533.9381   2.522263
#> 121 502.1534   2.675074
#> 122 500.1483   2.555984
#> 123 500.5968   2.508003
#> 124 661.9351   2.553020
#> 125 584.2469   2.856848
#> 126 546.0446   2.536215
#> 127 519.4762   2.653749
#> 128 509.7156   2.617169
#> 129 500.0167   2.702829
#> 130 500.0389   3.083378
#> 131 546.5948   3.497678
#> 132 622.5116   2.781986
#> 133 506.0750   2.519523
#> 134 501.0243   3.659063
#> 135 500.0181   2.674932
#> 136 537.0978   2.972509
#> 137 500.0653   2.530122
#> 138 507.6967   2.616094
#> 139 508.6110   2.840433
#> 140 655.8807   2.645105
#> 141 502.2376   3.015538
#> 142 500.0719   3.208930
#> 143 500.8904   2.593398
#> 144 509.1643   2.566828
#> 145 524.3992   2.597453
#> 146 500.0088   2.549569
#> 147 501.7507   2.739872
#> 148 503.0100   3.936601
#> 149 500.5226   3.230243
#> 150 500.2271   3.686111
#> 151 504.8919   2.577926
#> 152 525.2574   3.538712
#> 153 625.6536   2.510883
#> 154 571.0189   2.739521
#> 155 652.9553   3.191443
#> 156 502.7555   2.693953
#> 157 500.0191   2.751181
#> 158 510.0801   3.218240
#> 159 516.5230   2.841224
#> 160 501.9228   2.899096
#> 161 516.4440   3.068816
#> 162 541.7012   3.249401
#> 163 511.2760   3.213864
#> 164 525.2004   2.631230
#> 165 500.6381   2.908869
#> 166 511.4826   2.585493
#> 167 500.2411   3.619980
#> 168 500.2866   2.937655
#> 169 508.6707   2.563216
#> 170 518.1436   3.485676
#> 171 504.8330   2.755092
#> 172 500.0537   2.585627
#> 173 520.8925   3.339879
#> 174 501.8788   2.548341
#> 175 501.5362   3.181470
#> 176 500.6540   2.652735
#> 177 500.0304   2.518494
#> 178 500.1506   2.500840
#> 179 514.8763   2.910901
#> 180 967.4769   3.040657
#> 181 519.9936   3.000170
#> 182 500.4093   3.090750
#> 183 534.2942   2.771041
#> 184 952.5066   2.898035
#> 185 500.6021   2.777039
#> 186 500.2603   2.602298
#> 187 520.1695   2.881419
#> 188 500.0351   2.666844
#> 189 500.1061   3.025214
#> 190 500.8091   4.005162
#> 191 500.1269   2.830444
#> 192 500.0464   2.530308
#> 193 525.1351   2.638572
#> 194 500.1967   2.753354
#> 195 986.7945   2.851891
#> 196 514.4500   2.533992
#> 197 646.1235   2.737897
#> 198 661.6828   3.064501
#> 199 505.9071   3.527262
#> 200 502.3609   2.612926
#> 201 500.4132   3.691064
#> 202 541.3304   2.804415
#> 203 847.6069   3.354305
#> 204 502.4021   2.971110
#> 205 500.3119   3.353778
#> 206 500.7243   2.609185
#> 207 649.5113   2.735616
#> 208 503.4690   2.537100
#> 209 902.1718   3.807498
#> 210 500.0049   3.253933
#> 211 501.8389   2.605166
#> 212 500.0751   2.515766
#> 213 500.1223   2.581118
#> 214 526.9684   2.644022
#> 215 889.8430   2.547284
#> 216 500.0578   2.577387
#> 217 654.1061   2.772958
#> 218 500.2091   2.977842
#> 219 564.6934   3.127375
#> 220 500.3481   2.502989
#> 221 500.0625   2.631870
#> 222 506.4825   2.864617
#> 223 645.3649   2.581717
#> 224 503.9964   2.619952
#> 225 502.8714   2.647114
#> 226 652.5427   2.646907
#> 227 513.8462   2.856762
#> 228 500.0376   3.607059
#> 229 501.3730   3.064412
#> 230 500.4809   3.432350
#> 231 501.9913   2.601209
#> 232 507.1715   2.697438
#> 233 522.2102   2.721419
#> 234 549.6655   3.118545
#> 235 518.6715   2.718401
#> 236 517.7323   2.932312
#> 237 504.1797   2.629981
#> 238 642.8297   2.552208
#> 239 771.2289   2.604138
#> 240 500.2045   2.905873
#> 241 500.2267   2.515090
#> 242 860.5469   2.626266
#> 243 509.2001   4.358617
#> 244 502.9562   3.026741
#> 245 500.1764   3.278881
#> 246 511.1692   4.549148
#> 247 500.1318   3.853749
#> 248 734.5638   2.551007
#> 249 513.3233   3.278006
#> 250 504.6617   2.981690
#> 251 527.0944   2.674107
#> 252 715.5034   3.100367
#> 253 533.7106   3.433804
#> 254 554.7604   2.709961
#> 255 506.4070   2.827766
#> 256 509.1879   3.204006
#> 257 504.3721   2.761775
#> 258 506.4486   2.880372
#> 259 509.5152   2.984451
#> 260 737.3279   2.547211
#> 261 500.3097   2.607846
#> 262 501.3090   3.116165
#> 263 503.3521   2.665663
#> 264 501.3341   2.576230
#> 265 568.7081   2.871013
#> 266 505.3847   2.517434
#> 267 558.9514   7.217448
#> 268 929.1680   3.144357
#> 269 504.0189   2.534450
#> 270 544.7557   2.695630
#> 271 502.5607   2.625657
#> 272 501.4473   2.782233
#> 273 500.0330   2.733434
#> 274 500.9433   2.523087
#> 275 503.1617   2.921100
#> 276 689.7256   2.574200
#> 277 500.1664   2.923810
#> 278 510.6895   2.801883
#> 279 500.1258   2.654773
#> 280 869.7649   2.931863
#> 281 853.4698   2.754524
#> 282 560.5206   2.525589
#> 283 688.6865   2.765532
#> 284 500.0423   2.879463
#> 285 500.2543   2.666599
#> 286 500.9317   2.592383
#> 287 504.6517   2.948517
#> 288 500.0290   2.896957
#> 289 501.6622   2.558807
#> 290 567.0009   2.860572
#> 291 525.7834   2.701420
#> 292 501.1318   2.512742
#> 293 500.3380   2.580295
#> 294 507.6889   2.696003
#> 295 502.8066   3.304708
#> 296 501.6506   3.252551
#> 297 500.2674   3.760935
#> 298 505.3405   3.028203
#> 299 504.8159   3.125954
#> 300 500.1035   2.812343
#> 301 926.8461   3.963166
#> 302 546.1879   2.567283
#> 303 505.4589   3.836316
#> 304 534.2149   2.691343
#> 305 501.6633   3.982576
#> 306 509.4819   3.041036
#> 307 908.3954   2.680950
#> 308 500.1200   3.266508
#> 309 874.4928   3.192016
#> 310 520.1994   3.191926
#> 311 500.3426   3.226868
#> 312 502.1685   3.371646
#> 313 540.0898   2.552330
#> 314 536.0129   2.709117
#> 315 500.1147   2.588582
#> 316 501.9045   2.500470
#> 317 508.8998   3.080124
#> 318 502.3087   3.529508
#> 319 502.0418   2.563029
#> 320 510.6800   2.619773
#> 321 505.3524   2.923496
#> 322 500.6504   2.833899
#> 323 921.5071   2.868661
#> 324 501.9321   2.726970
#> 325 500.0702   2.520018
#> 326 500.0528   2.591596
#> 327 502.2486   2.689458
#> 328 517.7782   2.797500
#> 329 500.1002   2.672623
#> 330 526.8833   2.889644
#> 331 544.9401   2.878901
#> 332 501.3692   2.732762
#> 333 500.3066   2.831761
#> 334 516.8059   3.300502
#> 335 500.6911   2.887437
#> 336 500.4304   2.824183
#> 337 500.9063   2.614792
#> 338 530.3646   2.630159
#> 339 500.0089   2.528021
#> 340 500.8664   2.588485
#> 341 510.1538   2.623397
#> 342 573.0164   3.147630
#> 343 513.0060   3.078839
#> 344 500.3004   2.787693
#> 345 504.8884   3.144827
#> 346 500.1772   2.710792
#> 347 501.2017   2.590906
#> 348 932.2991   2.549209
#> 349 838.0644   3.259669
#> 350 500.0392   2.765717
#> 351 501.2166   2.612298
#> 352 618.7341   2.518644
#> 353 500.9513   2.839447
#> 354 500.0626   4.351115
#> 355 754.2024   2.554801
#> 356 501.7060   2.634978
#> 357 500.6756   2.514241
#> 358 856.9716   2.514295
#> 359 688.2392   2.511082
#> 360 502.4498   3.467062
#> 361 801.1412   2.570030
#> 362 500.0477   4.365252
#> 363 560.6111   2.601297
#> 364 511.5549   2.595302
#> 365 500.6736   3.358362
#> 366 504.0527   3.761214
#> 367 500.9704   2.524205
#> 368 648.2021   2.813412
#> 369 500.4743   2.970905
#> 370 502.4759   2.740327
#> 371 502.9496   3.372292
#> 372 509.5336   3.361916
#> 373 917.0684   2.575418
#> 374 500.3410   2.939580
#> 375 740.7833   3.456902
#> 376 500.3289   2.500118
#> 377 500.0392   2.544346
#> 378 722.9006   2.553098
#> 379 547.8281   2.667889
#> 380 500.0374   3.010895
#> 381 521.6994   3.709516
#> 382 502.0422   3.026714
#> 383 502.6189   2.644486
#> 384 779.1174   3.330318
#> 385 835.1406   2.696121
#> 386 500.5011   2.851769
#> 387 500.0999   2.733088
#> 388 541.8235   3.024492
#> 389 500.0708   2.904667
#> 390 500.0137   2.538095
#> 391 970.5987   3.379420
#> 392 731.8357   2.573727
#> 393 501.8726   3.463944
#> 394 617.5042   3.785497
#> 395 500.2399   2.509994
#> 396 510.9873   3.240924
#> 397 609.3989   2.538441
#> 398 500.2113   2.555225
#> 399 518.8493   3.003932
#> 400 526.6460   2.705881
#> 401 512.4917   2.633633
#> 402 501.6364   3.675269
#> 403 554.2355   3.641883
#> 404 502.0309   2.505176
#> 405 501.2764   4.396991
#> 406 521.2741   2.698361
#> 407 500.5338   2.815300
#> 408 500.0173   3.675566
#> 409 500.0158   2.637632
#> 410 503.8685   3.095691
#> 411 504.8359   2.566896
#> 412 500.5289   2.573393
#> 413 500.1402   2.704010
#> 414 500.1504   2.784213
#> 415 669.4596   2.557556
#> 416 500.0828   3.179059
#> 417 505.7041   2.518247
#> 418 503.4679   2.547798
#> 419 625.8718   2.648915
#> 420 500.5424   2.539689
#> 421 500.2951   3.743255
#> 422 500.0458   2.557728
#> 423 503.8148   2.541581
#> 424 523.7566   2.559875
#> 425 503.7260   2.626301
#> 426 500.4897   2.776789
#> 427 507.6445   2.654512
#> 428 502.2131   2.632238
#> 429 500.0368   2.788892
#> 430 500.9513   3.525446
#> 431 501.7082   2.520814
#> 432 501.6275   3.175086
#> 433 500.0377   2.825989
#> 434 500.0922   2.927227
#> 435 596.4158   3.663523
#> 436 510.2180   2.694054
#> 437 500.5927   3.213407
#> 438 500.1129   2.867045
#> 439 546.2779   2.625191
#> 440 576.5376   3.394706
#> 441 501.8751   2.769544
#> 442 501.5803   3.551358
#> 443 514.0682   2.889838
#> 444 516.6627   3.412531
#> 445 505.2714   2.723183
#> 446 500.6516   2.981914
#> 447 568.6357   2.815143
#> 448 500.2037   2.530603
#> 449 500.0502   3.765869
#> 450 500.0166   3.378283
#> 451 534.2285   2.805153
#> 452 858.7437   3.082643
#> 453 500.6927   2.664981
#> 454 500.5863   3.040543
#> 455 500.5263   2.549165
#> 456 504.0160   2.724535
#> 457 516.2363   3.555683
#> 458 508.1819   3.497634
#> 459 653.9294   2.666961
#> 460 524.2271   2.986588
#> 461 525.1159   2.610507
#> 462 523.6268   2.572993
#> 463 541.3976   4.445404
#> 464 523.3589   2.552173
#> 465 587.2080   2.567287
#> 466 503.5565   2.865575
#> 467 501.2615   3.357656
#> 468 684.9327   2.610681
#> 469 533.3920   2.787080
#> 470 500.0980   2.501079
#> 471 526.6187   2.646238
#> 472 720.3729   2.519246
#> 473 802.6008   2.602197
#> 474 510.6294   2.502788
#> 475 500.7657   2.599414
#> 476 518.4042   2.557920
#> 477 500.0337   2.535551
#> 478 617.3097   4.462250
#> 479 966.7051   2.973604
#> 480 500.0403   3.374179
#> 481 848.6732   2.601149
#> 482 505.7688   2.659290
#> 483 545.1876   2.780803
#> 484 500.0094   2.601060
#> 485 500.2260   2.578258
#> 486 697.1475   2.864713
#> 487 502.1154   2.507953
#> 488 500.0405   2.708122
#> 489 500.7217   2.870450
#> 490 603.2301   2.851641
#> 491 564.8818   2.509620
#> 492 500.1837   2.623431
#> 493 501.7971   2.696170
#> 494 500.2526   2.615866
#> 495 500.0809   2.827954
#> 496 504.2062   2.983167
#> 497 560.7814   2.630074
#> 498 508.7635   2.504777
#> 499 502.7556   2.680179
#> 500 508.6183   3.707508
#> 501 500.0993   2.573635
#> 502 500.0474   2.869881
#> 503 529.6265   3.140237
#> 504 508.4928   2.649917
#> 505 507.4436   2.646905
#> 506 500.2051   2.678190
#> 507 504.6137   2.627494
#> 508 500.0628   2.551732
#> 509 506.2913   3.459865
#> 510 511.3104   3.464955
#> 511 500.2927   2.698168
#> 512 919.1353   2.591031
#> 513 569.7535   2.578566
#> 514 500.1369   2.511894
#> 515 501.1316   2.565444
#> 516 595.8472   3.202402
#> 517 500.1660   2.625528
#> 518 527.4771   2.500491
#> 519 503.2682   2.593001
#> 520 515.4730   2.627688
#> 521 588.8294   3.929591
#> 522 501.2842   2.941594
#> 523 501.4052   3.096732
#> 524 506.1216   3.040631
#> 525 504.8683   3.013395
#> 526 500.7580   2.733439
#> 527 503.1160   3.068393
#> 528 500.0068   2.501902
#> 529 500.0242   2.884222
#> 530 592.3682   3.238796
#> 531 500.3707   2.720509
#> 532 519.6358   2.620822
#> 533 501.1768   2.563890
#> 534 621.6048   2.517305
#> 535 501.7517   3.341792
#> 536 611.9786   2.916933
#> 537 500.6897   3.788097
#> 538 516.8953   2.751957
#> 539 503.4783   2.698647
#> 540 503.0260   3.248649
#> 541 503.9757   3.310529
#> 542 502.0885   2.557435
#> 543 778.9303   3.892075
#> 544 536.0734   2.575235
#> 545 531.7294   2.784924
#> 546 894.6624   2.761014
#> 547 519.2632   3.115946
#> 548 502.0418   2.675978
#> 549 533.5661   2.526662
#> 550 503.2044   2.569249
#> 551 668.9498   3.523300
#> 552 500.2873   2.692840
#> 553 557.5612   2.534577
#> 554 500.3173   2.520677
#> 555 500.9773   2.909432
#> 556 522.6069   2.627776
#> 557 501.3411   3.412926
#> 558 542.0203   2.593389
#> 559 500.0689   2.526676
#> 560 856.7079   2.699369
#> 561 500.5640   2.632837
#> 562 525.8985   2.927613
#> 563 870.9533   2.887240
#> 564 502.2278   2.802448
#> 565 686.9763   2.595296
#> 566 500.7889   3.651323
#> 567 502.0649   2.538490
#> 568 501.8677   3.131406
#> 569 500.0484   2.599717
#> 570 514.6373   3.242911
#> 571 500.6580   2.622305
#> 572 500.3621   2.676807
#> 573 855.1283   2.688418
#> 574 500.0066   2.676753
#> 575 512.7049   2.890110
#> 576 500.3541   2.814430
#> 577 500.0349   4.104629
#> 578 598.8521   3.258947
#> 579 500.0085   2.746187
#> 580 515.1241   2.519680
#> 581 506.0880   2.575345
#> 582 533.7121   2.676419
#> 583 501.5062   3.894266
#> 584 503.3677   2.681788
#> 585 908.9794   2.961690
#> 586 583.6376   2.709015
#> 587 990.3426   3.076497
#> 588 500.3431   2.554946
#> 589 501.9385   2.554274
#> 590 996.0036   3.251834
#> 591 500.5335   3.041913
#> 592 500.7698   2.618027
#> 593 501.4424   2.817561
#> 594 501.4042   2.654490
#> 595 501.3052   2.823017
#> 596 510.3457   3.517801
#> 597 630.3753   4.702397
#> 598 588.6659   3.042782
#> 599 500.0560   2.551574
#> 600 508.9844   3.176156
#> 601 951.4754   4.232236
#> 602 500.4628   2.543757
#> 603 559.6795   2.903666
#> 604 671.4142   5.235950
#> 605 540.9630   2.664216
#> 606 562.0386   3.280439
#> 607 508.5488   2.958419
#> 608 510.0701   2.998090
#> 609 503.3785   3.158923
#> 610 525.8735   2.542363
#> 611 503.7397   3.045896
#> 612 500.0681   2.516771
#> 613 500.3610   2.640578
#> 614 509.9138   2.667719
#> 615 500.6329   2.614953
#> 616 500.7346   3.254007
#> 617 625.8363   2.665095
#> 618 500.0113   2.862501
#> 619 501.5065   2.873866
#> 620 500.8222   2.574368
#> 621 502.7171   2.610168
#> 622 552.0043   2.530398
#> 623 500.6762   2.537815
#> 624 500.3215   3.151303
#> 625 500.6737   2.503120
#> 626 555.2773   3.351021
#> 627 535.9687   3.007452
#> 628 956.7313   4.411418
#> 629 501.5879   2.500554
#> 630 500.4078   2.685520
#> 631 588.2135   2.670921
#> 632 500.9097   3.314156
#> 633 503.9921   3.051159
#> 634 503.8886   3.117780
#> 635 503.6190   3.285017
#> 636 541.2652   2.715859
#> 637 604.3758   3.014812
#> 638 500.3935   2.521061
#> 639 908.5864   2.659165
#> 640 501.0784   2.983222
#> 641 564.8706   3.494247
#> 642 566.5028   2.539323
#> 643 503.2095   3.063741
#> 644 500.1501   2.641069
#> 645 500.0596   2.511308
#> 646 512.9873   2.626514
#> 647 579.0355   3.983975
#> 648 500.1263   3.428522
#> 649 500.0682   2.582615
#> 650 502.8706   2.855054
#> 651 523.7579   2.605342
#> 652 500.0215   2.510798
#> 653 513.4015   2.663506
#> 654 739.2999   2.872404
#> 655 500.3476   3.120336
#> 656 500.4668   2.791211
#> 657 503.4227   2.728996
#> 658 503.5488   3.091296
#> 659 500.8783   2.869486
#> 660 500.0754   4.452651
#> 661 500.1909   2.654660
#> 662 500.6287   2.787145
#> 663 500.8972   2.790067
#> 664 500.0201   2.742242
#> 665 500.1079   3.569598
#> 666 561.7945   2.924953
#> 667 500.5805   2.826291
#> 668 527.0428   2.633864
#> 669 643.3762   2.904634
#> 670 502.0792   2.598698
#> 671 513.4158   2.751255
#> 672 500.1738   2.748896
#> 673 514.5901   4.108971
#> 674 500.2499   2.538250
#> 675 894.2354   2.577728
#> 676 500.9114   2.577735
#> 677 500.7203   2.553599
#> 678 500.2267   2.976461
#> 679 500.2503   2.596657
#> 680 500.1854   3.089454
#> 681 565.4867   4.208579
#> 682 500.1475   2.842477
#> 683 501.1808   2.552640
#> 684 501.4324   2.736130
#> 685 503.0941   2.976399
#> 686 501.0833   2.677522
#> 687 500.4520   2.564874
#> 688 955.9567   3.187586
#> 689 503.2418   2.581354
#> 690 500.6365   2.763634
#> 691 769.4937   3.249322
#> 692 514.6735   2.522905
#> 693 507.9828   2.843623
#> 694 502.7003   3.234042
#> 695 749.6399   3.316284
#> 696 500.8897   2.515624
#> 697 500.3130   3.653761
#> 698 525.1706   3.124952
#> 699 500.1806   4.294012
#> 700 500.6900   3.278770
#> 701 500.4860   2.695039
#> 702 514.7688   3.073072
#> 703 604.7869   3.205953
#> 704 530.9025   2.624022
#> 705 546.2718   3.376834
#> 706 848.4087   2.825207
#> 707 503.8770   2.862803
#> 708 503.9133   2.632814
#> 709 791.6175   2.505064
#> 710 500.2918   2.800000
#> 711 698.3148   2.635870
#> 712 502.9952   2.774047
#> 713 506.4920   2.509566
#> 714 516.5033   3.209426
#> 715 813.8199   3.088742
#> 716 500.6438   2.847428
#> 717 500.7160   2.532726
#> 718 503.5226   2.505419
#> 719 500.3796   2.682063
#> 720 500.5531   2.656568
#> 721 500.1581   2.659912
#> 722 511.0223   2.853668
#> 723 500.0982   3.206437
#> 724 500.4662   2.913813
#> 725 500.1055   2.577605
#> 726 501.2091   2.634952
#> 727 500.0581   3.346700
#> 728 501.0387   2.609366
#> 729 505.2815   3.067362
#> 730 567.8628   2.797447
#> 731 542.7450   2.949086
#> 732 851.5384   2.701829
#> 733 546.2318   3.005019
#> 734 501.2445   3.008332
#> 735 573.2843   3.066167
#> 736 503.8115   2.756524
#> 737 697.7486   2.835171
#> 738 501.6949   2.794245
#> 739 570.2628   2.740847
#> 740 500.2094   2.701732
#> 741 500.9252   2.702252
#> 742 519.0708   2.947362
#> 743 652.1395   2.632954
#> 744 976.4338   2.759505
#> 745 707.9913   2.915147
#> 746 508.6292   3.246333
#> 747 505.8151   2.570289
#> 748 500.3732   2.977059
#> 749 590.9520   2.577742
#> 750 500.2776   2.523146
#> 751 539.0720   2.557387
#> 752 500.0020   2.521371
#> 753 736.8132   2.944813
#> 754 500.0990   2.522510
#> 755 500.8524   2.545563
#> 756 506.3175   2.975595
#> 757 503.2161   2.835178
#> 758 507.1910   2.712401
#> 759 506.6799   2.586472
#> 760 551.7502   3.235077
#> 761 524.5997   2.536917
#> 762 838.9539   3.552206
#> 763 643.7872   2.773327
#> 764 558.7350   2.792998
#> 765 500.1482   2.544744
#> 766 500.0524   2.583990
#> 767 501.1608   2.584194
#> 768 754.4054   2.693871
#> 769 502.3368   3.239843
#> 770 500.4017   3.230576
#> 771 893.4079   3.050852
#> 772 505.5450   2.632510
#> 773 501.9115   3.008951
#> 774 504.0026   2.944667
#> 775 525.9706   2.734689
#> 776 500.4058   2.556175
#> 777 500.3961   2.550974
#> 778 524.0333   3.006725
#> 779 767.6373   2.751902
#> 780 504.7404   2.741991
#> 781 500.7833   2.666069
#> 782 549.6412   2.503191
#> 783 531.9912   3.640432
#> 784 504.3592   2.891504
#> 785 732.7290   2.550207
#> 786 513.6575   2.617771
#> 787 500.5013   3.002424
#> 788 519.8709   2.664712
#> 789 500.1274   2.860826
#> 790 642.5249   3.851267
#> 791 526.9707   2.664199
#> 792 500.0239   2.895543
#> 793 501.2156   4.181533
#> 794 577.1880   2.720806
#> 795 500.0374   4.442396
#> 796 500.1417   2.803335
#> 797 500.0076   3.503552
#> 798 502.8693   3.642793
#> 799 509.4596   2.573964
#> 800 730.1018   3.624597
#> 801 500.0600   2.633793
#> 802 502.9871   2.942388
#> 803 551.6955   2.669923
#> 804 500.0328   3.592106
#> 805 500.0121   3.428463
#> 806 510.7019   2.749389
#> 807 577.1557   2.847766
#> 808 500.0437   3.375687
#> 809 503.8401   2.870599
#> 810 500.6626   3.136797
#> 811 500.3360   2.955727
#> 812 501.6986   3.109144
#> 813 500.1679   2.596376
#> 814 500.0692   3.302968
#> 815 709.5950   3.006209
#> 816 647.4535   2.577239
#> 817 870.8202   2.671973
#> 818 502.2331   3.326008
#> 819 502.1181   3.175581
#> 820 981.8334   2.729095
#> 821 510.2323   3.143899
#> 822 500.5750   2.786385
#> 823 518.6549   2.679787
#> 824 519.3955   2.621378
#> 825 500.2175   2.831338
#> 826 500.0106   2.939806
#> 827 500.9061   2.595990
#> 828 500.7380   2.618877
#> 829 502.8991   2.640240
#> 830 500.6380   2.524195
#> 831 705.9212   2.839708
#> 832 501.0249   2.905433
#> 833 501.7229   4.318120
#> 834 500.5722   2.881292
#> 835 500.2816   2.545316
#> 836 500.2322   3.110035
#> 837 551.9867   3.220366
#> 838 508.9637   2.593453
#> 839 500.0732   2.949396
#> 840 500.7533   2.557834
#> 841 500.0967   2.822935
#> 842 901.4966   2.877495
#> 843 500.4503   3.616953
#> 844 500.0054   2.749776
#> 845 500.8285   2.973911
#> 846 507.8919   3.268260
#> 847 802.4094   2.861841
#> 848 623.3016   3.203786
#> 849 500.0272   2.629182
#> 850 500.0866   2.852841
#> 851 615.4028   3.186130
#> 852 500.1632   2.618760
#> 853 500.3953   3.205935
#> 854 500.1982   2.852754
#> 855 510.0243   3.412213
#> 856 502.1742   3.151344
#> 857 501.3327   3.128229
#> 858 525.5627   2.801669
#> 859 598.5871   2.804880
#> 860 510.9189   4.661854
#> 861 500.2333   2.540307
#> 862 501.7217   2.526112
#> 863 978.4459   2.608404
#> 864 557.4299   3.533032
#> 865 501.4107   2.689453
#> 866 528.6552   2.538631
#> 867 502.2576   3.220862
#> 868 500.4693   2.796856
#> 869 873.5309   2.700960
#> 870 504.0034   3.096317
#> 871 500.2566   3.320746
#> 872 500.8255   2.568434
#> 873 548.0577   3.281506
#> 874 500.1646   2.978749
#> 875 500.3582   3.288216
#> 876 502.2512   2.826842
#> 877 524.4623   2.622887
#> 878 518.8459   2.658846
#> 879 535.2730   2.762966
#> 880 500.6276   2.929612
#> 881 506.4206   2.714171
#> 882 506.3159   2.642509
#> 883 505.3762   2.502611
#> 884 504.5143   3.339937
#> 885 500.0127   2.972359
#> 886 500.1303   3.654938
#> 887 502.1136   2.680900
#> 888 500.6060   2.898234
#> 889 560.5573   2.929006
#> 890 500.2956   2.501465
#> 891 812.4267   2.873199
#> 892 500.4580   2.521213
#> 893 500.4045   3.371604
#> 894 503.3471   2.799051
#> 895 541.9417   2.529589
#> 896 582.0305   2.898990
#> 897 513.8946   2.649153
#> 898 501.4716   3.009552
#> 899 505.0674   2.712962
#> 900 886.4363   2.675339
#> 901 516.8226   2.784456
#> 902 945.1634   2.537013
#> 903 500.3282   2.523417
#> 904 504.8419   2.652969
#> 905 500.3228   2.600695
#> 906 503.0936   2.926483
#> 907 557.6147   3.698718
#> 908 500.0676   2.807640
#> 909 500.2487   2.712914
#> 910 500.3039   2.739495
#> 911 502.6426   3.912362
#> 912 500.0617   2.519190
#> 913 502.4405   2.725866
#> 914 500.3669   2.661208
#> 915 533.1250   2.987079
#> 916 502.7486   2.577838
#> 917 500.3652   3.779003
#> 918 500.2155   2.524499
#> 919 512.5380   3.537536
#> 920 505.7153   2.565459
#> 921 505.1448   2.808705
#> 922 763.4445   2.662533
#> 923 501.2456   2.849157
#> 924 500.0177   2.818244
#> 925 565.8331   3.699957