Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   500.7605   2.930985
#> 2   501.1970   2.618127
#> 3   537.7771   2.647316
#> 4   821.7276   2.677680
#> 5   501.6092   2.756032
#> 6   501.7713   2.591493
#> 7   521.1706   3.415213
#> 8   500.0352   3.165652
#> 9   776.6610   2.579943
#> 10  549.5208   4.348280
#> 11  508.6197   2.937116
#> 12  500.5331   2.730712
#> 13  501.6712   2.582719
#> 14  712.5166   3.575193
#> 15  948.9483   2.621986
#> 16  500.1243   2.549154
#> 17  500.0719   2.982804
#> 18  500.2642   2.950538
#> 19  504.1822   2.513267
#> 20  660.6681   2.783916
#> 21  573.7027   2.717412
#> 22  501.0032   3.608407
#> 23  500.4316   2.575153
#> 24  500.1831   3.585415
#> 25  501.2044   2.764161
#> 26  500.2368   2.711998
#> 27  500.9821   2.992522
#> 28  500.7591   3.871434
#> 29  589.9787   2.712338
#> 30  501.8462   3.060902
#> 31  521.5507   2.534252
#> 32  542.7640   2.872787
#> 33  501.4792   3.194304
#> 34  502.3339   2.536439
#> 35  575.9376   3.167279
#> 36  500.0502   3.426218
#> 37  569.6219   2.904947
#> 38  630.7166   3.397623
#> 39  517.6345   2.986921
#> 40  705.2378   2.582624
#> 41  502.9938   2.733608
#> 42  500.1386   2.792730
#> 43  979.8766   2.607919
#> 44  500.2590   4.943848
#> 45  500.3921   2.897601
#> 46  730.0560   3.826041
#> 47  509.3321   2.933630
#> 48  963.3569   2.635770
#> 49  525.8042   2.510232
#> 50  593.2909   3.223874
#> 51  529.9022   3.648703
#> 52  629.1781   2.701580
#> 53  500.1667   3.636890
#> 54  500.7907   3.118541
#> 55  504.0652   2.818578
#> 56  500.7153   2.708277
#> 57  504.8597   3.119617
#> 58  592.8334   2.583781
#> 59  512.8545   2.601552
#> 60  500.2576   2.517640
#> 61  500.2099   2.674022
#> 62  500.3925   2.513595
#> 63  500.0096   3.045669
#> 64  500.9313   2.887421
#> 65  516.0955   2.727046
#> 66  501.2930   2.706182
#> 67  507.4727   2.716056
#> 68  721.1789   3.009306
#> 69  501.8311   2.675669
#> 70  523.8113   2.983866
#> 71  500.3024   2.784507
#> 72  518.3228   2.894329
#> 73  501.2879   3.251238
#> 74  506.1210   2.813306
#> 75  500.0554   3.215253
#> 76  501.0946   2.864735
#> 77  524.6109   2.779504
#> 78  620.4297   2.952220
#> 79  828.7731   2.608024
#> 80  500.0775   3.013122
#> 81  500.6868   3.464526
#> 82  503.5652   2.799032
#> 83  537.3958   2.501432
#> 84  500.3010   3.100452
#> 85  526.5569   2.702325
#> 86  503.5219   2.895045
#> 87  512.2624   3.257564
#> 88  862.7786   2.509804
#> 89  538.9431   3.235566
#> 90  500.0273   2.806914
#> 91  510.1202   2.798289
#> 92  627.0694   3.416344
#> 93  504.6004   2.832957
#> 94  500.1303   2.567900
#> 95  547.9731   3.665979
#> 96  802.6323   3.613421
#> 97  905.5880   2.555828
#> 98  500.1908   2.955728
#> 99  669.0826   2.723275
#> 100 500.3005   2.606957
#> 101 894.1823   2.949047
#> 102 512.1254   2.503451
#> 103 501.0013   2.593819
#> 104 500.0806   2.771408
#> 105 505.9383   2.560984
#> 106 518.7703   3.283144
#> 107 500.1434   3.316189
#> 108 502.8437   3.678022
#> 109 700.5432   2.918397
#> 110 500.1761   2.633618
#> 111 507.8014   2.521607
#> 112 500.8603   2.602177
#> 113 612.0070   2.619796
#> 114 500.0033   3.495602
#> 115 503.8950   2.865686
#> 116 529.7120   2.513518
#> 117 710.3306   2.514434
#> 118 735.8094   2.695803
#> 119 521.8519   2.786552
#> 120 690.6515   2.969117
#> 121 567.7186   3.225912
#> 122 509.0637   2.799828
#> 123 649.1177   2.665070
#> 124 500.9103   2.693661
#> 125 500.0370   2.653234
#> 126 501.3646   3.328485
#> 127 500.4157   2.575980
#> 128 525.4528   2.656289
#> 129 501.2740   2.613803
#> 130 524.6218   2.544192
#> 131 506.7496   2.576851
#> 132 502.8058   3.982894
#> 133 532.7004   2.827637
#> 134 506.4526   3.056443
#> 135 518.7719   2.611677
#> 136 501.0201   2.530951
#> 137 501.8797   3.165860
#> 138 500.0427   3.215978
#> 139 504.6436   2.892369
#> 140 505.5110   3.239299
#> 141 503.1678   2.602349
#> 142 500.0322   2.560179
#> 143 502.2543   2.529271
#> 144 500.1107   2.650179
#> 145 512.5444   2.612079
#> 146 587.3460   2.656219
#> 147 501.2692   3.616821
#> 148 563.4604   2.882390
#> 149 503.7082   3.240421
#> 150 528.6976   2.510688
#> 151 503.2817   2.706042
#> 152 727.0360   2.582509
#> 153 500.4652   2.532549
#> 154 500.3224   2.724568
#> 155 500.0491   3.087683
#> 156 500.0040   2.630437
#> 157 500.4054   2.899703
#> 158 500.0928   2.730662
#> 159 500.1322   2.535885
#> 160 745.0785   2.921674
#> 161 505.0892   2.900475
#> 162 500.1130   2.809220
#> 163 734.3172   2.770305
#> 164 547.2798   3.210892
#> 165 502.5056   2.955475
#> 166 922.7531   2.825586
#> 167 575.4163   3.371120
#> 168 730.7674   2.515615
#> 169 516.1094   2.638519
#> 170 635.4378   2.754374
#> 171 509.9094   3.542337
#> 172 927.2507   2.724763
#> 173 516.4336   2.652504
#> 174 500.0030   3.842570
#> 175 501.4656   2.500645
#> 176 503.2185   2.578484
#> 177 500.2338   2.590329
#> 178 863.4807   2.939449
#> 179 518.5534   3.351753
#> 180 501.7633   2.754531
#> 181 827.7290   2.766094
#> 182 782.0552   2.557152
#> 183 500.3709   2.866490
#> 184 740.9550   2.667460
#> 185 500.0001   2.769667
#> 186 534.0307   3.088422
#> 187 500.6766   2.864190
#> 188 888.5423   2.825282
#> 189 786.8228   2.519598
#> 190 502.7828   2.642588
#> 191 504.3441   3.288485
#> 192 725.1861   3.274368
#> 193 507.0329   2.713222
#> 194 501.5019   2.623608
#> 195 500.6695   2.537813
#> 196 509.7920   2.865968
#> 197 548.6412   2.516626
#> 198 500.0235   2.579864
#> 199 500.1753   2.771370
#> 200 509.5525   3.019963
#> 201 527.0164   2.968474
#> 202 502.5188   2.519361
#> 203 502.0618   2.700575
#> 204 500.3252   3.024603
#> 205 503.9195   2.799147
#> 206 500.2548   3.305884
#> 207 500.2914   2.602539
#> 208 515.7295   3.331209
#> 209 502.3456   3.270810
#> 210 515.3310   2.538902
#> 211 507.0273   3.219932
#> 212 787.5818   3.670621
#> 213 559.9324   2.526154
#> 214 500.3928   2.600581
#> 215 510.5565   2.865346
#> 216 500.4529   2.502753
#> 217 501.9382   2.563782
#> 218 550.3644   2.632348
#> 219 501.5071   3.329690
#> 220 502.9593   2.673633
#> 221 604.8608   3.520069
#> 222 607.8992   2.664727
#> 223 500.1007   2.732486
#> 224 505.2199   3.259114
#> 225 532.3031   2.560648
#> 226 500.0692   2.649343
#> 227 713.5157   2.860707
#> 228 500.0059   2.731342
#> 229 812.8997   2.634906
#> 230 941.1836   2.609178
#> 231 501.0871   2.868210
#> 232 502.2948   3.938885
#> 233 500.1628   3.198262
#> 234 500.0951   2.694611
#> 235 513.8536   2.902566
#> 236 500.3694   3.028005
#> 237 500.0492   2.556015
#> 238 626.1233   2.629404
#> 239 504.0146   2.513403
#> 240 588.0073   3.299368
#> 241 500.1439   3.124287
#> 242 500.1572   2.852964
#> 243 531.7669   2.565986
#> 244 503.3589   3.018435
#> 245 500.9313   3.337656
#> 246 500.2708   2.902764
#> 247 500.0827   2.857497
#> 248 525.7941   3.740532
#> 249 500.5427   3.005823
#> 250 533.8611   2.731074
#> 251 500.2868   2.901570
#> 252 501.0095   2.981948
#> 253 622.9796   2.823899
#> 254 500.2113   2.553929
#> 255 689.2172   2.928630
#> 256 500.0378   2.506558
#> 257 584.4318   4.575222
#> 258 501.0814   3.197039
#> 259 903.8294   3.063979
#> 260 654.9418   2.758373
#> 261 532.7808   2.849451
#> 262 500.6111   2.528852
#> 263 656.3951   2.556711
#> 264 651.7616   3.102741
#> 265 505.1589   2.513273
#> 266 500.4578   2.710761
#> 267 546.3822   3.046479
#> 268 500.2842   2.914896
#> 269 500.7255   2.579889
#> 270 511.4142   3.053386
#> 271 816.2619   3.300730
#> 272 500.0924   2.613403
#> 273 626.7513   3.674762
#> 274 501.5090   2.551856
#> 275 500.0842   2.616951
#> 276 502.3979   2.621101
#> 277 501.4707   2.844258
#> 278 500.2445   2.759152
#> 279 503.0833   4.206893
#> 280 528.3477   2.878949
#> 281 505.5365   2.739928
#> 282 500.2905   4.025108
#> 283 500.6492   2.975262
#> 284 500.5621   2.605100
#> 285 502.3455   2.872220
#> 286 500.3319   2.702452
#> 287 506.7873   3.361738
#> 288 502.7653   4.410346
#> 289 500.6329   2.616580
#> 290 898.3403   2.671167
#> 291 501.7950   3.320841
#> 292 500.0736   2.757245
#> 293 509.5020   2.708680
#> 294 609.8306   2.629908
#> 295 500.7842   3.071222
#> 296 500.1414   2.979258
#> 297 925.4746   2.932694
#> 298 507.9821   2.647361
#> 299 897.1405   2.905351
#> 300 747.3222   4.203572
#> 301 500.5537   3.687337
#> 302 926.4462   2.915881
#> 303 733.9841   3.100060
#> 304 842.0686   2.916641
#> 305 514.8742   2.844276
#> 306 500.0393   2.601872
#> 307 532.2579   2.814920
#> 308 500.0187   2.510763
#> 309 500.3178   2.931900
#> 310 609.9475   2.507033
#> 311 501.0842   2.870971
#> 312 500.4593   3.032079
#> 313 501.9784   3.070302
#> 314 500.0800   3.163905
#> 315 501.9817   2.866130
#> 316 516.1028   4.085648
#> 317 500.2959   2.991869
#> 318 503.6209   2.507963
#> 319 534.3860   2.578956
#> 320 500.3553   3.525316
#> 321 504.8022   2.834736
#> 322 505.1186   2.866629
#> 323 547.5133   3.203321
#> 324 525.7982   2.594283
#> 325 532.6211   3.053134
#> 326 507.2558   2.591336
#> 327 502.5787   3.167227
#> 328 896.7969   2.595932
#> 329 500.3369   2.699738
#> 330 602.0321   2.531801
#> 331 527.5228   2.604383
#> 332 515.5309   2.994924
#> 333 500.1226   2.665175
#> 334 515.9513   2.780341
#> 335 518.3713   2.652298
#> 336 503.6897   2.831189
#> 337 643.6897   3.608066
#> 338 509.1772   2.528164
#> 339 500.0762   2.549450
#> 340 853.0040   2.788704
#> 341 504.4536   3.262837
#> 342 514.5311   3.391322
#> 343 621.6707   2.676312
#> 344 500.0231   2.556399
#> 345 567.7811   2.545055
#> 346 874.7080   2.740015
#> 347 503.6532   2.846933
#> 348 500.0996   3.146500
#> 349 500.0161   4.409617
#> 350 635.5185   2.723028
#> 351 609.2380   2.694275
#> 352 688.7478   2.588485
#> 353 501.7085   2.871286
#> 354 501.7338   3.084907
#> 355 500.0046   2.554438
#> 356 502.3823   2.551831
#> 357 500.2662   2.548442
#> 358 514.6895   2.621498
#> 359 504.2215   2.608242
#> 360 723.0915   2.667578
#> 361 543.1857   2.943230
#> 362 502.7831   2.753636
#> 363 689.8825   3.163730
#> 364 500.1862   3.066926
#> 365 500.7671   2.936573
#> 366 523.1232   3.293156
#> 367 913.0438   2.611378
#> 368 800.1097   3.300067
#> 369 563.7622   2.599417
#> 370 501.3864   2.576356
#> 371 501.3351   3.337058
#> 372 500.2862   2.758412
#> 373 671.0089   3.618939
#> 374 500.1862   3.865318
#> 375 535.0703   2.644646
#> 376 500.6951   2.598063
#> 377 633.0174   2.534748
#> 378 529.3454   2.557043
#> 379 500.2158   2.851452
#> 380 821.5265   2.563595
#> 381 501.0941   2.859042
#> 382 532.6622   2.558974
#> 383 501.3077   2.707951
#> 384 500.0315   2.905566
#> 385 500.0470   3.716239
#> 386 500.0555   3.272602
#> 387 636.8770   2.966837
#> 388 526.7998   3.051633
#> 389 903.5381   2.693676
#> 390 500.2425   3.190981
#> 391 508.3216   3.051415
#> 392 505.1480   2.825065
#> 393 500.2296   2.740502
#> 394 512.7134   3.134839
#> 395 501.8295   4.607834
#> 396 500.4850   4.699832
#> 397 830.8102   2.514725
#> 398 500.3893   2.730195
#> 399 500.0280   2.581330
#> 400 500.5358   2.750958
#> 401 502.2476   2.621517
#> 402 998.2880   2.614976
#> 403 520.6064   2.762087
#> 404 500.0649   2.509449
#> 405 668.2064   4.670135
#> 406 626.0125   3.165866
#> 407 500.3499   2.952584
#> 408 523.1524   2.941754
#> 409 501.8748   3.065098
#> 410 502.0219   2.849674
#> 411 500.6258   3.550127
#> 412 500.7974   2.787174
#> 413 500.0204   2.614953
#> 414 500.1913   2.541713
#> 415 682.1220   2.951516
#> 416 500.0746   2.624455
#> 417 535.8979   2.642434
#> 418 500.6826   3.641527
#> 419 515.9225   2.778342
#> 420 500.0111   2.607362
#> 421 503.7420   2.754133
#> 422 500.2602   2.705291
#> 423 500.1209   2.503793
#> 424 505.0515   2.824571
#> 425 506.6262   2.605372
#> 426 836.5890   3.103974
#> 427 554.6921   2.956972
#> 428 502.5991   2.971845
#> 429 701.6046   2.634461
#> 430 500.9089   2.513591
#> 431 927.9050   2.535640
#> 432 524.6922   2.827628
#> 433 547.2241   3.843430
#> 434 500.0202   2.688805
#> 435 935.3906   2.517520
#> 436 573.7910   2.832438
#> 437 704.3372   2.637784
#> 438 939.3920   4.157791
#> 439 507.7251   2.677312
#> 440 500.0839   2.517086
#> 441 579.3256   2.642268
#> 442 529.3437   2.809729
#> 443 501.2069   2.671667
#> 444 500.2098   2.748201
#> 445 818.7185   2.820376
#> 446 500.0112   2.713307
#> 447 549.6652   3.469792
#> 448 500.0250   2.723357
#> 449 515.4682   2.687630
#> 450 617.9594   2.631563
#> 451 503.0306   2.821465
#> 452 502.5615   2.999068
#> 453 500.1407   2.547918
#> 454 694.5060   2.553498
#> 455 709.9519   3.472374
#> 456 504.6974   2.925153
#> 457 500.6346   3.259513
#> 458 501.6555   3.564175
#> 459 508.6152   2.962379
#> 460 975.2801   3.953280
#> 461 500.7765   2.822702
#> 462 502.7621   2.913619
#> 463 500.2689   2.684407
#> 464 501.6713   2.666717
#> 465 500.1844   2.586506
#> 466 506.6599   3.233946
#> 467 512.8240   3.177607
#> 468 500.4280   2.591799
#> 469 500.0813   2.769182
#> 470 500.2525   2.574062
#> 471 502.6425   2.687786
#> 472 762.5532   2.861186
#> 473 500.0245   2.589645
#> 474 500.9070   2.991836
#> 475 503.1678   2.637109
#> 476 500.2825   2.665329
#> 477 500.2940   3.534126
#> 478 500.2164   2.851873
#> 479 549.5576   2.886444
#> 480 522.7568   2.535087
#> 481 501.7385   2.507618
#> 482 500.2348   3.719166
#> 483 500.2218   2.660414
#> 484 500.0834   2.502934
#> 485 500.0287   2.518297
#> 486 500.1845   2.608855
#> 487 592.7420   2.842435
#> 488 501.0816   2.873055
#> 489 501.1746   3.173791
#> 490 523.1182   3.759086
#> 491 501.0676   3.064117
#> 492 517.9549   3.121454
#> 493 500.3951   2.695240
#> 494 501.8678   2.941634
#> 495 500.4163   2.841799
#> 496 500.0051   2.621793
#> 497 500.5359   2.742100
#> 498 525.1389   3.463454
#> 499 500.0755   2.548804
#> 500 506.1134   2.834687
#> 501 507.9786   3.019621
#> 502 565.6117   3.198930
#> 503 515.0539   2.609041
#> 504 533.6131   2.835539
#> 505 531.9641   3.158831
#> 506 524.3947   2.656604
#> 507 502.9116   2.778612
#> 508 664.8289   2.660886
#> 509 685.3068   2.583654
#> 510 500.1297   2.596818
#> 511 500.1401   2.640677
#> 512 500.0536   2.587248
#> 513 500.2323   3.192529
#> 514 500.0110   2.788866
#> 515 500.9773   2.734181
#> 516 975.4517   2.696942
#> 517 594.2097   2.776356
#> 518 522.0627   2.951573
#> 519 500.0366   3.230530
#> 520 501.4116   3.443641
#> 521 622.8567   3.196720
#> 522 558.7041   2.630162
#> 523 500.4544   2.623408
#> 524 536.3267   2.580469
#> 525 500.0191   2.608948
#> 526 500.0045   3.407592
#> 527 504.8489   3.127801
#> 528 501.3657   2.530671
#> 529 504.2447   3.312313
#> 530 500.1148   2.529837
#> 531 500.5503   2.523593
#> 532 502.9671   3.096075
#> 533 500.0514   2.657140
#> 534 798.8722   2.991413
#> 535 500.8973   2.637670
#> 536 502.8885   2.654091
#> 537 500.4346   3.519993
#> 538 503.8941   2.754260
#> 539 500.2577   3.841684
#> 540 939.0404   3.673215
#> 541 514.1906   2.959732
#> 542 863.0775   3.296473
#> 543 537.2629   2.545015
#> 544 502.2844   3.090629
#> 545 505.0980   3.069669
#> 546 882.4408   2.523259
#> 547 809.9842   2.819310
#> 548 502.8688   2.858546
#> 549 500.0682   3.093537
#> 550 531.6458   2.916346
#> 551 501.5100   2.649683
#> 552 515.9054   2.714808
#> 553 516.1882   2.808518
#> 554 500.0063   2.516587
#> 555 609.2839   2.602333
#> 556 815.0790   3.118463
#> 557 500.0760   2.789499
#> 558 519.1484   2.642272
#> 559 500.1229   2.892924
#> 560 504.5235   2.789419
#> 561 788.7496   3.050113
#> 562 502.4351   2.560658
#> 563 541.6503   3.228372
#> 564 511.1870   3.027084
#> 565 500.0135   2.841163
#> 566 500.0695   2.934276
#> 567 500.0128   2.814523
#> 568 502.0546   2.983457
#> 569 570.1480   2.546460
#> 570 509.7519   2.860673
#> 571 506.2965   3.803596
#> 572 860.3736   2.513533
#> 573 508.6076   2.825061
#> 574 500.1149   2.631733
#> 575 500.4991   2.573384
#> 576 516.7699   3.071444
#> 577 509.0770   3.721845
#> 578 529.2531   2.577191
#> 579 536.0167   2.863564
#> 580 501.6962   2.619398
#> 581 500.8242   4.581940
#> 582 532.2465   4.013115
#> 583 510.9230   2.505579
#> 584 897.9820   2.843079
#> 585 625.3335   2.954960
#> 586 500.6840   2.907353
#> 587 500.2617   2.913689
#> 588 507.6090   2.950820
#> 589 500.3629   4.235782
#> 590 734.4908   2.830964
#> 591 500.1165   3.405429
#> 592 503.3041   2.639926
#> 593 908.8819   2.574851
#> 594 528.9488   3.159191
#> 595 500.3017   2.959635
#> 596 509.9248   2.582284
#> 597 508.0264   2.703871
#> 598 501.6146   2.821830
#> 599 500.0803   2.592282
#> 600 500.4305   4.108174
#> 601 507.3901   3.660979
#> 602 504.0498   2.781325
#> 603 500.2160   3.923904
#> 604 500.0068   3.383587
#> 605 502.7215   2.763662
#> 606 714.6531   3.019633
#> 607 513.1159   3.560312
#> 608 500.0909   2.916813
#> 609 704.4550   2.643847
#> 610 502.8228   3.585632
#> 611 520.5646   3.015585
#> 612 502.3631   2.748308
#> 613 526.6645   2.575363
#> 614 500.1628   3.235974
#> 615 500.8594   3.285938
#> 616 500.5300   2.588084
#> 617 616.8871   2.785950
#> 618 641.2069   2.870180
#> 619 500.4066   2.544842
#> 620 505.2754   3.344845
#> 621 534.9482   2.823103
#> 622 555.2761   2.808787
#> 623 809.0938   3.233693
#> 624 502.8806   2.939931
#> 625 500.1187   3.241871
#> 626 520.6774   2.561637
#> 627 506.2369   2.738194
#> 628 500.0834   2.803683
#> 629 507.4040   3.004109
#> 630 501.4255   2.664622
#> 631 519.4461   2.574388
#> 632 505.6871   3.274428
#> 633 500.0565   3.390203
#> 634 907.3100   2.698429
#> 635 500.6668   2.830561
#> 636 532.1735   3.230211
#> 637 611.2640   3.252883
#> 638 500.0975   3.188406
#> 639 501.6198   2.566133
#> 640 504.9722   2.546356
#> 641 500.2609   2.512728
#> 642 500.2219   3.312573
#> 643 500.0098   2.584318
#> 644 522.1014   2.779969
#> 645 500.2159   2.673154
#> 646 502.0382   2.912173
#> 647 500.3166   3.119889
#> 648 501.1520   2.550315
#> 649 500.0745   2.613670
#> 650 580.8613   2.736292
#> 651 507.3437   2.558052
#> 652 503.8467   2.938432
#> 653 500.3689   2.722682
#> 654 571.2384   2.598914
#> 655 582.2012   3.145672
#> 656 551.1544   4.097576
#> 657 512.3816   2.507651
#> 658 521.8052   3.440385
#> 659 500.0022   3.319131
#> 660 510.7176   2.788174
#> 661 524.7607   2.850875
#> 662 500.7214   3.298790
#> 663 540.9753   3.027989
#> 664 610.6877   3.433504
#> 665 500.1147   2.832917
#> 666 500.1016   2.975664
#> 667 637.8715   2.935968
#> 668 504.9758   2.562899
#> 669 508.8780   2.734332
#> 670 500.7614   2.772341
#> 671 500.2922   2.731133
#> 672 928.9058   2.618550
#> 673 500.8542   3.232957
#> 674 513.5437   3.576810
#> 675 694.9428   3.350880
#> 676 555.8248   2.712341
#> 677 500.8632   3.106043
#> 678 500.0102   2.661183
#> 679 505.0965   2.614344
#> 680 552.6090   3.087567
#> 681 544.4914   2.902634
#> 682 500.0535   2.658333
#> 683 672.8515   2.586195
#> 684 523.3568   2.696797
#> 685 646.9927   2.706511
#> 686 500.2779   2.525313
#> 687 512.5967   2.585662
#> 688 510.2503   3.824068
#> 689 500.8305   2.601581
#> 690 500.0969   2.515422
#> 691 500.4943   3.215453
#> 692 502.0506   3.017967
#> 693 500.0173   2.941300
#> 694 502.2836   2.661441
#> 695 667.8779   2.910231
#> 696 539.1693   2.513537
#> 697 516.0331   5.990178
#> 698 500.0570   3.229276
#> 699 525.7353   2.690647
#> 700 531.4309   3.770501
#> 701 500.1216   2.699846
#> 702 500.0578   2.625175
#> 703 500.3447   2.588033
#> 704 505.6630   3.073640
#> 705 649.6781   2.572819
#> 706 523.1583   2.514620
#> 707 526.9764   2.609299
#> 708 500.1830   3.320740
#> 709 505.0567   2.755979
#> 710 501.0978   2.705688
#> 711 523.0709   2.840127
#> 712 500.2211   3.426978
#> 713 641.9477   2.571293
#> 714 506.3951   2.994565
#> 715 506.0001   3.427563
#> 716 542.3555   2.503194
#> 717 500.4175   3.567522
#> 718 504.9231   2.711230
#> 719 508.1184   4.060230
#> 720 500.0790   3.269272
#> 721 558.9194   3.223760
#> 722 500.2726   2.636289
#> 723 842.6916   2.876833
#> 724 500.0965   3.088983
#> 725 500.3415   2.518457
#> 726 500.4444   2.691759
#> 727 500.1877   2.989884
#> 728 518.4594   2.616487
#> 729 500.1632   3.527733
#> 730 513.7935   3.622482
#> 731 501.8421   3.430983
#> 732 500.2568   2.638605
#> 733 503.3336   2.926374
#> 734 500.0948   3.293497
#> 735 500.1959   2.720141
#> 736 521.4715   3.664819
#> 737 500.9127   2.989376
#> 738 503.7846   2.700667
#> 739 500.2625   2.553697
#> 740 500.0154   3.077965
#> 741 500.1154   2.520686
#> 742 582.3239   3.946776
#> 743 500.1437   2.817830
#> 744 507.2040   2.639842
#> 745 562.9963   2.510355
#> 746 501.3369   2.934846
#> 747 500.3294   2.822252
#> 748 500.6092   2.739141
#> 749 500.0266   2.834912
#> 750 503.5171   2.966431
#> 751 500.0545   2.544779
#> 752 617.5801   2.804327
#> 753 599.3427   2.530205
#> 754 502.1893   2.892688
#> 755 501.4075   3.078776
#> 756 501.8799   2.802423
#> 757 508.8261   2.511930
#> 758 505.2542   2.558250
#> 759 500.1000   2.719840
#> 760 510.6787   2.837170
#> 761 537.8499   2.627413
#> 762 537.8503   2.981649
#> 763 500.1159   2.570113
#> 764 559.6200   2.613029
#> 765 549.6705   3.503641
#> 766 505.9949   3.997619
#> 767 500.2427   2.570800
#> 768 500.1987   2.590000
#> 769 500.5194   2.500366
#> 770 522.2816   2.978790
#> 771 548.5760   2.519220
#> 772 500.0046   2.524733
#> 773 694.3500   2.545657
#> 774 543.7972   3.247376
#> 775 501.4890   2.543864
#> 776 500.0408   2.760590
#> 777 500.2077   3.357561
#> 778 500.3148   3.123345
#> 779 500.0387   2.690540
#> 780 540.0076   2.518484
#> 781 500.3345   2.513659
#> 782 555.3046   2.661480
#> 783 500.1473   2.927249
#> 784 500.3715   2.612145
#> 785 705.6351   3.708615
#> 786 500.1277   3.821301
#> 787 502.0211   2.897898
#> 788 500.0907   2.510962
#> 789 509.8629   2.631300
#> 790 515.0611   4.677631
#> 791 522.6998   3.463934
#> 792 579.6938   2.772440
#> 793 515.4068   2.754764
#> 794 502.1289   3.074741
#> 795 505.9255   2.707310
#> 796 517.6028   2.565970
#> 797 502.5635   2.514024
#> 798 780.6538   3.099755
#> 799 606.0902   3.102740
#> 800 517.8836   2.818027
#> 801 580.0128   3.639565
#> 802 630.1105   2.817695
#> 803 500.0876   2.693724
#> 804 502.9763   2.600554
#> 805 501.7533   4.015728
#> 806 588.0163   2.624764
#> 807 501.1509   2.974797
#> 808 510.9422   2.975048
#> 809 500.2467   2.601154
#> 810 580.7719   3.081802
#> 811 509.6491   2.500548
#> 812 601.9262   2.517225
#> 813 513.2746   2.710752
#> 814 500.3369   2.897107
#> 815 816.6739   3.494109
#> 816 535.3670   2.796859
#> 817 520.5004   2.693694
#> 818 500.0099   2.825297
#> 819 504.1946   2.997862
#> 820 501.1635   2.687830
#> 821 500.0605   2.707278
#> 822 500.0207   2.859359
#> 823 505.1587   2.532444
#> 824 502.6485   2.618477
#> 825 528.6927   2.542875
#> 826 503.6375   3.159734
#> 827 507.4104   3.634880
#> 828 514.2347   2.824029
#> 829 500.0119   2.589952
#> 830 509.1656   2.984611
#> 831 529.8721   3.722344
#> 832 513.1147   3.046739
#> 833 753.4581   2.583150
#> 834 527.5975   2.718358
#> 835 505.6735   3.728057
#> 836 500.2926   2.742800
#> 837 500.1211   3.649151
#> 838 500.0481   2.672979
#> 839 500.0189   2.898666
#> 840 500.1082   2.529417
#> 841 659.7998   2.664416
#> 842 600.1858   2.855851
#> 843 507.5173   3.817416
#> 844 509.8568   2.507480
#> 845 503.1252   3.001189
#> 846 500.0829   2.584430
#> 847 501.7677   2.566145
#> 848 500.5152   2.593451
#> 849 546.3226   3.106517
#> 850 504.0517   3.525985
#> 851 502.2485   2.705943
#> 852 500.6769   3.437918
#> 853 500.1756   3.632204
#> 854 500.1228   3.092473
#> 855 695.2653   2.740594
#> 856 500.1153   2.702786
#> 857 500.3500   2.768072
#> 858 501.5622   2.720762
#> 859 500.4264   3.113458
#> 860 978.1323   3.102159
#> 861 501.5561   2.552627
#> 862 500.5254   3.141280
#> 863 504.5197   2.523230
#> 864 500.0531   2.932543
#> 865 500.9842   2.676224
#> 866 513.1865   4.294447
#> 867 510.8845   2.860789
#> 868 599.4087   2.888989
#> 869 509.1014   4.226244
#> 870 502.6177   2.649583
#> 871 500.8540   2.641361
#> 872 500.7776   2.787662
#> 873 500.0079   3.004431
#> 874 501.3262   2.657203
#> 875 505.4632   3.451363
#> 876 501.8344   2.671610
#> 877 796.7825   2.691618
#> 878 500.0592   3.042037
#> 879 511.1087   2.690848
#> 880 500.2551   3.361475
#> 881 501.0188   2.550464
#> 882 587.3240   3.493224
#> 883 511.2252   2.604782
#> 884 500.1325   3.068184
#> 885 501.1966   3.294247
#> 886 508.7074   3.455461
#> 887 541.5603   2.580018
#> 888 500.8968   3.254124
#> 889 500.2044   2.577690
#> 890 516.6527   2.530933
#> 891 509.1863   3.078045
#> 892 508.9177   2.521330
#> 893 723.6274   3.004261
#> 894 506.2974   2.524093
#> 895 505.3736   3.504740
#> 896 500.2872   2.708712
#> 897 523.7651   2.746829
#> 898 502.5524   2.578954
#> 899 501.5934   2.954334
#> 900 523.7276   3.130517
#> 901 502.5329   2.735944
#> 902 500.0405   2.518878
#> 903 500.0187   2.599993
#> 904 503.7565   2.703101
#> 905 500.0736   4.323748
#> 906 501.6583   2.903046
#> 907 504.3864   3.095836
#> 908 500.1767   2.655760
#> 909 502.7922   3.008389
#> 910 505.9565   2.633022
#> 911 500.9785   2.815006
#> 912 733.5868   2.815541
#> 913 500.3178   2.638257
#> 914 502.3698   3.438151
#> 915 501.3455   2.944028
#> 916 507.6109   2.525817
#> 917 501.6233   3.216177
#> 918 500.6015   2.655715
#> 919 503.8392   3.967012
#> 920 505.1783   2.656482
#> 921 500.0077   2.714116
#> 922 509.9844   3.081996
#> 923 500.7792   2.647510
#> 924 503.8224   2.866361
#> 925 511.2650   2.576049
#> 926 596.2432   3.596244
#> 927 584.3033   3.344300
#> 928 504.2523   2.920993
#> 929 500.6300   2.860256
#> 930 578.3767   2.654668
#> 931 669.4243   2.517522
#> 932 506.0540   3.197724
#> 933 500.1044   2.728468
#> 934 500.1700   2.890524
#> 935 501.2259   2.535853
#> 936 500.4930   2.630042
#> 937 500.0162   2.936410