Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   500.6724   2.708629
#> 2   503.3743   3.035758
#> 3   903.7651   3.311891
#> 4   500.7319   2.605522
#> 5   500.2062   2.662880
#> 6   751.2403   3.431450
#> 7   697.3092   2.556834
#> 8   500.5329   2.582954
#> 9   501.6957   2.569172
#> 10  500.2183   4.129657
#> 11  806.8861   2.999269
#> 12  502.9061   3.303382
#> 13  500.0034   2.824883
#> 14  501.2348   3.027526
#> 15  528.2554   2.724215
#> 16  581.1571   2.998670
#> 17  500.7798   2.520054
#> 18  501.8617   2.943400
#> 19  500.2084   3.341067
#> 20  501.1042   2.676308
#> 21  588.3652   2.527724
#> 22  501.4970   3.322811
#> 23  583.5989   2.515212
#> 24  500.8545   2.567047
#> 25  500.3640   2.804496
#> 26  504.5733   3.868972
#> 27  502.5058   2.562528
#> 28  501.0820   4.007470
#> 29  609.0785   2.729125
#> 30  503.9707   3.062063
#> 31  516.7518   2.577766
#> 32  500.1511   2.867621
#> 33  510.2353   2.613154
#> 34  695.6870   2.800043
#> 35  500.4669   2.541890
#> 36  672.1933   2.536930
#> 37  544.5170   2.595021
#> 38  513.2818   4.171092
#> 39  514.4766   3.631831
#> 40  516.4735   2.750274
#> 41  500.2599   2.841817
#> 42  513.2031   2.670202
#> 43  500.2751   2.812867
#> 44  506.1178   2.991585
#> 45  853.2555   2.820182
#> 46  568.6071   2.612689
#> 47  513.0050   2.513372
#> 48  505.4027   2.501235
#> 49  500.2435   2.620517
#> 50  606.4963   2.546267
#> 51  505.1408   2.834724
#> 52  505.4747   2.912811
#> 53  536.8700   2.597887
#> 54  526.8193   2.580413
#> 55  514.9607   3.187467
#> 56  511.3805   2.731061
#> 57  500.4205   2.990417
#> 58  938.8417   2.585341
#> 59  500.6336   3.069241
#> 60  501.5576   2.605470
#> 61  501.9243   3.461591
#> 62  500.2682   3.030465
#> 63  506.4515   2.813887
#> 64  500.4649   2.822353
#> 65  524.5214   2.805940
#> 66  559.1335   3.328070
#> 67  500.4626   2.765715
#> 68  509.8236   2.592038
#> 69  503.9029   2.724874
#> 70  672.0207   2.729931
#> 71  500.4580   2.760105
#> 72  500.3673   2.874094
#> 73  501.9116   2.668816
#> 74  516.9487   3.246806
#> 75  502.6713   3.254389
#> 76  500.2236   2.654047
#> 77  500.7041   2.707288
#> 78  791.7461   2.788880
#> 79  526.3349   2.998788
#> 80  556.0315   2.838795
#> 81  500.0508   2.563076
#> 82  500.6370   3.280636
#> 83  502.2877   2.715572
#> 84  647.2617   2.853019
#> 85  509.5401   2.610021
#> 86  500.0587   2.811701
#> 87  512.4737   2.749630
#> 88  501.9203   2.930663
#> 89  593.7675   2.607314
#> 90  510.5376   2.589648
#> 91  500.0314   2.628158
#> 92  524.4908   3.265501
#> 93  507.4901   2.870807
#> 94  522.3791   2.721762
#> 95  569.8066   3.039438
#> 96  515.9565   2.527881
#> 97  557.2326   3.185063
#> 98  536.5104   3.290030
#> 99  500.8550   2.700876
#> 100 589.8733   2.907768
#> 101 500.6806   3.013497
#> 102 500.0650   3.221763
#> 103 516.4759   2.598039
#> 104 507.9262   2.799218
#> 105 500.4119   2.717233
#> 106 504.4172   2.563494
#> 107 517.4032   3.137490
#> 108 516.3944   3.944976
#> 109 503.1537   2.629354
#> 110 500.7344   2.625130
#> 111 500.0210   2.657414
#> 112 621.2062   2.539467
#> 113 598.2305   3.040439
#> 114 507.9300   3.144344
#> 115 505.7969   2.618811
#> 116 514.1491   2.682666
#> 117 537.8998   2.702271
#> 118 503.7727   2.668616
#> 119 500.0726   2.995234
#> 120 503.3440   2.577432
#> 121 837.7515   2.514657
#> 122 557.8585   2.601463
#> 123 807.0264   3.334803
#> 124 571.2620   3.425313
#> 125 504.0742   3.002305
#> 126 500.2070   2.808451
#> 127 503.8322   3.078980
#> 128 740.9757   2.729106
#> 129 503.2413   2.907092
#> 130 500.5458   2.697156
#> 131 503.5508   2.599324
#> 132 524.3462   2.648944
#> 133 501.4065   2.507283
#> 134 500.1636   2.720653
#> 135 645.7993   3.066712
#> 136 500.8613   3.191054
#> 137 527.1155   2.750098
#> 138 630.9933   2.538204
#> 139 965.5361   2.558061
#> 140 501.5938   2.655516
#> 141 510.2668   2.918899
#> 142 500.3787   2.609360
#> 143 500.6203   3.206459
#> 144 519.5632   2.504623
#> 145 507.0292   3.186984
#> 146 500.1822   2.946099
#> 147 500.0876   2.891447
#> 148 664.8272   2.797199
#> 149 500.2394   3.083286
#> 150 521.2709   4.265923
#> 151 528.6098   2.987476
#> 152 500.0150   2.863693
#> 153 664.2159   2.551694
#> 154 503.0994   2.534883
#> 155 507.8676   2.610181
#> 156 556.6167   3.083355
#> 157 501.6688   2.986971
#> 158 515.3478   2.618495
#> 159 500.4640   2.913370
#> 160 515.5393   3.838574
#> 161 509.6644   3.455941
#> 162 529.9184   2.551756
#> 163 510.8141   3.316142
#> 164 712.0922   2.513855
#> 165 503.4669   2.569146
#> 166 768.6725   3.914338
#> 167 516.1121   3.578684
#> 168 500.0819   2.879277
#> 169 506.1065   2.551611
#> 170 523.1839   2.656998
#> 171 501.3056   3.682916
#> 172 500.0372   2.875501
#> 173 500.0108   3.110171
#> 174 507.4589   4.242936
#> 175 524.2649   3.304029
#> 176 618.6736   2.746345
#> 177 500.7302   3.057265
#> 178 532.3653   4.026684
#> 179 507.5706   2.991409
#> 180 500.2540   2.595546
#> 181 500.2192   2.517112
#> 182 500.4586   2.739408
#> 183 503.3238   2.524513
#> 184 867.4410   2.967890
#> 185 500.3131   2.528390
#> 186 501.7979   2.587870
#> 187 500.0712   2.610472
#> 188 727.5810   2.621670
#> 189 886.2140   2.669801
#> 190 950.3493   3.384449
#> 191 504.7903   2.587673
#> 192 500.1909   3.640459
#> 193 608.7287   2.570462
#> 194 500.0420   2.553040
#> 195 501.2477   2.830917
#> 196 522.6440   2.873026
#> 197 902.1673   3.229768
#> 198 505.9660   2.905981
#> 199 532.6702   3.459637
#> 200 500.2504   2.596438
#> 201 500.0588   3.172116
#> 202 505.1115   2.603400
#> 203 527.9379   2.698035
#> 204 502.4126   2.825738
#> 205 500.4355   2.597170
#> 206 500.0589   3.425120
#> 207 500.4231   3.396272
#> 208 500.3540   2.581793
#> 209 500.4228   2.569346
#> 210 500.0573   2.730879
#> 211 554.6908   3.033575
#> 212 501.2373   2.558835
#> 213 600.8938   3.148752
#> 214 500.0154   3.428675
#> 215 518.0970   2.677501
#> 216 975.0125   2.677553
#> 217 500.0250   3.295848
#> 218 500.0811   2.828706
#> 219 500.3914   3.019915
#> 220 501.2757   2.893447
#> 221 506.6980   2.982557
#> 222 522.7299   2.907532
#> 223 500.1289   3.744428
#> 224 656.9166   2.522847
#> 225 503.1195   3.005328
#> 226 500.0005   3.289459
#> 227 541.2362   2.931255
#> 228 511.1224   2.820816
#> 229 511.9218   2.549946
#> 230 506.2940   2.573208
#> 231 501.8271   4.126407
#> 232 545.7071   2.531180
#> 233 505.4703   2.517101
#> 234 501.5999   2.513484
#> 235 500.7332   3.083532
#> 236 633.9385   2.603605
#> 237 500.1925   2.757253
#> 238 501.6121   2.666640
#> 239 500.0275   2.732392
#> 240 500.1394   2.524678
#> 241 501.3208   2.779324
#> 242 591.9939   4.357148
#> 243 550.3651   2.821012
#> 244 646.8113   3.408840
#> 245 538.5786   3.007492
#> 246 502.7696   2.540413
#> 247 793.9283   2.509404
#> 248 542.2192   2.555927
#> 249 501.2505   2.596906
#> 250 881.5512   2.636764
#> 251 507.5519   4.181507
#> 252 525.7598   2.650078
#> 253 503.9205   2.905406
#> 254 500.1968   2.837912
#> 255 507.6013   3.180229
#> 256 500.9778   2.882921
#> 257 500.6616   3.687951
#> 258 514.2530   2.504532
#> 259 500.8269   3.021648
#> 260 500.0640   3.540070
#> 261 998.0243   2.669426
#> 262 512.1281   3.885353
#> 263 567.3461   3.308651
#> 264 799.0245   2.568769
#> 265 507.2578   2.644925
#> 266 500.0196   3.164207
#> 267 911.7675   2.792014
#> 268 500.0555   2.968654
#> 269 506.6183   2.642500
#> 270 534.9595   2.636945
#> 271 500.4700   2.619258
#> 272 744.5182   3.043305
#> 273 500.9919   2.949728
#> 274 963.2863   3.359491
#> 275 500.0783   2.818265
#> 276 590.0360   2.884005
#> 277 500.8320   3.116399
#> 278 500.9915   2.583686
#> 279 503.1224   3.000570
#> 280 505.5316   2.911053
#> 281 591.0200   2.607473
#> 282 605.1713   3.542551
#> 283 600.0183   3.193057
#> 284 636.2717   2.704506
#> 285 529.4167   2.712643
#> 286 506.9402   2.837866
#> 287 501.6032   3.024691
#> 288 915.0326   3.387341
#> 289 502.3773   2.740072
#> 290 513.5015   2.621601
#> 291 511.6979   2.689577
#> 292 502.5340   2.741489
#> 293 500.1274   2.686147
#> 294 500.0013   2.594464
#> 295 500.0091   4.011659
#> 296 511.3117   3.058404
#> 297 500.5571   2.627703
#> 298 501.5608   3.096649
#> 299 500.9049   3.381504
#> 300 502.8022   2.680136
#> 301 624.3356   2.851878
#> 302 631.2688   2.586066
#> 303 505.8103   2.800679
#> 304 506.3750   2.946708
#> 305 528.3151   3.075382
#> 306 500.4847   2.586280
#> 307 501.0075   2.657976
#> 308 503.0130   2.711272
#> 309 502.7886   3.685075
#> 310 559.2017   2.728339
#> 311 508.0012   3.246399
#> 312 514.2321   2.751564
#> 313 500.8383   2.672941
#> 314 505.4327   2.637714
#> 315 500.6150   2.602978
#> 316 500.3027   2.589461
#> 317 501.0137   2.837399
#> 318 505.4030   2.629334
#> 319 503.9870   2.912950
#> 320 500.7577   2.574221
#> 321 656.5935   2.852332
#> 322 532.0638   2.524583
#> 323 501.1935   2.667079
#> 324 502.6032   3.413950
#> 325 501.5249   2.703664
#> 326 501.2018   2.597042
#> 327 500.2066   2.995815
#> 328 882.0515   2.819512
#> 329 504.0618   2.550525
#> 330 558.9003   3.512837
#> 331 512.5893   2.541887
#> 332 620.4905   2.633130
#> 333 500.0789   2.978569
#> 334 501.7587   3.428198
#> 335 500.0135   3.178192
#> 336 504.4995   3.347027
#> 337 500.1575   2.559173
#> 338 520.0729   2.662246
#> 339 516.8751   3.797009
#> 340 588.9195   2.681112
#> 341 500.4660   2.612463
#> 342 598.3706   2.698756
#> 343 519.4791   2.609252
#> 344 561.3420   3.379207
#> 345 596.0402   3.191158
#> 346 500.0235   2.623063
#> 347 536.5403   2.652483
#> 348 964.3189   2.807839
#> 349 500.1652   2.966371
#> 350 502.4632   2.920972
#> 351 768.3533   2.562743
#> 352 500.6011   3.032186
#> 353 511.4990   3.406780
#> 354 565.5220   3.001085
#> 355 500.0011   2.650032
#> 356 500.3980   2.658008
#> 357 530.6108   2.771283
#> 358 501.0042   2.534145
#> 359 503.2781   2.679589
#> 360 877.2250   2.744200
#> 361 500.4180   3.221419
#> 362 500.0899   3.083986
#> 363 500.0690   2.962652
#> 364 596.4117   2.502487
#> 365 539.5414   3.124487
#> 366 501.8727   2.556826
#> 367 509.8955   2.823231
#> 368 500.9729   2.659412
#> 369 501.2471   2.611363
#> 370 500.4263   2.883520
#> 371 627.5472   3.061440
#> 372 500.2319   2.967079
#> 373 500.4676   2.635375
#> 374 500.0602   2.622797
#> 375 512.6170   4.307877
#> 376 502.9438   3.199935
#> 377 508.7036   2.623697
#> 378 500.1100   2.699847
#> 379 878.2280   3.315359
#> 380 500.1422   2.507904
#> 381 500.3968   4.960207
#> 382 507.9951   2.603332
#> 383 505.6461   2.641593
#> 384 527.2362   3.324302
#> 385 500.1841   3.620992
#> 386 505.5665   2.514331
#> 387 500.1146   2.610932
#> 388 717.9010   2.654232
#> 389 503.5804   3.884605
#> 390 501.2785   2.713411
#> 391 502.9292   3.669403
#> 392 509.6911   2.553473
#> 393 526.9676   2.597252
#> 394 509.6879   2.623914
#> 395 503.4332   2.699351
#> 396 501.4653   2.609225
#> 397 500.3596   2.867765
#> 398 519.9431   2.653868
#> 399 500.2690   2.642301
#> 400 516.2742   2.687620
#> 401 514.5532   2.644013
#> 402 609.5650   2.705960
#> 403 500.0502   3.167896
#> 404 500.3440   2.684809
#> 405 500.3632   2.542024
#> 406 501.4042   2.741070
#> 407 694.3672   2.829597
#> 408 500.0658   3.774535
#> 409 509.6482   2.687983
#> 410 583.3148   2.879373
#> 411 500.9596   3.847890
#> 412 500.0319   2.536294
#> 413 519.2628   3.599724
#> 414 977.9685   2.961788
#> 415 503.2305   4.136993
#> 416 586.1006   3.060339
#> 417 712.3477   2.657387
#> 418 504.0444   3.094019
#> 419 713.6774   3.625464
#> 420 505.9589   3.221316
#> 421 875.1471   3.027644
#> 422 576.5796   2.701525
#> 423 722.2015   2.826606
#> 424 820.8002   2.503732
#> 425 500.4287   2.748120
#> 426 500.3554   2.880733
#> 427 500.4676   3.798512
#> 428 550.2936   2.618138
#> 429 500.7770   3.503527
#> 430 642.9605   2.875850
#> 431 501.3480   4.785830
#> 432 500.0119   2.849201
#> 433 984.6168   3.152119
#> 434 858.7142   2.761083
#> 435 541.8252   2.513636
#> 436 721.7042   2.589181
#> 437 531.1727   2.894863
#> 438 500.6629   4.151134
#> 439 500.0699   3.286937
#> 440 659.8638   2.723999
#> 441 500.8587   2.645384
#> 442 891.7555   2.771957
#> 443 500.0203   2.839996
#> 444 662.2178   2.569394
#> 445 500.2522   2.569013
#> 446 500.1027   3.143065
#> 447 503.2781   3.565160
#> 448 502.4894   3.446914
#> 449 563.1241   3.785457
#> 450 534.4662   2.647846
#> 451 514.9079   3.144840
#> 452 542.1847   3.706590
#> 453 522.3971   2.556245
#> 454 501.8668   2.898033
#> 455 525.8290   2.557792
#> 456 501.9801   2.512580
#> 457 500.3558   2.746641
#> 458 544.5946   2.765243
#> 459 500.2256   3.186598
#> 460 500.2738   2.724400
#> 461 500.0745   3.138669
#> 462 500.1443   2.604859
#> 463 745.9574   2.870946
#> 464 505.9452   4.588334
#> 465 500.4817   2.676725
#> 466 502.5795   2.661812
#> 467 727.6669   2.628191
#> 468 500.0069   3.397642
#> 469 514.6011   3.184709
#> 470 513.2818   2.792691
#> 471 505.7213   3.374586
#> 472 500.0680   3.208607
#> 473 501.4574   2.983090
#> 474 503.5118   2.634354
#> 475 500.3018   2.699302
#> 476 594.3726   4.119537
#> 477 500.7996   2.680801
#> 478 500.5622   3.582187
#> 479 500.1547   3.060740
#> 480 500.7774   3.975613
#> 481 542.3784   2.950609
#> 482 500.0699   3.124788
#> 483 509.9902   2.535581
#> 484 500.0240   3.135836
#> 485 503.8236   2.790805
#> 486 500.0286   3.150818
#> 487 503.4034   4.005875
#> 488 522.8647   5.157879
#> 489 500.0516   3.432636
#> 490 500.5289   3.176686
#> 491 500.8340   3.299659
#> 492 836.7263   2.716317
#> 493 501.3149   2.676937
#> 494 500.0406   2.548873
#> 495 723.2936   4.345036
#> 496 534.6017   2.625360
#> 497 935.7390   2.529008
#> 498 503.7581   3.018524
#> 499 509.7782   2.794469
#> 500 503.5220   2.588037
#> 501 508.9602   2.555942
#> 502 504.6024   2.599583
#> 503 500.2440   2.707109
#> 504 505.2431   3.459024
#> 505 552.8313   2.781474
#> 506 565.1252   2.965287
#> 507 520.2491   2.503821
#> 508 500.3577   2.714239
#> 509 514.9448   3.040832
#> 510 525.3483   2.854743
#> 511 502.0319   2.641153
#> 512 500.6231   2.914443
#> 513 500.0697   2.535703
#> 514 531.2606   2.581011
#> 515 647.5005   3.369636
#> 516 630.0327   2.679751
#> 517 501.2253   2.573520
#> 518 512.9739   2.759781
#> 519 974.4880   2.716630
#> 520 501.8442   2.674238
#> 521 532.9222   2.543050
#> 522 500.2093   2.742033
#> 523 533.7802   3.038436
#> 524 690.9386   3.284522
#> 525 500.8263   2.511898
#> 526 500.3571   2.960683
#> 527 500.0943   2.873625
#> 528 520.0899   2.570956
#> 529 500.0483   2.965845
#> 530 605.8099   3.282461
#> 531 515.6870   2.541269
#> 532 500.9521   3.144185
#> 533 519.6804   3.182363
#> 534 660.9818   3.781972
#> 535 503.0979   2.811361
#> 536 503.6775   2.603077
#> 537 506.4184   2.842961
#> 538 637.0095   2.655370
#> 539 500.2552   3.150569
#> 540 500.8857   2.762077
#> 541 723.2075   3.647792
#> 542 500.6736   2.591916
#> 543 502.4172   3.243465
#> 544 505.3221   2.521888
#> 545 503.6940   2.589644
#> 546 915.2731   2.617397
#> 547 500.7417   3.087276
#> 548 500.2786   2.749746
#> 549 647.1846   2.530906
#> 550 513.7734   2.566038
#> 551 501.8664   3.301652
#> 552 564.7712   3.441779
#> 553 537.3658   2.838262
#> 554 521.5808   3.291217
#> 555 505.0608   2.504160
#> 556 504.9777   2.702067
#> 557 706.6791   3.829892
#> 558 529.0712   2.636088
#> 559 862.3350   3.081330
#> 560 735.6121   2.809170
#> 561 516.8187   2.551208
#> 562 514.9688   3.126540
#> 563 519.0800   2.655980
#> 564 500.3362   3.136182
#> 565 500.0502   4.138182
#> 566 666.9816   2.545606
#> 567 528.1470   2.721899
#> 568 500.8334   3.005677
#> 569 529.2807   3.819396
#> 570 501.9294   2.642996
#> 571 500.2584   4.134674
#> 572 519.6756   2.980795
#> 573 500.0251   2.599317
#> 574 535.2119   3.144476
#> 575 540.8347   3.700314
#> 576 500.2862   3.382663
#> 577 500.0050   2.568756
#> 578 517.6797   2.513217
#> 579 533.5920   2.649710
#> 580 567.3783   2.834729
#> 581 500.1774   2.618617
#> 582 590.6660   2.985387
#> 583 500.4049   2.811679
#> 584 509.2250   3.452725
#> 585 650.4360   2.635487
#> 586 500.2238   2.511455
#> 587 545.8986   2.593512
#> 588 506.9796   2.642780
#> 589 502.1841   3.040315
#> 590 503.3176   2.591594
#> 591 503.0582   2.620476
#> 592 500.0643   3.007221
#> 593 505.5994   2.762912
#> 594 500.0043   2.712806
#> 595 500.2876   3.042221
#> 596 641.8828   3.528945
#> 597 504.7594   2.636233
#> 598 607.6098   2.806035
#> 599 606.6774   2.595178
#> 600 888.7990   2.872661
#> 601 500.9948   2.586341
#> 602 500.4387   2.640546
#> 603 502.4182   3.400327
#> 604 500.0011   3.399547
#> 605 535.8057   3.385732
#> 606 515.8493   3.127882
#> 607 849.4778   2.969207
#> 608 508.1391   2.640762
#> 609 555.3236   3.545528
#> 610 503.2670   2.535712
#> 611 504.9566   2.662463
#> 612 661.4500   2.515365
#> 613 504.8764   2.757277
#> 614 500.1673   2.873914
#> 615 502.7159   3.054059
#> 616 500.2339   2.831593
#> 617 587.2233   2.909679
#> 618 500.9311   3.267832
#> 619 502.1858   4.156871
#> 620 529.0922   2.545032
#> 621 500.3233   3.134901
#> 622 533.4596   2.983243
#> 623 574.0554   2.848096
#> 624 512.5318   2.702856
#> 625 500.9758   3.518720
#> 626 523.8455   3.871579
#> 627 500.5440   2.609316
#> 628 594.4016   2.562490
#> 629 501.7749   3.048464
#> 630 500.5948   2.500171
#> 631 500.8548   2.940363
#> 632 500.0590   2.657719
#> 633 610.3991   2.505657
#> 634 508.2000   2.614554
#> 635 504.2407   2.561335
#> 636 699.8998   2.530609
#> 637 586.6900   3.250895
#> 638 556.5709   3.142798
#> 639 501.8596   2.523418
#> 640 500.1086   3.091181
#> 641 501.0618   2.617372
#> 642 500.1364   2.580420
#> 643 500.0211   2.573286
#> 644 503.8083   2.535716
#> 645 875.8918   2.844923
#> 646 703.1688   2.568759
#> 647 640.1145   2.862304
#> 648 500.3256   3.550981
#> 649 500.8408   2.558588
#> 650 511.0021   3.527751
#> 651 537.8831   4.101897
#> 652 500.0064   3.450580
#> 653 506.6380   4.497502
#> 654 500.2733   3.877759
#> 655 517.1036   3.218884
#> 656 500.3367   2.785277
#> 657 500.0960   2.824206
#> 658 517.4491   2.750724
#> 659 554.4924   2.654182
#> 660 500.0813   2.763450
#> 661 502.8701   3.162190
#> 662 500.7859   3.591253
#> 663 500.1438   2.967496
#> 664 511.4741   3.398391
#> 665 879.2662   2.621537
#> 666 503.9942   2.965237
#> 667 712.9242   2.624186
#> 668 508.0369   2.528146
#> 669 615.4467   3.339219
#> 670 585.0751   2.515101
#> 671 500.1402   2.698544
#> 672 536.6735   2.540695
#> 673 500.0317   2.633630
#> 674 508.0844   2.833360
#> 675 500.0091   2.846773
#> 676 500.4840   4.083464
#> 677 500.2866   2.933347
#> 678 717.5902   3.737066
#> 679 615.7069   2.825134
#> 680 634.1153   2.648624
#> 681 510.6277   3.691724
#> 682 500.4131   3.847947
#> 683 505.6330   3.034432
#> 684 509.9282   2.679338
#> 685 500.1509   2.620331
#> 686 557.3505   2.683202
#> 687 504.8652   2.721648
#> 688 721.9471   2.697325
#> 689 500.1942   2.715738
#> 690 551.6782   2.570821
#> 691 500.0308   2.749835
#> 692 500.0466   2.557853
#> 693 501.8061   3.101487
#> 694 537.4835   2.520439
#> 695 500.5396   2.913859
#> 696 500.2643   2.709150
#> 697 573.2812   2.645995
#> 698 504.1292   2.965489
#> 699 502.3354   2.686096
#> 700 503.5180   2.899609
#> 701 517.7922   2.567816
#> 702 589.3426   2.562752
#> 703 896.1393   2.506699
#> 704 617.4682   3.003161
#> 705 779.3305   2.721895
#> 706 501.4389   2.765883
#> 707 500.1262   2.572656
#> 708 595.8092   2.536915
#> 709 500.0144   3.163561
#> 710 692.4988   2.529199
#> 711 500.0841   2.521590
#> 712 501.1204   3.265618
#> 713 500.1152   2.955055
#> 714 501.5888   2.731806
#> 715 502.0256   3.350556
#> 716 500.1813   2.664861
#> 717 500.2748   2.575625
#> 718 655.7417   2.572274
#> 719 656.7358   3.158459
#> 720 500.1786   2.652584
#> 721 501.6792   2.841593
#> 722 505.0471   2.778322
#> 723 791.6147   2.738326
#> 724 500.3054   2.734930
#> 725 507.2308   2.935081
#> 726 512.2888   2.725249
#> 727 510.3620   2.626473
#> 728 501.7085   2.742731
#> 729 602.1849   2.543808
#> 730 600.8561   2.744394
#> 731 532.9031   2.616811
#> 732 535.0984   2.505816
#> 733 572.4187   3.202310
#> 734 500.0136   2.635461
#> 735 501.1319   3.445607
#> 736 517.5626   2.592364
#> 737 511.1183   2.576481
#> 738 513.9911   2.676789
#> 739 537.0662   2.531106
#> 740 501.2397   2.670343
#> 741 500.3794   2.535403
#> 742 981.5004   3.126655
#> 743 510.7850   2.589416
#> 744 500.2314   2.798979
#> 745 501.4619   3.138398
#> 746 550.2780   2.541854
#> 747 500.5126   2.711298
#> 748 522.4314   2.800403
#> 749 561.8916   3.627712
#> 750 500.1502   2.769626
#> 751 536.4246   3.772875
#> 752 501.2839   3.048609
#> 753 501.6673   2.671091
#> 754 501.2257   3.323144
#> 755 547.5348   2.583070
#> 756 506.7082   2.655804
#> 757 500.6036   2.577332
#> 758 582.4258   2.614312
#> 759 505.4558   3.271449
#> 760 500.0615   2.634334
#> 761 527.0048   2.983885
#> 762 552.2874   2.685252
#> 763 500.3243   2.683790
#> 764 509.2757   2.645029
#> 765 503.3086   3.168704
#> 766 500.1383   2.583995
#> 767 616.4028   2.729809
#> 768 543.1517   2.811696
#> 769 500.8671   2.938541
#> 770 508.0031   2.904672
#> 771 500.1216   3.154245
#> 772 503.5640   2.778392
#> 773 500.1131   2.981911
#> 774 500.8859   2.663628
#> 775 500.9820   2.579683
#> 776 500.5808   2.574159
#> 777 500.4592   2.715600
#> 778 500.9553   2.689411
#> 779 500.1212   3.180490
#> 780 520.3050   2.516033
#> 781 500.0018   3.448409
#> 782 512.0381   2.976199
#> 783 500.0421   2.657153
#> 784 510.3068   3.568591
#> 785 500.9194   3.277813
#> 786 500.1027   2.633331
#> 787 759.1266   2.582335
#> 788 502.6489   2.516959
#> 789 500.3812   3.491941
#> 790 507.3611   2.538606
#> 791 500.6036   2.763418
#> 792 504.5065   3.253243
#> 793 501.5145   2.538073
#> 794 768.2635   2.912224
#> 795 508.2346   2.750631
#> 796 500.1206   2.971486
#> 797 500.7038   3.597857
#> 798 500.0837   2.901661
#> 799 741.1864   2.716523
#> 800 533.4889   2.769003
#> 801 856.2060   2.717180
#> 802 506.3229   2.704909
#> 803 500.0129   4.125607
#> 804 501.1207   2.598050
#> 805 500.0925   2.543785
#> 806 502.6341   3.411139
#> 807 509.8217   2.515357
#> 808 606.4102   2.715320
#> 809 591.4260   2.954487
#> 810 504.3470   3.471190
#> 811 751.9974   3.089428
#> 812 917.5854   2.752699
#> 813 799.1566   2.690390
#> 814 504.3299   3.112488
#> 815 561.6272   2.797929
#> 816 545.3689   2.598829
#> 817 934.9320   2.606196
#> 818 500.4096   3.040230
#> 819 509.1342   3.764050
#> 820 506.9030   2.524157
#> 821 508.4161   3.186670
#> 822 500.0470   2.755392
#> 823 517.1831   2.640999
#> 824 506.0924   3.169281
#> 825 532.5986   2.667106
#> 826 548.9308   4.594128
#> 827 506.0241   2.616338
#> 828 558.7766   2.742278
#> 829 501.0282   3.376925
#> 830 505.4254   3.587273
#> 831 500.2590   2.877865
#> 832 524.9862   2.540965
#> 833 527.1657   2.646087
#> 834 822.4392   2.699642
#> 835 501.8117   2.574821
#> 836 502.0889   3.023783
#> 837 500.1407   3.014481
#> 838 502.2576   2.571404
#> 839 824.6222   2.921669
#> 840 510.6923   2.881397
#> 841 502.6025   2.597217
#> 842 500.1236   2.640324
#> 843 503.9212   2.762005
#> 844 500.1468   2.519572
#> 845 500.3392   3.430177
#> 846 503.3599   2.985831
#> 847 503.2586   3.245373
#> 848 500.5455   3.520758
#> 849 906.0414   3.166848
#> 850 603.1405   2.740454
#> 851 842.4848   2.503246
#> 852 501.4284   2.919886
#> 853 500.4366   2.685310
#> 854 501.1846   2.638754
#> 855 500.4755   3.287633
#> 856 666.3597   2.791352
#> 857 500.6887   2.554595
#> 858 500.5921   3.211149
#> 859 520.1947   3.219226
#> 860 532.7253   2.811907
#> 861 508.4325   2.606546
#> 862 933.9092   2.542986
#> 863 500.1413   3.117340
#> 864 500.3216   3.479859
#> 865 500.2901   2.899408
#> 866 778.8449   2.614022
#> 867 500.2293   2.584768
#> 868 814.6585   2.564714
#> 869 501.0748   2.585687
#> 870 500.2172   2.536041
#> 871 500.3447   3.598979
#> 872 503.7328   2.987213
#> 873 830.4188   2.524041
#> 874 504.6346   3.140137
#> 875 502.1276   2.614638
#> 876 509.4042   2.531208
#> 877 546.3143   2.570324
#> 878 501.3334   2.651944
#> 879 983.8137   3.299595
#> 880 500.0647   3.486124
#> 881 501.9374   3.034322
#> 882 500.8931   2.530876
#> 883 530.6347   2.791842
#> 884 507.8694   2.709044
#> 885 671.5675   2.625100
#> 886 953.7387   2.567892
#> 887 536.0720   2.754780
#> 888 504.7948   2.772467
#> 889 852.6082   3.328853
#> 890 501.2625   3.900333
#> 891 968.5932   4.013416
#> 892 659.9914   2.623520
#> 893 500.1797   3.026900
#> 894 667.7392   3.547567
#> 895 502.6231   2.937088
#> 896 501.8580   2.931076
#> 897 509.1092   3.688203
#> 898 500.3412   2.776301
#> 899 500.0970   2.952328
#> 900 548.6123   3.319378
#> 901 515.5489   3.486775
#> 902 500.1423   2.504724
#> 903 508.8942   2.758224
#> 904 780.7263   2.787857
#> 905 504.8452   2.581489
#> 906 523.4632   2.906523
#> 907 500.1169   2.627120
#> 908 500.3615   2.534457
#> 909 526.1614   2.817208
#> 910 502.9242   2.922817
#> 911 500.1171   3.354207
#> 912 500.1023   2.504442
#> 913 500.4602   2.524801