Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Take all previous parent events from Ht=data.frame[ts, magnitudes] and generates their daughters events using the ETAS model

Usage

sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)

Arguments

theta

ETAS parameters list(mu=mu, K=K, alpha=alpha, c=c, p=p).

beta.p

Slope of GR relation: beta = b ln(10).

Ht

The set of parent events in the form data.frame[ts, magnitudes]

M0

The minimum earthquake magnitude in the synthetic catalogue.

T1

The start time for the synthetic catalogue [days].

T2

The end time for the synthetic catalogue [days].

ncore

The number of compute cores to use

Value

Return one generation of daughters from the parents in Ht in the form data.frame(t_i, M_i).

Examples

# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
#>           ts magnitudes
#> 1   568.9823   3.571084
#> 2   501.6132   4.195077
#> 3   547.1766   2.568503
#> 4   504.7473   2.521689
#> 5   503.4542   4.233638
#> 6   505.2505   3.639895
#> 7   504.5681   2.534407
#> 8   510.8449   2.939831
#> 9   500.5047   2.899531
#> 10  508.5568   2.876008
#> 11  531.0334   3.206555
#> 12  527.5014   2.689592
#> 13  501.6430   2.645184
#> 14  504.3248   2.847346
#> 15  550.1624   2.981998
#> 16  500.0102   2.540607
#> 17  524.6027   3.465704
#> 18  825.2411   3.619555
#> 19  501.3773   3.195532
#> 20  750.0041   2.869526
#> 21  501.2127   2.559516
#> 22  922.1379   2.564709
#> 23  542.4783   3.430827
#> 24  500.0621   3.115632
#> 25  500.4423   2.606427
#> 26  640.0025   3.212720
#> 27  500.0754   3.178629
#> 28  509.3582   2.758883
#> 29  516.9962   2.625780
#> 30  540.4579   2.612371
#> 31  752.1046   2.756486
#> 32  500.8359   2.672342
#> 33  503.5036   2.542340
#> 34  571.9815   3.115868
#> 35  500.0516   3.443139
#> 36  504.0256   3.369268
#> 37  515.7341   3.398761
#> 38  501.0367   3.583824
#> 39  581.2237   2.790491
#> 40  500.1395   2.770859
#> 41  624.6424   2.984191
#> 42  500.1953   2.611274
#> 43  880.2577   2.557681
#> 44  500.0389   3.883764
#> 45  500.2780   2.583771
#> 46  500.5495   2.606705
#> 47  894.6501   2.824446
#> 48  560.3669   2.788403
#> 49  501.3637   3.532001
#> 50  596.1971   2.538083
#> 51  502.0672   2.752503
#> 52  500.0003   2.992684
#> 53  540.9952   2.575844
#> 54  502.8188   2.857265
#> 55  503.3119   2.662173
#> 56  914.0170   3.606424
#> 57  500.3185   2.734220
#> 58  679.2224   2.722900
#> 59  500.4337   3.015816
#> 60  750.2296   2.569378
#> 61  529.4233   2.750477
#> 62  501.8631   2.500055
#> 63  958.6478   2.513099
#> 64  500.3121   2.819980
#> 65  511.1411   2.692258
#> 66  506.5509   2.762440
#> 67  501.6925   2.721331
#> 68  501.1553   2.550041
#> 69  500.2938   2.507405
#> 70  503.0771   2.553292
#> 71  507.6143   2.884339
#> 72  519.3656   2.571679
#> 73  500.3227   3.132621
#> 74  500.1216   2.707311
#> 75  500.0150   3.353578
#> 76  537.7748   2.814075
#> 77  517.9402   2.505893
#> 78  510.6593   2.632384
#> 79  515.1295   3.055799
#> 80  519.9130   2.600111
#> 81  500.0017   2.975212
#> 82  567.7657   2.771351
#> 83  691.8183   2.955642
#> 84  500.0177   3.116502
#> 85  540.5214   2.608555
#> 86  500.4486   4.095367
#> 87  826.1509   3.112885
#> 88  915.4451   2.927353
#> 89  502.5466   3.140796
#> 90  550.3068   2.579726
#> 91  500.8185   2.810194
#> 92  513.1037   2.629917
#> 93  590.7787   2.584948
#> 94  525.8976   3.262103
#> 95  500.5935   2.867619
#> 96  500.7425   3.123486
#> 97  590.2159   2.949956
#> 98  541.7764   2.715938
#> 99  500.6785   3.133957
#> 100 531.9215   2.656030
#> 101 550.4086   2.865521
#> 102 500.0354   3.233796
#> 103 500.0066   2.532568
#> 104 500.2953   2.560338
#> 105 522.1749   2.958621
#> 106 549.0387   3.101611
#> 107 504.9722   2.971178
#> 108 500.2278   3.042614
#> 109 705.8404   2.846283
#> 110 500.2352   3.028046
#> 111 500.4995   3.683854
#> 112 590.1885   2.574263
#> 113 500.6906   2.730033
#> 114 500.8335   3.016933
#> 115 500.5921   4.027185
#> 116 780.5037   2.600313
#> 117 512.2416   2.661596
#> 118 514.5095   2.646561
#> 119 578.9043   4.382930
#> 120 732.0464   2.614316
#> 121 500.2600   3.404093
#> 122 552.1508   3.501725
#> 123 500.3452   2.598963
#> 124 500.0003   2.537903
#> 125 722.8181   2.585477
#> 126 519.3249   2.716889
#> 127 514.0056   2.530742
#> 128 500.0142   2.520045
#> 129 521.7444   2.652577
#> 130 601.0811   2.660474
#> 131 500.2353   2.548375
#> 132 500.4402   2.829996
#> 133 500.7879   2.635611
#> 134 500.1360   2.775442
#> 135 601.2080   2.590403
#> 136 500.6325   3.089935
#> 137 500.1048   2.700661
#> 138 500.0940   3.296928
#> 139 875.3529   2.701723
#> 140 500.2722   2.762149
#> 141 843.0954   2.541621
#> 142 501.2425   2.641850
#> 143 712.3800   3.082475
#> 144 540.8985   4.729332
#> 145 582.0047   2.861439
#> 146 572.6558   2.984165
#> 147 500.2636   2.665880
#> 148 500.1010   3.209210
#> 149 509.4075   2.640070
#> 150 500.5758   2.731759
#> 151 507.4532   3.227068
#> 152 500.1551   2.594431
#> 153 501.1322   2.591375
#> 154 566.0507   2.994101
#> 155 590.0928   2.579971
#> 156 568.4979   2.833903
#> 157 963.5837   2.567560
#> 158 987.8796   2.814949
#> 159 500.0586   2.806556
#> 160 978.8747   2.658268
#> 161 500.0577   2.888408
#> 162 508.1981   2.537305
#> 163 500.7401   3.512551
#> 164 525.6363   2.719045
#> 165 505.1939   2.979774
#> 166 500.8057   4.775833
#> 167 500.1925   2.605518
#> 168 506.3037   2.515254
#> 169 501.3087   3.185430
#> 170 505.2489   2.606282
#> 171 813.8504   2.723348
#> 172 537.2746   2.579971
#> 173 705.6745   2.663778
#> 174 511.7862   2.869994
#> 175 551.4036   3.281889
#> 176 568.2286   2.822496
#> 177 502.0937   2.906893
#> 178 500.1547   2.633927
#> 179 605.4820   2.995960
#> 180 676.5892   2.785790
#> 181 590.1805   3.212880
#> 182 500.2993   2.668695
#> 183 500.0508   2.798492
#> 184 514.5401   3.612169
#> 185 505.4730   2.980479
#> 186 650.8769   2.935285
#> 187 503.3181   3.047178
#> 188 685.6507   3.522431
#> 189 505.8890   2.893467
#> 190 501.4271   3.050493
#> 191 750.2105   2.627832
#> 192 550.8861   2.807952
#> 193 500.0287   2.664760
#> 194 504.6608   2.899706
#> 195 503.3691   3.646221
#> 196 508.3797   2.765250
#> 197 501.2569   2.935134
#> 198 500.0619   4.163144
#> 199 831.6994   2.604259
#> 200 505.4669   2.674713
#> 201 501.4579   2.694778
#> 202 500.4112   4.162172
#> 203 501.0240   2.753731
#> 204 620.4843   2.573188
#> 205 510.3983   3.041942
#> 206 500.0063   3.139836
#> 207 646.9685   2.664097
#> 208 500.3053   3.651473
#> 209 500.1044   2.571694
#> 210 501.5655   2.996857
#> 211 595.3877   3.048990
#> 212 580.8845   4.055246
#> 213 508.3826   3.135580
#> 214 658.7078   2.622632
#> 215 538.6816   2.974267
#> 216 506.7794   3.266579
#> 217 506.1773   3.866827
#> 218 501.3964   2.745371
#> 219 647.2369   2.683812
#> 220 500.3113   3.033421
#> 221 545.7064   2.587468
#> 222 531.8465   2.693284
#> 223 500.9547   2.924283
#> 224 502.7311   2.696316
#> 225 500.2341   2.512714
#> 226 545.8966   2.547465
#> 227 531.7225   2.890750
#> 228 503.2852   2.835128
#> 229 500.4602   2.521403
#> 230 500.2518   2.535738
#> 231 501.5674   3.183848
#> 232 500.2308   3.896704
#> 233 500.0710   3.567945
#> 234 500.4331   2.968305
#> 235 530.7153   2.853052
#> 236 500.0459   3.917904
#> 237 574.9032   4.455052
#> 238 500.1755   3.169148
#> 239 553.9544   2.715531
#> 240 506.8559   3.701032
#> 241 508.5093   3.165552
#> 242 504.7081   2.796098
#> 243 503.3034   3.096147
#> 244 509.2662   3.748115
#> 245 506.4226   3.021727
#> 246 500.0578   3.647980
#> 247 593.2003   2.836850
#> 248 505.8844   2.876117
#> 249 506.3057   2.680538
#> 250 530.5462   2.702505
#> 251 500.1432   2.612505
#> 252 502.6651   3.312583
#> 253 760.7159   3.433803
#> 254 500.3205   3.173166
#> 255 619.5636   2.580891
#> 256 570.5999   3.385022
#> 257 514.7466   2.747817
#> 258 500.0461   3.452601
#> 259 524.9954   2.993046
#> 260 500.0330   2.786464
#> 261 556.5555   2.933634
#> 262 500.6579   2.560190
#> 263 500.2014   3.137009
#> 264 500.0017   2.500802
#> 265 500.4731   2.558538
#> 266 500.0604   2.507956
#> 267 580.4000   3.100059
#> 268 500.5412   2.585799
#> 269 523.5995   2.640797
#> 270 926.5920   2.598129
#> 271 506.9074   2.703244
#> 272 501.0925   2.531166
#> 273 500.4142   3.350208
#> 274 507.8010   3.012806
#> 275 500.1146   2.957038
#> 276 508.0422   4.591027
#> 277 553.9816   3.882462
#> 278 516.6605   2.587920
#> 279 512.2715   2.749889
#> 280 540.4516   3.110152
#> 281 559.3900   2.725380
#> 282 500.6620   2.856227
#> 283 505.2601   2.576938
#> 284 727.2723   2.673443
#> 285 500.5918   3.372023
#> 286 613.2330   2.801618
#> 287 500.5619   2.703206
#> 288 507.6437   2.693144
#> 289 500.5674   2.877556
#> 290 652.5930   2.733916
#> 291 664.0534   2.986668
#> 292 500.3983   2.511235
#> 293 501.3044   3.087116
#> 294 503.9285   2.603144
#> 295 501.0454   2.986634
#> 296 501.2020   2.550106
#> 297 500.0342   2.653454
#> 298 520.6503   2.500636
#> 299 574.0589   3.387329
#> 300 667.5386   3.356186
#> 301 520.5117   2.788943
#> 302 500.0154   3.755199
#> 303 503.1260   3.669608
#> 304 500.3345   2.503280
#> 305 510.8156   2.653517
#> 306 500.1956   2.954449
#> 307 500.7446   3.589084
#> 308 500.7287   2.722968
#> 309 500.5778   3.105072
#> 310 534.2621   2.600850
#> 311 500.4248   2.517935
#> 312 504.1246   3.016751
#> 313 585.0614   2.573710
#> 314 504.7918   2.512841
#> 315 549.7202   3.374774
#> 316 509.1438   2.542150
#> 317 500.1959   3.842629
#> 318 545.2980   3.125518
#> 319 500.5018   2.647575
#> 320 505.9403   4.110985
#> 321 519.3103   3.537671
#> 322 500.3637   2.526553
#> 323 521.7779   3.277566
#> 324 503.2533   2.890103
#> 325 557.0180   2.755509
#> 326 596.7924   3.008392
#> 327 500.0477   2.913713
#> 328 502.6479   2.857514
#> 329 505.4139   2.824953
#> 330 505.5730   2.746956
#> 331 527.7944   2.587424
#> 332 502.8540   2.610486
#> 333 504.4323   2.715255
#> 334 524.1255   3.378028
#> 335 500.7429   3.031576
#> 336 514.0084   2.797313
#> 337 512.7867   3.319386
#> 338 513.7899   2.512022
#> 339 570.4073   2.572041
#> 340 568.8095   2.746479
#> 341 608.0399   2.710278
#> 342 500.0223   2.876269
#> 343 500.0556   2.752436
#> 344 512.4638   2.621690
#> 345 919.2393   2.858400
#> 346 512.9155   2.525459
#> 347 500.3382   2.645255
#> 348 560.7649   2.712852
#> 349 881.6562   2.620719
#> 350 617.2197   2.826452
#> 351 503.7156   2.775688
#> 352 674.6840   2.861382
#> 353 500.5655   2.668295
#> 354 558.5941   2.765672
#> 355 500.1382   3.201721
#> 356 503.1103   2.554469
#> 357 501.2754   2.862377
#> 358 525.4132   3.260554
#> 359 680.5986   2.938246
#> 360 524.0317   3.047735
#> 361 501.3396   2.728364
#> 362 508.3179   2.529709
#> 363 510.2565   2.653560
#> 364 500.8972   2.827451
#> 365 502.8582   2.578287
#> 366 518.1064   2.913415
#> 367 688.4261   3.807740
#> 368 500.0680   2.500318
#> 369 500.0450   2.510023
#> 370 500.9948   2.953089
#> 371 502.8450   2.841622
#> 372 501.6947   2.816107
#> 373 669.9958   2.669850
#> 374 500.8259   2.666173
#> 375 500.0880   2.701622
#> 376 500.1658   2.872259
#> 377 506.3913   2.528072
#> 378 535.6650   2.846835
#> 379 501.0481   3.014713
#> 380 500.0333   3.662848
#> 381 501.4261   2.885157
#> 382 504.3418   2.506894
#> 383 501.4221   3.593296
#> 384 509.4730   2.927406
#> 385 537.7948   3.713617
#> 386 626.6777   2.982188
#> 387 502.4126   2.747505
#> 388 500.9692   2.767530
#> 389 501.2285   3.079618
#> 390 507.1011   2.649633
#> 391 509.1872   3.653579
#> 392 511.0565   2.930467
#> 393 500.0857   2.750830
#> 394 513.6530   2.779635
#> 395 505.8284   3.379999
#> 396 500.0610   2.956936
#> 397 500.0069   2.647825
#> 398 780.5001   2.540002
#> 399 501.4659   4.117396
#> 400 728.9297   2.985585
#> 401 504.9773   2.948013
#> 402 500.0009   2.505041
#> 403 504.3409   2.644689
#> 404 500.2372   3.176981
#> 405 502.9089   2.519787
#> 406 500.0818   2.576830
#> 407 634.8547   2.737042
#> 408 714.5492   2.735751
#> 409 500.6189   2.809499
#> 410 623.5251   2.607447
#> 411 500.7254   2.751787
#> 412 529.0805   2.618868
#> 413 504.2855   2.507095
#> 414 523.8322   2.596383
#> 415 826.4350   3.253780
#> 416 519.2448   3.318754
#> 417 536.2473   2.732149
#> 418 502.7658   3.008141
#> 419 503.9836   2.638270
#> 420 500.0962   2.614245
#> 421 882.6240   3.015393
#> 422 544.6658   3.294045
#> 423 500.0532   3.499098
#> 424 780.1423   3.786203
#> 425 500.6717   2.683024
#> 426 500.0208   3.165313
#> 427 500.3449   2.665321
#> 428 546.8098   2.561179
#> 429 500.0246   2.543332
#> 430 941.9744   3.035184
#> 431 500.6764   2.922448
#> 432 582.7252   3.183201
#> 433 500.2295   3.092151
#> 434 502.6046   3.166107
#> 435 506.1002   3.865595
#> 436 500.5697   3.542154
#> 437 500.0085   2.779299
#> 438 551.7065   2.683066
#> 439 544.9427   2.509382
#> 440 768.9548   4.201775
#> 441 500.5114   3.158265
#> 442 500.0466   2.698487
#> 443 546.7406   2.738140
#> 444 509.0171   3.393598
#> 445 501.3838   2.515528
#> 446 504.0623   3.558348
#> 447 754.4747   3.231693
#> 448 721.3554   2.779162
#> 449 584.4608   2.980386
#> 450 500.1458   3.044115
#> 451 503.1328   2.717699
#> 452 501.2462   2.608355
#> 453 509.6203   2.693930
#> 454 532.8721   2.571885
#> 455 520.7342   2.563215
#> 456 504.5753   3.174452
#> 457 501.4186   3.196484
#> 458 507.0280   2.586621
#> 459 500.5788   2.538717
#> 460 513.5843   3.645502
#> 461 838.9829   2.876305
#> 462 502.6384   2.611726
#> 463 684.6636   2.500861
#> 464 503.5097   2.698404
#> 465 587.2548   3.160870
#> 466 500.1937   3.145723
#> 467 500.2197   2.606638
#> 468 501.2589   2.639747
#> 469 500.0132   2.822187
#> 470 500.1634   2.586775
#> 471 500.6587   3.670285
#> 472 500.3466   2.921128
#> 473 500.7691   3.893921
#> 474 603.2179   2.834851
#> 475 504.3195   3.397162
#> 476 617.1595   3.056998
#> 477 500.0254   2.519732
#> 478 500.5861   2.646219
#> 479 501.4342   2.517549
#> 480 500.4736   2.759974
#> 481 555.6735   2.833864
#> 482 500.3496   2.935198
#> 483 723.2730   2.760208
#> 484 500.6180   3.795559
#> 485 503.8679   2.583578
#> 486 502.9298   2.717126
#> 487 548.9471   2.547910
#> 488 511.1993   2.638411
#> 489 500.3105   2.602734
#> 490 500.0888   2.646167
#> 491 500.0623   2.955243
#> 492 503.2205   2.664556
#> 493 500.2808   2.908132
#> 494 836.2568   2.744470
#> 495 500.7090   2.528426
#> 496 503.4996   2.856212
#> 497 805.9792   3.142967
#> 498 500.3014   2.548344
#> 499 500.2673   2.553035
#> 500 526.4579   2.797545
#> 501 506.8760   3.114150
#> 502 512.6385   2.903473
#> 503 500.7157   3.033698
#> 504 500.0016   2.765251
#> 505 614.6656   3.321813
#> 506 500.7407   2.893213
#> 507 628.5908   2.940543
#> 508 625.4836   2.535939
#> 509 500.2777   2.508432
#> 510 566.8974   2.572293
#> 511 501.6268   2.973886
#> 512 506.4006   3.118579
#> 513 701.4642   2.967690
#> 514 500.4555   2.594136
#> 515 509.5567   3.065034
#> 516 500.0777   2.661013
#> 517 501.5241   2.704928
#> 518 500.5826   2.824858
#> 519 531.5295   2.909306
#> 520 500.0673   2.982265
#> 521 536.5921   2.507244
#> 522 954.0549   2.551112
#> 523 500.5905   2.772570
#> 524 500.9147   2.651417
#> 525 500.0127   2.788476
#> 526 716.5376   2.597724
#> 527 508.7732   2.531214
#> 528 500.5238   3.562368
#> 529 500.7747   2.500266
#> 530 502.8845   2.734621
#> 531 500.1323   2.706824
#> 532 500.0131   2.619952
#> 533 605.7994   4.566206
#> 534 661.9626   2.676655
#> 535 686.5707   3.401334
#> 536 500.1310   2.608826
#> 537 553.3297   2.821481
#> 538 523.3868   3.023782
#> 539 546.9937   2.552743
#> 540 500.7089   3.005545
#> 541 505.4960   3.078197
#> 542 500.2519   2.680751
#> 543 500.0914   2.944665
#> 544 500.9653   2.687688
#> 545 534.8958   3.779671
#> 546 501.6709   2.584003
#> 547 509.8001   2.629952
#> 548 630.7055   2.761251
#> 549 539.7158   2.687372
#> 550 599.7156   3.095893
#> 551 577.2486   2.764458
#> 552 501.2795   2.666145
#> 553 500.0253   2.659268
#> 554 505.2095   3.566049
#> 555 500.4101   2.708662
#> 556 500.4895   2.964636
#> 557 500.2700   2.688552
#> 558 717.6707   2.535009
#> 559 526.2423   2.595883
#> 560 500.0645   2.691732
#> 561 504.9918   2.552017
#> 562 500.5273   2.549239
#> 563 636.8841   2.687154
#> 564 513.9786   3.347984
#> 565 689.1690   2.568718
#> 566 501.8010   2.653060
#> 567 500.5463   3.897157
#> 568 514.2256   3.542730
#> 569 500.2067   2.907732
#> 570 580.7533   2.747526
#> 571 544.4698   2.627321
#> 572 500.0085   2.577565
#> 573 539.5513   2.763757
#> 574 505.5850   3.030761
#> 575 500.0812   4.084279
#> 576 507.0841   2.554328
#> 577 547.8616   2.611448
#> 578 563.0728   2.910812
#> 579 510.4727   3.852974
#> 580 500.2061   2.878643
#> 581 500.0776   3.885181
#> 582 706.5343   2.549750
#> 583 511.1242   2.614019
#> 584 838.9295   3.351448
#> 585 572.8731   2.516347
#> 586 510.4471   2.544226
#> 587 508.6629   3.715637
#> 588 501.0572   2.764879
#> 589 501.0602   2.904203
#> 590 783.3279   2.774488
#> 591 501.3446   3.189750
#> 592 500.3091   3.523547
#> 593 500.1081   3.333791
#> 594 501.2979   2.786026
#> 595 532.8897   3.237354
#> 596 511.4869   3.696172
#> 597 805.9310   2.677596
#> 598 500.1345   2.836011
#> 599 506.1720   2.754177
#> 600 500.0965   2.507704
#> 601 590.1806   2.724867
#> 602 692.3528   2.610762
#> 603 500.3081   2.623712
#> 604 547.2829   3.621466
#> 605 559.9517   3.766439
#> 606 500.8006   4.191982
#> 607 500.4437   2.642919
#> 608 500.0093   2.534446
#> 609 500.1034   3.423614
#> 610 504.7068   2.567776
#> 611 500.9890   2.536184
#> 612 503.8529   2.533468
#> 613 506.3467   2.728281
#> 614 500.4746   2.788780
#> 615 500.9197   2.883359
#> 616 503.7347   3.584542
#> 617 500.2401   2.584169
#> 618 501.2927   3.240176
#> 619 500.4237   3.412126
#> 620 500.4378   3.154166
#> 621 518.4589   2.702724
#> 622 732.6064   2.938127
#> 623 622.7009   2.590627
#> 624 500.5151   2.604525
#> 625 637.3112   2.808462
#> 626 500.3015   2.521683
#> 627 525.6247   3.101894
#> 628 765.2274   3.461382
#> 629 585.0759   2.520788
#> 630 500.2764   3.155980
#> 631 501.7961   2.647576
#> 632 500.2108   2.845330
#> 633 573.6853   2.676775
#> 634 531.5648   2.646311
#> 635 500.3332   2.711585
#> 636 500.2151   2.568276
#> 637 843.7693   2.541370
#> 638 500.0540   2.650274
#> 639 538.6082   3.222060
#> 640 501.0876   2.708190
#> 641 501.4499   2.747214
#> 642 504.4847   3.356844
#> 643 560.4462   2.917754
#> 644 504.3563   2.595410
#> 645 704.2734   2.845407
#> 646 978.7527   2.505609
#> 647 502.2701   2.581528
#> 648 511.9424   2.651800
#> 649 551.9580   2.628666
#> 650 501.0618   2.759299
#> 651 512.5319   2.507480
#> 652 500.0160   2.614030
#> 653 691.5753   2.722631
#> 654 505.0516   3.519787
#> 655 502.0115   2.586174
#> 656 609.6325   2.845581
#> 657 501.2083   2.869168
#> 658 505.3503   2.657142
#> 659 500.9769   2.727222
#> 660 501.0745   2.784135
#> 661 502.1776   2.866579
#> 662 624.9305   2.740782
#> 663 500.4803   2.534302
#> 664 501.0162   2.719790
#> 665 500.1543   2.505849
#> 666 504.7094   3.780974
#> 667 546.7618   3.432994
#> 668 500.5659   2.612704
#> 669 500.6692   2.690696
#> 670 501.1627   2.628473
#> 671 508.1326   3.042545
#> 672 500.4969   2.990485
#> 673 500.0678   2.622989
#> 674 501.7825   3.036800
#> 675 500.8484   3.192698
#> 676 591.0722   2.749026
#> 677 506.6103   2.639764
#> 678 501.1666   2.968105
#> 679 500.7355   2.562945
#> 680 500.2147   2.731736
#> 681 515.5513   2.758951
#> 682 520.0069   2.523314
#> 683 503.3577   3.276863
#> 684 537.9474   4.005114
#> 685 500.0406   2.707432
#> 686 500.2471   2.886446
#> 687 500.3167   3.399039
#> 688 519.7754   2.936126
#> 689 501.7615   2.586182
#> 690 545.1943   4.146289
#> 691 504.9898   3.110412
#> 692 521.4052   3.016299
#> 693 500.0526   2.851686
#> 694 500.3650   2.883663
#> 695 522.2021   2.532882
#> 696 500.1624   2.883237
#> 697 500.0315   2.943816
#> 698 500.5087   3.283193
#> 699 500.0212   2.715874
#> 700 500.7161   2.606702
#> 701 500.0345   3.137508
#> 702 500.1759   3.347519
#> 703 500.8890   2.759194
#> 704 502.9817   2.665628
#> 705 500.7351   2.500707
#> 706 568.6571   2.777145
#> 707 510.8403   2.765374
#> 708 500.1046   2.500299
#> 709 500.2099   3.218308
#> 710 772.5443   2.761094
#> 711 515.8424   2.911110
#> 712 519.4289   2.597860
#> 713 525.1456   2.523839
#> 714 500.7346   4.407878
#> 715 506.4153   3.068556
#> 716 741.3224   2.945852
#> 717 664.9061   3.173209
#> 718 501.6068   3.915631
#> 719 503.2968   2.566968
#> 720 500.5622   2.863476
#> 721 705.9217   2.573330
#> 722 501.0472   2.710057
#> 723 582.0248   3.003850
#> 724 500.6681   2.730444
#> 725 500.1625   3.223644
#> 726 570.6787   2.806964
#> 727 502.4148   2.825846
#> 728 500.5141   2.756838
#> 729 516.1918   2.864927
#> 730 500.0199   2.529297
#> 731 500.1955   3.140418
#> 732 609.5374   2.735643
#> 733 505.1069   2.692849
#> 734 656.6111   2.544741
#> 735 595.1060   3.513832
#> 736 500.6276   2.853531
#> 737 501.4030   3.494748
#> 738 657.2825   2.988834
#> 739 500.1660   2.665985
#> 740 502.0998   2.678886
#> 741 500.2386   2.511615
#> 742 565.8782   2.777054
#> 743 500.0790   3.033027
#> 744 510.0082   2.916036
#> 745 500.1768   2.619928
#> 746 597.8511   2.773487
#> 747 500.0649   2.992516
#> 748 501.0587   2.644089
#> 749 500.1436   2.563809
#> 750 515.9662   2.694327
#> 751 501.8194   2.825095
#> 752 500.0108   2.674786
#> 753 505.7435   2.621280
#> 754 500.8668   2.542153
#> 755 501.7955   3.257364
#> 756 551.8339   3.855967
#> 757 508.8650   3.172971
#> 758 500.4624   2.700597
#> 759 540.9532   3.223469
#> 760 793.0851   4.000110
#> 761 536.6775   3.207357
#> 762 501.2586   3.250648
#> 763 500.0903   2.604546
#> 764 504.1460   2.708525
#> 765 500.0487   3.010192
#> 766 501.6327   2.629615
#> 767 898.4507   3.176267
#> 768 530.8770   2.538429
#> 769 562.5612   2.719592
#> 770 508.0399   2.706377
#> 771 553.3339   2.749056
#> 772 500.0545   3.192384
#> 773 508.3547   2.971367
#> 774 562.6271   2.610115
#> 775 649.9422   3.448149
#> 776 500.1916   2.622460
#> 777 500.0110   2.750556
#> 778 515.1394   2.666938
#> 779 500.2811   3.579364
#> 780 500.1026   2.987024
#> 781 504.8516   3.558496
#> 782 502.5167   2.859529
#> 783 501.0779   2.518585
#> 784 534.5351   4.236018
#> 785 502.1313   3.099278
#> 786 676.8646   3.309505
#> 787 500.5507   3.008546
#> 788 660.2677   2.629752
#> 789 501.7125   2.840299
#> 790 500.2244   3.690315
#> 791 533.0190   2.637869
#> 792 520.6926   2.990314
#> 793 887.5639   3.006361
#> 794 500.5646   3.858062
#> 795 526.3441   3.112165
#> 796 500.5699   2.782368
#> 797 515.7453   2.755005
#> 798 504.7598   2.956868
#> 799 884.6984   2.819861
#> 800 502.2050   3.264584
#> 801 500.1557   3.661517
#> 802 537.3029   2.548081
#> 803 512.8154   2.524353
#> 804 501.2859   2.791313
#> 805 567.6169   2.927045
#> 806 500.0271   4.431179
#> 807 598.8919   2.594735
#> 808 506.0293   2.556545
#> 809 502.2753   3.378337
#> 810 505.5890   2.634249
#> 811 500.1475   2.727542
#> 812 547.4561   2.825702
#> 813 500.0821   2.867490
#> 814 500.2385   3.080445
#> 815 512.0148   2.587194
#> 816 503.7814   2.555804
#> 817 500.2250   3.094739
#> 818 502.1064   2.778461
#> 819 500.4123   2.557458
#> 820 505.8514   2.710999
#> 821 787.8333   2.804254
#> 822 500.2180   2.540998
#> 823 500.3000   2.576794
#> 824 514.5097   2.604443
#> 825 500.3201   3.704516
#> 826 503.4048   2.965689
#> 827 500.0203   3.000390
#> 828 821.3850   2.999062
#> 829 870.3343   2.579211
#> 830 512.9195   3.055944
#> 831 508.0753   3.062517
#> 832 501.8002   3.868288
#> 833 503.6246   2.603057
#> 834 583.2956   3.128958
#> 835 500.2909   2.839252
#> 836 502.8566   2.684773
#> 837 572.3445   2.863016
#> 838 500.2579   3.918607
#> 839 506.7956   3.037360
#> 840 571.6353   2.611823
#> 841 500.0001   2.867251
#> 842 503.9123   2.769217
#> 843 692.7508   2.976870
#> 844 791.3786   3.340282
#> 845 526.6993   3.154895
#> 846 690.9388   3.907819
#> 847 503.6175   2.767720
#> 848 500.0724   2.536076
#> 849 535.5728   2.860395
#> 850 500.0667   2.545929
#> 851 503.7528   3.030451
#> 852 500.6131   3.292931
#> 853 712.4128   3.513516
#> 854 500.7514   3.297836
#> 855 500.1410   2.558465
#> 856 503.3357   3.170726
#> 857 608.9527   4.180683
#> 858 507.4866   2.756444
#> 859 520.5437   2.573660
#> 860 501.0111   2.610529
#> 861 546.7574   2.614482
#> 862 501.1241   3.015094
#> 863 502.4033   2.769930
#> 864 500.0169   4.908887
#> 865 636.8039   2.550467
#> 866 642.2685   3.007551
#> 867 601.5656   3.540795
#> 868 500.2776   2.596579
#> 869 815.7571   2.768388
#> 870 500.0171   2.510979
#> 871 630.4733   3.179657
#> 872 537.5609   3.417155
#> 873 500.5265   3.687914
#> 874 500.5946   2.528846
#> 875 879.1474   2.686387
#> 876 809.8637   2.775787
#> 877 518.7687   2.912526
#> 878 504.7613   2.614359
#> 879 508.4200   3.046009
#> 880 500.1779   2.703143
#> 881 657.7053   2.584435
#> 882 503.7142   4.059824
#> 883 500.3722   2.803716
#> 884 509.2028   2.576854
#> 885 631.7004   2.869045
#> 886 500.7665   3.127808
#> 887 508.5991   3.051118
#> 888 593.9552   3.013905
#> 889 500.4408   4.165105
#> 890 559.9050   2.589295
#> 891 507.8912   3.293942
#> 892 707.5549   2.574450
#> 893 502.7623   3.249047
#> 894 500.3542   3.107017
#> 895 507.8316   2.800601
#> 896 502.9581   3.091935
#> 897 500.6454   2.739502
#> 898 500.2017   2.593926
#> 899 500.0126   2.652415
#> 900 501.2363   4.181065
#> 901 500.6820   3.206326
#> 902 533.7296   2.813317
#> 903 502.7566   2.776028
#> 904 615.9794   2.645072
#> 905 500.1736   2.614771
#> 906 501.3402   2.934960
#> 907 500.0762   2.819225
#> 908 557.5220   3.424666
#> 909 810.7328   3.499746
#> 910 500.4064   3.418232
#> 911 964.0447   3.256235
#> 912 536.0875   2.521922
#> 913 502.5170   2.717864